% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activateAssay}
\alias{activateAssay}
\alias{activeAssay}
\title{Default assay}
\usage{
activateAssay(object, assay_name, verbose = NULL)

activeAssay(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\code{activateAssay()}: Updated \code{SPATA2} object.
\code{activeAssay()}: Character value. Name of the default assay.
}
\description{
Sets and extracts the active (default) assay. Only relevant if the
\code{SPATA2} object contains more than one molecular assay.
}
\seealso{
\code{\link{MolecularAssay}}
}
