% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activateGrouping}
\alias{activateGrouping}
\alias{activeGrouping}
\title{Default grouping}
\usage{
activateGrouping(object, grouping, verbose = NULL)

activeGrouping(object, verbose = NULL, arg = "across")
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable that is
supposed to be used by default within all functions that need one.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\code{activateGrouping()}: Updated \code{SPATA2} object.
\code{activeGrouping()}: Character value. Name of the default grouping variable.
}
\description{
Sets and extracts the active (default) grouping. Useful to save typing
in functions that require a grouping variable as input. (Usually referred to
via arguments \code{across} or \code{grouping} / \code{grouping_variable}).
}
