% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{activateHistoImage}
\alias{activateHistoImage}
\alias{activateHistoImage,HistoImaging-method}
\title{Activate \code{HistoImage}}
\usage{
activateHistoImage(object, ...)

\S4method{activateHistoImage}{HistoImaging}(object, name, empty = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{empty}{Logical value. If \code{TRUE}, ensures that @image slots of
the inactive registered images are empty to prevent the \code{HistoImaging}
object from becoming too big.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Makes a \code{HistoImage} active while deactivating the previously
active \code{HistoImage}.
}
