% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activateImage}
\alias{activateImage}
\alias{activateImage,SPATA2-method}
\alias{activateImage,SpatialData-method}
\alias{activateImageInt}
\alias{activateImageInt,SPATA2-method}
\alias{activateImageInt,SpatialData-method}
\title{Activate an image}
\usage{
activateImage(object, ...)

\S4method{activateImage}{SPATA2}(
  object,
  img_name,
  load = TRUE,
  unload = TRUE,
  verbose = TRUE,
  ...
)

\S4method{activateImage}{SpatialData}(
  object,
  img_name,
  load = TRUE,
  unload = TRUE,
  verbose = TRUE,
  ...
)

activateImageInt(object, ...)

\S4method{activateImageInt}{SPATA2}(object, img_name, load = FALSE)

\S4method{activateImageInt}{SpatialData}(object, img_name, load = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{unload}{Logical value. If \code{TRUE}, ensures that @image slots of
the inactive registered images are empty to prevent the input object
from becoming too big.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
Sets the active image of the input object which is
then used by default in image dependent functions.
}
\note{
\code{activateImageInt()} exists mainly for internal use. It works
the same way \code{activateImage()} works but never unloads and is always
silent.
}
\seealso{
\code{\link[=activeImage]{activeImage()}}
}
