% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activateMatrix}
\alias{activateMatrix}
\alias{activeMatrix}
\title{Default matrix name}
\usage{
activateMatrix(
  object,
  mtr_name,
  assay_name = activeAssay(object),
  verbose = NULL
)

activeMatrix(object, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{Character value. The name of the matrix to activate as
the default matrix.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\itemize{
\item{activateMatrix()}: Updated \code{SPATA2} object.
\item{activeMatrix()}: Character value. Name of the currently active matrix in the respective assay.
}
}
\description{
Sets and extracts the active (default) matrix name of a \code{\link{MolecularAssay}}.
}
\examples{

library(SPATA2)
library(ggplot2)

data("example_data")

object <- example_data$object_UKF275T_diet

object <- normalizeCounts(object, mtr_name = "LogNormalize")

p1 <- plotSurface(object, color_by = "METRN") + labs(subtitle = activeMatrix(object))

object <- activateMatrix(object, mtr_name = "LogNormalize")

p2 <- plotSurface(object, color_by = "METRN") + labs(subtitle = activeMatrix(object))

plot(p1)
plot(p2)

}
\seealso{
\code{\link[=getMatrix]{getMatrix()}}, \code{\link[=getMatrixNames]{getMatrixNames()}}
}
