% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activeImage}
\alias{activeImage}
\alias{activeImage,SPATA2-method}
\alias{activeImage,SpatialData-method}
\title{Obtain name of active content}
\usage{
activeImage(object, ...)

\S4method{activeImage}{SPATA2}(object)

\S4method{activeImage}{SpatialData}(object)
}
\arguments{
\item{object}{An object that contains activated aspects such as
assays, images and matrices.}
}
\value{
Character value.
}
\description{
Handy functions to quickly access the name of currently
activated content.
}
\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

getImageNames(object)

activeImage(object)
plotImage(object)

getCoordsDf(object) # with image1

object <- activateImage(object, img_name = "very_low_res")
activeImage(object)
plotImage(object)

# note how x_orig, y_orig remain the same, but x and y differ
getCoordsDf(object) # with very_low_res

}
