% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_dbscan_variable}
\alias{add_dbscan_variable}
\alias{add_tissue_section_variable}
\title{Add tissue section variable to coordinate data.frame}
\usage{
add_dbscan_variable(
  coords_df,
  eps,
  minPts = 3,
  name = "dbscan",
  x = "x",
  y = "y",
  min_cluster_size = 1,
  ...
)

add_tissue_section_variable(coords_df, ccd, minPts = 3, ...)
}
\arguments{
\item{coords_df}{Data.frame with \emph{x} and \emph{y} variable.}

\item{eps}{Distance measure. Given to \code{eps} of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines
the size (radius) of the epsilon neighborhood.}

\item{minPts}{Numeric value. Given to \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines the
number of minimum points required in the eps neighborhood for core points
(including the point itself)}

\item{name}{Name of the added variable.}

\item{...}{To silently drop deprecated arguments.}
}
\value{
Data.frame with additional variable containing numbers. 0 means
that the spot is not connected to any other spot (probably artefact). 1-n
corresponds to the tissue sections.
}
\description{
Leverages \code{dbscan::dbscan()} to identify tissue sections
on the slide and to group barcode spots accordingly. Required to approximate
the outline of the tissue section(s).
}
\note{
\code{add_dbscan_variable()} is the working horse. \code{add_tissue_section_variable()}
has specific defaults.
}
\examples{

library(SPATA2)
library(EBImage)
library(tidyverse)

data("example_data")

object <- example_data$object_LMU_MCI_diet

coords_df <-
 getCoordsDf(object) \%>\%
 add_dbscan_variable(eps = recDbscanEps(object), name = "section")

plotSurface(coords_df, color_by = "section")

}
