% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_grid_variable}
\alias{add_grid_variable}
\title{Add Grid Variable to Coordinate Data Frame}
\usage{
add_grid_variable(coords_df, nr, grid_name = "grid", keep_temp = FALSE)
}
\arguments{
\item{coords_df}{A data frame containing x and y coordinates.}

\item{nr}{The number of grid cells required. The square root of this number determines
the number of breaks for the x and y axes.}

\item{grid_name}{The name of the new grid variable to be
added to the data frame. Default is "grid".}

\item{keep_temp}{Logical. If TRUE, temporary variables
used in the process (x_bins.temp and y_bins.temp) are retained in the output data frame. Default is FALSE.}
}
\value{
A data frame with the added grid variable.
}
\description{
This function adds a grid variable to a data frame containing x and y coordinates.
The grid variable represents the grid cell in which each coordinate point falls.
}
