% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{add_models}
\alias{add_models}
\title{Title}
\usage{
add_models(
  input_df,
  var_order,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  verbose = TRUE
)
}
\arguments{
\item{input_df}{Data.frame with at least three columns. \emph{values}
contain the actual values. \emph{variables} contain the variable belonging
of the values. \emph{\code{var_order}} contains the integers from 1 to n
corresponding to the ordering of the values.}

\item{var_order}{Character value. The variable that corresponds to the order
of the values.}

\item{verbose}{}
}
\value{

}
\description{
Title
}
