% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_noise_to_model}
\alias{add_noise_to_model}
\title{Add Noise to a Model}
\usage{
add_noise_to_model(model, random, nl)
}
\arguments{
\item{model}{A numeric vector representing the original model.}

\item{random}{A numeric vector representing random noise to be added.}

\item{nl}{A numeric value specifying the noise level as a percentage.}
}
\value{
A numeric vector representing the model with added noise.
}
\description{
This function adds adjustable noise to a model while maintaining a specified
level of randomness.
}
\details{
This function combines a model and random noise while allowing control
over the degree of randomness in the resulting data. The 'nl' parameter defines
the noise level as a percentage, where a higher value adds more randomness to
the model.

The function scales the model and random noise vectors based on the specified
noise level, ensuring that the original model remains a significant component.
It then merges these scaled vectors to create the final data with the desired
noise level.
}
