% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_outline_variable}
\alias{add_outline_variable}
\title{Add outline variable}
\usage{
add_outline_variable(input_df, id_var = "barcodes")
}
\arguments{
\item{input_df}{A data.frame with two numeric variables called \emph{x} and \emph{y}
and a variable as denoted in \code{id_var}.}

\item{id_var}{Character. Variable that identifies each observation.}
}
\value{
Input data.frame with additional logical variable \emph{outline}.
}
\description{
Adds a variable called \emph{outline} to the input data.frame
that tells if the observation belongs to the points that lie on the
edge of the covered area.
}
\examples{

 library(ggplot2)

 object <- downloadPubExample("313_T")

 pt_size <- getDefault(objet, "pt_size")

 coords_df <- getCoordsDf(object)[, c("barcodes", "x", "y")]

 head(coords_df)

 ggplot(data = coords_df) +
  geom_point_fixed(mapping = aes(x = x, y = y), size = pt_size) +
  theme_void()

 coords_df2 <- add_outline_variable(coords_df, id_var = "barcodes")

 ggplot(data = coords_df2) +
  geom_point_fixed(mapping = aes(x = x, y = y, color = outline), size = pt_size) +
  theme_void()

}
