% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_wh}
\alias{add_wh}
\alias{add_xy}
\title{Add width and height or x and y}
\usage{
add_wh(df, hrange = NULL)

add_xy(df, x = "x", y = "y")
}
\arguments{
\item{df}{A data.frame that contains at least the two numeric variables
\emph{x} and \emph{y} or \emph{width} and \emph{height}.}
}
\value{
Input data.frame with two additional variables named \emph{width} and \emph{height}
or \emph{x} and \emph{y}.
}
\description{
Adds image dimensions \emph{width} and \emph{height} as variables to
a data.frame of \emph{x} and \emph{y} corodiantes or vice versa.
}
\keyword{internal}
