% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addGeneFeatures}
\alias{addGeneFeatures}
\title{Add new gene features}
\usage{
addGeneFeatures(
  object,
  gene_df,
  feature_names = NULL,
  mtr_name = NULL,
  overwrite = FALSE,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gene_df}{A data.frame that contains the variables specified by name
in the argument \code{feature_names} and the key variable \emph{genes} by
which the feature variables are joined to the already existing
gene meta data.frame.}

\item{feature_names}{Character vector or NULL. See details for more.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}

\item{overwrite}{Logical. If the specified feature names already exist in the
current spata-object this argument must be set to TRUE in order to overwrite them.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
An updated spata-object.
}
\description{
This function allows to savely add features to the
gene meta data.frame of an expression matrix of choice.
}
\details{
If you are only interested in adding specific features to the spata-object
you can specify those with the \code{feature_names}-argument. If no variables
are specified this way all variables found in the input data.frame for argument
\code{gene_df} are taken. (Apart from the key variable \emph{genes}).

Eventually the new features are joined via \code{dplyr::left_join()} over the
key-variables \emph{genes}. Additional steps secure
the joining process.
}
