% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addHistoImage}
\alias{addHistoImage}
\alias{addHistoImage,SpatialData-method}
\title{Add object of class \code{HistoImage}}
\usage{
addHistoImage(object, hist_img, ...)

\S4method{addHistoImage}{SpatialData}(object, hist_img, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{hist_img}{An object of class \code{HistoImage} created with \code{createHistoImage()}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Adds objects of class \code{HistoImage} to list of
registered histology images. Should only be used within \code{\link[=registerHistoImage]{registerHistoImage()}}.
}
\keyword{internal}
