% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{addHistologyImage}
\alias{addHistologyImage}
\alias{addHistologyImage,HistologyImagingNew-method}
\title{Add object of class \code{HistologyImage}}
\usage{
addHistologyImage(object, hist_img, ...)

\S4method{addHistologyImage}{HistologyImagingNew}(object, hist_img, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{hist_img}{An object of class \code{HistologyImage} created with \code{createHistologyImage()}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Adds objects of class \code{HistologyImage} to list of
registered histology images. Should only be used within \code{registerHistologyImage()}.
}
