% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{addImageAnnotation}
\alias{addImageAnnotation}
\title{Deprecated}
\usage{
addImageAnnotation(object, ...)
}
\description{
Deprecated in favor of \code{\link[=addSpatialAnnotation]{addSpatialAnnotation()}}.
}
\keyword{internal}
