% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addImageDir}
\alias{addImageDir}
\title{Add individual image directories}
\usage{
addImageDir(object, dir, name, check = TRUE, overwrite = FALSE, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{dir}{Character value. Directory to specific image. Should end
with either \emph{.png}, \emph{.jpeg} or \emph{.tiff}. (Capital endings work, too.)}

\item{name}{Character value. Name with which to refer to this image.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Adds specific image directories beyond \emph{lowres}
\emph{highres} and \emph{default} with a simple name.
}
