% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addInnerBorder}
\alias{addInnerBorder}
\alias{addInnerBorder,SPATA2-method}
\alias{addInnerBorder,SpatialAnnotation-method}
\title{Add holes to spatial annotations}
\usage{
addInnerBorder(object, ...)

\S4method{addInnerBorder}{SPATA2}(object, id, border_df, new_id = FALSE, overwrite = FALSE, ...)

\S4method{addInnerBorder}{SpatialAnnotation}(object, border_df, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value. The ID of the spatial annotation to which the
hole is added.}

\item{border_df}{A data.frame that contains the x- and y- positions of
the vertices of the polygon that corresponds to the borders of the
whole. See details for input requirements.}

\item{new_id}{If character value, stores the resulting spatial annotatiton
under the specified ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
These methods allow the addition of an inner border to spatial annotations, creating
holes.
}
\details{
If used on a \code{\link{SpatialAnnotation}} directly, the variables of \code{border_df} should
be called \emph{x_orig} and \emph{y_orig} and should be scaled correspondingly. If
used with the \code{\link{SPATA2}} object, the variables can be called \emph{x} and \emph{y}, too.
In that case, the function assumes that the coordinates are scaled to the
image that is currently active and creates \emph{x_orig} and \emph{y_orig} accordingly.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

create_circle_polygon <- function(p, r, n) {

 angles <- seq(0, 2 * pi, length.out = n + 1)

 x_coords <- p[1] + r * cos(angles)
 y_coords <- p[2] + r * sin(angles)

 polygon_df <- data.frame(x = x_coords, y = y_coords)

 return(polygon_df)

}

center <-
 getCoordsDf(object)[c("x", "y")] \%>\%
 map_dbl(.f = mean)

area_circle <- create_circle_polygon(center, r = 75, n = 100)

object <- addSpatialAnnotation(object, area = list(outer = area_circle), id = "circle")

plotSpatialAnnotations(object, ids = "circle")

hole <- create_circle_polygon(center, r = 25, n = 100)

object <- addInnerBorder(object, id = "circle", border_df = hole, new_id = "circle_with_hole")

plotSpatialAnnotations(object, ids = c("circle", "circle_with_hole"))


}
\seealso{
\code{\link[=activeImage]{activeImage()}}, \code{\link{SpatialAnnotation}}
}
