% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{addMetaDataMol}
\alias{addMetaDataMol}
\title{Add meta variables for molecular data}
\usage{
addMetaDataMol(
  object,
  meta_var_df,
  var_names = NULL,
  assay_name = activeAssay(object),
  na_warn = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{meta_var_df}{A data frame containing new metadata variables to be added.
This data frame must contain a column named \code{molecule} which is used as the key
for merging. Other columns should represent the metadata variables to be added.}

\item{var_names}{A character vector specifying which columns from \code{meta_var_df} should be added as metadata variables.
If \code{NULL}, all columns except \emph{molecule} and \emph{<assay_name>} will be added. Default is \code{NULL}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{na_warn}{Logical value indicating whether to issue warnings if NAs are
introduced in the new metadata variables. Default is \code{TRUE}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function adds metadata variables from a given data frame to
an object, aligning the data with
existing \link[=concept_molecular_modalites]{molecular} \link[=concept_variables]{variables}.
}
\details{
The input \code{meta_var_df} must satisfy the following requirements:
\itemize{
\item It must be a data frame.
\item It must contain a column named \code{molecule}, which will be used as the key for merging the metadata variables with the existing molecular observations.
\item Any other columns can represent the metadata variables to be added. These columns must not be named \emph{"molecule"} or the value of \emph{assay_name}.
\item The columns specified in \code{var_names} must be present in \code{meta_var_df}.
}

The function checks for the existence of the specified metadata variables in the object.
If \code{overwrite} is \code{FALSE}, it ensures that no existing variables are overwritten.
If NAs are introduced in the new metadata variables, warnings will be issued if \code{na_warn} is \code{TRUE}.
}
