% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addMolecularAssay}
\alias{addMolecularAssay}
\title{Add a molecular assay}
\usage{
addMolecularAssay(
  object,
  omic,
  active_mtr = NULL,
  count_mtr = Matrix::Matrix(),
  mtr_proc = list(),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{omic}{Character value. Should best describe the molecular type of the count matrix.
E.g. \emph{'transcriptomics'}, \emph{'proteomics'}, \emph{'metabolomics'}. Additionally, defines the
\link[=MolecularAssay]{assay} name, that is created with the count matrix.}

\item{active_mtr}{Character value. The name of the matrix chosen as
the \link[=concept_active]{active} matrix. If \code{mtr_proc} is an empty
list, this value defaults to \emph{'counts'}}

\item{count_mtr}{A count matrix. Column names correspond to the barcodes of the \link[=concept_observations]{observations}.
Rownames correspond to the names of the molecular features (genes, proteins, metabolites etc.).}

\item{mtr_proc}{A list of processed matrices set in slot @mtr_proc.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{...}{Gives access to set remaining slots of the \code{\link{MolecularAssay}}
object.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Creates and adds an object of class \code{\link{MolecularAssay}}
to the \code{\link{SPATA2}} object.
}
