% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{addPolygonBase}
\alias{addPolygonBase}
\title{Add polygons to a base plot}
\usage{
addPolygonBase(x, y, poly = NULL, color = "black", size = 2, scale_fct = 1)
}
\arguments{
\item{x, y}{Numeric vectors representing x- and y-coordinates of the
vertices of the polygon.}

\item{poly}{Data.frame of at least two variables named \emph{x} and \emph{y}
representing the coordinates of the vertices of the polygon. If
variable \emph{section} exists multiple polygons are plotted based
on the number of different groups in this variable. Overwrites \code{x}
and \code{y}.}

\item{color}{Color of the lines.}

\item{size}{Width of the lines.}

\item{scale_fct}{A factor with which the vertice positions are scaled.}
}
\value{
Output of \code{graphics::polygon()} is directly plotted.
}
\description{
Adds polygons to a base plot.
}
