% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addProcessedMatrix}
\alias{addProcessedMatrix}
\title{Add a processed matrix}
\usage{
addProcessedMatrix(
  object,
  proc_mtr,
  mtr_name,
  assay_name = activeAssay(object),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{A character value that denotes the name of the matrix with
which one can refer to it in subsequent functions via \code{mtr_name}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{expr_mtr}{A matrix in which the rownames correspond to the feature names and the
column names correspond to the barcodes.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Adds a processed matrix to the chosen molecular assay of the object.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

library(Seurat)

scaled_mtr <-
 CreateSeuratObject(getCountMatrix(object)) \%>\%
 NormalizeData() \%>\%
 ScaleData() \%>\%
 GetAssayData(layer = "scale.data")

object <- addProcessedMatrix(object, proc_mtr = scaled_mtr, mtr_name = "scaled")

p1 <- plotSurface(object, color_by = "METRN")

object <- activateMatrix(object, mtr_name = "scaled")

p2 <- plotSurface(object, color_by = "METRN")

plot(p1)
plot(p2)

}
