% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addGeneSet}
\alias{addGeneSet}
\alias{addMetaboliteSet}
\alias{addProteinSet}
\alias{addSignature}
\title{Add molecular signature}
\usage{
addGeneSet(object, class, name, genes, overwrite = FALSE, check = TRUE)

addMetaboliteSet(
  object,
  class,
  name,
  metabolites,
  overwrite = FALSE,
  check = TRUE
)

addProteinSet(object, class, name, proteins, overwrite = FALSE, check = TRUE)

addSignature(
  object,
  class,
  name,
  molecules,
  assay = activeAssay(object),
  overwrite = FALSE,
  check = TRUE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{class}{Character value. The class (heritage) of the signature. Must not
contain a \emph{_}.}

\item{name}{Character value. The actual name of the signature.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{check}{Logical value. If \code{TRUE}, the functions checkes whether the specified
molecules exist in the raw count matrix of the assay and throw an error if some
are not. Defaults to \code{TRUE}.}

\item{molecules, genes, proteins, metabolites}{Character vector. The molecules the signature consists of.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Adds a \link[=concept_molecular_signatures]{molecular signature} to the \code{SPATA2} object.
}
\details{
These functions allow to add molecular signatures to the provided object.

\itemize{
\item{\code{addSignature()}}{: Adds a signature to the assay specified in \code{assay_name}.}
\item{\code{addGeneSet()}}{:  Adds a signature to the assay with @modality = 'gene' (\code{assay_name = 'gene'}).}
\item{\code{addMetaboliteSet()}}{: Adds a signature to the assay with @modality = 'metabolite' (\code{assay_name = 'metabolite'}).}
\item{\code{addProteinSet()}}{: Adds a signature to the assay with @modality = 'protein' (\code{assay_name = 'protein'}).}
}
}
\examples{

library(SPATA2)

data("example_data")
object <- example_data$object_UKF269T_diet

genes <-
 c('CYFIP1', 'SLC16A3', 'AKAP5', 'ADCY8', 'CALB2', 'GRIN1', 'NLGN4X', 'NLGN1',
   'ITGA3', 'NLGN4Y', 'ELFN1', 'BSN', 'CNTN6', 'PDE4B', 'DGKI', 'LRRTM2', 'LRRTM1',
   'SRPX2', 'SHANK1', 'SLC17A7')

# both, opt1 and opt2, have the same effect (addSignature() just allows to specify the
# assay of interest, which is fixed to 'gene' for `adGeneSet()`).
# opt1
object <- addSignature(object, name = "EXCITATORY_SYNAPSE", class = "HM", molecules = genes)

# opt2
# object <- addGeneSet(object, name = "EXCITATORY_SYNAPSE", class = "HM", genes = genes)

gs <- "HM_EXCITATORY_SYNAPSE"

# visualize signature expression
plotSurface(object, color_by = gs)

# extract genes of signature
getGenes(object, signatures = gs)

}
