% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addSpatialTrajectory}
\alias{addSpatialTrajectory}
\title{Add a spatial trajectory with data input}
\usage{
addSpatialTrajectory(
  object,
  id,
  width = NULL,
  traj_df = NULL,
  start = NULL,
  end = NULL,
  comment = base::character(1),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The id of the spatial trajectory.}

\item{width}{Distance measure. The width of the spatial trajectory.}

\item{start, end}{Numeric vectors of length two. Can be provided instead of
\code{segment_df}. If so, \code{start} corresponds to \emph{x} and \emph{y} and \code{end} corresponds to
\emph{xend} and \emph{yend} of the segment.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{segment_df}{Data.frame with \emph{x} and \emph{y} as variables corresponding
to the vertices of the trajectory. IN case of more than three rows the
trajectory is assumed to have a curve.}

\item{vertices}{List of numeric vectors of length two or \code{NULL}. If list,
sets additional vertices along the trajectory.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Adds a \link[=SpatialTrajectory]{spatial trajectory} using coordinate input.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

object <-
   addSpatialTrajectory(
     object = object,
     id = "cross_sample",
     width = "1.5mm",
     start = c(x = "1.35mm", y = "4mm"),
     end = c(x = "6.25mm", y = "4mm"),
     overwrite = TRUE
   )

 plotSpatialTrajectories(object, ids = "cross_sample")

}
\seealso{
\code{\link[=createSpatialTrajectories]{createSpatialTrajectories()}}
}
