% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTissueOutlineBase}
\alias{addTissueOutlineBase}
\alias{addTissueOutlineBase,HistologyImage-method}
\alias{addTissueOutlineBase,data.frame-method}
\title{Add tissue outline to base surface plot}
\usage{
addTissueOutlineBase(object, ...)

\S4method{addTissueOutlineBase}{HistologyImage}(
  object,
  by_section = FALSE,
  persp = "coords",
  linealpha = 0.9,
  linecolor = "black",
  linesize = 1,
  linetype = "solid",
  ...
)

\S4method{addTissueOutlineBase}{data.frame}(
  object,
  persp = "coords",
  linealpha = 0.9,
  linecolor = "black",
  linesize = 1,
  linetype = "solid",
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{linesize}{Numeric value. The size of the line(s) plotted.}
}
\value{
Output is directly plotted.
}
\description{
Adds tissue outline layer in form of polygons to the tissue
plotted with R base plotting.
}
