% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{addVarToCoords}
\alias{addVarToCoords}
\alias{addVarToCoords,HistoImaging-method}
\title{Add variable to coordinates data.frame}
\usage{
addVarToCoords(object, ...)

\S4method{addVarToCoords}{HistoImaging}(object, var_df, vars)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{var_df}{Data.frame with the variables to merge.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Adds variables to the coordinates data.frame in slot @coordinates.
}
