% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{adjustDefaultInstructions}
\alias{adjustDefaultInstructions}
\title{Deprecated}
\usage{
adjustDefaultInstructions(...)
}
\description{
Deprecated in favor of \code{\link[=setDefault]{setDefault()}}.
}
\keyword{internal}
