% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-family.R
\name{adjustGeneSetDf}
\alias{adjustGeneSetDf}
\title{Filter gene-set data.frame}
\usage{
adjustGeneSetDf(object, limit = 50)
}
\arguments{
\item{object}{A valid spata-object.}

\item{limit}{Numeric value between 1 and 100. The minimum percentage of gene-set genes
that have to exist in the given expression matrix in order for a gene set to stay in the
gene-set data.frame.}
}
\value{
An updated spata-object and an informative message about how many
gene-sets have been discarded and how many gene-sets remain.
}
\description{
Checks the objects gene-set data.frame for gene-sets that
are composed of genes that exist in the given expression matrix.
}
\details{
E.g.: Gene-set 'x' is composed of 30 genes. The expression matrix
however contains only 15 of them. If argument \code{limit} is set to 75 gene-set 'x'
is removed since the percentage of genes of which the given expression matrix
contains information about is only 50.
}
