% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{adjustGeneSetList}
\alias{adjustGeneSetList}
\title{Adjust Gene Set List}
\usage{
adjustGeneSetList(object, limits = 50)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{limits}{A numeric value representing the threshold percentage for gene set inclusion (default: 50).}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function adjusts the gene set list (GSL) of a given object based on specified limits.
}
\details{
This function calculates the proportion of genes in each gene set relative to the total number
of genes in the object. Gene sets with a proportion greater than or equal to the specified limit
are retained, while others are removed.
}
\seealso{
\code{\link[=getGeneSetList]{getGeneSetList()}}, \code{\link[=getGenes]{getGenes()}}, \code{\link[=getAssay]{getAssay()}}, \code{\link[=setAssay]{setAssay()}}
}
