% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{alignImage}
\alias{alignImage}
\alias{alignImage,HistoImaging-method}
\alias{alignImage,HistoImage-method}
\alias{alignImageAuto}
\alias{alignImageAuto,HistoImaging-method}
\alias{alignImageInteractive}
\alias{alignImageInteractive,HistoImaging-method}
\title{Align histology images}
\usage{
alignImage(object, ...)

\S4method{alignImage}{HistoImaging}(
  object,
  name,
  opt = "set",
  angle = 0,
  flip_h = logical(),
  flip_v = logical(),
  transl_h = 0,
  transl_v = 0
)

\S4method{alignImage}{HistoImage}(
  object,
  opt = "set",
  angle = 0,
  flip_h = logical(),
  flip_v = logical(),
  transl_h = 0,
  transl_v = 0
)

alignImageAuto(object, ...)

\S4method{alignImageAuto}{HistoImaging}(
  object,
  name,
  step = 0.01,
  stop_at = 25,
  plot_progress = TRUE,
  verbose = TRUE
)

alignImageInteractive(object, ...)

\S4method{alignImageInteractive}{HistoImaging}(object, window_size = "800px")
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{step}{Numeric value specifying the accuracy of the alignment
via vertical and horizontal translation. If \code{step >= 1}, it is interpreted
as a pixel value. For example, \code{step = 2} translates the image 2 pixels to the right,
then 4 pixels to the right, and so on. If \code{step < 1}, the final step value is
calculated as \code{round(side.length * step, digits = 0)} where \code{side.length} is
equal to the height and width of the \strong{reference} image. See details for more.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{stop_after}{Numeric value specifying the maximum number of consecutive iterations
during optimization of the image translation without improvement. If \code{stop_at >= 1}, it
is interpreted as an absolute number of attempts. For instance, setting
\code{stop_after = 25} makes the function stop after 25 iterations without any improvement.
If \code{stop_at < 1}, the maximum number of consecutive iterations without any improvement
allowed is calculated by the total number of translations possible times \code{stop_at}.
See details for more.}

\item{add}{Logical value. If \code{TRUE}, numeric values are added to the current values
instead of replacing them. E.g. if \code{angle = 90} and the image is already rotated by
90째 the saved transformation would be to rotate the image with 180째. If \code{FALSE}, input
values are simply set. E.g. if \code{angle = 90} the resulting saved transformation would
be to rotate the image with 90째 regardless of the previous setting.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Aligns an image with the reference image. See details for
more information about the process.
}
\details{
\code{alignImageAuto()} aligns the image specified in \code{name} with the reference
image obtained from \code{getHistoImageRef()}.

The alignment process consists of several steps:
\enumerate{
\item Scaling and translation: The outline of the tissue in the image to be aligned
(referred to as the "tissue outline") is scaled to match the dimensions of
the reference outline. It is then translated to ensure that its centroid aligns
with the centroid of the reference outline.
\item Flipping and rotation: The function iterates over all possible combinations of
vertical and horizontal flipping, along with rotation angles between 0-359째.
Each iteration evaluates the overlap between the tissue outline and the reference
outline. The combination with the highest overlap is selected, and the tissue
outline is transformed accordingly.
\item Optimization: The overlap is further optimized through consecutive horizontal
and vertical translations of the tissue outline. The outline is shifted horizontally
by the value specified by the \code{step} argument. After each horizontal step, the
outline is shifted vertically upwards by the \code{step} value. If there is no improvement
in the overlap after a certain number of consecutive vertical shifts (controlled by
\code{stop_at}), the outline is shifted downwards. This process continues until the
maximum number of shifts without improvment is reached. Then, another step to the right
is taken until the maximum number of shifts to the right is reached. The same procedure
is conducted for shifts to the left. The optimized translation values are then
applied to the tissue outline.
\item Image transformations: All the spatial transformations required to produce the final
aligned image are stored as a list obtained from \code{getImageTransformations()}. These
transformations can be applied during data extraction or visualization if the \code{transform}
argument is set to \code{TRUE} (the default behavior).
}

The resulting aligned image and the list of transformations are returned by the function
for further use.
}
