% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{alignImageAnnotation}
\alias{alignImageAnnotation}
\alias{alignSpatialTrajectory}
\title{Align image annotation}
\usage{
alignImageAnnotation(img_ann, image_object)

alignSpatialTrajectory(spat_traj, image_object)
}
\arguments{
\item{img_ann}{An object of class \code{ImageAnnotation}.}

\item{image_object}{An object of class \code{HistologyImaging} to which the image
annotation is aligned.}
}
\value{
Aligned input for \code{img_ann}.
}
\description{
Aligns an image annotation with the current image justification.
}
\details{
Information of the current justification of the image annotation
is stored in slot @info. This function aligns justification regarding
horizontal and vertical flipping, scaling and rotation.
}
\seealso{
Read documentation on \code{?ImageAnnotation} and \code{?HistologyImaging}
for more information.
}
