% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{all_bcsp_distances}
\alias{all_bcsp_distances}
\title{Obtain a all barcode-spots distances}
\usage{
all_bcsp_distances(scale_fct = "lowres")
}
\arguments{
\item{scale_fct}{If character, \emph{'lowres'} or \emph{'hires'}. If numeric,
value of length one. Determines the factor with which \emph{imagecol} and
\emph{imagerow} of the original visium coordinates are scaled to x- and
y-coordinates.}
}
\value{
A data.frame with all possible barcode-spot pairs
and their distance to each other.
}
\description{
Obtain a all barcode-spots distances
}
