% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-functions.R
\name{all_color_palettes}
\alias{all_color_palettes}
\alias{all_color_spectra}
\title{Color palette names}
\usage{
all_color_palettes()

all_color_spectra()
}
\value{
A named list.
}
\description{
Returns all currently valid color panels or -spectra.
}
\details{
Discrete color panels derive from the ggsci-package. Continuous
colorspectra derive from the colorspace-package.
}
