% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-functions.R
\name{allColorpalettes}
\alias{allColorpalettes}
\alias{allColorspectra}
\title{Valid colorpanels & -spectra}
\usage{
allColorpalettes()

allColorspectra()
}
\description{
Valid colorpanels & -spectra
}
