% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{append_polygon_df}
\alias{append_polygon_df}
\title{Append polygon df}
\usage{
append_polygon_df(lst, plg, allow_intersect = TRUE, in_outer = TRUE, ...)
}
\arguments{
\item{lst}{Polygon list the new polygon is appended to.}

\item{plg}{New polygon data.frame.}
}
\description{
Appends df to list of polygon data.frames and names it
accordingly in case of complex polygons.
}
\keyword{internal}
