% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{are_all_area_or_dist}
\alias{are_all_area_or_dist}
\alias{are_all_dist}
\title{Test area or distance input}
\usage{
are_all_area_or_dist(input, error = FALSE)

are_all_dist(input, error = FALSE)
}
\arguments{
\item{input}{Character vector. Elements must match the requirements of
the \code{SPATA2} area framework. See details for more information.}
}
\value{
Logical vector of the same length as input and/or an error if \code{verbose}
is \code{TRUE}.
}
\description{
Tests if input can be safely converted to distance
or to area values.
}
\note{
Only returns \code{TRUE} if all values are valid distance inputs
or all values are valid area inputs.
}
\examples{
are_all_area_or_dist(c("2mm", "2cm", "3px", "4.5mm2")) # TRUE

are_all_area_or_dist(c("2mm", "2CCm", "3px", "4.5mm2")) # FALSE

are_all_area_or_dist(c("2mm", "2CCm", "3px", "4.5mm2"), error = T) # FALSE -> error

}
