% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{as_SPATA2_dist}
\alias{as_SPATA2_dist}
\title{Distance transformation}
\usage{
as_SPATA2_dist(input)
}
\arguments{
\item{input}{Character or numeric vector. Elements must match the
requirements of the \code{SPATA2} distance framework. See details
for more information.}
}
\value{
Character vector of the same length as \code{input}.
}
\description{
Ensures that distance input can be read by \code{SPATA2} functions
that convert European units of length to pixels and vice versa.
}
\details{
Several functions in \code{SPATA2} have arguments that take \emph{distance input}.
To specifically refer to a distance the unit must be specified. There are
three ways to create valid input for these arguments.

\bold{1. Distance in pixel}:

There are two valid input options to specify the distance in pixel:

\itemize{
\item{numeric:}{ Single numeric values, e.g. \code{arg_input <- c(2, 3.554, 69, 100.67)}. If no unit
is specified the input will be interpreted as pixels.}
\item{character:}{ Suffixed with \emph{'px'}, e.g. \code{arg_input <- c('2px', '3.554px', '69px', '100.67px')}}
}

\bold{2. Distance in European units of length (euol)}:

Specifying distances in European units of length e.g. \code{arg_input <- c('2mm', '4mm')} etc.
requires the input to be a character as the unit must be provided as suffix. Between the numeric
value and the unit must be no empty space! Unit suffixes must be one of
\emph{'m', 'dm', 'cm', 'mm', 'um', 'nm'}.

\strong{3. As vectors of class \code{unit}:}

Behind the scenes \code{SPATA2} works with the \code{units}-package. Input
is converted into vectors of class \code{units}. Therefore, input can be directly
provided this way: \code{arg_input <- units::set_unit(x = c(2,4), value = 'mm')}
Note that \emph{pixel} is not a valid unit in the \code{units} package. If you want
to specify the input in pixel you have to use input option 1. Distance in pixel.
}
