% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asGiotto}
\alias{asGiotto}
\title{Transform SPATA2 object to Giotto object}
\usage{
asGiotto(object, transfer_features = TRUE, verbose = NULL)
}
\arguments{
\item{object}{Objects of classes for which a method has been defined.}

\item{transfer_features, transfer_meta_data}{Logical or character. Specifies
if meta/feature, e.g clustering, data from the input object is transferred
to the output object. If TRUE, all variables of the feature/meta data.frame
are transferred. If character, named variables are transferred. If FALSE,
none are transferred.}

\item{verbose}{Logical. If TRUE, progress messages will be printed. Default is TRUE.}
}
\value{
An object of class \code{Giotto}.
}
\description{
Transforms an \code{SPATA2} object object to an object of class
\code{Giotto}. See details for more information.
}
\details{
The object is created using the count matrix of the input as
well as coordinates. If an image is found it is transferred, too. \bold{No}
further processing is done (e.g. \code{Giotto::normalizeGiotto()},
\code{Giotto::runPCA()}).
}
