% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{assessTrajectoryTrends}
\alias{assessTrajectoryTrends}
\alias{assessTrajectoryTrends2}
\alias{assessTrajectoryTrendsCustomized}
\alias{assessTrajectoryTrendsCustomized2}
\title{Trajectory trend analysis}
\usage{
assessTrajectoryTrends(
  object,
  trajectory_name,
  variables,
  binwidth = 5,
  whole_sample = FALSE,
  summarize_with = "mean",
  verbose = TRUE,
  of_sample = NA
)

assessTrajectoryTrends2(stdf, verbose = TRUE)

assessTrajectoryTrendsCustomized(
  object,
  trajectory_name,
  customized_trends,
  variables,
  binwidth = 5,
  whole_sample = FALSE,
  summarize_with = "mean",
  verbose = TRUE,
  of_sample = NA
)

assessTrajectoryTrendsCustomized2(stdf, customized_trends, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{variables}{Character vector. The variables of interest:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
 }}

\item{binwidth}{Numeric value. Specifies the accuracy with which the transcriptomic spots
are binned based on their projection length on the trajectory. Given to argument
\code{accuracy} of \code{base::floor()}.}

\item{whole_sample}{Logical. If TRUE, normalization of the values used
takes place in the light of the complete sample.}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}. Specifies the
function with which the numeric values of each variable are summarized by bin.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{stdf}{A summarized trajectory data.frame. (e.g. obtained by
\code{getTrajectoryDf()}).}

\item{customized_trends}{A data.frame or a named list. All numeric variables are considered to correspond to customized trends
the trajectory of interest might adopt. The names of the respective variables will correspond to the name
with which you want to refer to the trend later on.}
}
\value{
A data.frame arranged by the residuals area-under-the-curve-values describing
how well a model fitted the expression trend of a gene or gene set.
}
\description{
Analyzes the trend of gene and gene-set-expressions along
trajectories by fitting a variety of mathematical models to them and
by assessing the quality of each fit.

\itemize{
 \item{\code{assessTrajectoryTrends()}: Takes a valid spata-object and assembles
 the needed summarized trajectory data.frame from scratch.}
 \item{\code{assessTrajectoryTrends2()}: Takes a summarized trajectory data.frame
 returned by \code{getTrajectoryDf()}.}
 \item{\code{assessTrajectoryTrendsCustomized()}: Takes a valid spata-object as well as
 a data.frame or list of customized models against which to fit the variables. It assembles
 the needed summarized trajectory data.frame from scratch.}
 \item{\code{assessTrajectoryTrendsCustomized2()}: Takes a summarized trajectory data.frame
 returned by \code{getTrajectoryDf()} as well as a data.frame or list of customized
 models against which to fit the variables.}
 }
}
