% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSeurat}
\alias{asSeurat}
\title{Transform \code{SPATA2} to \code{Seurat}}
\usage{
asSeurat(
  object,
  process = TRUE,
  transfer_features = TRUE,
  assay_name = "Spatial",
  image_name = "slice1",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{process}{Logical value. If \code{TRUE}, count matrix is processed.
See details for more.

Use \code{getInitiationInfo()} to obtain argument input of your \code{SPATA2} object
initiation.}

\item{assay_name, image_name}{Character values. Define the name with which
to refer to the assay or the image in the \code{Seurat} object. Defaults to
the default of the \code{Seurat} package.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{Seurat}.
}
\description{
Transforms an \code{SPATA2} object to an object of class \code{Seurat}.
See details for more information.
}
\details{
If you have used \code{initiateSpataObject_10X()}, chances are that you have
already specified input for various processing functions. \code{asSeurat()}
creates a \code{Seurat} object from scratch. It has to, because even though
many processing steps are run with the Seurat object as background \code{SPATA2}
does not net all its content and to keep \code{SPATA2} objects as small as
possible not everything is transferred from the \code{Seurat} object.

If \code{process = TRUE}, the input you've given to \code{initiateSpataObject_10X()} is taken to
conduct the same processing. To check what you have defined as input, you
can use the function \code{getInititationInfo()}.
}
