% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{assignToTissueSection}
\alias{assignToTissueSection}
\alias{assignToTissueSection,spata2-method}
\alias{assignToTissueSection,HistoImaging-method}
\title{Assign observations to tissue sections}
\usage{
assignToTissueSection(object, ...)

\S4method{assignToTissueSection}{spata2}(object, buffer = 0)

\S4method{assignToTissueSection}{HistoImaging}(object, img_name = NULL, buffer = 0)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Based on the results of \code{identifyTissueOutline()} observations
like cells or spots are assigned to the tissue sections or fragments
they exist on or are labelled as artefacts.
}
\details{
The results are stored in an additional variable in the coordinates
data.frame as obtained by \code{getCoordsDf()} named \emph{section}.
}
