% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSPATA2}
\alias{asSPATA2}
\alias{asSPATA2,giotto-method}
\alias{asSPATA2,Seurat-method}
\title{Transform miscellaneous objects to SPATA2 objects}
\usage{
asSPATA2(object, ...)

\S4method{asSPATA2}{giotto}(
  object,
  sample_name,
  coordinates,
  image_ebi,
  spatial_method,
  transfer_meta_data = TRUE,
  verbose = TRUE,
  ...
)

\S4method{asSPATA2}{Seurat}(
  object,
  sample_name,
  platform = "Undefined",
  assay_name = NULL,
  image_name = NULL,
  image_dir = NULL,
  transfer_meta_data = TRUE,
  transfer_dim_red = TRUE,
  count_mtr_name = "counts",
  proc_mtr_name = "scale.data",
  modality = "gene",
  scale_with = "lowres",
  verbose = TRUE
)
}
\arguments{
\item{object}{Objects of classes for which a method has been defined.}

\item{sample_name}{Character. The name of the sample.}

\item{transfer_meta_data}{Logical. If TRUE, the metadata will be transferred. Default is TRUE.}

\item{verbose}{Logical. If TRUE, progress messages will be printed. Default is TRUE.}

\item{platform}{Character. The platform used for the experiment. Should be one
of \code{names(spatial_methods)}.}

\item{assay_name}{Character. The name of the Seurat assay containing the matrices of interest.
If NULL, Seurat's default assay is used.}

\item{image_name}{Character. The name of the image in the Seurat object to be transferred.
If NULL, the function will attempt to use the only available image or throw an error if multiple images are found.}

\item{image_dir}{Character. The directory where the image is stored. Default is NULL.}

\item{transfer_dim_red}{Logical. If TRUE, dimensionality reduction data (PCA, t-SNE, UMAP) will be transferred. Default is TRUE.}

\item{count_mtr_name}{Character. The name of the count matrix layer in the Seurat object. Default is "counts".}

\item{proc_mtr_name}{Character. The name of the processed matrix layer in the Seurat object. Default is "scale.data".}

\item{modality}{Character. The modality of the data (e.g., "gene"). Default is "gene".}

\item{scale_with}{Character. The scale factor to use for image scaling. Default is "lowres".}
}
\value{
A \code{SPATA2} object containing the converted data.
}
\description{
This S4 generic converts miscellaneous objects into a \code{SPATA2} object,
transferring relevant data and metadata.
}
\examples{

# ----- example for Seurat conversion
library(SPATA2)
library(Seurat)

seurat_object <- example_data$seurat_object

spata_object <-
 asSPATA2(
  object = seurat_object,
    sample_name = "gbm",
    image_name = "slice1",
    platform = "VisiumSmall",
   modality = "gene"
  )

# with Seurat
SpatialFeaturePlot(seurat_object, "nCounts_Spatial")

# with SPATA2
plotSurface(spata_object, "nCounts_Spatial")

}
