% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSummarizedExperiment}
\alias{asSummarizedExperiment}
\title{Transform SPATA2 object to SummarizedExperiment object}
\usage{
asSummarizedExperiment(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{SummarizedExperiment}.
}
\description{
Transforms an \code{SPATA2} object to an object of class
\code{SummarizedExperiment}. See details for more information.
}
\details{
Output object contains the count matrix in slot @assays and
feature data.frame combined with barcode-spot coordinates
in slot @colData.

Slot @metadata is a list that contains the image object in slot $image.
}
