% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-family.R
\name{asVisium}
\alias{asVisium}
\alias{asVisium,VisiumV1-method}
\title{Convert to class \code{Visium}}
\usage{
asVisium(object, ...)

\S4method{asVisium}{VisiumV1}(object)
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
An object of class \code{Visium}.
}
\description{
Coverts objects of specific classes to objects
of class \code{Visium}.
}
