% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asVisiumV1}
\alias{asVisiumV1}
\title{Transform \code{HistologyImage} to \code{VisiumV1}}
\usage{
asVisiumV1(object, name = "slice1")
}
\arguments{
\item{object}{An object of class \code{HistologyImage}.}

\item{name}{Name of the \code{VisiumV1} object. Suffixed with \emph{_} to fill
slot @key.}
}
\value{
An object of class \code{VisiumV1} from the \code{Seurat} package.
}
\description{
Transforms an \code{HistologyImage} obejct to an object of
class \code{VisiumV1} from the \code{Seurat} package.
}
