% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{average_genes}
\alias{average_genes}
\title{average_genes}
\usage{
average_genes(average_genes)
}
\arguments{
\item{average_genes}{Logical value. If set to TRUE the expression values
of all specified genes are averaged instead of considered separately.
If the output of the function is a data.frame the variable in which the results
are stored is named \emph{mean_genes}.}
}
\description{
average_genes
}
