% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{bin_by_angle}
\alias{bin_by_angle}
\title{Bin barcode-spots by angle}
\usage{
bin_by_angle(
  coords_df,
  center,
  n_bins_angle = 12,
  angle_span = c(0, 360),
  min_bins_circle = NULL,
  rename = FALSE,
  remove = FALSE,
  drop = TRUE
)
}
\arguments{
\item{coords_df}{The coordinates data.frame whose barcode-spots are supposed
to be binned.}

\item{center}{Numeric vector of length two that is named. Value named \emph{x}
provides position on the x-axis. Value named \emph{y} provides position on
the y-axis.}

\item{min_bins_circle}{Numeric value or NULL. Indiates the minimum
number of circle bins the angle bin groups must have in order not
to be renamed or removed. Ignored if NULL.}

\item{rename}{Logical value. If TRUE, barcode-spots that fall into
angle bins that feature less circle bins than input for argument
\code{min_bins_circle} are renamed to \emph{'Outside'}. Ignored if \code{min_bins_circle} is NULL.
Set \code{remove} to FALSE in order not to remove the renamed
barcode-spots.}

\item{remove}{Logical value. If TRUE, barcode-spots that fall into
angle bins that do feature less circle bins than input for
argument \code{min_bins_circle} are removed. Ignored if
\code{min_bins_circle} is NULL.}

\item{drop}{Logical value. If TRUE, unused levels of the \emph{bins_circle}
variables are dropped.}
}
\description{
Bins barcode-spots according to their angle towards the position
specified with argument \code{center}.
}
