% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{bin_by_area}
\alias{bin_by_area}
\title{Bin barcode-spots by area extension}
\usage{
bin_by_area(
  coords_df,
  area_df,
  binwidth,
  n_bins_circle,
  remove = FALSE,
  drop = TRUE
)
}
\arguments{
\item{coords_df}{The coordinates data.frame whose barcode-spots are supposed
to be binned.}

\item{area_df}{Data.frame with variables \emph{x} and \emph{y} describing the
vertices of the polygon that encircles the area based on which the barcode-spots
are binned. E.g. slot @area of \code{ImageAnnotation}-objects.}

\item{binwidth}{Numeric value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the distance
between the barcode-spots: \code{binwidth = getBarcodeSpotsDistance(object)}.
(See details for more.)}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
 (See details for more.)}

\item{remove}{Character or logical. If character, denotes circle bins that
are removed. If TRUE, bins \emph{'Core' and 'Outside'} are removed. If FALSE,
ignored.}

\item{drop}{Logical value. If TRUE, unused levels of the \emph{bins_circle}
variables are dropped.}
}
\description{
Bins barcode-spots by consecutively expanding a polygon.
}
