% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{bin_by_expansion}
\alias{bin_by_expansion}
\title{Bin barcode-spots by area extension}
\usage{
bin_by_expansion(
  coords_df,
  area_df,
  binwidth,
  n_bins_circle,
  remove = FALSE,
  bcs_exclude = NULL,
  core = TRUE,
  periphery = TRUE,
  drop = TRUE,
  arrange = TRUE,
  ...
)
}
\arguments{
\item{coords_df}{The coordinates data.frame whose barcode-spots are supposed
to be binned.}

\item{area_df}{Data.frame with variables \emph{x} and \emph{y} describing the
vertices of the polygon that encircles the area based on which the barcode-spots
are binned. E.g. slot @area of \code{SpatialAnnotation}-objects.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{remove}{Character or logical. If character, denotes circle bins that
are removed. If TRUE, bins \emph{'Core' and 'Outside'} are removed. If FALSE,
ignored.}

\item{bcs_exclude}{Character value containing name(s) of data points to be excluded from the analysis.}

\item{drop}{Logical value. If TRUE, unused levels of the \emph{bins_circle}
variables are dropped.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
Bins barcode-spots by consecutively expanding a polygon.
}
