% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-sub-functions.R
\name{binarize_matrix}
\alias{binarize_matrix}
\title{Binarize expression matrix}
\usage{
binarize_matrix(
  object,
  mtr_name = "scaled",
  genes_subset = NULL,
  method_kmeans = "Hartigan-Wong",
  kmeans = list(),
  ...,
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{genes_subset}{Chararcter vector or NULL. If character, the matrix is subsetted by
this vector such that only genes of the input are kept before the binarization process
starts.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A binarized matrix.
}
\description{
Uses kmeans algorithm of choice on every gene expression variable
with k = 2. The value 1 is assigned to observations that belong to the cluster
with the higher mean. The value 0 is assigned to observations that belong to the
cluster with the lower mean.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
