% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{binwidth_dummy}
\alias{binwidth_dummy}
\title{binwidth}
\usage{
binwidth_dummy(binwidth)
}
\arguments{
\item{binwidth}{Numeric value. Denotes the binwidth to use for the histogram.}
}
\description{
binwidth
}
