% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{ccDist}
\alias{ccDist}
\title{Obtain center to center distance}
\usage{
ccDist(object = NULL, unit = "pixel", platform = "Visium")
}
\arguments{
\item{object}{An object of class \code{SPATA2} or NULL. If NULL,
\code{unit} must not be \emph{'pixel'}.}
}
\value{
If \code{unit} is \emph{'pixel'} a numeric value that scales
the center to center distance of barcode spots to the current image.
Else an object of class \code{unit}.
}
\description{
Returns center to center distance depending on the
platform used. Output can be adjusted to the unit needed.
}
