% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{centerSpatialAnnotation}
\alias{centerSpatialAnnotation}
\alias{centerSpatialAnnotation,spata2-method}
\alias{centerSpatialAnnotation,HistoImaging-method}
\title{Center the Borders of a Spatial Annotation}
\usage{
centerSpatialAnnotation(object, ...)

\S4method{centerSpatialAnnotation}{spata2}(
  object,
  id,
  center_x,
  center_y,
  new_id = FALSE,
  overwrite = FALSE
)

\S4method{centerSpatialAnnotation}{HistoImaging}(
  object,
  id,
  center_x,
  center_y,
  new_id = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{center_x, center_y}{Distance measures. The new center of the
spatial annotation.}

\item{new_id}{Character value or \code{FALSE}. If character, the resulting
spatial annotation is stored under a new ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Shifts the borders of a spatial annotation in a way that
it's center corresponds to the input of \code{c(center_x, center_y)}.
}
\seealso{
\code{\link[=expandSpatialAnnotation]{expandSpatialAnnotation()}}, \code{\link[=shiftSpatialAnnotation]{shiftSpatialAnnotation()}},
\code{\link[=smoothSpatialAnnotation]{smoothSpatialAnnotation()}}, \code{\link{SpatialAnnotation}}
}
