% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_all_barcodes}
\alias{check_all_barcodes}
\title{Check barcodes of all matrices and data.frames}
\usage{
check_all_barcodes(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\description{
Check barcodes of all matrices and data.frames
}
