% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_assign}
\alias{check_assign}
\title{Check assign input}
\usage{
check_assign(assign = FALSE, assign_name = character(1))
}
\arguments{
\item{assign}{Logical. If set to TRUE a named list will be assigned to the global
environment. This list contains data and information to rebuild or additionally
customize the output plot of this function.}

\item{assign_name}{The name the assigned list is supposed to have specified as
a single character value.}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
