% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_availability}
\alias{check_availability}
\title{Gives feedback if object was not found}
\usage{
check_availability(test, ref_x, ref_fns)
}
\arguments{
\item{test}{Logical test that checks if something was found.}

\item{ref_x}{Object reference.}

\item{ref_fns}{Function(s) that needs}
}
\description{
Gives feedback if object was not found
}
