% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_color_to}
\alias{check_color_to}
\title{Check color to}
\usage{
check_color_to(
  color_to,
  color_by,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = NULL
)
}
\arguments{
\item{all_features}{The valid features specified as a character vector.}

\item{all_gene_sets}{The valid gene sets specified as a character vector.}

\item{all_genes}{The valid genes specified as a character vector.}
}
\description{
A member of the \code{adjusting-check_()}-family. Takes a character
vector and sorts its elements into a list depending on whether it was found in
the input of \code{all_features}, \code{all_genes} or \code{all_gene_sets}.

Returns a list with one slot named \emph{features}, \emph{genes} or \emph{gene_sets}
containing the respective found/valid input of \code{color_to}.
}
\keyword{internal}
