% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{check_compiled_trajectory_df}
\alias{check_compiled_trajectory_df}
\title{Check compiled trajectory data.frame}
\usage{
check_compiled_trajectory_df(ctdf)
}
\arguments{
\item{ctdf}{A compiled trajectory data.frame containing the variables
\emph{'barcodes', 'sample', 'x', 'y', 'projection_length', 'trajectory_part'}.}
}
\description{
Check compiled trajectory data.frame
}
\keyword{internal}
