% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_coordinate_variables}
\alias{check_coordinate_variables}
\title{Check coordinate variables}
\usage{
check_coordinate_variables(data, x = "x", y = "y")
}
\arguments{
\item{data}{A data.frame containing the variables of interest as well
as the variables needed to map onto the x- and y axis of the plot.}

\item{x}{The name of the numeric variable to be plotted on the x axis.}

\item{y}{The name of the numeric variable to be plotted on the y axis.}
}
\description{
Check coordinate variables
}
