% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_coords_df}
\alias{check_coords_df}
\title{Check coords data.frame}
\usage{
check_coords_df(coords_df)
}
\arguments{
\item{coords_df}{A data.frame containing information about every \link[=concept_observations]{observation}. Must contain the variables:
\itemize{
\item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every observation.}
\item{\emph{sample} Character. The sample belonging of every observation.}
\item{\emph{x_orig} Numeric. The unscaled x-coordinates of every observation.}
\item{\emph{y_orig} Numeric. The unscaled y-coordinates of every observation.}
}}
}
\description{
Check coords data.frame
}
\keyword{internal}
