% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_customized_trends}
\alias{check_customized_trends}
\title{Check customized trend list}
\usage{
check_customized_trends(length_trajectory, customized_trends)
}
\arguments{
\item{length_trajectory}{Numeric value. Length of trajectory according to which the trends have been
customized.}

\item{customized_trends}{A data.frame or a named list. All numeric variables are considered to correspond to customized trends
the trajectory of interest might adopt. The names of the respective variables will correspond to the name
with which you want to refer to the trend later on.}
}
\description{
Check customized trend list
}
