% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_display}
\alias{check_display}
\title{Check display input}
\usage{
check_display(display_title = FALSE, display_image = FALSE)
}
\arguments{
\item{display_title}{Logical. If set to TRUE an informative title is displayed.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}
}
\description{
Check display input
}
\keyword{internal}
