% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_feature_df}
\alias{check_feature_df}
\title{Check feature data.frame}
\usage{
check_feature_df(feature_name, feature_df)
}
\arguments{
\item{feature_name}{Character value. The name of the feature that is to be added
to the object.}

\item{feature_df}{A data.frame that contains the feature and the key-variables.}
}
\description{
Check feature data.frame
}
\keyword{internal}
