% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_features}
\alias{check_features}
\title{Check feature variables input}
\usage{
check_features(
  object,
  features,
  valid_classes = NULL,
  max_length = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{features}{The features of interest specified as a character vector.}

\item{valid_classes}{The feature-classes that are allowed.}

\item{max_length}{The maximum number of features allowed.}
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of feature names, checks which of the features exist and checks additionally
if these features match the class requirements of \code{valid_classes}.

Returns an adjusted features-vector or raises an error.
}
