% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_method}
\alias{check_method}
\title{Check method input}
\usage{
check_method(
  method_aggl = NULL,
  method_csr = NULL,
  method_de = NULL,
  method_dist = NULL,
  method_dr = NULL,
  method_gs = NULL,
  method_hclust = NULL,
  method_ovl = NULL,
  method_padj = NULL
)
}
\arguments{
\item{method_aggl}{Character value or vector (see details for more). Denotes the agglomerative methods (e.g. \emph{'ward.D'}) to be used. Run \code{validAgglomerativeMethods()}
to obtain all valid input options.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{method_dist}{Character value or vector (see details for more). Denotes the distance methods (e.g. \emph{'euclidean'}) to be used. Run \code{validDistanceMethods()}
to obtain all valid input options.}

\item{method_dr}{Character value. The dimensional reduction method of
interest specified as a single character value. (Currently
either \emph{'pca'}, \emph{'umap'} or \emph{'tsne'}).}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{method_hclust}{Character value. Denotes the hierarchical clustering method  according
to which the clustering is performed. Valid input options are \emph{'ward.D', 'ward.D2', 'single',
'complete', 'average', 'mcquitty', 'median'} and \emph{'centroid'}.}

\item{method_ovl}{Character value. One of \emph{'classic', 'bayesian'}. Decides
according to which method the spatial overlap is calculated.}

\item{method_padj}{Character value. The method according to which the adjusted p-values will
be calculated. Given to \code{stats::p.adjust()}. Run \code{stats::p.adjust.methods} to obtain
all valid input options.}
}
\description{
Check method input
}
\keyword{internal}
