% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_monocle_input}
\alias{check_monocle_input}
\title{Check monocle input}
\usage{
check_monocle_input(
  preprocess_method,
  reduction_method,
  cluster_method,
  k = base::numeric(1),
  num_iter = base::numeric(1)
)
}
\arguments{
\item{preprocess_method}{Monocle3 - description:

A string specifying the initial dimension method to use,
currently either PCA or LSI. For LSI (latent semantic
indexing), it converts the (sparse) expression matrix into
tf-idf matrix and then performs SVD to decompose the gene
expression / cells into certain modules / topics. Default
is "PCA".}

\item{reduction_method}{Monocle3 - description:

A character string specifying the algorithm to use for
dimensionality reduction. Currently "UMAP", "tSNE", "PCA"
and "LSI" are supported.}

\item{cluster_method}{Monocle3 - description:

String indicating the clustering method to use. Options are
"louvain" or "leiden". Default is "leiden". Resolution parameter
is ignored if set to "louvain".}

\item{k}{Monocle3 - description:

Integer number of nearest neighbors to use when creating the k
nearest neighbor graph for Louvain/Leiden clustering. k is related
to the resolution of the clustering result, a bigger k will result
in lower resolution and vice versa. Default is 20.}

\item{num_iter}{Monocle3 - description:

Integer number of iterations used for Louvain/Leiden clustering.
The clustering result giving the largest modularity score will be
used as the final clustering result. Default is 1. Note that if
num_iter is greater than 1, the random_seed argument will be ignored
for the louvain method.}
}
\value{

}
\description{
Check monocle input
}
\details{
With respect to the arguments \code{preprocess_method},
\code{reduction_method} and \code{cluster_method}:

If you provide a vector instead of a single character value (string)
the function will iterate over all inputs via a for-loop to compute all
valid combinations.
}
