% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_monocle_packages}
\alias{check_monocle_packages}
\title{Makes sure that both packages are installed}
\usage{
check_monocle_packages()
}
\description{
Makes sure that both packages are installed
}
