% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_object}
\alias{check_object}
\title{Check spata object input}
\usage{
check_object(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\description{
Check spata object input
}
\keyword{internal}
