% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-wrappers.R
\name{check_paths}
\alias{check_paths}
\title{Check candidate paths}
\usage{
check_paths(candidate_paths, target, verbose = TRUE)
}
\arguments{
\item{candidate_paths}{Character vector. A list of candidate paths to check.}

\item{target}{Character. The resource name being searched for (e.g., "conda.sh").}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A character string with the first candidate path that exists.
}
\description{
Iterates over a vector of candidate paths and returns the first path that exists.
If none of the candidate paths exist, an error is raised.
}
\keyword{internal}
