% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_pattern}
\alias{check_pattern}
\title{Check pattern input}
\usage{
check_pattern(object, patterns = "", method_pr = "hotspot", of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of pattern names and checks which of these exist.

Returns an adjusted gene-set-vector or raises an error.
}
\keyword{internal}
