% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_pt}
\alias{check_pt}
\title{Check pt input}
\usage{
check_pt(
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_clrp = NULL,
  pt_clr = NULL
)
}
\arguments{
\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clr}{Character value. Specifies the color of all points.}
}
\description{
Check pt input
}
