% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_sample}
\alias{check_sample}
\title{Check sample input}
\usage{
check_sample(object, of_sample = "", desired_length = NULL, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{desired_length}{The length the input must have.}
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of sample names and checks which of these exist.

Returns an adjusted sample-vector or raises an error.
}
