% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_saving}
\alias{check_saving}
\title{Check saving input}
\usage{
check_saving(output_path, file_name)
}
\arguments{
\item{output_path}{Character vector or NULL. Specifies the folder in which to store
the object if the directory is valid. If set to NULL saving is skipped.}

\item{file_name}{Character value. The name-suffix for the file name under which the
spata-object is stored if \code{output_path} is a valid directory. Is prefixed with
\emph{'spata-obj-'} and suffixed with \emph{'.RDS'}.}
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a folder directory
and assembles the final output directory including the specified filename.

Returns the final directory if it is valid or raises an error if not or NULL if \code{output_path}
is set to NULL.
}
