% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{check_slot_coordinates}
\alias{check_slot_coordinates}
\title{Check spata slots}
\usage{
check_slot_coordinates(object)
}
\arguments{
\item{object}{A spata-object.}
}
\value{
A character string. (Call \code{base::writeLines()} with that
string as input in order to format it.)
}
\description{
Functions that provide a report regarding the validity of
the respective slot.
}
\keyword{internal}
