% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_smooth}
\alias{check_smooth}
\title{Check smooth input}
\usage{
check_smooth(
  df = NULL,
  smooth = NULL,
  smooth_span = NULL,
  smooth_method = NULL,
  smooth_se = NULL
)
}
\arguments{
\item{df}{A data.frame that is to be smoothed spatially. That data frame must have
numeric \emph{x}- and \emph{y}-variables.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{smooth_method}{The smoothing method that will be used specified as a
single character value (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be}

\item{smooth_clr}{Character value. The color with which to display the smoothed
model.}
}
\description{
Check smooth input
}
