% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_spata_df}
\alias{check_spata_df}
\title{Check coordinate data.frame}
\usage{
check_spata_df(spata_df)
}
\arguments{
\item{spata_df}{A data.frame containing information of barcode-spots. Must contain the variables.

\describe{
\item{\emph{barcodes}}{Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
\item{\emph{sample}}{Character. The sample belonging of every barcode-spot.}
}}
}
\description{
Check coordinate data.frame
}
