% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_stdf}
\alias{check_stdf}
\title{Check summarized trajectory data.frame}
\usage{
check_stdf(stdf, shift = NULL)
}
\arguments{
\item{stdf}{A data.frame containing information about a spatial trajectory. Must contain the variables:

 \describe{
  \item{\emph{trajectory_part}}{Character. The trajectory's subparts.}
  \item{\emph{trajectory_part_order}}{Numeric. Denotes every trajectory-bin's position along the trajectory's subpart.}
  \item{\emph{trajectory_order}}{Numeric. Denotes every trajectory-bin's position along the whole trajectory.}
  \item{\emph{variables}}{Character. The genes, gene-sets and features that have been summarized along the trajectory.}
  \item{\emph{values}}{Numeric. The binned values of every gene, gene-set and feature that has been summarized along the trajectory.}
  }}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
