% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_summarized_trajectory_df}
\alias{check_summarized_trajectory_df}
\title{Check stdf-input}
\usage{
check_summarized_trajectory_df(stdf)
}
\arguments{
\item{stdf}{A summarized trajectory data.frame}
}
\description{
Check stdf-input
}
