% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_to}
\alias{check_to}
\title{Gives feedback about input validity}
\usage{
check_to(object, to)
}
\arguments{
\item{to}{Character value. Denotes the object to which
the directory leads. Must be one of \emph{'cell_data_set', 'seurat_object'}
or \emph{'spata_object'}.}
}
\description{
Gives feedback about input validity
}
