% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_trajectory_binwidth}
\alias{check_trajectory_binwidth}
\title{Check trajectory binwdith input}
\usage{
check_trajectory_binwidth(binwidth)
}
\arguments{
\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}
}
\description{
Check trajectory binwdith input
}
