% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_uniform_genes}
\alias{check_uniform_genes}
\title{Check uniform genes input}
\usage{
check_uniform_genes(uniform_genes)
}
\arguments{
\item{uniform_genes}{Character value. If set to \emph{'discard'} genes that are
uniformly expressed across all barcode-spots of the specified coordinates
data.frame are discarded. If set to \emph{'keep'} they are kept.}
}
\value{

}
\description{
Check uniform genes input
}
