% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_variables}
\alias{check_variables}
\title{Check variables}
\usage{
check_variables(
  variables,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = Inf,
  max_slots = 3,
  simplify = FALSE
)
}
\arguments{
\item{variables}{Character vector. The variables of interest:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
 }}

\item{all_features}{Valid features.}

\item{all_gene_sets}{Valid gene sets.}

\item{all_genes}{Valid genes.}

\item{max_length}{Max number of variable input.}

\item{max_slots}{Max number of different aspects.}

\item{simplify}{If set to TRUE the \code{check_variables()}-output is a vector.}
}
\description{
A member of the \code{adjusting-check_()}-family. Takes a character
vector and sorts its elements into a list depending on whether it was found in
the input of \code{all_features}, \code{all_genes} or \code{all_gene_sets}.

Returns a list with one slot named \emph{features}, \emph{genes} or \emph{gene_sets}
containing the respective found/valid input of \code{color_to}.
}
