% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clustering}
\alias{clustering}
\title{A list of data.frames that contain clustering variables for specific \code{SPATA2}
objects - available via the \code{SPATAData} package. Names of the list slots
correspond to the sample names in variable  \emph{sample} of \code{SPATAData::source_df}.
Names of the cluster variables in the data.frames corresponds to the algorithm used.}
\format{
An object of class \code{list} of length 1.
}
\usage{
clustering
}
\description{
@docType data
@usage data(snn_clustering)
}
\keyword{datasets}
