% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused.R
\name{clusterSpCorResults}
\alias{clusterSpCorResults}
\title{Cluster genes according to their expression profile}
\usage{
clusterSpCorResults(
  object,
  of_sample = "",
  method_hclust = "complete",
  k = NULL,
  h = NULL
)
}
\description{
Cluster genes according to their expression profile
}
