% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnv_ref}
\alias{cnv_ref}
\title{A list of reference data for copy number variation analysis (CNV)}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(cnv_ref)
}
\description{
The list contains three named slots:
}
\details{
\describe{
   \item{\emph{annotation}:}{ A data.frame denoting the barcodes of the count matrix in slot \emph{mtr} as reference barcodes.}
   \item{\emph{mtr}:}{ A count matrix that can be used as reference data for cnv analysis. The data derived from stRNA-seq of healthy human brain tissue.}
   \item{\emph{regions}:}{ A data.frame containing information about start and end points of chromosomes.}
 }
}
\keyword{datasets}
