% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cnv_regions_df}
\alias{cnv_regions_df}
\title{A data.frame necessary for cnv-analysis. Gives information about start and
end points of chromosomes.}
\format{
An object of class \code{data.frame} with 44 rows and 4 columns.
}
\usage{
data(cnv_regions_df)
}
\description{
A data.frame necessary for cnv-analysis. Gives information about start and
end points of chromosomes.
}
\keyword{datasets}
