% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{compute_angle_between_two_points}
\alias{compute_angle_between_two_points}
\title{Compute angle between two points}
\usage{
compute_angle_between_two_points(p1, p2)
}
\arguments{
\item{p1, p2}{Numeric vectors of length two, named \emph{x} and \emph{y}.}
}
\description{
Computes the angle between two points. 0° is aligned
with the y-axis.
}
