% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{compute_curve_irregularity}
\alias{compute_curve_irregularity}
\title{Compute Curve Irregularity}
\usage{
compute_curve_irregularity(curve)
}
\arguments{
\item{curve}{A numeric vector representing a curve or sequence of values.}
}
\value{
A numeric value indicating the irregularity of the curve.
}
\description{
Calculate the irregularity of a curve based on the total variation of its values.
}
\details{
This function computes the irregularity of a given curve by summing
the absolute differences between adjacent values in the curve. A lower irregularity
value suggests a smoother, less irregular curve, while a higher value indicates
a more irregular pattern.
}
\examples{
curve <- c(1, 2, 3, 2, 1)
irregularity <- compute_curve_irregularity(curve)

}
