% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute the distance between to points}
\usage{
compute_distance(starting_pos, final_pos)
}
\arguments{
\item{starting_pos, final_pos}{Numeric vector of length two. Denotes the two positions
between which the distance is calculated}
}
\value{
A numeric value.
}
\description{
Compute the distance between to points
}
\keyword{internal}
