% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{compute_expression_estimates}
\alias{compute_expression_estimates}
\title{Compute}
\usage{
compute_expression_estimates(coords_df)
}
\arguments{
\item{coords_df}{A coordinates data.frame as obtained by \code{\link[=getCoordsDfSA]{getCoordsDfSA()}}
or \code{\link{getCoordsDfST}}.}
}
\value{
A numeric vector representing position-based expression estimates.
}
\description{
This function computes position-based expression estimates given the minimum
and maximum distances and the average minimum center-to-center distance (AMCCD).
}
\keyword{internal}
