% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{compute_img_scale_fct}
\alias{compute_img_scale_fct}
\title{Compute scale factor of two images}
\usage{
compute_img_scale_fct(hist_img1, hist_img2)
}
\arguments{
\item{hist_img1, hist_img2}{Objects of class \code{HistologyImage}.}
}
\value{
Numeric value.
}
\description{
Computes the factor with which the dimensions
of \strong{image 1} must be multiplied in order to equal dimensions of
image 2.
}
