% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{compute_lds}
\alias{compute_lds}
\title{Compute loess deviation score}
\usage{
compute_lds(gradient, span = 0.5)
}
\arguments{
\item{span}{Given to \code{span} of \code{stats::loess()}.}

\item{y}{Expression gradient.}
}
\value{
Numeric value.
}
\description{
Fits a loess model to a curve and quantifies its noisiness by
averaging the absolute residuals of the curve to the fit.
}
