% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{compute_positions_expression_estimates}
\alias{compute_positions_expression_estimates}
\title{Compute Position-Based Expression Estimates}
\usage{
compute_positions_expression_estimates(min_dist, max_dist, amccd)
}
\arguments{
\item{min_dist}{Minimum distance for estimation.}

\item{max_dist}{Maximum distance for estimation.}

\item{amccd}{Average Minimum Center-to-Center Distance (AMCCD).}
}
\value{
A numeric vector representing position-based expression estimates.

A numeric vector representing positions for expression estimates.
}
\description{
This function computes position-based expression estimates given the minimum
and maximum distances and the average minimum center-to-center distance (AMCCD).
}
\note{
This function validates that the units of \code{amccd}, \code{min_dist},
and \code{max_dist} match to ensure consistent unit measurements.
}
