% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{compute_sgs_pvalue}
\alias{compute_sgs_pvalue}
\title{Compute p-value based on curve irregularity scores}
\usage{
compute_sgs_pvalue(gradient, random, span = 0.5)
}
\arguments{
\item{random}{A vector of randomly generated irregularity scores
against which to compare the observed one.}

\item{observed}{Numeric vector. The observed gradient.}
}
\value{
A numeric value ranging between 0-1 (inclusive).
}
\description{
Compute p-value based on curve irregularity scores
}
\examples{

random_lds <-
   map_dbl(
     .x = 1:1000,
     .f = function(i){

        set.seed(123*i)

        rg <- runif(20, min = 0, max = 1)

        compute_lds(rg)

        })

 gradient <- scales::rescale(1:20, to = c(0,1))

 gradient_lds <- compute_lds(gradient)

 compute_sgs_pvalue(gradient, random = random_lds)

}
