% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{computeChromosomalDamage}
\alias{computeChromosomalDamage}
\title{Compute chromosomal damage}
\usage{
computeChromosomalDamage(object, chr_vars = stringr::str_c("Chr", 1:22))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{chr_vars}{Character vector. Names of the meta features that contain the
copy number variation scores for each chromosome.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Estimates the degree of chromosomal damage of each \link[=concept_observations]{observation}
by computing the variance of copy number variations across chromosomes 1-22.

(Requires the results of \code{\link[=runCNV]{runCNV()}}).
}
