% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{computeCnvByChrArm}
\alias{computeCnvByChrArm}
\title{Compute CNV by chromosome arm}
\usage{
computeCnvByChrArm(object, summarize_with = "mean", overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}
}
\description{
Extension to \code{runCnvAnalysis()}. Uses the results
of \code{runCnvAnalysis()} to compute chromosomal by chromosome arm instead
of only by chromosome.
}
\details{
\code{runCnvAnalysis()} computes chromosomal alterations and, among
other things, adds the results in form of numeric variables to the feature
data.frame. Depending on the prefixed used (default \emph{'Chr'}) chromosomal alterations of e.g.
chromosome 7 are then accessible as numeric variables. E.g.
\code{plotSurface(object, color_by = 'Chr7')}.

\code{computeCnvByChrArm()} adds additional variables to the data.frame that
contain information about the alterations in chromosome \bold{arms} and
are named accordingly \emph{Chr7p}, \emph{Chr7q}.
}
