% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{computeCountPercentage}
\alias{computeCountPercentage}
\title{Compute count percentage}
\usage{
computeCountPercentage(
  object,
  regex = NULL,
  molecules = NULL,
  var_name = NULL,
  assay_name = activeAssay(object),
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{regex}{Character value. A regular expression with which to create the
set of molecules (e.g. '^MT-.*' to subset human mitochondrial genes).}

\item{molecules}{Character vector. Instead of providing a regular expression
the set of molecules can be specified directly.}

\item{var_name}{Character value. The name of the new meta feature.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Calculates the percentage contribution of a specified set of molecules to the total counts
within the count matrix of the given assay.
}
\details{
The equivalent of \code{\link[Seurat:PercentageFeatureSet]{Seurat::PercentageFeatureSet()}}. Usage differs. See examples.
}
\examples{

library(SPATA2)
library(SPATAData)
library(patchwork)

object <- downloadSpataObject("MGH258")

# compute the percentage contribution of mitochondrial genes
object <- computeCountPercentage(object, regex = "MT-.*", var_name = "perc_mit")

plotSurface(object, color_by = "perc_mit") +
plotDensityPlot(object, variables = "perc_mit")

# keep only spots with less than 20\% mitochondrial counts
object <- filterSpataObject(object, perc_mit < 20)

# new outlier identification necessary?
plotSurface(object, color_by = "tissue_section")

}
\seealso{
\code{\link[=filterSpataObject]{filterSpataObject()}}
}
