% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{computePixelScaleFactor}
\alias{computePixelScaleFactor}
\alias{computePixelScaleFactor,SPATA2-method}
\alias{computePixelScaleFactor,SpatialData-method}
\title{Compute pixel scale factor}
\usage{
computePixelScaleFactor(object, ...)

\S4method{computePixelScaleFactor}{SPATA2}(object, verbose = TRUE, ...)

\S4method{computePixelScaleFactor}{SpatialData}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
Computes the pixel scale factor. Only possible for methods
that have a fixed center to center distance between their
observational units (e.g. Visium).
}
\seealso{
\code{\link[=containsCCD]{containsCCD()}}, \code{\link[=getPixelScaleFactor]{getPixelScaleFactor()}}, \code{\link[=setPixelScaleFactor]{setPixelScaleFactor()}}
}
