% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-documentation.R
\name{concept_active}
\alias{concept_active}
\title{Active / Inactive}
\description{
SPATA2 aims to facilitate vertical integration of multiple
data modalities. Therefore, the \code{\link{SPATA2}} object allows to store multiple
aspects of data at the same time. To know what to extract by default, the
concept \emph{active} and \emph{inactive} has been integrated.

For instance, multiple images can be \link[=registerImage()]{registered}.
Which image to use can always be specified using the argument \code{img_name}. By
default, the functions will pick the image that was denoted as the active image
by the function \code{\link[=activateImage]{activateImage()}}. Which image is currently active can
be checked with \code{\link[=activeImage]{activeImage()}}.

This concept applies to every other aspect, like \link[=MolecularAssays]{assays}
and data matrices. It only becomes relevant if the \code{SPATA2} object contains
more than one aspect.
}
\seealso{
\code{\link[=activeAssay]{activeAssay()}}, \code{\link[=activeImage]{activeImage()}}, \code{\link[=activeMatrix]{activeMatrix()}}
}
\keyword{internal}
