% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-documentation.R
\name{concept_molecular_modalities}
\alias{concept_molecular_modalities}
\title{Molecular Modalities}
\description{
SPATA2 was developed with the Visium platform in mind which revolves around
spatial gene expression. With SPATA2 v3.0.0 we aim to expand the package to
include different platforms and more molecular modalities to analyze spatial
distribution of proteins or metabolites, too. Hence, when creating  objects
with \link[=MolecularAssay]{molecular assays} the modality must be specified.

To ensure that inbuilt functions of SPATA2 like \link[=runGSEA]{gene set enrichment} or
\link[=runCNV]{copy number variation} analysis work seamlessly, the modality must be
specified \emph{"correctly"}.

SPATA2 knows three molecular modalities for which specific functions have been
written like the ones linked above.

\enumerate{
\item{Gene expression}{: Use \code{modality = 'gene'}}
\item{Protein expression}{: Use \code{modality = 'protein'}}
\item{Metabolites expression}{: Use \code{modality = 'metabolite'}}
}

Depending on the modality of an assay, specific functions can be used or not.
For instance, \code{\link[=runCNV]{runCNV()}} only works if \code{\link{SPATA2}} object contains an assay
of data modality \emph{gene} (not \emph{genes}, \emph{rna}, \emph{mRNA} or anything else). This
extends to the inbuilt concept of \link[=concept_molecular_signatures]{molecular signatures}.
It is not forbidden, of course, to create molecular assays with modalities
differing from the ones SPATA2 knows. It is just that you won't be able to
use certain functions with the created assay.
}
\section{Molecular Modality == Assay Name!}{


The molecular modality of an assay also defines its name! Hence, if you
encounter the parameter \code{assay_name} it can be thought of defining the molecular
modality of interest! And the output of \code{\link[=activeAssay]{activeAssay()}} can be thought of
the output of \emph{active molecular modality}.
}

\seealso{
\code{\link[=createMolecularAssay]{createMolecularAssay()}}, \code{\link[=containsModality]{containsModality()}}, \code{\link[=getAssayModalities]{getAssayModalities()}}
}
\keyword{internal}
