% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-documentation.R
\name{concept_molecular_signatures}
\alias{concept_molecular_signatures}
\title{Molecular Signatures}
\description{
Molecular signatures are sets of molecules (such as genes or proteins) that are
associated with specific biological states, processes, or conditions. In SPATA2
a molecular signature is represented as a vector in a named list, where the character
values are the molecules of which the signature consists.
}
\section{SPATA2 inbuilt signatures}{


SPATA2 knows three \link[=concept_molecular_modalities]{molecular modalities}.

\enumerate{
\item{Gene expression}{: With \code{modality = 'gene'}}
\item{Protein expression}{: With \code{modality = 'protein'}}
\item{Metabolites expression}{: With \code{modality = 'metabolite'}}
}

Included in the package is a list named \code{\link{signatures}}, with corresponding
slots \code{signatures$gene}, \code{signatures$protein}, \code{signatures$metabolite}. This list is where
default signatures are stored for the respective data modality. Depending
on how the \code{\link{SPATA2}} object is initiated, the created \link[MolecularAssay]{molecular assay(s)}
already contain the respective signatures in slot @signatures.
}

\section{Signature names}{


In SPATA2 a signature name corresponds of two parts:

\emph{class}_\emph{biological function}

For instance, the gene set \emph{HM_HYPOXIA} is of class \emph{HM} (short for Hallmark) and contains
genes associated with increased presence of or response to hypoxic circumstances.
The class indicates the source from where the signature derives and is separated
from the biological function part with the \strong{first} _. Underscores afterwards
are ignored and interpreted as part of the biological function as in \emph{RCTM_TCR_SIGNALING}
(class: \emph{RCTM}; biological function: \emph{TCR_SIGNALING}).
}

\seealso{
\code{\link[=addSignature]{addSignature()}}, \code{\link[=getSignature]{getSignature()}}, \code{\link[=addGeneSet]{addGeneSet()}}, \code{\link[=getGeneSet]{getGeneSet()}}
}
\keyword{internal}
