% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-documentation.R
\name{concept_observations}
\alias{concept_observations}
\title{Observations (Data points)}
\description{
In the context of SPATA2, the term \emph{observations} refers to the
data points or entities for which spatial information (x- and y-coordinates)
as well as molecular data (e.g. gene expression) and meta data (e.g. cluster
assignment) is captured. The term \emph{observational unit} is the term that best
characterizes these data points as a whole. E.g. \emph{cells} for platform MERFISH,
\emph{barcoded spots} for platform Visium, \emph{barcoded beads} for platform SlideSeq
and so on.

Throughout documentation the term \emph{data points} and \emph{observations} are used
synonymously. Usually, when data regarding these data points is extracted as
with \code{\link[=getCoordsDf]{getCoordsDf()}}, it comes in form of data.frames where each row
corresponds to an observation and each column corresponds to a \emph{\link[=concept_variables]{variable}}
of this observation.
}
\seealso{
\code{\link{concept_variables}}
}
\keyword{internal}
