% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-documentation.R
\name{concept_spatial_trajectory}
\alias{concept_spatial_trajectory}
\title{Spatial Gradient Screening}
\description{
Spatial Gradient Screening conceptualizes changes in gene expression or,
generally speaking, changes in expression of numeric features as a continuum in
two dimensional space. It uses \link[=concept_spatial_annotation]{spatial annotations}
or \link[=concept_spatial_trajectory]{spatial trajectories} as references that
are indicative of biological forces to analyse gene expression changes dependent
on the distance to these forces.

Please refer to \emph{Kueckelhaus J., Frerichs S. et al., 2024} for a detailed
explanation.

A spatial trajectory is a linear segment with a start and an
end point which indicates a \emph{'direction'}. It can be used within the
framework of \link[=concept_spatial_gradient_screening]{Spatial Gradient Screening (SGS)}
as a spatial reference to abstract biological forces such as tumorous growth,
increasing cellular density or proximity to a border.
}
\keyword{internal}
