% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept-documentation.R
\name{concept_variables}
\alias{concept_variables}
\title{Variables (Features)}
\description{
In the context of SPATA2, the term \emph{variables} refers to the
features that characterize \emph{\link[=concept_observations]{observations}}.

Throughout documentation the term \emph{variables} and \emph{features} are used
synonymously. We work closely with the tidyverse which proposes the concept
of tidy data, which structures data.frames in observations and variables.
Therefore, we tend to stick to the term \emph{variables}

\strong{Note:} In previous versions of SPATA2 we used the term features and feature
data.frame and the slot @fdata to refer to variables that were not related
to molecular counts like gene expression or gene sets. This resulted in
confusion as many other platforms such as Seurat use the term features in general
to refer to what we refer to as variables. Therefore, the slot @fdata has been renamed
to @meta_obs and the corresponding data.frame has been renamed to meta data.frame,
as obtained by \code{\link[=getMetaDf]{getMetaDf()}}.

Next to the obligatory variable \emph{barcodes} - which uniquely identifies each observation -
different kind of variables exist in the \code{\link{SPATA2}} object.
}
\section{Numeric variables}{


Numeric variables represent continuous or numerical data. These variables
can take on numeric values and are typically used to represent quantitative
measurements counts. When working with SPATA2 numeric variables are
conceptually subdivided.

\itemize{
\item{\emph{spatial}:}{ Numeric variables used to position the observations in two dimensional
space. Stored in the coordinates data.frame as obtained by \code{\link[=getCoordsDf]{getCoordsDf()}}.
E.g. \emph{x}, \emph{x_orig}, \emph{y} and \emph{y_orig}.}
\item{\emph{molecular}:}{ Numeric variables used to quantify molecular expression of an
observation. Stored in the count and processed matrices of the \code{\link{MolecularAssay}}
objects. E.g. \emph{GFAP}, \emph{VEGFA}, \emph{LDH}}
\item{\emph{signature}:}{ Specific scores or mean expression based on other numeric (mainly
molecular) data variables. E.g. gene signatures like \emph{HM_HYPOXIA} or cell cycling scores.}
\item{\emph{meta}:}{ Numeric variables that do not fit in any of the descriptions above
and often correspond to meta data. E.g. the number of molecule counts per observation.}
}
}

\section{Categorical / Grouping variables}{


Categorical or grouping variables represent qualitative data that can take on a limited number
of distinct categories or levels. These variables are used to categorize or group observations
into distinct groups or classes. When working with SPATA2 grouping variables are
conceptually subdivided.

\itemize{
\item{\emph{cluster}:}{ Results of clustering algorithms. E.g. \code{\link[=runBayesSpaceClustering]{runBayesSpaceClustering()}}}
\item{\emph{segmentation}:}{ Results of manual, spatial segmentation via \code{\link[=createSpatialSegmentation]{createSpatialSegmentation()}}}
\item{\emph{meta}:}{ Categorical variables that do not fit in any of the descriptions above.
E.g. group assignment by \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}.}
}

Grouping variables are stored as factors in the meta data.frame of slot @meta_obs.
}

\seealso{
\code{\link[=joinWithVariables]{joinWithVariables()}}
}
\keyword{internal}
