% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsAssay}
\alias{containsAssay}
\title{Check availability of an assay}
\usage{
containsAssay(object, assay_name, error = FALSE)
}
\value{
TRUE if the assay is found in the object, FALSE otherwise.
}
\description{
Checks if the provided object contains a specific assay.
}
\seealso{
\code{\link[=getAssayNames]{getAssayNames()}}, \code{\link{MolecularAssay}}
}
