% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsCaptureArea}
\alias{containsCaptureArea}
\alias{containsCaptureArea,SPATA2-method}
\alias{containsCaptureArea,SpatialData-method}
\title{Check availability of capture area}
\usage{
containsCaptureArea(object, ...)

\S4method{containsCaptureArea}{SPATA2}(object, error = FALSE)

\S4method{containsCaptureArea}{SpatialData}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the object contains a capture area data.frame
as obtained by \code{\link[=getCaptureArea]{getCaptureArea()}}.
}
