% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsCCD}
\alias{containsCCD}
\alias{containsCCD,ANY-method}
\title{Check availability of center to center distance}
\usage{
containsCCD(object, ...)

\S4method{containsCCD}{ANY}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the object contains a center to center
distance as obtained by \code{\link[=getCCD]{getCCD()}}.
}
