% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsCells}
\alias{containsCells}
\alias{containsCells,ANY-method}
\title{Check availability of cells}
\usage{
containsCells(object, ...)

\S4method{containsCells}{ANY}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the object revolves around a spatial method
with single cells as the observational unit.
}
\seealso{
\code{\link[=containsSpots]{containsSpots()}}
}
