% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{containsCNV}
\alias{containsCNV}
\alias{containsHistologyImage}
\title{Check availability of miscellaneous content}
\usage{
containsCNV(object)

containsHistologyImage(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Logical value.
}
\description{
Logical tests that check if content exists in the \code{spata2} object.
}
