% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsHistoImages}
\alias{containsHistoImages}
\alias{containsHistoImages,SPATA2-method}
\alias{containsHistoImages,SpatialData-method}
\title{Check availability of image containers}
\usage{
containsHistoImages(object, ...)

\S4method{containsHistoImages}{SPATA2}(object, error = FALSE, ...)

\S4method{containsHistoImages}{SpatialData}(object, error = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the input object contains any \code{\link{HistoImage}} objects.

Note to confuse with \code{\link[=containsImage]{containsImage()}}.
}
