% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{containsHistologyImage}
\alias{containsHistologyImage}
\title{Check availability of \code{HistologyImage} object}
\usage{
containsHistologyImage(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Logical value.
}
\description{
Checks if slot @images contains an object
of class \code{HistologyImage} or if it is empty.
}
