% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R, R/deprecated_a-i.R
\name{containsImageObject}
\alias{containsImageObject}
\alias{containsHistologyImaging}
\title{Check availability of \code{HistologyImaging} object}
\usage{
containsImageObject(object)

containsHistologyImaging(object)
}
\description{
Deprecated in favor of \code{\link[=containsHistoImaging]{containsHistoImaging()}}.
}
\keyword{internal}
