% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{containsInnerBorders}
\alias{containsInnerBorders}
\alias{containsInnerBorders,spata2-method}
\alias{containsInnerBorders,SpatialAnnotation-method}
\alias{containsInnerBorders,data.frame-method}
\title{Check for Inner Borders in a Spatial Annotation}
\usage{
containsInnerBorders(object, ...)

\S4method{containsInnerBorders}{spata2}(object, id, ...)

\S4method{containsInnerBorders}{SpatialAnnotation}(object, ...)

\S4method{containsInnerBorders}{data.frame}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}
}
\value{
Logical value.
}
\description{
Checks whether a \code{SpatialAnnotation} object contains any inner borders.
}
\seealso{
\code{\link{SpatialAnnotation}}
}
