% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsModality}
\alias{containsModality}
\title{Check availability of molecular modality}
\usage{
containsModality(object, modality, error = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{modality}{Character value. The \link[=concept_molecular_modalities]{molecular modality}
of interst.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Logical value.
}
\description{
Tests if the input object contains a molecular assay
of a certain modality.
}
