% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsPixelContent}
\alias{containsPixelContent}
\alias{containsPixelContent,SpatialData-method}
\alias{containsPixelContent,HistoImage-method}
\title{Check availability pixel content}
\usage{
containsPixelContent(object, ...)

\S4method{containsPixelContent}{SpatialData}(object, img_name, error = FALSE)

\S4method{containsPixelContent}{HistoImage}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{\link{HistoImage}} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if slot @pxl_content of a \code{HistoImage} object
contains the results of \code{identifyPixelContent()}.
}
\seealso{
\code{\link[=identifyPixelContent]{identifyPixelContent()}}
}
