% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{containsPseudoImage}
\alias{containsPseudoImage}
\alias{containsPseudoImage,ANY-method}
\title{Check if the object contains only a pseudo image}
\usage{
containsPseudoImage(object, ...)

\S4method{containsPseudoImage}{ANY}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if the object only contains a pseudo image which
makes it not suitable for image depending processes.
}
