% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{containsScaleFactor}
\alias{containsScaleFactor}
\alias{containsScaleFactor,HistoImaging-method}
\title{Check availability of specific scale factors}
\usage{
containsScaleFactor(object, ...)

\S4method{containsScaleFactor}{HistoImaging}(object, fct_name, img_name = NULL, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if specifics scale factors are set or not.
}
