% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsSectionVariable}
\alias{containsSectionVariable}
\alias{containsSectionVariable,ANY-method}
\title{Check availability of section variable}
\usage{
containsSectionVariable(object, ...)

\S4method{containsSectionVariable}{ANY}(object, error = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if the results of \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}} exist
in the object, namely a variable called \emph{section} in the coordinates
data.frame.
}
\seealso{
\code{\link[=containsSpatialOutliers]{containsSpatialOutliers()}} to check if the section variable
contains any identified spatial outliers.
}
