% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsSpatialOutliers}
\alias{containsSpatialOutliers}
\title{Check if spatial outliers exist}
\usage{
containsSpatialOutliers(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Logical value.
}
\description{
Checks if \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}} has identified any
spatial outliers.
}
\seealso{
\code{\link[=removeSpatialOutliers]{removeSpatialOutliers()}} to exclude spatial outliers from further
analysis.
}
