% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsTissueOutline}
\alias{containsTissueOutline}
\alias{containsTissueOutline,SPATA2-method}
\alias{containsTissueOutline,HistoImaging-method}
\alias{containsTissueOutline,HistoImage-method}
\title{Check availability of tissue outline}
\usage{
containsTissueOutline(object, ...)

\S4method{containsTissueOutline}{SPATA2}(object, img_name = activeImage(object), error = FALSE)

\S4method{containsTissueOutline}{HistoImaging}(object, img_name = activeImage(object), error = FALSE)

\S4method{containsTissueOutline}{HistoImage}(object, img_name = activeImage(object), error = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if the object contains tissue outline
as identified by \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}.
}
