% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{countImageAnnotationTags}
\alias{countImageAnnotationTags}
\title{Count image annotation tags}
\usage{
countImageAnnotationTags(object, tags = NULL, collapse = " & ")
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{tags}{Character vector or list or NULL. If character vector only image
annotations that pass the "tag test" are included in the counting process. If
list, every slot should be a character vector of tag names that are counted
as combinations.}

\item{collapse}{Characer value. Given to argument \code{collapse} of
\code{sttringr::str_c()} if input for argument \code{tags} is a list.}
}
\value{
A data.frame with two variables: \emph{tags} and \emph{n}
}
\description{
Counts image annotations by tags. See details for more
information.
}
