% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_model_df}
\alias{create_model_df}
\title{Create model data.frame}
\usage{
create_model_df(
  input,
  var_order = NULL,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  noise_level = 0,
  noise = NULL,
  seed = 123,
  range = c(0, 1),
  verbose = TRUE
)
}
\arguments{
\item{var_order}{Character. The name of the variable that is supposed to
indicate the direction.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Data.frame.
}
\description{
Creates the data.frame that contains the models
for spatial gradient screening algorithms.
}
