% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{createHistoImage}
\alias{createHistoImage}
\title{Create an object of class \code{HistoImage}}
\usage{
createHistoImage(
  dir,
  name,
  sample,
  scale_factors = list(coords = 1),
  reference = FALSE,
  tissue_outline = TRUE,
  use_greyscale = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir}{Character value. The directory from where to retrieve the image.}

\item{name}{Character value. The name of the image.}

\item{sample}{Character value. The sample name to which the image belongs.
Should be equal to slot @sample of the \code{HistoImaging} object in which
the \code{HistoImage} is stored.}

\item{scale_factors}{list. Sets slot @scale_factors,}

\item{reference}{Logical value. If \code{TRUE}, the \code{HistoImage} is
treated as the reference image for all other registered images in
the \code{HistoImaging} object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{HistoImage}
}
\description{
Official constructor function of the S4 class \code{HistoImage}.
}
\seealso{
\code{\linkS4class{HistoImage}}
}
