% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{createHistoImaging}
\alias{createHistoImaging}
\alias{createHistoImagingVisium}
\title{Create an object of class \code{HistoImaging}}
\usage{
createHistoImaging(
  sample,
  hist_img_ref = NULL,
  hist_imgs = list(),
  active = NULL,
  unload = TRUE,
  coordinates = tibble::tibble(),
  meta = list(),
  method = SpatialMethod(),
  misc = list(),
  verbose = TRUE,
  ...
)

createHistoImagingVisium(
  dir,
  sample,
  img_ref = "lowres",
  img_active = "lowres",
  meta = list(),
  misc = list(),
  verbose = TRUE
)
}
\arguments{
\item{sample}{Character value. The sample name of the tissue.}

\item{hist_img_ref}{The \code{HistoImaging} serving as the reference image.
Should be created with \code{createHistoImage()}.}

\item{hist_imgs}{List of additional \code{HistoImaging} objects for slot @images.}

\item{active}{Character value. Name of the \code{HistoImage} that is set
to the active image. Defaults to the reference image.}

\item{coordinates}{Data.frame of at least three variables:

\itemize{
\item{\emph{barcodes}: }{Character variable with unique IDs for each observation.}
\item{\emph{x_orig}: }{Numeric variable representing x-coordinates in a cartesian coordinate system.}
\item{\emph{y_orig}: }{Numeric variable representing y-coordinates in a cartesian coordinate system.}
}

Coordinates should align with the tissue outline of the reference \code{HistoImage} after being
multiplied withe its coordinate scale factor in slot @scale_factors$coords.}

\item{meta}{List of meta data regarding the tissue.}

\item{misc}{List of miscellaneous information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{dir}{The directory to the output folder of the platform.}

\item{img_ref, img_active}{Character values specifying which of the images to register and how to register
them. See details of \code{\link{HistoImaging}} for more information about the definitions
of the reference image and the active image. Setting both arguments to the same
value results in the function to register the specified image as the active
as well as the reference image. Additional images can be registered later on
at any time using the funciton \code{\link[=registerImage]{registerImage()}}. Valid input options depend
on the platform used:

\itemize{
\item{\emph{Visium}:}{ Either \emph{'lowres'} or \emph{'hires'}.}
}}

\item{empty_image_slots}{Logical value. If \code{TRUE}, content of slot @image
of all \code{HistoImage} objects is emptied except for the active one.}
}
\value{
An object of class \code{HistoImaging}
}
\description{
Official constructor function of the S4 class \code{HistoImaging}.
Functions suffixed by the platform name are wrappers written for their
standardized output folder.
}
\seealso{
\code{\link[=createHistoImage]{createHistoImage()}}, \code{\link[=registerHistoImage]{registerHistoImage()}}
}
