% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{createHistologyImage}
\alias{createHistologyImage}
\title{Create an object of class \code{HistologyImage}}
\usage{
createHistologyImage(
  dir,
  name,
  sample,
  reference = FALSE,
  use_greyscale = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dir}{Character value. The directory from where to retrieve the image.}

\item{name}{Character value. The name of the image.}

\item{sample}{Character value. The sample name to which the image belongs.
Should be equal to slot @sample of the \code{HistologyImaging} object in which
the \code{HistologyImage} is stored.}

\item{reference}{Logical value. If \code{TRUE}, the \code{HistologyImage} is
treated as the reference image for all other registered images in
the \code{HistologyImaging} object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{HistologyImage}
}
\description{
Official constructor function of the S4 class \code{HistologyImage}.
}
