% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createHistologyImaging}
\alias{createHistologyImaging}
\title{Create S4 Image object for SPATA2}
\usage{
createHistologyImaging(
  image,
  id,
  img_scale_fct = 1,
  meta = list(),
  pxl_scale_fct = NULL,
  coordinates = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{image}{Image input or character value. If character, input is interpreted as a directory
to a file or to an URL and is read with \code{EBImage::readImage()}. The read image
should be of type \emph{.png}, \emph{.jpeg} or \emph{.tiff}. Capital letters work, too.

If not character, the function ensures that the input is - or is convertible - to
class \code{Image} via \code{EBimage::as.Image()}. If that fails, an error is thrown.}

\item{id}{Character value. Name of the \code{HistologyImaging} object.}

\item{img_scale_fct}{Numeric value between 0 and 1. If lower than 1, is used
to downscale the image before setting it.}

\item{coordinates}{A data.frame of observational units that underlie the image
in case of spatially resolved multi-omic studies. Should contain at least the
two variables: \emph{x}, \emph{y} and a variable that identifies the observational units (e.g. \emph{barcodes}).}
}
\value{
An object of class \code{HistologyImaging}.
}
\description{
Creates an object of class \code{HistologyImaging} that is used to
store the image, image meta data and image annotations.

Located in slot @images in the \code{SPATA2} object.
}
\seealso{
\code{?HistologyImaging} for the documentation of all slots.
}
