% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{createHistologyImagingNew}
\alias{createHistologyImagingNew}
\title{Create an object of class \code{HistologyImaging}}
\usage{
createHistologyImagingNew(
  sample,
  hist_img_ref,
  hist_img_active = NULL,
  hist_imgs = list(),
  empty_image_slots = TRUE,
  coordinates = tibble::tibble(),
  coordinates_id = NULL,
  meta = list(),
  misc = list(),
  ...
)
}
\arguments{
\item{sample}{Character value. The sample name of the tissue.}

\item{hist_img_ref}{The \code{HistologyImaging} object for slot @image_reference.
Should be created with \code{createHistologyImage()}.}

\item{hist_img_active}{The \code{HistologyImaging} object for slot @image_active.
Should be created with \code{createHistologyImage()}. If \code{NULL}, input of \code{hist_img_ref} is taken.}

\item{hist_imgs}{List of additional images for slot @images_registerd.}

\item{empty_image_slots}{Logical value. If \code{TRUE}, content of slot @image
of all \code{HistologyImage} objects is emptied except for the one set in
slot @image_active - to minimize object size.}

\item{coordinates}{Data.frame of five variables.
\itemize{
\item{\code{coordinates_id}}{Character variable with unique IDs for each observation.}
\item{\emph{x}:}{Numeric variable representing x-coordinates in a Cartesian coordinate system.}
\item{\emph{y}:}{Numeric variable representing y-coordinates in a Cartesian coordinate system.}
\item{\emph{width}:}{Numeric variable representing x-coordinates as image width. (Equal to \emph{x}).}
\item{\emph{height}:}{Numeric variable representing y-coordinates as image height.}
}}

\item{coordinates_id}{Character value. The name of the ID variable of the data.frame
from \code{coordinates}.}

\item{meta}{List of meta data regarding the tissue.}

\item{misc}{List of miscellaneous information.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{HistologyImagingNew}
}
\description{
Official constructor function of the S4 class \code{HistologyImaging}.
}
\seealso{
\code{\link[=add_xy]{add_xy()}}, \code{\link[=add_wh]{add_wh()}}
}
