% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family.R
\name{createImageObject}
\alias{createImageObject}
\title{Create S4 Image object for SPATA2}
\usage{
createImageObject(image, image_class = "HistologyImage", ...)
}
\arguments{
\item{image}{Image of class \code{EBImage}.}

\item{image_class}{Name of the S4 object. One of \code{validImageClasses()}.}

\item{...}{}
}
\value{
An object of class \code{image_class}.
}
\description{
Creates the basic S4 image object that is put
in slot @images in the \code{SPATA2} object.
}
