% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{createSegmentation}
\alias{createSegmentation}
\title{Deprecated in favor of \code{createSpatialSegmentation()}}
\usage{
createSegmentation(...)
}
\description{
Deprecated in favor of \code{createSpatialSegmentation()}
}
