% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{createSegmentation2}
\alias{createSegmentation2}
\title{Spatial Segmentation (old version)}
\usage{
createSegmentation2(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated version of the spata-object specified as \code{object}
now containing the information about all drawn segments.
}
\description{
The function \code{createSegmentation2()} provides access to an
interactive mini-shiny application that allows to separate a sample into
several segments.
}
