% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createSpatialData}
\alias{createSpatialData}
\alias{createSpatialDataMERFISH}
\alias{createSpatialDataSlideSeqV1}
\alias{createSpatialDataVisium}
\alias{createSpatialDataVisiumHD}
\alias{createSpatialDataXenium}
\title{Create an object of class \code{SpatialData} from raw output}
\usage{
createSpatialData(
  sample,
  hist_img_ref = NULL,
  hist_imgs = list(),
  active = NULL,
  unload = TRUE,
  coordinates = tibble::tibble(),
  meta = list(),
  method = SpatialMethod(),
  scale_factors = list(),
  misc = list(),
  verbose = TRUE,
  ...
)

createSpatialDataMERFISH(
  dir,
  sample,
  file_coords = NULL,
  meta = list(),
  misc = list(),
  verbose = TRUE
)

createSpatialDataSlideSeqV1(
  dir,
  sample,
  file_coords = NULL,
  meta = list(),
  misc = list()
)

createSpatialDataVisium(
  dir,
  sample,
  img_ref = "lowres",
  img_active = "lowres",
  meta = list(),
  misc = list(),
  verbose = TRUE
)

createSpatialDataVisiumHD(
  dir,
  sample,
  img_ref = "lowres",
  img_active = "lowres",
  meta = list(),
  misc = list(),
  verbose = TRUE
)

createSpatialDataXenium(dir, sample, meta = list(), misc = list())
}
\arguments{
\item{sample}{Character value. The sample name of the tissue.}

\item{hist_img_ref}{The \code{SpatialData} serving as the \code{\link[=concept_images]{reference image}}.
Should be created with \code{createHistoImage()}.}

\item{hist_imgs}{List of additional \code{HistoImage} objects for slot @images.}

\item{active}{Character value. Name of the \code{HistoImage} that is set
to the active image. Defaults to the reference image.}

\item{coordinates}{Data.frame of at least three variables:

\itemize{
\item{\emph{barcodes}: }{Character variable with unique IDs for each observation.}
\item{\emph{x_orig}: }{Numeric variable representing x-coordinates in a cartesian coordinate system.}
\item{\emph{y_orig}: }{Numeric variable representing y-coordinates in a cartesian coordinate system.}
}

Coordinates should align with the tissue outline of the reference \code{HistoImage} after being
multiplied withe its coordinate scale factor in slot @scale_factors$coords.}

\item{meta}{List of meta data regarding the tissue.}

\item{misc}{List of miscellaneous information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{dir}{The directory to the output folder of the platform.}

\item{file_coords}{Character value or \code{NULL}. If character, specifies the filename
\strong{within} the directory \code{dir} that leads to the coordinates .csv file. If \code{NULL}
the expected filename is tried:

\itemize{
\item{\emph{MERFISH}:}{ File that contains \emph{'cell_metadata'} and ends with \emph{'.csv'}}
\item{\emph{SlideSeqV1}:}{ File that ends with \emph{'MatchedBeadLocation.csv'}}
\item{\emph{VisiumSmall/VisiumLarge}:}{ File named \emph{'tissue_positions_list.csv'} or \emph{'tissue_positions.csv'}}
\item{\emph{VisiumHD}:}{ File named \emph{'tissue_positions.parquet'}}
\item{\emph{Xenium}:}{ File named \emph{'cells.csv.gz'}.}
}}

\item{img_ref, img_active}{Character values specifying which of the images to register and how to register
them. Click \code{\link[=concept_images]{here}} for more information about the definitions
of the reference image and the active image. Setting both arguments to the same
value results in the function to register the specified image as the active
as well as the reference image. Valid input options depend
on the platform used:

\itemize{
\item{\emph{Visium}:}{ Either \emph{'lowres'} or \emph{'hires'}.}
}}
}
\value{
An object of class \code{SpatialData}
}
\description{
Official constructor function of the S4 class \code{\link{SpatialData}}.
Functions suffixed by the platform name are wrappers written for their
standardized output folder.
}
\seealso{
\code{\link[=registerImage]{registerImage()}} to register images afterwards.
}
