% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createSpatialMethod}
\alias{createSpatialMethod}
\title{Create an object of class \code{SpatialMethod}}
\usage{
createSpatialMethod(...)
}
\arguments{
\item{...}{Input for the respective slots. Use \code{slotNames(SpatialMethod)} to
have them printed in the console.}
}
\value{
An object of class \code{SpatialMethod}.
}
\description{
A simple wrapper to construct objects of class \code{\link{SpatialMethod}}.
Input is directly given to the basic constructer, but slot \verb{@version} is set
automatically.
}
