% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createSpatialSegmentation}
\alias{createSpatialSegmentation}
\title{Interactive sample segmentation}
\usage{
createSpatialSegmentation(
  object,
  height = 500,
  break_add = NULL,
  box_widths = c(4, 4, 4)
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\description{
Gives access to an interactive user interface where barcode-spots
can be interactively annotated.
}
\details{
Segmentation variables are grouping variables that are stored in
the feature data.frame of the \code{SPATA2} object (such as clustering variables).
They differ from clustering variables in so far as that they are not the result
of unsupervised cluster algorithms but from group assignment the researcher
conducts him/herself (e.g. histological classification).

Therefore, all segmentation variables can be extracted via \code{getFeatureNames()}
as they are part of those. To specifically extract variables that were created
with \code{createSpatialSegmentation()} use \code{getSegmentationVariableNames()}. To remove
annotations you no longer need use \code{discardSegmentationVariable()}.
}
\note{
The interface allows to zoom in on the sample. This is useful if your
\code{SPATA2} object contains an HE-image as background and you want to classify
barcode spots based on the histology. As these images are displayed by pixels
the resolution decreases the more you zoom in. Many experiments (such as
the Visium output) contain high resolution images. You can use the function
\code{exchangeImage()} to read in images of higher resolution for a better
histological classification.
}
\seealso{
exchangeImage()
}
