% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialTrajectories.R
\name{createSpatialTrajectories}
\alias{createSpatialTrajectories}
\title{Spatial Trajectories}
\usage{
createSpatialTrajectories(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated \code{SPATA2} object.
}
\description{
Provides access to an interactive shiny application
where trajectories can be drawn..
}
