% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{createTrajectories}
\alias{createTrajectories}
\title{Deprecated}
\usage{
createTrajectories(object)
}
\description{
Deprecated in favor of \code{\link[=createSpatialTrajectories]{createSpatialTrajectories()}}.
}
\keyword{internal}
