% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{createTrajectoryManually}
\alias{createTrajectoryManually}
\title{Deprecated}
\usage{
createTrajectoryManually(...)
}
\description{
Deprecated in favor of \code{\link[=adSpatialTrajectory]{adSpatialTrajectory()}}.
}
\keyword{internal}
