% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{currentSpata2Version}
\alias{currentSpata2Version}
\title{The current version of SPATA2}
\usage{
currentSpata2Version()
}
\value{
List of three numeric slots: \emph{major}, \emph{minor}, \emph{patch}
}
\description{
Outputs the current version of the package.
}
