% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{data_counts-class}
\alias{data_counts-class}
\alias{data_counts}
\title{data_counts object}
\value{
S4 object
}
\description{
data_counts object
}
\section{Slots}{

\describe{
\item{\code{counts}}{A sparse matrix containing the original counts.}

\item{\code{norm_exp}}{A processed expression matrix. Rownames must be the gene-names.
Column names must be the barcodes.}
}}

