% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{defaultGeomObs}
\alias{defaultGeomObs}
\title{Default observations geometry}
\usage{
defaultGeomObs(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character value.
}
\description{
Extracts the default ggplot2 geometry to visualize \link[=concept_observations]{observations}
in 2D space on surface plots.
}
\keyword{internal}
