% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{dim_red-class}
\alias{dim_red-class}
\alias{dim_red}
\title{dim_red object}
\value{
S4 object
}
\description{
dim_red object
}
\section{Slots}{

\describe{
\item{\code{UMAP}}{A data.frame containing the variables \emph{'barcdoes', 'sample', 'umap1', 'umap2'}}

\item{\code{TSNE}}{A data.frame containing the variables \emph{'barcodes', 'smaple', 'tsne1', 'tsne2'}}
}}

