% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{discardFeatures}
\alias{discardFeatures}
\title{Discard features}
\usage{
discardFeatures(object, feature_names, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_names}{Character vector. Specifies the features to be discarded.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
An updated spata-object.
}
\description{
Discards the features of choice.
}
