% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{discardGeneFeatures}
\alias{discardGeneFeatures}
\title{Discard gene features}
\usage{
discardGeneFeatures(object, feature_names, mtr_name = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_names}{Character vector. Specifies the gene features to be discarded.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}
}
\value{
An updated spata-object.
}
\description{
Discards the features of choice of the gene meta data.
}
