% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{discardGenes}
\alias{discardGenes}
\title{Discard genes}
\usage{
discardGenes(
  object,
  genes = NULL,
  regex = NULL,
  include_dea = TRUE,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{genes}{Character vector or NULL. If character vector, specifies the genes
to be discarded by name.}

\item{regex}{Character value or NULL. If character value, specifies the
regular expression according to which genes are discarded.}

\item{include_dea}{Logical value. If set to TRUE the results of de-analysis
are included. If set to FALSE de-analysis results are skipped during the
discarding steps.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
An updated spata-object.
}
\description{
This function takes a vector of genes or
a regular expression and discards genes from the object's
data matrices, gene meta data.frames and de-analysis results
that match the input.
}
