% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{downloadSpataObject}
\alias{downloadSpataObject}
\title{Download a SPATA2 object}
\usage{
downloadSpataObject(
  sample_name,
  overwrite = FALSE,
  file = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name of the sample you want to
download. Use \code{validSampleNames()} to obtain all valid input options.}

\item{overwrite}{Logical. Must be set to \code{TRUE} if file directories
under which downloaded files are to be saved already exist.}

\item{file}{If you want to save the object on disc:
The filename of the \code{SPATA2} object. Must end with \emph{'.RDS'} if provided
as a character. If \code{NULL}, the function saves the object
under the sample name with an \emph{'.RDS'} suffix. If \code{FALSE}, the saving
is skipped and the object is simply returned.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The downloaded \code{SPATA2} object.
}
\description{
Downloads a single \code{SPATA2} object and returns it.
}
\note{
Imported from the package \code{SPATAData}.
}
\examples{
# download & assign (no saving on the disk)
object <- downloadSpataObject(sample_name = "UKF275T")

# download, assign and save on disk
# -> stores the file under ~/UKF275T.RDS (where '~' is your working directory)
object <- downloadSpataObject(sample_name = "UKF275T", file = TRUE)

# download, assign and save on disk in a specified directory
object <- downloadSpataObject(sample_name = "UKF275T", file = "my/path/to/spata_object.RDS")
}
\seealso{
For convenient downloads of multiple \code{SPATA2} objects
check out \code{\link[SPATAData:downloadSpataObjects]{downloadSpataObjects()}}.
}
