% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{dropped_df_dummy}
\alias{dropped_df_dummy}
\title{dropped_df}
\usage{
dropped_df_dummy(dropped_df)
}
\arguments{
\item{dropped_df}{A data.frame with no NAs. (Result of \code{tidyr::drop_na()}).}
}
\description{
dropped_df
}
