% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{emptyImageSlot}
\alias{emptyImageSlot}
\alias{emptyImageSlot,HistologyImage-method}
\alias{emptyImageSlot,HistologyImagingNew-method}
\title{Empty image slot}
\usage{
emptyImageSlot(object, ...)

\S4method{emptyImageSlot}{HistologyImage}(object, ...)

\S4method{emptyImageSlot}{HistologyImagingNew}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Removes the image from slot @image of a \code{HistologyImage}.
Useful for efficient data storing.
}
\seealso{
\code{\link[=loadImage]{loadImage()}}
}
