% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-data-frame-functions.R
\name{examineClusterResults}
\alias{examineClusterResults}
\title{Examine clustering results}
\usage{
examineClusterResults(cluster_df)
}
\arguments{
\item{cluster_df}{A data.frame containing the character variable \emph{barcodes}
as well as additional character variables representing different clustering-results.

E.g. the return value of \code{findMonocleClusters()}}
}
\value{
A list in which every slot represents a cluster variable and it's content
the unique clusters (groups) it contains.
}
\description{
Gives an overwiew of the cluster results of e.g. `findMonocleClusters()`.
}
