% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{examineTrajectoryAssessment}
\alias{examineTrajectoryAssessment}
\title{Examine trajectory-moddeling results}
\usage{
examineTrajectoryAssessment(
  atdf,
  limits = c(0, 10),
  plot_type = "histogram",
  binwidth = 0.5,
  clrp = "milo",
  ...
)
}
\arguments{
\item{limits}{The minimum and maximum auc-values to include. Given to
\code{ggplot2::scale_x_continuous()}.}

\item{plot_type}{One of \emph{'histogram', 'density', and 'ridgeplot'}.}

\item{...}{additional arguments given to \code{ggplot2::facet_wrap()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes the distribution of the assessment-scores
(residuals-area-under-the-curve) of a trajectory.
}
