% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{A data.frame necessary for cnv-analysis. Contains information about the gene positions
on chromosomes. Contains the following variables:}
\format{
An object of class \code{list} of length 11.
}
\usage{
example_data
}
\description{
\describe{
\item{\emph{ensembl_gene_id}:}{ Character. ENSEMBL encoding of gene names.}
\item{\emph{hgnc_symbol}:}{ Character. Gene names in HUGO format.}
\item{\emph{chromosome_name}:}{ Character. Name of the chromosome.}
\item{\emph{start_position}:}{ Integer. Starting position of the gene.}
\item{\emph{end_position}:}{ Integer. Ending positiong of the gene.}
}
}
\details{
@docType data
@usage data(example_data)
}
\keyword{datasets}
