% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{exclude}
\alias{exclude}
\alias{exclude,HistoImaging-method}
\alias{excludeSpatialOutliers}
\alias{excludeSpatialOutliers,spata2-method}
\alias{excludeSpatialOutliers,HistoImaging-method}
\alias{excludeTissueFragments}
\alias{excludeTissueFragments,HistoImaging-method}
\title{Exclude observations}
\usage{
exclude(object, ...)

\S4method{exclude}{HistoImaging}(object, ids, reason)

excludeSpatialOutliers(object, ...)

\S4method{excludeSpatialOutliers}{spata2}(object)

\S4method{excludeSpatialOutliers}{HistoImaging}(object)

excludeTissueFragments(object, ...)

\S4method{excludeTissueFragments}{HistoImaging}(object, fragments = "all")
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{ids}{Character vector. The IDs to exclude.}

\item{reason}{Character value. The reasoning for exclusion.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Excludes observations from further integration in analysis
or plots by setting their \emph{exclude} value to \code{TRUE}. Depending on the
suffix of the function exclusion happens based on results
of previous algorithms.
}
\note{
\code{excludeSpatialOutliers()} and \code{excludeTissueFragments()} require the output
of \code{identifyTissueOutline()}.
}
