% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{expandSpatialAnnotation}
\alias{expandSpatialAnnotation}
\title{Expand the outline of spatial annotations}
\usage{
expandSpatialAnnotation(object, id, expand, new_id = FALSE, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The ID of the spatial annotation of interest.}

\item{expand}{Distance measure with which to expand the border. Negative
values shrink the outline.}

\item{new_id}{Character value or \code{FALSE}. If character, the resulting
spatial annotation is stored under a new ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Expands or shrinks the outer outline of a spatial annotation.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- loadExampleObject("UKF275T")

plotImage(object) + ggpLayerSpatAnnOutline(object, ids = "vessel1", line_color = "red")
plotSpatialAnnotations(object, "vessel1")

object <- expandSpatialAnnotation(object, id = "vessel1", expand = "50um", new_id = "vessel1_exp")

plotSpatialAnnotations(object, ids = c("vessel1", "vessel1_exp"))

}
\seealso{
\code{\link[=smoothSpatialAnnotation]{smoothSpatialAnnotation()}}, \code{\link[=shiftSpatialAnnotation]{shiftSpatialAnnotation()}}, \code{\link{SpatialAnnotation}}
}
