% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R, R/unused.R
\name{subsetIAS}
\alias{subsetIAS}
\alias{export}
\title{dummy}
\usage{
subsetIAS(
  ias,
  angle_span = NULL,
  angle_bins = NULL,
  variables = NULL,
  verbose = TRUE
)

export()
}
\description{
dummy
}
\keyword{internal}
