% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{extract_row_col_vars_visiumHD}
\alias{extract_row_col_vars_visiumHD}
\title{Extract Row and Column Variables for VisiumHD Barcodes}
\usage{
extract_row_col_vars_visiumHD(
  coords_df,
  name_bcs = "barcodes",
  name_row = "row",
  name_col = "col"
)
}
\arguments{
\item{coords_df}{A dataframe containing VisiumHD barcodes.}

\item{name_bcs}{The column name where the barcodes are stored (default is "barcodes").}

\item{name_row}{The name to assign to the row variable (default is "row").}

\item{name_col}{The name to assign to the column variable (default is "col").}
}
\value{
A dataframe with added row and column information.
}
\description{
This function extracts row and column information from VisiumHD barcodes in a provided dataframe.
}
\keyword{internal}
