% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{extract_var_names}
\alias{extract_var_names}
\title{Extract var names from filter expression}
\usage{
extract_var_names(filter_expr)
}
\arguments{
\item{filter_expr}{A list of quosures representing logical expressions.}
}
\value{
A character vector containing the unique variable names used in the expressions.
}
\description{
This function takes a list of quosures representing logical expressions and
returns the variable names used in these expressions.
}
\keyword{internal}
