% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\name{filterTrends}
\alias{filterTrends}
\alias{filterTrajectoryTrends}
\title{Filter variables of a certain trend}
\usage{
filterTrends(atdf, limit = 5, trends = "all", variables_only = TRUE)

filterTrajectoryTrends(atdf, limit = 5, trends = "all", variables_only = TRUE)
}
\arguments{
\item{atdf}{An assessed trajectory data.frame (easily accessed via
\code{assessTrajectoryTrends()}).}

\item{limit}{Numeric value. The maximum area-under-the-curve value the
trajectory-trend-assessment might have.}

\item{variables_only}{Logical. If set to TRUE a character of variable-names is returned.
If set to FALSE the filtered data.frame is returned.}

\item{trend}{Character vector. The patterns of interest.}
}
\value{
A character vector of gene or gene-set names that follow the specified
patterns to the specified degree.
}
\description{
Extracts the genes or gene sets that follow a desired trend.
}
