% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-wrappers.R
\name{find_path}
\alias{find_path}
\title{Find a Resource Path}
\usage{
find_path(target, verbose = TRUE)
}
\arguments{
\item{target}{Character. The name of the resource to search for (e.g., "conda.sh").}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A character string containing the full path to the resource.
}
\description{
Searches for the specified resource (currently only "conda.sh" is supported)
by checking a predefined set of candidate paths.
}
\keyword{internal}
