% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-implementation.R
\name{findHierarchicalCluster}
\alias{findHierarchicalCluster}
\alias{findHierarchicalClusterCnv}
\title{Cluster sample via hclust}
\usage{
findHierarchicalCluster(
  object,
  method_dist,
  method_aggl,
  k = NULL,
  h = NULL,
  cluster_prefix = "",
  of_sample = NA
)

findHierarchicalClusterCnv(
  object,
  method_dist,
  method_aggl,
  k = NULL,
  h = NULL,
  cluster_prefix = "",
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_dist}{Character value or vector (see details for more). Denotes the distance methods (e.g. \emph{'euclidean'}) to be used. Run \code{validDistanceMethods()}
to obtain all valid input options.}

\item{method_aggl}{Character value or vector (see details for more). Denotes the agglomerative methods (e.g. \emph{'ward.D'}) to be used. Run \code{validAgglomerativeMethods()}
to obtain all valid input options.}

\item{cluster_prefix}{Character value. The character string with which the cluster names
are prefixed.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A tidy spata-data.frame containing the cluster variables.
}
\description{
Iterates over all valid combinations of arguments \code{method_dist, method_aggl, k}
and \code{h} and stores the resulting clustering variables in a data.frame. The function's
suffix indicates the data type from which the clustering derived.
}
\details{
The arguments mentioned in the description can be specifie as
vectors of any length. The variable names of the resulting data.frame
are glued together as an unambiguous combination. (e.g. \code{method_dist} = \emph{'euclidean'},
\code{method_aggl} = \emph{'complete'}, \code{k} = \emph{4} results in a clustering variable
named \emph{euclidean_complete_k_4})
}
