% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{findNearestNeighbourClusters}
\alias{findNearestNeighbourClusters}
\title{Cluster sample via nearest neighbour analysis}
\usage{
findNearestNeighbourClusters(
  object,
  n_pcs = 30,
  k = 50,
  searchtype = "priority",
  treetype = "bd",
  radius = 0,
  eps = 0,
  verbose = TRUE,
  of_sample = NA
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{k}{The maximum number of nearest neighbours to compute. The default value
is set to the smaller of the number of columnns in data.}

\item{treetype}{Character vector. Character vector specifying the standard
\emph{'kd'} tree or a \emph{'bd'} (box-decomposition, AMNSW98) tree which
may perform better for larger point sets.}

\item{eps}{Distance measure. Given to \code{eps} of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines
the size (radius) of the epsilon neighborhood.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{searchtypes}{Character value. Either \emph{'priority', 'standard'} or \emph{'radius '}. See details for more.}
}
\value{
A tidy spata-data.frame containing the cluster variables.
}
\description{
Cluster sample via nearest neighbour analysis
}
\details{
Search types: priority visits cells in increasing order of distance from the
query point, and hence, should converge more rapidly on the true nearest neighbour,
but standard is usually faster for exact searches. radius only searches for neighbours
within a specified radius of the point. If there are no neighbours then nn.idx will
contain 0 and nn.dists will contain 1.340781e+154 for that point.
}
\keyword{internal}
