% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{findOptimalDistanceSAS}
\alias{findOptimalDistanceSAS}
\title{Optimize \code{distance}-input}
\usage{
findOptimalDistanceSAS(
  object,
  id,
  variables,
  core = FALSE,
  start = recBinwidth(object) * 20,
  max = distToEdge(object, id),
  step = recBinwidth(object),
  sign_var = "fdr",
  sign_threshold = 0.05,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
This function suggests a value for the \code{distance}-parameter
of the \code{spatialAnnotationScreening()} algorithm by consecutively running it
while consecutively expanding the distance with the \code{binwidth} input. This is
done till the maximum distance is reached. For every run, the number of
valid best fits is counted. The distance with the most best fits is returned.
}
