% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdegs.R
\name{findSDEGS}
\alias{findSDEGS}
\title{Spatial Differential Gene Expression}
\usage{
findSDEGS(
  object,
  id,
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  genes_rm = character(0),
  variable.features.n = 3000,
  method_de = "wilcox",
  base = 2,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The image annotation of interest.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An S4 object of class \code{SDEGS}.
}
\description{
Identifies spatially differentially expressed genes (SDEGs) as
suggested by \emph{Zeng et al. (2023)}.
}
\references{
Zeng, H., Huang, J., Zhou, H. et al. Integrative in situ mapping of single-cell
transcriptional states and tissue histopathology in a mouse model of Alzheimer's
disease. Nat Neurosci 26, 430-446 (2023).
}
