% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{findSeuratClusters}
\alias{findSeuratClusters}
\title{Cluster sample via Seurat}
\usage{
findSeuratClusters(
  object,
  NormalizeData = list(),
  ScaleData = list(),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A tidy spata-data.frame containing the cluster variables.
}
\description{
Cluster sample via Seurat
}
\keyword{internal}
