% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{flexible_call_dummy}
\alias{flexible_call_dummy}
\title{flexible_call_dummy}
\usage{
flexible_call_dummy(...)
}
\arguments{
\item{...}{Allows to manipulate functions that are called 'flexibly'. Denote
the function name with the argument name and the way you want to manipulate
the way it is called with a named list of arguments. E.g. \code{facet_wrap =
list(drop = TRUE)}.

Use \code{validFlexiblyCalls()} to see all functions you can manipulate this
way.}
}
\description{
flexible_call_dummy
}
