% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{flip_coords_df}
\alias{flip_coords_df}
\title{Flip coordinate variables}
\usage{
flip_coords_df(
  df,
  axis,
  ranges,
  xvars = c("x", "xend", "col", "imagecol"),
  yvars = c("y", "yend", "row", "imagerow"),
  verbose = FALSE,
  error = FALSE,
  ...
)
}
\arguments{
\item{df}{Data.frame with numeric coordinate variables.}

\item{axis}{Character value. Denotes the axis around which the coordinates are flipped.
Either \emph{x} or \emph{y} to adress the coordinate
variables specifically or \emph{h} (horizontal, flips y-coords) or \emph{v} (vertical -
flips x-coords).}

\item{ranges}{A named list as returned by \code{getImageRange()}. Must at least
have one slot that is named like input for \code{axis}. This slot should
be a numeric vector of length two. First value being the axis minimum and
the second value being the axis maximum.}

\item{xvars, yvars}{Character vector. Names of the data.frame variables that
contain axis coordinates. If some of the names are not present in the
input data.frame: Depending on the input of \code{verbose} and \code{error}
the functions silently skips flipping, gives feedback or throws an error.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Adjusted data.frame.
}
\description{
Flip coordinate variables in a data.frame.
}
\keyword{internal}
