% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{flipCoordinates}
\alias{flipCoordinates}
\alias{flipCoordsDf}
\alias{flipImageAnnotations}
\alias{flipSpatialTrajectories}
\title{Flip coordinates}
\usage{
flipCoordinates(object, axis, verbose = FALSE)

flipCoordsDf(object, axis, verbose = NULL)

flipImageAnnotations(object, axis, verbose = NULL)

flipSpatialTrajectories(object, axis, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{axis}{Character value. Either \emph{'x'} or \emph{'y'}. Denotes the axis
around which the coordinates are flipped.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Flips coordinates to align with image in case
of non matching image and coordinates.
}
\details{
\code{flipCoordinates()} flips every spatial unit in the \code{SPATA2}
object around the denoted axis. \code{flipCoordsDf()}, \code{flipImageAnnotations()}
and \code{flipTrajectories()} flip single spatial aspects without affecting
the rest.
}
\seealso{
\code{flipImage()}
}
