% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-handling.R
\name{flipImage}
\alias{flipImage}
\title{Flip object image}
\usage{
flipImage(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated spata-object.
}
\description{
Flips image dimensions to align with coordinates in case
of non matching image and coordinates.
}
