% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-handling.R
\name{flipImageAndCoords}
\alias{flipImageAndCoords}
\title{Mirror invert the surface}
\usage{
flipImageAndCoords(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated spata-object.
}
\description{
Flips both and coordinates which results in mirror inverted plots.
}
