% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{get_names_dummy}
\alias{get_names_dummy}
\title{get_names}
\usage{
get_names_dummy()
}
\value{
Character vector of names. If no content is present the character
vector is of length 0.
}
\description{
get_names
}
