% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{get_spat_ann_helper}
\alias{get_spat_ann_helper}
\title{Obtain image annotation screening data.frame}
\usage{
get_spat_ann_helper(
  object,
  id,
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  n_bins_angle = 1,
  variables = NULL,
  method_gs = NULL,
  summarize_by = FALSE,
  summarize_with = "mean",
  normalize_by = "sample",
  normalize = FALSE,
  remove_circle_bins = FALSE,
  remove_angle_bins = FALSE,
  rename_angle_bins = FALSE,
  bcsp_exclude = NULL,
  drop = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{normalize_by}{Character value or FALSE. If character, there are two options:
\itemize{
\item{\code{normalize_by} = \emph{'sample'}:}{ Values are normalized across the whole sample.}
\item{\code{normalize_by} = \emph{'bins_angle'}:}{
Values are normalized within each angle bin. This only has an effect if \code{n_bins_angle}
is bigger than 1.
}
}}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{drop}{Logical value. If TRUE, unused levels of the \emph{bins_circle}
variables are dropped.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The final output depends on the input for \code{variables} and
\code{summarize_by}.

By default (both arguments are NULL) the returned data.frame contains
barcode-spots as observations/rows and variables that describe their position
to the image annotation denoted with \code{id}. This includes the variables
\emph{bins_circle}, \emph{bins_order}, \emph{angle}, \emph{bins_angle}. Their
content depends on the set up via the arguments \code{distance}, \code{binwidth}
and \code{n_bins_circle}.

\bold{Coordinates data.frame vs. Inferred expression changes}:

If argument \code{variables} is a character the denoted variables are
joined to the data.frame via \code{joinWith()}. If the set of variables
contains only numeric ones (genes, gene-sets and numeric features) the
function argument \code{summarize_by} can be set up in three different ways:

\itemize{
\item{\code{summarize_by} = \code{FALSE}:}{ Values are not summarized. The output
is a coordinates data.frame with each observation/row corresponding to
a barcode spots with additional information of its relation to the image
annotation denoted in \code{id}.}
\item{\code{summarize_by} = \emph{'bins_circle'}}{ Values of each variable
area summarized by each circular expansion of the polygon. This results
in data.frame with a column named \emph{bins_circle} containing the names of the bin
(\emph{Core, Circle 1, Circle 2, Circle 3, ..., Circle n, Outside}) and 1 column
per variable that contain the summarized expression value by circle bin. Visualization
of the concept can be obtained using \code{plotIasLineplot(..., facet_by = 'variables')}
}
\item{\code{summarize_by} = \emph{c('bins_circle', 'bins_angle'))}}{ Values of
each area are summarized by each circular expansion as well as by angle-bin.
Output data.frame is similar to \code{summarize_by} = \emph{'bins_circle'} apart
from having an extra column identifying the angle-bins. Adding \emph{'bins_circle'}
is only useful if \code{n_bins_circle} is bigger than 1. Visualization
of the concept can be obtained by using \code{plotIasLineplot(..., facet_by = 'bins_angle')}.
}}

Normalization in case of \code{normalize_by} != \code{FALSE} happens after the
summary step.
}
\description{
Extracts a data.frame that contains information about barcode-spots
needed for analysis related to \code{spatialAnnotationScreening()}.
}
\keyword{internal}
