% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getActive}
\alias{getActive}
\alias{getActive,SPATA2-method}
\title{Obtain name of active content}
\usage{
getActive(object, ...)

\S4method{getActive}{SPATA2}(object, what)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character value.
}
\description{
Gets the name of currently active content in the object.
}
