% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getActiveExpressionMatrixName}
\alias{getActiveExpressionMatrixName}
\title{Deprecated}
\usage{
getActiveExpressionMatrixName(...)
}
\description{
Deprecated in favor of \code{\link[=activeMatrix]{activeMatrix()}}.
}
\keyword{internal}
