% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getAnnotationNames}
\alias{getAnnotationNames}
\alias{addAnnotationVariable}
\alias{discardAnnotationVariable}
\title{Obtain annotation names}
\usage{
getAnnotationNames(object, fdb_fn = "message", ...)

addAnnotationVariable(object, name, verbose = NULL, ...)

discardAnnotationVariable(object, name, verbose = NULL, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Character vector.
}
\description{
Extracts tha names of the variables that have been created
via \code{addAnnotationInteractively()}.
}
