% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getAssay}
\alias{getAssay}
\title{Obtain molecular assay}
\usage{
getAssay(object, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Assay data corresponding to the specified name.
}
\description{
Retrieves an object of class \code{\link{MolecularAssay}} from the provided object.
}
\details{
This function retrieves assay data from the provided object based on the specified assay name. It internally calls \code{\link[=containsAssay]{containsAssay()}} to ensure that the assay exists in the object.
}
\seealso{
\code{\link[=activeAssay]{activeAssay()}}
}
