% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getAssayModalities}
\alias{getAssayModalities}
\alias{getAssayNames}
\title{Obtain assay names/modalites}
\usage{
getAssayModalities(object)

getAssayNames(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
A character vector containing the names of assays.
}
\description{
Retrieves the names and modalities of assays present in the provided object.

Since the name of an assay should be identical with its molecular modality both functions
should return the same output. If they don't, something is wrong.
}
\seealso{
\code{\link[=getAssay]{getAssay()}}
}
