% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getAssayNames}
\alias{getAssayNames}
\title{Obtain assay names}
\usage{
getAssayNames(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
A character vector containing the names of assays.
}
\description{
Retrieves the names of assays present in the provided object.
}
\seealso{
\code{\link[=getAssay]{getAssay()}}
}
