% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{getAutoencoderAssessment}
\alias{getAutoencoderAssessment}
\title{Obtain information about the optimal neural network set up}
\usage{
getAutoencoderAssessment(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A data.frame containing the total variance measured by \code{irlba::prcomp_irlba()} after each
combination of activations/bottlenecks.
}
\description{
Extracts the results from \code{assessAutoencoderOptions()}.
}
\keyword{internal}
