% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{getAutoencoderSetUp}
\alias{getAutoencoderSetUp}
\title{Obtain information on neural network}
\usage{
getAutoencoderSetUp(object, mtr_name, of_sample = NA)
}
\value{
A named list.
}
\description{
Returns the argument input that was chosen to construct the
neural network that generated the matrix denoted in \code{mtr_name}.
}
