% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getBackgroundColor}
\alias{getBackgroundColor}
\alias{getBackgroundColor,spata2-method}
\alias{getBackgroundColor,HistoImaging-method}
\alias{getBackgroundColor,HistoImage-method}
\title{Obtain background color}
\usage{
getBackgroundColor(object, ...)

\S4method{getBackgroundColor}{spata2}(object, img_name = NULL, default = "white", ...)

\S4method{getBackgroundColor}{HistoImaging}(object, img_name = NULL, default = "white", ...)

\S4method{getBackgroundColor}{HistoImage}(object, default = "white")
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{default}{Color to default to if no background color is set.}
}
\value{
Character value.
}
\description{
Extracts results of \code{\link[=identifyBackgroundColor]{identifyBackgroundColor()}}.
}
