% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Obtain barcodes}
\usage{
getBarcodes(object, across = NULL, across_subset = NULL, simplify = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}
}
\value{
Named character vector or list.
}
\description{
Returns a character vector of barcode names. See details for more.
}
\details{
If argument \code{across} is specified the output is named according
to the group membership the variable specified assigns the barcode spots to.
If simplify is set to FALSE a list is returned.

Not specifying \code{across} makes the function return an unnamed character
vector containing all barcodes.
}
