% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getBarcodesInPolygon}
\alias{getBarcodesInPolygon}
\title{Obtain barcodes in polygon}
\usage{
getBarcodesInPolygon(object, polygon_df, strictly = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{polygon_df}{A data.frame that contains the vertices of the polygon
in form of two variables: \emph{x} and \emph{y}.}

\item{strictly}{Logical value. If \code{TRUE}, only barcode spots that are strictly
interior to the polygon are returned. If \code{FALSE}, barcodes that are
on the relative interior the polygon border or that are vertices themselves
are returned, too.}
}
\value{
Character vector.
}
\description{
Extracts barcodes of barcode-spots that fall in a given
polygon. Works closely with \code{sp::point.in.polygon()}.
}
