% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getBarcodeSpotDistances}
\alias{getBarcodeSpotDistances}
\title{Obtain barcode spot distances}
\usage{
getBarcodeSpotDistances(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data.frame.
}
\description{
Computes the distance from every barcode spot to every other
barcode spot.
}
\details{
The output data.frame has a number of rows that is equal to
\code{nBarcodes(object)^2}
}
