% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCCD}
\alias{getCCD}
\title{Obtain center to center distance}
\usage{
getCCD(object, unit = NULL, as_numeric = FALSE, round = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{unit}{Character value or \code{NULL}. If character, specifies
the unit in which the distance is supposed to be returned.
Use \code{validUnitsOfLength()} to obtain  all valid input options.}
}
\value{
Character value.
}
\description{
Extracts the center to center distance from
barcode-spots depending on the method used.
}
