% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-analysis.R, R/getA-G.R
\name{getCnvFeatureNames}
\alias{getCnvFeatureNames}
\title{Obtain features names under which cnv-analysis results are stored.}
\usage{
getCnvFeatureNames(object, ...)

getCnvFeatureNames(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character vector.

Character vector.
}
\description{
Returns a character vector of feature names referring to the
barcode-spots chromosomal gains and losses as computed by \code{runCnvAnalysis()}.

Returns a character vector of feature names referring to the
barcode-spots chromosomal gains and losses as computed by \code{runCnvAnalysis()}.
}
