% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCnvGenesDf}
\alias{getCnvGenesDf}
\title{Obtain CNV results by gene}
\usage{
getCnvGenesDf(object, add_meta = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{add_meta}{Logical value. If TRUE, meta information obtained by
\code{getGenePosDf()} every gene is added to the data.frame}
}
\value{
Data.frame.
}
\description{
Extracts CNV results in form of barcode ~ pairs in a data.frame.
}
