% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/get-family.R, R/getA-G.R
\name{getSegmentDf}
\alias{getSegmentDf}
\alias{getCoordsDf}
\title{Obtain spatial coordinates}
\usage{
getSegmentDf(object, segment_names, of_sample = NA)

getCoordsDf(object, type = "both", ...)

getSegmentDf(object, segment_names, of_sample = NA)

getCoordsDf(object, type = "both", ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{segment_names}{Character vector. Specifies the segments of interest.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A data.frame containing the variables \emph{barcods, sample, x, y}
(and \emph{segmentation} in case of \code{getSegmentDf()}).

A data.frame containing the variables \emph{barcods, sample, x, y}
(and \emph{segmentation} in case of \code{getSegmentDf()}).
}
\description{
Obtain spatial coordinates

Obtain spatial coordinates
}
\keyword{internal}
