% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCoordsDfSA}
\alias{getCoordsDfSA}
\title{Relate points to spatial annotations}
\usage{
getCoordsDfSA(
  object,
  id = idSA(object),
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  verbose = NULL,
  dist_unit = "px",
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{\link[=joinWithVariables]{joinWithVariables()}}. Only used
if not empty.}
}
\value{
Data.frame.
}
\description{
Adds the spatial relation to a spatial
annotation to the coordinates data.frame. See details for more.
}
\details{
The coordinates data.frame as returned by \code{\link[=getCoordsDf]{getCoordsDf()}} with five
additional variables:

\itemize{
\item{\emph{dist}:}{ Numeric. The distance of the data point to the outline of the spatial annotation.}
\item{\emph{bins_dist}:}{ Factor. The bin the data point was assigned to based on its \emph{dist} value and the \code{binwidth}.}
\item{\emph{angle}:}{ Numeric. The angle of the data point to the center of the spatial annotation.}
\item{\emph{bins_angle}:}{ Factor. The bin the data point was assigned to based on its \emph{angle} value.}
\item{\emph{rel_loc}:}{ Character. Possible values are \emph{'core'}, if the data point lies inside the spatial annotation,
\emph{'periphery'} if the data point lies outside of the boundaries of the spatial annotation but inside
the area denoted via \code{distance} and \emph{outside}, if the data point lies beyond the screening area (it's
distance to the spatial annotation boundaries is bigger than the value denoted in \code{distance}).}
}
}
