% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCoordsMtr}
\alias{getCoordsMtr}
\title{Obtain coordinates matrix}
\usage{
getCoordsMtr(object, img_name = activeImage(object), orig = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{img_name}{Character value. The name of the image the coordinates are
scaled to. If \code{NULL}, defaults to the active image.}

\item{orig}{Logical value. If \code{TRUE}, the coordinates are not scaled to any
image.}
}
\value{
A matrix.
}
\description{
Wraps the coordinates in a matrix with column names \emph{x} and \emph{y}
and rownames that correspond to the barcodes.
}
\details{
In contrast to \code{\link[=getCoordsDf]{getCoordsDf()}}, column names of the output matrix
are always named \emph{x} and \emph{y}, regardless of whether they correspond to the
original x- and y-coordiantes (\emph{x_orig}, \emph{y_orig}) or if they are scaled
to the image specified with \code{img_name}. The input for argument \code{orig}
decides!
}
