% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getCoordsRange}
\alias{getCoordsRange}
\title{Obtain coordinate range}
\usage{
getCoordsRange(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A list of two vectors each of length 2.
}
\description{
Extracts the range of the x- and y-coordinates.
}
