% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getN-Z.R
\name{getCountMatrix}
\alias{getCountMatrix}
\alias{getProcessedMatrix}
\title{Obtain data matrices}
\usage{
getCountMatrix(object, ...)

getProcessedMatrix(object, mtr_name)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A matrix with rownames corresponding to the features and
column names corresponding to the barcodes.
}
\description{
Extracts count or processed data matrices.
}
