% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getDeaLfcName}
\alias{getDeaLfcName}
\title{Obtain LFC name}
\usage{
getDeaLfcName(object, across = getDefaultGrouping(object), method_de = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}
}
\value{
Character value.
}
\description{
Extracts name of variable that contains log fold change results
of DEA.
}
