% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R
\name{getDeaOverview}
\alias{getDeaOverview}
\title{Obtain info on de-analysis storage}
\usage{
getDeaOverview(object)

getDeaOverview(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A summarizing list.

A summarizing list.
}
\description{
Obtain info on de-analysis storage

Obtain info on de-analysis storage
}
