% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getDeaGenes}
\alias{getDeaGenes}
\alias{getDeaResultsDf}
\title{Obtain DEA results}
\usage{
getDeaGenes(
  object,
  across = getDefaultGrouping(object),
  across_subset = NULL,
  method_de = "wilcox",
  max_adj_pval = NULL,
  min_lfc = 0,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  flatten = TRUE,
  assay_name = activeAssay(object),
  ...
)

getDeaResultsDf(
  object,
  across = getDefaultGrouping(object),
  across_subset = NULL,
  relevel = FALSE,
  method_de = "wilcox",
  max_adj_pval = NULL,
  min_lfc = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  stop_if_null = TRUE,
  assay_name = activeAssay(object),
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{max_adj_pval}{Numeric value. Sets the threshold for adjusted p-values. All genes
with adjusted p-values above that threshold are ignored.}

\item{min_lfc}{Numeric value. Sets the threshold for average log fold change. All genes
with an average log fold change below that threshold are ignored.}

\item{n_highest_lfc}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{n_lowest_pval}{Numeric value. Affects the total number of genes that are kept. See details.}
}
\value{
A data.frame:

\itemize{
\item{\emph{gene}} Character. The differentially expressed genes.
\item{\emph{'across'}} Character. The grouping across which the analysis was performed. The variable/column name is
equal to the input for argument \code{across}.
\item{\emph{avg_logFC}} Numeric. The average log-fold change to which the belonging gene was differentially expressed..
\item{\emph{p_val}} Numeric. The p-values.
\item{\emph{p_val_adj}} Numeric. The adjusted p-values.
}
}
\description{
Extracts differential expression
analysis results. Function \code{getDeaGenes()} is a wrapper around
\code{getDeaResultsDf()} and returns only gene names in a character vector.
}
\details{
The de-data.frame is processed such that the following steps are performed for every experimental
group.

\enumerate{
\item{Discards genes with \emph{avg_logFC}-values that are either infinite or negative}
\item{Discards genes with adjusted p-values above the threshold set with \code{max_adj_pval}}
\item{Discard genes with average log fold change below the treshold set with \code{min_lfc}}
\item{Slices the data.frame in order that for every experimental group:}
\enumerate{
\item{the n genes with the highest \emph{avg_logFC}-values are kept where n = \code{n_highest_lfc}}
\item{the n genes with the lowest \emph{p_val_adj}-values are kept where n = \code{n_lowest_pval}}
}
\item{Arranges the genes according to the highest \emph{avg_logFC}-values}
}
}
