% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getDefaultGrouping}
\alias{getDefaultGrouping}
\title{Deprecated}
\usage{
getDefaultGrouping(object, ...)
}
\description{
Deprecated in favor of \code{\link[=activeGrouping]{activeGrouping()}}.
}
\keyword{internal}
