% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getDefaultUnit}
\alias{getDefaultUnit}
\title{Obtain default unit}
\usage{
getDefaultUnit(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character value.
}
\description{
Extracts the default unit of the spatial method the
\code{spata2} object relies on.
}
