% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getN-Z.R
\name{getDimRedDf}
\alias{getDimRedDf}
\alias{getPcaDf}
\alias{getPcaMtr}
\alias{getTsneDf}
\alias{getUmapDf}
\title{Obtain dim red data.frame}
\usage{
getDimRedDf(object, method_dr = c("pca", "tsne", "umap"), of_sample = NA)

getPcaDf(object, n_pcs = 30, of_sample = NA)

getPcaMtr(object, n_pcs = 30, of_sample = NA)

getTsneDf(object, of_sample = NA)

getUmapDf(object, of_sample = NA)
}
\description{
Obtain dim red data.frame
}
