% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\title{Deprecated}
\usage{
getExpressionMatrix(object, ...)
}
\description{
Deprecated in favor of \code{\link[=getProcessedMatrix]{getProcessedMatrix()}}.
}
\keyword{internal}
