% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getExpressionMatrixNames}
\alias{getExpressionMatrixNames}
\title{Obtain names of stored expression matrices}
\usage{
getExpressionMatrixNames(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
Character vector.
}
\description{
Obtain names of stored expression matrices
}
