% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getFeatureDf}
\alias{getFeatureDf}
\title{Obtain feature data}
\usage{
getFeatureDf(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
The feature data data.frame of the specified object and sample(s).
}
\description{
Obtain feature data
}
