% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getFeatureNames}
\alias{getFeatureNames}
\title{Obtain feature names}
\usage{
getFeatureNames(object, of_class = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{of_class}{Character vector. Specify the classes a feature must be of for
it's name to be returned.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A named character vector of the variables in the feature data slot.
}
\description{
Extracts names of features \strong{from the meta data.frame}.
}
