% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureValues}
\alias{getFeatureValues}
\title{Obtain unique categorical feature values}
\usage{
getFeatureValues(object, features, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{features}{The features of interest specified as a character vector.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A vector or a named list according to the length of \code{features}.
}
\description{
Extracts the unique values of discrete features.
}
