% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-family.R
\name{getFromSeurat}
\alias{getFromSeurat}
\title{Safe extraction}
\usage{
getFromSeurat(return_value, error_handling, error_value, error_ref)
}
\arguments{
\item{return_value}{Whatever needs to be extracted.}

\item{error_handling}{Either \emph{'warning} or \emph{'stop'}.}

\item{error_value}{What is supposed to be returned if extraction fails.}

\item{error_ref}{The reference for the feedback message.}
}
\description{
A wrapper around \code{base::tryCatch()} with predefined error handling
messages if extraction from seurat-object failed.
}
