% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R
\name{getGeneCounts}
\alias{getGeneCounts}
\title{Obtain total number of gene counts}
\usage{
getGeneCounts(object, of_sample = NA, return = "tibble")

getGeneCounts(object, of_sample = NA, return = "tibble")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{return}{Character value. One of \emph{'data.frame', 'tibble' or 'vector'}.
Specifies the output class.}
}
\value{
Depends on input for argument \code{return}.

Depends on input for argument \code{return}.
}
\description{
Obtain total number of gene counts

Obtain total number of gene counts
}
