% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/unused.R
\name{getGeneDistMtr}
\alias{getGeneDistMtr}
\title{Obtain distance measurements of spatially correlated genes}
\usage{
getGeneDistMtr(object, of_sample = NA)

getGeneDistMtr(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A data.frame or a distance matrix.

A data.frame or a distance matrix.
}
\description{
Obtain distance measurements of spatially correlated genes

Obtain distance measurements of spatially correlated genes
}
\keyword{internal}
