% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGeneFeatureNames}
\alias{getGeneFeatureNames}
\title{Obtain feature names of the gene meta data}
\usage{
getGeneFeatureNames(object, mtr_name = NULL, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}
}
\description{
Convenient way to obtain the column names of the output data.frame
of \code{getGeneMetaDf()}.
}
