% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGeneMetaData}
\alias{getGeneMetaData}
\alias{getGeneMetaDf}
\title{Obtain gene meta data}
\usage{
getGeneMetaData(object, mtr_name = NULL, only_df = FALSE, ...)

getGeneMetaDf(object, mtr_name = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}

\item{only_df}{Logical. If set to TRUE only the data.frame is returned.
If set to FALSE (the default) the whole list is returned.}
}
\value{
A data.frame from \code{getMetaDataDf()} or a list from \code{getGeneMetaData()}.
}
\description{
Obtain gene meta data
}
