% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternCluster}
\alias{getGenePatternCluster}
\title{Title}
\usage{
getGenePatternCluster(
  object,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  k = NULL,
  h = NULL,
  cluster_prefix = "cluster_",
  pattern_suffix = TRUE,
  return = "list",
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
