% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternCorrelations}
\alias{getGenePatternCorrelations}
\title{Title}
\usage{
getGenePatternCorrelations(
  object,
  threshold_sim = 0.25,
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
