% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternDf}
\alias{getGenePatternDf}
\alias{getGenePatternCoordsDf}
\title{Title}
\usage{
getGenePatternDf(
  object,
  genes_subset = NULL,
  pattern_subset = NULL,
  barcode_info = TRUE,
  unnest = TRUE,
  verbose = NULL
)

getGenePatternCoordsDf(object, genes = NULL, verbose = NULL)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
