% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternDistances}
\alias{getGenePatternDistances}
\alias{getGenePatternSimilarities}
\title{Obtain similarity scores between identified gene patterns}
\usage{
getGenePatternDistances(
  object,
  return = "matrix",
  threshold_dist = 0.5,
  of_sample = NA
)

getGenePatternSimilarities(
  object,
  return = "matrix",
  threshold_sim = 0.5,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{return}{Character value. If set to \emph{'matrix'} a similarity matrix
is returned. If set to \emph{'data.frame' or 'tibble'} the respective type
is returned.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{threshold_sim}{Numeric value. Makes sure that only gene-patterns
are kept that are at least to one other pattern equal or more similar
than the threshold determines. 0 effectively skips filtering.}
}
\value{
Similarity matrix or data.frame.
}
\description{
Obtain similarity scores between identified gene patterns
}
