% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternEvalDf}
\alias{getGenePatternEvalDf}
\title{Title}
\usage{
getGenePatternEvalDf(
  object,
  genes = NULL,
  gene_patterns = NULL,
  pattern_subset = NULL,
  of_sample = NA
)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
