% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternExtentDf}
\alias{getGenePatternExtentDf}
\title{Title}
\usage{
getGenePatternExtentDf(
  object,
  genes = NULL,
  gene_patterns = NULL,
  pattern_subset = NULL,
  join_with_expr = TRUE,
  smooth = NULL,
  smooth_span = NULL,
  verbose = NULL,
  dbscan_display = FALSE,
  dbscan_remove = FALSE,
  of_sample = NA
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
