% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-helper-functions.R
\name{getGenePatternsAboveThreshold}
\alias{getGenePatternsAboveThreshold}
\title{Title}
\usage{
getGenePatternsAboveThreshold(
  object,
  threshold_sim,
  of_sample = NA,
  verbose = FALSE
)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
