% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGenePosDf}
\alias{getGenePosDf}
\title{Obtain gene information}
\usage{
getGenePosDf(object, keep = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{keep}{Logical value, TRUE the columns \emph{ensemble_gene_id} and
\emph{hgnc_symbol} are included. The content of \emph{hgnc_symbol} is
identical to the content of column \emph{genes}.}
}
\value{
Data.frame.
}
\description{
Extracts information regarding gene positioning
on chromosomes and/or chromosome arms.
}
