% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPositions.R
\name{getGenePositions}
\alias{getGenePositions}
\title{Receive genomic coordinates of a gene list}
\usage{
getGenePositions(
  gene_names = character(0),
  ensembl_version = "https://feb2023.archive.ensembl.org",
  species = "human",
  ignoreAlt = F
)
}
\arguments{
\item{gene_names}{A vector of gene names in HUGO format.}

\item{ensembl_version}{Version of the ENSEMBL database used to quantify gene expression data. Default: v109.}

\item{ignoreAlt}{If set to TRUE: Ignore if multiple loci are reported for a gene, pick the one from the primary assembly.}
}
\description{
This function allows to receive the genomic positions of a vector of genes in HUGO format.
}
\examples{
getGenePositions(gene_names = c("EGFR", "PDGFRA"))
}
\keyword{Chromosomal}
\keyword{positions}
