% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGenes}
\alias{getGenes}
\alias{getGenesInteractive}
\title{Obtain gene names}
\usage{
getGenes(
  object,
  of_gene_sets = NULL,
  similar_to = NULL,
  top_n = 25,
  simplify = TRUE,
  of_sample = NA,
  in_sample = NA
)

getGenesInteractive(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_gene_sets}{A character vector specifying the gene sets from which to
return the gene names.}

\item{similar_to}{Character value. If specified, n genes with the highest
spatial correlation similarity to the specified gene are returned where n
is equal to the input for argument \code{top_n}.}

\item{top_n}{Numeric value. Denotes the number of genes to be returned if argument
\code{similar_to} is specified.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{in_sample}{Deprecated.}
}
\value{
A list of character vectors or a single character vector.
}
\description{
A convenient way to extract gene names.
}
\details{
If neither \code{of_gene_sets} nor \code{similar_to} is specified all
genes found in the active expression matrix are returned in a character vector.

If \code{of_gene_sets} is specified a list named according to the input of
\code{of_gene_sets} is returned in which each element is a character vector
containing the names of genes the specific gene set is composed of. Is simplified
to a vector if \code{simplify} is set to TRUE.

If \code{of_gene_sets} is not specified but argument \code{similar_to} is
a gene name a character vector of genes featuring the highest
similarity to the specified gene is returned. The number of genes depends on
the input of argument \code{top_n}.
}
