% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneSetOverview}
\alias{getGeneSetOverview}
\title{Overview about the current gene sets}
\usage{
getGeneSetOverview(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A data.frame with two variables \emph{Class} and \emph{Available Gene
Sets} indicating the number of different gene sets the classes contain.
}
\description{
Overview about the current gene sets
}
