% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneSets}
\alias{getGeneSets}
\alias{getGeneSetsInteractive}
\title{Obtain gene set names}
\usage{
getGeneSets(object, of_class = "all", index = NULL, simplify = TRUE)

getGeneSetsInteractive(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_class}{A character vector indicating the classes from which to obtain
the gene set names. (Which classes exist in the current gene set data.frame can
be obtained e.g. with \code{printGeneSetOverview()}). If set to \emph{"all"} all
gene sets are returned.}

\item{index}{A regular expression according to which the gene set names to be returned
are filtered again.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}
}
\value{
A list named according to the input of argument \code{of_class}. Each element of
the returned list is a character vector containing the names of gene sets of the specified classes.
The list is coalesced to an unnamed vector if \code{simplify} is set to TRUE.
}
\description{
Obtain gene set names
}
