% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGridVisiumHD}
\alias{getGridVisiumHD}
\title{Obtain grid information}
\usage{
getGridVisiumHD(object, res, img_name = activeImage(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{res}{A \link[=concept_distance_measure]{distance value} specifying the desired
resolution for the grid in micrometers (e.g., "32um"). The resolution must be greater
than or equal to the current resolution and divisible by the current resolution.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
A data frame with the following columns:
\itemize{
\item{idx}{: A character vector representing the identifier for each segment, typically indicating the row or column number (e.g., "row_9").}
\item{x}{: A numeric vector representing the x-coordinate of the starting point of the segment.}
\item{y}{: A numeric vector representing the y-coordinate of the starting point of the segment.}
\item{xend}{: A numeric vector representing the x-coordinate of the ending point of the segment.}
\item{yend}{: A numeric vector representing the y-coordinate of the ending point of the segment.}
\item{just}{: A character vector indicating the orientation of the segment (e.g., "horizontal").}
}
}
\description{
Generates a data.frame of grid elements for visualizing a grid overlay
on VisiumHD spatial data.
}
\keyword{internal}
