% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v3_1_0.R
\name{getGridVisiumHDx}
\alias{getGridVisiumHDx}
\title{Generate grid segments for Visium HD data}
\usage{
getGridVisiumHDx(object, res, img_name = activeImage(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{res}{Character string specifying the desired resolution (e.g., "10um").
The resolution must be lower or equal to the current resolution of the data and divisible by the current resolution.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
A data frame containing the grid segments with their start and end coordinates (x, y, xend, yend),
segment type (horizontal or vertical), and other relevant information.
}
\description{
Generates grid segments at a specified resolution for Visium HD data.
The grid can be used to visualize or process the spatial layout of the Visium HD data.
}
\details{
The function performs the following steps:
\itemize{
\item Verifies the desired resolution is compatible with the current resolution of the Visium HD data.
\item Computes the grid segments based on the resolution and spatial layout of the data.
\item Scales the grid segments to match the image's scale factor.
}
}
\examples{
\dontrun{
object <- downloadSpataObject("HumanPancreasHD")
grid_df <- getGridVisiumHD(object, res = "32um")
plotGrid(grid_df) # Assuming a function that plots the grid
}

}
