% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\title{Obtain variable names that group data points}
\usage{
getGroupingOptions(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector.
}
\description{
Extracts the names of the features of class \emph{factor} which
are valid input options for the arguments \code{grouping}, \code{grouping_variable},
and \code{across}.
}
