% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupNames}
\alias{getGroupNames}
\title{Obtain group names a grouping variable contains}
\usage{
getGroupNames(object, grouping_variable, of_sample = NA, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
Character vector
}
\description{
Obtain group names a grouping variable contains
}
\examples{
#Not run:

 # obtain all group names the variable 'my_cluster'
 # contains

 getGroupNames(object = object, grouping_variable = "my_cluster")

}
