% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGseaResults}
\alias{getGseaResults}
\title{Obtain enrichment results}
\usage{
getGseaResults(
  object,
  across = getDefaultGrouping(object, verbose = TRUE, "across"),
  across_subset = NULL,
  method_de = NULL,
  flatten = TRUE,
  stop_if_null = TRUE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}
}
\value{
A list or an object of class \code{hyp}.
}
\description{
Extracts the results from gene set enrichment analysis
in form of either a list (if \code{reduce} was set to TRUE) or
an object of class \code{hyp} (if \code{reduce was set to FALSE}).
}
