% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-implementation.R
\name{getHclustObject}
\alias{getHclustObject}
\alias{getHclustObjectCnv}
\title{Obtain hierarchical cluster results}
\usage{
getHclustObject(object, method_dist = NULL, method_aggl = NULL, of_sample = NA)

getHclustObjectCnv(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_dist}{Character value or vector (see details for more). Denotes the distance methods (e.g. \emph{'euclidean'}) to be used. Run \code{validDistanceMethods()}
to obtain all valid input options.}

\item{method_aggl}{Character value or vector (see details for more). Denotes the agglomerative methods (e.g. \emph{'ward.D'}) to be used. Run \code{validAgglomerativeMethods()}
to obtain all valid input options.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
An object of class \emph{hclust}.
}
\description{
Convenient access to results of hierarchical
clustering analysis. The functions suffix denotes the data type
from which the clustering derived.
}
\details{
Input for \code{method_*}-arguments must be character values.
}
