% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getHistoImage}
\alias{getHistoImage}
\alias{getHistoImage,HistoImaging-method}
\alias{getHistoImageActive}
\alias{getHistoImageActive,HistoImaging-method}
\alias{getHistoImageRef}
\alias{getHistoImageRef,HistoImaging-method}
\title{Obtain object of class \code{HistoImage}}
\usage{
getHistoImage(object, ...)

\S4method{getHistoImage}{HistoImaging}(object, name = NULL, ...)

getHistoImageActive(object, ...)

\S4method{getHistoImageActive}{HistoImaging}(object)

getHistoImageRef(object, ...)

\S4method{getHistoImageRef}{HistoImaging}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{}
}
\description{
Extracts the S4-containers of registered images. Note that
slot @image might be empty. Use \code{loadImage()} in that case.

\itemize{
\item{\code{getHistoImage()}:}{ Extracts object by name. If \code{name = NULL} the active histology image is returned.}
\item{\code{getHistoImageActive()}:}{ Extracts the active \code{HistoImage} object.}
\item{\code{getHistoImageRef()}:}{ Extracts the reference \code{HistoImage} object.}
}
}
