% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getHistoImageNames}
\alias{getHistoImageNames}
\alias{getHistoImageNames,HistoImaging-method}
\title{Obtain names of registered \code{HistoImage} objects}
\usage{
getHistoImageNames(object, ...)

\S4method{getHistoImageNames}{HistoImaging}(object, ref = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{ref}{Logical value. If \code{FALSE}, name of the reference image is not
included.}
}
\value{
Character vector.
}
\description{
Extracts the names of the \code{HistoImage} objects currently
registered in the object.
}
