% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getHistoImaging}
\alias{getHistoImaging}
\alias{getImageObject}
\title{Obtain object of class \code{HistoImaging}}
\usage{
getHistoImaging(object)

getImageObject(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Object of class \code{HistoImaging}.
}
\description{
Extracts the S4-object used as a container for
images.
}
\note{
\code{getImageObject()} is deprecated as of version v3.0.0 in favor
of \code{getHistoImaging()}.
}
\seealso{
\code{\link[=getImage]{getImage()}},\code{\link[=getHistoImage]{getHistoImage()}}
}
\keyword{internal}
