% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getHistologyImage}
\alias{getHistologyImage}
\alias{getHistologyImage,HistologyImagingNew-method}
\alias{getHistologyImageActive}
\alias{getHistologyImageActive,HistologyImagingNew-method}
\alias{getHistologyImageRef}
\alias{getHistologyImageRef,HistologyImagingNew-method}
\title{Obtain object of class \code{HistologyImage}}
\usage{
getHistologyImage(object, ...)

\S4method{getHistologyImage}{HistologyImagingNew}(object, name = NULL, ...)

getHistologyImageActive(object, ...)

\S4method{getHistologyImageActive}{HistologyImagingNew}(object)

getHistologyImageRef(object, ...)

\S4method{getHistologyImageRef}{HistologyImagingNew}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{}
}
\description{
Extracts the S4-containers of registered images. Note that
slot @image might be empty. Use \code{loadImage()} in that case.

\itemize{
\item{\code{getHistologyImage()}:}{ Extracts object by name. If \code{name = NULL} the active histology image is returned.}
\item{\code{getHistologyImageActive()}:}{ Extracts the active histology image.}
\item{\code{getHistologyImageRef()}:}{ Extracts the reference histology image.}
}
}
