% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getIasExpansion}
\alias{getIasExpansion}
\title{Obtain expanded image annotation polygons}
\usage{
getIasExpansion(
  object,
  id,
  distance = NA_integer_,
  binwidth = getCCD(object),
  n_bins_circle = NA_integer_,
  direction = "outwards",
  inc_outline = TRUE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{distance}{Distance value. Specifies the distance from the border of the
image annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{binwidth}{Distance value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the center-center
distance: \code{binwidth = getCCD(object)}. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
(See details for more.)}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
List of data.frames.
}
\description{
Expands polygons of image annotations according
to \code{distance}, \code{binwidth} and \code{n_bins_circle} input.
}
