% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageAnnotation}
\alias{getImageAnnotation}
\title{Obtain object of class \code{ImageAnnotation}}
\usage{
getImageAnnotation(object, id, add_image = TRUE, expand = 0, square = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See details
for more information.}

\item{square}{Logical value. Most image annotations come in variable shapes and
have different horizontal and vertical diameters. Therefore, height and width of the image
section are usually not equal. If \code{square = TRUE}, the cropped section of the image that
contains the annotated structure is forced into a square: the bigger diameter of both is taken
as default. E.g. if the horizontal diameter of the image annotation is 1mm and the
vertical diameter is 1.5mm, the output image will have height and width of 1.5mm. That is,
in terms of coordinates, an x-range and a y-range of 1.5mm.

Processing of the image output depending on argument \code{expand} happens afterwards.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts object of class \code{ImageAnnotaion} by
its id.
}
