% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImageAnnotationAreaDf}
\alias{getImageAnnotationAreaDf}
\title{Deprecated}
\usage{
getImageAnnotationAreaDf(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnOutlineDf]{getSpatAnnOutlineDf()}}.
}
\keyword{internal}
