% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageAnnotationAreaSize}
\alias{getImageAnnotationAreaSize}
\title{Obtain area of image annotation}
\usage{
getImageAnnotationAreaSize(
  object,
  unit,
  ids = NULL,
  tags = NULL,
  test = "any",
  as_numeric = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{unit}{Character value. Specifies the desired unit.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags, test}{Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are multiple options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags. Can contain tags that are not specified.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags. Can not be tagged with anything else.
\item Argument \code{test} set to \emph{not_identical} or \emph{4}: To be included, an image
annotation must \strong{not} be tagged with the combination of input tags.
\item Argument \code{test} set to \emph{'none'} or \emph{5}: To be included, an image annotation
must \strong{not} contain any of the input tags.
}

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Numeric vector of the same length as \code{ids}. Named accordingly.
Contains the area of the image annotations in the unit that is specified in \code{unit}.
The unit is attached to the output as an attribute named \emph{unit}. E.g. if
\verb{unit = *mm2*} the output value has the unit \emph{mm^2}.
}
\description{
Computes the area of an image annotation in units of area.
}
\details{
First, the area of each pixel is calculated by using the center
to center distance of the underlying method as ground truth to scale
the current number of pixels in the image.

Second, the number of pixels that fall in the area of the image annotation
is computed with \code{sp::point.in.polygon()} using the area data.frame of
the image annotation as the polygon.

Third, the number of pixels that fall in the area is multiplied with
the area per pixel.
}
\seealso{
\code{getImageAnnotationAreaDf()}, \code{getCCD()}, \code{as_unit()}
}
