% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationCenterBcsp}
\alias{getImageAnnotationCenterBcsp}
\title{Obtain center barcode-spot}
\usage{
getImageAnnotationCenterBcsp(object, id)
}
\arguments{
\item{object}{A valid spata-object.}

\item{id}{Character value. The ID of the image annotation of interest.}
}
\value{
Data.frame as returned by \code{getCoordsDf()} with one row.
}
\description{
Extracts the barcode spot that lies closest
to the center of the image annotation.
}
