% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationDf}
\alias{getImageAnnotationDf}
\title{Obtain image annotation data.frame}
\usage{
getImageAnnotationDf(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  sep = " & ",
  last = " & "
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}
}
\value{
A data.frame that contains the grouping variables \emph{id} and \emph{tags}
and the numeric variables \emph{x} and \emph{y}. The returned data.frame can contain
the spatial extent of more than just
one annotated structure, depending on the input of arguments \code{ids}, \code{tags}
and \code{test}. The variable \emph{ids} of the returned data.frame is used to
uniquely mark the belonging of each x- and y-coordinate.
}
\description{
Extracts the coordinates of the polygon that was drawn to
annotate structures in the histology image in a data.frame.
}
\details{
After the filtering by image annotation ID you can use a combination
of the arguments \code{tags} and \code{test} to filter for the image annotations of
interest. Input for argument \code{tags} specifies the tags of interest. With argument
\code{test} set to \emph{'any'} you make the function include all image annotations
that were tagged with at least one tag of the input for argument \code{tags}.
If \code{test} is set to \emph{'all'} an image annotation must contain all
tags of \code{tags} to be included.
}
\note{
The variables \emph{x} and \emph{y} correspond to the coordinates
with which the annotated structure was denoted in \code{annotateImage()}. These
coordinates are not to be confused with coordinates of barcode spots that might
fall in to the area of the polygon! To obtain a data.frame of barcode spots that fall in to
the spatial extent of the annotated structure along with their coordinates
use \code{getImageAnnotationBarcodes()}.
}
