% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageAnnotationRange}
\alias{getImageAnnotationRange}
\title{Obtain image annotations range}
\usage{
getImageAnnotationRange(object, id)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}
}
\value{
List of length two. Named with \emph{x} and \emph{y}. Each slot
contains a vector of length two with the minima and maxima in pixel.
}
\description{
Extracts the minimum and maximum x- and y-coordinates
of the image annotation border.
}
