% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotations}
\alias{getImageAnnotations}
\title{Obtain list of \code{ImageAnnotation}-objects}
\usage{
getImageAnnotations(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  add_image = TRUE,
  padding = 20,
  flatten = FALSE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{add_image}{Logical. If TRUE, the area of the histology image that
is occupied by the annotated structure is added to the \code{ImageAnnotation}
object in slot @image.}

\item{padding}{Numeric value. If bigger than 0 creates a padding around the cropped image that never
exceeds the actual limits of the image. Barcode spots are not affected.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts a list of objects of class \code{ImageAnnotaion}.
}
\details{
After the filtering by image annotation ID you can use a combination
of the arguments \code{tags} and \code{test} to filter for the image annotations of
interest. Input for argument \code{tags} specifies the tags of interest. With argument
\code{test} set to \emph{'any'} you make the function include all image annotations
that were tagged with at least one tag of the input for argument \code{tags}.
If \code{test} is set to \emph{'all'} an image annotation must contain all
tags of \code{tags} to be included.
}
