% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{getImageAnnotationScreeningDf}
\alias{getImageAnnotationScreeningDf}
\title{Obtain IAS screending data.frame}
\usage{
getImageAnnotationScreeningDf(
  object,
  id,
  distance = NA_integer_,
  binwidth = NA_integer_,
  n_bins_circle = NA_integer_,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  variables = NULL,
  normalize = TRUE,
  smooth = FALSE,
  smooth_span = 0.2,
  remove_circle_bins = FALSE,
  remove_angle_bins = FALSE,
  rename_angle_bins = FALSE,
  drop = TRUE,
  summarize_by_circles = TRUE,
  summarize_with = "mean",
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{distance}{Numeric value. Specifies the distance from the border of the
image annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.)}

\item{binwidth}{Numeric value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the distance
between the barcode-spots: \code{binwidth = getBarcodeSpotsDistance(object)}.
(See details for more.)}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
 (See details for more.)}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the image annotation.
 (See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{remove_circle_bins, remove_angle_bins, remame_angle_bins}{Logical values.
Given to the corresponding arguments of \code{bin_by_area()}
and \code{bin_by_angle()}.}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}.
Specifies the function with which the bins are summarized.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Bins and annotates barcode-spots in the same way that
\code{imageAnnotationScreening()} does.
}
