% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageAnnotationSize}
\alias{getImageAnnotationSize}
\title{Obtain area of image annotation}
\usage{
getImageAnnotationSize(object, id, unit = NULL, as_numeric = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{unit}{Character value. Specifies the desired unit.}
}
\value{
Single numeric value, the area of the image annotation in the squared
unit of length that is specified in \code{unit}. The unit is attached to the
output as an attribute named \emph{unit}. E.g. if \verb{unit = *mm*} the output value
has the unit \emph{mm^2}.
}
\description{
Computes the area of an image annotation in units of area.
}
\details{
First, the area of each pixel is calculated by using the center
to center distance of the underlying method as ground truth to scale
the current number of pixels in the image.

Second, the number of pixels that fall in the area of the image annotation
is computed with \code{sp::point.in.polygon()} using the area data.frame of
the image annotation as the polygon.

Third, the number of pixels that fall in the area is multiplied with
the area per pixel.
}
\seealso{
\code{getImageAnnotationAreaDf()}, \code{geCCD()}, \code{as_unit()}
}
