% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImageAnnotationTags}
\alias{getImageAnnotationTags}
\title{Deprecated}
\usage{
getImageAnnotationTags(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnTags]{getSpatAnnTags()}}.
}
\keyword{internal}
