% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getImageCenter}
\alias{getImageCenter}
\alias{getImageCenter,spata2-method}
\alias{getImageCenter,HistoImaging-method}
\alias{getImageCenter,HistoImage-method}
\title{Obtain image center}
\usage{
getImageCenter(object, ...)

\S4method{getImageCenter}{spata2}(object)

\S4method{getImageCenter}{HistoImaging}(object, name = NULL)

\S4method{getImageCenter}{HistoImage}(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Numeric vector of length two.
}
\description{
Computes and extracts center of the image frame.
}
