% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getImageDf}
\alias{getImageDf}
\alias{getImageDf,spata2-method}
\alias{getImageDf,HistoImaging-method}
\alias{getImageDf,HistoImage-method}
\alias{getImageDf,SpatialAnnotation-method}
\alias{getImageDf,Image-method}
\title{Obtain image as a data.frame}
\usage{
getImageDf(object, ...)

\S4method{getImageDf}{spata2}(object, img_name = NULL, transform = TRUE, scale_fct = 1, ...)

\S4method{getImageDf}{HistoImaging}(object, img_name = NULL, transform = TRUE, scale_fct = 1)

\S4method{getImageDf}{HistoImage}(object, transform = TRUE, scale_fct = 1)

\S4method{getImageDf}{SpatialAnnotation}(object, rescale_axes = TRUE, scale_fct = 1)

\S4method{getImageDf}{Image}(object, scale_fct = 1)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{rescale_axes}{Logical value. If \code{TRUE}, rescales the pixel positions
(height/width) to the position in the original image.

The image annotation contains a crop of the original image that only shows
the area of the image annotation (plus \code{expand}, see \code{\link[=getSpatialAnnotation]{getSpatialAnnotation()}}).}
}
\value{
Data.frame with three variables.

\itemize{
\item{\emph{width}:}{ Numeric. Width value of the pixel (position on horizontal axis).}
\item{\emph{height}:}{ Numeric. Height value of the pixel (position on vertical axis).}
\item{\emph{color}:}{ Character. HEX-code of the color the pixel carries.}
}
}
\description{
Extracts a data.frame in which each row corresponds
to a pixel in the image. (Faster than \code{getPixelDf()}, though without
any further options.)
}
\seealso{
\code{\link[=getPixelDf]{getPixelDf()}}
}
