% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getImageDims}
\alias{getImageDims}
\alias{getImageRange}
\title{Obtain image dimensions/ranges}
\usage{
getImageDims(object, xrange = NULL, yrange = NULL)

getImageRange(object, xrange = NULL, yrange = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{xrange, yrange}{Vectors of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200 up to x-coordinate 500.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details section of \code{?is_dist} for more information.}
}
\description{
Obtain image dimensions/ranges
}
