% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImageDirDefault}
\alias{getImageDirDefault}
\title{Deprecated}
\usage{
getImageDirDefault(object, ...)
}
\description{
Deprecated in favor of \code{\link[=getImageDir]{getImageDir()}}.
}
\keyword{internal}
