% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImageDirHighres}
\alias{getImageDirHighres}
\title{Deprecated}
\usage{
getImageDirHighres(object, fdb_fn = "warning", check = FALSE, ...)
}
\description{
Deprecated in favor of \code{\link[=getImageDir]{getImageDir()}}.
}
\keyword{internal}
