% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageInfo}
\alias{getImageInfo}
\title{Obatain image information}
\usage{
getImageInfo(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
List that contains information of slots @image_info and @justification
of the \code{HistologyImaging} object.
}
\description{
Extracts a list of information about the currently set
image.
}
