% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageObject}
\alias{getImageObject}
\title{Obtain object of class \code{HistologyImaging}}
\usage{
getImageObject(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Object of class \code{HistologyImaging}
}
\description{
Extracts the S4-object. Do not confuse with \code{getImage()}
}
