% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getImageOrigin}
\alias{getImageOrigin}
\title{Obtain image origin}
\usage{
getImageOrigin(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Either a directory or \emph{Global.Env.} if it was read in from
the global environment.
}
\description{
Extrats the origin of the image that is currently set.
}
