% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getImageSectionsByBarcode}
\alias{getImageSectionsByBarcode}
\title{Obtain image sections by barcode spot}
\usage{
getImageSectionsByBarcode(object, barcodes = NULL, verbose = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{barcodes}{Characte vector or NULL. If character, subsets the barcodes
of interest. If NULL, all barcodes are considered.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A named list. Each slot is named after one barcode. The content is
another list that contains the barcode specific image section as well
as the x- and y-ranges that were used to crop the section.
}
\description{
Cuts out the area of the image that is covered by each barcode.
}
