% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getImageTransformations}
\alias{getImageTransformations}
\alias{getImageTransformations,HistoImaging-method}
\alias{getImageTransformations,HistoImage-method}
\title{Obtain image transformation instructions}
\usage{
getImageTransformations(object, ...)

\S4method{getImageTransformations}{HistoImaging}(object, img_name = NULL, ...)

\S4method{getImageTransformations}{HistoImaging}(object, img_name = NULL, ...)

\S4method{getImageTransformations}{HistoImage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
A list with the following structure:
\itemize{
\item{\emph{angle}:}{ Numeric value that ranges from 0-359. Indicates the angle in degrees
b y which the image needs to be rotated in \strong{clockwise} direction. Defaults to 0.}
\item{\emph{flip}:}{ List of two logical values named \emph{horizontal} and \emph{vertical}. Both default to \code{FALSE}}
\item{\emph{scale}:}{ Numeric value that ranges from 0.01-1. Defaults to 1.}
\item{\emph{translate}:}{ Vector of two numeric values named \emph{horizontal} and \emph{vertical}. Indicate
the number of pixels the image needs to be translated. Positive values shift the image
\strong{downwards} or to the right, respectively. Negative values shift the image \strong{upwards}
or to the left, respectively. Both default to 0.}
}
}
\description{
Extracts a list that contains information regarding required
image transformations to ensure alignment.
}
