% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImgAnnBorderDf}
\alias{getImgAnnBorderDf}
\title{Deprecated}
\usage{
getImgAnnBorderDf(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnOutlineDf]{getSpatAnnOutlineDf()}}.
}
\keyword{internal}
