% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImgAnnCenterBcsp}
\alias{getImgAnnCenterBcsp}
\title{Obtain center barcode-spot}
\usage{
getImgAnnCenterBcsp(object, id)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}
}
\value{
Data.frame as returned by \code{getCoordsDf()} with one row.
}
\description{
Extracts the barcode spot that lies closest
to the center of the image annotation.
}
