% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImgAnnCenters}
\alias{getImgAnnCenters}
\title{Deprecated}
\usage{
getImgAnnCenters(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnCenters]{getSpatAnnCenters()}}.
}
\keyword{internal}
