% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImgAnnOutlineDf}
\alias{getImgAnnOutlineDf}
\title{Deprecated}
\usage{
getImgAnnOutlineDf(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnOutlineDf]{getSpatAnnOutlineDf()}}.
}
\keyword{internal}
