% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImgAnns.R
\name{getImgAnnRange}
\alias{getImgAnnRange}
\alias{getImgAnnRange,spata2-method}
\alias{getImgAnnRange,HistoImaging-method}
\title{Obtain image annotations range}
\usage{
getImgAnnRange(object, ...)

\S4method{getImgAnnRange}{spata2}(object, id, scale_fct = 1)

\S4method{getImgAnnRange}{HistoImaging}(object, id, scale_fct = 1)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value. The ID of the image annotation of interest.}
}
\value{
List of length two. Named with \emph{x} and \emph{y}. Each slot
contains a vector of length two with the minima and maxima in pixel.
}
\description{
Extracts the minimum and maximum x- and y-coordinates
of the image annotation border.
}
