% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImgAnnTags}
\alias{getImgAnnTags}
\title{Obtain image annotations tags}
\usage{
getImgAnnTags(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts all unique tags with which image annotations
have been tagged.
}
