% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getInitiationInfo}
\alias{getInitiationInfo}
\alias{getInitiationInput}
\title{Obtain information about object initiation}
\usage{
getInitiationInfo(object)

getInitiationInput(object, verbose = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A list. See description.
}
\description{
Information about the object's initiation is stored in
a list of three slots:

\itemize{
 \item{\emph{init_fn}: Contains the name of the initation function as a character value.}
 \item{\emph{input}: Contains a list of which every slot refers to the input of one argument with which the
 initiation function has been called.}
 \item{\emph{time}: Contains the time at which the object was initiated.}
 }

 \code{getInitiationInput()} returns only slot \emph{input}.
}
\details{
\code{initiateSpataObject_CountMtr()} and \code{initiateSpataObject_ExprMtr()} each require
a matrix and a coordinate data.frame as input. These are not included in the output
of this function but can be obtained via \code{getCoordsDf()} and \code{getCountMtr()} or \code{getExpressionMtr()}.
}
