% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getLogfileDf}
\alias{getLogfileDf}
\title{Obtain object logfile}
\usage{
getLogfileDf(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
A data.frame with variables \emph{fn_name}, \emph{date_time}, \emph{args_input}, \emph{pkg_version}.
See the vignette on the \link[=concept_logfile]{logfile} for more information.
}
\description{
Extracts the logfile which stores information on the progress
made with the provided \code{\link{SPATA2}} object.
}
