% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getMatrixNames}
\alias{getMatrixNames}
\title{Obtain matrix names}
\usage{
getMatrixNames(object, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A character vector containing the names of matrices.
}
\description{
Retrieves the names of matrices present in the specified assay of the provided object.
}
\seealso{
\code{\link[=getMatrix]{getMatrix()}}
}
