% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getMetaDf}
\alias{getMetaDf}
\title{Obtain meta data.frame for observations}
\usage{
getMetaDf(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
The meta data frame.
}
\description{
Retrieves the meta data frame from the provided object which
contains \link[=concept_variables]{meta information} about the objects's
\link[=concept_observations]{observations}.
}
