% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{getMetaVarDf}
\alias{getMetaVarDf}
\title{Obtain molecular meta data.frame}
\usage{
getMetaVarDf(object, assay_name = activeAssay(object), verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A data frame containing metadata variables for the specified assay.
}
\description{
Retrieves the metadata variable data frame for a specified assay
in the given object. If the metadata variable data frame is empty, it creates
a new one based on the molecule names.

Do not confuse with \code{\link[=getMetaDf]{getMetaDf()}} which contains meta variables for
the \link[=concept_observations]{observations}.
}
