% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getMethod}
\alias{getMethod}
\title{Obtain unit of method}
\usage{
getMethod(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Character value.
}
\description{
Extracts the European unit of length in which the size of
the fiducial frame of the underlying spatial method is specified.
}
