% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getMethodSpecifics}
\alias{getMethodSpecifics}
\title{Obtain method specifics}
\usage{
getMethodSpecifics(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
A named list.
}
\description{
Extracts a list of method specifics from the \code{SpatialMethod} of
the \code{SPATA2} object.
}
