% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getModality}
\alias{getModality}
\alias{getModality,SPATA2-method}
\alias{getModality,MolecularAssay-method}
\title{Obatin molecular modality}
\usage{
getModality(object, ...)

\S4method{getModality}{SPATA2}(object, assay_name = activeAssay(object), ...)

\S4method{getModality}{MolecularAssay}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Character value.
}
\description{
Extracts the molecular modality of chosen \link[=MolecularAssay]{molecular assay}.
}
\examples{

library(SPATA2)

object <- loadExampleObject("UKF275T")

getModality(object)

}
