% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getModelEvaluationDf}
\alias{getModelEvaluationDf}
\alias{getModelEvaluationDf,SpatialAnnotationScreening-method}
\alias{getModelEvaluationDf,SpatialTrajectoryScreening-method}
\title{Obtain model evaluation}
\usage{
getModelEvaluationDf(object, ...)

\S4method{getModelEvaluationDf}{SpatialAnnotationScreening}(object, smrd = TRUE)

\S4method{getModelEvaluationDf}{SpatialTrajectoryScreening}(object, ...)
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
Data.frame.
}
\description{
Extracts the data.frame that contains the variable-model-fit
evaluation containing.
}
\keyword{internal}
