% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{getMoleculeCoordinates}
\alias{getMoleculeCoordinates}
\title{Obtain molecule coordinates}
\usage{
getMoleculeCoordinates(
  object,
  molecules = NULL,
  assay_name = activeAssay(object)
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{molecules}{Character or \code{NULL}. If character, specifies the molecules
of interest and the output data.frame is filtered accordingly.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Data.frame with variables \emph{molecule}, \emph{x_orig}, \emph{x}, \emph{y_orig}, \emph{y}.
}
\description{
Extracts the molecule coordinates of a specfific assay.
}
