% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R, R/getN-Z.R
\name{getGenes}
\alias{getGenes}
\alias{getGenesInteractive}
\alias{getMetabolites}
\alias{getMolecules}
\alias{getProteins}
\title{Obtain molecule names}
\usage{
getGenes(object, signatures = NULL, simplify = TRUE, ...)

getGenesInteractive(object)

getMetabolites(object, signatures = NULL, simplify = TRUE, ...)

getMolecules(
  object,
  signatures = NULL,
  simplify = TRUE,
  assay_name = activeAssay(object)
)

getProteins(object, signatures = NULL, simplify = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{signatures}{Character vector or \code{NULL}. If character, specifies the name of
signatures to filter the molecules for.}

\item{simplify}{Only relevant if \code{signatures} is not \code{NULL}. If \code{TRUE}, all molecule
names are merged in to one character vector of unique molecule names. If \code{FALSE},
a named list of character vectors is returned (names correspond to signatures).}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest (and thus the
active \link[=concept_molecular_modalities]{molecular modality}. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A character vector or a named list of character vectors.
}
\description{
Retrieves the list of molecules present in the given object,
optionally filtered by a specific signature.
}
\details{
This function retrieves the list of molecules from the provided object.
If the \code{signatures} argument is provided, it filters the molecules based on the
specified signatures. If 'signature' is \code{NULL}, it returns all molecules from the
active assay in the object.
}
\note{
Molecule names are picked from the currently active matrix of their assay.
If the processed matrix does not contain some molecules the raw count matrix
contains the do not appear.
}
\examples{
# Get all molecules from the object
all_molecules <- getMolecules(object)

# Get molecules associated with a specific signature
signature_molecules <- getMolecules(object, signature = "signature_name")

}
\seealso{
\code{\link[=getMatrix]{getMatrix()}}, \code{\link[=getSignatures]{getSignatures()}}
}
