% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getMoleculeTypeList}
\alias{getMoleculeTypeList}
\title{Obtain a list of molecules}
\usage{
getMoleculeTypeList(object, molecules = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{molecules}{A character vector specifying the subset of molecules to include in the output.
By default, all molecules are included.}
}
\value{
A list containing the names of molecules categorized by type.
}
\description{
Retrieves a list of molecules sorted by \link[=concept_molecular_modality]{molecular modality}.
}
\details{
This function categorizes molecules into different types based on the provided object.
If the 'molecules' argument is provided as a character vector, the function returns only the
specified molecules categorized by type. Otherwise, it returns all molecules categorized by type.
}
\examples{
# Get molecular type list for all molecules in the object
mol_types <- getMolTypeList(object)

# Get molecular type list for specific molecules
mol_types <- getMolTypeList(object, molecules = c("mol1", "mol2"))

}
\seealso{
\code{\link[=getAssayModalities]{getAssayModalities()}}
}
