% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getPixelScaleFactor}
\alias{getPixelScaleFactor}
\alias{getPixelScaleFactor,spata2-method}
\alias{getPixelScaleFactor,HistoImaging-method}
\title{Obtain scale factor for pixel to SI conversion}
\usage{
getPixelScaleFactor(object, ...)

\S4method{getPixelScaleFactor}{spata2}(
  object,
  unit,
  switch = FALSE,
  force = FALSE,
  add_attr = TRUE,
  verbose = NULL,
  ...
)

\S4method{getPixelScaleFactor}{HistoImaging}(
  object,
  unit = NULL,
  img_name = NULL,
  switch = FALSE,
  force = FALSE,
  add_attr = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{unit}{The desired unit. Defaults to the unit
in which the original size of the image of the spatial method is
provided. Obtain valid input options with \code{validUnitsOfLengthSI()}.}

\item{switch}{Logical value. If \code{TRUE}, the unit of the output is switched.
See details for more.}

\item{force}{Logical value. If \code{TRUE}, the scale factor is computed
regardless of what the function finds in the respective slot.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A single numeric value with the unit defined in attribute \emph{unit}.
}
\description{
Extracts or computes the side length of pixel sides depending
on the current resolution of the image.
}
\details{
If \code{switch} is \code{FALSE}, the default, the output is to be interpreted as
unit/pixel. E.g. an output of \emph{15 'um/px'} means that under the current resolution
of the image height and width of one pixel corresponds to \emph{15 um} in height and
width in the original tissue.

If \code{switch} is \code{TRUE}, the output is to be interpreted as pixel/unit.  E.g.
an output value of \emph{0.07 'px/um'} means that under the current image resolution
one micrometer corresponds to 0.07 pixel in the image.
}
\seealso{
\code{\link[=computePixelScaleFactor]{computePixelScaleFactor()}}, \code{\link[=setScaleFactor]{setScaleFactor()}}
}
