% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getProcessedMatrix}
\alias{getProcessedMatrix}
\alias{getProcessedMatrix,SPATA2-method}
\alias{getProcessedMatrix,MolecularAssay-method}
\title{Obtain processed data matrix}
\usage{
getProcessedMatrix(object, ...)

\S4method{getProcessedMatrix}{SPATA2}(object, mtr_name, assay_name = activeAssay(object), ...)

\S4method{getProcessedMatrix}{MolecularAssay}(object, mtr_name, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{mtr_name}{Character value. The name of the processed matrix of interest.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest (and thus the
active \link[=concept_molecular_modalities]{molecular modality}. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A numeric matrix with rownames corresponding to molecule names and column
names corresponding to barcodes.
}
\description{
Extracts a processed data matrix.
}
\note{
The argument \code{mtr_name} must be specified in contrast to \code{getMatrix()}.
}
\seealso{
\code{\link[=getCountMatrix]{getCountMatrix()}}, \code{\link[=getMatrix]{getMatrix()}}, [\verb{getProcessedMatrixNames()'], [}getMatrixNames()`]
}
