% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getProcessedMatrixNames}
\alias{getProcessedMatrixNames}
\title{Obtain names of processed matrices}
\usage{
getProcessedMatrixNames(object, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Character vector of names. If no content is present the character
vector is of length 0.
}
\description{
Extract names of processed matrices.
}
