% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{getProjectionDf}
\alias{getProjectionDf}
\title{Obtain trajectory projection}
\usage{
getProjectionDf(object, id, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\description{
Extracts the projection data.frame of a trajectory. If \code{variables}
is specified
}
