% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getPrResults}
\alias{getPrResults}
\alias{getPatternNames}
\title{Obtain pattern recognition results}
\usage{
getPrResults(object, method_pr = "hspa", of_sample = NA)

getPatternNames(object, method_pr = "hotspot", of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
The list containing all information the respective pattern
recognition algorithm returns.

\itemize{
 \item{\code{getPrResults()}: List containing all information the respective
 method returns}
 \item{\code{getPrSuggestion()}: List containing the actual pattern suggestions.}
 \item{\code{getPatternNames()}: Character vector of pattern names.}}
}
\description{
Obtain pattern recognition results
}
