% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-Z.R
\name{getResultsDf}
\alias{getResultsDf}
\alias{getResultsDf,ImageAnnotationScreening-method}
\alias{getResultsDf,SpatialTrajectoryScreening-method}
\title{Obtain results stored in data.frames}
\usage{
getResultsDf(object, ...)

\S4method{getResultsDf}{ImageAnnotationScreening}(
  object,
  eval = "ias_score",
  pval = "p_value_mean_adjusted",
  arrange_by = eval,
  threshold_eval = 0,
  threshold_pval = 1,
  model_subset = NULL,
  model_remove = NULL,
  best_only = FALSE
)

\S4method{getResultsDf}{SpatialTrajectoryScreening}(
  object,
  eval = "sts_score",
  pval = "p_value",
  arrange_by = eval,
  threshold_eval = 0,
  threshold_pval = 1,
  model_subset = NULL,
  model_remove = NULL,
  best_only = FALSE
)
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
Data.frame.
}
\description{
Extracts content of slot @results of screening S4 objects. For
a more detailed explanation of what the slot contains check the description
of the respective S4 class. E.g. with \code{?SpatialTrajectoryScreening}.
}
\details{
Without any argument specification the function \code{getResultsDf()} returns
the complete data.frame. The arguments can be used to filter the results. Filtering
works as follows:

\enumerate{
 \item{}{ Model-fits are filtered according to the input of \code{model_subset} and \code{model_remove}. }
 \item{}{ Model-fits are filtered according to the \code{threshold_} arguments. }
 \item{}{ If \code{best_only} is set to TRUE, model-fits are filtered such that the best model-fit
  (among the remaining models from 1.) for every gene remains. E.g. if gene GFAP fits to model
 \emph{linear_descending} with a score of 0.9 and to \emph{immediate_descending} with a score of
  0.75 the model-fit \emph{GFAP-linear_descending} remains in the output.
  }
 }

The output is arranged by the evaluation.
}
