% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getResultsVec}
\alias{getResultsVec}
\alias{getResultsVec,ImageAnnotationScreening-method}
\alias{getResultsVec,SpatialTrajectoryScreening-method}
\title{Obtain screening results stored in vectors}
\usage{
getResultsVec(object, ...)

\S4method{getResultsVec}{ImageAnnotationScreening}(
  object,
  eval = "ias_score",
  pval = "p_value_mean_adjusted",
  arrange_by = eval,
  threshold_eval = 0.5,
  threshold_pval = 0.05,
  model_subset = NULL,
  model_remove = NULL
)

\S4method{getResultsVec}{SpatialTrajectoryScreening}(
  object,
  eval = "sts_score",
  pval = "p_value",
  arrange_by = eval,
  threshold_eval = 0.5,
  threshold_pval = 0.05,
  model_subset = NULL,
  model_remove = NULL
)
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
Named character vector. Values are the variable/gene names. Names
correspond to the model that fitted best.
}
\description{
Extracts results in form of character vectors.
}
\details{
Extraction works similar to \code{getResultsDf()}. Argument \code{best_only}, however,
is always set to TRUE.
}
