% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getSampleAreaSize}
\alias{getSampleAreaSize}
\title{Deprecated}
\usage{
getSampleAreaSize(...)
}
\description{
Deprecated in favor of \code{\link[=getTissueArea]{getTissueArea()}}.
}
\keyword{internal}
