% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSampleName}
\alias{getSampleName}
\title{Obtain name of SPATA2 object}
\usage{
getSampleName(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
A character value.
}
\description{
Extracts the name/ID of the \code{SPATA2} object
in form of a single character value.
}
