% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Obtain sample names}
\usage{
getSampleNames(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A character vector.
}
\description{
Obtain sample names
}
