% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getSasBinAreas}
\alias{getSasBinAreas}
\title{Calculate IAS bin area}
\usage{
getSasBinAreas(
  object,
  id,
  distance = NA_integer_,
  n_bins_dist = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  n_bins_angle = 1,
  area_unit = NULL,
  use_outline = TRUE,
  remove_circle_bins = "Outside",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{n_bins_dist}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}. Not required if \code{distance} and \code{binwidth} is
specified. (See details for more.)}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{use_outline}{Logical value. If \code{TRUE}, uses the outline variable
set with \code{setOutlineVarName()} or if none is set DBSCAN to identify the
outline of the tissue section or sections in case of multiple tissue sections
on one Visium slide to only compute the area of circle bins that covers the
tissue section.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Data.frame in which each observation corresponds to a circular bin.
}
\description{
Computes the area circular bins of the IAS algorithm
cover.
}
\details{
Approximates the area each circular bin covers
by assigning each pixel to the circular bin it falls into.
Afterwards the number of pixels per bin is multiplied
with the area scale factor as is obtained by \code{getPixelScaleFactor(object, unit = unit)}
where unit is the squared unit of input for argument \code{binwidth}. E.g.
if \code{binwidth} = \emph{'0.1mm'} then \code{unit} = \emph{mm2}.
}
\note{
If multiple tissue sections are located on the Visium slide use \code{createSpatialSegmentation()}
to encircle each section and set the variable name in which you saved the
encircling via \code{setOutlineVarName()}. Else the areas of the circle bins might
include space that is not covered by tissue. This might distort computation
results.
}
