% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSasDf}
\alias{getSasDf}
\title{Obtain spatial annotation screening data.frame}
\usage{
getSasDf(
  object,
  ids,
  distance = "dte",
  resolution = recSgsRes(object),
  core = FALSE,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  variables = NULL,
  unit = getDefaultUnit(object),
  ro = c(0, 1),
  format = "wide",
  bcs_exclude = NULL,
  outlier_rm = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector. Specifies the IDs of the spatial annotations of interest.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance measure}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. Defaults to a distance
that covers the whole tissue section the spatial annotation is located
on using \code{\link[=distToEdge]{distToEdge()}}. (This distance must not be exceeded.)}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of its closest spatial annotation.}

\item{variables}{Character vector. The numeric variables to be included in
the screening process. Makre sure that the correct matrix is active in the
respective assays.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}

\item{ro}{The numeric range to which the output gradients is scaled. Defaults
to c(0,1).}

\item{bcs_exclude}{Character value containing the barcodes of observations to be excluded
from the analysis.}

\item{outlier_rm}{Deprecated.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame of inferred gradients of numeric
variables as a function of distance to spatial annotations.
}
