% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getSegmentationNames}
\alias{getSegmentationNames}
\title{Deprecated}
\usage{
getSegmentationNames(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatSegmVarNames]{getSpatSegmVarNames()}}.
}
\keyword{internal}
