% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getSegmentNames}
\alias{getSegmentNames}
\title{Obtain segment names}
\usage{
getSegmentNames(object, simplify = TRUE, of_sample = NA, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of segments which were drawn for the
specific sample.
}
\description{
Obtain segment names
}
