% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSgsResultsDf}
\alias{getSgsResultsDf}
\alias{getSgsResultsDf,SpatialAnnotationScreening-method}
\alias{getSgsResultsDf,SpatialTrajectoryScreening-method}
\title{Obtain spatial gradient screening results}
\usage{
getSgsResultsDf(object, ...)

\S4method{getSgsResultsDf}{SpatialAnnotationScreening}(
  object,
  eval = "mae",
  pval = "fdr",
  arrange_by = eval,
  threshold_eval = 1,
  threshold_pval = 1,
  model_subset = NULL,
  model_remove = NULL,
  best_only = FALSE
)

\S4method{getSgsResultsDf}{SpatialTrajectoryScreening}(
  object,
  eval = "mae",
  pval = "fdr",
  arrange_by = eval,
  threshold_eval = 1,
  threshold_pval = 1,
  model_subset = NULL,
  model_remove = NULL,
  best_only = FALSE
)
}
\arguments{
\item{object}{Any object for which a method has been defined.}

\item{eval}{Character value. The evaluation variable to use. Either \emph{'mae'} (Mean
Absolute Error) or \emph{'rmse'} (Root Mean Squared Error).}

\item{pval}{Character value. The p-value variable. Defaults to \emph{'fdr'} (False Discovery
Rate).}

\item{threshold_pval, threshold_eval}{Numeric values. The threshold with which
the results are filtered. Default is 1. Since p-values and model fit evaluation
scores range from 0-1 (with 1 being worst), the default includes everything.}

\item{best_only}{Logical value. If \code{TRUE}, only the best gradient-model fit according
to the chosen evaluation metric (\code{eval}) for each screened variable is kept.}
}
\value{
A data.frame with results of the spatial gradient screening conducted.
Column names are:

\itemize{
\item{variables}{ The name of the variable to which the row corresponds.}
\item{models}{ The name of the model which fits best to the inferred gradient.}
\item{mae}{ The mean absolute error of the gradient-model fit.}
\item{rmse}{ The root mean squared error of the gradient-model fit.}
\item{p_value}{ The p-value regarding the hypothesis whether such a gradient
can be obtained under random circumstances (= the spatial reference feature
has no impact on the spatial expression pattern.)
}
\item{fdr}{ The adjusted p-value using false discovery rate.}
}

Data.frame.
}
\description{
Extracts content of slot @results of screening S4 objects. For
a more detailed explanation of what the slot contains see the documentation
of \code{\link{SpatialGradientScreening}}.
}
\details{
Without any argument specification the function \code{getSgsResultsDf()} returns
the complete data.frame. The arguments can be used to filter the results. Filtering
works as follows:

\enumerate{
\item{}{ Model-fits are filtered according to the input of \code{model_subset} and \code{model_remove}. }
\item{}{ Model-fits are filtered according to the \code{threshold_} arguments. }
\item{}{ If \code{best_only} is set to TRUE, model-fits are filtered such that the best model-fit
(among the remaining models from 1.) for every gene remains. E.g. if gene GFAP fits to model
\emph{descending_linear} with a score of 0.2 and to \emph{descending_gradual} with an MAE score of
0.15 the model-fit \emph{GFAP-descending_gradual} remains in the output.
}
}

The output is arranged by the evaluation score.
}
