% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSgsResultsVec}
\alias{getSgsResultsVec}
\alias{getSgsResultsVec,SpatialGradientScreening-method}
\title{Obtain spatial gradient screening results}
\usage{
getSgsResultsVec(object, ...)

\S4method{getSgsResultsVec}{SpatialGradientScreening}(
  object,
  eval = "mae",
  pval = "fdr",
  arrange_by = eval,
  threshold_eval = 0.5,
  threshold_pval = 0.05,
  model_subset = NULL,
  model_remove = NULL
)
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
Named character vector. Values are the variable/gene names. Names
correspond to the model that fitted best.
}
\description{
Extracts results in form of character vectors.
}
\details{
Extraction works similar to \code{getSgsResultsDf()}. Argument \code{best_only}, however,
is always set to TRUE.
}
