% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R, R/getN-Z.R
\name{getGeneSet}
\alias{getGeneSet}
\alias{getMetaboliteSet}
\alias{getProteinSet}
\alias{getSignature}
\title{Obtain molecular signature}
\usage{
getGeneSet(object, gene_set, ...)

getMetaboliteSet(object, metabolite_set, ...)

getProteinSet(object, protein_set, ...)

getSignature(object, signature, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{signature}{Character value. The signature of interest.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Character vector.
}
\description{
Extracts a character vector of molecule names making up a molecular signature.
}
\details{
These functions retrieve molecule names of single signatures.

\itemize{
\item{\code{getSignature()}}{: A character vector of molecule names from the signature specified.}
\item{\code{getGeneSet()}}{:  A character vector of gene names from the gene set specified.}
\item{\code{getMetaboliteSet()}}{: A character vector metabolite names from the metabolite set specified.}
\item{\code{getProteinSet()}}{: A character vector of protein names from the protein set specified.}
}
}
