% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R, R/getN-Z.R
\name{getGeneSetList}
\alias{getGeneSetList}
\alias{getMetaboliteSetList}
\alias{getProteinSetList}
\alias{getSignatureList}
\title{Obtain molecular signatures}
\usage{
getGeneSetList(object, class = NULL)

getMetaboliteSetList(object, class = NULL)

getProteinSetList(object, class = NULL)

getSignatureList(object, class = NULL, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{class}{Character vector of signature classes with which to subset the output.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A named list of character vectors.
}
\description{
Retrieves the list of \link[=concept_molecular_signatures]{molecular signatures}
stored in a \link[=MolecularAssay]{molecular assay}. While \code{getSignatureList()} allows
to extract signatures of all assays, \code{getGeneSetList()}, \code{getMetaboliteSetList()} and \code{getProteinSetList()}
are quick wrappers. But they require assays of specific names/\link[=concept_molecular_modalities]{molecular modalities}.
}
\details{
These functions retrieve the signatures from the provided object.

\itemize{
\item{\code{getSignatureList()}}{: The list of signatures from the assay specified in \code{assay_name}.}
\item{\code{getGeneSetList()}}{:  The list of signatures from the assay with @modality = 'gene' (\code{assay_name = 'gene'}).}
\item{\code{getMetaboliteSetList()}}{: The list of signatures from the assay with @modality = 'metabolite' (\code{assay_name = 'metabolite'}).}
\item{\code{getProteinSetLit()}}{: The list of signatures the assay with @modality = 'protein' (\code{assay_name = 'protein'}).}
}
}
\examples{

library(SPATA2)

object <- example_data$object_UKF269T_diet

## how the different functions work
getAssayNames(object)

# opt 1
activeAssay(object)
gene_sets <- getSignatureList(object)

head(gene_sets)

# opt 2 (equal to opt 1, cause active and only assay is 'gene')
gene_sets <- getSignatureList(object, assay_name = "gene")

head(gene_sets)

# opt 3
gene_sets <- getGeneSetList(object)

head(gene_sets)

# opt 4 - fails cause no 'protein' assay
protein_sets <- getProteinSetList(object)

## using class argument

hm_gene_sets <- getGeneSetList(object, class = "HM")

head(hm_gene_sets)
tail(hm_gene_sets)

two_kinds_of_gene_sets <- getGeneSetList(object, class = c("HM", "RCTM"))

head(two_kinds_of_gene_sets)
tail(two_kinds_of_gene_sets)

}
\seealso{
Documentation of slot @signatures in the \code{\link{MolecularAssay}}-class.
To extract character vectors of molecule names \code{\link[=getMolecules]{getMolecules()}} or \code{\link[=getGenes]{getGenes()}}, ...
To extract character vectors of signature names \code{\link[=getSignatureNames]{getSignatureNames()}}, \code{\link[=getGeneSets]{getGeneSets()}}, ...
To add signatures \code{\link[=addSignature]{addSignature()}} or \code{\link[=addGeneSet]{addGeneSet()}}, ...
}
