% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R, R/getN-Z.R
\name{getGeneSetOverview}
\alias{getGeneSetOverview}
\alias{getMetaboliteOverview}
\alias{getProteinSetOverview}
\alias{getSignatureOverview}
\title{Overview about the current signature collection}
\usage{
getGeneSetOverview(object, ...)

getMetaboliteOverview(object, ...)

getProteinSetOverview(object, ...)

getSignatureOverview(object, ..., assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Additional selection helpers from the tidyselect package that match names according to a given pattern.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A data.frame with two variables \emph{Class} and \emph{Available Signatures}
indicating the number of different signatures the classes contain.
}
\description{
Counts the number of signatures by class - after subsetting if desired.
}
