% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSignatures}
\alias{getSignatures}
\title{Obtain molecular signatures}
\usage{
getSignatures(object, assay_name = activeAssay(object), class = NULL)
}
\arguments{
\item{object}{An object containing molecular data.}

\item{assay_name}{The name of the assay containing the molecular data (default: active assay in the object).}
}
\value{
A list of character vectors.
}
\description{
Retrieves the signatures present in the \code{\link{SPATA2}} object.
}
\details{
This function retrieves the signatures from the provided object. It returns a list of signatures associated with the specified assay in the object.
}
\examples{
# Get signatures from the object
signatures <- getSignatures(object)

}
\seealso{
Documentation of slot @signatures in the \code{\link{MolecularAssay}}-class.
}
