% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSignatureTypeList}
\alias{getSignatureTypeList}
\title{Obtain a list of signatures}
\usage{
getSignatureTypeList(object, signatures = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{signatures}{A character vector specifying the subset of signatures to include in the output (default: NULL).}
}
\value{
A list containing the names of signatures categorized by assay type.
}
\description{
Retrieves a list of signatures sorted by molecular type as
present in the given object.
}
\details{
This function categorizes signatures into different types based on the provided object.
If the 'signatures' argument is provided as a character vector, the function returns only the specified
signatures categorized by assay type. Otherwise, it returns all signatures categorized by type.
}
\examples{
# Get signature type list for all signatures in the object
signature_types <- getSignatureTypeList(object)

# Get signature type list for specific signatures
signature_types <- getSignatureTypeList(object, signatures = c("sig1", "sig2"))

}
\seealso{
Documentation of slot @signatures in the \code{\link{MolecularAssay}}-class.
}
