% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getSmrdResultsDf}
\alias{getSmrdResultsDf}
\title{Deprecated}
\usage{
getSmrdResultsDf(ias, ...)
}
\description{
Deprecated in favor of \code{\link[=getSgsResults]{getSgsResults()}}.
}
\keyword{internal}
