% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpataDir}
\alias{getSpataDir}
\title{Obtain SPATA2 object directory}
\usage{
getSpataDir(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Character value or an error if no directory is set.
}
\description{
Extracts the file directory under which the \code{SPATA2} object
is saved by default with \code{saveSpataObject()}.
}
\seealso{
\code{\link[=setSpataDir]{setSpataDir()}}
}
