% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnCenter}
\alias{getSpatAnnCenter}
\alias{getSpatAnnCenter,SPATA2-method}
\alias{getSpatAnnCenter,HistoImaging-method}
\alias{getSpatAnnCenter,SpatialAnnotation-method}
\alias{getSpatAnnCenters}
\alias{getSpatAnnCenters,SPATA2-method}
\alias{getSpatAnnCenters,HistoImaging-method}
\alias{getSpatAnnCenters,SpatialAnnotation-method}
\title{Obtain center of an spatial annotation}
\usage{
getSpatAnnCenter(object, ...)

\S4method{getSpatAnnCenter}{SPATA2}(object, id)

\S4method{getSpatAnnCenter}{HistoImaging}(object, id)

\S4method{getSpatAnnCenter}{SpatialAnnotation}(object)

getSpatAnnCenters(object, ...)

\S4method{getSpatAnnCenters}{SPATA2}(object, id, outer = TRUE, inner = TRUE)

\S4method{getSpatAnnCenters}{HistoImaging}(object, id, outer = TRUE, inner = TRUE)

\S4method{getSpatAnnCenters}{SpatialAnnotation}(object, outer = TRUE, inner = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{outer}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the outer border of the image annotation
is not included in the output.}

\item{inner}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the inner borders of the image annotation
are not included in the output.}
}
\value{
Numeric vector of length two or a list of these. Values are named \emph{x} and \emph{y}.
}
\description{
\code{getSpatAnnCenter()} computes the
x- and y- coordinates of the center of the outer border, returns
a numeric vector of length two. \code{getSpatAnnCenters()} computes the center of the outer
and every inner border and returns a list of numeric vectors of length two.
}
