% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnCenterBcsp}
\alias{getSpatAnnCenterBcsp}
\title{Obtain center data point}
\usage{
getSpatAnnCenterBcsp(object, id)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}
}
\value{
Data.frame as returned by \code{getCoordsDf()} with one row.
}
\description{
Extracts the barcode spot (data point) that lies closest
to the center of the spatial annotation.
}
