% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnIds}
\alias{getSpatAnnIds}
\alias{getSpatAnnIds,ANY-method}
\alias{getSpatAnnIds,HistoImaging-method}
\title{Obtain IDs of spatial annotations}
\usage{
getSpatAnnIds(object, ...)

\S4method{getSpatAnnIds}{ANY}(object, ids = NULL, tags = NULL, test = "any", class = NULL)

\S4method{getSpatAnnIds}{HistoImaging}(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  class = NULL,
  error = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of spatial annotations} for more information.}

\item{test}{Character value. One of \emph{any}. \emph{all}, \emph{identical}, \emph{not_identical} and
\emph{none}. Specifies how input for \code{tags} is used to select image annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{class}{Character vector or \code{NULL}. If character, defines the subtypes
of spatial annotations to consider. Must be a combination of \emph{c('Group', 'Image'
'Numeric')}.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Character vector. If no spatial annotations are returned the character
vector is of length 0. If this is because no spatial annotations have been
stored yet, the functions remains silent. If this is due to the selection
options, the function throws a warning.
}
\description{
Extracts spatial annotation IDs as a character vector.
}
\section{Selection of spatial annotations}{


Selection of spatial annotations via the arguments \code{ids}, \code{class}, \code{tags} and
\code{test} works in three steps:

First, if \code{ids} is a character it prefilters the annotations by ID and only
the specified ones are submitted to the next steps. If it is \code{NULL}, all
annotations are submitted to the next steps.

Secondd, if \code{class} is a character it filters the annotations remaining
after the first step by their class. If \code{NULL}, the step is skipped.

Third, if \code{tags} is a character it is used in combination with \code{test} to select
from the spatial annotations that remain after the second step based on the meta data
they are tagged with. There are multiple options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags. Can contain tags that are not specified.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags. Can not be tagged with anything else.
\item Argument \code{test} set to \emph{not_identical} or \emph{4}: To be included, an image
annotation must \strong{not} be tagged with the combination of input tags.
\item Argument \code{test} set to \emph{'none'} or \emph{5}: To be included, an image annotation
must \strong{not} contain any of the input tags.
}

If \code{tags} is \code{NULL}, the step is skipped. Therefore, if \code{ids}, \code{class} and \code{tags}
are all NULL, which is the default, all annotations are selected as all subsetting
steps are skipped. Eventually, the remaining spatial annotations are submitted to
whatever the respective function does.
}

\seealso{
S4-classes \code{\link{SpatialAnnotation}}, \code{\link{GroupAnnotation}}, \code{\link{ImageAnnotation}},
\code{\link{NumericAnnotation}}
}
