% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnOutlineDf}
\alias{getSpatAnnOutlineDf}
\alias{getSpatAnnOutlineDf,SPATA2-method}
\alias{getSpatAnnOutlineDf,HistoImaging-method}
\alias{getSpatAnnOutlineDf,SpatialAnnotation-method}
\title{Obtain spatial annotation border data.frame}
\usage{
getSpatAnnOutlineDf(object, ...)

\S4method{getSpatAnnOutlineDf}{SPATA2}(
  object,
  ids = NULL,
  class = NULL,
  tags = NULL,
  test = "any",
  outer = TRUE,
  inner = TRUE,
  incl_edge = FALSE,
  add_tags = FALSE,
  sep = " & ",
  last = " & "
)

\S4method{getSpatAnnOutlineDf}{HistoImaging}(
  object,
  ids = NULL,
  class = NULL,
  tags = NULL,
  test = "any",
  outer = TRUE,
  inner = TRUE,
  incl_edge = FALSE,
  add_tags = FALSE,
  sep = " & ",
  last = " & "
)

\S4method{getSpatAnnOutlineDf}{SpatialAnnotation}(
  object,
  add_tags = TRUE,
  sep = " & ",
  last = " & ",
  expand_outline = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of spatial annotations} for more information.}

\item{test}{Character value. One of \emph{any}. \emph{all}, \emph{identical}, \emph{not_identical} and
\emph{none}. Specifies how input for \code{tags} is used to select image annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{outer}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the outer border of the image annotation
is not included in the output.}

\item{inner}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the inner borders of the image annotation
are not included in the output.}

\item{incl_edge}{Logical value. If \code{TRUE}, the function includes the tissue edge
(see \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}) in the visualization and removes the parts of
the outline that transgress the tissues edge.}

\item{expand_outline}{Distance measure by which the outline of the area is expanded.}
}
\value{
A data.frame that contains variables \emph{id}, \emph{border},
and the numeric variables \emph{x}, \emph{y} and \emph{tags}.
}
\description{
Extracts the coordinates of the vertices of the polygon that represents
the borders of the spatial annotation.
}
\details{
The variables \emph{x} and \emph{y} give the position of the vertices of the polygon
that was drawn to used the area via \code{\link[=createGroupAnnotations]{createGroupAnnotations()}},
\code{\link[=createImageAnnotations]{createImageAnnotations()}} or \code{\link[=createNumericAnnotations]{createNumericAnnotations()}}. These vertices
correspond to the border of the annotation.
}
\section{Selection of spatial annotations}{


Selection of spatial annotations via the arguments \code{ids}, \code{class}, \code{tags} and
\code{test} works in three steps:

First, if \code{ids} is a character it prefilters the annotations by ID and only
the specified ones are submitted to the next steps. If it is \code{NULL}, all
annotations are submitted to the next steps.

Secondd, if \code{class} is a character it filters the annotations remaining
after the first step by their class. If \code{NULL}, the step is skipped.

Third, if \code{tags} is a character it is used in combination with \code{test} to select
from the spatial annotations that remain after the second step based on the meta data
they are tagged with. There are multiple options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags. Can contain tags that are not specified.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags. Can not be tagged with anything else.
\item Argument \code{test} set to \emph{not_identical} or \emph{4}: To be included, an image
annotation must \strong{not} be tagged with the combination of input tags.
\item Argument \code{test} set to \emph{'none'} or \emph{5}: To be included, an image annotation
must \strong{not} contain any of the input tags.
}

If \code{tags} is \code{NULL}, the step is skipped. Therefore, if \code{ids}, \code{class} and \code{tags}
are all NULL, which is the default, all annotations are selected as all subsetting
steps are skipped. Eventually, the remaining spatial annotations are submitted to
whatever the respective function does.
}

