% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnRange}
\alias{getSpatAnnRange}
\alias{getSpatAnnRange,spata2-method}
\alias{getSpatAnnRange,HistoImaging-method}
\title{Obtain spatial annotations range}
\usage{
getSpatAnnRange(object, ...)

\S4method{getSpatAnnRange}{spata2}(object, id, expand = 0, scale_fct = 1, ...)

\S4method{getSpatAnnRange}{HistoImaging}(object, id, scale_fct = 1)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}
}
\value{
List of length two. Named with \emph{x} and \emph{y}. Each slot
contains a vector of length two with the minima and maxima in pixel.
}
\description{
Extracts the minimum and maximum x- and y-coordinates
of the spatial annotation border.
}
