% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnSf}
\alias{getSpatAnnSf}
\title{Obtain simple feature}
\usage{
getSpatAnnSf(object, id, img_name = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The ID of the spatial annotation of interest.}
}
\value{
An object of class \code{POLYGON} from the \code{sf} package.
}
\description{
Exracts an object as created by \code{sf::st_polygon()} that
corresponds to the image annotation.
}
