% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getSpatAnnSummaryDf}
\alias{getSpatAnnSummaryDf}
\title{Obtain spatial annotation summary}
\usage{
getSpatAnnSummaryDf(
  object,
  ids = NULL,
  area = TRUE,
  unit_area = "mm2",
  center = TRUE,
  unit_center = "px",
  genes = NULL,
  summarize_with = "mean",
  tags_to_lgl = TRUE,
  tags_keep = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{area}{Logical. If \code{TRUE}, the area of each spatial annotation
is added in a variable named \emph{area}.}

\item{unit_area}{The unit of the \emph{area} variable.}

\item{center}{Logical. If \code{TRUE}, two variables named \emph{center_x} and
\emph{center_y} area added providing the center coordinates of the image
annotation.}

\item{unit_center}{The unit of the center variables.}

\item{genes}{Character value or \code{NULL}. If character, the gene expression
of the named genes is summarized among all barcode spots that fall in the
area of the spatial annotation and are added as a variable.}

\item{summarize_with}{Character value. The summarizing function with
which the gene expression values are summarized.}

\item{tags_to_lgl}{Logical. If \code{TRUE}, tag information is displayed in logical
variables where each variable is named like one of the unique tags and
every value is either \code{TRUE} if the annotation cotnains the tag or \code{FALSE}
if not.}

\item{tags_keep}{Logical. If \code{TRUE}, variable \emph{tags} is not removed if
\code{tags_to_lgl} is \code{TRUE}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Data.frame in which each row corresponds to an spatial annotation identified
by the variable \emph{id}.
}
\description{
Extracts information about spatial annotations in a
data.frame.
}
\section{Selection of spatial annotations}{


Selection of spatial annotations via the arguments \code{ids}, \code{class}, \code{tags} and
\code{test} works in three steps:

First, if \code{ids} is a character it prefilters the annotations by ID and only
the specified ones are submitted to the next steps. If it is \code{NULL}, all
annotations are submitted to the next steps.

Secondd, if \code{class} is a character it filters the annotations remaining
after the first step by their class. If \code{NULL}, the step is skipped.

Third, if \code{tags} is a character it is used in combination with \code{test} to select
from the spatial annotations that remain after the second step based on the meta data
they are tagged with. There are multiple options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags. Can contain tags that are not specified.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags. Can not be tagged with anything else.
\item Argument \code{test} set to \emph{not_identical} or \emph{4}: To be included, an image
annotation must \strong{not} be tagged with the combination of input tags.
\item Argument \code{test} set to \emph{'none'} or \emph{5}: To be included, an image annotation
must \strong{not} contain any of the input tags.
}

If \code{tags} is \code{NULL}, the step is skipped. Therefore, if \code{ids}, \code{class} and \code{tags}
are all NULL, which is the default, all annotations are selected as all subsetting
steps are skipped. Eventually, the remaining spatial annotations are submitted to
whatever the respective function does.
}

