% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getSpatialData}
\alias{getSpatialData}
\title{Obtain object of class \code{SpatialData}}
\usage{
getSpatialData(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Object of class \code{SpatialData}.
}
\description{
Extracts the S4-object used as a container for
images.
}
\note{
\code{getImageObject()} is deprecated as of version v3.0.0 in favor
of \code{getSpatialData()}.
}
\seealso{
\code{\link[=getImage]{getImage()}},\code{\link[=getHistoImage]{getHistoImage()}}
}
