% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getSpatialMethod}
\alias{getSpatialMethod}
\title{Obtain spatial method}
\usage{
getSpatialMethod(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Character value.
}
\description{
Extracts an S4 object of class \code{SpatialMethod} that contains
meta data about the set up of the protocol that was followed to create
the data used for the \code{SPATA2} object.
}
\seealso{
\code{?SpatialMethod}
}
