% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatialTrajectory}
\alias{getSpatialTrajectory}
\alias{getSpatialTrajectories}
\title{Obtain spatial trajectories}
\usage{
getSpatialTrajectory(object, id)

getSpatialTrajectories(object, ids = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}

\item{ids}{Character vector. Denotes the spatial trajectories
of interest.}
}
\value{
An object of class \code{SpatialTrajectory} in case of \code{getSpatialTrajectory()}
or a named list of \code{SpatialTrajectory} objects in case of \code{getSpatialTrajectories()}.
An empty list if \code{nSpatialTrajectories() == 0}.
}
\description{
Extracts objects of class \code{\link{SpatialTrajectory}}.
}
