% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatialTrajectoryIds}
\alias{getSpatialTrajectoryIds}
\title{Obtain spatial trajectory IDs}
\usage{
getSpatialTrajectoryIds(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts IDs of spatial trajectories that were
drawn with \code{createSpatialTrajectories()}
}
