% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused.R
\name{getSpCorCluster}
\alias{getSpCorCluster}
\alias{getSpCorClusterNames}
\title{Obtain cluster results based on spatial correlation analysis}
\usage{
getSpCorCluster(object, method_hclust = "complete", of_sample = NA)

getSpCorClusterNames(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_hclust}{Character value. Denotes the method that was used to generate the
clustering results you want to extract.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
The list containing all information about the clustering results.
}
\description{
Obtain cluster results based on spatial correlation analysis
}
