% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unused.R
\name{getSpCorResults}
\alias{getSpCorResults}
\title{Obtain spatial correlation results}
\usage{
getSpCorResults(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
The list containing all information about the clustering results.
}
\description{
Obtain spatial correlation results
}
\keyword{internal}
