% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpotSize}
\alias{getSpotSize}
\alias{getSpotSize,spata2-method}
\alias{getSpotSize,HistoImaging-method}
\title{Obtain spot size}
\usage{
getSpotSize(object, ...)

\S4method{getSpotSize}{spata2}(object, ...)

\S4method{getSpotSize}{HistoImaging}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Numeric value.
}
\description{
Extracts the spot size with which to display
the barcoded spots in surface plots.
}
