% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getStsDf}
\alias{getStsDf}
\title{Obtain spatial trajectory screening data.frame}
\usage{
getStsDf(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  methods_gs = NULL,
  smooth_span = 0,
  format = "wide",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame of inferred gradients related to the
course of a trajectory.
}
