% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getTissueOutlineCentroid}
\alias{getTissueOutlineCentroid}
\alias{getTissueOutlineCentroid,HistoImaging-method}
\alias{getTissueOutlineCentroid,HistoImage-method}
\title{Obtain tissue outline centroid}
\usage{
getTissueOutlineCentroid(object, ...)

\S4method{getTissueOutlineCentroid}{HistoImaging}(
  object,
  name = NULL,
  transform = TRUE,
  persp = "ccs",
  ...
)

\S4method{getTissueOutlineCentroid}{HistoImage}(object, transform = TRUE, persp = "ccs", ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Numeric vector of length two.
}
\description{
Extracts the centroid of the polygon used to outline
the whole tissue.
}
