% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getTrajectory}
\alias{getTrajectory}
\title{Deprecated}
\usage{
getTrajectory(object, id, ...)
}
\description{
Deprecated in favor of \code{\link[=getSpatialTrajectory]{getSpatialTrajectory()}}.
}
\keyword{internal}
