% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generic-functions.R
\name{getTrajectoryComment,spata-method}
\alias{getTrajectoryComment,spata-method}
\title{Obtain a trajectory comment}
\usage{
\S4method{getTrajectoryComment}{spata}(object, trajectory_name, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object or a valid spatialTrajectory-object.}

\item{trajectory_name}{The trajectory specified as a character value.}

\item{of_sample}{The sample from which to extract the content.}
}
\description{
Obtain a trajectory comment
}
