% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generic-functions.R
\name{getTrajectoryComment}
\alias{getTrajectoryComment}
\alias{getTrajectoryComment,spatial_trajectory-method}
\title{Obtain a trajectory comment}
\usage{
getTrajectoryComment(object, ...)

\S4method{getTrajectoryComment}{spatial_trajectory}(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the content.}

\item{trajectory_name}{The trajectory specified as a character value.}
}
\description{
Obtain a trajectory comment
}
