% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getN-Z.R
\name{getTrajectoryDf}
\alias{getTrajectoryDf}
\title{Obtain a summarized trajectory data.frame}
\usage{
getTrajectoryDf(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  method_gs = NULL,
  normalize = TRUE,
  summarize_with = FALSE,
  smooth_span = 0,
  format = "wide",
  verbose = NULL,
  ...
)

getTrajectoryDf(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  method_gs = NULL,
  normalize = TRUE,
  summarize_with = FALSE,
  smooth_span = 0,
  format = "wide",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}

\item{variables}{Character vector. The variables of interest:

\itemize{
\item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
\item{ \strong{Genes}: Must be in \code{getGenes()}}
\item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
}}

\item{binwidth}{Distance value. The width of the bins to which
the barcode-spots are assigned. Defaults to the center-center
distance: \code{binwidth = getCCD(object)}.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{shift_wider}{Logical. If set to TRUE the trajectory data.frame is
shifted to it's wider format. Formats can be changed via \code{shiftTrajectoryDf()}.}
}
\value{
A summarized trajectory data.frame.

Data.frame. (See details for more.)
}
\description{
Computes the expression trends of all specified variables
along the direction of the spatial trajectory.

Extracts a data.frame that contains information about barcode-spots
needed for analysis related to \code{spatialTrajectoryScreening()}.
}
\details{
Initially the compiled trajectory data.frame of the specified trajectory
is joined with the respective input of variables via \code{joinWithVariables()}.

The argument \code{binwidth} refers to the amount of which the barcode-spots of the
given trajectory will be summarized with regards to the trajectory's direction:
The amount of \code{binwidth} and the previously specified 'trajectory width' in \code{createTrajectories()}
determine the length and width of the sub-rectangles in which the rectangle the
trajectory embraces are splitted and in which all barcode-spots are binned.
Via \code{dplyr::summarize()} the variable-means of every sub-rectangle are calculated.
These mean-values are then arranged according to the trajectory's direction.

Eventually the data.frame is shifted via \code{tidyr::pivot_longer()} to a data.frame in which
every observation refers to the mean-value of one of the specified variable-elements (e.g. a specified
gene set) of the particular sub-rectangle. The returned data.frame contains the following variables:

\itemize{
\item{\emph{trajectory_part}: Character. Specifies the trajectory's sub-part of the observation. (Negligible if there is
only one trajectory part.)}
\item{\emph{trajectory_part_order}: Numeric. Indicates the order within the trajectory-part. (Negligible if there is
only one trajectory part.)}
\item{\emph{trajectory_order}: Numeric. Indicates the order within the whole trajectory.}
\item{\emph{gene_sets, genes or features}: Character. The respective gene sets, gene or feature the value refers to.}
\item{\emph{values}: Numeric. The actual summarized values.}}
}
