% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getTrajectoryIds}
\alias{getTrajectoryIds}
\title{Deprecated}
\usage{
getTrajectoryIds(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatialTrajectoryIds]{getSpatialTrajectoryIds()}}.
}
\keyword{internal}
