% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrajectoryLength}
\alias{getTrajectoryLength}
\title{Obtain the length of a trajectory}
\usage{
getTrajectoryLength(object, id)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Numeric value.
}
\description{
This function returns the length (the number of bins) of a trajectory
depending on the chosen \code{binwidth}.
}
