% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{getTrajectoryObject}
\alias{getTrajectoryObject}
\title{Deprecated}
\usage{
getTrajectoryObject(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatialTrajectory]{getSpatialTrajectory()}}.
}
\keyword{internal}
