% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTrajectorySegmentDf}
\alias{getTrajectorySegmentDf}
\title{Obtain trajectory course}
\usage{
getTrajectorySegmentDf(object, id = idST(object), ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Data.frame.
}
\description{
Extracts data.frame that contains the course
of a spatial trajectory.
}
