% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getVariableNames}
\alias{getVariableNames}
\title{Obtain variable names}
\usage{
getVariableNames(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Character vector.
}
\description{
Extracts a character vector of variable names that are currently
known to the \code{spata2} object.
}
