% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getVarTypeList}
\alias{getVarTypeList}
\title{Get variable type list}
\usage{
getVarTypeList(object, variables = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{A character vector specifying the subset of variables
to include in the output. By default, all variables known to the
object are returned in the output list.}
}
\value{
A list containing the names of variables categorized by type.
}
\description{
Retrieves a list of variable types present in the \code{SPATA2} object.
}
\details{
This function categorizes variables into different types,
including spatial coordinates, molecules, signatures, meta features,
and additional information like barcodes and sample identifiers. If
the 'variables' argument is provided as a character vector,
the function returns only the specified variables categorized by type.
Otherwise, it returns all variables categorized by type.
}
\examples{
# Get variable type list for all variables in the object
var_types <- getVarTypeList(object)

# Get variable type list for specific variables
var_types <- getVarTypeList(object, variables = c("var1", "var2"))

}
\seealso{
\code{getCoordsDf()}, \code{getMetaDf()}
}
