% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{getWindowSize}
\alias{getWindowSize}
\alias{getWindowSize,HistologyImage-method}
\title{Obtain window size of padded image}
\usage{
getWindowSize(object, ...)

\S4method{getWindowSize}{HistologyImage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Numeric value.
}
\description{
Extracts the window size of the padded image in pixel.
}
