% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpInit}
\alias{ggpInit}
\title{Initiate ggplot2 layering}
\usage{
ggpInit(object = "object", theme = "void", data = "coords")
}
\arguments{
\item{object}{A valid spata-object.}

\item{theme}{Character value. String that denotes the default
theme. Defaults to \code{void}}
}
\value{
An empty ggplot.
}
\description{
Initiates a ggplot object to which \code{ggpLayer}-
functions can be added for individual plotting ideas.
}
