% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerAxesClean}
\alias{ggpLayerAxesClean}
\title{Display clean axes}
\usage{
ggpLayerAxesClean(..., object = NULL)
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Removes axis text, -ticks and -titles (labs) from the plot.
}
