% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerAxesEUOL}
\alias{ggpLayerAxesEUOL}
\title{Display axes with European Units of Length}
\usage{
ggpLayerAxesEUOL(
  object,
  eUOL = getMethodUnit(object),
  which = c("x", "y"),
  frame_by = "coords",
  breaks_x = NULL,
  breaks_y = NULL,
  add_labs = TRUE,
  round = 2
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{eUOL}{The desired unit in wich the axes are displayed. Defaults to the unit
in which the original size of the image of the spatial method is
provided. Obtain valid input options with \code{validUnitsOfLength()}.}

\item{which}{One or two of \emph{'x'} and \emph{'y'}. Specifies
for which axes the transformation is performed. Defaults to both.}

\item{frame_by}{Either \emph{'coords'} or \emph{'image'} or \code{NULL}.
If specified, sets the plot frame accordingly.}

\item{breaks_x, breaks_y}{Vector of distance inputs. Can be pixel or European
units of lengths. If European unit of lengths, input is transformed to pixels as
the plot is plotted with pixel-based coordinates. If \code{NULL}, is set
automatically to five breaks equally distributed along the axis.}

\item{add_labs}{Logical. If \code{TRUE}, adds informative x- and y-labs to
the plot.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Performs necessary transformations to display axes of
surface plots with European Units of Length.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
