% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerCaptureArea}
\alias{ggpLayerCaptureArea}
\title{Add capture area to surface plot}
\usage{
ggpLayerCaptureArea(
  object,
  opt = c("outline"),
  line_clr = "black",
  line_type = "solid",
  line_size = 1,
  expand_x = ggplot2::waiver(),
  expand_y = ggplot2::waiver(),
  img_name = activeImage(object)
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{opt}{Combination of \emph{'rect'} and/or \emph{'crop'}.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{expand_x, expand_y}{Given to \code{expand} of \code{ggplot2::scale_x/y_continuous()}.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
List of ggpLayer outputs.
}
\description{
Adds the capture area as a rectangular and/or
crops the frame of the plot accordingly.
}
\examples{

library(SPATA2)

data("example_data")

object <- loadExampleObject("UKF275T")

plotImage(object) +
 ggpLayerCaptureArea(object, opt = "outline")

plotImage(object) +
 ggpLayerCaptureArea(object, opt = "crop")

}
\seealso{
\code{\link[=getCaptureArea]{getCaptureArea()}}
}
