% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerColorGroupScale}
\alias{ggpLayerColorGroupScale}
\title{Add group specific color spectrum}
\usage{
ggpLayerColorGroupScale(
  object,
  grouping,
  group,
  clrp,
  clrp_adjust = NULL,
  low = ggplot2::alpha("white", 0),
  aes = "color",
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{clrp, clrp_adjust}{The colorpalette and adjustment used to visualize the grouping.}

\item{low}{The color against which to plot.}

\item{aes}{Either \emph{'color'} or \emph{'fill'}.}

\item{...}{Additional arguments given to \code{ggplot2::scale_color_gradient()}}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Creates a color spectrum from the color used to
represent a group to transparent white (can be changed) to maintain
a consistent color scheme.
}
