% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{ggpLayerCoords}
\alias{ggpLayerCoords}
\alias{ggpLayerCoords,HistoImaging-method}
\title{Adds identified entities to the surface plot}
\usage{
ggpLayerCoords(object, ...)

\S4method{ggpLayerCoords}{HistoImaging}(
  object,
  name = NULL,
  pt_alpha = 0.9,
  pt_clr = "lightgrey",
  pt_size = NULL,
  xrange = NULL,
  yrange = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Plots identified or known entities of the object
to the plot based on their x- and y-coordinates.
}
