% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{ggpLayerEncirclingGroups}
\alias{ggpLayerEncirclingGroups}
\title{Deprecated}
\usage{
ggpLayerEncirclingGroups(...)
}
\description{
Deprecated in favor of \code{\link[=ggpLayerGroupOutline]{ggpLayerGroupOutline()}}.
}
\keyword{internal}
