% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerEncirclingIAS}
\alias{ggpLayerEncirclingIAS}
\title{Add IAS area expansion}
\usage{
ggpLayerEncirclingIAS(
  object,
  id,
  buffer,
  n_bins_circle,
  linecolor = "black",
  linesize = 1
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}.}

\item{n_bins_circle}{Numeric value (or vector of length two, see details
for more). Specifies how many times the area is buffere with the value
denoted in \code{buffer}. The combination of both results in area that is
screened.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Adds the circular expansion used by the IAS-algorithm
of the area of  an image annotation to a surface plot.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
