% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{ggpLayerEncirclingSAS}
\alias{ggpLayerEncirclingSAS}
\title{Deprecated}
\usage{
ggpLayerEncirclingSAS(...)
}
\description{
Deprecated in favor of \code{\link[=ggpLayerExprEstimatesSAS]{ggpLayerExprEstimatesSAS()}}.
}
\keyword{internal}
