% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerFrame}
\alias{ggpLayerFrame}
\alias{ggpLayerFixFrame}
\title{Fix ggplot frame}
\usage{
ggpLayerFrame(object, xrange, yrange, expand = FALSE)

ggpLayerFixFrame(object)
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Fixes the frame of an surface plot based
on the coordinates range of the \code{SPATA2} object in
case of \code{ggpLayerFixFrame()} or based on specific distance
inputs in case of \code{ggpLayerFrame()}.
}
