% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByImage}
\title{Set plot limits}
\usage{
ggpLayerFrameByCoords(object = "object")

ggpLayerFrameByImage(object = "object")
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
List.
}
\description{
Sets the limits on the x- and y-axis of a ggplot based on the coordinate
range or the image range.
}
\note{
Always adds \code{ggplot2::coord_equal()}.
}
