% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerHorizonSAS}
\alias{ggpLayerHorizonSAS}
\title{Add SAS screening horizon}
\usage{
ggpLayerHorizonSAS(
  object,
  id,
  distance = distToEdge(object, id),
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1.5,
  line_type = "solid",
  incl_edge = TRUE,
  incr_vert = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance measure}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. Defaults to a distance
that covers the whole tissue section the spatial annotation is located
on using \code{\link[=distToEdge]{distToEdge()}}. (This distance must not be exceeded.)}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. The size with which to display lines
of the expression estimates.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{incl_edge}{Logical. If \code{TRUE}, makes use of \code{SPATA2} automatic tissue outline algorithm.}

\item{incr_vert}{Logical value. If \code{TRUE}, the number of vertices of the polygon
to be plotted is increased by linear interpolation. This can lead to smoother
visualization but also to increased runtime.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Visualizes the border between screened tissue (environment)
and everything beyond that is not included in the screening (periphery).
}
\examples{

library(SPATA2)
library(tidyverse)

object <- loadExampleObject("UKF275T")

object <-
 createNumericAnnotations(
   object = object,
   variable = "HM_HYPOXIA",
   threshold = "kmeans_high",
   id = "hypoxia_ann",
   force1 = TRUE
   )

plotSurface(object, color_by = "HM_HYPOXIA", outline = T, pt_clrsp = "Reds 3") +
 ggpLayerHorizonSAS(object, id = "hypoxia_ann", distance = "3mm") +
 ggpLayerScaleBarSI(object, sb_pos = c("3mm", "2mm"))

}
