% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerImage}
\alias{ggpLayerImage}
\title{Add histology image}
\usage{
ggpLayerImage(object = "object")
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
List.
}
\description{
Creates ggplot2 layer with the histology image
as a raster annotation.
}
\note{
The returned list contains an additional \code{ggplot2::geom_point()}
layer with invisible barcode spots coordinates (alpha = 0) to enable the
image plotting.
}
