% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{ggpLayerImageAnnotation}
\alias{ggpLayerImageAnnotation}
\title{Deprecated}
\usage{
ggpLayerImageAnnotation(...)
}
\description{
Deprecated in favor of \code{\link[=ggpLayerSpatAnnOutline]{ggpLayerSpatAnnOutline()}}.
}
\keyword{internal}
