% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{ggpLayerImgAnnBorder}
\alias{ggpLayerImgAnnBorder}
\title{Deprecated}
\usage{
ggpLayerImgAnnBorder(...)
}
\description{
Deprecated in favor of \code{\link[=ggpLayerSpatAnnOutline]{ggpLayerSpatAnnOutline()}}.
}
\keyword{internal}
