% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/g.R
\name{ggpLayerImgAnnBorder}
\alias{ggpLayerImgAnnBorder}
\alias{ggpLayerImgAnnOutline}
\title{Add borders of annotated structures}
\usage{
ggpLayerImgAnnBorder(...)

ggpLayerImgAnnOutline(
  object = "object",
  ids = NULL,
  tags = NULL,
  test = "any",
  alpha = 0.5,
  fill = NA,
  line_color = "black",
  line_size = 1.5,
  line_type = "solid",
  use_colors = FALSE,
  inner = FALSE,
  ...
)
}
\arguments{
\item{...}{Used to absorb deprecated arguments or functions.}

\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of image annotation with tags} for more information.}

\item{test}{Character value. One of \emph{any}. \emph{all}, \emph{identical}, \emph{not_identical} and
\emph{none}. Specifies how input for \code{tags} is used to select image annotations.
See section \emph{Selection of image annotation with tags} for more information.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{use_colors}{Logical value. If \code{TRUE}, the color aesthetic is used to display
each image annotation in a different color while providing a legend.}

\item{inner}{Logical value. If \code{FALSE}, only outer borders of the image annotation
are displayed.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds ggplot2 layer of polygons of structures that were annotated within the image
with \code{createImageAnnotations()}.
}
\note{
Adds two additional layers to set the scales for the color- and
fill aesthetic of the plot.
}
\keyword{internal}
