% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerImgAnnPointer}
\alias{ggpLayerImgAnnPointer}
\title{Add pointer towards image annotations}
\usage{
ggpLayerImgAnnPointer(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  color_by = NULL,
  ptr_angles = 45,
  ptr_labels = NULL,
  ptr_lengths = "250um",
  ptr_alpha = 0.9,
  ptr_arrow = NULL,
  ptr_color = "black",
  ptr_size = 1,
  text_alpha = 0.9,
  text_color = "black",
  text_dist = 0,
  text_nudge_x = 0,
  text_nudge_y = 0,
  text_size = 4,
  point_at = "center",
  seed = NULL,
  clrp = NULL,
  clrp_adjust = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of image annotation with tags} for more information.}

\item{test}{Character value. One of \emph{any}. \emph{all}, \emph{identical}, \emph{not_identical} and
\emph{none}. Specifies how input for \code{tags} is used to select image annotations.
See section \emph{Selection of image annotation with tags} for more information.}

\item{color_by}{Character value or \code{NULL}. If character, one of \emph{'id'} or \emph{'label'}
which colors the the pointers accordingly.}

\item{ptr_angles, ptr_lengths}{Numeric value of length 1 or of length equal to the number
of image annotations. Specifies the angle from which the segments points
towards the image annotation as well as their length. \code{ptr_lengths} works
within the SPATA2 distance framework. See section \emph{Distance measures} for more
information.}

\item{ptr_labels}{Specifies if and how the pointers are labeled. If \code{NULL},
the default, the image annotations are labeled by their ID. If character,
specifies the exact label of each image annotation and should be of length 1
or of length equal to the number of image annotations. If \code{FALSE}, no text
is displayed.}

\item{ptr_alpha}{Numeric value. Specifies the transparency of the pointers.}

\item{ptr_arrow}{\code{NULL} or \code{arrow} as displayed by \code{grid::arrow()}.}

\item{ptr_color}{Character value. Specifies the color of the pointers if
\code{color_by} is not a character.}

\item{ptr_size}{Numeric value. Specifies the size (thickness) of the pointers.}

\item{text_dist}{Distance measure. Specifies the distance from the text to
the pointer.}

\item{point_at}{Character value. If \emph{'center'}, the pointer is directed at
the center of the image annotation. If \emph{'border'}, the pointer points
at a random point of the image annotation border - recommended if the
image annotation is big.}

\item{seed}{Numeric value or \code{NULL}. If numeric, sets seed before picking
a random point of the image annotation border if \code{point_at = 'border'}.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds segments and, if desired, labels to the surface plot that
point towards and highlight the position of image annotations.
}
\section{Distance measures}{


Several functions in \code{SPATA2} have arguments that take \emph{distance input}.
To specifically refer to a distance the unit must be specified. There are
three ways to create valid input for these arguments.

\strong{1. In pixel:}

There are two valid input options to specify the distance in pixel:

\itemize{
\item{numeric:}{ Single numeric values, e.g. \code{arg_input = c(2, 3.554, 69, 100.67)}. If no unit
is specified the input will be interpreted as pixels.}
\item{character:}{ Suffixed with \emph{'px'}, e.g. \code{arg_input = c('2px', '3.554px', '69px', '100.67px')}}
}

Note: The unit pixel (px) is used for distances as well as for areas. If pixel
refers to a distance the pixel side length is meant. If pixel refers to an area the
number of pixels is meant.

\strong{2. According to the Systeme international d`unites (SI):}

Specifying distances in SI units e.g. \code{arg_input = c('2mm', '4mm')} etc.
requires the input to be a character as the unit must be provided as suffix.
Between the numeric value and the unit must be no empty space! Valid suffixes
can be obtained using the function \code{validUnitsOfLengthSI()}.

\strong{3. As vectors of class \code{unit}:}

Behind the scenes \code{SPATA2} works with the \code{units} package. Input
is converted into vectors of class \code{units}. Therefore, input can be directly
provided this way: \code{arg_input = units::set_unit(x = c(2,4), value = 'mm')}
Note that \emph{pixel} is not a valid unit in the \code{units} package. If you want
to specify the input in pixel you have to use input option 1. In pixel.
}

