% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerRect}
\alias{ggpLayerRect}
\title{Add a rectangular to the plot}
\usage{
ggpLayerRect(
  object = "object",
  xrange,
  yrange,
  alpha = 0,
  color = "black",
  size = 1,
  expand = 0,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{xrange, yrange}{Numeric vectors of length 2. Specify the range
from the rectangular on the x- and y-axis.}

\item{alpha, color, fill, size}{Given to \code{ggplot2::geom_rect()}.}

\item{expand}{Numeric value or vector of length two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the xspan and the second
value is used to adjust the yspan. E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original xspan to the xspan of the image section and does not adjust
 the yspan.}

\item{...}{Additional arguments given to \code{ggplot2::geom_rect()}.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Adds a rectangular to the plot.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
