% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{ggpLayerSampleMask}
\alias{ggpLayerSampleMask}
\title{Deprecated}
\usage{
ggpLayerSampleMask(...)
}
\description{
Deprecated in favor of \code{\link[=ggpLayerTissueOutline]{ggpLayerTissueOutline()}}.
}
\keyword{internal}
