% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerScaleBarEUOL}
\alias{ggpLayerScaleBarEUOL}
\title{Add a European units of length scale bar}
\usage{
ggpLayerScaleBarEUOL(
  object,
  pos = "top_right",
  dist_sb = "1mm",
  sgmt_alpha = 1,
  sgmt_color = "black",
  sgmt_size = 1.25,
  sgmt_type = "solid",
  text_alpha = 1,
  text_color = "black",
  text_nudge_x = 0,
  text_nudge_y = NULL,
  text_size = 5.5,
  text_type = 0.9,
  xrange = NULL,
  yrange = NULL,
  offset = c(0.9, 0.9),
  theme_opt = "none"
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{pos}{Character value or vector of length two. If character,
one of \emph{top_right}, \emph{top_left}, \emph{bottom_right} or \emph{bottom_left}. The scale
bar is positioned accordingly. If vector of length two, distance measures
that specify the positioning. First value is taken for positioning on x-
and second value is taken for positioning on the y-axis.}

\item{dist_sb}{The distance in European units of length that the scale bar
illustrates.}

\item{sgmt_alpha, sgmt_color, sgmt_size, sgmt_type}{Parameters given to
\code{ggplot2::geom_segment()} that control the appearance of segments
of the plot.}

\item{text_nudge_y}{Numeric value or \code{NULL}. Moves the scale bar text away from
the scale bar. If \code{NULL}, nudging is computed based on the input of \code{yrange}.
Set manually, if the positioning fails.}

\item{xrange, yrange}{The range of the image that is considered if the positioning
of the scale is calculated via \code{pos} as one of \emph{top_right}, \emph{top_left}, \emph{bottom_right}
or \emph{bottom_left}. Defaults to the image range.}

\item{offset}{Numeric vector of length two. Used to move the position of
the scale bar away from the center. Values should range from 0 to 1. First
value is used to move along the x-axis. Second value is used for the y-axis.}
}
\description{
Adds a scale bar to the surface plot that visualizes
distance in European units of length. Segment of the line is drawn
with \code{ggplot2::geom_segment()}. Text is drawn with \code{ggplot2::geom_text()}.
}
\details{
If \code{pos} is one of \emph{top_right}, \emph{top_left}, \emph{bottom_right}
or \emph{bottom_left}, the position of the scale bar is computed in combination
with the input for argument \code{offset}. Argument \code{offset} is used to repel
the scale bar away from the center into the corner specified in \code{pos}. Thus,
if \code{offset = c(0,0)}, the scale bar is positioned in the center of the plot
regardless of the specification of \code{pos}. Offset values specify the percentage
of the distanec between the center of the plot and its limits. For instance,
if \code{pos = c(0.5, 0.75)} and \code{pos = 'top_right'} to the right (50\% of the distance
to the border - x-axis) and to the top (75\% of the distance between the center
and the ceiling).
}
