% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerScreeningDirectionSAS}
\alias{ggpLayerScreeningDirectionSAS}
\title{Add SAS screening direction}
\usage{
ggpLayerScreeningDirectionSAS(
  object,
  ids,
  distance = "dte",
  line_alpha = 1,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector. Specifies the IDs of the spatial annotations of interest.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance measure}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. Defaults to a distance
that covers the whole tissue section the spatial annotation is located
on using \code{\link[=distToEdge]{distToEdge()}}. (This distance must not be exceeded.)}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
Visualizes the screenining direction of an \link[=spatialAnnotationScreening]{SAS}
set up on top of a surface plots. See examples.
}
\details{
\strong{In contrast} to \code{\link[=ggpLayerExprEstimates]{ggpLayerExprEstimates()}}, which visualizes
the precise positions of the expression estimates, \code{ggpLayerScreeningDiretionSAS()} visualizes
only the concept, the idea, of the screening direction. This is particularly useful, if the screening
set up includes multiple annotations in one tissue section, where \code{ggpLayerExprEstimates()} fails
to visualize the combined expression estimates.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

ids <- getSpatAnnIds(object, tags = c("necrotic", "compr"))

plotImage(object) +
  ggpLayerSpatAnnOutline(object, ids = ids, fill = "lightgrey") +
  ggpLayerScreeningDirectionSAS(object, ids = ids, line_size = 0.5)


}
