% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerSpatAnnPointer}
\alias{ggpLayerSpatAnnPointer}
\title{Add pointer towards spatial annotations}
\usage{
ggpLayerSpatAnnPointer(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  color_by = NULL,
  ptr_angles = 45,
  ptr_labels = NULL,
  ptr_lengths = "250um",
  ptr_alpha = 0.9,
  ptr_arrow = NULL,
  ptr_color = "black",
  ptr_size = 1,
  text_alpha = 0.9,
  text_color = "black",
  text_dist = 0,
  text_nudge_x = 0,
  text_nudge_y = 0,
  text_size = 4,
  point_at = "center",
  seed = NULL,
  clrp = NULL,
  clrp_adjust = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of spatial annotations} for more information.}

\item{test}{Character value. One of \emph{any}. \emph{all}, \emph{identical}, \emph{not_identical} and
\emph{none}. Specifies how input for \code{tags} is used to select image annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{color_by}{Character value or \code{NULL}. If character, one of \emph{'id'} or \emph{'label'}
which colors the the pointers accordingly.}

\item{ptr_angles, ptr_lengths}{Numeric value of length 1 or of length equal to the number
of spatial annotations. Specifies the angle from which the segments points
towards the spatial annotation as well as their length. \code{ptr_lengths} works
within the SPATA2 distance framework. See section \emph{Distance measures} for more
information.}

\item{ptr_labels}{Specifies if and how the pointers are labeled. If \code{NULL},
the default, the spatial annotations are labeled by their ID. If character,
specifies the exact label of each spatial annotation and should be of length 1
or of length equal to the number of spatial annotations. If \code{FALSE}, no text
is displayed.}

\item{ptr_alpha}{Numeric value. Specifies the transparency of the pointers.}

\item{ptr_arrow}{\code{NULL} or \code{arrow} as displayed by \code{grid::arrow()}.}

\item{ptr_color}{Character value. Specifies the color of the pointers if
\code{color_by} is not a character.}

\item{ptr_size}{Numeric value. Specifies the size (thickness) of the pointers.}

\item{text_dist}{Distance measure. Specifies the distance from the text to
the pointer.}

\item{point_at}{Character value. If \emph{'center'}, the pointer is directed at
the center of the spatial annotation. If \emph{'border'}, the pointer points
at a random point of the spatial annotation border - recommended if the
spatial annotation is big.}

\item{seed}{Numeric value or \code{NULL}. If numeric, sets seed before picking
a random point of the spatial annotation border if \code{point_at = 'border'}.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds segments and, if desired, labels to the surface plot that
point towards and highlight the position of spatial annotations.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

