% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerSpatialTrajectories}
\alias{ggpLayerSpatialTrajectories}
\title{Add spatial trajectory layer}
\usage{
ggpLayerSpatialTrajectories(
  object = "object",
  ids,
  arrow = ggplot2::arrow(length = ggplot2::unit(x = 0.125, "inches")),
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector. The IDs of the trajectories
that should be plotted.}

\item{arrow}{A list of arguments given to \code{ggplot2::arrow()}. Based
on which the trajectories are plotted.}

\item{...}{Additional arguments given to \code{ggplot2::geom_segment()}.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds spatial trajectories in form of arrows to a surface plot.
}
