% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{ggpLayerSpots}
\alias{ggpLayerSpots}
\alias{ggpLayerSpots,spata2-method}
\alias{ggpLayerSpots,HistoImaging-method}
\alias{ggpLayerSpots,data.frame-method}
\title{Adds barcoded spots to the surface plot}
\usage{
ggpLayerSpots(object, ...)

\S4method{ggpLayerSpots}{spata2}(
  object,
  img_name = NULL,
  alpha_by = NULL,
  color_by = NULL,
  spot_alpha = 0.9,
  spot_clr = "lightgrey",
  spot_size = getSpotSize(object),
  scale_spot_size = TRUE,
  smooth = FALSE,
  smooth_span = 0.2,
  normalize = NULL,
  method_gs = NULL,
  xrange = NULL,
  yrange = NULL,
  unit = NULL,
  breaks = NULL,
  expand = FALSE,
  scale_fct = 1,
  add_labs = TRUE
)

\S4method{ggpLayerSpots}{HistoImaging}(
  object,
  img_name = NULL,
  alpha_by = NULL,
  color_by = NULL,
  spot_alpha = 0.9,
  spot_clr = "lightgrey",
  spot_size = getSpotSize(object),
  scale_spot_size = TRUE,
  xrange = NULL,
  yrange = NULL,
  unit = NULL,
  breaks = NULL,
  expand = FALSE,
  scale_fct = 1,
  add_labs = TRUE
)

\S4method{ggpLayerSpots}{data.frame}(
  object,
  alpha_by = NULL,
  color_by = NULL,
  spot_alpha = 0.9,
  spot_clr = "lightgrey",
  spot_size = 1,
  scale_fct = 1
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{spot_alpha, spot_size, spot_clr}{Parameters to set the aesthetics
alpha, size, and color of the spots. Arguments \code{alpha_by} and \code{color_by}
are prioritized.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}

\item{unit}{The desired unit. Defaults to the unit
in which the original size of the image of the spatial method is
provided. Obtain valid input options with \code{validUnitsOfLengthSI()}.}

\item{breaks}{Specifies where the breaks are set. Labels are plotted in the unit
specified in \code{unit}. Valid input:

\itemize{
\item{\code{NULL}:}{ No specification. Five breaks are set equally distributed. Does not work with STS/IAS related plots as
the range is taken from the whole image.}
\item{\code{vector}:}{ Vector of distance measures. Breaks are set for axes denoted in \code{which}. (Defaults to both, x and y.)}
\item{\code{list}:}{ List with slots \emph{x} and \emph{y}. Vector of distance measures to set each axis specifically.}
}}

\item{expand}{Specifies how the axis are expanded. Using \code{expand} of \code{ggplot2::scale_x/y_continuous()}.
Valid input:

\itemize{
\item{\code{NULL} or \code{TRUE}:}{ No specification. Default is used.}
\item{\code{FALSE}:}{ No expansion.}
\item{\code{vector}:}{ Numeric vector of length two. Input is set for axes denoted in \code{which}. (Defaults to both, x and y.)}
\item{\code{list}:}{ List with slots \emph{x} and \emph{y}. Numeric vector of length two, used for each axis specifically.}
}}

\item{add_labs}{Logical. If \code{TRUE}, adds x- and y-labs to the plot.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Plots the barcoded spots of \emph{Visium} or \emph{SlideSeq}
experiments in space.
}
\details{
Method for \code{HistoImaging} and \code{spata2}.
}
