% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerTissueOutline}
\alias{ggpLayerTissueOutline}
\title{Add a hull that outlines the tissue}
\usage{
ggpLayerTissueOutline(
  object,
  line_color = "grey",
  line_size = 0.5,
  expand = getCCD(object, unit = "px"),
  inc_outline = TRUE,
  minPts = 1,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{expand}{Given to \code{ggforce::geom_mark_hull()}.}

\item{inc_outline}{Employs the function \code{add_outline_variable()} which uses
\code{dbscan::dbscan()} to identify incoherent tissue sections.}

\item{...}{Additional arguments given to \code{ggforce::geom_mark_hull()}}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a hull that encircles the sample. Useful, if you want
to plot numeric variables by color against white.
}
\examples{

object <- donwloadPubExample("MCD_LMU")

plotImageGgplot(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = TRUE)

plotImageGgplot(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = FALSE)


}
