% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerTrajectoryFrame}
\alias{ggpLayerTrajectoryFrame}
\title{Add the screening frame of a spatial trajectory}
\usage{
ggpLayerTrajectoryFrame(
  object,
  id,
  width = getTrajectoryLength(object, id),
  rect_alpha = 1,
  rect_color = "black",
  rect_linesize = 1,
  rect_linetype = "solid"
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{The identifier of the trajectory segment within the \code{spata2} object.}

\item{width}{Distance measure. The width of the trajectory frame, defaulting to
the trajectory length.}

\item{rect_alpha}{The alpha transparency level of the rectangle's border}

\item{rect_color}{The color of the rectangle's border.}

\item{rect_linesize}{The size of the line used for the rectangle's border.}

\item{rect_linetype}{The type of line used for the rectangle's border.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
This function generates a ggplot layer representing a trajectory
frame based on a specified trajectory segment in a \code{spata2} object. It creates
a rectangular frame around the trajectory segment, with customizable appearance.
}
