% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerZoom}
\alias{ggpLayerZoom}
\title{Set plot limits manually}
\usage{
ggpLayerZoom(xrange = NULL, yrange = NULL)
}
\arguments{
\item{xrange, yrange}{Numeric vectors of length two or NULL. If numeric, specifies the coordinates
with which the image is subsetted. E.g. xrange = c(200, 500) results in the image being cropped
from x-coordinate 200 up to x-coordinate 500.}

\item{xlim, ylim}{Numeric vector of length 2 or NULL. If specified, sets
the range of the plot that is displayed on the x- or y-axis
respectively.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Sets the limits on the x- and y-axis of a ggplot based on
manual input.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
