% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{ggplot_family}
\alias{ggplot_family}
\title{ggplot_family}
\usage{
ggplot_family()
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
ggplot_family
}
