% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{gradientToModelIAS}
\alias{gradientToModelIAS}
\title{Create input for \code{model_add}}
\usage{
gradientToModelIAS(
  object,
  id,
  variables,
  distance = NA_integer_,
  binwidth = getCCD(object),
  n_bins_circle = NA_integer_,
  include_area = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. ID of the spatial trajectory or the image annotation
of interest.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{distance, binwidth, n_bins_circle, n_bins}{The input given to the desired
screening- or visualization functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Generates appropriate input for argument \code{model_add}
of functions related to Spatial Trajectory Screening (STS) or
Image Annotation Screening (IAS). To screen for gradient cooexpression.
}
