% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{GroupAnnotation-class}
\alias{GroupAnnotation-class}
\alias{GroupAnnotation}
\title{The \code{GroupAnnotation} - Class}
\description{
An S4 class designed to represent the spatial extent of data points,
such as cells or barcoded spots, by filtering and outlining them based on predefined groups.
This class allows for the creation of annotations that highlight specific spatial clusters,
areas, or patterns identified through grouping techniques. It provides a means to focus on
regions of interest within spatial multi-omic datasets using predefined categorizations.
}
\details{
This class is an extension of class \code{\link{SpatialAnnotation}} and
inherits all of its slots.
}
\section{Slots}{

\describe{
\item{\code{grouping}}{Character value. The name of the grouping variable that
contains the outlined group.}

\item{\code{group}}{Character value. The name of the outlined group.}

\item{\code{parameters}}{list. A named list of the parameters set up to compute
the spatial outline.}
}}

