% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{HistologyImaging-class}
\alias{HistologyImaging-class}
\alias{HistologyImaging}
\title{The \code{HistologyImaging} - Class}
\description{
Deprecated in favor of \code{SpatialData}.
}
\keyword{internal}
