% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{HistologyImagingNew-class}
\alias{HistologyImagingNew-class}
\alias{HistologyImagingNew}
\title{The \code{HistologyImaging} - Class}
\description{
S4 class that represents a set of histological images from one
tissue slide or several consecutive slides of one and the same tissue portion.
}
\section{Slots}{

\describe{
\item{\code{annotations}}{list. List of objects of class \code{ImageAnnotation}.}

\item{\code{coordinates}}{data.frame. Data.frame that stores information about identified
or known entities located on the imaged tissue, such as cells or capture spots.}

\item{\code{coordinates_id}}{character. The name of the variable of the data.frame
in @slot coordinates that uniquely identifies each observation.}

\item{\code{image_active}}{HistologyImage. The image that is used by default.}

\item{\code{image_reference}}{HistologyImage. The image that is used as a reference for aligning
every additional image in slot @images_registered.}

\item{\code{images_registered}}{list. List of objects of class \code{HistologyImage}. Leaving
slot @image empty can or should be done for more efficient utilization of memory.}

\item{\code{justification}}{list. List containing two sub-slots that track justification
changes applied to the entire tissue. These justification changes are implemented
uniformly across all images within the dataset, following their reading and specific
transformations defined in the respective '@transformation' slot.
\itemize{
\item{\emph{angle}:}{ Numeric value that ranges from 0-359. Defaults to 0.}
\item{\emph{flipped}:}{ List of two logical values named \emph{horizontal} and \emph{vertical}. Both default to \code{FALSE}.}
}}

\item{\code{meta}}{list. List for meta data regarding the imaged tissue portion.}

\item{\code{misc}}{list. A flexible list for miscellaneous input.}

\item{\code{sample}}{character. String to identify the imaged tissue.}
}}

