% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_add_barcode_suffix}
\alias{hlpr_add_barcode_suffix}
\title{Make sure that barcodes are spata-like}
\usage{
hlpr_add_barcode_suffix(input, sample_name)
}
\arguments{
\item{input}{A matrix with columns = barcodes, or a data.frame with a barcode-variable}

\item{sample_name}{Character value.}
}
\description{
Make sure that barcodes are spata-like
}
