% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_add_models}
\alias{hlpr_add_models}
\alias{hlpr_add_residuals}
\alias{hlpr_add_residuals_customized}
\alias{hlpr_summarize_residuals}
\alias{hlpr_name_models}
\alias{hlpr_filter_trend}
\title{Helper functions for trajectory ranking}
\usage{
hlpr_add_models(df, custom_fit = NULL)

hlpr_add_residuals(df, pb = NULL, custom_fit = NULL)

hlpr_add_residuals_customized(df, customized_trends_df, pb = NULL)

hlpr_summarize_residuals(df, pb = NULL)

hlpr_name_models(names)

hlpr_filter_trend(atdf, limit, poi)
}
\arguments{
\item{df}{A data.frame.}

\item{customized_trends_df}{A data.frame of only numeric variables that represent the trends a trajectory
might adopt.}
}
\value{
If used within \code{purrr::map()} a list of data.frames.
}
\description{
Functions to use within \code{purrr::map()} again in \code{dplyr::mutate()} in order to
create a nested ranked trajectory data.frame.

\itemize{
 \item{\code{hlpr_add_models()}: Returns a data.frame of variables corresponding to
 mathematical curves.}
 \item{\code{hlpr_add_residuals(): Calculates the residuals of the variable \emph{values} with respect
 to each mathematical curve.}}
 \item{\code{hlpr_summarise_residuals(): Calculates the area under the curve for every residual in order to
 access the fit of the respective expression trend to the fitted curve.}}}
}
