% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_add_old_coords}
\alias{hlpr_add_old_coords}
\title{Adds old coordinates}
\usage{
hlpr_add_old_coords(object, plot_df, complete)
}
\arguments{
\item{object}{A valid spata-object.}

\item{plot_df}{The plot_df.}

\item{complete}{Logical.}
}
\value{

}
\description{
Adds old coordinates of subsetted object to
plot_df in \code{plotSurface()}.
}
