% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{hlpr_add_residuals2}
\alias{hlpr_add_residuals2}
\title{Title}
\usage{
hlpr_add_residuals2(
  df,
  pb = NULL,
  column_order = "bins_order",
  column_values = "values",
  shift_longer = FALSE
)
}
\arguments{
\item{df}{data.frame that contains a column with the values against which the residuals for
every pattern (remaining columns) are computed.}
}
\description{
Title
}
