% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_adjust_legend_size}
\alias{hlpr_adjust_legend_size}
\title{Adjusts the size of discrete legend points}
\usage{
hlpr_adjust_legend_size(variable, aes, pt_size)
}
\arguments{
\item{variable}{The variable mapped to the color/fill aesthetic.}

\item{aes}{The aesthetic used.}

\item{pt_size}{Numeric value. Specifies the size of all points.}
}
\description{
Adjusts the size of discrete legend points
}
