% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_assign_arguments}
\alias{hlpr_assign_arguments}
\title{Assign arguments to calling function}
\usage{
hlpr_assign_arguments(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\description{
Mediates between the input of the calling function and
the default instructions within the spata object. Every 'default'-argument
of the calling function that is NULL is replaced by what is defined as
the default.
}
