% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_compare_samples}
\alias{hlpr_compare_samples}
\title{Compare samples within an object}
\usage{
hlpr_compare_samples(object, df, messages)
}
\arguments{
\item{object}{A valid spata-object.}

\item{df}{A data.frame}

\item{messages}{The message vector that is generated by the function in which
\code{hlpr_compare_samples()} is called in.}
}
\value{
Updated message vector.
}
\description{
Checks whether the sample column in \code{df}
complies with the sample names of the provided object.
}
\details{
This function is to be used within the \code{SPATA::validateSpataObject()}-
functions. It expands the message-vector that is being generated by the calling
function and returns it.
}
