% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_compile_trajectory}
\alias{hlpr_compile_trajectory}
\title{Compiles a trajectory data.frame}
\usage{
hlpr_compile_trajectory(
  segment_trajectory_df,
  trajectory_width,
  object,
  sample
)
}
\arguments{
\item{segment_trajectory_df}{A data.frame specifying each segment of the whole
trajectory with variables \code{x, y, xend, yend}.}

\item{trajectory_width}{Numeric value that determines the width of the
trajectory.}

\item{object}{A valid spata-object.}
}
\value{
A data.frame containing the variables \emph{barcodes, sample, x, y}
as well as
\itemize{
 \item{\emph{projection_length}: indicating the position of every barcode-spot
 with respect to the direction of the trajectory-part. The higher the barcode-spots
 value is the farther away it is from the starting point of the trajectory-part
 it belongs to. }
 \item{\emph{trajectory_part}: indicating the part of the trajectory the barcode-spot
  belongs to.}
  }
}
\description{
Compiles a trajectory data.frame
}
