% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_display_title}
\alias{hlpr_display_title}
\alias{hlpr_display_subtitle}
\title{No documentation needed}
\usage{
hlpr_display_title(display_title, title)

hlpr_display_subtitle(display_subtitle, subtitle)
}
\description{
No documentation needed
}
