% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_dist_mtr_to_df}
\alias{hlpr_dist_mtr_to_df}
\title{Convert distance matrix to distance data.frame}
\usage{
hlpr_dist_mtr_to_df(dist_mtr, varnames = c("gene1", "gene2"))
}
\arguments{
\item{dist_mtr}{A distance matrix gene expression - gene expression distances}
}
\description{
Reshapes a distance matrix with a wrapper around \code{reshape2::melt()}.
}
