% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_gene_set_name}
\alias{hlpr_gene_set_name}
\title{Removes the class part of a gene set string}
\usage{
hlpr_gene_set_name(string)
}
\arguments{
\item{string}{Gene sets as a character vector}
}
\value{
Gene set name
}
\description{
Removes the class part of a gene set string
}
