% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{hlpr_geom_trajectory_fit}
\alias{hlpr_geom_trajectory_fit}
\title{Easy switch between geom line and and geom smooth}
\usage{
hlpr_geom_trajectory_fit(
  smooth,
  smooth_span,
  plot_df,
  ref_model,
  ref_variable,
  linesize,
  linealpha,
  smooth_se = FALSE
)
}
\description{
To be used in plotTrajectoryFit()/-Customized()
}
