% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{hlpr_image_add_on}
\alias{hlpr_image_add_on}
\alias{hlpr_image_add_on2}
\alias{hlpr_image_to_df}
\title{Provides the image as ggplot background}
\usage{
hlpr_image_add_on(object, display_image, ...)

hlpr_image_add_on2(image)

hlpr_image_to_df(img)
}
\arguments{
\item{object}{A valid spata-object.}

\item{display_image}{Logical value.}

\item{image}{Image input.}
}
\value{
Either null or a ggplot2::geom_annotation_raster
}
\description{
Provides the image as ggplot background
}
