% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{hlpr_join_with_color_by}
\alias{hlpr_join_with_color_by}
\title{Join with single value}
\usage{
hlpr_join_with_color_by(object, df, color_by = NULL, variables = NULL, ...)
}
\description{
Join with single value
}
