% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_labs_add_on}
\alias{hlpr_labs_add_on}
\title{Return customized ggplot:labs()}
\usage{
hlpr_labs_add_on(input, input_str, color_str, display_title)
}
\arguments{
\item{input}{The color_to argument}

\item{input_str}{Title-prefix. Should be one of \emph{'Genes:', 'Gene set:'} or \code{'Feature:'}.}

\item{color_str}{Legend title}

\item{display_title}{Logical. If set to FALSE only the legend-title will be specified.}
}
\value{
A customized \code{ggplot2::labs()}-function.
}
\description{
Helper function
}
