% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_normalize_imap}
\alias{hlpr_normalize_imap}
\title{Normalize gene or gene set values}
\usage{
hlpr_normalize_imap(variable, var_name, aspect, subset)
}
\arguments{
\item{variable}{The variable to normalize (if matches requirements).}

\item{var_name}{The name of the variable to smooth.}

\item{aspect}{Gene or Gene set}

\item{subset}{A character vector of variable names that are to be normalized}

\item{pb}{An R6 progress bar object.}
}
\value{
A normalized variable (data.frame within \code{purrr::imap()})
}
\description{
Helper function to use within \code{purrr::imap()}
}
