% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_normalize_vctr}
\alias{hlpr_normalize_vctr}
\title{Normalize gene or gene set values}
\usage{
hlpr_normalize_vctr(variable)
}
\arguments{
\item{variable}{The variable to normalize (if matches requirements).}
}
\value{
A normalized variable
}
\description{
Normalize gene or gene set values
}
