% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_one_distinct}
\alias{hlpr_one_distinct}
\title{Number of distinct values equal to 1?}
\usage{
hlpr_one_distinct(x, rna_assay, pb = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Input variable.}

\item{rna_assay}{Expression matrix}

\item{pb}{Progress bar object}
}
\value{

}
\description{
Helper function to use within purrr::map_lgl. Returns
TRUE if the gene row of the provided rna assay is uniformly expressed.
}
