% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_process_spatial_correlation_cluster}
\alias{hlpr_process_spatial_correlation_cluster}
\title{Process spatial correlation results}
\usage{
hlpr_process_spatial_correlation_cluster(cutree_df, dist_df, input)
}
\arguments{
\item{cutree_df}{Data.frame of two variables: gene and cluster-belonging.}

\item{dist_df}{Data.frame of three variables: gene1, gene2, distance}
}
\description{
Helper function to be used in \code{clusterSpatialCorrelationResults()}.
Takes a cutree data.frame and a distance data.frame and returns a named list:

\itemize{
 \item{\emph{assessment_df}: A data.frame that attempts to evaluate all cluster's
  quality by providing the average distance between it's genes.}
  \item{\emph{distances_list}: A named list of data.frames. Each data.frame contains the
  gene-gene distances between all genes the cluster it corresponds to contains.}
  \item{\emph{gene_names_list}: A named list of character vectors. Each vector contains the
  unique gene names of the cluster it corresponds to.}
  \item{\emph{k}: k-value}
  \item{\emph{h}: h-value}
  }
}
