% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{hlpr_run_cnva_pca}
\alias{hlpr_run_cnva_pca}
\title{run pca on the cnv-matrix}
\usage{
hlpr_run_cnva_pca(object, n_pcs = 30, of_sample = NA, ...)
}
\arguments{
\item{object}{spata-object}

\item{n_pcs}{number of pcs to be calculated}

\item{of_sample}{sample of interest}

\item{...}{arguments given to the pca algorithm}
}
\description{
run pca on the cnv-matrix
}
