% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_subset_across}
\alias{hlpr_subset_across}
\title{Subset the across-variables}
\usage{
hlpr_subset_across(data, across, across_subset)
}
\arguments{
\item{data}{A data.frame that contains the variable specified in \code{across}.}

\item{across}{Character value. Denotes the discrete variable in the data.frame
across which something is to be analyzed or displayed.}

\item{across_subset}{Character vector. The groups of interest that the \code{across}-
variable contains.}
}
\value{
A filtered data.frame, informative messages or an error.
}
\description{
Checks across and across_subset input and if at least one
of the across_subset values exists filters the data accordingly.
}
