% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_summarize_trajectory_df}
\alias{hlpr_summarize_trajectory_df}
\title{Join and summarize compiled trajectory data.frames}
\usage{
hlpr_summarize_trajectory_df(
  object,
  ctdf,
  binwidth = 5,
  variables,
  method_gs = "mean",
  verbose = TRUE,
  normalize = FALSE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{ctdf}{A compiled trajectory data.frame containing the variables
\emph{'barcodes', 'sample', 'x', 'y', 'projection_length', 'trajectory_part'}.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{variables}{Character vector. The variables of interest:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
 }}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}
}
\description{
Joins a compiled trajectory data.frame with
the desired information and summarizes those.
}
\details{
Initially the compiled trajectory data.frame of the specified trajectory
is joined with the respective input of variables via \code{joinWithVariables()}.

The argument \code{binwidth} refers to the amount of which the barcode-spots of the
given trajectory will be summarized with regards to the trajectory's direction:
The amount of \code{binwidth} and the previously specified 'trajectory width' in \code{createTrajectories()}
determine the length and width of the sub-rectangles in which the rectangle the
trajectory embraces are splitted and in which all barcode-spots are binned.
Via \code{dplyr::summarize()} the variable-means of every sub-rectangle are calculated.
These mean-values are then arranged according to the trajectory's direction.

Eventually the data.frame is shifted via \code{tidyr::pivot_longer()} to a data.frame in which
every observation refers to the mean-value of one of the specified variable-elements (e.g. a specified
gene set) of the particular sub-rectangle. The returned data.frame contains the following variables:

\itemize{
 \item{\emph{trajectory_part}: Character. Specifies the trajectory's sub-part of the observation. (Negligible if there is
 only one trajectory part.)}
 \item{\emph{trajectory_part_order}: Numeric. Indicates the order within the trajectory-part. (Negligible if there is
 only one trajectory part.)}
 \item{\emph{trajectory_order}: Numeric. Indicates the order within the whole trajectory.}
 \item{\emph{gene_sets, genes or features}: Character. The respective gene sets, gene or feature the value refers to.}
 \item{\emph{values}: Numeric. The actual summarized values.}}
}
