% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_widen_trajectory_df}
\alias{hlpr_widen_trajectory_df}
\title{Widen trajectory data.frame}
\usage{
hlpr_widen_trajectory_df(stdf, variable)
}
\arguments{
\item{stdf}{A summarized trajectory data.frame - output
from \code{hlpr_summarize_trajectory_df()}.}

\item{variable}{Character value. The variable of \code{stdf}:
\emph{'gene_sets', 'genes'} or \emph{'features'}.}
}
\value{
A widened data.frame in which each every observation is a
trajectory and every variable describes the value of the trajectory
at a specific position.
}
\description{
Widen trajectory data.frame
}
