% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-main-functions.R
\name{hspaClusterGenePatterns}
\alias{hspaClusterGenePatterns}
\title{Construct hcluster-tree from correlated gene pattern similarities}
\usage{
hspaClusterGenePatterns(
  object,
  threshold_sim = 0.25,
  correlated = TRUE,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  force = FALSE,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{threshold_sim}{Numeric value. Makes sure that only gene-patterns
are kept that are at least to one other pattern equal or more similar
than the threshold determines. 0 effectively skips filtering.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\description{
Construct hcluster-tree from correlated gene pattern similarities
}
