% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-main-functions.R
\name{hspaDataBinarization}
\alias{hspaDataBinarization}
\title{Title}
\usage{
hspaDataBinarization(
  object,
  genes_subset = NULL,
  mtr_name = "scaled",
  method_binarization = "kmeans",
  method_kmeans = "Hartigan-Wong",
  kmeans = list(),
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
