% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-main-functions.R
\name{hspaPatternIdentification}
\alias{hspaPatternIdentification}
\title{Title}
\usage{
hspaPatternIdentification(object, verbose = NULL, of_sample = NA)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
