% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-sub-functions.R
\name{identify_gene_patterns_dbscan}
\alias{identify_gene_patterns_dbscan}
\title{Title}
\usage{
identify_gene_patterns_dbscan(
  marked_df,
  verbose = TRUE,
  pb = NULL,
  object = NULL
)
}
\arguments{
\item{pb}{}
}
\value{

}
\description{
Title
}
