% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{identifyPixelContent}
\alias{identifyPixelContent}
\alias{identifyPixelContent,HistoImaging-method}
\alias{identifyPixelContent,HistoImage-method}
\alias{identifyPixelContent,Image-method}
\title{Identify pixel content}
\usage{
identifyPixelContent(object, ...)

\S4method{identifyPixelContent}{HistoImaging}(
  object,
  percentile = 99,
  compactness_factor = 10,
  superpixel = 1000,
  dbscan_eps = 0.005,
  dbscan_minPts = 3,
  frgmt_threshold = c(0.005, 0.02),
  plot_progress = plot_progress,
  verbose = TRUE
)

\S4method{identifyPixelContent}{HistoImage}(
  object,
  percentile = 99,
  compactness_factor = 10,
  superpixel = 1000,
  dbscan_eps = 0.005,
  dbscan_minPts = 3,
  frgmt_threshold = c(0.05, 0.02),
  plot_progress = FALSE,
  verbose = TRUE
)

\S4method{identifyPixelContent}{Image}(
  object,
  percentile = 99,
  compactness_factor = 10,
  superpixel = 1000,
  frgmt_threshold = c(5e-04, 0.01),
  dbscan_eps = 0.005,
  dbscan_minPts = 3,
  plot_progress = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The method for class \code{Image} returns a data.frame of the following
variables.
\itemize{
\item{\emph{pixel}:}{ character. Pixel index.}
\item{\emph{width}:}{ numeric. Pixel position on horizontal axis of the image.}
\item{\emph{height}:}{ numeric. Pixel position on the vertical axis of the image.}
\item{\emph{clusterK2}:}{ character. Either \emph{'background'} or \emph{'tissue'}.}
\item{\emph{colTiss#} :}{ numeric. Numeric variables that correspond to the color dimensions
of the image mask based on which the clustering of \emph{clusterK2} was conducted.}
\item{\emph{clusterDBSCAN}:}{ character. Cluster results of dbscan::dbscan() after removal
of background pixels.}
\item{\emph{clusterDBSCAN_size}:}{numeric. Size of each dbscan cluster.}
\item{\emph{content}:}{ character. The identified content of each pixel.}
}

Methods for S4-classes serving as containers the input object is returned with
the results stored in the corresponding slots.
}
\description{
Determines the type of content displayed by each pixel in the image,
categorizing it as tissue from tissue segments or fragments, artifacts, or background.
}
\seealso{
\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}
}
