% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifySpatialOutliers}
\alias{identifySpatialOutliers}
\title{Identify spatial outliers}
\usage{
identifySpatialOutliers(
  object,
  method = "obs",
  img_name = activeImage(object),
  buffer = 0,
  eps = recDbscanEps(object),
  minPts = recDbscanMinPts(object),
  min_section = nBarcodes(object) * 0.05,
  test = "any",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{method}{Character vector. The method(s) to use. A combination of \emph{'obs'}
and/or \emph{'image'}. Defaults to \emph{'obs'}. See details for more.}

\item{img_name}{Character value. The name of the image whose tissue outline
is used if \code{method} contains \emph{'image'}.}

\item{buffer}{Numeric value. Expands the tissue outline to include observations
that lie on the edge of the outline and are mistakenly removed.}

\item{eps, minPts}{Given to the corresponding arguments of
\code{\link[dbscan:dbscan]{dbscan::dbscan()}} if \code{method} contains \emph{'obs'}.}

\item{min_section}{Numeric value. The minimum number of observations a spatial cluster must
contain such that the whole cluster is identified as a contiguous tissue section.}

\item{test}{Character value. Only required if \code{method = c('obs', 'image')}. If
\emph{'any'}, spots are labeled as outliers if at least one method identifies them
as outliers. If \emph{'all'}, spots are labeled as outliers if both methods identify
them as outliers.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Labels data points as spatial outliers depending on certain
criteria with a new meta variable called \emph{sp_outlier}. See details for more.

Requires the results of \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}.
}
\details{
In case of \code{method = 'obs'}, the results from \code{identifyTissueOutline(object, method = 'obs')}
are directly transferred to the new meta feature \emph{sp_outlier}, which declares
whether an observation is a spatial outlier or not.

In case of \code{method = 'image'}, the image based tissue outline from the
\code{identifyTissueOutline()} function is used. This function has created polygons that
outline the tissue or tissue sections identified in the image. For each data point,
the function checks which polygon it falls within and assigns it to the corresponding
group. If an observation does not fall within any of the tissue polygons, it is
considered a spatial outlier.

If \code{method = c('obs', 'image')}, the results of both methods are used Whether a
data point is considered a spatial outlier depends on the \code{test} argument:

\itemize{
\item{\code{test = 'any'}:} The data point is considered a spatial outlier if
either of the two tests classifies it as an outlier.
\item{\code{test = 'all'}:} The data point is considered a spatial outlier
only if both tests classify it as an outlier.
}

The results can be visualized using \code{plotSurface(object, color_by = "sp_outlier")}.
In case of bad results the function can be run over and over again with
changing parameters as the results are simply overwritten.
}
\seealso{
\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}, \code{\link[=mergeTissueSections]{mergeTissueSections()}}
}
