% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyTissueOutline}
\alias{identifyTissueOutline}
\title{Identify tissue outline}
\usage{
identifyTissueOutline(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
An updated \code{spata2} object. The coordinates data.frame
as obtained by \code{getCoordsDf()} contains an additional, logical
variable named \emph{outline} indicating whether the spot belongs
to the outline spots of the respective tissue section indicated by
variable \emph{section}.
}
\description{
Identifies the barcode-spots that lie on the edge
of each tissue section and, thus, outline it. Requires \code{identifyTissueSections()}
results.
}
