% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{identifyTissueSections}
\alias{identifyTissueSections}
\title{Deprecated}
\usage{
identifyTissueSections(object, ...)
}
\description{
Deprecated in favor of \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}.
}
\keyword{internal}
