% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyVariableMolecules}
\alias{identifyVariableMolecules}
\title{Identify variable molecules}
\usage{
identifyVariableMolecules(
  object,
  n_mol = 2500,
  method = "vst",
  assay_name = activeAssay(object),
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{n_mol}{Numeric value. The number of molecules to be identified. Given
to \code{nfeatures} of \code{\link[Seurat:FindVariableFeatures]{Seurat::FindVariableFeatures()}}.}

\item{method}{Character value. The identification method. One of \emph{'vst'},
\emph{'mean.var.plot'} or \emph{'dispersion'}. Given to \code{selection.method} of
\code{\link[Seurat:FindVariableFeatures]{Seurat::FindVariableFeatures()}}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Identfies molecules that are outliers on a mean-variability plot as proposed by
\code{Seurat}.
}
\examples{

library(SPATA2)
data("example_data")

object <- loadExampleObject("UKF269T")

object <- identifyVariableMolecules(object, n_mol = 2000)
object <- identifyVariableMolecules(object, n_mol = 3000, method = "mean.var.plot")

vars_vst <- getVariableMolecules(object, method = "vst")
vars_mvp <- getVariableMolecules(object, method = "mean.var.plot")

length(vars_vst)
length(vars_mvp)

}
