% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{idSA}
\alias{idSA}
\alias{idST}
\title{Quick access to IDs}
\usage{
idSA(object)

idST(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character value.
}
\description{
Handy functions to access the ID of a spatial annotation
or a spatial trajectory if there exist only one of each in the object. Mostly
used to define the default of dependent functions. Return an error if there
are no or more than one IDs found.
}
