% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPositions.R
\name{ignoreAlternative}
\alias{ignoreAlternative}
\title{Remove alternative chromosomes, X chromosome, Y chromosome, and mitochondrial genome from a gene positions dataframe}
\usage{
ignoreAlternative(gene_positions_df)
}
\arguments{
\item{gene_positions_df}{A data frame containing gene positions.}
}
\value{
A modified gene positions dataframe with alternative chromosomes, X chromosome, Y chromosome, and mitochondrial chromosome removed, sorted in numeric order.
}
\description{
This function removes alternative chromosomes, X chromosome, Y chromosome, and mitochondrial genome from a gene positions dataframe.
It also removes any duplicated genes, sorts the dataframe by chromosome column in numeric order, and returns the modified dataframe.
}
\examples{
gene_positions <- data.frame(
  ensembl_gene_id = c("ENSG00000261846", "ENSG00000197953", "ENSG00000262466"),
  hgnc_symbol = c("AADACL2", "AADACL2", "AADACL2-AS1"),
  chromosome_name = c("CHR_HSCHR3_1_CTG2_1", "3", "CHR_HSCHR3_1_CTG2_1"),
  start_position = c(151744454, 151733916, 151761981),
  end_position = c(151770036, 151761339, 151765669)
)
ignoreAlternative(gene_positions)

}
