% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-generic-functions.R
\name{image}
\alias{image}
\alias{exprMtr}
\alias{countMtr}
\alias{coordinates}
\alias{coordinates<-}
\alias{featureData}
\alias{featureData<-}
\alias{samples}
\alias{trajectory}
\alias{ctdf}
\alias{exprMtr,spata-method}
\alias{countMtr,spata-method}
\alias{coordinates,spata-method}
\alias{coordinates<-,spata-method}
\alias{featureData,spata-method}
\alias{featureData<-,spata-method}
\alias{samples,spata-method}
\alias{trajectory,spata-method}
\alias{ctdf,spatial_trajectory-method}
\title{Generics to extract a slots content}
\usage{
image(object, of_sample = "")

exprMtr(object, of_sample = "")

countMtr(object, of_sample = "")

coordinates(object, of_sample = "")

coordinates(object) <- value

featureData(object, of_sample = "")

featureData(object) <- value

samples(object)

trajectory(object, trajectory_name, of_sample = "")

ctdf(t_obj)

\S4method{exprMtr}{spata}(object, of_sample = "")

\S4method{countMtr}{spata}(object, of_sample = "")

\S4method{coordinates}{spata}(object, of_sample = "")

\S4method{coordinates}{spata}(object) <- value

\S4method{featureData}{spata}(object, of_sample = "")

\S4method{featureData}{spata}(object) <- value

\S4method{samples}{spata}(object)

\S4method{trajectory}{spata}(object, trajectory_name, of_sample = "")

\S4method{ctdf}{spatial_trajectory}(t_obj)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample from which to extract the content.}
}
\value{
The respective slots content.
}
\description{
Generics to extract a slots content
}
