% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{image_annotations}
\alias{image_annotations}
\title{A list of lists of image annotations. Every slot in list \code{image_annotations}
is named by the sample the image annotations belong to.}
\format{
An object of class \code{list} of length 3.
}
\usage{
image_annotations
}
\description{
@docType data
 @usage data(image_annotations)
}
\keyword{datasets}
