% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{image_dummy}
\alias{image_dummy}
\title{image_dummy}
\usage{
image_dummy(image)
}
\arguments{
\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}
}
\description{
image_dummy
}
