% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{ImageAnnotation-class}
\alias{ImageAnnotation-class}
\alias{ImageAnnotation}
\title{The \code{ImageAnnotation} - Class}
\description{
An S4 class designed to capture spatial annotations by outlining areas
of interest on images. This class provides a flexible framework for creating annotations
that visually highlight specific regions within images, such as histological structures,
cellular patterns, or other histomorphological features in images from spatial multi-omic
studies.
}
\details{
This class is an extension of class \code{\link{SpatialAnnotation}} and
inherits all of its slots.
}
\section{Slots}{

\describe{
\item{\code{parent_name}}{Character string. The name of the image this annotation
was drawn on.}
}}

