% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{ImageAnnotationScreening-class}
\alias{ImageAnnotationScreening-class}
\alias{ImageAnnotationScreening}
\title{The \code{ImageAnnotationScreening} - Class}
\description{
S4 class that contains input for and output of the
function \code{imageAnnotationScreening()}.
}
\section{Slots}{

\describe{
\item{\code{angle_span}}{numeric. Vector of length two. Confines the area of interest
by angle relative to the center of the image annotation.}

\item{\code{binwidth}}{numeric. The value with which the polygon that encircles
the image annotation is consecutively expanded via \code{sf::st_buffer()},}

\item{\code{coords}}{data.frame. Coordinates data.frame of the sample.}

\item{\code{dea}}{data.frame. Results of the DE-Analysis between the image annotation
area and the rest of the sample.}

\item{\code{img_annotation}}{ImageAnnotation. The \code{ImageAnnotation}-object of
the image annotation chosen for the screening.}

\item{\code{method_padj}}{character. The method with which p-values were adjusted.}

\item{\code{n_bins_angle}}{numeric. Number of bins that were created anglewise.}

\item{\code{n_bins_circle}}{numeric. Number of bins that were created circlewise.}

\item{\code{results}}{data.frame. Data.frame that contains the results of
the model fitting per angle bin.}

\item{\code{results_smrd}}{data.frame. Data.frame that contains the summary of
all gene-model fits across all angle bins.}

\item{\code{sample}}{character. The sample name.}
}}

