% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{imageAnnotationScreening}
\alias{imageAnnotationScreening}
\title{Implementation of the IAS-algorithm}
\usage{
imageAnnotationScreening(
  object,
  id,
  variables,
  n_bins_circle,
  buffer,
  angle_span = c(0, 360),
  n_bins_angle = 12,
  summarize_with = "mean",
  method_padj = "fdr",
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  with_de = TRUE,
  method_de = "wilcox",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{buffer}.
 (See details for more.)}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}. (See details for more.)}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the image annotation.
 (See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}.
Specifies the function with which the bins are summarized.}

\item{method_padj}{Character value. The method with which adjusted p-values are
calculated. Use \code{validPadjMethods()} to obtain all valid input options.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{ImageAnnotationScreening}. See documentation
with \code{?ImageAnnotationScreening} for more information.
}
\description{
Screens the sample for numeric variables that stand
in meaningful, spatial relation to annotated structures/areas of
the sample. For a detailed explanation on how to define the parameters
\code{n_bins_circle}, \code{buffer}, \code{angle_span} and \code{n_bins_angle}
see details section.
}
\details{
In conjunction with argument \code{id} which provides the
ID of the image annotation of interest the arguments \code{n_bins_circle},
\code{buffer}, \code{angle_span} and \code{n_bins_angle} can be used
to specify the exact area that is screened as well as the resolution of the screening.

Note that the function \code{plotSurfaceIas()} allows to visually check
if your input results in the desired screening.
}
