% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{imageAnnotationToSegmentation}
\alias{imageAnnotationToSegmentation}
\title{Convert spatial annotation to segmentation}
\usage{
imageAnnotationToSegmentation(
  object,
  ids,
  segmentation_name,
  inside = "inside",
  outside = "outside",
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector. Specifies the spatial annotation(s) of interest.
data points that fall into the area of these annotations are labeled
with the input for argument \code{inside}.}

\item{segmentation_name}{Character value. The name of the new segmentation variable.}

\item{inside}{Character value. The group name for the data points that
are located inside the area of the spatial annotation(s).}

\item{outside}{Character value. The group name for the data points that
are located outside the area of the spatial annotation(s).}

\item{overwrite}{Logical. Set to TRUE to overwrite existing variables with
the same name.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Converts one or more spatial annotations to a binary
segmentation variable in the feature data.frame.
}
