% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{include_tissue_outline}
\alias{include_tissue_outline}
\title{Include spatial extent of tissue sections in analysis}
\usage{
include_tissue_outline(
  input_df,
  outline_df = NULL,
  coords_df = NULL,
  spat_ann_center = NULL,
  sas_circles = FALSE,
  ccd = NULL,
  outside_rm = TRUE,
  inside_if = c(1, 2),
  opt = "concaveman",
  buffer = 0,
  ...
)
}
\arguments{
\item{input_df}{A data.frame that contains at least numeric \emph{x} and \emph{y}
variables.}

\item{outline_df}{A data.frame that contains the ouline/hull of all tissue sections.
Must contain variables \emph{x}, \emph{y} and \emph{section}.}

\item{sas_circles}{Logical value. If \code{TRUE}, input data.frame is assumed
to contain polygon coordinates of the expanded spatial annotation encircling
and sorts them after filtering for those that lie inside the tissue section
in order to plot them via \code{ggplot2::geom_path()}.}

\item{opt}{Either \emph{'concaveman'}' or \emph{'chull'}. Defines with which function
the tissue outline is computed.}

\item{buffer}{Distance measure with which to buffer the tissue outline data.frame.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Filtered input data.frame.
}
\description{
Ensures section specific processing of observations
in relation by identifying the outline of the tissue section
(or -sections in case of multiple tissue sections per sample). Additionally,
allows to relate observations to the spatial position and extent of image
annotations.
}
\keyword{internal}
