% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{increase_n_data_points}
\alias{increase_n_data_points}
\title{Increase the Number of Data Points in a Coordinate Data Frame}
\usage{
increase_n_data_points(coords_df, fct = 10, cvars = c("x", "y"))
}
\arguments{
\item{coords_df}{A data frame containing the original coordinates.
Must include columns specified in \code{cvars}.}

\item{fct}{A numeric factor indicating how many times the number of data points should be increased.
The function will round this number up to the nearest integer.}

\item{cvars}{A character vector specifying the column names of the coordinates in \code{coords_df}.}
}
\value{
A data frame with the increased number of data points.
}
\description{
This function artificially increases the number of data points in a given coordinate data frame.
It generates additional points by adding random noise to the existing coordinates,
ensuring the new points stay within the original range of the data.
}
