% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{increase_polygon_vertices}
\alias{increase_polygon_vertices}
\title{Increase the Number of Vertices in a Polygon}
\usage{
increase_polygon_vertices(polygon_df, avg_dist, skip = FALSE)
}
\arguments{
\item{polygon_df}{A data frame representing the polygon, with columns for x and y coordinates.}

\item{avg_dist}{The average distance between vertices after interpolation.}

\item{skip}{Logical; if TRUE, the function skips the interpolation process and returns the original polygon.}
}
\value{
A data frame representing the polygon with increased vertex density.
}
\description{
This function interpolates additional vertices in a polygon data frame to increase its vertex density.
It ensures that the polygon remains closed and that the new vertices are evenly distributed.
}
\keyword{internal}
