% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObject_10X}
\alias{initiateSpataObject_10X}
\title{Initiate a spata-object from 10X Visium}
\usage{
initiateSpataObject_10X(
  directory_10X,
  sample_name,
  image_name = "tissue_lowres_image.png",
  directory_spata = NULL,
  directory_seurat = NULL,
  add_wd = "/",
  gene_set_path = NULL,
  SCTransform = FALSE,
  NormalizeData = list(normalization.method = "LogNormalize", scale.factor = 1000),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  ScaleData = TRUE,
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  RunTSNE = TRUE,
  RunUMAP = list(dims = 1:30),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{directory_10X}{Character value. Specifies the 10X visium-folder from
which to load the information. This folder must contain the following sub directories:

\itemize{
\item{\emph{'/filtered_feature_bc_matrix.h5'}}
\item{\emph{'/spatial/*.jpg}}
}

(It is no longer required that the folder contains an \emph{/outs/} sub directory.)}

\item{sample_name}{Character value. The sample name with which to refer to the
respective sample. Should start with a letter.}

\item{image_name}{Character value. The filename of the image that is read in
as the default image. Should be in subdirectory \emph{directory_10X/spatial/}.}

\item{gene_set_path}{Character value (or NULL). Specifies the path to a
.RDS-file containing a data.frame that is to be used as input for slot @used_genesets.

Must have the character-variables

\itemize{
\item{\emph{'ont'}: The gene set names.}
\item{\emph{'gene'}: The belonging gene names.}
}

If set to NULL the default gene-set data.frame will used. Run \code{?gsdf} to get more information.}

\item{SCTransform}{A named list of arguments given to \code{Seurat::SCTransform()}, TRUE or FALSE.}

\item{NormalizeData}{A named list of arguments given to \code{Seurat::NormalizeData()}, TRUE or FALSE.}

\item{FindVariableFeatures}{A named list of arguments given to \code{Seurat::FindVariableFeatures()}, TRUE or FALSE.}

\item{ScaleData}{A named list of arguments given to \code{Seurat::ScaleData()}, TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument \code{features} input
for argument \code{features} is set to \code{base::rownames(seurat_object)}.}

\item{RunPCA}{A named list of arguments given to \code{Seurat::RunPCA()}, TRUE or FALSE.}

\item{FindNeighbors}{A named list of arguments given to \code{Seurat::FindNeighbors()}, TRUE or FALSE.}

\item{FindClusters}{A named list of arguments given to \code{Seurat::FindClusters()}, TRUE or FALSE.}

\item{RunTSNE}{A named list of arguments given to \code{Seurat::RunTSNE()}, TRUE or FALSE.}

\item{RunUMAP}{A named list of arguments given to \code{Seurat::RunUMAP()}, TRUE or FALSE.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A spata-object.
}
\description{
Creates, saves and returns an object of class spata
from 10X Visium results. See details for more information.
}
\details{
The loading and preprocessing of the spata-object  currently relies on the Seurat-package. Before any pre processing function is applied
mitochondrial and stress genes are discarded. For more advanced users the arguments above starting with a capital letter allow to
manipulate the way the spata-object is processed. For all of these arguments apply the following instructions:

\itemize{
\item{If set to FALSE the processing function is skipped.}
\item{If set to TRUE the respective function is called with it's default argument settings. Note: \code{RunUMAP()} needs
additional input!}
\item{If a named list is provided the respective function is called whereby the named list will provide the argument-input (the seurat-object to be constructed must not be named and will be
passed to every function as the first argument!!!.)}
}

Note that certain listed functions require previous functions! E.g. if \code{RunPCA} is set to FALSE \code{RunTSNE()}
will result in an error. (\code{base::tryCatch()} will prevent the function from crashing but the respective slot
is going to be empty.) Skipping functions might result in an incomplete spata-object. Use \code{validateSpataObject()} after
initiating it in order to see which slots are valid and which are not.
}
