% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate-family.R
\name{initiateSpataObject_MALDI}
\alias{initiateSpataObject_MALDI}
\title{Initiate a spata-object from MALDI Experiments}
\usage{
initiateSpataObject_MALDI(
  coords_df,
  intensity_mtr,
  sample_name,
  gene_set_path = NULL,
  output_path = NULL,
  file_name = NULL,
  pca_comp = 30,
  nn = 50,
  tsne_perplexity = 30,
  verbose = TRUE
)
}
\arguments{
\item{coords_df}{Data.frame containing information about the positions of all
barcode-spots in form of a numeric \emph{x}- and \emph{y}-variable. The key-variable
\emph{barcodes} needs to be of type character and must be identical to the column names
of the input matrix (\code{expr_mtr}).}

\item{intensity_mtr}{A numeric matrix to be used as the expression matrix. Rownames must
correspond to the genes and column names must correspond to the barcodes.}

\item{sample_name}{Character value. The future input for SPATA's \code{of_sample}-argument.}

\item{gene_set_path}{Character value (or NULL). Specifies the path to a
.RDS-file containing a data.frame that is to be used as input for slot @used_genesets.

 Must have the character-variables

   \itemize{
    \item{\emph{'ont'}: The gene set names.}
    \item{\emph{'gene'}: The belonging gene names.}
    }

If set to NULL the default gene-set data.frame will used. Run \code{?gsdf} to get more information.}

\item{nn}{Numeric value. Given to argument \code{k} of function \code{RANN::nn2()}: Determines to maximum number
of nearest neighbours to compute. (\code{nn} is deprecated.)}
}
\value{
A spata-object.
}
\description{
Initiate a spata-object from MALDI Experiments
}
