% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{initiateSpataObject_Visium}
\alias{initiateSpataObject_Visium}
\title{Initiate \code{spata2} object from platform}
\usage{
initiateSpataObject_Visium(
  directory_visium,
  sample_name,
  mtr = "filtered",
  img_active = "lowres",
  img_ref = "hires",
  directory_spata = NULL,
  gene_set_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{directory_visium}{Character value. Directory to a visium folder. Should contain
the subdirectory \emph{'.../spatial'}.}

\item{sample_name}{Character value. Name of the sample.}

\item{mtr}{The matrix to load. One of \code{c("filtered", "raw")}.}

\item{img_ref, img_active}{Character values specifying the active and reference images. Choose between \emph{'lowres'} and \emph{'hires'}.
Setting both \code{img_ref} and \code{img_active} to the same value indicates that only one image is used,
and the other image will not be read automatically. In such cases, manual registration of the
unselected image is required using the \code{registerHistoImage()} function.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{spata2}.
}
\description{
A colleciton of functions that initiate \code{spata2} objects
from standardized output folders.
}
\seealso{
\code{\link{createHistoImagingVisium}}
}
