% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObjectMERFISH}
\alias{initiateSpataObjectMERFISH}
\title{Initiate a \code{spata2} object from platform MERFISH}
\usage{
initiateSpataObjectMERFISH(
  directory_merfish,
  sample_name,
  file_counts = NULL,
  file_cell_meta = NULL,
  verbose = TRUE
)
}
\arguments{
\item{directory_merfish}{Character value. Directory to a MERFISH folder
that should contain a .csv file called \emph{cell_by_gene.csv} and a .csv file
called \emph{cell_metadata.csv}. Deviating filenames can be specified using
arguments \code{file_counts} and \code{file_cell_meta}, respectively.}

\item{file_counts}{Character value or \code{NULL}. If character, specifies
the filename of .csv file that contains the gene counts by cell. Use only
if filename deviates from the default.}

\item{file_cell_meta}{Character value or \code{NULL}. If character, specifies
the filename of the .csv file that contains cell meta data, in particular,
spatial location via the variables \emph{center_x} and \emph{center_y}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{spata2}.
}
\description{
Wrapper function around the necessary content to create a
\code{spata2} object from the standardized output of the MERFISH platform.
}
\details{
MERFISH output does not come with an image. However, many spatial information such
as coordinates, coordinate scale factors or spatial annotations are stored
in class \code{\link{HistoImaging}} and \code{\link{HistoImage}}. The \code{spata2} object is equipped
with a \code{HistoImaging} object that contains an empty \code{HistoImage} called \emph{pseudo}.

MERFISH works in micron space. The coordinates of the cellular centroids are
provided in unit um. Therefore no pixel scale factor must be computed or set
to work with SI units.
}
