% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObjectSlideSeqV1}
\alias{initiateSpataObjectSlideSeqV1}
\title{Initiate an object of class \code{SPATA2} from platform SlideSeq}
\usage{
initiateSpataObjectSlideSeqV1(
  sample_name,
  directory_slide_seq,
  file_counts = NULL,
  file_barcodes = NULL,
  file_genes = NULL,
  file_coords = NULL,
  verbose = TRUE
)
}
\arguments{
\item{sample_name}{Character value. The name with which to identify the \code{SPATA2} object.}

\item{directory_slide_seq}{Character value. Directory to a SlideSeq folder
that contains a count matrix and bead locations.}

\item{file_counts}{Character value or \code{NULL}. If character, specifies
the filename of the count matrix. If \code{NULL}, the SlideSeq folder is skimmed
for a file ending with \emph{.mtx}.}

\item{file_barcodes}{Character value or \code{NULL}. If character, specifies
the filename of the barcode names for the count matrix if it does not
contain column names. If \code{NULL}, the SlideSeq folder is skimmed
for a file ending with \emph{barcodes.tsv}.}

\item{file_genes}{Character value or \code{NULL}. If character, specifies
the filename of the gene names for the count matrix if it does not
contain row names. If \code{NULL}, the SlideSeq folder is skimmed
for a file ending with \emph{genes.tsv}.}

\item{file_coords}{Character value or \code{NULL}. If character, specifies
the filename of the coordinates. If \code{NULL}, the SlideSeq folder is skimmed
for a file ending with \emph{MatchedBeadLocation.csv}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{SPATA2}.
}
\description{
Wrapper function around the necessary content to create a
\code{SPATA2} object from the standardized output of the SlideSeq platform.
}
