% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObjectVisium}
\alias{initiateSpataObjectVisium}
\title{Initiate an object of class \code{SPATA2} from platform Visium}
\usage{
initiateSpataObjectVisium(
  sample_name,
  directory_visium,
  mtr = "filtered",
  img_active = "lowres",
  img_ref = "lowres",
  verbose = TRUE
)
}
\arguments{
\item{sample_name}{Character value. The name with which to identify the \code{SPATA2} object.}

\item{directory_visium}{Character value. Directory to a visium folder. Should contain
the subdirectory \emph{'.../spatial'}.}

\item{mtr}{The matrix to load. One of \code{c("filtered", "raw")}.}

\item{img_ref, img_active}{Character values specifying which of the images to register and how to register
them. Click \code{\link[=concept_images]{here}} for more information about the definitions
of the reference image and the active image. Setting both arguments to the same
value results in the function to register the specified image as the active
as well as the reference image. Valid input options depend
on the platform used:

\itemize{
\item{\emph{Visium}:}{ Either \emph{'lowres'} or \emph{'hires'}.}
}}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{SPATA2}.
}
\description{
Wrapper function around the necessary content to create a
\code{SPATA2} object from standardized output of the Visium platform. See details
section for more.
}
\details{
This function expects \code{directory_visium} to lead to a folder in which the
following subdirectories exist:

\itemize{
\item{\emph{spatial/}}{: The folder in which image and coordinates are located.}
\item{\emph{filtered_feature_bc_matrix.h5}}{: The filtered count matrix.}
\item{\emph{raw_feature_bc_matrix.h5}}{: The filtered count matrix.}
}

Depending on the input for \code{mtr} only the requested file has to exist. Given
the content and subsequent filenames, the function differentiates between

\itemize{
\item{VisiumSmall}{ Visium data set with capture area of 6.5mm x 6.5mm.}
\item{VisiumLarge}{ Visium data set with capture area of 11mm x 11m. }
}

In any case, the output is an object of class \code{SPATA2}.
}
\seealso{
\code{\link[=createSpatialDataVisium]{createSpatialDataVisium()}}
}
