% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObjectVisiumHD}
\alias{initiateSpataObjectVisiumHD}
\title{Initiate an object of class \code{SPATA2} from platform VisiumHD}
\usage{
initiateSpataObjectVisiumHD(
  sample_name,
  directory_visium,
  square_res = "16um",
  mtr = "filtered",
  img_active = "lowres",
  img_ref = "lowres",
  verbose = TRUE
)
}
\arguments{
\item{sample_name}{Character. The name of the sample.}

\item{directory_visium}{Character. The directory containing the Visium output files.}

\item{square_res}{Character. The square resolution from which to load the data. One of *c('16um', '8um', '2um').
See details for more.}

\item{mtr}{Character. Specifies which matrix to use, either "filtered" or "raw". Default is "filtered".}

\item{img_active}{Character. The active image to use, either "lowres" or "hires". Default is "lowres".}

\item{img_ref}{Character. The reference image to use, either "lowres" or "hires". Default is "lowres".}

\item{verbose}{Logical. If TRUE, progress messages are printed. Default is TRUE.}
}
\value{
A \code{SPATA2} object the VisiumHD platform.
}
\description{
This function initiates a \code{\link{SPATA2}} object with data generated
using the 10x Genomics \code{\link{VisiumHD}} platform.
}
\details{
The function requires a directory containing the output files from a 10x Genomics VisiumHD experiment
specified with the argument \code{directory_visium}. This directory (below denoted as \strong{~})  must include the following
sub-directories:

\itemize{
\item \emph{~/binned_outputs}: A folder with the following subdirectories.
\itemize{
\item \emph{~/binned_outputs/square_002um}: The folder containing the data for \code{square_res = '2um'}.
\item \emph{~/binned_outputs/square_008um}: The folder containing the data for \code{square_res = '8um'}.
\item \emph{~/binned_outputs/square_016um}: The folder containing the data for \code{square_res = '16um'}.
}
}

Depending on your input for \code{square_res} only the corresponding subfolder is required. This subfolder should
contain the following files and sub-directories:

\itemize{
\item \emph{~/binned_outputs/<square_res>/filtered_feature_bc_matrix.h5} or \emph{~/raw_feature_bc_matrix.h5}: The HDF5 file containing the filtered or raw feature-barcode matrix, respectively.
\item \emph{~/binned_outputs/<square_res>/spatial/tissue_lowres_image.png} or \emph{~/spatial/tissue_hires_image.png}: The low-resolution or high-resolution tissue image.
\item \emph{~/binned_outputs/<square_res>/spatial/scalefactors_json.json}: A JSON file containing the scale factors for the images.
\item \emph{~/binned_outputs/<square_res>/spatial/tissue_position.parquet} A .parguet file containing the tissue positions and spatial coordinates.
}

The function will check for these files and process them to create a \code{SPATA2} object. It reads the count matrix, loads the spatial data,
and initializes the \code{SPATA2} object with the necessary metadata and settings.
}
\note{
It is crucial to install the package \code{arrow} in a way that \code{\link[arrow:read_parquet]{arrow::read_parquet()}} works. There
are several ways. Installing the package with \code{install.packages('arrow', repos = 'https://apache.r-universe.dev')}
worked reliably for us.
}
