% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObjectXenium}
\alias{initiateSpataObjectXenium}
\title{Initiate an object of class \code{SPATA2} from platform Xenium}
\usage{
initiateSpataObjectXenium(sample_name, directory_xenium, verbose = TRUE)
}
\arguments{
\item{directory_xenium}{Character value. Directory to a Xenium output folder.}
}
\description{
Wrapper function around the necessary content to create a
\code{SPATA2} object from standardized output of the Xenium platform.
}
\details{
The function requires a directory containing the output files from a 10x Genomics Xenium experiment
specified with the argument \code{directory_xenium}. This directory (below denoted as \strong{~}) must include the following
sub-directories:

\itemize{
\item \emph{~/cell_feature_matrix}: A folder that contains the following files:
\itemize{
\item \emph{~/cell_feature_matrix/barcodes.tsv.gz}: The cell barcodes.
\item \emph{~/cell_feature_matrix/features.tsv.gz}: The feature (gene) names.
\item \emph{~/cell_feature_matrix/matrix.mtx.gz}: The raw count data matrix.
}
\item \emph{~/cells.csv.gz}: The cellular positions.
}

MERFISH works in micron space. The coordinates of the cellular centroids are
provided in unit um. Therefore no pixel scale factor must be computed or set
to work with SI units.
}
