% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{interpolate_points_along_path}
\alias{interpolate_points_along_path}
\title{Interpolate points along path}
\usage{
interpolate_points_along_path(data, max_distance = 1)
}
\arguments{
\item{data}{Data.frame of x- and y-coordinates.}
}
\description{
Ensures equally distributed number of points along a
curved trajectory.
}
\keyword{internal}
