% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{intersect_polygons}
\alias{intersect_polygons}
\title{Test polygon intersection}
\usage{
intersect_polygons(a, b, strictly = FALSE)
}
\arguments{
\item{a, b}{Matrix or data.frame with columns \emph{x} and \emph{y}.}

\item{strictly}{Logical value. If \code{TRUE}, only barcode spots that are strictly
interior to the polygon are returned. If \code{FALSE}, barcodes that are
on the relative interior the polygon border or that are vertices themselves
are returned, too.}
}
\value{
Logical vector of the same length as the number of rows in \code{a}.
}
\description{
Tests which vertices of polygon \code{a} lay inside polygon \code{b}.
}
