% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{is_outlier}
\alias{is_outlier}
\title{Check for Outliers in a Numeric Vector}
\usage{
is_outlier(x, ...)
}
\arguments{
\item{x}{A numeric vector for which outliers need to be identified.}

\item{...}{Additional arguments passed to \code{boxplot.stats}.}
}
\value{
A logical vector of the same length as \code{x}, where \code{TRUE} indicates that
the corresponding element in \code{x} is an outlier.
}
\description{
This function identifies outliers in a numeric vector \code{x}.
}
\details{
An element is considered an outlier if it falls outside the range as defined
by \code{boxplot.stats} from the \code{grDevices} package. The function is useful
for quickly flagging or filtering outliers in a dataset.
}
\examples{
# Example vector
data <- c(-10, 1, 2, 3, 4, 5, 100)

# Check for outliers
is_outlier(data)

# Using the function to filter out outliers
data[!is_outlier(data)]

}
