% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-family.R
\name{is_subsetted_by_segment}
\alias{is_subsetted_by_segment}
\title{Check if object has been subsetted by segment}
\usage{
is_subsetted_by_segment(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
TRUE or FALSE
}
\description{
Check if object has been subsetted by segment
}
