% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{is_unit_dist}
\alias{is_unit_dist}
\title{Test unit of length input}
\usage{
is_unit_dist(input, error = FALSE)
}
\arguments{
\item{input}{Character vector of distance units. Obtain valid
input options with \code{validUnitsOfLength()}.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value and/or error if argument \code{error} is \code{TRUE}.
}
\description{
Tests if input is a valid unit of distance.
}
