% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{isDirVisium}
\alias{isDirVisium}
\title{Directory tests}
\usage{
isDirVisium(dir, error = FALSE)
}
\arguments{
\item{dir}{Character value. The directory to check.}
}
\value{
Logical value.
}
\description{
Tests if input directories meet the requirements of specific
functions specifically written for reading data from standardized output
folders.
}
\keyword{internal}
