% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{isSPATA2}
\alias{isSPATA2}
\title{General test for SPATA2 object}
\usage{
isSPATA2(object, warn = TRUE, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{warn, error}{Logical values, allowing the function to throw warnings
or errors. If \code{FALSE}, the default, only a logical value is returned.}
}
\value{
Logical value.
}
\description{
Logical test of an object being interpretable as a \code{\link{SPATA2}} object.
}
