% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j.R
\name{joinWith}
\alias{joinWith}
\alias{joinWithFeatures}
\alias{joinWithGenes}
\alias{joinWithGeneSets}
\title{Join barcodes with additional variables}
\usage{
joinWith(
  object,
  spata_df = getCoordsDf(object),
  features = NULL,
  gene_sets = NULL,
  method_gs = NULL,
  genes = NULL,
  smooth = FALSE,
  smooth_span = NULL,
  verbose = NULL,
  normalize = NULL,
  ...
)

joinWithFeatures(
  object,
  spata_df = getCoordsDf(object),
  features,
  smooth = FALSE,
  smooth_span = 0.02,
  verbose = TRUE,
  ...
)

joinWithGenes(
  object,
  spata_df = getCoordsDf(object),
  genes,
  average_genes = FALSE,
  uniform_genes = "keep",
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = NULL,
  ...
)

joinWithGeneSets(
  object,
  spata_df = getCoordsDf(object),
  gene_sets,
  method_gs = "mean",
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE,
  ignore = T,
  ...
)
}
\description{
These functions have been deprecated in favor of \code{\link[=joinWithVariables]{joinWithVariables()}}.
}
\keyword{internal}
