% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{lastImageAnnotation}
\alias{lastImageAnnotation}
\title{Deprecated}
\usage{
lastImageAnnotation(...)
}
\description{
Deprecated in favor of \code{\link[=lastSpatialAnnotation]{lastSpatialAnnotation()}}.
}
\keyword{internal}
