% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{legendBottom}
\alias{legendBottom}
\alias{legendColor}
\alias{legendLeft}
\alias{legendNone}
\alias{legendRight}
\alias{legendTop}
\title{Ggplot add on wrapper}
\usage{
legendBottom(...)

legendColor(...)

legendLeft(...)

legendNone(...)

legendRight(...)

legendTop(...)
}
\description{
Ggplot add on wrapper
}
