% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{loadExampleObject}
\alias{loadExampleObject}
\title{Load example SPATA2 object}
\usage{
loadExampleObject(sample_name, process = FALSE, meta = FALSE, verbose = TRUE)
}
\arguments{
\item{sample_name}{Character value. One of \code{c("UKF275T", "UKF313T", "UKF269T", "LMU_MIC")}.}

\item{process}{Logical value. If \code{TRUE}, the object is processed via:
\itemize{
\item{\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}}
\item{\code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}}
\item{\code{\link[=removeSpatialOutliers]{removeSpatialOutliers()}}}
\item{\code{\link[=normalizeCounts]{normalizeCounts()}}}
\item{\code{\link[=computeMetaFeatures]{computeMetaFeatures()}}}
}
using their default setup (in this precise order).}

\item{meta}{Logical value. If \code{TRUE}, adds some meta features like clustering, histological
segmentation or copy number variation results to the meta data.frame that has been
computed prior to reducing the object to the example version.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A reduced diet object of class \code{SPATA2}.
}
\description{
Loads a diet example \code{\link{SPATA2}} object.
}
\details{
The example objects only contain a subsetted count matrix. If you want to work
with the complete data sets, use \code{\link[=downloadSpataObject]{downloadSpataObject()}}.
}
\examples{

library(SPATA2)

# simply load
object <- loadExampleObject(sample_name = "UKF269T")

# load, process and add some meta data
object_proc <- loadExampleObject(sample_name = "UKF269T", process = TRUE, meta = TRUE)

getMatrixNames(object)
getMatrixNames(object_proc)

plotSurface(object, color_by = "histology") # fails
plotSurface(object_proc, color_by = "histology")

}
