% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadGSDF}
\alias{loadGSDF}
\title{Original load gene set data.frame}
\usage{
loadGSDF(gene_set_path = NULL, verbose = TRUE)
}
\arguments{
\item{gene_set_path}{Character value (or NULL). Specifies the path to a
.RDS-file containing a data.frame that is to be used as input for slot @used_genesets.

 Must have the character-variables

   \itemize{
    \item{\emph{'ont'}: The gene set names.}
    \item{\emph{'gene'}: The belonging gene names.}
    }

If set to NULL the default gene-set data.frame will used. Run \code{?gsdf} to get more information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Not exported due to naming issues. Kept as it is used in several
loading functions.
}
