% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{loadImage}
\alias{loadImage}
\alias{loadImage,SPATA2-method}
\alias{loadImage,SpatialData-method}
\alias{loadImage,HistoImage-method}
\alias{loadImages}
\alias{loadImages,SPATA2-method}
\alias{loadImages,SpatialData-method}
\title{Load image}
\usage{
loadImage(object, ...)

\S4method{loadImage}{SPATA2}(object, img_name, ...)

\S4method{loadImage}{SpatialData}(object, img_name, verbose = TRUE)

\S4method{loadImage}{HistoImage}(object, verbose = TRUE)

loadImages(object, ...)

\S4method{loadImages}{SPATA2}(object, verbose = TRUE, force = FALSE)

\S4method{loadImages}{SpatialData}(object, verbose = TRUE, force = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{\link{HistoImage}} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{force}{Logical value. If \code{TRUE}, image is loaded even if
the image slot is not empty.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Loads the image based on the directory stored in slot @dir
of the \code{HistoImage} object.
}
\seealso{
\code{\link[=unloadImage]{unloadImage()}},\code{\link[=unloadImages]{unloadImages()}}
}
