% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{loadSpataObject}
\alias{loadSpataObject}
\title{Load SPATA2 object}
\usage{
loadSpataObject(directory_spata, verbose = TRUE, update = TRUE)
}
\arguments{
\item{directory_spata}{Character value. The directory from which to load the \link{`SPATA2'} object.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{update}{Logical value. If \code{TRUE}, calls \code{\link[=updateSpataObject]{updateSpataObject()}} after loading to check whether
the loaded object is up to date.}
}
\value{
The loaded object. If not of class \code{\link{SPATA2}}, a warning is printed.
}
\description{
Reads an .RDS file, checks if it is an object of class \code{\link{SPATA2}}
and checks whether it must be updated.
}
\details{
\code{loadSpataObject()} is a wrapper around \code{base::readRDS()} with which
you could load your \code{\link{SPATA2}} object as well.
}
