% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{make_orthogonal_segment}
\alias{make_orthogonal_segment}
\title{Compute an orthogonal vector}
\usage{
make_orthogonal_segment(sp, ep, out_length, inters_loc = "m")
}
\arguments{
\item{sp, ep}{Numeric vectors of length two. Correspond to the start- and end
point of the geometrical vector for which to compute the orthogonal vector.}

\item{out_length}{The length/magnitude of the orthogonal vector.}

\item{inters_loc}{The location of the intersection. Valid input options are
\emph{'sp'}, \emph{'ep'} or \emph{'m'}.}
}
\value{
List of two slots named sp and ep with vectors of length two that
correspond to the start- and end point of the crossing vector.
}
\description{
Computes the start and enpoint of a vector that crosses an
input vector orthogonally at its start point, end point or at its midth section.
}
