% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{make_orthogonal_segments}
\alias{make_orthogonal_segments}
\title{Make orthogonal segments}
\usage{
make_orthogonal_segments(sp, ep, binwidth, out_length)
}
\arguments{
\item{sp, ep}{Numeric vectors of x- and y-coordinates that correspond to the
start- and end point of the vector for which the orthogonal segment is
computed.}

\item{out_length}{The length of the orthogonal vector.}
}
\value{
A list of two slots named \emph{sp} and \emph{ep}. Both contain
numeric vectors of length two that correspond to the start and
end point of the orthogonal vector.
}
\description{
Make orthogonal segments
}
\keyword{internal}
