% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{makeContent.resizingSegmentsGrob}
\alias{makeContent.resizingSegmentsGrob}
\title{Make content for segments grob}
\usage{
\method{makeContent}{resizingSegmentsGrob}(x)
}
\description{
Used in conjunction with GeomSegmentFixed
}
\keyword{internal}
