% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{makeContent.resizingTextGrob}
\alias{makeContent.resizingTextGrob}
\title{Make content for text grob}
\usage{
\method{makeContent}{resizingTextGrob}(x)
}
\description{
Used in conjunction with GeomTextFixed
}
\keyword{internal}
