% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{map_to_tissue_section}
\alias{map_to_tissue_section}
\title{Map observations to tissue sections}
\usage{
map_to_tissue_section(object, coords_df, cvars = c("x", "y"))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{coords_df}{A data frame containing coordinates to be mapped. Must contain columns specified in \code{cvars}
and a column named \emph{variables}.}

\item{cvars}{A character vector of length 2 specifying the column names for x and y coordinates in \code{coords_df}. Default is \code{c("x", "y")}.}
}
\value{
A data frame with the input coordinates and an additional column \code{tissue_section} indicating the mapped tissue section.
}
\description{
Maps observations in a data frame to their respective tissue sections
based on the results of \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}.
}
