% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{mapImageAnnotationTags}
\alias{mapImageAnnotationTags}
\title{Image annotation and barcode intersection}
\usage{
mapImageAnnotationTags(
  object,
  ids = NULL,
  tags = NULL,
  merge = TRUE,
  merge_name = "img_annotations",
  merge_missing = "none",
  merge_sep = "_",
  merge_drop = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of image annotation with tags} for more information.}

\item{merge}{Logical value. If TRUE, the results are merged in a single variable.}

\item{merge_name}{Character value. The name of the merged variable.}

\item{merge_missing}{Character value. The value that is assigned to barcodes that
do not fall in the extent of any image annotation.}

\item{merge_sep}{Character value. The string with which the image annotation tags
are separated with while being merged.}

\item{merge_drop}{Logical value. If TRUE and \code{merge} is TRUE, all image-annotation-
tag-variables are dropped.}
}
\value{
A data.frame.
}
\description{
Creates a data.frame that maps the tags of image annotations
to the barcodes that were covered by the spatial extent of the respective
image annotation.
}
