% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{matrix_dummy}
\alias{matrix_dummy}
\title{matrix_dummy}
\usage{
matrix_dummy()
}
\value{
A numeric matrix with rownames corresponding to molecule names and column
names corresponding to barcodes.
}
\description{
matrix_dummy
}
