% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-spata-objects.R
\name{mergeSpataObjects}
\alias{mergeSpataObjects}
\title{Merge Spata Objects}
\usage{
mergeSpataObjects(objects, gsdf_input = NULL, verbose = TRUE)
}
\arguments{
\item{objects}{A list of valid spata-objects. All sample names must be unique.}

\item{gsdf_input}{Determines the final input for slot @used_genesets:

If set to \emph{'merge'} all gene-set data.frames of all objects are joined
and unique gene-sets are kept. Gene-sets with the same name but different
genes are merged!

If a directory is specified the directory is given to \code{loadGSDF()}.

If a data.frame is specified that data.frame is used.

If set to NULL the standard \code{SPATA::gsdf} is used.}
}
\value{
A merged spata object.
}
\description{
Takes an arbitrary number of spata-objects and merges them into one.
}
