% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{mergeSpatialAnnotations}
\alias{mergeSpatialAnnotations}
\title{Merge spatial annotations}
\usage{
mergeSpatialAnnotations(
  object,
  ids,
  id,
  tags = NULL,
  tags_expand = TRUE,
  concavity = 2,
  remove_old = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector of ids from spatial annotations to merge.}

\item{id}{Character value. The ID of the new spatial annotation that results
from the merging.}

\item{tags}{A character vector of tags for the spatial annotation.}

\item{tags_expand}{Logical value. If \code{TRUE}, the tags with which the image
annotations are tagged are expanded by the unsuffixed \code{id}, the \code{grouping},
the \code{group} and \emph{'createGroupAnnotations'}.}

\item{concavity}{Numeric value. Given to argument \code{concavity} of
\code{\link[concaveman:concaveman]{concaveman::concaveman()}}. Determines the relative measure of concavity.
1 results in a relatively detailed shape, Infinity results in a convex hull.
You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{remove_old}{Logical value. If \code{TRUE}, the \emph{old} spatial annotations
denoted in \code{ids} are removed from the object.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Merges the spatial extent of two or more spatial annotations
into one.
}
\examples{

library(SPATA2)
library(tidyverse)

object <- loadExampleObject("UKF275T")

r <- getSpatAnnRange(object, id = "img_ann_1")

plotImage(object) +
 ggpLayerSpatAnnOutline(object, ids = c("vessel2", "img_ann_1"), use_colors = T)

plotImage(object, xrange = r$x, yrange = r$y) +
 ggpLayerSpatAnnOutline(object, ids = c("vessel2", "img_ann_1"), use_colors = T)

object <-
 mergeSpatialAnnotations(
   object = object,
   ids = c("img_ann_1", "vessel2"),
   id = "new_img_ann",
   )

plotSpatialAnnotations(object)

}
\seealso{
\code{\link[=getSpatAnnIds]{getSpatAnnIds()}}
}
