% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{mergeTissueSections}
\alias{mergeTissueSections}
\title{Merge tissue sections}
\usage{
mergeTissueSections(object, sections, section_new, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{sections}{Character vector. The names of the tissue sections to be merged.}

\item{section_new}{Character value. The name of the resulting tissue section.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Merges tissue sections that have been mistakenly identified
as two non-contiguous sections.
}
\seealso{
\code{\link[=identifyTissueSections]{identifyTissueSections()}}, \code{\link[=getTissueSections]{getTissueSections()}}
}
