% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{mergeWithTissueOutline}
\alias{mergeWithTissueOutline}
\title{Integrate tissue outline in spatial annotation}
\usage{
mergeWithTissueOutline(object, id, new_id = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The ID of the spatial annotation whose outline
is supposed to be cut at the tissue edge.}

\item{new_id}{If character, gives the resulting spatial annotation a new
id. If \code{NULL}, the spatial annotaiton is effectively overwritten.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Ensures that the outline of a spatial annotaiton does not
transgresses the outline of the tissue.
}
