% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{method_hclust}
\alias{method_hclust}
\title{method_hclust}
\usage{
method_hclust(method_hclust)
}
\arguments{
\item{method_hclust}{Character value. Denotes the method that was used to generate the
clustering results you want to extract.}
}
\description{
method_hclust
}
