% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{model_ascending}
\alias{model_ascending}
\alias{model_descending}
\title{Generate Model-Based Ascending or Descending Sequence}
\usage{
model_ascending(input, incl = 1, ro = range(input))

model_descending(input, dcl = 1, ro = range(input))
}
\arguments{
\item{input}{A numeric vector serving as the basis for generating the sequence.}

\item{incl, dcl}{An optional parameter controlling the inclination/declination of the sequence.}

\item{ro}{A numeric vector of length 2 specifying the range of values for the output sequence.
Default is the range of the input vector.}
}
\value{
A numeric vector representing the generated ascending or descending sequence.
}
\description{
Generates a sequence of values based on a model for ascending or descending patterns.
}
\details{
This function generates a sequence of values based on the input vector and
inclination parameter. It can produce either an ascending or descending sequence
depending on the sign of the inclination parameter. You can also specify a custom
range for the output sequence using the 'ro' parameter.
}
