% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{model_peak}
\alias{model_peak}
\alias{model_trough}
\title{Model a peaking pattern}
\usage{
model_peak(input, dos = 100, pp = 0.5, ro = range(input))

model_trough(input, dos = 100, pp = 0.5, ro = range(input))
}
\arguments{
\item{input}{Numeric vector of length greater than 5.}

\item{dos}{Numeric value. Degree of smoothness. The higher the value the
smoother the peak. The lower the value the sharper the peak. Should range
between 1-100 (if <1 is multiplied with 100 to rescale).}

\item{pp}{Numeric value. Peak position. Determines the position of the
peak either as an index (>= 1) or as a percentage of length (<1).}

\item{ro}{Numeric vector of length two. The range of the output vector.
Defaults to the range of the input.}
}
\value{
Numeric vector of the same length and range as the input
vector that contains a peaking pattern based on the adjustments
of \code{dos} and \code{pp}.
}
\description{
Models a peaking pattern based on an input vector.
}
