% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{moduleAddGeneSetsServer}
\alias{moduleAddGeneSetsServer}
\title{Server of the add gene sets module}
\usage{
moduleAddGeneSetsServer(id, object)
}
\arguments{
\item{id}{The namespace id.}

\item{object}{A valid spata-object.}
}
\value{
An updated spata-object.
}
\description{
Server of the add gene sets module
}
