% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-add-gene-sets.R
\name{moduleAddGeneSetsUI}
\alias{moduleAddGeneSetsUI}
\title{UI of the add gene sets module}
\usage{
moduleAddGeneSetsUI(id)
}
\arguments{
\item{id}{The namespace id.}
}
\description{
UI of the add gene sets module
}
