% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-surface-plot.R
\name{moduleSurfacePlotServer}
\alias{moduleSurfacePlotServer}
\title{Server of the surface plot module}
\usage{
moduleSurfacePlotServer(
  id,
  object,
  final_plot,
  reactive_object,
  highlighted = shiny::reactive(FALSE)
)
}
\arguments{
\item{id}{The namespace id.}

\item{object}{A valid spata-object.}

\item{final_plot}{The final plot that is to be displayed. (See details.).}

\item{reactive_object}{A valid (reactive) spata-object.}
}
\value{
A reactive list with several slots:
 \enumerate{
  \item $assembled_plot() The surface plot as a ggplot-object.
  \item $dblclick() A list containing information regarding the double clicked position in the plot.
  \item $current_setting() A list with information about the settings of \code{assembled_plot} (e.g. sample, color_to, smooth, smoothing_span ...)}
}
\description{
Server of the surface plot module
}
\details{
The argument \code{final_plot} takes a ggplot object as input which is going to be displayed as the final plot. This allows to
adjust the output of \code{$assembled_plot()} outside of the module. If no further adjustment is needed determine \code{final_plot} as:
\code{shiny::reactive(*module_return_variable*()$assembled_plot())}
}
