% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{moduleSurfacePlotUI}
\alias{moduleSurfacePlotUI}
\title{UI of the surface plot module}
\usage{
moduleSurfacePlotUI(id)
}
\arguments{
\item{id}{The namespace id.}
}
\description{
UI of the surface plot module
}
\keyword{internal}
