% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{MolecularAssay-class}
\alias{MolecularAssay-class}
\alias{MolecularAssay}
\title{The \code{MolecularAssay} - Class}
\description{
A class to represent molecular assay data, including analysis results, metrics, and
statistical summaries. The \code{MolecularAssay} class encapsulates various components
of assay data including raw and processed metrics, analytical results,
and associated metadata like the assay name and molecular type.

(In further documentation the shorter term assay is used to refer to molecular
assays.)
}
\section{Slots}{

\describe{
\item{\code{active_mtr}}{Character string indicating which matrix to extract and
use by default.}

\item{\code{analysis}}{List of analysis results where each element can represent
a different analysis aspect such as \link[=runDEA]{DEA} or \link[=runCNV]{CNV}.}

\item{\code{meta_var}}{Data.frame of meta data for the variables of this assay - the molecules.}

\item{\code{mtr_counts}}{Matrix object storing raw counts from the assay. Rownames
should corresponds to the molecule names. Colnames should correspond to the
barcodes (IDs) of the observations to which the molecule counts were mapped.}

\item{\code{mtr_proc}}{List of processed data matrices.}

\item{\code{modality}}{Character value. A string that best characterizes the type of molecular data
the assay carries (e.g., "gene", "protein"). More on molecular modalities in SPATA2
can be found \link[=concept_molecular_modalities]{here}.}

\item{\code{signatures}}{Named list of character vectors.

Molecular signatures are sets of molecules (such as genes or proteins) that are
associated with specific biological states, processes, or conditions. This slot stores the molecular
signatures detected in the assay data. Each signature is represented as a vector of molecule names
in a named list, where the names of the list slots correspond to the signature. Read more
on molecular signatures \link[=concept_molecular_signatures]{here}.}
}}

