% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nGenes}
\alias{nGenes}
\title{Number of genes}
\usage{
nGenes(object, mtr_name = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeriv value.
}
\description{
Returns the number of genes in the active matrix.
}
