% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nImages}
\alias{nImages}
\title{Number of images}
\usage{
nImages(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeric value.
}
\description{
Returns the number of images in the \code{SPATA2} object.
}
