% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nGenes}
\alias{nGenes}
\alias{nMolecules}
\alias{nMetabolites}
\alias{nProteins}
\title{Number of molecules}
\usage{
nGenes(object)

nMolecules(object, assay_name = activeAssay(object))

nMetabolites(object)

nProteins(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Numeric value.
}
\description{
Returns the number of molecules in the raw count matrix of the chosen
assay.
}
\details{
The functions \code{nGenes()}, \code{nProteins()}, \code{nMetabolites()} are wrappers for
objects that contain the corresponding \link[=concept_molecular_modalities]{molecular modality}
and do not have an \code{assay_name} argument.
}
