% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nObs}
\alias{nObs}
\title{Number of observations}
\usage{
nObs(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeric value.
}
\description{
Returns the number of \link[=concept_observations]{observations}
in the sample.
}
