% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(normalize)
}
\arguments{
\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}
}
\description{
normalize
}
