% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Normalize raw counts}
\usage{
normalizeCounts(
  object,
  method = "LogNormalize",
  mtr_name_new = method,
  sct_clip_range = c(-sqrt(x = ncol(x = umi)/30), sqrt(x = ncol(x = umi)/30)),
  activate = TRUE,
  assay_name = activeAssay(object),
  overwrite = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{method}{Character value. The normalization method. One of c(\emph{'LogNormalize'},
\emph{'CLR'}, \emph{'RC'}).}

\item{mtr_name_new}{Character value. The name under which the new processed matrix
is stored in the \code{SPATA2} object.}

\item{activate}{Logical. If \code{TRUE}, the created matrix is activated via \code{activateMatrix()}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Normalizes the count matrix of a molecular assay.
}
\details{
The function creates a temporary \code{Seurat} object and calls \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}}
with the corresponding method. Afterwards, the normalized matrix is extracted and
stored in the \code{SPATA2} object.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

object <- normalizeData(object)

}
