% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{NumericAnnotation-class}
\alias{NumericAnnotation-class}
\alias{NumericAnnotation}
\title{The \code{NumericAnnotation} - Class}
\description{
An S4 class designed to represent the spatial extent
of data points, such as cells or barcoded spots, by filtering and outlining them
according to their values for a specific numeric variable. This class is particularly
suitable for creating annotations that highlight areas of interest based on continuous
characteristics like gene expression or other numeric attributes derived from
spatial multi-omic datasets.
}
\details{
This class is an extension of class \code{\link{SpatialAnnotation}} and
inherits all of its slots.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{list. A named list of the parameters set up to compute
the spatial outline.}

\item{\code{threshold}}{character. The threshold parameter by which the data points were filtered.}

\item{\code{variable}}{character. The name of the numeric variable based on which the
annotation was created.}
}}

