% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{object_dummy}
\alias{object_dummy}
\title{object}
\usage{
object_dummy()
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\description{
object
}
\keyword{internal}
