% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o.R
\name{obsUnit}
\alias{obsUnit}
\alias{obsUnit,ANY-method}
\alias{obsUnit,SpatialMethod-method}
\title{The observational unit}
\usage{
obsUnit(object, ...)

\S4method{obsUnit}{ANY}(object)

\S4method{obsUnit}{SpatialMethod}(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character value.
}
\description{
Extracts the observational unit of the \link[\code{SpatialMethod}]{spatial method (platform)}
of the object.
}
