% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o.R
\name{obtain_inferred_gradient}
\alias{obtain_inferred_gradient}
\title{Obtain Inferred Expression Gradient}
\usage{
obtain_inferred_gradient(
  coords_df,
  variable,
  amccd,
  expr_est_pos = NULL,
  weighted = FALSE,
  delta = NULL,
  npts = 200,
  span = NULL,
  iterations = 4,
  ro = c(0, 1)
)
}
\arguments{
\item{coords_df}{A dataframe containing coordinates with distance and expression values.}

\item{variable}{The variable of interest for which the inferred gradient is to be calculated.}

\item{amccd}{The average minimal center-to-center distance (AMCCD) of the data points, indicating the spatial resolution.}

\item{expr_est_pos}{Optional. A numeric vector specifying the positions where expression estimates should be obtained. If not provided, it is calculated based on the distance and AMCCD.}

\item{weighted}{Logical. If TRUE, weighting is applied based on distance from the core. Default is FALSE.}

\item{delta}{Optional. The smoothing parameter for the loess model. Default is NULL.}

\item{npts}{The number of points to be used for loess smoothing. Default is 200.}

\item{span}{Optional. The span for loess smoothing. If not provided, it is calculated based on AMCCD and distance.}

\item{iterations}{The number of iterations for loess smoothing. Default is 4.}

\item{ro}{A numeric vector of length 2 specifying the range of scaling for the inferred gradient. Default is c(0, 1).
If \code{NULL}, no rescaling is done.}
}
\value{
A dataframe containing the inferred expression gradient with distances and scaled variable values.
}
\description{
This function calculates the inferred expression gradient based on a provided dataframe of coordinates and a specific variable of interest.
}
\details{
This function calculates the inferred expression gradient by first ensuring that the units of the provided average
minimal center-to-center distance (\code{amccd}) match the units of the distances in the input coordinates
dataframe (\code{coords_df}). It then computes the span for the loess smoothing if not provided, based
on the ratio of \code{amccd} and the maximum distance in \code{coords_df}.

The expression estimates positions (\code{expr_est_pos}) are either provided as input
or calculated using the \code{compute_positions_expression_estimates} function.

The function uses the \code{limma::weightedLowess} function for loess smoothing,
allowing customization through parameters such as \code{delta}, \code{npts},
and \code{iterations}.

The gradient is calculated by interpolating expression values at specified positions
using \code{stats::approx}, and the result is returned as a dataframe containing
distances and scaled variable values, where scaling is determined by the \code{ro}
argument.

For additional details on how the positions for expression estimates are calculated,
refer to the \code{compute_positions_expression_estimates} function.
}
\seealso{
\code{\link{compute_positions_expression_estimates}}

\code{\link{compute_positions_expression_estimates}}
}
