% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{pheatmap}
\alias{pheatmap}
\title{de_df}
\usage{
pheatmap(de_df)
}
\arguments{
\item{dea_df}{A data.frame containing information about differentially expressed genes.
This includes the numeric variables \emph{p_val, avg_logFC, p_val_adj} and the character
variables \emph{cluster, gene}.}
}
\description{
de_df
}
