% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH-M.R
\name{plot_model_comparison_dotplot}
\alias{plot_model_comparison_dotplot}
\title{Plot Model Comparison Dotplot}
\usage{
plot_model_comparison_dotplot(
  data,
  eval = "mae",
  pval = "p_value",
  model_subset = NULL,
  model_remove = NULL,
  scale_factor = 1,
  pt_size = 1.5,
  label_vars = 2,
  label_size = 4,
  threshold_pval = 0.05,
  label_color = "#4d4d4d",
  x_label = "Gene-model correlation"
)
}
\arguments{
\item{data}{Output of \code{spatialAnnotationScreening()} or \code{spatialTrajectoryScreening()}
in the form of \code{screening_results@results}.}

\item{model_remove}{(Optional) A character vector specifying models to remove from the plot.}

\item{scale_factor}{A numeric value to scale the point sizes. Default is 1.}

\item{pt_size}{The size of the points in the plot. Default is 0.1.}

\item{label_vars}{The number of top variables to label for each model. Default is 2.}

\item{label_size}{The size of the labels. Default is 4.}

\item{threshold_pval}{The p-value threshold for coloring points. Default is 0.05.}

\item{label_color}{The color of labels. Default is "#4d4d4d".}

\item{x_label}{The label for the x-axis. Default is "Gene-model correlation".}
}
\value{
A dotplot comparing model screening results.
}
\description{
Overview dotplot to compare screening results of selected models.
}
\examples{
# Example usage:
pl <- plot_model_comparison_dotplot(screening_1@results, model_remove = c("peak"), label_vars = 3)
pl + coord_cartesian(xlim = c(0.5, 1))

}
