% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_.R
\name{plot_screening_evaluation}
\alias{plot_screening_evaluation}
\title{Helper}
\usage{
plot_screening_evaluation(
  df,
  variables,
  var_order,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_size = 1,
  line_alpha = 0.9,
  line_color = "blue",
  line_size = 1,
  display_se = FALSE,
  display_corr = FALSE,
  corr_p_min = 5e-05,
  corr_pos_x = NULL,
  corr_pos_y = NULL,
  corr_text_sep = "\\n",
  corr_text_size = 1,
  force_grid = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{force_grid}{Logical value. If \code{TRUE}, \code{facet_grid()} is used regardless
of \code{variables} being of length 1.}
}
\description{
Helper
}
