% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{plot_type_dummy}
\alias{plot_type_dummy}
\title{plot_type}
\usage{
plot_type_dummy(plot_type)
}
\arguments{
\item{plot_type}{Character value. Specifies the type of plot to use to
visualize the results. If valid input options are not mentioned in the
description use \code{validPlotTypess()} to obtain all valid input options.}
}
\description{
plot_type
}
