% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{plotAutoencoderAssessment}
\alias{plotAutoencoderAssessment}
\title{Plot total variance of different neural networks}
\usage{
plotAutoencoderAssessment(
  object,
  activation_subset = NULL,
  clrp = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
ggplot_family return
}
\description{
Visualizes the results of \code{assessAutoencoderOptions()} by displaying the
total variance of each combination of an activation function (such as \emph{relu, sigmoid})
and the number of bottleneck neurons.

The results depend on further adjustments like number of layers, dropout and number of epochs.
}
\keyword{internal}
