% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-plotting-functions.R
\name{plotAutoencoderResults}
\alias{plotAutoencoderResults}
\title{Plot scaled vs. denoised expression}
\usage{
plotAutoencoderResults(
  object,
  genes,
  mtr_name = "denoised",
  normalize = NULL,
  scales = NULL,
  pt_alpha = NULL,
  pt_clrp = NULL,
  pt_size = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{genes}{Character vector of length two. Denotes the genes to be used for the validation plot.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Compares the distribution of the expression levels of \code{genes}
between the scaled matrix and the denoised matrix.
}
\details{
This function requires a denoised matrix in slot @data generated
by \code{runAutoEncoderDenoising()} as well as a scaled matrix.
}
