% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-statistics.R
\name{plotBoxplot}
\alias{plotBoxplot}
\alias{plotDensityplot}
\alias{plotHistogram}
\alias{plotRidgeplot}
\alias{plotViolinplot}
\title{Plot numeric distribution and statistical tests}
\usage{
plotBoxplot(
  object,
  variables,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clrp = NULL,
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  vjust = 0,
  display_facets = NULL,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  n_bcsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_size = NULL,
  pt_shape = NULL,
  method_gs = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotDensityplot(
  object,
  variables,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clrp = NULL,
  clrp_adjust = NULL,
  display_facets = NULL,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  method_gs = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotHistogram(
  object,
  variables,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clrp = NULL,
  clrp_adjust = NULL,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  method_gs = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotRidgeplot(
  object,
  variables,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  alpha = 0.8,
  clrp = NULL,
  clrp_adjust = NULL,
  display_facets = NULL,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  method_gs = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotViolinplot(
  object,
  variables,
  across = NULL,
  across_subset = NULL,
  relevel = NULL,
  clrp = NULL,
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  display_facets = NULL,
  vjust = 0,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  n_bcsp = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_size = NULL,
  pt_shape = NULL,
  method_gs = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{variables}{Character vector. The variables of interest:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
 }}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the colors with which they are displayed.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{test_groupwise}{Character value or NULL. Specifies the groupwise statistical
test to be conducted. If character, one of \emph{'anova', 'kruskal.test'}. If set
to NULL the testing is skipped.}

\item{test_pairwise}{Character value or NULL. Specifies the pairwise statistical
test to be conducted. If character, one of \emph{'t.test', 'wilcox.test'}. If set
to NULL the testing is skipped.}

\item{ref_group}{Character value or NULL. Specifies the reference group against
which all other groups are compared in the test denoted in \code{test_groupwise}
is conducted. If set to NULL the first group found is taken.}

\item{step_increase}{Numeric value. Denotes the increase in fraction of total
height for every additional comparison to minimize overlap.}

\item{vjust}{Numeric value. Denotes the relative, vertical position of the results of
the test denoted in \code{test.groupwise}. Negative input highers, positive
input lowers the position.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{scales}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{nrow}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{ncol}{Given to \code{ggplot2::facet_wrap()}. Affects the way the subplots
are displayed.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{n_bcsp}{Numeric value. Specifies the sample size of barcode-spots and
can be set to prevent overplotting.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to the respective \code{ggplot2::geom_<plot_type>()}
function. E.g. \code{plotViolinplot()} relies on \code{ggplot2::geom_violin()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
These functions display the distribution of numeric
variables for the whole sample or in a comparative manner if argument
\code{across} is specified. \code{plotViolinplot()} and \code{plotBoxplot()}
allow for statistical tests such as t-test or ANOVA.
}
