% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-gradient-screening.R
\name{plotClockplot}
\alias{plotClockplot}
\alias{plotClockplot,spata2-method}
\alias{plotClockplot,SpatialGradientScreening-method}
\title{Plot Clockplots}
\usage{
plotClockplot(object, ...)

\S4method{plotClockplot}{spata2}(
  object,
  id,
  variables,
  n_circles,
  buffer,
  n_angle_bins = 12,
  pattern_subset = NULL,
  pattern_remove = NULL,
  pattern_add = NULL,
  layout = 1,
  switch = NULL,
  fill = "steelblue",
  ...
)

\S4method{plotClockplot}{SpatialGradientScreening}(
  object,
  variables,
  n_angle_bins = 12,
  pattern_subset = NULL,
  pattern_remove = NULL,
  layout = 1,
  switch = NULL,
  fill = "steelblue",
  ...
)
}
\description{
Plot Clockplots
}
