% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA-D.R
\name{plotCnvLineplot}
\alias{plotCnvLineplot}
\title{Plot CNV Lineplot}
\usage{
plotCnvLineplot(
  object,
  across = NULL,
  across_subset = NULL,
  arm_subset = c("p", "q"),
  chrom_subset = 1:22,
  chrom_separate = 1:22,
  chrom_arm_subset = NULL,
  smooth_span = 0.08,
  line_alpha = 0.9,
  line_color = "blue",
  line_size = 1,
  display_ribbon = TRUE,
  ribbon_alpha = 0.25,
  ribbon_fill = "lightgrey",
  vline_alpha = 0.75,
  vline_color = "black",
  vline_size = 0.5,
  vline_type = "dashed",
  summarize_with = "mean",
  nrow = NULL,
  ncol = NULL,
  breaks_y = c(0.9, 0.95, 1, 1.05, 1.1),
  labels_y = breaks_y,
  limits_y = base::range(breaks_y),
  expand_y = ggplot2::waiver(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{arm_subset}{Character vector. A combination of \emph{'p'} and/or \emph{'q'}.
Denotes which chromosome arms are included. Defaults to both.}

\item{chrom_subset}{Character or numeric vector. Denotes the chromosomes that
are included. Defaults to all 1-22.}

\item{chrom_separate}{Character or numeric vector. Denotes the chromosomes that
are separated from their neighbors by vertical lines. Defaults to all 1-22. If FALSE or NULL,
no vertical lines are drawn. Requires \code{display_vlines} to be set to TRUE.}

\item{chrom_arm_subset}{Character vector. Denotes the exact chromosome-arm combinations
that are included.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{display_ribbon}{Logical value. If TRUE, a ribbon is displayed around
the main line of the plot visualizing uncertainty using standard deviation.}

\item{ribbon_alpha, ribbon_fill}{Parameters given to \code{ggplot2::geom_ribbion()}
that control the appearance of the ribbon around the main line of the plot.}

\item{vline_alpha, vline_color, vline_size, vline_type}{Parameters given to
\code{ggplot2::geom_vline()} that control the appearance of vertical lines
of the plot.}

\item{summarize_with}{Character value. Name of the function with which to summarize. Either
\emph{'mean'} or \emph{'median'}.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{breaks_y, labels_y, limits_y, expand_y}{Given to the corresponding arguments
of \code{ggplot2::scale_y_continuous()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A ggplot.
}
\description{
Plots the results of \code{runCnvAnalysis()} in form of a lineplot.
Use arguments \code{across} and \code{across_subset} to visualize CNV differences
between subgroups of cluster variables or other grouping variables (e.g. based on
histology created with \code{createSpatialSegmentation()}).
}
