% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-analysis.R
\name{plotCnvResults}
\alias{plotCnvResults}
\title{Visualize copy-number-variations results}
\usage{
plotCnvResults(...)
}
\arguments{
\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays a smoothed lineplot indicating
chromosomal gains and losses. Requires the results of \code{runCnvAnalysis()}.
}
