% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotCsrCutoffSimulations}
\alias{plotCsrCutoffSimulations}
\title{Title}
\usage{
plotCsrCutoffSimulations(object, of_sample = NA)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Visualizes the number of genes that would remain depending on the
cutoff value chosen for the cluster tendency results.
}
