% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotCustomizedTrajectoryTrends}
\alias{plotCustomizedTrajectoryTrends}
\title{Plot customized trajectory trends}
\usage{
plotCustomizedTrajectoryTrends(
  customized_trends,
  smooth = TRUE,
  smooth_span = 0.2,
  smooth_se = FALSE,
  clrp = "milo",
  ...
)
}
\arguments{
\item{customized_trends}{A data.frame or a named list. All numeric variables are considered to correspond to customized trends
the trajectory of interest might adopt. The names of the respective variables will correspond to the name
with which you want to refer to the trend later on.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.
displayed.}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes the trajectory trends you set up yourself.
}
