% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDeaDotPlot}
\alias{plotDeaDotPlot}
\title{Plot differentially expressed genes}
\usage{
plotDeaDotPlot(
  object,
  across,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  by_group = TRUE,
  max_adj_pval = NULL,
  min_lfc = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  genes = NULL,
  x = c("avg_logFC", "p_val_adj")[1],
  alpha_by = NULL,
  alpha_trans = "identity",
  color_by = "avg_logFC",
  color_trans = "identity",
  size_by = "p_val_adj",
  size_trans = "reverse",
  pt_alpha = 0.9,
  pt_size = 2,
  pt_color = "blue4",
  pt_clrsp = "plasma",
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  transform_with = NULL,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character value or NULL. If character, specifies the name of the grouping variable
that contains the group assignment across which results are displayed. If NULL,
no grouping is done.}

\item{across_subset}{Character vector or NULL. If character, specifies groups of the
grouping variable that are of interest. Prefixing group names with \emph{'-'} excludes them.
If NULL, all groups are considered.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{by_group}{Logical value. If TRUE for every group in the grouping
variable a single dot plot is created. If FALSE one plot for all groups and all
gene sets is created.}

\item{genes}{Character vector or NULL. If character, vector of gene names
that determines which genes are included. If NULL, genes are taken according
to the threshold input for average log fold change and adjusted p-value.}

\item{x}{Character value. Specifies what is plotted on the x-axis.
If \emph{p_val_adj} the scale is reversed. Ignored if \code{by_group} = FALSE.}

\item{alpha_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{alpha_trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{color_by}{Character value or NULL. If character, specifies the
variable that is displayed by color.}

\item{color_trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{size_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{size_trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_color}{Character value. Specifies the color with which all points are displayed .}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{scales}{Affects the way the subplots
are displayed.}

\item{nrow}{Affects the way the subplots
are displayed.}

\item{ncol}{Affects the way the subplots
are displayed.}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{...}{Additional arguments given to \code{scale_color_add_on()}.}
}
\value{
A ggplot.
}
\description{
Visualizes results of DE analysis with
dot plots.
}
