% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDeaHeatmap}
\alias{plotDeaHeatmap}
\title{Plot differentially expressed genes}
\usage{
plotDeaHeatmap(
  object,
  across,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  max_adj_pval = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  breaks = NULL,
  genes = NULL,
  n_bcsp = NULL,
  clrp = NULL,
  colors = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the colors with which they are displayed.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{max_adj_pval}{Numeric value. Sets the maximal threshold for adjusted p-values allowed. All genes
with adjusted p-values above that threshold are ignored.}

\item{n_highest_lfc}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{n_lowest_pval}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{breaks}{Denotes the colorspectrum breaks. If set to NULL the breaks are set automatically. If a
numeric vector is specified it is taken as input. If a function is specified the expression matrix is
passed to it as the first argument and the length of \code{colors} as the second argument.}

\item{genes}{Character vector or NULL. If you want to display specific genes irrespective of de-anaylsis results you
can specifiy them in \code{genes}. If \code{genes} is specified that way arguments referring to de-anylsis results are
ignored and only the genes specified are taken and displayed.}

\item{n_bcsp}{The number of barcode-spots belonging to each cluster you want to
include in the matrix. Should be lower than the total number of barcode-spots of every cluster
and can be deployed in order to keep the heatmap clear and aesthetically pleasing.

If set to NULL (the default) it is automatically computed according to the number of genes that
are displayed in the heatmap.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to \code{pheatmap::pheatmap()}.}
}
\value{
A heatmap of class 'pheatmap'.
}
\description{
Visualizes the expression of genes across subgroups in a heatmap. It either takes the results
from previously conducted de-analysis or uses the expression information of specific genes to plot a heatmap.
}
\details{
The de-data.frame is processed such that the following steps are performed for every experimental
group.

\enumerate{
 \item{Discards genes with \emph{avg_logFC}-values that are either infinite or negative}
 \item{Discards genes with adjusted p-values above the threshold set with \code{max_adj_pval}}
 \item{Slices the data.frame in order that for every experimental group:}
 \enumerate{
  \item{the n genes with the highest \emph{avg_logFC}-values are kept where n = \code{n_highest_lfc}}
  \item{the n genes with the lowest \emph{p_val_adj}-values are kept where n = \code{n_lowest_pval}}
  }
 \item{Arranges the genes according to the highest \emph{avg_logFC}-values}
 }
}
