% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R, R/unused.R
\name{plotDendrogramCnv}
\alias{plotDendrogramCnv}
\alias{plotDendrogram}
\title{dummy}
\usage{
plotDendrogramCnv(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_size = 1,
  clrp = NULL,
  clrp_adjust = NULL,
  display_legend = NULL,
  display_title = NULL,
  ncol = NULL,
  nrow = NULL,
  of_sample = NA,
  verbose = NULL
)

plotDendrogram()
}
\description{
dummy
}
