% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDistribution}
\alias{plotDistribution}
\alias{plotDistribution2}
\alias{plotDistributionAcross}
\alias{plotDistributionAcross2}
\alias{plotDistributionDiscrete}
\title{Distribution of continuous values (Deprecated)}
\usage{
plotDistribution(
  object,
  variables,
  plot_type = "histogram",
  method_gs = NULL,
  clrp = NULL,
  binwidth = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotDistribution2(
  df,
  variables = "all",
  plot_type = "histogram",
  clrp = "milo",
  binwidth = 0.05,
  verbose = TRUE,
  ...
)

plotDistributionAcross(
  object,
  variables,
  across,
  across_subset = NULL,
  binwidth = 0.05,
  method_gs = NULL,
  plot_type = NULL,
  clrp = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotDistributionAcross2(
  df,
  variables = "all",
  across,
  across_subset = NULL,
  plot_type = "violin",
  binwidth = 0.05,
  clrp = "milo",
  ...,
  normalize = TRUE,
  assign = FALSE,
  assign_name,
  verbose = TRUE
)

plotDistributionDiscrete(
  object,
  features,
  feature_compare = NULL,
  clrp = NULL,
  position = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{variables}{Character vector. The numeric variables of interest. Must be inside:

\itemize{
  \item{ \strong{Gene sets} Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} Must be in \code{getGenes()}}
  \item{ \strong{Features} Must be in \code{getFeatureNames(..., of_class = "numeric")}}
  }}

\item{plot_type}{Character value. One of \emph{'histogram', 'density', 'violin', 'boxplot' and 'ridgeplot'}.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{binwidth}{The binwidth to use if \code{plot_type} is specified as \emph{'histogram'}.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{additional arguments to \code{ggplot2::facet_wrap()}}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{assign}{Logical. If set to TRUE a named list will be assigned to the global
environment. This list contains data and information to rebuild or additionally
customize the output plot of this function.}

\item{assign_name}{The name the assigned list is supposed to have specified as
a single character value.}
}
\description{
These functions are deprecated in favor of \code{plotDensityplot(),
plotHistogram(), plotRidgplot(), plotBoxplot(), plotViolinplot()} and \code{plotBarchart()}.
}
