% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA-E.R
\name{plotExprVsDistSA}
\alias{plotExprVsDistSA}
\title{Plot Expression as a Function of Distance to a Spatial Annotation}
\usage{
plotExprVsDistSA(
  object,
  variables,
  core = FALSE,
  id = idSA(object),
  distance = distToEdge(object, id = id),
  binwidth = recBinwidth(object),
  pt_alpha = 0.1,
  pt_color = "black",
  pt_clrp = NULL,
  pt_size = 1.5,
  line_alpha = 0.9,
  line_color = "forestgreen",
  line_size = 1.75,
  border_linealpha = 1,
  border_linecolor = "black",
  border_linesize = 1,
  border_linetype = "solid",
  se_fill = ggplot2::alpha("lightgrey", 0.5),
  normalize = FALSE,
  unit = getDefaultUnit(object),
  ggpLayers = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. All numeric variables (meaning genes, gene-sets
and numeric features) that are supposed to be plotted.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{A numeric vector of distances to the annotation's edge.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{se_fill}{The fill color for the smoothing line's standard error area (default: "lightgrey").}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A ggplot.
}
\description{
Generates a scatterplot to visualize the relationship between gene expression and
the distance of data points to an annotation's outline.
}
