% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotFourStates}
\alias{plotFourStates}
\alias{plotFourStates2}
\title{Gene set state plot}
\usage{
plotFourStates(
  object,
  states,
  color_by = NULL,
  method_gs = NULL,
  average_genes = NULL,
  pt_alpha = NULL,
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  display_labels = NULL,
  verbose = NULL,
  of_sample = NA
)

plotFourStates2(
  data,
  states,
  color_by = NULL,
  pt_size = 1.5,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  pt_clrp = "milo",
  display_labels = TRUE
)
}
\arguments{
\item{states}{The gene sets defining the four states specified as a character vector
of length 4.}

\item{color_by}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{data}{A data.frame containing at least the variables \emph{barcodes, \code{states.}}.
Whereby the states-variables contain the respective expression values of the specified
gene sets.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Takes four gene sets and visualizes the relative
expression of these four gene sets for every barcode by computing it's respective
x- and y- coordinates in the state plot. (See details.)

\itemize{
 \item{ \code{plotFourStates()} Takes the spata-object as the starting point and creates
 the necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotFourStates2()} Takes a data.frame as input.}
 }
}
