% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotGeneMetaData}
\alias{plotGeneMetaData}
\title{Plot gene meta data}
\usage{
plotGeneMetaData(
  object,
  variables,
  plot_type = "scatter",
  mtr_name = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_fill = NULL,
  pt_size = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{variables}{Character vector. The variables among the gene meta data
that you want to visualize.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
This function visualizes variables of the @gdata slot that
contains information about the expression profile of each gene in
the expression matrix specified via the \code{mtr_name}-argument.

If the input vector of argument \code{variables} is of length two and
argument \code{plot_type} is set to \emph{'scatter'} a scatterplot is
plotted. Else, depending on the input for \code{plot_type} it is
either a histogram, a densityplot, a ridgeplot, a violinplot or a
boxplot.
}
