% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotGenePattern}
\alias{plotGenePattern}
\alias{plotGenePatterns}
\title{Title}
\usage{
plotGenePattern(
  object,
  genes = NULL,
  pattern = NULL,
  facet_by = c("genes", "gene_pattern"),
  background = "coords",
  pt_alpha = NULL,
  pt_size = NULL,
  nrow = NULL,
  ncol = NULL
)

plotGenePatterns(
  object,
  genes = NULL,
  gene_patterns = NULL,
  pattern_subset = NULL,
  smooth = TRUE,
  smooth_span = 0.2,
  hull_clr = "red",
  hull_clrp = "milo",
  hull_expand = 1,
  hull_size = 1,
  pt_clrsp = "inferno",
  pt_alpha = NULL,
  pt_size = NULL,
  pt_size_fixed = NULL,
  geom_mark_hull = list(),
  na_rm = TRUE,
  of_sample = NA,
  ...
)
}
\arguments{
\item{hull_expand}{}
}
\value{

}
\description{
Title
}
