% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotGenePatternCluster}
\alias{plotGenePatternCluster}
\title{Title}
\usage{
plotGenePatternCluster(
  object,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  h = NULL,
  k = NULL,
  cluster_subset = NULL,
  ncol = NULL,
  nrow = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  normalize = NULL,
  smooth = NULL,
  smooth_span = NULL,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
