% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotGenePatternDendrogram}
\alias{plotGenePatternDendrogram}
\title{Title}
\usage{
plotGenePatternDendrogram(
  object,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  of_sample = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
