% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotGenePatternHeatmap}
\alias{plotGenePatternHeatmap}
\title{Title}
\usage{
plotGenePatternHeatmap(
  object,
  threshold_sim = 0.25,
  verbose = NULL,
  of_sample = NA,
  method_dist = "euclidean",
  method_aggl = "ward.D",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
