% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotGenePatterns}
\alias{plotGenePatterns}
\title{Title}
\usage{
plotGenePatterns(
  object,
  genes = NULL,
  gene_patterns = NULL,
  pattern_subset = NULL,
  smooth = TRUE,
  smooth_span = 0.2,
  hull_clr = "red",
  hull_clrp = "milo",
  hull_expand = 1,
  hull_size = 1,
  split_by = "genes",
  pt_clrsp = "inferno",
  pt_size = 2.3,
  geom_mark_hull = list(),
  of_sample = NA,
  ...
)
}
\arguments{
\item{hull_expand}{}
}
\value{

}
\description{
Title
}
