% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-set-enrichment.R
\name{plotGseaDotPlot}
\alias{plotGseaDotPlot}
\title{Plot gene set enrichment}
\usage{
plotGseaDotPlot(
  object,
  across = getDefaultGrouping(object, verbose = TRUE, "across"),
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  by_group = TRUE,
  n_gsets = 20,
  signif_var = "fdr",
  signif_threshold = 0.05,
  alpha_by = NULL,
  alpha_trans = "reverse",
  color_by = "fdr",
  color_trans = "reverse",
  size_by = "fdr",
  size_trans = "reverse",
  pt_alpha = 0.9,
  pt_size = 2,
  pt_color = "blue4",
  pt_clrsp = "plasma",
  remove = "^.+?(?=_)",
  remove_gsets = NULL,
  replace = c("_", " "),
  arrange_gsets = TRUE,
  reverse = TRUE,
  reverse_within = FALSE,
  scientific = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  transform_with = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character value or NULL. If character, specifies the name of the grouping variable
that contains the group assignment across which results are displayed. If NULL,
no grouping is done.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{by_group}{Logical value. If TRUE for every group in the grouping
variable a single dot plot is created. If FALSE one plot for all groups and all
gene sets is created.}

\item{n_gsets}{Numeric value. Maximal number of gene sets whose results are included.
The first \code{n_gsets} are included starting with the one with the lowest significance value.}

\item{signif_var}{Character value. Determines what to be considered while checking
for significance. Either \emph{'pval'} (p-Value) or \emph{'fdr'} (False Discovery Rate).}

\item{signif_threshold}{Numeric value. Significance values below \code{signif_threshold}
are not included.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{remove}{Character value or NULL. If character, regular expression given
to \code{pattern} of \code{stringr::str_remove_all()}. Used to adjust gene set names.}

\item{remove_gsets}{Character value or NULL. If character, regular expression.
All gene set names that match the regular expression are not included in
the plot.}

\item{replace}{Character vector of length 2 or NULL. If character vector, two regular expressions.
First value is given to argument \code{pattern} and second is given to argument \code{replacement}
of \code{stringr::str_replace_all()}. Used to adjust gene set names.}

\item{arrange_gsets}{Logical. If TRUE gene sets are arranged by their group
belonging. Making the appearance of the plots tidier.}

\item{reverse}{Logical. If TRUE the gene sets are arranged from top to bottom.
If FALSE they are arranged from bottom to top.}

\item{scales, ncol, nrow}{Affects the way the subplots
are displayed.}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{scale_color_add_on()}.}

\item{reverse_whitin}{Logical. If TRUE the gene sets are displayed in a reversed
order within the groups.}
}
\value{
A ggplot.
}
\description{
Visualizes results of gene set enrichment analysis with
dot plots.
}
