% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotHspaSummary}
\alias{plotHspaSummary}
\title{Title}
\usage{
plotHspaSummary(object, plot_type = "histogram", of_sample = NA, ...)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
