% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIasHeatmap.R
\name{plotIasHeatmap}
\alias{plotIasHeatmap}
\title{Plot Image Annotation Screening Heatmap}
\usage{
plotIasHeatmap(
  object,
  id,
  variables,
  buffer,
  n_bins_circle,
  include_area = FALSE,
  arrange_rows = "input",
  method_gs = "mean",
  smooth_span = 0.4,
  multiplier = 10,
  clrsp = "inferno",
  .cols = dplyr::everything(),
  summarize_with = "mean",
  .f = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}. (See details for more.)}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{buffer}.
 (See details for more.)}

\item{include_area}{Logical. If TRUE, the visualization
includes the area the image annotation covered. If FALSE,
visualization only includes the surrounding of the area of the image
annotation.}

\item{arrange_rows}{Alter the way the rows of the heatmap
are displayed in order to highlight patterns. Currently either \emph{'maxima'},
\emph{'minima'} or \emph{'input'}. If \emph{'input'}, variables are displayed
in the same order as they are provided in the argument \code{variables}.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{multiplier}{Numeric value. For better visualization the transient pattern
is smoothed with a loess fit. The total number of predicted values (via \code{stats::predict()})
is the number of bins multiplied with the input for this argument.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}.
Specifies the function with which the bins are summarized.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap()}. If argument
\code{with_ggplot} is TRUE given to \code{scale_color_add_on()}.}
}
\value{
A ggplot.
}
\description{
Plot Image Annotation Screening Heatmap
}
\details{
In conjunction with argument \code{id} which provides the
ID of the image annotation of interest the arguments \code{n_bins_circle},
\code{buffer}, \code{angle_span} and \code{n_bins_angle} can be used
to specify the exact area that is screened as well as the resolution of the screening.

Note that the function \code{plotSurfaceIas()} allows to visually check
if your input results in the desired screening.
}
