% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIasHeatmap.R
\name{plotIasLineplot}
\alias{plotIasLineplot}
\title{Plot Image Annotation Screening Lineplot}
\usage{
plotIasLineplot(
  object,
  id,
  buffer,
  n_bins_circle,
  n_bins_angle,
  variables,
  include_area = FALSE,
  method_gs = NULL,
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = TRUE,
  clrp = NULL,
  clrp_adjust = NULL,
  alpha = 0.4,
  linecolor = "black",
  linesize = 1.5,
  facet_by = "variables",
  summarize_with = "mean",
  nrow = NULL,
  ncol = NULL,
  display_axis_text = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}.}

\item{n_bins_circle}{Numeric value (or vector of length two, see details
for more). Specifies how many times the area is buffered with the value
denoted in \code{buffer}. The combination of both results in the area that is
screened.}

\item{variables}{Character vector. All numeric variables, meaning genes,
gene-sets and numeric features that are supposed to be included in
the screening process.}

\item{include_area}{Logical. If TRUE, the visualization
includes the area the image annotation covered. If FALSE,
visualization only includes the surrounding of the area of the image
annotation.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{smooth_method}{The smoothing method that will be used specified as a
single character value (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{linesize}{Numeric value. Specifies the thicknes of the lines with which
the trajectory dynamics are displayed.}

\item{facet_by}{Either \emph{'variables'} or \emph{'bins_angle'}.}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}.
Specifies the function with which the bins are summarized.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap()}. If argument
\code{with_ggplot} is TRUE given to \code{scale_color_add_on()}.}
}
\value{
A ggplot.
}
\description{
Plot Image Annotation Screening Lineplot
}
