% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH-M.R
\name{plotIasLineplotFitted}
\alias{plotIasLineplotFitted}
\title{Plot IAS model fitting}
\usage{
plotIasLineplotFitted(
  object,
  id,
  variables,
  distance,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  method_gs = NULL,
  lineorder = c(1, 2, 3),
  linesize = 1,
  linecolors = c("forestgreen", "blue4", "red3"),
  linetypes = c("solid", "solid", "dotted"),
  display_residuals = TRUE,
  area_alpha = 0.25,
  nrow = NULL,
  ncol = NULL,
  force_grid = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{distance}{Distance value. Specifies the distance from the border of the
image annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{binwidth}{Distance value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the center-center
distance: \code{binwidth = getCCD(object)}. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{linesize}{Numeric value. The size of the line(s) plotted.}

\item{linecolors, linetypes}{The colors and types of the three lines. First value stands for the
values of the variable, second on for the models, third one for the residuals.}

\item{display_residuals}{Logical value. If TRUE, the residuals curve is displayed.}

\item{area_alpha}{Numeric value. The alpha value for the area under the curve
of the resiudals.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An invisible list of ggplots.
}
\description{
Plots an IAS lineplot in combination with models
fitted to the binned distance towards the image annotation.
}
\details{
The method for class \code{ImageAnnotationScreening} (the output of
the function \code{imageAnnotationScreening()}) can be used
to show the area on which the results base. Therefore, it does not have
arguments \code{binwidth}, \code{n_bins_circle} and \code{n_bins_angle}.
}
