% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH-M.R
\name{plotImage}
\alias{plotImage}
\alias{plotImage,spata2-method}
\alias{plotImage,HistologyImagingNew-method}
\alias{plotImage,HistologyImage-method}
\alias{plotImage,Image-method}
\title{Plot histology image}
\usage{
plotImage(object, ...)

\S4method{plotImage}{spata2}(object, xrange = NULL, yrange = NULL, ...)

\S4method{plotImage}{HistologyImagingNew}(
  object,
  name = NULL,
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1,
  ...
)

\S4method{plotImage}{HistologyImage}(object, xrange = NULL, yrange = NULL, scale_fct = 1, ...)

\S4method{plotImage}{Image}(object, scale_fct = 1, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}
}
\value{
A plot that is immediately plotted.
}
\description{
Plots the histology image as a raster.
}
