% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{plotImageAnnotations}
\alias{plotImageAnnotations}
\title{Deprecated}
\usage{
plotImageAnnotations(...)
}
\description{
Deprecated in favor of \code{\link[=plotSpatialAnnotations]{plotSpatialAnnotations()}}
}
\keyword{internal}
