% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{plotImageBase}
\alias{plotImageBase}
\alias{plotImageBase,spata2-method}
\alias{plotImageBase,HistoImaging-method}
\alias{plotImageBase,HistoImage-method}
\alias{plotImageBase,Image-method}
\title{Plot histology image}
\usage{
plotImageBase(object, ...)

\S4method{plotImageBase}{spata2}(object, xrange = NULL, yrange = NULL, ...)

\S4method{plotImageBase}{HistoImaging}(
  object,
  img_name = NULL,
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1,
  axes = TRUE,
  ...
)

\S4method{plotImageBase}{HistoImage}(
  object,
  img_name = NULL,
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1,
  axes = TRUE,
  ...
)

\S4method{plotImageBase}{Image}(
  object,
  scale_fct = 1,
  xrange = NULL,
  yrange = NULL,
  axes = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
A plot that is immediately plotted.
}
\description{
Plots the histology image as a raster.
}
