% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{plotImageMask}
\alias{plotImageMask}
\alias{plotImageMask,spata2-method}
\alias{plotImageMask,HistoImaging-method}
\alias{plotImageMask,HistoImage-method}
\alias{plotPixelContent}
\alias{plotPixelContent,HistoImaging-method}
\alias{plotPixelContent,HistoImage-method}
\title{Plot pixel content}
\usage{
plotImageMask(object, ...)

\S4method{plotImageMask}{spata2}(object, img_name = NULL, clr_fg = "black", clr_bg = "white")

\S4method{plotImageMask}{HistoImaging}(object, img_name = NULL, clr_fg = "black", clr_bg = "white")

\S4method{plotImageMask}{HistoImage}(object, clr_fg = "black", clr_bg = "white")

plotPixelContent(object, ...)

\S4method{plotPixelContent}{HistoImaging}(
  object,
  clrp = "sifre",
  clr_bg = "white",
  clr_fragments = "red",
  clr_tissue = "forestgreen",
  clr_artefact = "blue",
  type = FALSE,
  clrp_adjust = NULL
)

\S4method{plotPixelContent}{HistoImage}(
  object,
  clrp = "sifre",
  clr_bg = "white",
  clr_fragments = "red",
  clr_tissue = "forestgreen",
  clr_artefact = "blue",
  type = FALSE,
  clrp_adjust = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{clr_fg, clr_bg}{Character values. Color with which to display
foreground and background of the mask.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clr_artefact, clr_fragments, clr_tissue}{Character values. Colors
with which to display the content type if \code{type = FALSE}.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}
}
\value{
A ggplot.
}
\description{
Visualizes the results of \code{\link[=identifyPixelContent]{identifyPixelContent()}}.
\itemize{
\item{\code{plotImageMask()}:}{ Distinguishes pixel in back- and foreground. Foreground being the tissue.}
\item{\verb{plotPixelContent():}}{ Visualizes the classification of each pixel in detail.}
}
}
