% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{plotImages}
\alias{plotImages}
\alias{plotImages,spata2-method}
\alias{plotImages,HistoImaging-method}
\title{Plot histology images (ggplot2)}
\usage{
plotImages(object, ...)

\S4method{plotImages}{spata2}(
  object,
  img_names = NULL,
  verbose = NULL,
  nrow = NULL,
  ncol = NULL,
  ...
)

\S4method{plotImages}{HistoImaging}(
  object,
  img_names = NULL,
  ncol = NULL,
  nrow = NULL,
  image = TRUE,
  outline = TRUE,
  outline_ref = FALSE,
  by_section = TRUE,
  fragments = FALSE,
  line_alpha = line_alpha_ref * 0.75,
  line_alpha_ref = 1,
  line_color = "black",
  line_color_ref = "red",
  line_size = 0.5,
  line_size_ref = line_size * 1.5,
  transform = TRUE,
  img_alpha = 1,
  against_ref = FALSE,
  alignment_eval = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Additionel arguments given to \code{plotImage()}.}

\item{img_names}{Character vector or \code{NULL}. If character, specifies the images
by name. If \code{NULL}, all images are plotted.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}
}
\value{
A ggplot assembled with via \code{patchwork::wrap_plots()}.
}
\description{
Reads in and plots all images known to the \code{SPATA2} object.
}
\section{Distance measures}{


Several functions in \code{SPATA2} have arguments that take \emph{distance input}.
To specifically refer to a distance the unit must be specified. There are
three ways to create valid input for these arguments.

\strong{1. In pixel:}

There are two valid input options to specify the distance in pixel:

\itemize{
\item{numeric:}{ Single numeric values, e.g. \code{arg_input = c(2, 3.554, 69, 100.67)}. If no unit
is specified the input will be interpreted as pixels.}
\item{character:}{ Suffixed with \emph{'px'}, e.g. \code{arg_input = c('2px', '3.554px', '69px', '100.67px')}}
}

Note: The unit pixel (px) is used for distances as well as for areas. If pixel
refers to a distance the pixel side length is meant. If pixel refers to an area the
number of pixels is meant.

\strong{2. According to the Systeme international d`unites (SI):}

Specifying distances in SI units e.g. \code{arg_input = c('2mm', '4mm')} etc.
requires the input to be a character as the unit must be provided as suffix.
Between the numeric value and the unit must be no empty space! Valid suffixes
can be obtained using the function \code{validUnitsOfLengthSI()}.

\strong{3. As vectors of class \code{unit}:}

Behind the scenes \code{SPATA2} works with the \code{units} package. Input
is converted into vectors of class \code{units}. Therefore, input can be directly
provided this way: \code{arg_input = units::set_unit(x = c(2,4), value = 'mm')}
Note that \emph{pixel} is not a valid unit in the \code{units} package. If you want
to specify the input in pixel you have to use input option 1. In pixel.
}

\seealso{
\code{\link[=getImageDirectories]{getImageDirectories()}}
}
