% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{plotMolecules2D}
\alias{plotMolecules2D}
\title{Plot molecules in 2D space}
\usage{
plotMolecules2D(
  object,
  molecules,
  pt_alpha = 0.5,
  pt_size = 1,
  pt_clrp = NULL,
  clrp_adjust = NULL,
  use_scattermore = TRUE,
  xrange = getCoordsRange(object)$x,
  yrange = getCoordsRange(object)$y,
  display_facets = TRUE,
  nrow = NULL,
  ncol = NULL,
  assay_name = activeAssay(object),
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{molecules}{Character vector. The molecules of interest.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{use_scattermore}{Logical value. If \code{TRUE}, data points are plotted with
\code{scattermore::geom_scattermore()} which allows quick plotting of several
thousand data points. If the number of data points plotted is bigger than
10.000 it is used anyway.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Visualizes the positions of molecules in 2D space on the sample.
}
