% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA-E.R
\name{plotPcaElbow}
\alias{plotPcaElbow}
\title{Plot PCA Elbow Plot}
\usage{
plotPcaElbow(object, elbow = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{elbow}{Logical. If TRUE, a vertical line is added to the plot indicating the elbow point.}
}
\value{
A ggplot.
}
\description{
This function generates an elbow plot for the principal
component analysis (PCA) of the given object.
}
\details{
This function calculates the standard deviation of each principal component and plots them.
If the \code{elbow} parameter is set to TRUE, a vertical line is added at the elbow point,
which is calculated using a helper function \code{\link[=find_elbow_point]{find_elbow_point()}}.
}
