% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN-S.R
\name{plotPcaVariation}
\alias{plotPcaVariation}
\title{Plot Pca Variation}
\usage{
plotPcaVariation(object, n_pcs = NULL, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays a scree plot of the current principal component
analysis data stored in the object.
}
