% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN-S.R
\name{plotRiverplot}
\alias{plotRiverplot}
\title{Plot riverplots}
\usage{
plotRiverplot(
  object,
  grouping_variables,
  fill_by = NULL,
  strata_alpha = 0,
  strata_color = "white",
  strata_fill = "white",
  strata_width = 1/3,
  allv_color = "white",
  allv_type = "xspline",
  allv_width = 1/3,
  clrp = NULL,
  clrp_adjust = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{grouping_variables}{Character vector. Names of the grouping variables
you want to include in the riverplot.}

\item{fill_by}{Character value or NULL. If character, denotes the grouping
variable that is visualized by color (fill) of the streamlines (alluvias)
between the stratas.}

\item{strata_alpha}{Numeric value. Denotes transparency of the stratas.}

\item{strata_color}{Character value. Denotes the color used for the borders
of all strata.}

\item{strata_fill}{Character value. Denotes the color used to fill all
strata.}

\item{strata_width, allv_width}{Numeric value. Denotes the width of each stratum, as a proportion of the distance between axes.}

\item{allv_type}{Character value. Denotes the type of the curve used to produce flows. Use \code{validAlluvialTypes()}
to obtain all valid input options.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{...}{Additional arguments given to \code{scale_color_add_on()}.}
}
\value{
A ggplot.
}
\description{
Visualizes overlapping proportions of multiple grouping variables.
See details for more information.
}
\details{
For an explanation of the vocabulary and essentials of
riverplots check out the website of the package \code{ggalluvial} at
\emph{https://corybrunson.github.io/ggalluvial/articles/ggalluvial.html}.
}
