% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{plotSasEvaluation}
\alias{plotSasEvaluation}
\title{Plot SAS evaluation per variable-model pair}
\usage{
plotSasEvaluation(
  object,
  variables,
  id = idSA(object),
  method_eval = "corr",
  core = TRUE,
  distance = distToEdge(object),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_size = 1,
  line_alpha = 0.9,
  line_color = "blue",
  line_size = 1,
  display_se = FALSE,
  display_eval = FALSE,
  eval_p_min = 5e-05,
  eval_pos_x = NULL,
  eval_pos_y = NULL,
  eval_text_sep = "\\n",
  eval_text_size = 3,
  force_grid = FALSE,
  bcsp_exclude = NULL,
  ncol = NULL,
  nrow = NULL,
  make_pretty = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{display_eval}{Logical. If TRUE, evaluation values are added to the plots.}

\item{eval_p_min}{Numeric value. Everything below is displayed as \emph{<eval_p_min}.}

\item{eval_pos_x, eval_pos_y}{Numeric vector of length two. The position of
the correlation text with x- and y-coordinates.}

\item{eval_text_sep}{Character value used to separate correlation value and
corresponding p-value.}

\item{eval_text_size}{Numeric value. Size of text.}

\item{force_grid}{Logical value. If \code{TRUE}, \code{facet_grid()} is used regardless
of \code{variables} being of length 1.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Plots inferred gene expression along the distance to an image
annotation against model values.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\keyword{internal}
