% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(
  object,
  encircle = TRUE,
  params_encircle = list(),
  segment_subset = NULL,
  pt_alpha = NULL,
  pt_size = NULL,
  pt_clrp = NULL,
  clrp_adjust = NULL,
  pt_size_fixed = TRUE,
  color_by = "segmentation",
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{encircle}{Logical. If set to TRUE the segments are enclosed in a polygon.}

\item{params_encircle}{Named list of arguments given to \code{ggforce::geom_mark_hull()}.}

\item{segment_subset}{Character vector or NULL. If character vector, denotes
the segments that are supposed to be highlighted.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{...}{Additional arguments given to \code{confuns::scale_color_add_on()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the segmentation of a specified sample that was drawn with
\code{SPATA::createSegmentation()}.
}
