% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{plotSummaryIAS}
\alias{plotSummaryIAS}
\title{Visualize summary of spatial fitting}
\usage{
plotSummaryIAS(
  ias,
  x = "corr_mean",
  y = "pvalue_mean",
  pt_alpha = 0.75,
  pt_color = "black",
  pt_size = 1,
  display_labels = FALSE,
  n_labels = 10,
  var_labels = "x",
  model_subset = NULL,
  model_remove = NULL,
  ...
)
}
\description{
Visualize summary of spatial fitting
}
