% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurface}
\alias{plotSurface}
\alias{plotSurface,spata2-method}
\alias{plotSurface,data.frame-method}
\alias{plotSurfaceInteractive}
\title{Plot the surface of the sample}
\usage{
plotSurface(object, ...)

\S4method{plotSurface}{spata2}(
  object,
  color_by = NULL,
  alpha_by = NULL,
  method_gs = NULL,
  normalize = NULL,
  smooth = FALSE,
  smooth_span = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  pt_size_fixed = NULL,
  clrp_adjust = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  display_image = NULL,
  display_title = NULL,
  complete = NULL,
  verbose = NULL,
  highlight_groups = NULL,
  transform_with = NULL,
  bcsp_rm = NULL,
  na_rm = FALSE,
  pt_size_legend = 10,
  order_by = NULL,
  order_desc = FALSE,
  ...
)

\S4method{plotSurface}{data.frame}(
  object,
  color_by = NULL,
  alpha_by = NULL,
  pt_alpha = 0.9,
  pt_clr = "lightgrey",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 2,
  image = NULL,
  clrp_adjust = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  order_by = NULL,
  order_desc = FALSE,
  ...
)

plotSurfaceInteractive(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{color_by}{The variable to be displayed by color:

\itemize{
\item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
\item{ \strong{Genes} as a character vector. If more than one gene is specified the average
expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
\item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
}}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{use_scattermore}{Logical value. If \code{TRUE}, data points are plotted with
\code{scattermore::geom_scattermore()} which allows quick plotting of several
thousand data points. If the number of data points plotted is bigger than
10.000 it is used anyway.}

\item{sctm_interpolate, sctm_pixels}{Given to the corresponding arguments
of \code{scattermore::geom_scattermore()}. Note: With increasing \code{sctm_pixels}
the point size must be adjusted with the argument \code{pt_size}.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{complete}{Logical. If the provided \code{SPATA2} object has been subsetted by
\code{subsetBySegment()} the original sample is completed with grey barcode
spots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{bcsp_rm}{Character vector or NULL. If character, specifies barcode-spots that
are removed before plotting.}

\item{order_by}{Character value or \code{NULL}. If character, the specified
variable is used to order the data points.}

\item{order_desc}{Logical value. If \code{TRUE}, reverses the arrangement specified
via \code{order_by} and/or \code{order}.}

\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}
}
\description{
Displays the spatial dimension of the sample and colors the
surface according to the expression of genes, gene sets or features.

\itemize{

\item{ \code{plotSurface()} Takes the \code{SPATA2} object or a data.frame.}
\item{ \code{plotSurfaceInteractive()} Takes only the \code{SPATA2} object and opens a shiny
application which allows for interactive plotting.}

}
}
