% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceAverage}
\alias{plotSurfaceAverage}
\title{Plot several surface plots colored by gene averages}
\usage{
plotSurfaceAverage(
  object,
  color_by,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  smooth = NULL,
  smooth_span = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{color_by}{A character vector of gene names or a
named list in which each element is a vector of gene names.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\description{
Displays the spatial dimension of the sample and colors the
surface according to the expression of averaged gene expression values.
For each element in the list specified in argument \code{color_by} a
surface plot is generated colored by the gene's average expression score.
}
\examples{
#Not run:

color_by_list <- list(Example1 = c("PGK1", "PDK1", "GBE1"),
                      Example2 = c("METRN", "GFAP")
                      )

plotSurfaceAverage(object = spata_obj,
                   color_by = color_by_list)

}
