% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceComparison}
\alias{plotSurfaceComparison}
\alias{plotSurfaceComparison2}
\title{Plot several surface plots colored by numeric variables}
\usage{
plotSurfaceComparison(
  object,
  color_by,
  method_gs = NULL,
  normalize = NULL,
  smooth = NULL,
  smooth_span = NULL,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  display_image = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotSurfaceComparison2(
  coords_df,
  color_by = NULL,
  pt_alpha = 0.9,
  pt_clrsp = "inferno",
  pt_size = 2,
  image = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.
displayed.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}

\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}
}
\description{
Displays a surface plot for every variable specified
in argument \code{color_by}.

\itemize{
 \item{ \code{plotSurfaceComparison()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotSurfaceComparison2()} Takes a data.frame as the starting point. }
 }
}
