% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotSurfaceHotspots}
\alias{plotSurfaceHotspots}
\title{Title}
\usage{
plotSurfaceHotspots(
  object,
  plot_type = "density_2d",
  clrp = "inferno",
  of_sample = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
