% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{plotSurfaceIAS}
\alias{plotSurfaceIAS}
\alias{plotSurfaceIAS,spata2-method}
\alias{plotSurfaceIAS,ImageAnnotationScreening-method}
\title{Visualize screening are of IAS-algorithm}
\usage{
plotSurfaceIAS(object, ...)

\S4method{plotSurfaceIAS}{spata2}(
  object,
  id,
  buffer,
  n_bins_circle,
  n_bins_angle = 12,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = NULL,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  direction = -1,
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)

\S4method{plotSurfaceIAS}{ImageAnnotationScreening}(
  object,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = 2.25,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  direction = -1,
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)
}
\description{
Visualize screening are of IAS-algorithm
}
