% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurfaceQuantiles}
\alias{plotSurfaceQuantiles}
\title{Deprecated}
\usage{
plotSurfaceQuantiles(
  object,
  color_by,
  alpha_by = NULL,
  n_qntls = 5,
  keep_qntls = 1:n_qntls,
  pt_alpha = NULL,
  pt_clrp = NULL,
  pt_size = NULL,
  smooth = NULL,
  smooth_span = NULL,
  display_image = NULL,
  bcsp_rm = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  verbose = NULL,
  ...
)
}
\description{
Deprecated in favor of \code{\link[=plotSurface]{plotSurface()}}.
}
\keyword{internal}
