% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurfaceSC}
\alias{plotSurfaceSC}
\title{Plot single cells on surface}
\usage{
plotSurfaceSC(
  object,
  sc_input,
  cell_types = NULL,
  line_alpha = 0.9,
  line_size = 1,
  pt_alpha = 1,
  pt_size = 2,
  display_density = TRUE,
  display_facets = TRUE,
  display_image = FALSE,
  display_points = TRUE,
  clrp = NULL,
  clrp_adjust = NULL,
  frame_by = "coords",
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{sc_input}{Data.frame that contains the results from single cell deconvolution.
Must have at least three columns:
\itemize{
\item{\emph{x} : }{ numeric. Position of cell on the x axis in pixel.},
\item{\emph{y} :}{ numeric. Position of cell on the y axis in pixel.},
\item{\emph{cell_type} :}{ character. Cell type of the cell.}
}}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{display_density}{Logical value. If \code{TRUE}, uses \code{ggplot2::geom_density2d()}
to visualize cell density with contours.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{display_image}{Logical value. If \code{TRUE}, adds the image of the
tissue.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{cell_type}{Character value or \code{NULL}. If character,
subsets the cell types that are included in the plots.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Plots single cell input on the sample surface.
}
\examples{

object <- loadPubExample("MCD_LMU")

data("sc_deconvolution")

plotSurfaceSC(
 object = object,
 sc_input = sc_deconvolution$MCD_LMU,
 cell_types = c("Astrocytes", "Neurons", "Microglia", "Macrophages/Monocytes"),
 clrp = "sifre"
 )

}
