% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-gradient-screening.R
\name{plotSurfaceSGS}
\alias{plotSurfaceSGS}
\alias{plotSurfaceSGS,spata2-method}
\alias{plotSurfaceSGS,SpatialGradientScreening-method}
\title{Visualize screening are of SGS-algorithm}
\usage{
plotSurfaceSGS(object, ...)

\S4method{plotSurfaceSGS}{spata2}(
  object,
  id,
  n_circles,
  buffer,
  n_angle_bins = 12,
  pt_alpha = NA_integer_,
  pt_clrp = "inferno",
  pt_size = NULL,
  color_core = ggplot2::alpha("grey", 0),
  direction = -1,
  display_image = FALSE,
  circle_plot = TRUE,
  angle_plot = TRUE,
  ggpLayers = list(),
  ...
)

\S4method{plotSurfaceSGS}{SpatialGradientScreening}(
  object,
  pt_clrp = "inferno",
  pt_size = 2.25,
  color_core = ggplot2::alpha("grey", 0),
  direction = -1,
  circle_plot = TRUE,
  angle_plot = TRUE,
  ggpLayers = list(),
  ...
)
}
\description{
Visualize screening are of SGS-algorithm
}
