% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{plotTrajectoryBarplot}
\alias{plotTrajectoryBarplot}
\title{Deprecated}
\usage{
plotTrajectoryBarplot(...)
}
\description{
Deprecated in favor of \code{\link[=plotStsBarplot]{plotStsBarplot()}}
}
\keyword{internal}
