% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotT-Z.R
\name{plotTrajectoryEvaluation}
\alias{plotTrajectoryEvaluation}
\title{Plot STS evaluation per variable-model pair}
\usage{
plotTrajectoryEvaluation(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins_circle = NA_integer_,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_size = 1,
  line_alpha = 0.9,
  line_color = "blue",
  line_size = 1,
  display_se = FALSE,
  display_corr = FALSE,
  corr_p_min = 5e-05,
  corr_pos_x = NULL,
  corr_pos_y = NULL,
  corr_text_sep = "\\n",
  corr_text_size = 1,
  force_grid = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{binwidth}{Distance value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the center-center
distance: \code{binwidth = getCCD(object)}. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
(See details for more.)}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{display_corr}{Logical. If TRUE, correlation values are added to the plots.}

\item{corr_p_min}{Numeric value. Everything below is displayed as \emph{<corr_p_min}.}

\item{corr_pos_x, corr_pos_y}{Numeric vector of length two. The position of
the correlation text with x- and y-coordinates.}

\item{corr_text_sep}{Character value used to separate correlation value and
corresponding p-value.}

\item{corr_text_size}{Numeric value. Size of text.}

\item{force_grid}{Logical value. If \code{TRUE}, \code{facet_grid()} is used regardless
of \code{variables} being of length 1.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Plots inferred gene expression along a spatial trajectory
against model values.
}
