% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{plotTrajectoryFit}
\alias{plotTrajectoryFit}
\alias{plotTrajectoryFitCustomized}
\title{Deprecated}
\usage{
plotTrajectoryFit(
  object,
  trajectory_name = getDefaultTrajectory(object, verbose = TRUE, "trajectory_name"),
  variable,
  binwidth = 5,
  method_gs = NULL,
  smooth = NULL,
  smooth_span = NULL,
  smooth_se = FALSE,
  linealpha = 0.75,
  linesize = 1,
  lineorder = c(1, 2, 3),
  display_residuals = NULL,
  display_auc = FALSE,
  auc_alpha = 0.5,
  auc_linetype = "dotted",
  display_auc_text = FALSE,
  colors = c("forestgreen", "blue4", "red3"),
  model_subset = validTrajectoryTrends(),
  ref_model = "Model",
  nrow = NULL,
  ncol = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotTrajectoryFitCustomized(
  object,
  trajectory_name = getDefaultTrajectory(object, verbose = TRUE, "trajectory_name"),
  variable,
  customized_trends,
  binwidth = 5,
  method_gs = NULL,
  smooth = NULL,
  smooth_span = NULL,
  linealpha = 0.75,
  linesize = 1,
  lineorder = c(1, 2, 3),
  display_residuals = NULL,
  display_auc = FALSE,
  auc_alpha = 0.5,
  auc_linetype = "dotted",
  colors = c("forestgreen", "blue4", "tomato"),
  ref_model = "Model",
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\description{
Deprecated
}
