% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{plotTrajectoryHeatmap}
\alias{plotTrajectoryHeatmap}
\title{Deprecated}
\usage{
plotTrajectoryHeatmap(...)
}
\description{
Deprecated in favor of \code{\link[=plotStsHeatmap]{plotStsHeatmap()}}
}
\keyword{internal}
