% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R, R/plotT-Z.R
\name{plotTrajectoryLineplot}
\alias{plotTrajectoryLineplot}
\alias{plotTrajectoryRidgeplot}
\title{Deprecated}
\usage{
plotTrajectoryLineplot(...)

plotTrajectoryRidgeplot(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  unit = getSpatialMethod(object)@unit,
  round = 2,
  method_gs = NULL,
  smooth_method = "loess",
  smooth_span = 0.2,
  smooth_se = TRUE,
  clrp = NULL,
  clrp_adjust = NULL,
  display_trajectory_parts = NULL,
  alpha = 0.9,
  fill = NULL,
  line_color = "black",
  line_size = 1.5,
  vlinecolor = "grey",
  vlinesize = 1,
  vlinetype = "dashed",
  x_nth = 7L,
  xi = NULL,
  yi = NULL,
  expand_x = c(0, 0),
  summarize_with = "mean",
  ncol = 1,
  nrow = NULL,
  overlap = 0.5,
  strip_pos = "right",
  display_model = NULL,
  verbose = NULL,
  ...
)
}
\description{
Deprecated in favor of \code{\link[=plotStsLineplot]{plotStsLineplot()}}
}
\keyword{internal}
