% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{plotTrajectoryLineplotFitted}
\alias{plotTrajectoryLineplotFitted}
\title{Title}
\usage{
plotTrajectoryLineplotFitted(
  object,
  id = getDefaultTrajectory(object),
  variable,
  binwidth = 5,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  method_gs = NULL,
  smooth = FALSE,
  smooth_span = 0.2,
  lineorder = c(1, 2, 3),
  linesize = 1,
  linecolors = c("forestgreen", "blue4", "red3"),
  linetypes = c("solid", "solid", "dotted"),
  display_residuals = NULL,
  area_alpha = 0.25,
  nrow = NULL,
  ncol = NULL,
  verbose = NULL
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
