% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{plotVolcano}
\alias{plotVolcano}
\alias{plotVolcano,ImageAnnotationScreening-method}
\alias{plotVolcano,SpatialTrajectoryScreening-method}
\title{Compare evaluation of spatially opposing fits}
\usage{
plotVolcano(object, ...)

\S4method{plotVolcano}{ImageAnnotationScreening}(
  object,
  eval = "corr_mean",
  pval = "p_value_mean",
  left = "linear_ascending",
  right = "linear_descending",
  display_thresholds = TRUE,
  threshold_eval = 0.5,
  threshold_pval = 0.05,
  threshold_colors = c("tomato", "lightgrey"),
  label_vars = NULL,
  label_alpha = 0.9,
  label_color = "black",
  label_size = 2,
  negative_log = TRUE,
  pt_alpha = 0.9,
  pt_size = 1,
  display_names = TRUE,
  hstep = 1.5,
  vstep = 1.2,
  best_only = FALSE,
  ...
)

\S4method{plotVolcano}{SpatialTrajectoryScreening}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{eval}{Character value. The variable to use for the x-axis.}

\item{pval}{Character value. The variable to use for the y-axis.}

\item{left, right}{Character value. The name of the model whose best-fit variables
go to the left or to the right, respectively. Defaults to \code{left} = \emph{'linear_ascending'}
and \code{right} = \emph{'linear_descending'}.}

\item{threshold_pval, threshold_eval}{Numeric values that set the thresholds below/above
which the points are highlighted.}

\item{threshold_colors}{Character vector of length two. First denotes
the color of the significant variables, second denotes the color
of the not-significant variables.}

\item{label_vars}{Character value, numeric value or NULL. Useful to highlight
the exact position/evalation of variables.

If character, specifies the variables that are labeled. If numeric, specifies
the top n of variables that are labeled. If NULL, ignored.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{hstep, vstep}{Adjust the position of the two labels that show the
model names on the left and on the right.}

\item{best_only}{Logical value. If TRUE, only variables are included in
the plot that have their best model fit in either the left or the right
model.}

\item{display_threshold}{Logical value. If TRUE, the thresholds set by
\code{treshold_pval} and \code{threshold_eval} are used to color the points
of the plot.}
}
\description{
Plots a volcano plot by using the model evaluation
of spatial fitting as implemented by \code{imageAnnotationScreening()}
and \code{spatialTrajectoryScreening()}.
}
