% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{prepare_coords_df_visium_hd}
\alias{prepare_coords_df_visium_hd}
\title{Prepare coordinates data.frame for decreasing resolution}
\usage{
prepare_coords_df_visium_hd(coords_df, fct)
}
\arguments{
\item{coords_df}{A data frame containing the original spatial coordinates.
It must include columns \code{row}, \code{col}, \code{x_orig}, and \code{y_orig} representing the
spatial grid positions and original coordinates.}

\item{fct}{An integer factor by which the dimensions of the grid should be divisible.
This ensures the grid can be evenly divided into larger squares.}
}
\value{
A data frame with adjusted spatial coordinates, new row and column groupings,
predicted coordinates, and updated barcodes.
}
\description{
This function prepares a data frame of spatial coordinates for Visium HD
analysis by adjusting the coordinates to form a grid of evenly sized squares
and grouping them into larger aggregate squares. It ensures the grid dimensions
are divisible by a given factor, and it predicts missing coordinates.
}
\details{
The function adjusts the input coordinates to form a perfect rectangular
grid of squares and groups them into larger squares based on the provided factor (\code{fct}).
It ensures that both the row and column counts are divisible by \code{fct} by expanding
the grid if necessary. The function then assigns new groupings to these aggregated
squares and predicts missing \code{x} and \code{y} coordinates using linear models based
on the original coordinates. The output is a data frame that includes these new
groupings and predicted coordinates.

The main steps include:
\itemize{
\item Calculating scaling factors for the columns and rows.
\item Adjusting the grid to ensure divisibility by \code{fct} and equal lengths of rows and columns.
\item Creating new row and column groups and generating corresponding barcodes.
\item Predicting missing \code{x} and \code{y} coordinates using linear models.
}
}
\keyword{internal}
