% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{printAutoencoderSummary}
\alias{printAutoencoderSummary}
\title{Print autoencoder summary}
\usage{
printAutoencoderSummary(object, mtr_name = "denoised", of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A human readable report of the issue of interest.
}
\description{
Prints a human readable summary about the set up of the last neural network that
was constructed to generate a denoised expression matrix.
}
\keyword{internal}
