% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-family.R
\name{printDeaOverview}
\alias{printDeaOverview}
\title{Print overview of all conducted de-analysis}
\usage{
printDeaOverview(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A human readable report of the issue of interest.
}
\description{
Print overview of all conducted de-analysis
}
