% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{printDefaultInstructions}
\alias{printDefaultInstructions}
\title{Print current default settings}
\usage{
printDefaultInstructions(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
A human readable report of the issue of interest.
}
\description{
Print current default settings
}
\keyword{internal}
