% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{printGeneSetOverview}
\alias{printGeneSetOverview}
\title{Print overview about the current gene sets}
\usage{
printGeneSetOverview(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A human readable report of the issue of interest.
}
\description{
Print overview about the current gene sets
}
