% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{process_expand_input}
\alias{process_expand_input}
\title{Process expand input}
\usage{
process_expand_input(expand)
}
\value{
Returns always a list of length two. Two slots named h (height)
and x (width).
}
\description{
Process expand input
}
\keyword{internal}
