% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{process_ranges}
\alias{process_ranges}
\title{Process input ranges}
\usage{
process_ranges(
  xrange = getImageRange(object)$x,
  yrange = getImageRange(object)$y,
  expand = 0,
  persp = "ccs",
  object = NULL,
  ranges = NULL,
  opt = 1
)
}
\arguments{
\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{expand}{Parameter to adjust how the image is expanded. See section
Image expansion for more information.}

\item{persp}{Determines the perspective of the output. Thus, if \emph{image}, it will flip
the values of \code{yrange}. If \emph{'ccs'}, the values of \code{yrange} are not flipped.}

\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
List of 4 slots. Named \emph{xmin}, \emph{xmax}, \emph{ymin} and \emph{ymax}. Adjusted range
in pixel.
}
\description{
Processes x- and y-ranges. The function assumes that x- and
yrange are given from the perspective of a cartesian coordinate system.
}
\keyword{internal}
