% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{process_seurat_object}
\alias{process_seurat_object}
\title{Wrapper around Seurat processing functions}
\usage{
process_seurat_object(
  seurat_object,
  assay_name = NULL,
  calculate_rb_and_mt = TRUE,
  SCTransform = FALSE,
  NormalizeData = TRUE,
  FindVariableFeatures = TRUE,
  ScaleData = TRUE,
  RunPCA = TRUE,
  FindNeighbors = TRUE,
  FindClusters = TRUE,
  RunTSNE = TRUE,
  RunUMAP = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{A valid seurat-object.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A processed seurat-object.
}
\description{
Wrapper around Seurat processing functions
}
\keyword{internal}
