% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{PseudoHistoImage}
\alias{PseudoHistoImage}
\title{Pseudo \code{HistoImage}}
\format{
An object of class \code{HistoImage} of length 1.
}
\usage{
PseudoHistoImage
}
\description{
A \code{\link{HistoImage}} object that serves as the container for some
spatial related functions that are applicable to spatial methods even if they
do not come with an image (e.g. MERFISH, SlideSeq).

This is necessary, as \code{SPATA2} stores a variety of spatial information
in the \code{\link{HistoImaging}} and \code{\link{HistoImage}} containers. Therefore, code relies
on the presence of a \code{HistoImage} object inside the \code{spata2} object even though
there is no actual image of the tissue.
}
\keyword{datasets}
