% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{read_coords_visium}
\alias{read_coords_visium}
\title{Read coordinate data.frames}
\usage{
read_coords_visium(dir_coords)
}
\arguments{
\item{dir_coords}{Character value. Directory to the coordinates data.frame.}
}
\value{
Data.frame of four columns:
\itemize{
\item{\emph{id}:}{ Character. Unique identifier of each observation.}
\item{\emph{exclude}:}{ Logical. Indicates whether to exclude the observation by default.}
\item{\emph{exclude_reason}:}{ Character. The reason for why to exclude the observation.}
\item{\emph{x_orig}:}{ Numeric. x-coordinates of the original input.}
\item{\emph{y_orig}:}{ Numeric. y-coordinates of the original input.}
}
}
\description{
Reads in coordinates data.frame from various platforms.
}
