% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{read_matrix_from_folder}
\alias{read_matrix_from_folder}
\title{Read Matrix from Folder}
\usage{
read_matrix_from_folder(dir)
}
\arguments{
\item{dir}{Character. The directory containing the matrix, barcodes, and features files.}
}
\value{
A sparse matrix with barcodes as column names and features as row names.
}
\description{
This function reads a matrix, barcodes, and features from a specified directory
and returns the matrix with appropriate row and column names.
}
\details{
The specified directory must contain the following files:
\itemize{
\item{Matrix file}: A file with the extension \code{.mtx.gz} or \code{.mtx}. This file contains the count matrix in Matrix Market format.
\item{Barcodes file}: A file with the name \code{barcodes.tsv.gz} or \code{barcodes.tsv}. This file contains the barcodes for the columns of the matrix.
\item{Features file}: A file with the name \code{features.tsv.gz} or \code{features.tsv}. This file contains the features (e.g., gene names) for the rows of the matrix.
}

The function will search for these files in the specified directory and read them using appropriate functions. The matrix will be returned with barcodes as column names and features as row names.
}
\examples{
\dontrun{
  matrix_dir <- "path/to/matrix/folder"
  matrix <- read_matrix_from_folder(matrix_dir)
  print(matrix)
}

}
