% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{recBinwidth}
\alias{recBinwidth}
\title{Platform dependent binwidth recommendation}
\usage{
recBinwidth(object, unit = getDefaultUnit(object))
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Distance measure.
}
\description{
Recommends a binwidth parameter for the spatial screening algorithms
based on the platform used.
}
\details{
For objects derived from the Visium platform we recommend a binwidth equal
to the center to center distance as obtained by \code{getCCD()}.

For objects derived from platforms that do not rely on a fixed grid of
data points (MERFISH, SlideSeq, etc.) we recommend the average minimal
distance between the data points.

\code{recBinwidth()} is a wrapper around these recommendations.
}
