% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{recDbscanEps}
\alias{recDbscanEps}
\alias{recDbscanMinPts}
\title{DBSCAN parameter recommendations}
\usage{
recDbscanEps(object)

recDbscanMinPts(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeric value in case of \code{recDbscanMinPts()}. Distance measure
in case of \code{recDbscanEps()}.
}
\description{
Suggests a value for DBSCAN applications within \code{SPATA2}.
}
\details{
For objects derived from the Visium platform with a fixed center to center
distance, we recommend to set \code{eps = getCCD(object, unit = "px")*1.25}
and \code{minPts = 3}.

For objects derived from platforms that do not rely on a fixed grid of
data points (MERFISH, SlideSeq, etc.) we recommend the average minimal
distance between the data points times 10 for \code{eps} and \code{minPts = 12}.

\code{recDbscanEps()} and \code{recDbscanMinPts()} are wrappers around these recommendations.
}
