% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{recDbscanEps}
\alias{recDbscanEps}
\alias{recDbscanMinPts}
\alias{recSgsRes}
\title{Platform dependent input recommendations}
\usage{
recDbscanEps(object)

recDbscanMinPts(object)

recSgsRes(object, unit = getDefaultUnit(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}
}
\value{
Single values of different classes depending on the function. See details for more.
}
\description{
A collection of functions that return the recommended default input
of certain arguments depending on the \link[=SpatialMethod]{spatial method}
(the platform) the \code{SPATA2} object derived from.
}
\details{
\itemize{
\item{\code{recDbscanEps}}{ The default input for the \code{eps} argument of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}.
For objects derived from the Visium platform, we recommend
a binwidth equal to the center-to-center distance as obtained by \code{\link[=getCCD]{getCCD()}} multiplied
by \emph{1.25}.
For objects derived from platforms that do not rely on a fixed grid of data points
(MERFISH, SlideSeq, etc.), we recommend the average minimal distance between the
data points multiplied by \emph{1.25}. }
\item{\code{recDbscanMinPts}}{ The default input for the \code{minPts} argument of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}.
For objects derived from the Visium platform, we recommend \code{minPts = 3}.
For objects derived from platforms that do not rely on a fixed grid of data points
(MERFISH, SlideSeq, etc.), we recommend \code{minPts = 12}. }
\item{\code{recSgsRes()}}{ The default input for the \code{resolution} argument of the spatial
gradient screening algorithm. For objects derived from the Visium platform, we recommend
a binwidth equal to the center-to-center distance as obtained by \code{\link[=getCCD]{getCCD()}}.
For objects derived from platforms that do not rely on a fixed grid of data points
(MERFISH, SlideSeq, etc.), we recommend the average minimal distance between the
data points.}
}
}
