% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{reduce_coords_df_visium_hd}
\alias{reduce_coords_df_visium_hd}
\title{Reduce resolution coordinates data.frame}
\usage{
reduce_coords_df_visium_hd(coords_df_red, fct)
}
\arguments{
\item{coords_df_red}{The output data.frame of the \code{prepare_coords_df_visium_hd()} function.}
}
\value{
A summarized data frame with one row per new barcode, containing aggregated spatial
coordinates and counts.
}
\description{
This function reduces a data frame of spatial coordinates for Visium HD samples
by summarizing and aggregating the data at the new barcode level.
}
\details{
The function aggregates the spatial coordinates based on the new barcodes created in the previous step.
It calculates summary statistics for each new barcode group, including:

\itemize{
\item \code{x_orig}, \code{y_orig}: The original \code{x} and \code{y} coordinates at the new barcode level.
\item \code{col}, \code{row}: The numeric values of the column and row groups.
\item \code{square_exp}: The theoretical number of original spots that could fall into the new aggregated spot.
\item \code{square_count}: The actual number of non-missing spots observed.
\item \code{square_perc}: The relative number of observed spots as a proportion of the theoretical number.
}
}
\keyword{internal}
