% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{reduceSpataObject}
\alias{reduceSpataObject}
\title{Reduce SPATA2 object}
\usage{
reduceSpataObject(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function reduces a \code{\link{SPATA2}} object to a minimal version by
removing analysis progress and other non-essential data.
}
\details{
The following components are removed or reduced:
\itemize{
\item{@assays:}{ The @analysis and @mtr_proc slots are cleared, and @meta_var is reduced to only include molecule names.}
\item{@data_add:}{ Cleared completely.}
\item{@dim_red:}{ Cleared completely.}
\item{@meta:}{ Only the \emph{barcodes} column is retained in the metadata.}
\item{@spatial:}{
\itemize{
\item{@annotations:}{ Cleared completely.}
\item{@images:}{ The @outline and @pixel_content slots are cleared.}
\item{@trajectories:}{ Cleared completely.}
\item{@outline:}{ Cleared completely.}
}
}

Furthermore, all images are unload with \code{\link[=unlaodImages]{unlaodImages()}}.

}
}
\examples{

# Assuming 'object' is a valid SPATA2 object
reduced_object <- reduceSpataObject(object)

}
