% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ref}
\alias{Ref}
\title{A list with a reference dataset [[1]] and the annotations [[2]].}
\format{
An object of class \code{list} with 2 data.frames
}
\usage{
data(Ref)
}
\description{
A list with a reference dataset [[1]] and the annotations [[2]].
}
\keyword{datasets}