% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{refImage}
\alias{refImage}
\alias{refImage,SPATA2-method}
\alias{refImage,SpatialData-method}
\title{Obtain name of reference content}
\usage{
refImage(object, ...)

\S4method{refImage}{SPATA2}(object)

\S4method{refImage}{SpatialData}(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character value.
}
\description{
Handy functions to quickly access the name of reference content.
}
