% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{registerImage}
\alias{registerImage}
\alias{registerImage,HistoImaging-method}
\title{Register images as \code{HistoImage}}
\usage{
registerImage(object, ...)

\S4method{registerImage}{HistoImaging}(object, dir, img_name, unload = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{dir}{Character value. The directory from where to retrieve the image.}

\item{img_name}{Character value. The name of the \code{HistoImage} with which
to refer to it via arguments \code{img_name} and \code{img_names}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Creates and registers new images by putting them in relation
to the reference image of the object.
}
