% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{relateToSpatialAnnotation}
\alias{relateToSpatialAnnotation}
\title{Relate observations to a spatial annotation}
\usage{
relateToSpatialAnnotation(
  object,
  id,
  input_df,
  input_id_var = NULL,
  distance = NA_integer_,
  binwidth = NA_integer_,
  n_bins_circle = NA_integer_,
  n_bins_angle = 12,
  calc_dist_to = "border",
  unit = "px",
  inc_outline = TRUE,
  verbose = NULL,
  ...
)
}
\description{
Deprecated in favor of \code{getCoordsDfSA(..., coords_df = input_df)}
}
\keyword{internal}
