% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeAnnotation}
\alias{removeAnnotation}
\title{Remove annotation}
\usage{
removeAnnotation(object, ann_var, groups)
}
\arguments{
\item{ann_var}{Character value. The annotation variable that contains
the barcode spot annotations you want to alter.}

\item{groups}{Character vector. The annotation / group names you want
to remove.}
}
\value{
An updated spata object.
}
\description{
Removes annotations within annotation variables.
}
\details{
As the default within every annotation variable is \emph{'unnamed'}
removing the annotation effectively renames the annotation back to \emph{'unnamed'}.
}
\examples{

  object <- createAnnotation(object)

  object <- removeAnnotation(object, ann_var = "cns_layer", groups = c("layer_1", "layer2")
}
