% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{removeFeatures}
\alias{removeFeatures}
\title{Discard features}
\usage{
removeFeatures(object, feature_names, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_names}{Character vector. Specifies the features to be discarded.}
}
\value{
An updated spata-object.
}
\description{
Discards the features of choice.
}
