% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeGenes}
\alias{removeGenes}
\alias{removeGenesStress}
\alias{removeGenesZeroCounts}
\title{Remove genes from the \code{spata2} object}
\usage{
removeGenes(object, genes, show_warnings = FALSE, verbose = NULL)

removeGenesStress(object, verbose = NULL)

removeGenesZeroCounts(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{genes}{Character vector. Names of genes to remove.}

\item{show_warnings}{Logical value. If \code{TRUE}, \strong{warnings} about genes that were not found
although they were mentioned in the vector of genes that are to be discarded
are suppressed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Functions that removes genes from the \code{spata2} object by removing
them from count matrix and all processed matrices of the assay.

\itemize{
\item{\code{removeGenes()}:}{ Removes user specified genes.}
\item{\code{removeGenesZeroCounts()}:}{ Removes genes that do not have a single count
across all data points.}
\item{\code{removeGenesStress()}:}{ Removes mitochondrial and stress related genes.}
}
}
\details{
This step affects the matrices of the object and thus all subsequent
analysis steps. Analysis steps that have already been conducted are not affected!
This includes clustering, DEA, GSEA etc. It is advisable to integrate this
step as early as possible in the processing pipeline.
}
