% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeMolecules}
\alias{removeMolecules}
\alias{removeGenes}
\alias{removeGenesStress}
\alias{removeGenesZeroCounts}
\title{Remove molecules from the \code{SPATA2} object}
\usage{
removeMolecules(
  object,
  molecules,
  show_warnings = FALSE,
  ref = "molecule",
  verbose = NULL
)

removeGenes(object, genes, show_warnings = FALSE, verbose = NULL)

removeGenesStress(object, verbose = NULL)

removeGenesZeroCounts(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{show_warnings}{Logical value. If \code{TRUE}, \strong{warnings} about genes that were not found
although they were mentioned in the vector of genes that are to be discarded
are suppressed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{genes}{Character vector. Names of genes to remove.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Functions that removes molecules from the \code{SPATA2} object by removing
them from count matrix and all processed matrices of the respective \link[MolecularAssay]{assay}.

\itemize{
\item{\code{removeMolecules()}:}{ Removes user specified molecules}
\item{\code{removeMoleculesZeroCounts()}:}{ Removes molecules that do not have a single count
across all observations.}
}

Wrappers for transriptomic assay:

\itemize{
\item{\code{removeGenes()}:}{ Removes user specified genes.}
\item{\code{removeGenesZeroCounts()}:}{ Removes genes that do not have a single count
across all observations.}
\item{\code{removeGenesStress()}:}{ Removes mitochondrial and stress related genes.}
}
}
\details{
This step affects the matrices of the object and thus all subsequent
analysis steps. Analysis steps that have already been conducted are not affected!
This includes clustering, DEA, GSEA etc. It is advisable to integrate this
step as early as possible in the processing pipeline.
}
