% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeObsNoCounts}
\alias{removeObsNoCounts}
\title{Remove observations with no counts}
\usage{
removeObsNoCounts(object, assay_name = activeAssay(object), verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Identifies and removes observations with no molecule counts
from the \code{SPATA2} object.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

# the function tells you if / how many observations were removed
object <- removeObsNoCounts(object, verbose = TRUE)

}
