% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeProcessedMatrix}
\alias{removeProcessedMatrix}
\title{Remove a processed matrix}
\usage{
removeProcessedMatrix(object, mtr_name, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Removes a processed matrix from the \code{SPATA2} object.
}
