% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeSignatures}
\alias{removeSignatures}
\title{Remove molecular signatures}
\usage{
removeSignatures(object, signatures, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{signatures}{Character vector. The names of the signatures to be removed from
the assay of choice.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes \link[=concept_molecular_signatures]{molecular signatures} from the \code{SPATA2} object.
}
\note{
GSEA results remain unaffected.
}
