% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeSpatialOutliers}
\alias{removeSpatialOutliers}
\title{Remove spatial outliers}
\usage{
removeSpatialOutliers(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes data points that were identified as spatial outliers
and all their related data. If no spatial outliers exist, the input object
is returned as is.
}
\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

# spatial outliers have not been labeled histologically (= NA)
plotSurface(object, color_by = "histology")

object <- identifyTissueOutline(object) # step 1

plotSurface(object, color_by = "tissue_section")

object <- identifySpatialOutliers(object) # step 2

plotSurface(object, color_by = "sp_outlier")

nObs(object) # before removal

object <- removeSpatialOutliers(object) # step 3

plotSurface(object, color_by = "histology")

nObs(object) # after removal


}
\seealso{
\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}, \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}, \code{\link[=containsSpatialOutliers]{containsSpatialOutliers()}}
}
