% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeSpatialTrajectories}
\alias{removeSpatialTrajectories}
\title{Remove spatial trajectories}
\usage{
removeSpatialTrajectories(object, ids)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{id}{Character vector. The IDs of the spatial trajectories to remove.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes spatial trajectories from the SPATA2 object.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

getTrajectoryIds(object)

object <- removeSpatialTrajectories(object, ids = "horizontal_mid")

getTrajectoryIds(object)

}
