% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeTissueFragments}
\alias{removeTissueFragments}
\title{Remove data points from tissue fragments}
\usage{
removeTissueFragments(object, fragments = NULL, fdb_fn = "message")
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{fragments}{Numeric vector, character vector or \code{NULL}. If \code{NULL},
all tissue fragments are removed. If numeric or character indicates the
fragments to be removed.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes data points that fall on tissue fragments as
identified by \code{\link[=identifyTissueOutline]{identifyTissueOutline()}} and \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}
}
\seealso{
\code{\link[=removeSpatialOutliers]{removeSpatialOutliers()}}, \code{\link[=identifyTissueOutline]{identifyTissueOutline()}} and \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}
}
