% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameImage}
\alias{renameImage}
\alias{renameImage,SPATA2-method}
\alias{renameImage,SpatialData-method}
\title{Rename an image}
\usage{
renameImage(object, ...)

\S4method{renameImage}{SPATA2}(object, img_name, new_img_name, ...)

\S4method{renameImage}{SpatialData}(object, img_name, new_img_name, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image to be renamed.}

\item{new_img_name}{Character value. The new name of the image.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Renames an image.
}
\examples{

library(SPATA2)

object <- example_data$object_UKF275T_diet

getImageNames(object)

plotImage(object, img_name = "normres") # fails, does not exist

object <- renameImage(object, img_name = "image1", new_img_name = "normres")

plotImage(object, img_name = "normres")

}
