% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameImgAnn}
\alias{renameImgAnn}
\title{Rename image annotation ID}
\usage{
renameImgAnn(object, id, new_id)
}
\arguments{
\item{id}{Character value. The current ID of the image annotation to be
renamed.}

\item{new_id}{Character value. The new ID of the image annotation.}

\item{inherit}{argument_dummy params}
}
\value{
An updates spata object.
}
\description{
Renames image annotation created with \code{createImageAnnotations()}.
}
