% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameMetaFeatures}
\alias{renameMetaFeatures}
\title{Rename features}
\usage{
renameMetaFeatures(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{The features to be renamed specified according to the following
syntax: \emph{'new_feature_name'} \code{=} \emph{'old_feature_name'}.}
}
\value{
An upated spata-object.
}
\description{
Allows to rename features stored inside the @fdata slot.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

getFeatureNames(object)

object <- renameMetaFeatures(object, "histology_new" = "histology")

getFeatureNames(object)

plotSurface(object, color_by = "histology") # fails
potSurface(object, color_by = "histology_new")

}
