% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameSpatAnn}
\alias{renameSpatAnn}
\title{Rename spatial annotation ID}
\usage{
renameSpatAnn(object, id, new_id)
}
\arguments{
\item{id}{Character value. The current ID of the spatial annotation to be
renamed.}

\item{new_id}{Character value. The new ID of the spatial annotation.}

\item{inherit}{argument_dummy params}
}
\value{
An updates spata object.
}
\description{
Renames spatial annotation created with \code{createImageAnnotations()}.
}
