% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameSpataObject}
\alias{renameSpataObject}
\title{Rename the object}
\usage{
renameSpataObject(object, sample_name)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{sample_name}{Character value. The new sample name.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Renames the \code{\link{SPATA2}} object.
}
\details{
Sets slot @sample of all S4 classes within the \code{SPATA2} object
to the new name.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

object <- renameSpataObject(object, sample_name = "my_new_name")

}
