% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameSpatialAnnotation}
\alias{renameSpatialAnnotation}
\title{Rename a Spatial Annotation}
\usage{
renameSpatialAnnotation(object, id, new_id, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The current ID of the spatial annotation to be
renamed.}

\item{new_id}{Character value. The new ID of the spatial annotation.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{inherit}{argument_dummy params}
}
\description{
Renames spatial annotation.
}
