% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameSpatialTrajectory}
\alias{renameSpatialTrajectory}
\title{Rename a spatial trajectory.}
\usage{
renameSpatialTrajectory(object, id, new_id, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The current ID of the spatial trajectory to be
renamed.}

\item{new_id}{Character value. The new ID of the spatial trajectory.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{inherit}{argument_dummy params}
}
\description{
Renames spatial trajectory.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

plotSpatialTrajectories(object)

object <- renameSpatialTrajectory(object, id = "horizontal_mid", new_id = "Horizontal_Mid_Cap")

plotSpatialTrajectories(object)

}
