% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{resetImageJustification}
\alias{resetImageJustification}
\title{Deprecated}
\usage{
resetImageJustification(object)
}
\description{
Deprecated in favor of \code{\link[=resetImageTransformations]{resetImageTransformations()}}.
}
\keyword{internal}
