% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{resetImageTransformations}
\alias{resetImageTransformations}
\alias{resetImageTransformations,SPATA2-method}
\alias{resetImageTransformations,SpatialData-method}
\alias{resetImageTransformations,HistoImage-method}
\title{Reset image transformations}
\usage{
resetImageTransformations(object, ...)

\S4method{resetImageTransformations}{SPATA2}(object, img_name, ...)

\S4method{resetImageTransformations}{SpatialData}(object, img_name, ...)

\S4method{resetImageTransformations}{HistoImage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Resets the transformation values of an image defined
by usage of \code{\link[=alignImage]{alignImage()}}, \code{\link[=alignImageAuto]{alignImageAuto()}} or \code{\link[=alignImageInteractive]{alignImageInteractive()}}.
}
\seealso{
\code{\link[=getImageTransformations]{getImageTransformations()}}
}
