% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{rotate_coords_df}
\alias{rotate_coords_df}
\title{Rotate coordinate variables pairs}
\usage{
rotate_coords_df(
  df,
  angle,
  ranges,
  clockwise = TRUE,
  coord_vars = list(pair1 = c("x", "y"), pair2 = c("xend", "yend")),
  verbose = FALSE,
  error = FALSE,
  ...
)
}
\arguments{
\item{df}{Data.frame with numeric coordinate variable pairs.}

\item{angle}{Numeric value. The angle by which the coordinates
are rotated. Should range from 1-359.}

\item{clockwise}{Logical value. If \code{TRUE}, rotation is performed
in clockwise direction. If \code{FALSE}, the other way round.}

\item{coord_vars}{Input that denotes the variable pairs. Can be
a vector of length two. Or a list of vectors of length two. First
element in vector sets name for the x-axis, second value sets name
for the y axis.

If a list is provided, each slot is checked and invalid slots
are removed from the iteration.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}

\item{...}{Additional arguments given to \code{give_feedback()}.}
}
\value{
Adjusted data.frame.
}
\description{
Rotates coordinate variable pairs in a data.frame.
}
\details{
Usually a data.frame that contains variables that refer
to x- and y-coordinates has one single pair of these. E.g. one
variable named \emph{x} and one variable named \emph{y}. If so, \code{coord_vars = c("x", "y")}
or \verb{coord_vars = list(pair1 = c("x", "y")} is appropriate (naming the list
is not necessary). If the data.frame contains several variables that
refer to the same axes but in different scales they can be adjusted altogether.
E.g. a data.frame that contains variable pair \emph{x} and \emph{y} as well as \emph{col}
and \emph{row} needs \verb{coord_vars = list(pair1 = c("x", "y"), pair2 = c("col", "row")}.
For a pair to be adjusted \strong{both} variables must be found, else the adjustment
is skipped and the function gives feedback if \code{verbose = TRUE} or throws an
error if \code{error = TRUE}. Default sets both to \code{FALSE} which results in
silent skipping.
}
