% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{rotateAll}
\alias{rotateAll}
\alias{rotateImage}
\alias{rotateCoordinates}
\alias{rotateCoordsDf}
\alias{rotateImageAnnotations}
\alias{rotateSpatialTrajectories}
\title{Rotate image and coordinates}
\usage{
rotateAll(object, angle, clockwise = TRUE)

rotateImage(object, angle, clockwise = TRUE, track = TRUE)

rotateCoordinates(object, angle, clockwise = TRUE, verbose = NULL)

rotateCoordsDf(object, angle, clockwise = TRUE, verbose = NULL)

rotateImageAnnotations(object, angle, clockwise = TRUE, verbose = NULL)

rotateSpatialTrajectories(object, angle, clockwise = TRUE, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{angle}{Numeric value. The angle by which the coordinates
are rotated. Should range from 1-359.}

\item{clockwise}{Logical value. If \code{TRUE}, rotation is performed
in clockwise direction. If \code{FALSE}, the other way round.}

\item{track}{Logical value. If \code{TRUE}, changes regarding the image
justification (rotations and flipping) are tracked. Assuming that
image versions of different resolution are stored on your device with the same
justification as the primarily read image these changes in justification
can be automatically applied if the image is exchanged via \code{exchangeImage()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
The \verb{rotate*()} family rotates the current image
or coordinates of spatial aspects or everything. See details
for more information.
}
\details{
The \verb{rotate*()} functions can be used to rotate the complete \code{SPATA2}
object content or to rotate single aspects.

\itemize{
\item{\code{rotateAll()}:}{ Rotates image as well as every single spatial aspect.
\strong{Always tracks the justification.}}
\item{\code{rotateImage()}:}{ Rotates the image.}
\item{\code{rotateCoordinates()}:}{ Rotates the coordinates data.frame, image annotations
and spatial trajectories.}
\item{\code{rotateCoordsDf()}:}{ Rotates the coordinates data.frame.}
\item{\code{rotateImageAnnotations()}:}{ Rotates image annotations.}
\item{\code{rotateSpatialTrajectories()}:}{ Rotates spatial trajectories.}
}

@seealso \code{\link[=flipAll]{flipAll()}}, \code{\link[=scaleAll]{scaleAll()}}
}
