% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{rotateCoords}
\alias{rotateCoords}
\title{Rotate image}
\usage{
rotateCoords(object, angle, stepwise = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{angle}{Numeric value. The angle by which the coordinates
are rotated.}

\item{stepwise}{Logical value. If TRUE, the function allows
only step wise rotation by values that can be divided
by 45.}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Rotates the coordinates clockwise. Can be used to align
with coordinates.
}
