% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{rotateSpatialAnnotation}
\alias{rotateSpatialAnnotation}
\alias{rotateSpatialAnnotation,spata2-method}
\alias{rotateSpatialAnnotation,HistoImaging-method}
\title{Rotate Borders of a Spatial Annotation}
\usage{
rotateSpatialAnnotation(object, ...)

\S4method{rotateSpatialAnnotation}{spata2}(
  object,
  id,
  angle,
  clockwise = TRUE,
  new_id = FALSE,
  overwrite = FALSE
)

\S4method{rotateSpatialAnnotation}{HistoImaging}(
  object,
  id,
  angle,
  clockwise = TRUE,
  new_id = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Additional arguments given to \code{give_feedback()}.}

\item{id}{Character value. The ID of the spatial annotation of interest.}

\item{angle}{Numeric value. The angle by which the coordinates
are rotated. Should range from 1-359.}

\item{clockwise}{Logical value. If \code{TRUE}, rotation is performed
in clockwise direction. If \code{FALSE}, the other way round.}

\item{new_id}{Character value or \code{FALSE}. If character, the resulting
spatial annotation is stored under a new ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Rotates the outline of a spatial annotation to a specific
degree.
}
\seealso{
\code{\link[=centerSpatialAnnotation]{centerSpatialAnnotation()}}, \code{\link[=expandSpatialAnnotation]{expandSpatialAnnotation()}}, \code{\link[=smoothSpatialAnnotation]{smoothSpatialAnnotation()}},
\code{\link[=shiftSpatialAnnotation]{shiftSpatialAnnotation()}}
}
