% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpaceWrapper.R
\name{runBayesSpaceClustering}
\alias{runBayesSpaceClustering}
\title{Clustering with BayesSpace}
\usage{
runBayesSpaceClustering(
  object,
  dirname,
  name = "bayes_space",
  qs = seq(2, 15),
  prefix = "",
  return_model = TRUE,
  empty_remove = FALSE,
  overwrite = FALSE,
  seed = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{dirname}{Path to spaceranger output directory (e.g. "sampleID/outs/").
This directory must contain the counts matrix and feature/barcode TSVs in
\code{filtered_feature_bc_matrix/}, and the spot positions at
\code{spatial/tissue_positions_list.csv}. (These are default locations for
spaceranger outputs.)}

\item{name}{Character value. The name the cluster variable has in
the feature data of the \code{SPATA2} object. Defaults to \emph{bayes_space}.}

\item{qs}{The values of q to evaluate.}

\item{prefix}{Character value. Prefix of the cluster groups.}

\item{overwrite}{Logical value. If TRUE, \code{name} overwrites features
in feature data of the \code{SPATA2} object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated \code{SPATA2} object.
}
\description{
A wrapper around the BayesSpace clustering pipeline introduced
by \emph{Zhao et al. 2021}.
}
\details{
This function is a wrapper around \code{readVisium()},
\code{spatialPreprocess()}, \code{qTune()} and \code{spatialCluster()}
of the BayesSpace package. The results are stored in form of a grouping
variable in the feature data.frame of the returned \code{SPATA2} object.
}
\author{
Zhao E, Stone MR, Ren X, Guenthoer J, Smythe KS, Pulliam T,
 Williams SR, Uytingco CR, Taylor SEB, Nghiem P, Bielas JH, Gottardo R.
 Spatial transcriptomics at subspot resolution with BayesSpace.
 Nat Biotechnol. 2021 Nov;39(11):1375-1384. doi: 10.1038/s41587-021-00935-2.
 Epub 2021 Jun 3. PMID: 34083791; PMCID: PMC8763026.
}
