% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runCIN}
\alias{runCIN}
\title{Compute chromosomal instability}
\usage{
runCIN(
  object,
  bin_size = 1e+06,
  window_k = 10,
  noise = 0.035,
  gene_pos_df = SPATA2::gene_pos_df,
  chrom_regions_df = SPATA2::cnv_regions_df,
  coverage_model = SPATA2::coverage_model,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{bin_size}{The size used for the chromosomal binning. Defaults 1.000.000
base pairs.}

\item{window_k}{Given to \code{\link{window.k}} of \code{\link[imputeTS:na_ma]{imputeTS::na_ma}}.}

\item{noise}{Numeric value. Sets the level of noise to add.}

\item{gene_pos_df}{Data.frame defining the positions of genes on the chromosomes.}

\item{chrom_regions_df}{Data.frame defining the chromosomal regions.}

\item{coverage_model}{A formula.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Computes \strong{c}hromosomal \strong{in}stability scores based on copy number
variation results according to \emph{Drews et al., 2022}. Requires the results
of \code{\link[=runCNV]{runCNV()}}. See details for more.
}
\details{
Adds a total of 18 new numeric variables to the meta data.frame.

\itemize{
\item{\emph{CX1}, \emph{CX2}, \emph{CX3}, ..., \emph{CX17}:}{ Chromosomal instability scores as described in the paper
referenced below.}
\item{\emph{ploidy}:}{ Quantification of abnormal ploidy.}
}

Adding these variables is forced! Existing variable with equal names will be overwritten!
}
\note{
Requires the packages \code{CINSignatureQuantification} and \code{imputeTS} to be installed.
}
\references{
Drews, R.M., Hernando, B., Tarabichi, M. et al.
A pan-cancer compendium of chromosomal instability. Nature 606, 976–983 (2022).
https://doi.org/10.1038/s41586-022-04789-9
}
