% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-expression-analysis.R
\name{runDeAnalysis}
\alias{runDeAnalysis}
\title{Find differently expressed genes}
\usage{
runDeAnalysis(
  object,
  across,
  method_de = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to \code{Seurat::FindAllMarkers()}}
}
\value{
A spata-object containing the results in slot @dea.
}
\description{
This function makes use of \code{Seurat::FindAllMarkers()} to compute
the differently expressed genes across the groups denoted in the argument \code{across}.
See details for more.
}
\details{
If \code{across} and/or \code{method_de} are vectors instead of single
values \code{runDeAnalysis()} iterates over all combinations in a for-loop and
stores the results in the respective slots. (e.g.: If \code{across} = \emph{'seurat_clusters'}
and \code{method_de} = \emph{c('wilcox', 'bimod')} the function computes the differently expressed genes
across all groups found in the feature variable \emph{seurat_clusters} according to method \emph{wilcox} and
stores the results in the respective slot. Then it does the same according to method \emph{bimod}.)

The results are obtainable via \code{getDeResults()} and \code{getDeGenes()}.
}
