% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-implementation.R
\name{runHclust}
\alias{runHclust}
\alias{runHclustCnv}
\title{Run hierarchical clustering}
\usage{
runHclust(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  p = 2,
  force = FALSE,
  of_sample = NA,
  verbose = NULL
)

runHclustCnv(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  p = 2,
  force = FALSE,
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_dist}{Character value or vector (see details for more). Denotes the distance methods (e.g. \emph{'euclidean'}) to be used. Run \code{validDistanceMethods()}
to obtain all valid input options.}

\item{method_aggl}{Character value or vector (see details for more). Denotes the agglomerative methods (e.g. \emph{'ward.D'}) to be used. Run \code{validAgglomerativeMethods()}
to obtain all valid input options.}

\item{p}{Numeric value. Given to arguement \code{p} of
function \code{stats::dist()}.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
An updated spata-object containing the results
in the respective slot.
}
\description{
Initiates hierarchical clustering based on PCA-results
of either gene expression or copy-number-variation results. (The latter
requires the results of \code{runCnvAnalysis()}).
}
\details{
Iterates over all valid combinations of input for the
\code{method_*}-arguments and stores the results in the respective slot.
}
\seealso{
getHclustObject(), findHierarchicalCluster(), plotDendrogram()
}
