% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runImagePipeline}
\alias{runImagePipeline}
\title{Run image processing pipeline}
\usage{
runImagePipeline(object, img_name = activeImage(object), verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Arguments passed to \code{\link[=identifyPixelContent]{identifyPixelContent()}}.}
}
\value{
The input object containing the added or computed
results.
}
\description{
A wrapper around the image processing functions:

\itemize{
\item{\code{\link[=identifyPixelContent]{identifyPixelContent()}}}{}
\item{\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}}{}
\item{\code{\link[=identifyBackgroundColor]{identifyBackgroundColor()}}}
}
}
