% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runNearestNeighborClustering}
\alias{runNearestNeighborClustering}
\title{Clustering with nearest neighbor search}
\usage{
runNearestNeighborClustering(
  object,
  n_pcs = NULL,
  k = 50,
  naming = "nn2_{priority}_{treetype}",
  searchtype = "priority",
  treetype = "bd",
  radius = 0,
  eps = 0,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{n_pcs}{Numeric value or \code{NULL}. If numeric, specifies the number of
principal components to use. If \code{NULL}, all principal components are used
(as computed with \code{\link[=runPCA]{runPCA()}}).}

\item{k}{The maximum number of nearest neighbours to compute. The default value
is set to the smaller of the number of columnns in data.}

\item{naming}{Character value. A glue expression for the new cluster variable name.}

\item{searchtype}{See details}

\item{treetype}{Character vector. Character vector specifying the standard
\emph{'kd'} tree or a \emph{'bd'} (box-decomposition, AMNSW98) tree which
may perform better for larger point sets.}

\item{radius}{Radius of search for searchtype='radius'}

\item{eps}{Error bound: default of 0.0 implies exact nearest neighbour search}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{searchtypes}{Character value. Either \emph{'priority', 'standard'} or \emph{'radius '}. See details for more.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
A wrapper around the \code{\link[RANN:nn2]{RANN::nn2()}} function - nearest neighbor clustering.
}
\details{
Search types: priority visits cells in increasing order of distance from the
query point, and hence, should converge more rapidly on the true nearest neighbour,
but standard is usually faster for exact searches. radius only searches for neighbours
within a specified radius of the point. If there are no neighbours then nn.idx will
contain 0 and nn.dists will contain 1.340781e+154 for that point.
}
\note{
Requires the \code{RANN} packge.
}
