% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runPca}
\alias{runPca}
\alias{runPca2}
\title{Run Principal Component Analysis}
\usage{
runPca(
  object,
  n_pcs = 30,
  mtr_name = activeMatrix(object),
  assay_name = activeAssay(object),
  ...
)

runPca2(object, n_pcs = 30, mtr_name = NULL, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be computed.}

\item{mtr_name}{Character value. The name of the matrix to extract. If
\code{NULL}, defaults to the active matrix of the provided or specified \link[=MolecularAssay]{assay}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{...}{Additional arguments given to \code{irlba::prcomp_irlba()}.}
}
\value{
\itemize{
\item{\code{runPca()}:}{ An updated \code{SPATA2} object containing the reduction variables in the pca data.frame.}
\item{\code{runPca2()}:}{ The direct output-object of \code{irlba::prcomp_irlba()}}.
}
}
\description{
Takes the expression matrix of choice and passes it to
\code{irlba::prcomp_irlba()}.
}
