% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runSDEA}
\alias{runSDEA}
\title{Run spatial differential expression analysis}
\usage{
runSDEA(
  object,
  interval,
  id = idSA(object),
  naming = "sdea_{id}",
  method_de = "wilcox",
  base = 2,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{interval}{Distance measure. The width of the spatial intervals in which
the data points are grouped starting from the boundary of the spatial annotation
till the edge of the tissue is reached.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{naming}{A \code{\link[glue:glue]{glue::glue()}} instruction on how to create the name of the
grouping variable.Providing a simple string without glue syntax works, too.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function conducts differential expression analysis (\emph{DEA})
where data points are grouped based on their distance to a \link{SpatialAnnotation}.
}
\details{
This function bases on the concept of \code{\link[=runDEA]{runDEA()}} where gene expression
is compared across different groups within a grouping variable. However, in contrast to
\code{runDEA()} where the grouping variable is denoted via \code{across}, the function \code{\link[=runSDEA]{runSDEA()}}
creates a grouping variable in which data points are grouped based on their distance
to a \code{\link{SpatialAnnotation}}. This approach aims to identify genes that are upregulated
within certain distance intervals to the spatial annotation of interest. The spatial
interval is defined via the argument \code{interval}. E.g. if \verb{interval = 500um} data points
are grouped in 500um intervals starting from the boundaries of the spatial annotation
until the edge of the tissue is reached. The names of the groups correspond
to the distance itself: \emph{500um}, \emph{1000um}, \emph{1500um}, etc. If \verb{interval = 0.5mm}, which
is equivalent to 500um, the grouping will be the same but the group names correspond
to \emph{0.5mm}, \emph{1mm}, \emph{1.5mm}, etc. Data points that lie within the boundaries
of the spatial annotation are assigned to group \emph{core}.

The grouping variable created this way is stored in the feature data.frame as
any other grouping variable and the DEA results are stored in slot @dea as all
other DEA results.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\seealso{
\code{\link[=createGroupAnnotations]{createGroupAnnotations()}}, \code{\link[=createImageAnnotations]{createImageAnnotations()}},
\code{\link[=createNumericAnnotations]{createNumericAnnotations()}} to create spatial annotations.

The function \code{\link[=getCoordsDfSA]{getCoordsDfSA()}} relates data points to a spatial annotation.
This information is used by \code{runSDEA()} to create the spatial grouping.

\code{\link[=plotDeaDotplot]{plotDeaDotplot()}} to visualize results.
}
\keyword{internal}
