% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runSeuratClustering}
\alias{runSeuratClustering}
\title{Clustering with Seurat}
\usage{
runSeuratClustering(
  object,
  name = "seurat_clusters",
  mtr_name = activeMatrix(object),
  assay_name = activeAssay(object),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8)
)
}
\arguments{
\item{FindVariableFeatures, RunPCA, FindNeighbors, FindClusters}{Each argument
takes a list of arguments that is given to the equivalent function.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
A wrapper around the Seurat clustering pipeline suggested by
\emph{Hao and Hao et al., 2021}.
}
\examples{

 object <- SPATAData::downloadSpataObject("275_T")

 object <- runSeuratClustering(object, name = "seurat_clusters")

 plotSurface(object, color_by = "seurat_clusters")

}
