% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R, R/runA-Z.R
\name{getSparkxGeneDf}
\alias{getSparkxGeneDf}
\alias{getSparkxGenes}
\alias{getSparkxResults}
\alias{runSparkx}
\title{Identify spatially significant features with SPARKX}
\usage{
getSparkxGeneDf(object, threshold_pval = 1, arrange_pval = TRUE)

getSparkxGenes(object, threshold_pval)

getSparkxResults(object, assay_name = activeAssay(object), test = TRUE)

runSparkx(
  object,
  assay_name = activeAssay(object),
  numCores = 1,
  option = "mixture",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{test}{Character value. One of \emph{any}. \emph{all}, \emph{identical}, \emph{not_identical} and
\emph{none}. Specifies how input for \code{tags} is used to select image annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
A wrapper around the algorithm introduced by \emph{Zhu et al. 2021}
to identify features with non-random spatial expression pattern with SPARK-X.
}
\author{
Zhu, J., Sun, S. & Zhou, X. SPARK-X: non-parametric modeling enables
scalable and robust detection of spatial expression patterns for large spatial
transcriptomic studies. Genome Biol 22, 184 (2021). https://doi.org/10.1186/s13059-021-02404-0
}
