% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-correlation.R
\name{runSpatialCorrelationAnalysis}
\alias{runSpatialCorrelationAnalysis}
\title{Initiate gene clustering analysis based on spatial patterns}
\usage{
runSpatialCorrelationAnalysis(
  object,
  of_sample = "",
  genes = 2000,
  genes_additional = NULL,
  threshold_stw = 0.5,
  threshold_stpv = 0.1,
  with_ties = TRUE,
  method_cor = "pearson",
  method_dist = "euclidean",
  mtr_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{genes}{A numeric value (integer) or a character vector. Determines which genes
are included in the correlation assessment. If specified as a numeric value
the genes are sorted in a decreasing fashion corresponding to their variance
across all barcode spots. Then the top n genes are included whereby n is equal
to the specified numeric value.

If specified as a character vector it's elements are considered to be gene
names and all valid inputs are included.}

\item{genes_additional}{Character vector of gene names. If \code{genes} is
specified as a numeric value but you want certain genes to be included irrespective
of their variance you can denote them here and they are added after the
variance evaluation.}

\item{threshold_stw, threshold_stpv}{Numeric values. Both values refer to the
results of the shapiro-wilkinson test results for each gene. Before beeing sorted
according to their variance you can use both arguments to filter for genes
with a \emph{W-value} bigger or equal to \code{threshold_stw} and a respective
p-value lower or equal to \code{threshold_stpv}.}

\item{with_ties}{Logical. If set to TRUE (the default) genes with equal
variances are kept even if the total number of genes}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}
}
\value{
An updated spata-object.
}
\description{
This function screens a subset of genes and evaluates their
spatial overlap by correlation- and subsequent clustering analysis. Results can be
conveniently obtained or processed with additional functions such as
\code{clusterSpCorResults()}, \code{getGenes()} or \code{getSpCorResults()}.
}
\details{
The overall expression matrix is filtered according to the input
of argument \code{genes}, transposed and given to \code{stats::cor()}. The returned
correlation matrix is given to \code{stats::dist()} to calculate the distance
matrix needed for subsequent cluster analysis.

Use \code{getGenes()} and it's argument \code{similar_to} in order to get genes
that feature a similar expression profile/pattern as a gene of interest.
}
