% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runUmap}
\alias{runUmap}
\alias{runUmap2}
\title{Run UMAP-Analysis}
\usage{
runUmap(object, n_pcs = 20, ...)

runUmap2(object, n_pcs = 20, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{n_pcs}{Numeric value. Denotes the number of principal components used. Must be
smaller or equal to the number of principal components the pca data.frame contains.}

\item{...}{Additional arguments given to \code{umap::umap()}.}
}
\value{
\itemize{
\item{\code{runUmap()}:}{ An updated \code{SPATA2} object containing the reduction variables in the umap data.frame.}
\item{\code{runUmap2()}:}{ The direct output-object of \code{umap::umap()}}
}
}
\description{
Takes the pca-data of the object up to the principal component denoted
in argument \code{n_pcs} and performs UMAP with it.
}
