% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{sample_name}
\alias{sample_name}
\title{sample_name}
\usage{
sample_name(sample_name)
}
\arguments{
\item{sample_name}{Character value. Denotes the name of the sample you are
analyzing with the spata-object. The future input for SPATA's \code{of_sample}-argument.}
}
\description{
sample_name
}
\keyword{internal}
