% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{saveGeneSetDf}
\alias{saveGeneSetDf}
\title{Save a gene set data.frame}
\usage{
saveGeneSetDf(object, directory)
}
\arguments{
\item{object}{A valid spata-object.}

\item{directory}{Character value.}
}
\value{
An invisible TRUE if saved successfully or an informative error message.
}
\description{
Extracts the gene-set data.frame and saves it as a .RDS-file.
}
