% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_deconvolution}
\alias{sc_deconvolution}
\title{List of data.frames of single cell deconvolution. Names of
the list correspond to the sample name. Obtain data.frame
using \verb{sc_deconvolution[[*<sample.name>*]]}}
\format{
An object of class \code{list} of length 1.
}
\usage{
sc_deconvolution
}
\description{
@docType data
@usage data(sc_deconvolution)
}
\keyword{datasets}
