% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imported-functions.R
\name{scale_color_add_on}
\alias{scale_color_add_on}
\title{ggplot2 - color add on}
\usage{
scale_color_add_on(
  aes = "color",
  variable = "numeric",
  clrsp = NULL,
  clrp = NULL,
  clrp.adjust = NULL,
  ...
)
}
\arguments{
\item{aes}{Character value. Either \emph{'color'} or \emph{'fill'}. Denotes the
used aesthetic.}

\item{variable}{The variable that is mapped onto the denoted aesthetic or a character
value - one of \emph{'numeric', 'discrete'}.}

\item{clrsp}{Character value. The color spectrum of choice. Run \code{confuns::all_color_spectra()}
to see all valid input choices.

(Ignored if \code{variable} is discrete)}

\item{clrp}{Character value. The color palette of choice.
Run \code{confuns::all_color_palettes()} to see all valid input choices.

(Ignored if \code{variable} is numeric)}

\item{clrp.adjust}{Named character vector or NULL. If character, it adjusts the
colors that are used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{...}{Additional arguments given to the respective function.

 \itemize{
  \item{\code{ggplot2::scale_<aes>_viridis_c(...)}: If \code{variable} is numeric and
  \code{clrsp} is one of \emph{'cividis', 'viridis', 'inferno', 'magma', 'plasma'}.}
  \item{\code{colorspace::scale_<aes>_continuous_sequential(...)}: If \code{variable} is numeric
  and \code{clrsp} is sequential.}
  \item{\code{colorspace::scale_<aes>_continuous_diverging(...)}: If \code{variable} is numeric
  and \code{clrsp} is diverging.}
  \item{\code{ggplot2::scale_<aes>_viridis_d(...)}: If \code{variable} is numeric
  and \code{clrp} is one of \emph{'cividis', 'viridis', 'inferno', 'magma', 'plasma'}.}
  \item{\code{ggplot2::scale_<aes>_greyscale(...)}: If \code{variable} is discrete
  and \code{clrp} is set to \emph{'greyscale'}.}
  \item{\code{ggplot2:.scale_<aes>_discrete()}: If \code{variable} is discrete
  and \code{clrp} is set to \emph{'default'}.}
  \item{\code{ggplot2::scale_<aes>_manual(...)}: If \code{variable} is discrete.}
  }}
}
\value{
An unnamed list containing the ggproto object.
}
\description{
A wrapper around a variety of \code{scale_color_-() / scale_fill_-()}
functions.
}
\details{
If the specified \code{clrp} does not contain enough colors to cover the
specified variable NULL is returned - which makes the call to \code{ggplot2::ggplot()}
use the default ggplot2 color panel.
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = qsec, y = wt)) +
geom_point(aes(color = mpg)) +
scale_color_add_on(aes = "color", clrsp = "Reds 3")
}
