% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scale_factors}
\alias{scale_factors}
\title{List of scale factors for Visium input.}
\format{
An object of class \code{scalefactors} of length 4.
}
\usage{
data(scale_factors)
}
\description{
List of scale factors for Visium input.
}
\keyword{datasets}
