% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{scaleCoordinates}
\alias{scaleCoordinates}
\alias{scaleCoordsDf}
\alias{scaleImageAnnotations}
\alias{scaleSpatialTrajectories}
\title{Scale coordinates}
\usage{
scaleCoordinates(object, scale_fct, verbose = NULL)

scaleCoordsDf(object, scale_fct, verbose = NULL)

scaleImageAnnotations(object, scale_fct, verbose = NULL)

scaleSpatialTrajectories(object, scale_fct, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{scale_fct}{Numeric vector of length two. Scale factors with which the coordinates
are multiplied. First value is used for x-, second value is used for y-coordinates.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Scales coordinates of spatial aspects in the \code{SPATA2} object.
}
