% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{segments_to_vertices}
\alias{segments_to_vertices}
\title{Transform segments to vertices}
\usage{
segments_to_vertices(arcs_df)
}
\arguments{
\item{arcs_df}{A data frame containing arc information with columns:
\itemize{
\item \code{x1}, \code{y1}: Starting coordinates of each segment
\item \code{x2}, \code{y2}: Ending coordinates of each segment
\item \code{idx}: An identifier for each segment
}}
}
\value{
A data frame (\code{outline_df}) containing the vertices (\code{x}, \code{y} coordinates and \code{idx})
that form the ordered outline of the polygon.
}
\description{
This function takes a data frame of arcs (segments) and constructs an ordered sequence of vertices
to outline a connected, non-intersecting polygon.
}
\details{
This function iteratively selects arcs that connect to form a polygon. If an arc’s endpoint does not
match any available arc’s start or endpoint, it selects the nearest arc using a nearest-neighbor search.
Segments that would cause intersections are removed. If the constructed polygon closes on itself,
it terminates the process.
}
\examples{

# Sample data frame of arcs

arcs_df <- data.frame(
  x1 = c(0, 1, 2),
  y1 = c(0, 1, 0),
  x2 = c(1, 2, 0),
  y2 = c(1, 0, 0),
  idx = c(1, 2, 3)
)
segments_to_vertices(arcs_df)


}
