% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{set_dummy}
\alias{set_dummy}
\title{set}
\usage{
set_dummy()
}
\value{
A spata object containing the set input.
}
\description{
set
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
