% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setActiveExpressionMatrix}
\alias{setActiveExpressionMatrix}
\title{Deprecated}
\usage{
setActiveExpressionMatrix(...)
}
\description{
Deprecated in favor of \code{\link[=activateMatrix]{activateMatrix()}}.
}
\keyword{internal}
