% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setActiveMatrix}
\alias{setActiveMatrix}
\title{Deprecated}
\usage{
setActiveMatrix(object, mtr_name, verbose = NULL)
}
\description{
Deprecated in favor of \code{\link[=activateMatrix]{activateMatrix()}}.
}
\keyword{internal}
