% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{setAutoencoderAssessment}
\alias{setAutoencoderAssessment}
\title{Set results of autoencoder assessment}
\usage{
setAutoencoderAssessment(object, assessment_list, of_sample = "")
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assessment_list}{Named list with slots \code{$df} and \code{$set_up}.}
}
\value{
A spata-object.
}
\description{
Set results of autoencoder assessment
}
\keyword{internal}
