% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setBarcodes}
\alias{setBarcodes}
\title{Set barcodes}
\usage{
setBarcodes(object, barcodes)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{barcodes}{Character vector. Barcode ids.}
}
\value{
A spata object containing the set input.
}
\description{
Sets the reference barcode ids.
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
