% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setCaptureArea}
\alias{setCaptureArea}
\alias{setCaptureArea,SPATA2-method}
\alias{setCaptureArea,SpatialData-method}
\title{Set capture area}
\usage{
setCaptureArea(object, ...)

\S4method{setCaptureArea}{SPATA2}(object, capture_area, ...)

\S4method{setCaptureArea}{SpatialData}(object, capture_area, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{capture_area}{Data.frame of vertices .with x_orig and y_orig variables}
}
\description{
Sets the capture area for objects from platforms with
a specific capture area / field of view.
}
\seealso{
\code{\link[=getCaptureArea]{getCaptureArea()}}
}
\keyword{internal}
