% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setCCD}
\alias{setCCD}
\title{Set center to center distance}
\usage{
setCCD(object, ccd)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ccd}{Distance measure of length one in SI units.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets center to center distance in slot \verb{$ccd} of the \code{SpatialMethod}
of the object.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\keyword{internal}
