% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setCoordsDf}
\alias{setCoordsDf}
\alias{setCoordsDf,SPATA2-method}
\alias{setCoordsDf,HistoImaging-method}
\title{Set the coordinates}
\usage{
setCoordsDf(object, ...)

\S4method{setCoordsDf}{SPATA2}(object, coords_df, force = FALSE)

\S4method{setCoordsDf}{HistoImaging}(object, coords_df, force = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\itemize{
\item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
\item{\emph{sample} Character. The sample belonging of every barcode-spot.}
\item{\emph{x} Numeric. The x-coordinates of every barcode.}
\item{\emph{y} Numeric. The y-coordinates of every barcode.}
}}
}
\value{
A spata object containing the set input.
}
\description{
Set the coordinates
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
