% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/setA-Z.R
\name{getDefaultGrouping}
\alias{getDefaultGrouping}
\alias{setDefaultGrouping}
\title{Default grouping}
\usage{
getDefaultGrouping(object, verbose = NULL, arg = "across")

setDefaultGrouping(object, grouping_variable, verbose = NULL)
}
\arguments{
\item{grouping_variable}{Character value. The grouping variable that is
used by default within all functions that need one.}
}
\value{
\code{setDefaultGrouping()}: Updated spata object. \code{getDefaultGrouping()}: Character value. Name
of the default grouping variable.
}
\description{
Sets and extracts the default grouping. Useful to save typing
in functions that require a grouping variable as input. (Usually referred to
via arguments \code{across} or \code{grouping_variable}).
}
