% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setDefaultInstructions}
\alias{setDefaultInstructions}
\alias{setDirectoryInstructions}
\title{Set default instructions}
\usage{
setDefaultInstructions(object, instructions = NULL)

setDirectoryInstructions(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Set default instructions
}
\keyword{internal}
