% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/setA-Z.R
\name{getDefaultTrajectory}
\alias{getDefaultTrajectory}
\alias{getDefaultTrajectoryId}
\alias{setDefaultTrajectory}
\alias{setDefaultTrajectoryId}
\title{Obtain dimensional reduction data}
\usage{
getDefaultTrajectory(object, ...)

getDefaultTrajectoryId(object, ...)

setDefaultTrajectory(object, id, verbose = NULL)

setDefaultTrajectoryId(object, id, verbose = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{id}{Character value.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be included.}
}
\value{
A data.frame that contains the unique identifiers
(keys): \emph{barcodes, sample} and:.

\itemize{
\item{ \code{getPcaDf()}: \emph{PC1, PC2, PC3, ...PCn}}
\item{ \code{getTsneDf()}: \emph{tsne1, tsne2}}
\item{ \code{getUmapDf()}: \emph{umap1, umap2}}
}

\code{setDefaultTrajectory()}: Updated spata object. \code{getDefaultTrajectory()}: Character value. Id
of the default trajectory.
}
\description{
Sets and extracts the default trajectory id. Useful to save typing
in functions that require a trajectory name as input.
}
