% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setDenoisedMatrix}
\alias{setDenoisedMatrix}
\title{Deprecated}
\usage{
setDenoisedMatrix(object, denoised_mtr, ...)
}
\description{
Deprecated in favor of \code{\link[=setProcessedMatrix]{setProcessedMatrix()}}.
}
\keyword{internal}
