% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setFeatureDf}
\alias{setFeatureDf}
\title{Set feature data.frame}
\usage{
setFeatureDf(object, feature_df, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_df}{A data.frame that contains the feature and the key-variables.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
set_dummy return details
}
\description{
Set feature data.frame
}
