% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setGeneSetDf}
\alias{setGeneSetDf}
\title{Set the gene-set data.frame}
\usage{
setGeneSetDf(object, gene_set_df)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gene_set_df}{A data.frame containing the gene names in
variable \emph{gene} and the gene set name in variable \emph{ont}.}
}
\value{
A spata object containing the set input.
}
\description{
Set the gene-set data.frame
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
