% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setHistoImaging}
\alias{setHistoImaging}
\alias{setImageObject}
\title{Set \code{HistoImaging}}
\usage{
setHistoImaging(object, imaging)

setImageObject(object, image_object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{imaging}{An object of class \code{\link{HistoImaging}}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
Sets the image container class \code{HistoImaging}
in the corresponding slot of the \code{spata2} object.
}
