% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{setHistologyImage}
\alias{setHistologyImage}
\alias{setHistologyImage,HistologyImagingNew-method}
\title{Set \code{HistologyImage}}
\usage{
setHistologyImage(object, ...)

\S4method{setHistologyImage}{HistologyImagingNew}(object, hist_img, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{hist_img}{An object of class \code{HistologyImage}.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets object of class \code{HistologyImage}.
}
