% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setImage}
\alias{setImage}
\title{Deprecated}
\usage{
setImage(object, image, of_sample = "")
}
\description{
Deprecated in favor of \code{\link[=registerImage]{registerImage()}}.
}
\keyword{internal}
