% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setImageAnnotation}
\alias{setImageAnnotation}
\alias{setImageAnnotations}
\title{Deprecated}
\usage{
setImageAnnotation(object, img_ann, ...)

setImageAnnotations(object, img_anns, ...)
}
\description{
Deprecated in favor of \code{\link[=setSpatialAnnotation]{setSpatialAnnotation()}}.
}
\keyword{internal}
