% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setImageDir}
\alias{setImageDir}
\alias{setImageDir,SPATA2-method}
\alias{setImageDir,SpatialData-method}
\alias{setImageDir,HistoImage-method}
\title{Set image directory}
\usage{
setImageDir(object, ...)

\S4method{setImageDir}{SPATA2}(object, img_name, dir)

\S4method{setImageDir}{SpatialData}(object, img_name, dir)

\S4method{setImageDir}{HistoImage}(object, dir)
}
\arguments{
\item{dir}{Character value. The directory of the new image.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets the file directory under which an image is read in when
activated.
}
\note{
The function also accepts directories that do not exist! It throws
a warnign if that is the case.
}
\seealso{
\code{\link[=activateImage]{activateImage()}}, \code{\link[=loadImage]{loadImage()}}
}
