% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setImageDirLowres}
\alias{setImageDirLowres}
\alias{setImageDirHighres}
\title{Set image directories}
\usage{
setImageDirLowres(object, dir_lowres, check = TRUE)

setImageDirHighres(object, dir_highres, check = TRUE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{dir_lowres, dir_highres}{Character value. The file directories to
the corresponding images.}

\item{check}{Logical value. If set to TRUE the input directory is checked
for validity and it is checked if the file actually exists.}
}
\value{
An updated spata object.
}
\description{
Sets image directories that facilitate image exchanges.
}
