% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setImageOrigin}
\alias{setImageOrigin}
\title{Set image origin}
\usage{
setImageOrigin(object, origin)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{origin}{Character value. Directory or name of the object
from the global environment.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Sets the origin info of the current image.
}
