% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{setImageTransformations}
\alias{setImageTransformations}
\alias{setImageTransformations,HistologyImagingNew-method}
\title{Set image transformation instructions}
\usage{
setImageTransformations(object, ...)

\S4method{setImageTransformations}{HistologyImagingNew}(object, name, transformations, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets image transformation instruction list.
}
