% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setInitiationInfo}
\alias{setInitiationInfo}
\title{Set initiation information}
\usage{
setInitiationInfo(object, additional_input = list())
}
\arguments{
\item{object}{A valid spata-object.}

\item{additional_input}{A list of named arguments provided by
... of the calling function.}
}
\value{
A spata object containing the set input.
}
\description{
Set initiation information
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
