% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setLogfileDf}
\alias{setLogfileDf}
\title{Set logfile data.frame}
\usage{
setLogfileDf(object, lf_df)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{lf_df}{The \link[=concept_logfile]{logfile data.frame}.}
}
\value{
A spata object containing the set input.
}
\description{
Set logfile data.frame
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
