% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setMetaDf}
\alias{setMetaDf}
\title{Set meta data.frame}
\usage{
setMetaDf(object, meta_df)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
set_dummy return details
}
\description{
Sets the meta data.frame.
}
\keyword{internal}
