% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{setMetaVarDf}
\alias{setMetaVarDf}
\title{Set molecular meta data.frame}
\usage{
setMetaVarDf(object, meta_var_df, assay_name = activeAssay(object))
}
\arguments{
\item{meta_var_df}{A data.frame for slot @meta_var of the molecular assay.}

\item{inherit}{update_dummy return}
}
\description{
Sets the metadata variable data frame for a specified assay in the given object.
}
