% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setPcaDf}
\alias{setPcaDf}
\alias{setTsneDf}
\alias{setUmapDf}
\title{Set dimensional reductions data.frames}
\usage{
setPcaDf(object, pca_df, fdb_fn = "stop")

setTsneDf(object, tsne_df, ...)

setUmapDf(object, umap_df, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{pca_df, tsne_df, umap_df}{The data.frame composed of the variables
\emph{barcodes}, \emph{sample} and the variables containing the barcode-
spot embedding.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A spata object containing the set input.
}
\description{
Safely adds data.frames containing the barcode-spot embedding
of different dimensional reduction techniques. Cell embedding variables
must be named as follows:

\itemize{
\item{ \code{pca_df}: \emph{PC1, PC2, PC3, ...}}
\item{ \code{tsne_df}: \emph{tsne1, tsne2, ...}}
\item{ \code{umap_df}: \emph{umap1, umap2, ...}}
}
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
