% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setPixelScaleFactor}
\alias{setPixelScaleFactor}
\title{Set pixel scale factor}
\usage{
setPixelScaleFactor(object, pxl_scale_fct = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{pxl_scale_fct}{Numeric value with an attribute named
\emph{unit} with the unit euol/px.}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Sets pixel scale factor.
}
