% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setScaleFactor}
\alias{setScaleFactor}
\alias{setScaleFactor,SPATA2-method}
\alias{setScaleFactor,SpatialData-method}
\alias{setScaleFactor,HistoImage-method}
\title{Set scale factors}
\usage{
setScaleFactor(object, ...)

\S4method{setScaleFactor}{SPATA2}(object, fct_name, value)

\S4method{setScaleFactor}{SpatialData}(object, fct_name, value)

\S4method{setScaleFactor}{HistoImage}(object, fct_name, value)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{fct_name}{Character value. Name of the scale factor.}

\item{value}{Value to set.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Sets scale factor values.
}
\details{
For S4 methods other than for \code{\link{HistoImage}}: Sets scale factors
\strong{for the reference image}. Corresponding scale factors for additionally
registered images (if there are any) are computed.

If there are no images registered, the scale factor is set in the corresponding list
slot of @scale_factors of the \code{SpatialData} object.
}
