% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{setSpataDir}
\alias{setSpataDir}
\title{Set SPATA2 directory}
\usage{
setSpataDir(object, dir, add_wd = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{dir}{Character value. The directory under which to store the
\code{SPATA2} object.}

\item{add_wd}{Logical value. If \code{TRUE}, the working directory is added to
the directory separated by \emph{'/'}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Sets a directory under which the \code{SPATA2} object is
always stored using the function \code{saveSpataObject()}.
}
