% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialAnnotation}
\alias{setSpatialAnnotation}
\alias{setSpatialAnnotation,spata2-method}
\alias{setSpatialAnnotations}
\title{Set spatial annotations}
\usage{
setSpatialAnnotation(object, ...)

\S4method{setSpatialAnnotation}{spata2}(object, spat_ann, ...)

setSpatialAnnotations(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{spat_ann}{An object of class \code{\link{SpatialAnnotation}}.}
}
\description{
Sets spatial annotations in the correct slot. Expects a
valid spatial annotation and does not conduct any further checks or
adjustments.
}
\note{
\code{\link{GroupAnnotation}}, \code{\link{NumericAnnotation}}, \code{\link{ImageAnnotation}} are
derivatives of the \code{SpatialAnnotation} class and are valid inputs, too!
}
