% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialMethod}
\alias{setSpatialMethod}
\title{Set the \code{SpatialMethod} object}
\usage{
setSpatialMethod(object, method)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{method}{An object of class \code{SpatialMethod}.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets the object of class \code{SpatialMethod}.
}
