% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialMethodInfo}
\alias{setSpatialMethodInfo}
\title{Set information of \code{SpatialMethod} object}
\usage{
setSpatialMethodInfo(object, slot, content)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{slot}{Character value. The list-slot of slot \verb{@info}
of the \code{SpatialMethod} object as in \verb{@info[[slot]]}.}

\item{content}{The content to set.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets information in slot \verb{@info} of the
\code{SpatialMethod}-object.
}
