% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialTrajectory}
\alias{setSpatialTrajectory}
\alias{setSpatialTrajectories}
\title{Set SpatialTrajectory objects}
\usage{
setSpatialTrajectory(object, trajectory, overwrite = FALSE)

setSpatialTrajectories(object, trajectories, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{trajectory}{An object of class \code{SpatialTrajectory.}}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{trajectories}{List of objects of class \code{SpatialTrajectory}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets trajectories in the correct slot.
}
