% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setSpCorResults}
\alias{setSpCorResults}
\title{Set results of spatial correlation analysis}
\usage{
setSpCorResults(object, sp_cor_list, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{sp_cor_list}{The list of information and results the
function \code{runSpatialCorrelationAnalysis()} returns.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A spata object containing the set input.
}
\description{
Set results of spatial correlation analysis
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
