% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setTissueOutline}
\alias{setTissueOutline}
\title{Set tissue outline}
\usage{
setTissueOutline(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\code{spata2} object with additional variables in coordinates data.frame.

\itemize{
\item{\emph{section} :}{ character. The identified tissue section. 0 means probable artefact spot.}
\item{\emph{outline} :}{logical. \code{TRUE} if identified as a spot that lies on the edge of the tissue.}
}
}
\description{
Sets tissue outline by calling \code{identifyTissueSections()}
and \code{identifyTissueOutline()}.
}
\keyword{internal}
