% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setTrajectory}
\alias{setTrajectory}
\alias{setTrajectories}
\title{Set trajectories}
\usage{
setTrajectory(object, trajectory, align = TRUE, overwrite = FALSE)

setTrajectories(object, trajectories, align = TRUE, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{trajectory}{An object of class \code{Trajectory.}}

\item{align}{Logical value. If \code{TRUE}, trajectories of class \code{SpatialTrajectory}
are aligned with image justification changes of the image of the
\code{SPATA2} object.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{trajectories}{List of objects of class \code{Trajectory}.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets trajectories in the correct slot.
}
