% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setUpMetaVar}
\alias{setUpMetaVar}
\alias{setUpMetaVar,SPATA2-method}
\alias{setUpMetaVar,MolecularAssay-method}
\title{Set up meta var slot}
\usage{
setUpMetaVar(object, ...)

\S4method{setUpMetaVar}{SPATA2}(object, assay_name = activeAssay(object), ...)

\S4method{setUpMetaVar}{MolecularAssay}(object, assay_name = activeAssay(object), ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets up slot @meta_var of a molecular assay if not already done.
}
\keyword{internal}
