% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{seurat_object_dummy}
\alias{seurat_object_dummy}
\title{Title}
\usage{
seurat_object_dummy(seurat_object)
}
\arguments{
\item{seurat_object}{A valid seurat-object. (from the Seurat platform)}
}
\description{
Title
}
