% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{shiftSpatialAnnotation}
\alias{shiftSpatialAnnotation}
\title{Shift the borders of a Spatial Annotation}
\usage{
shiftSpatialAnnotation(
  object,
  id,
  outer = TRUE,
  inner = TRUE,
  shift_x = 0,
  shift_y = 0,
  new_id = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{outer}{Logical; if TRUE, the outer border of the annotation is affected.}

\item{inner}{Logical or numeric; if TRUE, all inner borders are affected.
If a numeric value is provided, only the specified inner borders are affected.}

\item{shift_x, shift_y}{Distance measure. The shift in the x- and/or y-direction. Negative
values are also accepted.}

\item{new_id}{Character value or \code{FALSE}. If character, the resulting
spatial annotation is stored under a new ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input object containing the added or computed
results.
}
\description{
This function moves a spatial annotation
either in the x or y direction, or both. It allows for selective shifting of
outer and/or inner borders of the annotation.
}
\seealso{
\code{\link[=expandSpatialAnnotation]{expandSpatialAnnotation()}}, \code{\link[=smoothSpatialAnnotation]{smoothSpatialAnnotation()}}, \code{\link{SpatialAnnotation}}
}
