% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{shinyModuleZoomingServer}
\alias{shinyModuleZoomingServer}
\alias{shinyModuleZoomingUI}
\title{Shiny module \emph{Zooming}}
\usage{
shinyModuleZoomingServer(
  id = "m1",
  default = list(),
  brushed_area = NULL,
  dims = NULL,
  persp = "ccs",
  trigger_zoom_out = NULL,
  ...
)

shinyModuleZoomingUI(id = "m1")
}
\arguments{
\item{id}{Character value. ID of the module.}

\item{brushed_area}{Reactive expression which, if called with \verb{<reactiveExpression>()}
returns a list of four slots named \emph{xmin}, \emph{xmax}, \emph{ymin} and \emph{ymax} each being
a numeric value.}

\item{dims}{Image dimensions of the original, complete image during zooming.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{object}{An object for which a method for \code{getImageRange()} is defined.}
}
\value{
A reactive expression which, if called with \verb{<reactiveExpression>()}
returns a list of three slots:
\itemize{
\item{\emph{x}:}{ Numeric vector of length two corresponding to the xrange of the zoom.}
\item{\emph{y}:}{ Numeric vector of length two corresponging to the yrange of the zoom.}
\item{\emph{dims}:}{ Numeric vector of length two. The dimensions of the image that was zoomed in on.}
}
}
\description{
The server logical for shiny module \emph{Zooming}.
}
