% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{The default collection of frequently used signatures}
\format{
An object of class \code{list} of length 1.
}
\usage{
data(gsdf)
}
\description{
Number of different gene-sets: 11654
}
\details{
Gene set classes:

\itemize{
\item{BP.GO: Biological processes - gene ontology}
\item{CC.GO: Cellular componenents - gene-ontology}
\item{MF.GO: Molecular functions - gene-ontology}
\item{HM: Hallmark gene sets}
\item{BC: Biocarta gene sets}
\item{RCTM: Reactome gene sets}
}
}
\keyword{datasets}
