% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{simulate_expression_pattern_sas}
\alias{simulate_expression_pattern_sas}
\alias{simulate_expression_pattern_sts}
\title{Expression pattern simulation}
\usage{
simulate_expression_pattern_sas(
  object,
  id,
  simulations,
  core,
  binwidth = recBinwidth(object),
  distance = distToEdge(object, id),
  angle_span = c(0, 360),
  noise_levels = seq(0, 100, length.out = 21),
  noise_types = c("ed", "ep", "fp", "cb"),
  range_sim = c(0, 1),
  range_random = base::range(range_sim),
  model_add = NULL,
  model_remove = NULL,
  seed = 123,
  npref = "SE",
  nsep = ".",
  verbose = TRUE
)

simulate_expression_pattern_sts(
  object,
  id,
  simulations,
  width,
  binwidth = recBinwidth(object),
  noise_levels = seq(0, 100, length.out = 21),
  noise_types = c("ed", "ep", "fp", "cb"),
  range_sim = c(0, 1),
  range_random = base::range(range_sim),
  model_add = NULL,
  seed = 123,
  npref = "SE",
  nsep = ".",
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{simulations}{A list of \emph{simulation lists}. A simulations list is a list
that contains at least three slots that provide instructions for a set of
simulations:

\itemize{
\item{id}{ Character value. The overall ID of the simulation.}
\item{n}{ Integer. The number of simulations.}
\item{model} Character value. The name of the model based on which you want
to simulate the expression.
}

A simulation list can have an additional slot called \emph{noise_levels} which,
if present, overwrites the input for \code{noise_levels} for the specific simulation.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{noise_levels}{A vector of integers between 0-100 (inclusive). Indicate
the respective levels of noise to be added to each simulation in percent.
0 indicates no noise. 100 indicates only noise.}

\item{range_sim}{The range of the output simulation values.}

\item{range_random}{The range of the values randomly assigned to areas
outside of the area of interest.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{seed}{Numeric value, given to \code{set.seed()}.}

\item{npref}{Prefix for the simulations. Defaults to SG (simulated gene.)}

\item{nsep}{Character value with which to separate the aspects that
build the name of each simulation.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Matrix with rownames corresponding to the simulation names and
colnames corresponding to the barcodes of the object.
}
\description{
Simulates expression pattern by modelling expression dependent
on the distance to a spatial annotation.
}
\details{
Simulation names are created according to the function call
\code{stringr::str_c(npref, nsep, "NP", nl, nsep, sim_id, nsep, n, sep = "")}.
Where \code{nl} is the noise level of the simulation, \code{sim_id} corresponds to
the slot \emph{id} of the simulation list and \code{n} is the index of the simulation
in case slot \emph{n} of the simulation instruction is bigger than one.
}
