% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{simulate_random_gradients}
\alias{simulate_random_gradients}
\title{Simulate Random Gradients}
\usage{
simulate_random_gradients(
  coords_df,
  span,
  expr_est_pos,
  amccd,
  n = 1000,
  seed = 123,
  coef = Inf,
  range = c(0, 1),
  fn = "runif",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{coords_df}{A data frame containing coordinates and distance information.}

\item{span}{The alpha parameter for the loess smoothing, controlling the degree of smoothness.}

\item{n}{The number of simulations to perform (default is 1000).}

\item{seed}{The random seed to ensure reproducibility (default is 123).}

\item{range}{A numeric vector specifying the range for generating random expression values (default is c(0, 1)).}

\item{verbose}{A logical value indicating whether to display progress messages (default is TRUE).}

\item{...}{Additional parameters given to \code{\link[=obtain_inferred_gradient]{obtain_inferred_gradient()}}.}

\item{pred_pos}{A numeric vector of positions for which the gradients are inferred.}
}
\value{
A list of length \code{n} containing information about the simulated gradients, Loess Deviation Scores (LDS),
loess models, and total variation for each simulation. The list is named with seed indices to trace the circumstances
under which each simulation was conducted.

The output list contains the following components:
\describe{
\item{gradient}{A numeric vector representing the inferred expression gradient.}
\item{lds}{A numeric value representing the Loess Deviation Score (LDS) for the inferred gradient.}
\item{model}{A loess model object fitted to the simulated data.}
\item{tot_var}{A numeric value representing the total variation of the inferred gradient.}
}
}
\description{
This function simulates random expression patterns, computes gradients, and calculates
various metrics for evaluating the inferred expression gradients.
}
\details{
This function generates random expression patterns with specified randomness levels, fits loess curves to the data,
computes gradients, LDS, loess models, and total variation for each simulation, and returns the results in a named list.
}
