% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{simulate_spatial_niches}
\alias{simulate_spatial_niches}
\title{Create a Spatial Niche in Coordinate Data Frame}
\usage{
simulate_spatial_niches(
  coords_df,
  origins,
  size,
  size_fct = 1.1,
  vt,
  vf = NULL,
  rr = c(0, 1),
  type = "small",
  seed = 123
)
}
\arguments{
\item{coords_df}{A data frame containing spatial coordinates and other variables.}

\item{origins}{A character vector specifying the barcodes of the origin points around which the niche will be created.}

\item{size}{A numeric value specifying the size of the niche around each origin point.}

\item{vt}{The name of the variable in \code{coords_df} to which values will be assigned.}

\item{vf}{The name of the variable in \code{coords_df} from which values will be taken. If NULL (default), random values will be generated and assigned.}

\item{rr}{A numeric vector of length 2 specifying the range for generating random values. Default is c(0, 1).}

\item{seed}{A numeric value specifying the seed for random number generation. Default is 123.}
}
\value{
A data frame with the modified values based on the created spatial niche.
}
\description{
This function creates a spatial niche around specified origin points in a coordinate data frame.
The niche is defined by a specified size and is used to modify values in the data frame based on proximity to the origin points.
}
\details{
The function works by calculating the distances from each origin point to all other points in \code{coords_df}.
Points that fall within the specified niche size around each origin are identified, and their values in the \code{vt} variable are modified based on the \code{vf} variable.
If \code{vf} is NULL, random values within the specified range (\code{rr}) are generated and used.
}
\examples{
\dontrun{
df <- data.frame(x = runif(100, 0, 10), y = runif(100, 0, 10), value = rnorm(100))
df_with_niche <- simulate_spatial_niches(df, origins = c("A", "B"), size = 5, vt = "value")
}

}
\keyword{internal}
