% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{SlideSeqV1}
\alias{SlideSeqV1}
\title{SlideSeqV1 Spatial Method}
\format{
An object of class \code{SpatialMethod} of length 1.
}
\usage{
SlideSeqV1
}
\description{
This object abstracts the SlideSeqV1 spatial method, which focuses on bead-level observations.
}
\details{
The \code{SlideSeqV1} object is based on the \code{SpatialMethod} class and contains the following specifications:
\itemize{
\item \code{method_specifics}: A list with the following element:
\itemize{
\item \code{diameter}: Diameter of each bead is "10um".
}
\item \code{name}: The name of the spatial method, which is "SlideSeqV1".
\item \code{observational_unit}: The unit of observation, which is "bead".
\item \code{unit}: The SI unit used, which is "mm".
\item \code{version}: The version of the SPATA2 package being used.
}
}
\keyword{datasets}
