% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-data-frame-functions.R
\name{smoothSpatially}
\alias{smoothSpatially}
\title{Smooth numeric variables spatially}
\usage{
smoothSpatially(
  coords_df,
  variables,
  smooth_span = 0.025,
  normalize = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\itemize{
 \item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
 \item{\emph{sample} Character. The sample belonging of every barcode-spot.}
 \item{\emph{x} Numeric. The x-coordinates of every barcode.}
 \item{\emph{y} Numeric. The y-coordinates of every barcode.}
}}

\item{variables}{Character vector. Specifies the numeric variables of the
input data.frame that are to be smoothed.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input data.frame containing the smoothed variables.
}
\description{
Uses a loess-fit model to smooth numeric variables spatially.
The variable names denoted in argument \code{variables} are overwritten.
}
