% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snn_clustering}
\alias{snn_clustering}
\title{A list of data.frames that each contain two variables: \emph{barcodes} and \emph{snn}.
The latter contains the grouping according to Shared Nearest Neighbour clustering.}
\format{
An object of class \code{list} of length 1.
}
\usage{
snn_clustering
}
\description{
@docType data
@usage data(snn_clustering)
}
\keyword{datasets}
