% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{spata-class}
\alias{spata-class}
\alias{spata}
\title{spata object}
\value{
S4 object
}
\description{
spata object
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{A data.frame containing information about every barcode-spot. Must contain the variables:

 \describe{
  \item{\emph{barcodes}}{Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
  \item{\emph{sample}}{Character. The sample belonging of every barcode-spot.}
  \item{\emph{x}}{Numeric. The x-coordinates of every barcode.}
  \item{\emph{y}}{Numeric. The y-coordinates of every barcode.}
 }}

\item{\code{data}}{See documentation for S4-object 'data'}

\item{\code{dea}}{A list in which every slot is named according to a discrete feature for which differential gene expression
analysis has been conducted (via \code{findDeGenes()}). Every slot contains a data.frame (output of \code{Seurat::FindAllMarkers()}).}

\item{\code{dim_red}}{See documentation for S4-object 'dim_red'}

\item{\code{fdata}}{A data.frame containing the additionally computed features. Must contain the variables:
\describe{
 \item{\emph{barcodes}}{Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
 \item{\emph{sample}}{Character. The sample belonging of every barcode-spot.}
}}

\item{\code{image}}{A list of images named according to the samples the object contains.}

\item{\code{samples}}{Character value. Contains the sample names.}

\item{\code{scvelo}}{Currently not in use.}

\item{\code{trajectories}}{A list named according to the samples the object contains. Each slot in
that list contains another list of all 'spatial_trajectory'-objects created for that sample.}

\item{\code{used_genesets}}{A data.frame containing the defined gene-sets. Must contain the variables:

\describe{
 \item{\emph{ont}}{Character. The gene-set name.}
 \item{\emph{gene}}{Character. The belonging genes.}
 }}

\item{\code{version}}{A list of four slots denoting the version of SPATA under which the object has been
created.}

\item{\code{compatibility}}{A list of miscellaneous information that mainly ensures compatibility between different
platforms.}
}}

