% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SPATA2-class}
\alias{SPATA2-class}
\alias{SPATA2}
\title{The \code{SPATA2} - class}
\description{
The \code{SPATA2} - class
}
\section{Slots}{

\describe{
\item{\code{assays}}{A list of assays containing molecular data.}

\item{\code{compatibility}}{A list detailing compatibility information.}

\item{\code{dim_red}}{A list containing dimensionality reduction results.}

\item{\code{images}}{An object of class 'HistoImaging' representing histological images.}

\item{\code{logfile}}{A data frame containing log file data.}

\item{\code{meta_obs}}{A data frame containing metadata for observations.}

\item{\code{meta_info}}{A list providing additional metadata information.}

\item{\code{method}}{An object of class 'SpatialMethod' detailing the spatial method used.}

\item{\code{obj_info}}{A list containing additional object information.}

\item{\code{sample}}{A character vector of sample identifiers.}

\item{\code{spatial}}{A list containing spatial data.}

\item{\code{version}}{A list specifying versioning details.}
}}

