% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SPATA2-class}
\alias{SPATA2-class}
\alias{SPATA2}
\title{The \code{SPATA2} - class}
\description{
This S4 class represents a spatial multiomics data object, containing various
assays, compatibility information, dimensionality reduction results, histological
images, log file data, metadata for observations, additional metadata information,
spatial method details, object information, sample identifiers, spatial data, and versioning details.
}
\section{Slots}{

\describe{
\item{\code{assays}}{A named list of \code{\link{MolecularAssay}} objects containing molecular data. Names
correspond to slot @omic of the respective omic.}

\item{\code{compatibility}}{A list for compatibility information and data.}

\item{\code{data_add}}{A list for additional data that has not yet a fixed slot in
\code{SPATA2} architecture.}

\item{\code{dim_red}}{A list containing dimensionality reduction results.}

\item{\code{logfile}}{A data frame containing log file data.}

\item{\code{meta_obs}}{A data frame containing metadata for the \link[=concept_observations]{observations}. This comprises all
\link[=concept_variables]{variables} that are \strong{not} found in the molecular assays including cluster and grouping variables,
summary statistics, etc.}

\item{\code{meta_sample}}{A list providing additional information about the tissue and the tissue donor.}

\item{\code{method}}{An object of class 'SpatialMethod' detailing the spatial method used.}

\item{\code{obj_info}}{A list containing object information such as default instructions and
directories.}

\item{\code{sample}}{A character value. The name of the sample and the \code{SPATA2} object.}

\item{\code{spatial}}{An object of class \code{\link{SpatialData}} storing coordinates of the observations,
images, required transformations for alignment, scale factors, \code{\link{SpatialAnnotation}} objects,
\code{\link{SpatialTrajectory}} objects, etc.}

\item{\code{version}}{A list specifying versioning details.}
}}

