% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{spatial_gradient_screening}
\alias{spatial_gradient_screening}
\title{Low level implementation of the spatial gradient screening}
\usage{
spatial_gradient_screening(
  coords_df,
  variables,
  resolution,
  cf = 1,
  rm_zero_infl = TRUE,
  n_random = 10000,
  sign_var = "fdr",
  sign_threshold = 0.05,
  skip_comp = FALSE,
  force_comp = FALSE,
  model_subset = NULL,
  model_add = NULL,
  model_remove = NULL,
  control = SPATA2::sgs_loess_control,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{coords_df}{A data.frame that contains at least a numeric variable named
\emph{dist} as well the numeric variables denoted in \code{variables}.}

\item{variables}{Character vector of numeric variable names that are integrated
in the screening process.}

\item{resolution}{Units value of the same unit of the \emph{dist} variable in
\code{coords_df}.}

\item{n_random}{Number of random permutations for the significance testing of step 2.}

\item{sign_var}{Either \emph{p_value} or \emph{fdr}. Defaults to \emph{fdr}.}

\item{sign_threshold}{The significance threshold. Defaults to 0.05.}

\item{control}{A list given to \code{control} of \code{\link[stats:loess]{stats::loess()}}.}

\item{seed}{Numeric value. Sets the random seed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A list of four slots:

\itemize{
\item{variables}: A character vector of the names of all variables included
in the screening.
\item{model_df}: A data.frame of the models used for step 3.
\item{loess_models}: A named list of loess models for all variables
integrated in the screening process. Names correspond to the variable names.
\item{pval}: Data.frame of three variables: \emph{variable}, \emph{lds}, \emph{p_value} and \emph{fdr}.
Contains the results of step 2. Each observation corresponds to the inferred
gradient of a variable.
\item{eval}: Data.frame of five variable: \emph{variable}, \emph{model}, \emph{corr}, \emph{mae}
\emph{rmse}. Contains the results of step 3. Each observation corresponds to a
gradient ~ model fit. Variables correspond to the evaluation metrics of
the fit.
}
}
\description{
Conducts spatial gradient screening. See details for more information.
}
