% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{spatial_methods}
\alias{spatial_methods}
\title{List of Spatial Methods}
\format{
An object of class \code{list} of length 7.
}
\usage{
spatial_methods
}
\description{
This list contains predefined spatial methods for various spatial biology platforms. Each entry is an instance of the \code{SpatialMethod} class with specifications tailored to a specific platform.
}
\details{
The \code{spatial_methods} list includes the following spatial methods:
\itemize{
\item \link{MERFISH}: A spatial method focusing on cell-level observations using the Multiplexed Error-Robust Fluorescence In Situ Hybridization technique.
\item \link{SlideSeqV1}: A spatial method focusing on bead-level observations using the Slide-seq V1 platform.
\item \code{Undefined}: A default spatial method with undefined properties.
\item \link{VisiumSmall}: A spatial method for small capture areas using the Visium platform.
\item \link{VisiumLarge}: A spatial method for large capture areas using the Visium platform.
\item \link{VisiumHD}: A high-density spatial method using the Visium platform.
\item \link{Xenium}: A spatial method focusing on cell-level observations using the Xenium platform.
}

Each spatial method in the list is an instance of the \code{SpatialMethod} class and contains specific configurations such as the unit of observation, SI unit, and method-specific details.
}
\examples{
\dontrun{
  # Access the VisiumLarge spatial method
  visium_large_method <- spatial_methods[["VisiumLarge"]]
  print(visium_large_method)
}

}
\keyword{datasets}
