% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_trajectories}
\alias{spatial_trajectories}
\title{A list of lists of spatial spatial_trajectories for specific \code{SPATA2}
objects - available via the \code{SPATAData} package. Names of the list slots
correspond to the sample names in variable  \emph{sample} of \code{SPATAData::source_df}.}
\format{
An object of class \code{list} of length 2.
}
\usage{
spatial_trajectories
}
\description{
Every sample specific slot is another list that contains S4 objects of class \code{SpatialTrajectory}.
Use \code{setSpatialTrajectory()} to add them to the corresponding object.
}
\details{
@docType data
@usage data(spatial_trajectories)
}
\keyword{datasets}
