% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{spatial_trajectory-class}
\alias{spatial_trajectory-class}
\alias{spatial_trajectory}
\title{The \code{spatial_trajectory}- Class}
\description{
Deprecatd in favor of \code{SpatialTrajectory}.
}
\keyword{internal}
