% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialAnnotation-class}
\alias{SpatialAnnotation-class}
\alias{SpatialAnnotation}
\title{The \code{SpatialAnnotation} - Class}
\description{
S4 class that represents annotations for spatial data,
allowing users to define and store polygons that outline areas of interest
within images or datasets. It serves as the overarching class for different
spatial annotation methods.
}
\details{
The following classes are derivatives of this class: \code{\link{GroupAnnotation}},
\code{\link{ImageAnnotation}}, \code{\link{NumericAnnotation}}
}
\section{Slots}{

\describe{
\item{\code{area}}{list. A named list of data.frames with the numeric variables \emph{x_orig} and \emph{y_orig}.
Observations correspond to the vertices of the polygons that are needed to represent the
image annotation. \strong{Must} contain a slot named \emph{outer} which sets the outer border
of the image annotation. \strong{Can} contain multiple slots named \emph{inner} (suffixed)
with numbers that correspond to inner polygons - holes within the annotation.}

\item{\code{id}}{character. String to identify the object in a list of multiple objects
of the same class.}

\item{\code{image}}{A cropped version of an image that focuses solely on the area containing the
annotation, along with an expansion margin. This slot is designed to remain
empty until the annotation object is explicitly extracted or used, contributing to efficient
data storage. Extraction through functions like \code{getSpatialAnnotation()} or
\code{getSpatialAnnotation()} populates this slot with the cropped image. The parameters
used for cropping the image are stored in the \verb{@image_info} slot.}

\item{\code{image_info}}{A list containing information related to the image stored in the @image slot.
This information pertains to the cropped image obtained and set using functions like
\code{getSpatialAnnotation()} or \code{getSpatialAnnotations()}. It serves as metadata
around the cropped image and may include parameters or details about the cropping process.}

\item{\code{misc}}{list. A flexible list for miscellaneous input.}

\item{\code{sample}}{Character string. The sample to which the annotation belongs.}

\item{\code{tags}}{character. Vector of arbitrary length. Contains tags that can be used
to group and select spatial annotations in different manners.}

\item{\code{version}}{A list of three slots denoting the version of \code{SPATA2} under
which the object has been created.}
}}

\section{Selection of spatial annotations with tags}{


Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the spatial annotations of interest. There are multiple options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags. Can contain tags that are not specified.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags. Can not be tagged with anything else.
\item Argument \code{test} set to \emph{not_identical} or \emph{4}: To be included, an image
annotation must \strong{not} be tagged with the combination of input tags.
\item Argument \code{test} set to \emph{'none'} or \emph{5}: To be included, an image annotation
must \strong{not} contain any of the input tags.
}

Note that the filtering process happens after the filtering by input for argument
\code{ids}. You can first select a group of annotations by naming their IDs
and then select among them via tags and test. If \code{ids} is \code{NULL}, you select
among all annotations via tags and test. And if \code{tags} is also \code{NULL},
the function uses all  annotations.
}

