% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialAnnotationScreening-class}
\alias{SpatialAnnotationScreening-class}
\alias{SpatialAnnotationScreening}
\title{The \code{SpatialAnnotationScreening} - Class}
\description{
S4 class that contains input for and output of the
function \code{\link[=spatialAnnotationScreening]{spatialAnnotationScreening()}}.
}
\details{
This class is an extension of class \code{\link{SpatialGradientScreening}} and inherits
all of its slots.
}
\section{Slots}{

\describe{
\item{\code{annotations}}{list. A list of the \code{\link{SpatialAnnotation}} objects used in
the screening.}
}}

