% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{spatialAnnotationScreening}
\alias{spatialAnnotationScreening}
\title{Implementation of the SAS-algorithm}
\usage{
spatialAnnotationScreening(
  object,
  variables,
  id = idSA(object),
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  core = TRUE,
  periphery = TRUE,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  mtr_name = NULL,
  bcs_exclude = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{n_bins_dist}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}. Not required if \code{distance} and \code{binwidth} is
specified. (See details for more.)}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{bcs_exclude}{Character value containing name(s) of data points to be excluded from the analysis.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{SpatialAnnotationScreening}. See documentation
with \code{?SpatialAnnotationScreening} for more information.
}
\description{
Screens the sample for numeric variables that stand
in meaningful, spatial relation to annotated structures/areas.
For a detailed explanation on how to define the parameters \code{distance},
\code{n_bins_dist}, \code{binwidth}, \code{angle_span} and \code{n_bins_angle}
see details section.
}
\details{
In conjunction with argument \code{id} which provides the
ID of the spatial annotation of interest the arguments \code{distance},
\code{binwidth}, \code{n_bins_dist}, \code{angle_span} and \code{n_bins_angle} can be used
to specify the exact area that is screened as well as the resolution of the screening.

\bold{How the algorithm works:} During the SAS-algorithm the data points are
binned according to their localisation to the spatial annotation. Every bin's mean
expression of a given gene is then aligned in an ascending order - mean expression
of bin 1, mean expression of bin 2, ... up to the last bin, the bin with the
data points that lie farest away from the spatial annotation. This allows to infer
the gene expression changes in relation to the spatial annotation and
to screen for genes whose expression changes resemble specific biological
behaviors. E.g. linear ascending: gene expression increases linearly with
the distance to the spatial annotation. E.g. immediate descending: gene expression
is high in close proximity to the spatial annotation and declines logarithmically
with the distance to the spatial annotation.

\bold{How circular binning works:}
To bin data points according to their localisation to the spatial annotation
three parameters are required:

\itemize{
\item{\code{distance}: The distance from the border of the spatial annotation to
the \emph{horizon} in the periphery up to which the screening is conducted.
}
\item{\code{binwidth}: The width of every bin.}
\item{\code{n_bins_dist}: The number of bins that are created.}
}

These three parameters stand in the following relation to each other:

\enumerate{
\item{\code{n_bins_dist} = \code{distance} / \code{binwidth}}
\item{\code{distance} = \code{n_bins_dist} * \code{binwidth}}
\item{\code{binwidth} = \code{distance} / \code{n_bins_dist}}
}

Therefore, only two of the three arguments must be specified as the remaining
one is calculated. We recommend to stick to the first option: Specifying
\code{distance} and \code{binwidth} and letting the function calculate
\code{n_bins_dist}.

Once the parameters are set and calculated the polygon that is used to
define the borders of the spatial annotation is repeatedly expanded by the distance
indicated by parameter \code{binwidth}. The number of times this expansion is
repeated is equal to the parameter \code{n_bins_dist}. Every time the
polygon is expanded, the newly enclosed data points are binned (grouped)
and the bin is given a number that is equal to the number of the expansion.
Thus, data points that are adjacent to the spatial annotation are binned into
bin 1, data points that lie a distance of \code{binwidth} away are binned into
bin 2, etc.

Note that \code{\link[=plotSurfaceSAS]{plotSurfaceSAS()}} and/or \code{\link[=ggpLayerExpansionsSAS]{ggpLayerExpansionsSAS()}} allow to
visually inspect if your input results in the desired screening.

\bold{How the screening works:} For every numeric variable (e.g. genes) that
is included in the screening process every bin's mean expression is calculated
and then aligned in an ascending order - mean expression of distance bin 1,
mean expression of distance bin 2, ... up to the last bin, namely the bin
with the data points that lie farest away from the spatial annotation. This
allows to infer the gene expression changes in relation to the spatial annotation and
to screen for genes whose expression changes resemble specific biological
behaviors. The gene expression change is fitted to every model that is included.
(Use \code{showModels()} to visualize the predefined models of \code{SPATA2}).
A gene-model-fit is evaluated twofold:

\itemize{
\item{Mean Absolute Error}: Description.
\item{Root Mean Squared Error}: Description.
\item{Pearson correlation}: The inferred expression changes is correlated
with the model. (Correlation as well as the corresponding p-value depend
on the number of bins!)
}
}
\seealso{
\code{\link[=createGroupAnnotations]{createGroupAnnotations()}}, \code{\link[=createImageAnnotations]{createImageAnnotations()}},
\code{\link[=createNumericAnnotations]{createNumericAnnotations()}} for how to create spatial annotations.

\code{\link[=getCoordsDfSA]{getCoordsDfSA()}} for how to obtain spatial relation of data points to
a spatial annotation.

\code{\link[=getSasDf]{getSasDf()}} for how to obtain inferred expression gradients as used in
spatial annotation screening.

\code{\link[=plotSasLineplot]{plotSasLineplot()}} for visualization of inferred expression gradients.
}
