% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{spatialAnnotationToGrouping}
\alias{spatialAnnotationToGrouping}
\title{Convert spatial annotation to grouping variable}
\usage{
spatialAnnotationToGrouping(
  object,
  ids,
  grouping_name,
  inside = "inside",
  outside = "outside",
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector. Specifies the spatial annotation(s) of interest.
data points that fall into the area of these annotations are labeled
with the input for argument \code{inside}.}

\item{grouping_name}{Character value. The name of the new segmentation variable.}

\item{inside}{Character value. The group name for the data points that
are located inside the area of the spatial annotation(s).}

\item{outside}{Character value. The group name for the data points that
are located outside the area of the spatial annotation(s).}

\item{overwrite}{Logical. Set to TRUE to overwrite existing variables with
the same name.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Converts one or more spatial annotations to a binary
segmentation variable in the feature data.frame.
}
\examples{

library(SPATA2)
library(patchwork)

data("example_data")

object <- example_data$object_UKF275T_diet

object <-
 createNumericAnnotations(
   object = object,
   variable = "HM_HYPOXIA",
   threshold = "kmeans_high",
   id = "hypoxia_ann",
   inner_borders = FALSE,
   force1 = TRUE
   )


 object <-
   spatialAnnotationToGrouping(
     object = object,
     ids = "hypoxia_ann",
     grouping_name = "hyp_group"
     )

 plotSurface(object, "hyp_group")

 plotBoxplot(object, variables = "HM_HYPOXIA", group_by = "hyp_group")

}
