% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialData-class}
\alias{SpatialData-class}
\alias{SpatialData}
\title{The \code{SpatialData} - Class}
\description{
S4 class that represents spatial data. This includes the spatial
positions of \emph{\link[=concept_observations]{observations}}, a set of histological images from one
tissue slide or several consecutive slides of one and the same tissue portion as well
as additional spatial aspects such as \emph{\link[=SpatialAnnotation]{spatial annotations}}
or \emph{\link[=SpatialTrajectory]{spatial trajectories}}.
}
\section{Slots}{

\describe{
\item{\code{annotations}}{list. List of objects of class \code{\link{SpatialAnnotation}}.}

\item{\code{coordinates}}{data.frame. Data.frame that stores information about identified
or known entities located on the imaged tissue, such as cells or capture spots.}

\item{\code{images}}{list. List of objects of class \code{\link{HistoImage}} - the container objects
for images.}

\item{\code{method}}{SpatialMethod. Object of class \code{\link{SpatialMethod}}.}

\item{\code{meta}}{list. List for meta data regarding the imaged tissue portion.}

\item{\code{misc}}{list. A flexible list for miscellaneous input.}

\item{\code{name_img_active}}{character. The name of the image that is currently active.}

\item{\code{name_img_ref}}{character. The name of the image that is used as a reference for aligning
every additional image in slot @images}

\item{\code{outline}}{list. List of two data.frames in which each row corresponds to
a vertice of the polygon required to outline all \link[=concept_observations]{observations}
marking the edge of tissue sections.

\itemize{
\item{\emph{tissue_whole}:}{ Data.frame of two variables \emph{x_orig} and \emph{y_orig}.}
\item{\emph{tissue_sections}:} {Data.frame of two variables \emph{x_orig}, \emph{y_orig} and \emph{section} to
outline the tissue section outlined.}
}

See the vignette about \code{\link[=concept_tissue_outline]{tissue outline}} for more information.}

\item{\code{sample}}{character. String to identify the imaged tissue.}

\item{\code{scale_factors}}{list. A list of scale factors. Only required if slot @images does not contain any image.
See \code{\link[=concept_scale_factors]{scale factors}} for more information.}

\item{\code{trajectories}}{list. A list of objects of class \code{\link{SpatialTrajectory}}.}
}}

