% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialGradientScreening-class}
\alias{SpatialGradientScreening-class}
\alias{SpatialGradientScreening}
\title{The \code{SpatialGradientScreening} - Class}
\description{
S4 class that contains input for and output of spatial gradient
screening implementation \code{\link[=spatialAnnotationScreening]{spatialAnnotationScreening()}} and \code{\link[=spatialTrajectoryScreening]{spatialTrajectoryScreening()}}.
}
\details{
The following classes are derivates of this class: \code{\link{SpatialAnnotationScreening}},
\code{\link{SpatialTrajectoryScreening}}.
}
\section{Slots}{

\describe{
\item{\code{coordinates}}{data.frame. A data.frame as obtained by \code{\link[=getCoordsDfSA]{getCoordsDfSA()}} or \code{\link[=getCoordsDfST]{getCoordsDfST()}},
respectively, with the setup used for the screening.}

\item{\code{models}}{data.frame. A data.frame in which columns correspond to the
modeled gradients used for the model fitting.}

\item{\code{qc}}{list. A named list that contains results from the quality control
steps. Reserved list slots are:

\itemize{
\item{cf}{ Numeric value. The correction factor used to adjust the span.}
\item{est_R2}{ List. Results from estimating the overal R2 between noise and total variation.}
\item{zero_infl_vars}{ Character vector. Names of variables that have been identified as zero inflated.}
}}

\item{\code{results}}{A named list that contains the screening results. Reserved list
slots are:

\itemize{
\item{significance}{ Data frame with results from the identification of non-random gradients.}
\item{model_fits}{ Data frame with results from gradient ~ model fits.}
}}

\item{\code{sample}}{Character value. The name of the \code{\link{SPATA2}} object with which
the screening was conducted.}

\item{\code{set_up}}{A list of parameters required to reproduce the set up of the
screening.}
}}

