% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialSegment-class}
\alias{SpatialSegment-class}
\alias{SpatialSegment}
\title{The \code{SpatialSegment} - class}
\description{
Abstracts the concept of an annotated segment within a spatial
segmentation.
}
\section{Slots}{

\describe{
\item{\code{info}}{list. Stores meta data and miscellaneous information regarding the
spatial segment. Slots that should always exist:
\itemize{
\item{image_origin:}{ Character string. Content of slot @info$origin of the \code{HistologyImaging} at the
time the segment was drawn.}
\item{parent_id:}{ Character string. The ID of the spatial segmentation this segment is part of.}
\item{pot:}{ POSIXct. The point of time when the segment was drawn. Used to handle overlapping
segments.}
\item{sample:}{ Character string. The name of the sample (slot @sample of the \code{SPATA2} object.)}
}}

\item{\code{label}}{character. Character string. The label that was given to the segment.
Corresponds to the group name of the barcode-spots that fall into the segment.}

\item{\code{polygons}}{list. List of data.frames with \emph{x} and \emph{y} variables that contain
the vertices of the polygon. The first polygon (should be named \emph{outer}) defines
the outer ring of the segment. Further polygons (should be named \emph{inner} suffixed
with a number) define holes within the segment.}
}}

