% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialSegmentation-class}
\alias{SpatialSegmentation-class}
\alias{SpatialSegmentation}
\title{The \code{SpatialSegmentation} - Class}
\description{
Abstracts the concept of manual segmentation/annotation
of the sample surface.
}
\section{Slots}{

\describe{
\item{\code{id}}{character. String to identify the object in a list of multiple objects of
the same class.}

\item{\code{info}}{list. Stores meta data and miscellaneous information regarding the
spatial segmentation. Slots that should always exist:
\itemize{
\item{sample:}{ Character string. The name of the sample (slot @sample of the \code{SPATA2} object.)}
}}

\item{\code{segments}}{list. A named, nested list. Named according to the labels
given to each segment. E.g. list of length two with slot \emph{necrosis} and \emph{vivid}.
Each named slot is a unnamed list. In this list each slot is a list of data.frames with
a \emph{x} and a \emph{y} variable. First data.frame, named \emph{exterior} corresponds to the exterior border of
the segment. Every following data.frame is named \emph{interior}-suffix where the suffix is a number
and corresponds to interior holes of the segment.}
}}

\keyword{internal}
