% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialTrajectory-class}
\alias{SpatialTrajectory-class}
\alias{SpatialTrajectory}
\title{The \code{SpatialTrajectory} - class}
\description{
Extension of the \code{Trajectory} for trajectories
that have been drawn in real space.
}
\section{Slots}{

\describe{
\item{\code{coords}}{data.frame. The coordinates data.frame of the sample.}
}}

