% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialTrajectoryScreening-class}
\alias{SpatialTrajectoryScreening-class}
\alias{SpatialTrajectoryScreening}
\title{The \code{SpatialTrajectoryScreening} - class}
\description{
S4 class that contains input for and output of the
function \code{spatialTrajectoryScreening()}.
}
\section{Slots}{

\describe{
\item{\code{binwidth}}{numeric. The width of the bins in which the barcode-spots
are put based on the projection length values.}

\item{\code{coords}}{data.frame. Coordinates data.frame of the sample.}

\item{\code{id}}{character. The ID of the screened trajectory.}

\item{\code{method_padj}}{character. The method with which p-values were adjusted.}

\item{\code{models}}{data.frame. The model data.frame that has been used for the
screening.}

\item{\code{n_bins}}{numeric. The number of bins in which the barcode-spots
are distributed based on their projection length.}

\item{\code{results}}{data.frame. A data.frame that contains the model evaluations}

\item{\code{sample}}{character. The sample name.}

\item{\code{summarize_with}}{character. The name of the function that has been
used to summarize the variables by bin.}

\item{\code{spatial_trajectory}}{SpatialTrajectory. The spatial trajectory based on
which the screening took place.}
}}

