% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{splitSpataObject}
\alias{splitSpataObject}
\title{Split SPATA2 object}
\usage{
splitSpataObject(object, grouping, reduce = FALSE, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{reduce}{A logical value indicating whether to reduce the \code{SPATA2} sub-objects after splitting.
Default is \code{FALSE}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A named list of \code{SPATA2} sub-objects split by the specified grouping.
}
\description{
This function splits a \code{\link{SPATA2}} object into multiple sub-objects based on a
specified grouping variable.
}
\examples{

library(SPATA2)

object <- example_data$object_lmu_mci_object

object <- identifyTissueOutline(object)

plotSurface(object, color_by = "tissue_section")

obj_list <- splitSpataObject(object, grouping = "tissue_section")

purrr::map(obj_list, .f = ~ .x)

}
