% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{subsetByBarcodes}
\alias{subsetByBarcodes}
\title{Simple subsetting by barcodes}
\usage{
subsetByBarcodes(object, barcodes, verbose = NULL)
}
\arguments{
\item{barcodes}{Character vector. The barcodes of the barcode spots that are
supposed to be \bold{kept}.}
}
\value{
An updated \code{spata2} object.
}
\description{
Removes unwanted barcode spots from the object without any significant
post processing.
}
\details{
Unused levels of factor variables in the feature data.frame are dropped
and directory settings are reset to NULL.
}
