% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{summarize_batch_reduce_visium_hd}
\alias{summarize_batch_reduce_visium_hd}
\title{Summarize and reduce Visium HD data by batch}
\usage{
summarize_batch_reduce_visium_hd(batch)
}
\arguments{
\item{batch}{A list containing the following elements:
\itemize{
\item \code{mtr}: A matrix (sparse or dense) of gene expression counts, with genes as rows and barcodes as columns.
\item \code{df}: A data frame containing barcode information, including a column named \emph{barcodes} and a column named \emph{barcodes_new} for the grouping operation.
}}
}
\value{
A matrix containing the aggregated gene expression data, with genes as columns and the new barcodes as rows.
The values in the matrix represent the sum of gene expression counts for each gene across the grouped barcodes.
}
\description{
This function processes and reduces spatial transcriptomics data from a Visium HD batch.
It aggregates gene expression data across specified barcodes, groups the data by a new set of barcodes,
and returns a summarized matrix.
}
\details{
The function performs the following steps:
\enumerate{
\item Subsets the count matrix to include only the barcodes listed in the data frame.
\item Converts the subsetted count matrix to a data frame and merges it with the barcode data frame.
\item Groups the data by the new barcode identifiers and sums the gene expression counts across these groups.
\item Converts the summarized data back into a matrix format for further analysis.
}
}
\keyword{internal}
