% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{summarize_formulas}
\alias{summarize_formulas}
\title{List of summarizing formulas}
\format{
An object of class \code{list} of length 5.
}
\usage{
summarize_formulas
}
\description{
List of summarizing formulas
}
\keyword{internal}
