% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{summarizeIAS}
\alias{summarizeIAS}
\title{Summarize IAS-results}
\usage{
summarizeIAS(ias, method_padj = "fdr")
}
\description{
Summarizes the results of the IAS-algorithm. Creates
the content of slot @results_smrd of the \code{ImageAnnotationScreening}-class.
}
\details{
Model fitting and evaluation happens within every angle-bin.
To get a single evaulation for every gene the results of every
angle-bin must be summarized.
}
