% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-add-ons.R
\name{surfaceFrame}
\alias{surfaceFrame}
\title{Add frame of the sample}
\usage{
surfaceFrame(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
ggproto object refering to the x- and y-scales
}
\description{
Add frame of the sample
}
