% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-dashboard-tab-items.R
\name{tab_create_segmentation_return}
\alias{tab_create_segmentation_return}
\title{Segmentation plot tab - return}
\usage{
tab_create_segmentation_return()
}
\description{
Segmentation plot tab - return
}
