% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-dashboard-tab-items.R
\name{tab_create_trajectories_return}
\alias{tab_create_trajectories_return}
\title{Trajectory plot tab - return}
\usage{
tab_create_trajectories_return()
}
\description{
Trajectory plot tab - return
}
