% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{tissueOutlineIdentified}
\alias{tissueOutlineIdentified}
\alias{tissueSectionsIdentfied}
\title{Check availability of tissue information}
\usage{
tissueOutlineIdentified(object)

tissueSectionsIdentfied(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Logical value.
}
\description{
Checks if \code{identifyTissueSections()} and \code{identifyTissueOutline()}
has been run successfully.
}
