% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{Trajectory-class}
\alias{Trajectory-class}
\alias{Trajectory}
\title{The \code{Trajectory} - Class}
\description{
S4 class that represents trajectories
drawn within real or latent space.
}
\section{Slots}{

\describe{
\item{\code{comment}}{character. A comment about why the trajectory was drawn.}

\item{\code{id}}{character. ID that uniquely identfies the trajectory in a sample.}

\item{\code{projection}}{data.frame. Data.frame that contains the length of the
projection of each barcode spot onto the trajectory.}

\item{\code{sample}}{character. The sample name.}

\item{\code{segment}}{data.frame. Contains the course of the trajetory in
form of a data.frame with the variables \emph{x, y, xend} and \emph{yend.}}

\item{\code{width}}{numeric. The width of the rectangle that was spanned along
the trajectory. (Length of the rectangle corresponds to the length of
the segment.)}
}}

\keyword{internal}
