% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\docType{data}
\name{trajectory_patterns}
\alias{trajectory_patterns}
\title{Trajectory patterns}
\format{
An object of class \code{character} of length 14.
}
\usage{
trajectory_patterns
}
\description{
Character vectors containing the names of valid trajectory patterns.
}
\keyword{datasets}
