% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{transfer_slot_content}
\alias{transfer_slot_content}
\title{Transfer S4 slot content}
\usage{
transfer_slot_content(recipient, donor, skip = character(0), verbose = TRUE)
}
\arguments{
\item{recipient}{Empty and new S4 object.}

\item{donor}{Old S4 object.}

\item{skip}{Slot names whose content is not transferred.}
}
\value{
Updated S4 object.
}
\description{
Transfers slot content from one S4 object (donor) to a newer
version (recepient).
}
