% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{transform_area_si_to_pixel}
\alias{transform_area_si_to_pixel}
\alias{transform_area_si_to_pixels}
\title{Convert area in SI units to pixel}
\usage{
transform_area_si_to_pixel(input, object, round = FALSE)

transform_area_si_to_pixels(input, object, round = FALSE, as_numeric = TRUE)
}
\arguments{
\item{input}{Area in SI units. See details for more information.}

\item{object}{A valid \code{SPATA2} object or \code{NULL}. If specified the
distance scaling is adjusted to the current resolution of the image inside
the object. If \code{NULL}, \code{image_dims} and \code{method} must be specified.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}

\item{as_numeric}{Logical value. If \code{TRUE}, forces the output to be numeric.
This means that the unit is not \bold{not} suffixed.}
}
\value{
Transformed input. Vector of the same length as input. Function
\code{transform_area_si_to_pixel()} always returns a single numeric value. Function
\code{transform_si_to_pixels()} returns a numeric vector by default. If \code{as_numeric}
is \code{FALSE}, the output is a string suffixed with \emph{px}.
}
\description{
Transforms area in SI units to pixel based on the current
resolution of the image in the \code{SPATA2} object.
}
\keyword{internal}
