% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{transform_coords}
\alias{transform_coords}
\title{Transform coordinates}
\usage{
transform_coords(coords_df, transformations, center, ranges, ...)
}
\arguments{
\item{transformations}{List of transformation instructions. See
slot @transformations of class \code{HistoImage}.}

\item{outline_df}{Data.frame with x- and y-coordinates.}
}
\value{
Transformed input.
}
\description{
Applies spatial linear transformations on a set of points
in a Cartesian coordinate system.
}
