% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{transform_eUOL_to_pixel}
\alias{transform_eUOL_to_pixel}
\alias{transform_eUOL_to_pixels}
\title{Transforms European Units of Length to pixels}
\usage{
transform_eUOL_to_pixel(
  input,
  object = NULL,
  image_dims = NULL,
  method = NULL,
  round = FALSE
)

transform_eUOL_to_pixels(
  input,
  object = NULL,
  image_dims = NULL,
  method = NULL,
  round = FALSE,
  as_numeric = FALSE
)
}
\arguments{
\item{input}{Distance as European unit of length. See details for more.}

\item{object}{A valid \code{SPATA2} object or \code{NULL}. If specified the
distance scaling is adjusted to the current resolution of the image inside
the object. If \code{NULL}, \code{image_dims} and \code{method} must be specified.}

\item{image_dims}{Numeric vector of length two. Specifies the dimensions
of the image to which the distance is scaled. First value corresponds to
the width, second value corresponds to the height of the image.

Ignored if \code{object} is specified (carries needed information).}

\item{method}{The spatial -omic method by name as a character value or S4 object
of class \code{SpatialMethod}. Specifies the method and thus the frame size of
the original image in European units of length. If character, must be one of \code{validSpatialMethods()}.

Ignored if \code{object} is specified (carries needed information).}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}

\item{as_numeric}{Logical value. If \code{TRUE}, forces the output to be numeric.
This means that the unit is not \bold{not} suffixed.}
}
\value{
Transformed input. Vector of the same length as \code{input}.
\code{transform_eUOL_to_pixel()} returns always a numeric value.
}
\description{
Transforms European units of length (e.g. \emph{'2mm'}, \emph{'400.50um'})
to pixel values depending on the original size of spatial -omic methods.
}
\details{
Input to specify a distance can be provided in two different
ways.

\bold{Distance in pixel}:

There are two valid input options to specify the distance in pixel:

\itemize{
\item{numeric:}{ Single numeric values, e.g. 2, 3.554, 69, 100.67. If no unit
is specified the input will be interpreted as pixels.}
\item{character:}{ Suffixed with \emph{'px'}, e.g. \emph{'2px'}, \emph{'3.554px'}}
}

\bold{Distance in European units of length (eUOL)}:

Specifying distances in European units of length e.g. \emph{'2mm'}, \emph{'400um'} etc.
requires the input to be a character as the unit must be provided as suffix. Between the numeric
value and the unit must be no empty space! Unit suffixes must be one of
\emph{'m', 'dm', 'cm', 'mm', 'um', 'nm'}.
}
\note{
\code{transform_eUOL_to_pixel()} transforms only single values. \code{transform_eUOL_to_pixels()}
transforms vectors of lengths one or more.
}
