% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{transform_image}
\alias{transform_image}
\title{Transform image}
\usage{
transform_image(image, transformations, bg_col = "white")
}
\arguments{
\item{image}{Image comptabible with the \code{EBImage}-package.}

\item{transformations}{List of transformation instructions. See
slot @transformations of class \code{HistoImage}.}
}
\value{
Transformed input.
}
\description{
Transforms the image or the tissue outline.
}
