% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{transform_outline}
\alias{transform_outline}
\title{Transform outline}
\usage{
transform_outline(outline_df, transformations, center, ranges)
}
\arguments{
\item{outline_df}{Data.frame with x- and y-coordinates.}

\item{transformations}{List of transformation instructions. See
slot @transformations of class \code{HistoImage}.}
}
\value{
Transformed input.
}
\description{
Transforms tissue outline.
}
