% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{transform_pixel_to_area_si}
\alias{transform_pixel_to_area_si}
\title{Converts from pixel to area in SI units}
\usage{
transform_pixel_to_area_si(input, unit, object, round = FALSE)
}
\arguments{
\item{input}{Area as pixel input. See details for more information.}

\item{unit}{The SI area unit. Use \code{validUnitsOfAreaSI()} to obtain all
valid input options.}

\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}
}
\value{
Transformed input. Vector of the same length as \code{input} and of class \code{units}.
}
\description{
Transforms pixel values to SI units (e.g. '\emph{7.5mm2'}, '20um2')
}
