% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R, R/unused.R
\name{transform_pixels_to_area_si}
\alias{transform_pixels_to_area_si}
\alias{transform_pixel_to_si}
\title{dummy}
\usage{
transform_pixels_to_area_si(input, unit, object, round = FALSE)

transform_pixel_to_si()
}
\description{
dummy
}
\keyword{internal}
