% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{transformSpataToCDS}
\alias{transformSpataToCDS}
\title{Deprecated.}
\usage{
transformSpataToCDS(
  object,
  preprocess_method = "PCA",
  reduction_method = c("PCA", "UMAP"),
  cluster_method = "leiden",
  estimate_size_factors = list(),
  preprocess_cds = list(),
  reduce_dimension = list(),
  cluster_cells = list(),
  learn_graph = list(),
  order_cells = list(),
  of_sample = NA,
  verbose = TRUE
)
}
\description{
Deprecated.
}
\keyword{internal}
