% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-family.R
\name{transformSpataToSeurat}
\alias{transformSpataToSeurat}
\title{Transform spata-object to a seurat-object}
\usage{
transformSpataToSeurat(
  object,
  assay_name = "Spatial",
  ...,
  SCTransform = FALSE,
  NormalizeData = list(normalization.method = "LogNormalize", scale.factor = 1000),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  ScaleData = TRUE,
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  RunTSNE = TRUE,
  RunUMAP = list(dims = 1:30),
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{Additional parameters given to \code{Seurat::CreateSeuratObject()}.}

\item{SCTransform}{A named list of arguments given to \code{Seurat::SCTransform()}, TRUE or FALSE.}

\item{NormalizeData}{A named list of arguments given to \code{Seurat::NormalizeData()}, TRUE or FALSE.}

\item{FindVariableFeatures}{A named list of arguments given to \code{Seurat::FindVariableFeatures()}, TRUE or FALSE.}

\item{ScaleData}{A named list of arguments given to \code{Seurat::ScaleData()}, TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument \code{features} input
for argument \code{features} is set to \code{base::rownames(seurat_object)}.}

\item{RunPCA}{A named list of arguments given to \code{Seurat::RunPCA()}, TRUE or FALSE.}

\item{FindNeighbors}{A named list of arguments given to \code{Seurat::FindNeighbors()}, TRUE or FALSE.}

\item{FindClusters}{A named list of arguments given to \code{Seurat::FindClusters()}, TRUE or FALSE.}

\item{RunTSNE}{A named list of arguments given to \code{Seurat::RunTSNE()}, TRUE or FALSE.}

\item{RunUMAP}{A named list of arguments given to \code{Seurat::RunUMAP()}, TRUE or FALSE.}

\item{assay}{Character value. The name under which the count- and expression matrix is to be saved.}
}
\value{
A seurat-object.
}
\description{
Takes the count matrix of your spata-object and creates a
Seurat-object with it. The spata-object's feature-data is passed as input
for the \code{meta.data}-argument of \code{Seurat::CreateSeuratObject()}.
If specified as TRUE or named list of arguments the respective functions are called in
order to pre process the object.
}
\details{
`transformSpataToSeurat()` is a convenient wrapper around all functions that preprocess a seurat-object
after it's initiation. The object is initiated by passing the spata-objects count-matrix and feature data to it whereupon
the functions are called in the order they are presented in this documentation. For all
pre processing functions apply the following instructions:

 \itemize{
  \item{If set to FALSE the processing function is skipped.}
  \item{If set to TRUE the respective function is called with it's default argument settings. Note: \code{RunUMAP()} needs
  additional input!}
  \item{If a named list is provided the respective function is called whereby the named list will provide the argument-input (the seurat-object to be constructed must not be named and will be
  passed to every function as the first argument!!!.)}
  }

Note that certain listed functions require previous functions! E.g. if \code{RunPCA} is set to FALSE \code{RunTSNE()}
will result in an error. (\code{base::tryCatch()} will prevent the function from crashing.)
}
