% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{translate_shape_string}
\alias{translate_shape_string}
\title{inspired from https://github.com/tidyverse/ggplot2/blob/main/R/geom-point.r}
\usage{
translate_shape_string(shape_string)
}
\description{
inspired from https://github.com/tidyverse/ggplot2/blob/main/R/geom-point.r
}
