% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{tutorial_hint_dummy}
\alias{tutorial_hint_dummy}
\title{Title}
\usage{
tutorial_hint_dummy()
}
\description{
Title
}
\section{Tutorials}{

Extensive tutorials for how to use this function can be found on our website
\url{https://themilolab.github.io/SPATA2/} .
}

\keyword{internal}
