% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{update_dummy}
\alias{update_dummy}
\title{update}
\usage{
update_dummy()
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
update
}
\keyword{internal}
