% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{updateS4}
\alias{updateS4}
\alias{updateS4,ImageAnnotation-method}
\alias{updateS4,SpatialMethod-method}
\alias{updateS4,spata2-method}
\title{Update S4 objects}
\usage{
updateS4(object, ...)

\S4method{updateS4}{ImageAnnotation}(object)

\S4method{updateS4}{SpatialMethod}(object, method_name)

\S4method{updateS4}{spata2}(
  object,
  method = NULL,
  sample_name = NULL,
  n_pcs = 30,
  verbose = TRUE,
  ....
)
}
\arguments{
\item{object}{The S4 object.}

\item{...}{}

\item{method_name}{Character value. Name of the used spatial method.}
}
\value{
An updated S4 object.
}
\description{
Methods for all S4 classes within \code{SPATA2} that keep S4 objects
up to date.
}
\keyword{internal}
