% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{updateSpataObject}
\alias{updateSpataObject}
\title{Update \code{SPATA2} object}
\usage{
updateSpataObject(
  object,
  method,
  sample_name = NULL,
  chr_to_fct = TRUE,
  n_pcs = 30,
  verbose = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method}{Character value. The name of the spatial method (platform) from
which the data derives. Should be one of \code{base::names(spatial_methods)}.}

\item{chr_to_fct}{Logical. SPATA2 recommends to store grouping variables as factors
in the slot @fdata. If set to TRUE, character variables (apart from \emph{barcodes, sample, segmentation})
of the old obejct's feature data are converted to factors.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be computed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Updates the \code{\link{SPATA2}} object to the newest version of the package.
}
