% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{useVarForTissueOutline}
\alias{useVarForTissueOutline}
\title{Use specified variable for tissue outline}
\usage{
useVarForTissueOutline(
  object,
  var_name,
  concavity = 2,
  min_obs = nObs(object) * 0.05
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{var_name}{A character string specifying the name of the variable in the metadata to be used for
the tissue outline.}

\item{concavity}{Numeric value. Given to argument \code{concavity} of
\code{\link[concaveman:concaveman]{concaveman::concaveman()}}. Determines the relative measure of concavity.
1 results in a relatively detailed shape, Infinity results in a convex hull.
You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{min_obs}{Numeric value. The minimal number of observations a group must have
to be considered a tissue section. Defaults to 5\% of the total number of observations. Must be higher than 3.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function sets a specified variable from the metadata of the given object to
be used if \code{\link[=identifyTissueOutline]{identifyTissueOutline()}} does not produce acceptable results.
}
\seealso{
\code{\link[=createSpatialSegmentation]{createSpatialSegmentation()}} to create the outline manually, then use the created
spatial segmentation variable as input for \code{var_name}.
}
