% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-validation.R
\name{validateSpataObject}
\alias{validateSpataObject}
\title{Validate a spata object}
\usage{
validateSpataObject(object)
}
\arguments{
\item{object}{A spata-object.}
}
\value{
A character string that is printed by \code{base::writeLines()}
}
\description{
Takes a spata object and checks whether all slots contain suitable
data. If not it attempts to provide a helpful report.
}
