% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{variable}
\alias{variable}
\title{variable}
\usage{
variable(variable)
}
\arguments{
\item{variable}{Character value. The name \link[=concept_variables]{variable} of interest.}
}
\description{
variable
}
\keyword{internal}
