% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{variable_num}
\alias{variable_num}
\title{variable_num}
\usage{
variable_num(variable)
}
\arguments{
\item{variable}{Character value. The name of the numeric \link[=concept_variables]{variable} of interest.}
}
\description{
variable_num
}
\keyword{internal}
