% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{variables_num}
\alias{variables_num}
\title{variables_num}
\usage{
variables_num(variables)
}
\arguments{
\item{variables}{Character vector. The name of the numeric \link[=concept_variables]{variables} of interest.}
}
\description{
variables_num
}
\keyword{internal}
