% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\docType{class}
\name{Visium-class}
\alias{Visium-class}
\alias{Visium}
\title{The Visium - Class}
\description{
S4 class that represents spatial information from 10X Genomics
Visium experiments.
}
\section{Slots}{

\describe{
\item{\code{annotations}}{list. List of objects of class \code{ImageAnnotation}.}

\item{\code{grid}}{data.frame. A data.frame that contains at least a the numeric variables
named \emph{x} and \emph{y} as well as the character variable \emph{barcodes}.}

\item{\code{dir_default}}{character. Directory to the default version of the image.}

\item{\code{dir_highres}}{character. Directory to the high resolution version of the image.}

\item{\code{grid}}{data.frame. A data.frame that contains at least a variable
named \emph{x} and a variable named \emph{y} representing a grid. Must contain
a character variable named \emph{barcodes}, too.}

\item{\code{image}}{Image.}
}}

