% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visium_coords}
\alias{visium_coords}
\title{A data.frame that contains all barcode spots of the Visium method including
their spatial positioning. Variable names: \emph{barcodes}, \emph{col}, \emph{row}, \emph{imagecol},
\emph{imagerow}, \emph{xlr}, \emph{ylr}, \emph{xhr}, \emph{yhr}. lr = low resolution, hr = high resolution}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4992 rows and 9 columns.
}
\usage{
visium_coords
}
\description{
@docType data
@usage data(visium_coords)
}
\keyword{datasets}
