% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visium_spots}
\alias{visium_spots}
\title{A list of a slot named \emph{VisiumSmall}, which contains a list of data.frames of all
barcoded spots of 6.5mm capture area and a slot named \emph{VisiumLarge} which a list of
data.frames that contain the barcoded spots of the 11mm capture area.}
\format{
An object of class \code{list} of length 2.
}
\usage{
visium_spots
}
\description{
@docType data
@usage data(visium_spots)
}
\keyword{datasets}
