% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{VisiumLarge}
\alias{VisiumLarge}
\title{VisiumLarge Spatial Method}
\format{
An object of class \code{SpatialMethod} of length 1.
}
\usage{
VisiumLarge
}
\description{
This object abstracts the VisiumLarge spatial method, which covers a larger capture area compared to other Visium methods.
}
\details{
The \code{VisiumLarge} object is based on the \code{SpatialMethod} class and contains the following specifications:
\itemize{
\item \code{capture_area}: A list specifying the coordinates of the capture area:
\itemize{
\item \code{x}: A vector with coordinates "0.75mm" and "11.75mm".
\item \code{y}: A vector with coordinates "0.75mm" and "11.75mm".
}
\item \code{method_specifics}: A list with the following elements:
\itemize{
\item \code{ccd}: Center to center distance of "100um".
\item \code{diameter}: Diameter of each spot is "55um".
\item \code{fiducial_frame}: A list with the following elements:
\itemize{
\item \code{x}: A vector with coordinates "0mm" and "12.5mm".
\item \code{y}: A vector with coordinates "0mm" and "12.5mm".
}
}
\item \code{observational_unit}: The unit of observation, which is "spot".
\item \code{unit}: The SI unit used, which is "mm".
\item \code{version}: The version of the SPATA2 package being used.
\item \code{name}: The name of the spatial method, which is "VisiumLarge".
}
}
\keyword{datasets}
