% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{VisiumSmall}
\alias{VisiumSmall}
\title{VisiumSmall Spatial Method}
\format{
An object of class \code{SpatialMethod} of length 1.
}
\usage{
VisiumSmall
}
\description{
This object abstracts the VisiumSmall spatial method, which is designed for smaller capture areas compared to other Visium methods.
}
\details{
The \code{VisiumSmall} object is based on the \code{SpatialMethod} class and contains the following specifications:
\itemize{
\item \code{method_specifics}: A list with the following elements:
\itemize{
\item \code{capture_area_frame}: A list specifying the coordinates of the capture area:
\itemize{
\item \code{x}: A vector with coordinates "0.75mm" and "7.25mm".
\item \code{y}: A vector with coordinates "0.75mm" and "7.25mm".
}
\item \code{ccd}: Center to center distance of "100um".
\item \code{diameter}: Diameter of each spot is "55um".
\item \code{fiducial_frame}: A list with the following elements:
\itemize{
\item \code{x}: A vector with coordinates "0mm" and "8mm".
\item \code{y}: A vector with coordinates "0mm" and "8mm".
}
}
\item \code{observational_unit}: The unit of observation, which is "spot".
\item \code{unit}: The SI unit used, which is "mm".
\item \code{version}: The version of the SPATA2 package being used.
\item \code{name}: The name of the spatial method, which is "VisiumSmall".
}
}
\keyword{datasets}
