% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{visiumSpotDistances}
\alias{visiumSpotDistances}
\title{Calculate Distances Between Visium Spots}
\usage{
visiumSpotDistances(
  type = c("small", "large"),
  bcs_o = NULL,
  bcs_n = NULL,
  nnn = NULL
)
}
\arguments{
\item{type}{A character vector specifying the type of Visium platform. One of "small" or "large". Default is "small".}

\item{bcs_o}{A character vector of barcodes specifying the origin spots. If NULL (default), all barcodes from the specified type are used.}

\item{bcs_n}{A character vector of barcodes specifying the neighbor spots. If NULL (default), all barcodes from the specified type are used.}

\item{nnn}{A numeric value specifying the number of nearest neighbors to consider. If NULL (default), all neighbors are considered.}
}
\value{
A data frame containing the pairwise distances between the specified Visium spots. The data frame contains the following variables:
\itemize{
\item {\strong{bcs_o}}: Barcode of the origin spot.
\item {\strong{bcs_n}}: Barcode of the neighbor spot.
\item {\strong{xo}}: x-coordinate of the origin spot.
\item {\strong{yo}}: y-coordinate of the origin spot.
\item {\strong{xn}}: x-coordinate of the neighbor spot.
\item {\strong{yn}}: y-coordinate of the neighbor spot.
\item {\strong{distance}}: Calculated distance between the origin and neighbor spots.
}
}
\description{
This function calculates the pairwise distances between specified Visium spots
based on their x and y coordinates.
}
