% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/helper-functions.R
=======
% Please edit documentation in R/v.R
>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
\name{vselect}
\alias{vselect}
\title{Select vector with tidyselect functions}
\usage{
<<<<<<< HEAD
vselect(input, ...)
=======
vselect(input, ..., out.fail = base::character(0))
>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
}
\arguments{
\item{input}{A character vector or a factor.}

\item{...}{Additional selection helpers from the \code{tidyselect} package that match
variable names according to a given pattern.}

\item{lst}{A named list. (Unnamed elements are discarded.)}
}
\value{
A subsetted version of the input.
}
\description{
A wrapper around the tidyselect functions that allows to use them
not only on data.frames but on vectors as well.
}
\seealso{
\code{starts_with()}, \code{ends_with()}, \code{contains()}, \code{matches()}
}
<<<<<<< HEAD
=======
\keyword{internal}
>>>>>>> f7196fe9dc51c1738b724567a2e437b9d4780e28
