% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-objects.R
\docType{data}
\name{Xenium}
\alias{Xenium}
\title{Xenium Spatial Method}
\format{
An object of class \code{SpatialMethod} of length 1.
}
\usage{
Xenium
}
\description{
This object abstracts the Xenium spatial method, which focuses on cell-level observations.
}
\details{
The \code{Xenium} object is based on the \code{SpatialMethod} class and contains the following specifications:
\itemize{
\item \code{name}: The name of the spatial method, which is "Xenium".
\item \code{observational_unit}: The unit of observation, which is "cell".
\item \code{unit}: The SI unit used, which is "mm".
\item \code{version}: The version of the SPATA2 package being used.
}
}
\keyword{datasets}
