% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activateAssay}
\alias{activateAssay}
\alias{activeAssay}
\title{Default assay name}
\usage{
activateAssay(object, assay_name, verbose = NULL)

activeAssay(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Character value. The name of the assay to activate as
the default assay}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\itemize{
\item{activateAssay()}: Updated \code{SPATA2} object.
\item{activateAssay()}: Character value. Name of the currently active assay.
}
}
\description{
Sets and extracts the active (default) assay name.
}
\examples{
data("example_data")

object <- example_data$object_UKF275T_diet

activeAssay(object)

}
\seealso{
\code{\link[=getAssay]{getAssay()}}, \code{\link[=getAssayNames]{getAssayNames()}}
}
