% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activateImage}
\alias{activateImage}
\alias{activateImage,spata2-method}
\alias{activateImage,HistoImaging-method}
\alias{activateImageInt}
\alias{activateImageInt,spata2-method}
\alias{activateImageInt,HistoImaging-method}
\title{Activate \code{HistoImage}}
\usage{
activateImage(object, ...)

\S4method{activateImage}{spata2}(
  object,
  img_name,
  load = TRUE,
  unload = TRUE,
  verbose = TRUE,
  ...
)

\S4method{activateImage}{HistoImaging}(
  object,
  img_name,
  load = TRUE,
  unload = TRUE,
  verbose = TRUE,
  ...
)

activateImageInt(object, ...)

\S4method{activateImageInt}{spata2}(object, img_name, load = FALSE)

\S4method{activateImageInt}{HistoImaging}(object, img_name, load = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{unload}{Logical value. If \code{TRUE}, ensures that @image slots of
the inactive registered images are empty to prevent the input object
from becoming too big.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets the active image of the input object which is
then used by default in image dependent functions if argument \code{img_name = NULL}.
}
\note{
\code{activateImageInt()} exists mainly for internal use. It works
the same way \code{activateImage()} works but never unloads and is always
silent.
}
