% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activeImage}
\alias{activeImage}
\alias{activeImage,SPATA2-method}
\alias{activeImage,SpatialData-method}
\alias{activateImage}
\alias{activateImage,SPATA2-method}
\alias{activateImage,SpatialData-method}
\alias{activateImageInt}
\alias{activateImageInt,SPATA2-method}
\alias{activateImageInt,SpatialData-method}
\title{Default image name}
\usage{
activeImage(object, ...)

\S4method{activeImage}{SPATA2}(object)

\S4method{activeImage}{SpatialData}(object)

activateImage(object, ...)

\S4method{activateImage}{SPATA2}(
  object,
  img_name,
  load = TRUE,
  unload = TRUE,
  verbose = TRUE,
  ...
)

\S4method{activateImage}{SpatialData}(
  object,
  img_name,
  load = TRUE,
  unload = TRUE,
  verbose = TRUE,
  ...
)

activateImageInt(object, ...)

\S4method{activateImageInt}{SPATA2}(object, img_name, load = FALSE)

\S4method{activateImageInt}{SpatialData}(object, img_name, load = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image to activate as
the default image.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\itemize{
\item{activateImage()}: Updated \code{SPATA2} object.
\item{activeImage()}: Character value. Name of the currently active image.
}
}
\description{
Sets and extracts the active (default) image name.
}
\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

getImageNames(object)

activeImage(object)
plotImage(object)

getCoordsDf(object) # with image1

object <- activateImage(object, img_name = "very_low_res")
activeImage(object)
plotImage(object)

# note how x_orig, y_orig remain the same, but x and y differ
getCoordsDf(object) # with very_low_res

}
\seealso{
\code{\link[=getImage]{getImage()}}, \code{\link[=getHistoImage]{getHistoImage()}}, \code{\link[=getImageNames]{getImageNames()}}
}
