% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{activeImage}
\alias{activeImage}
\alias{activeImage,spata2-method}
\alias{activeImage,HistoImaging-method}
\title{Obtain name of active content}
\usage{
activeImage(object, ...)

\S4method{activeImage}{spata2}(object)

\S4method{activeImage}{HistoImaging}(object)
}
\arguments{
\item{object}{An object that contains activated aspects such as
assays, images and matrices.}
}
\value{
Character value.
}
\description{
Handy functions to quickly access the name of currently
activated content.
}
