% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{activeImage}
\alias{activeImage}
\title{Obtain name of active content}
\usage{
activeImage(object)
}
\arguments{
\item{object}{An object that contains activated aspects such as
assays, images and matrices.}
}
\value{
Character value.
}
\description{
Handy functions to quickly access the name of currently
activated content.
}
