% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_dbscan_variable}
\alias{add_dbscan_variable}
\alias{add_tissue_section_variable}
\title{Add tissue section variable}
\usage{
add_dbscan_variable(
  coords_df,
  eps,
  minPts = 3,
  name = "dbscan",
  x = "x",
  y = "y",
  ...
)

add_tissue_section_variable(coords_df, ccd, minPts = 3, ...)
}
\arguments{
\item{coords_df}{Data.frame with \emph{x} and \emph{y} variable.}

\item{eps}{size (radius) of the epsilon neighborhood. Can be omitted if
\code{x} is a frNN object.}

\item{minPts}{number of minimum points required in the eps neighborhood for
core points (including the point itself).}

\item{name}{Name of the added variable.}

\item{x}{a data matrix, a data.frame, a \link{dist} object or a \link[dbscan]{frNN} object with
fixed-radius nearest neighbors.}

\item{...}{To silently drop deprecated arguments.}

\item{ccd}{Center to center distance in pixel units.}
}
\value{
Data.frame with additional variable containing numbers. 0 means
that the spot is not connected to any other spot (probably artefact). 1-n
corresponds to the tissue sections.
}
\description{
Leverages \code{dbscan::dbscan()} to identify tissue sections
on the slide and to group barcode spots accordingly. Required to approximate
the outline of the tissue section(s).
}
\note{
\code{add_dbscan_variable()} is the working horse. \code{add_tissue_section_variable()}
has specific defaults.
}
\examples{

# --- identify tissue sections
object <- downloadPubExample("MCI_LMU", verbose = FALSE)

coords_df <- getCoordsDf(object)

coords_df <- add_tissue_section_variable(coords_df, ccd = getCCD(object, "px"))

plotSurface(coords_df, color_by = "section")

# --- identify artefact spots
object <- SPATAData::downloadSpataObject("269_T", verbose = FALSE)

coords_df <- getCoordsDf(object)

coords_df <- add_tissue_section_variable(coords_df, ccd = getCCD(object, "px"))

plotSurface(coords_df, color_by = "section")

}
