% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_edge_variable}
\alias{add_edge_variable}
\title{Add edge variable}
\usage{
add_edge_variable(coords_df, id_var = "barcodes")
}
\arguments{
\item{id_var}{Character. Variable that identifies each observation.}

\item{input_df}{A data.frame with two numeric variables called \emph{x} and \emph{y}
and a variable as denoted in \code{id_var}.}
}
\value{
Input data.frame with additional logical variable \emph{edge}.
}
\description{
Adds a variable called \emph{edge} to the input data.frame
that tells if the observation belongs to the points that lie on the
edge of the covered area.
}
\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

coords_df <-
 getCoordsDf(object) \%>\%
 add_edge_variable()

plotSurface(coords_df, color_by = "edge")

}
