% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_models}
\alias{add_models}
\title{Add models to a data.frame}
\usage{
add_models(
  input_df,
  var_order,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  verbose = TRUE
)
}
\arguments{
\item{input_df}{Data.frame with at least three columns. \emph{values}
contains the actual values. \emph{variables} contains the variable belonging
of the values. \emph{\code{var_order}} contains the integers from 1 to n
corresponding to the ordering of the values.}

\item{var_order}{Character value. The variable that corresponds to the order
of the values.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}
}
\description{
Add models to a data.frame
}
\keyword{internal}
