% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAutoencoderSetUp}
\alias{addAutoencoderSetUp}
\title{Add the set up of a neural network}
\usage{
addAutoencoderSetUp(object, mtr_name, set_up_list, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}

\item{set_up_list}{A named list with slots \code{$activation, $bottleneck, $dropout, $epochs, $layers}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A \code{SPATA2} object.
}
\description{
Add the set up of a neural network
}
\keyword{internal}
