% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addAutoencoderSetUp}
\alias{addAutoencoderSetUp}
\title{Add the set up of a neural network}
\usage{
addAutoencoderSetUp(object, mtr_name, set_up_list, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{set_up_list}{A named list with slots \code{$activation, $bottleneck, $dropout, $epochs, $layers}.}
}
\value{
A \code{SPATA2} object.
}
\description{
Add the set up of a neural network
}
\keyword{internal}
