% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addAutoencoderSetUp}
\alias{addAutoencoderSetUp}
\title{Title}
\usage{
addAutoencoderSetUp(object, mtr_name, set_up_list, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{set_up_list}{A named list with slots \code{$activation, $bottleneck, $dropout, $epochs, $layers}.}
}
\value{
A spata-object.
}
\description{
Title
}
