% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addExpressionMatrix}
\alias{addExpressionMatrix}
\title{Add an expression matrix}
\usage{
addExpressionMatrix(object, expr_mtr, mtr_name, overwrite = FALSE, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{expr_mtr}{A matrix in which the rownames correspond to the gene names and the
column names correspond to the barcode-spots.}

\item{mtr_name}{A character value that denotes the name of the exprssion matrix with
which one can refer to it in subsequent functions.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Adds an expression matrix to the object's data slot and
makes it available for all SPATA-intern function. Use \code{setActiveExpressionMatrix()}
to denote it as the default to use.
}
