% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{addExpressionMatrix}
\alias{addExpressionMatrix}
\title{Deprecated}
\usage{
addExpressionMatrix(...)
}
\description{
Deprecated in favor of \code{\link[=addProcessedMatrix]{addProcessedMatrix()}}.
}
\keyword{internal}
