% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{addExpressionMatrix}
\alias{addExpressionMatrix}
\title{Add an expression matrix}
\usage{
addExpressionMatrix(object, expr_mtr, mtr_name, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{expr_mtr}{A matrix in which the rownames correspond to the gene names and the
column names correspond to the barcode-spots.}

\item{mtr_name}{A character value that denotes the name of the exprssion matrix with
which one can refer to it in subsequent functions.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
An updated spata-object.
}
\description{
Adds an expression matrix to the object's data slot and
makes it available for all SPATA-intern function. Use \code{setActiveExpressionMatrix()}
to denote it as the default to use.
}
