% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discuss.R
\name{addGeneSet}
\alias{addGeneSet}
\alias{addGeneSetsInteractive}
\title{Add a new gene set}
\usage{
addGeneSet(
  object,
  class_name,
  gs_name,
  genes,
  overwrite = FALSE,
  check_genes = TRUE
)

addGeneSetsInteractive(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{class_name}{Character value. The class the gene set belongs to..}

\item{gs_name}{Character value. The name of the new gene set.}

\item{genes}{The genes of interest specified as a character vector.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Stores a new gene set in the \code{SPATA2} object.
}
\details{
Combines \code{class_name} and \code{gs_name} to the final gene set name.
Gene set classes and gene set names are separated by '\emph{' and handled like this
in all additional gene set related functions which is why \code{class_name} must
not contain any '}'.
}
\examples{

# ----- prepare
library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet
}
