% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addImageAnnotation}
\alias{addImageAnnotation}
\alias{addImageAnnotation,HistoImaging-method}
\title{Add an image annotation manually}
\usage{
addImageAnnotation(object, ...)

\S4method{addImageAnnotation}{HistoImaging}(object, tags, area, parent_name, id = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{tags}{A character vector of tags for the image annotation.}

\item{area}{A named list of data.frames with the numeric variables \emph{x} and \emph{y}.
Observations correspond to the vertices of the polygons that are needed to represent the
image annotation. \strong{Must} contain a slot named \emph{outer} which sets the outer border
of the image annotation. \strong{Can} contain multiple slots named \emph{inner} (suffixed)
with numbers that correspond to inner polygons - holes within the annotation.}

\item{parent_name}{Character value. The name of the image on
which the annotation was drawn.}

\item{id}{Character value. The ID of the image annotation.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Adds image annotations manually by assembling the object
based on the function input.
}
