% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{addImageAnnotation}
\alias{addImageAnnotation}
\title{Add image annotation}
\usage{
addImageAnnotation(object, tags, area_df, id = NULL)
}
\arguments{
\item{area_df}{A data.frame that contains at least two numeric variables named
\emph{x} and \emph{y}.}
}
\value{
An updated spata object.
}
\description{
Creates and adds an object of class \code{ImageAnnotation}.
}
