% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addPointsBase}
\alias{addPointsBase}
\title{Add points to base surface plot}
\usage{
addPointsBase(
  object,
  color_by,
  alpha_by = NULL,
  pt_alpha = 0.75,
  pt_size = 1,
  pt_clrp = "default",
  pt_clrsp = "inferno",
  clrp_adjust = NULL,
  smooth = NULL,
  smooth_span = NULL,
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{color_by}{Character value. The \link[=concept_variables]{variables} by which
to color the data points.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}
}
\value{
Plots right into the plotting window.
}
\description{
Adds a point layer to a base surface plot.
}
\keyword{internal}
