% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R, R/adj-aZ.R
\name{addPointsBase}
\alias{addPointsBase}
\title{Add points to base surface plot}
\usage{
addPointsBase(
  object,
  color_by,
  alpha_by = NULL,
  pt_alpha = 0.75,
  pt_size = 1,
  pt_clrp = "default",
  pt_clrsp = "inferno",
  clrp_adjust = NULL,
  smooth = NULL,
  smooth_span = NULL,
  xrange = NULL,
  yrange = NULL
)

addPointsBase(
  object,
  color_by,
  alpha_by = NULL,
  pt_alpha = 0.75,
  pt_size = 1,
  pt_clrp = "default",
  pt_clrsp = "inferno",
  clrp_adjust = NULL,
  smooth = NULL,
  smooth_span = NULL,
  xrange = NULL,
  yrange = NULL
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{scale_alpha}{}
}
\description{
Adds a point layer to a base surface plot.

Adds a point layer to a base surface plot.
}
