% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addExpressionMatrix}
\alias{addExpressionMatrix}
\alias{addProcessedMatrix}
\title{Add a processed matrix}
\usage{
addExpressionMatrix(...)

addProcessedMatrix(
  object,
  proc_mtr,
  mtr_name,
  assay_name = activeAssay(object),
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{...}{Used to absorb deprecated arguments or functions.}

\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{A character value that denotes the name of the matrix with
which one can refer to it in subsequent functions via \code{mtr_name}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{expr_mtr}{A matrix in which the rownames correspond to the feature names and the
column names correspond to the barcodes.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Adds a processed matrix to the chosen molecular assay of the object.
}
