% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addSegmentationVariable}
\alias{addSegmentationVariable}
\title{Add Segmentation Variable}
\usage{
addSegmentationVariable(object, name, verbose = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{name}{A character string specifying the name of the segmentation variable to add.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
This function adds a segmentation variable to the metadata of the given object.
}
\details{
This function checks if the provided segmentation variable name is not already used
by a feature, gene, or gene set in the object. If the name is unique, it adds the segmentation variable
to the object's metadata.
}
\seealso{
\code{\link[=getSegmentationNames]{getSegmentationNames()}}, \code{\link[=getFeatureNames]{getFeatureNames()}}, \code{\link[=getMetaDf]{getMetaDf()}}, \code{\link[=setMetaDf]{setMetaDf()}}
}
