% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addSpatialAnnotation}
\alias{addSpatialAnnotation}
\alias{addSpatialAnnotation,spata2-method}
\alias{addSpatialAnnotation,HistoImaging-method}
\title{Add an spatial annotation manually}
\usage{
addSpatialAnnotation(object, ...)

\S4method{addSpatialAnnotation}{spata2}(
  object,
  tags,
  area,
  parent_name,
  id = NULL,
  overwrite = FALSE,
  class = "SpatialAnnotation",
  ...
)

\S4method{addSpatialAnnotation}{HistoImaging}(
  object,
  tags,
  area,
  id = NULL,
  overwrite = FALSE,
  class = "SpatialAnnotation",
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{tags}{A character vector of tags for the spatial annotation.}

\item{area}{A named list of data.frames with the numeric variables \emph{x_orig} and \emph{y_orig}.
Observations correspond to the vertices of the polygons that are needed to represent the
spatial annotation. \strong{Must} contain a slot named \emph{outer} which sets the outer border
of the spatial annotation. \strong{Can} contain multiple slots named \emph{inner} (suffixed)
with numbers that correspond to inner polygons - holes within the annotation.

The scale of the two variables must correspond to the scale of the \emph{x_orig} and
\emph{y_orig} variables of the coordinates data.frame.}

\item{id}{Character value. The ID of the spatial annotation.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Adds spatial annotations manually by assembling the object
based on the function input.
}
\seealso{
\code{\link[=getCoordsDf]{getCoordsDf()}}
}
