% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addTissueOutlineBase}
\alias{addTissueOutlineBase}
\alias{addTissueOutlineBase,HistoImage-method}
\title{Add polygons to a base plot}
\usage{
addTissueOutlineBase(object, ...)

\S4method{addTissueOutlineBase}{HistoImage}(
  object,
  by_section = FALSE,
  persp = "coords",
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  scale_fct = 1,
  init = list(),
  rect = FALSE,
  ...
)
}
\arguments{
\item{scale_fct}{A factor with which the vertice positions are scaled.}

\item{x, y}{Numeric vectors representing x- and y-coordinates of the
vertices of the polygon.}

\item{poly}{Data.frame of at least two variables named \emph{x} and \emph{y}
representing the coordinates of the vertices of the polygon. If
variable \emph{section} exists multiple polygons are plotted based
on the number of different groups in this variable. Overwrites \code{x}
and \code{y}.}

\item{color}{Color of the lines.}

\item{size}{Width of the lines.}
}
\value{
Output of \code{graphics::polygon()} is directly plotted.
}
\description{
Adds polygons to a base plot.
}
