% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addVarToCoords}
\alias{addVarToCoords}
\alias{addVarToCoords,SpatialData-method}
\title{Add variable to coordinates data.frame}
\usage{
addVarToCoords(object, ...)

\S4method{addVarToCoords}{SpatialData}(object, var_df, vars, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{var_df}{Data.frame with the variables to merge.}

\item{vars}{Character vector. Subset of variables to add.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Adds variables to the coordinates data.frame in slot @coordinates.
}
\keyword{internal}
