% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust-family.R
\name{adjustDefaultInstructions}
\alias{adjustDefaultInstructions}
\alias{adjustDirectoryInstructions}
\title{Adjust default instructions}
\usage{
adjustDefaultInstructions(object, ...)

adjustDirectoryInstructions(object, to, directory_new, combine_with_wd = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{Named arguments whoose default input you want to override.}

\item{to}{Character value. Denotes the platform for which a new storage
directory is to be created. Must be either \emph{'cell_data_set', 'seurat_object'}
or \emph{'spata_object'}.}

\item{directory_new}{Character value. The new directory under which
to store the object of interest. Overwrites the stored default directory.
Use \code{getDefaultDirectory()} to obtain the current set up.}

\item{combine_with_wd}{Character value or FALSE. If specified with a
character value (default: \emph{'/'}) the input of \code{new_directory}
is considered to be a relative directory and is combined with the
current working directory (\code{base::getwd()}) separated with the character string
specified. If set to FALSE the input of \code{new_directory}
is taken as is.}
}
\value{
An updated spata object.
}
\description{
Adjust default instructions
}
\examples{

 # Not run

 object <- adjustDefaultInstructions(object, pt_size = 4, smooth = FALSE)
}
