% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy(
  clrp,
  clrsp,
  display_points,
  display_facets,
  scales,
  ncol,
  nrow,
  verbose
)
}
\arguments{
\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{scales, ncol, nrow}{Affects the way the subplots
are displayed.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{bcsp_rm}{Character vector or NULL. If character, specifies barcode-spots that
are removed before plotting.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{dir}{Character value. The chosen directory. See details for possible
requirements.}

\item{discrete_feature}{Character value. Specifies the name of the grouping variable
of interest. Use \code{getGroupingOptions()} to obtain all valid input options.}

\item{display_ribbon}{Logical value. If TRUE, a ribbon is displayed around
the main line of the plot visualizing uncertainty using standard deviation.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{expand}{Vector of length one or two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the x-span and the second
value is used to adjust the y-span E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original x-span to the x-span of the image section and does not adjust
the y-span.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{h}{Numeric value or vector or NULL (see details for more). Denotes the height at which
the dendrogom is cut.}

\item{hline_alpha, hline_color, hline_size, hline_type}{Parameters given to
\code{ggplot2::geom_hline()} that control the appearance of vertical lines
of the plot.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{k}{Numeric value or vector or NULL (see details for more). Denotes the number of clusters
in which the hierarchical tree is supposed to be split.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{method_padj}{Character value. The method with which adjusted p-values are
calculated. Use \code{validPadjMethods()} to obtain all valid input options.}

\item{n_bcsp}{Numeric value. Specifies the sample size of barcode-spots and
can be set to prevent overplotting.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{n_gsets}{Numeric value. Maximal number of gene sets whose results are included.
The first \code{n_gsets} are included starting with the one with the lowest significance value.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{line_alpha, line_color, line_size, line_type}{Parameters given to
\code{ggplot2::geom_line()}, \code{ggplot2::geom_path()} or \code{ggplot2::geom_smooth()}
that control the appearance of the main line(s) of the plot.}

\item{linesize}{Numeric value. The size of the line(s) plotted.}

\item{object}{An object of class \code{SPATA2}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the colors with which they are displayed.}

\item{signif_var}{Character value. Determines what to be considered while checking
for significance. Either \emph{'pval'} (p-Value) or \emph{'fdr'} (False Discovery Rate).}

\item{signif_threshold}{Numeric value. Significance values below \code{signif_threshold}
are not included.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. X- and yspan of the square are equal to the span
that is the biggest. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{test}{Character value. Specifies how the input of \code{tags} is used to
subset the image annotations. One of \emph{'any'}, \emph{'all'} or \emph{'identical'}.
See details for more information.}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{vline_alpha, vline_color, vline_size, vline_type}{Parameters given to
\code{ggplot2::geom_vline()} that control the appearance of vertical lines
of the plot.}

\item{whole_sample}{Logical. If TRUE, normalization of the values used
takes place in the light of the complete sample.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}
}
\description{
Argument dummy
}
