% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy()
}
\arguments{
\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{angle_span}{Numeric vector of length two. Defines the range of angle values
an observation might have when related to the center of the spatial annotation,
to be included.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{bcsp_rm}{Character vector or \code{NULL.} If character, specifies barcode-spots that
are removed before analysis or plotting. (Deprecated in favor of \code{bcs_rm}).}

\item{bcs_rm}{Character vector or \code{NULL}. If character, specifies the observations
to be removed prior to analysis or visualization by their barcode.}

\item{by_section}{Logical value. If \code{TRUE}, the outline is used which acknowledges
the potential for multiple tissue sections. If \code{FALSE}, the outline is used which
simply outlines everything with one single polygon.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{color_by}{Character value. The \link[=concept_variable]{variables} by which
to color the data points.}

\item{concavity}{Numeric value. Given to argument \code{concavity} of
\code{\link[concaveman:concaveman]{concaveman::concaveman()}}. Determines the relative measure of concavity.
1 results in a relatively detailed shape, Infinity results in a convex hull.
You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{dir}{Character value. The chosen directory. See details for possible
requirements.}

\item{discrete_feature}{Character value. Specifies the name of the grouping variable
of interest. Use \code{getGroupingOptions()} to obtain all valid input options.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display_ribbon}{Logical value. If TRUE, a ribbon is displayed around
the main line of the plot visualizing uncertainty using standard deviation.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{eps}{Distance measure. Given to \code{eps} of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines
the size (radius) of the epsilon neighborhood.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}

\item{expand_x, expand_y}{Given to argument \code{expand} of \code{ggplot2:scale_x/y_continuous()}.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}

\item{expand_outline}{Distance measure by which the outline of the area is expanded.}

\item{format}{Character value. Either \emph{'long'} or \emph{'wide'}. Defaults to
\emph{'wide'}. If \emph{'wide'} each variable gets a column.  If \emph{'long'}, the data.frame
is organized such that a column called \emph{variables} contains the variable names
and a column called \emph{'values'} contains the values.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{grouping}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{grouping_variable}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{h}{Numeric value or vector or NULL (see details for more). Denotes the height at which
the dendrogom is cut.}

\item{hline_alpha, hline_color, hline_size, hline_type}{Parameters given to
\code{ggplot2::geom_hline()} that control the appearance of vertical lines
of the plot.}

\item{img_alpha}{Numeric value. Sets the transparency for the image.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{img_names}{Character vector. The names of the images of interest.}

\item{incl_edge}{Logical value. If \code{TRUE}, the function includes the tissue edge
(see \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}) in the visualization and removes the parts of
the outline that transgress the tissues edge.}

\item{incr_vert}{Logical value. If \code{TRUE}, the number of vertices of the polygon
to be plotted is increased by linear interpolation. This can lead to smoother
visualization but also to increased runtime.}

\item{inner}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the inner borders of the image annotation
are not included in the output.}

\item{k}{Numeric value or vector or NULL (see details for more). Denotes the number of clusters
in which the hierarchical tree is supposed to be split.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{method_padj}{Character value. The method with which adjusted p-values are
calculated. Use \code{validPadjMethods()} to obtain all valid input options.}

\item{minPts}{Numeric value. Given to \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines the
number of minimum points required in the eps neighborhood for core points
(including the point itself)}

\item{modality}{Character value. The \link[=concept_molecular_modalities]{molecular modality}
of interst.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}

\item{n_bins_angle}{Numeric value. The number of bins in which observations
are categorized in the variable \emph{bins_angle}.}

\item{n_bcsp}{Numeric value. Specifies the sample size of barcode-spots and
can be set to prevent overplotting.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{n_gsets}{Numeric value. Maximal number of gene sets whose results are included.
The first \code{n_gsets} are included starting with the one with the lowest significance value.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{n_pcs}{Numeric value or \code{NULL}. If numeric, specifies the number of
principal components to use. If \code{NULL}, all principal components are used
(as computed with \code{\link[=runPCA]{runPCA()}}).}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{linesize}{Numeric value. The size of the line(s) plotted.}

\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{order}{Logical value. If \code{TRUE}, data points are ordered according
to their values before beeing plotted.}

\item{order_by}{Character value or \code{NULL}. If character, the specified
variable is used to order the data points.}

\item{order_desc}{Logical value. If \code{TRUE}, reverses the arrangement specified
via \code{order_by} and/or \code{order}.}

\item{outer}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the outer border of the image annotation
is not included in the output.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{rect_alpha}{The alpha transparency level of the rectangle's border}

\item{rect_color}{The color of the rectangle's border.}

\item{rect_linesize}{The size of the line used for the rectangle's border.}

\item{rect_linetype}{The type of line used for the rectangle's border.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the colors with which they are displayed.}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{sc_input}{Data.frame that contains the results from single cell deconvolution.
Must have at least three columns:
\itemize{
\item{\emph{x} : }{ numeric. Position of cell on the x axis in pixel.},
\item{\emph{y} :}{ numeric. Position of cell on the y axis in pixel.},
\item{\emph{cell_type} :}{ factor Cell type of the cell.}
}}

\item{scales, ncol, nrow}{Affects the way the subplots
are displayed.}

\item{sctm_interpolate, sctm_pixels}{Given to the corresponding arguments
of \code{scattermore::geom_scattermore()}. Note: With increasing \code{sctm_pixels}
the point size must be adjusted with the argument \code{pt_size}.}

\item{sgmt_alpha, sgmt_color, sgmt_size, sgmt_type}{Parameters given to
\code{ggplot2::geom_segment()} that control the appearance of segments
of the plot.}

\item{signif_var}{Character value. Determines what to be considered while checking
for significance. Either \emph{'pval'} (p-Value) or \emph{'fdr'} (False Discovery Rate).}

\item{signif_threshold}{Numeric value. Significance values below \code{signif_threshold}
are not included.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{square}{Logical value. Most image annotations come in variable shapes and
have different horizontal and vertical diameters. Therefore, height and width of the image
section are usually not equal. If \code{square = TRUE}, the cropped section of the image that
contains the annotated structure is forced into a square: the bigger diameter of both is taken
as default. E.g. if the horizontal diameter of the image annotation is 1mm and the
vertical diameter is 1.5mm, the output image will have height and width of 1.5mm. That is,
in terms of coordinates, an x-range and a y-range of 1.5mm.

Processing of the image output depending on argument \code{expand} happens afterwards.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{spatial_method}{Character value. The name of the spatial method that underlies
the experiment. Must be one of \code{validSpatialMethods()}. Defaults to \emph{'Unknown'}.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of spatial annotations} for more information.}

\item{test}{Character value. One of \emph{c('any'. 'all', 'identical', 'not_identical', 'none')}.
Specifies how input for \code{tags} is used to select spatial annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{text_alpha, text_color, text_nudge_x, text_nudge_y, text_size, text_type}{Parameters
given to \code{ggplot2::geom_text()} that control the appearance of text of the plot.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there are any images registered
in the \code{SPATA2} object.}

\item{transform_with}{List or \code{NULL.} If list, can be used to transform continuous variables before usage.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
E.g if the variable of interest is \emph{GFAP} gene expression, the following would work:

\itemize{
\item{Single function:}{ \code{transform_with = log10}},
\item{Multiple functions:}{ \verb{transform_with = list(GFAP = list(log10, log2)}}
}

In case of plotting:
Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}

\item{use_scattermore}{Logical value. If \code{TRUE}, data points are plotted with
\code{scattermore::geom_scattermore()} which allows quick plotting of several
thousand data points. If the number of data points plotted is bigger than
10.000 it is used anyway.}

\item{variables}{Character vector. The names of the \emph{\link[=concept_variables]{data variables}}
of interest.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{vline_alpha, vline_color, vline_size, vline_type}{Parameters given to
\code{ggplot2::geom_vline()} that control the appearance of vertical lines
of the plot.}

\item{whole_sample}{Logical. If TRUE, normalization of the values used
takes place in the light of the complete sample.}

\item{x_nth}{Numeric value. If the number of breaks/labels on the
x-axis becomes too high \code{x_nth} can be used to reduce it. If \code{x_nth} is 1,
every label is kept. If 2, every second label is kept. If 3, every
third label is kept. And so on.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
Argument dummy
}
\keyword{internal}
