% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy(
  clrp,
  clrsp,
  display_points,
  display_facets,
  scales,
  ncol,
  nrow,
  verbose
)
}
\arguments{
\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{display_facets}{Logical value. If set to TRUE the plot is split via
\code{ggplot2::facet_wrap()} such that each variable gets it's own subplot.}

\item{scales, ncol, nrow}{Affects the way the subplots
are displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{discrete_feature}{Character value. Specifies the name of the grouping variable
of interest. Use \code{getGroupingOptions()} to obtain all valid input options.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{expand}{Numeric value or vector of length two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the xspan and the second
value is used to adjust the yspan. E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original xspan to the xspan of the image section and does not adjust
 the yspan.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{h}{Numeric value or vector or NULL (see details for more). Denotes the height at which
the dendrogom is cut.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{k}{Numeric value or vector or NULL (see details for more). Denotes the number of clusters
in which the hierarchical tree is supposed to be split.}

\item{method_padj}{Character value. The method with which adjusted p-values are
calculated. Use \code{validPadjMehods()} to obtain all valid input options.}

\item{n_bcsp}{Numeric value. Specifies the sample size of barcode-spots and
can be set to prevent overplotting.}

\item{n_gsets}{Numeric value. Maximal number of gene sets whose results are included.
The first \code{n_gsets} are included starting with the one with the lowest significance value.}

\item{object}{A valid spata-object.}

\item{signif_var}{Character value. Determines what to be considered while checking
for significance. Either \emph{'pval'} (p-Value) or \emph{'fdr'} (False Discovery Rate).}

\item{signif_threshold}{Numeric value. Significance values below \code{signif_threshold}
are not included.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. X- and yspan of the square are equal to the span
that is the biggest. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{whole_sample}{Logical. If TRUE, normalization of the values used
takes place in the light of the complete sample.}

\item{xrange, yrange}{Numeric vectors of length two or NULL. If numeric, specifies the coordinates
with which the image is subsetted. E.g. xrange = c(200, 500) results in the image being cropped
from x-coordinate 200 up to x-coordinate 500.}
}
\description{
Argument dummy
}
