% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{as_meter}
\alias{as_meter}
\alias{as_meter2}
\alias{as_micrometer}
\alias{as_micrometer2}
\alias{as_millimeter}
\alias{as_millimeter2}
\alias{as_nanometer}
\alias{as_nanometer2}
\alias{as_pixel}
\alias{as_unit}
\alias{as_centimeter}
\alias{as_centimeter2}
\alias{as_decimeter}
\alias{as_decimeter2}
\title{Transform distance and area values}
\usage{
as_meter(input, ...)

as_meter2(input, ...)

as_micrometer(input, ...)

as_micrometer2(input, ...)

as_millimeter(input, ...)

as_millimeter2(input, ...)

as_nanometer(input, ...)

as_nanometer2(input, ...)

as_pixel(input, object = NULL, ..., add_attr = TRUE)

as_unit(input, unit, object = NULL, round = FALSE, verbose = FALSE, ...)

as_centimeter(input, ...)

as_centimeter2(input, ...)

as_decimeter(input, ...)

as_decimeter2(input, ...)
}
\arguments{
\item{input}{Values that represent spatial measures.}

\item{...}{Needed arguments that depend on the input/unit combination. If
one of both is \emph{'px'}, argument \code{object} must be specified.}

\item{object}{An object of class \code{SPATA2}.}

\item{unit}{Character value. Specifies the desired unit.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
All functions return an output vector of the same length as the input
vector.

If argument \code{unit} is among \code{validUnitsOfLengthSI()} or \code{validUnitsOfAreaSI()}
the output vector is of class \code{units}. If argument \code{unit} is \emph{'px'}, the output
vector is a character vector or numeric vector if \code{as_numeric} is \code{TRUE}.
}
\description{
Collection of functions to transform distance and area values.
If pixels are involved, additional \code{SPATA2} specific content is needed.
}
\details{
For more information about area values, see details of \code{?is_area}. Fore
more information about distance values, see details of \code{?is_dist}.
}
\examples{

library(SPATA2)
library(SPATAData)

object <- downloadSpataObject("269_T")

pixel_values <- c(200, 450, 500)

si_dist_values <- c("2mm", "400mm", "0.2mm")

# spata object must be provided to scale based on current image resolution
as_millimeter(input = pixel_values, object = object, round = 2)

as_micrometer(input = pixel_values, object = object, round = 4)

as_pixel(input = si_dist_values, object = object)

# spata object must not be provided
as_micrometer(input = si_dist_values)


}
