% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-family.R
\name{asHistologyImage}
\alias{asHistologyImage}
\alias{asHistologyImage,VisiumV1-method}
\title{Convert to class \code{Visium}}
\usage{
asHistologyImage(object, ...)

\S4method{asHistologyImage}{VisiumV1}(object, scale_with = "lowres")
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
An object of class \code{Visium}.
}
\description{
Coverts objects of specific classes to objects
of class \code{Visium}.
}
