% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{asHistologyImaging}
\alias{asHistologyImaging}
\alias{asHistologyImaging,VisiumV1-method}
\title{Convert to class \code{HistologyImaging}}
\usage{
asHistologyImaging(object, ...)

\S4method{asHistologyImaging}{VisiumV1}(object, id, scale_with = "lowres", verbose = TRUE)
}
\arguments{
\item{object}{Any object for which a method has been defined.}
}
\value{
An object of class \code{HistologyImaging}.
}
\description{
Coverts objects of specific classes to objects
of class \code{HistologyImaging}.
}
