% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSeurat}
\alias{asSeurat}
\title{Transform SPATA2 object to Seurat object}
\usage{
asSeurat(
  object,
  process = TRUE,
  transfer_features = TRUE,
  assay_name = "Spatial",
  image_name = "slice1",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{process}{Logical value. If \code{TRUE}, count matrix is processed.
See details for more.

Use \code{getInitiationInfo()} to obtain argument input of your \code{SPATA2} object
initiation.}

\item{assay_name, image_name}{Character values. Define the name with which
to refer to the assay or the image in the \code{Seurat} object. Defaults to
the default of the \code{Seurat} package.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{Seurat}.
}
\description{
Transforms an \code{SPATA2} object to an object of class \code{Seurat}.
See details for more information.
}
\examples{
library(SPATA2)
library(Seurat)

data("example_data")

object <- example_data$object_UKF275T_diet

seurat_obj <- asSeurat(object)

class(seurat_obj)

}
