% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSingleCellExperiment}
\alias{asSingleCellExperiment}
\title{Transform to \code{SingleCellExperiment}}
\usage{
asSingleCellExperiment(
  object,
  assay_name = activeAssay(object),
  bayes_space = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{bayes_space}{Logical. If \code{TRUE}, the function creates an object
fitted to the requirements for the BayesSpace pipeline.}

\item{...}{The features to be renamed specified according to
the following syntax: 'new_feature_name' = 'old_feature_name'. This applies
to coordinates, too. E.g. ... ~ \emph{'image_col' = 'x', 'image_row' = 'y'}
renames the coordinate variables to \emph{'image_col'} and \emph{'image_row'}.}
}
\value{
An object of class \code{SingleCellExperiment}.
}
\description{
Transforms an \code{SPATA2} object to an object of class
\code{SingleCellExperiment}. See details for more information.
}
\details{
Output object contains the count matrix in slot @assays and
feature data.frame combined with coordinates
in slot @colData.

Slot @metadata is a list that contains the image object.
}
\examples{
library(SPATA2)
library(SingleCellExperiment)

data("example_data")

object <- example_data$object_UKF275T_diet

sce <- asSingleCellExperiment(object)

class(sce)
}
