% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSPATA2}
\alias{asSPATA2}
\alias{asSPATA2,giotto-method}
\alias{asSPATA2,Seurat-method}
\alias{asSPATA2,AnnDataR6-method}
\title{Transform to \code{spata2} object object}
\usage{
asSPATA2(object, ...)

\S4method{asSPATA2}{giotto}(object, sample_name, transfer_meta_data = TRUE, verbose = TRUE, ...)

\S4method{asSPATA2}{Seurat}(
  object,
  sample_name,
  spatial_method,
  assay_name = "Spatial",
  image_name = "slice1",
  transfer_meta_data = TRUE,
  transfer_dim_red = TRUE,
  count_mtr_name = "counts",
  scaled_mtr_name = "scale.data",
  verbose = TRUE
)

\S4method{asSPATA2}{AnnDataR6}(
  object,
  sample_name,
  count_mtr_name = "counts",
  normalized_mtr_name = "normalized",
  scaled_mtr_name = "scaled",
  transfer_meta_data = TRUE,
  transfer_dim_red = TRUE,
  image_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of either one of the following classes: \code{Seurat}, \code{SingleCellExperiment}, \code{AnnDataR6}}

\item{...}{Additional arguments given to \code{initiateSpataObject_CountMtr()}.}

\item{sample_name}{A character string specifying the name of the sample}

\item{transfer_meta_data}{Logical or character. Specifies
if meta data, e.g clustering, from the input object is transferred
to the output object. If TRUE, all variables of the meta data.frame
are transferred. If character, named variables are transferred. If FALSE,
none are transferred.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{spatial_method}{Character value. The name of the spatial method that underlies
the experiment. Must be one of \code{validSpatialMethods()}. Defaults to \emph{'Unknown'}.}

\item{transfer_dim_red}{A logical specifying whether to transfer dimensional reduction data (PCA, UMAP,
tSNE) from the input object to the output object.}

\item{count_mtr_name}{A character string specifying the name of the count matrix}

\item{scaled_mtr_name}{A character string specifying the name of the scaled matrix}

\item{normalized_mtr_name}{A character string specifying the name of the normalized matrix (anndata only currently)}
}
\value{
An object of class \code{spata2} object.
}
\description{
Transforms input object to object of class \code{spata2} object.
}
