% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSPATA2}
\alias{asSPATA2}
\alias{asSPATA2,giotto-method}
\alias{asSPATA2,Seurat-method}
\title{Transform to \code{spata2} object object}
\usage{
asSPATA2(object, ...)

\S4method{asSPATA2}{giotto}(object, sample_name, transfer_meta_data = TRUE, verbose = TRUE, ...)

\S4method{asSPATA2}{Seurat}(
  object,
  sample_name,
  assay_name = "Spatial",
  image_name = "slice1",
  transfer_meta_data = TRUE,
  transfer_dim_red = TRUE,
  count_mtr_name = "counts",
  scaled_mtr_name = "scale.data",
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Additional arguments given to \code{initiateSpataObject_CountMtr()}.}

\item{sample_name}{Character value. Denotes the name of the sample you are
analyzing with the spata-object. The future input for SPATA's \code{of_sample}-argument.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{transfer_features, transfer_meta_data}{Logical or character. Specifies
if meta/feature, e.g clustering, data from the input object is transferred
to the output object. If TRUE, all variables of the feature/meta data.frame
are transferred. If character, named variables are transferred. If FALSE,
none are transferred.}
}
\value{
An object of class \code{spata2} object.
}
\description{
Transforms input object to object of class \code{spata2} object.
}
