% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSPATA2}
\alias{asSPATA2}
\alias{asSPATA2,giotto-method}
\title{Transform to \code{SPATA2} object}
\usage{
asSPATA2(object, ...)

\S4method{asSPATA2}{giotto}(object, sample_name, transfer_meta_data = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{...}{Additional arguments given to \code{initiateSpataObject_CountMtr()}.}

\item{sample_name}{Character value. Denotes the name of the sample you are
analyzing with the spata-object. The future input for SPATA's \code{of_sample}-argument.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{transfer_features, transfer_meta_data}{Logical or character. Specifies
if meta/feature, e.g clustering, data from the input object is transferred
to the output object. If TRUE, all variables of the feature/meta data.frame
are transferred. If character, named variables are transferred. If FALSE,
none are transferred.}
}
\value{
An object of class \code{SPATA2}.
}
\description{
Transforms input object to object of class \code{SPATA2}.
}
