% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSPATA2}
\alias{asSPATA2}
\alias{asSPATA2,giotto-method}
\alias{asSPATA2,Seurat-method}
\title{Transform to SPATA2 object object}
\usage{
asSPATA2(object, ...)

\S4method{asSPATA2}{giotto}(
  object,
  sample_name,
  coordinates,
  image_ebi,
  spatial_method,
  transfer_meta_data = TRUE,
  verbose = TRUE,
  ...
)

\S4method{asSPATA2}{Seurat}(
  object,
  sample_name,
  spatial_method,
  assay_name = "Spatial",
  image_name = "slice1",
  transfer_meta_data = TRUE,
  transfer_dim_red = TRUE,
  count_mtr_name = "counts",
  scaled_mtr_name = "scale.data",
  verbose = TRUE
)
}
\value{
An object of class \code{\link{SPATA2}}.
}
\description{
Transforms input object to object of class \code{SPATA2} object.
}
