% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{asSpatialTrajectory}
\alias{asSpatialTrajectory}
\alias{asSpatialTrajectory,spatial_trajectory-method}
\title{Title}
\usage{
asSpatialTrajectory(object, ...)

\S4method{asSpatialTrajectory}{spatial_trajectory}(object, ...)
}
\description{
Title
}
