% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSpatialTrajectory}
\alias{asSpatialTrajectory}
\title{Transform to \code{SpatialTrajectory}}
\usage{
asSpatialTrajectory(object, ...)
}
\description{
Transforms old spatial trajectory class to new one.
}
\keyword{internal}
