% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-aZ.R
\name{asSummarizedExperiment}
\alias{asSummarizedExperiment}
\title{Transform SPATA2 object to SummarizedExperiment object}
\usage{
asSummarizedExperiment(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{The features to be renamed specified according to
the following syntax: 'new_feature_name' = 'old_feature_name'. This applies
to coordinates, too. E.g. ... ~ \emph{'image_col' = 'x', 'image_row' = 'y'}
renames the coordinate variables to \emph{'image_col'} and \emph{'image_row'}.}
}
\value{
An object of class \code{SummarizedExperiment}.
}
\description{
Transforms an \code{SPATA2} object to an object of class
\code{SummarizedExperiment}. See details for more information.
}
\details{
Output object contains the count matrix in slot @assays and
feature data.frame combined with barcode-spot coordinates
in slot @colData.

Slot @metadata is a list that contains the image object in slot $image.
}
