% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{bin_by_expansion}
\alias{bin_by_expansion}
\title{Bin barcode-spots by area extension}
\usage{
bin_by_expansion(
  coords_df,
  area_df,
  binwidth,
  n_bins_circle,
  remove = FALSE,
  bcsp_exclude = NULL,
  drop = TRUE
)
}
\arguments{
\item{coords_df}{The coordinates data.frame whose barcode-spots are supposed
to be binned.}

\item{area_df}{Data.frame with variables \emph{x} and \emph{y} describing the
vertices of the polygon that encircles the area based on which the barcode-spots
are binned. E.g. slot @area of \code{ImageAnnotation}-objects.}

\item{binwidth}{Distance value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the center-center
distance: \code{binwidth = getCCD(object)}. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
(See details for more.)}

\item{remove}{Character or logical. If character, denotes circle bins that
are removed. If TRUE, bins \emph{'Core' and 'Outside'} are removed. If FALSE,
ignored.}

\item{bcsp_exclude}{Character value containing name(s) of barcode-spots to be excluded from the analysis.}

\item{drop}{Logical value. If TRUE, unused levels of the \emph{bins_circle}
variables are dropped.}
}
\description{
Bins barcode-spots by consecutively expanding a polygon.
}
