% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{buffer_area}
\alias{buffer_area}
\title{Buffer area}
\usage{
buffer_area(df, buffer)
}
\arguments{
\item{df}{Data.frame with two variables, named \emph{x}
and \emph{y} corresponding to the vertices of the polygon.}
}
\description{
Buffers the area of a polygon.
}
