% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{buffer_area}
\alias{buffer_area}
\title{Buffer area}
\usage{
buffer_area(df, buffer)
}
\arguments{
\item{df}{Data.frame with variables \emph{x} and \emph{y} describing the
vertices of the polygon that encircles the area based on which the barcode-spots
are binned. E.g. slot @area of \code{ImageAnnotation}-objects.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}. (See details for more.)}
}
\description{
Buffers the area of a polygon.
}
