% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{centerTissueOutline}
\alias{centerTissueOutline}
\alias{centerTissueOutline,HistoImage-method}
\title{Center tissue}
\usage{
centerTissueOutline(object, ...)

\S4method{centerTissueOutline}{HistoImage}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Computes the necessary translations in order to center
the identified tissue outline in the center of the image.
}
