% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_atdf}
\alias{check_atdf}
\alias{check_rtdf}
\title{Check assessed trajectory data.frame}
\usage{
check_atdf(atdf)

check_rtdf(rtdf, variable = NULL)
}
\arguments{
\item{atdf}{A data.frame containing the results of trajectory-modelling. Must contain the variables:

 \describe{
  \item{\emph{variables}}{Character. The genes, gene-sets and features that have been assessed.}
  \item{\emph{pattern}}{Character. The respective trajectory pattern.}
  \item{\emph{auc}}{Numeric. The assessment value which is the residual's area under the curve.}
 }}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
