% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_color_to}
\alias{check_color_to}
\title{Check color to}
\usage{
check_color_to(
  color_to,
  color_by,
  all_features = character(),
  all_gene_sets = character(),
  all_genes = character(),
  max_length = NULL
)
}
\arguments{
\item{color_by}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{all_features}{The valid features specified as a character vector.}

\item{all_gene_sets}{The valid gene sets specified as a character vector.}

\item{all_genes}{The valid genes specified as a character vector.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_()}-family. Takes a character
vector and sorts its elements into a list depending on whether it was found in
the input of \code{all_features}, \code{all_genes} or \code{all_gene_sets}.

Returns a list with one slot named \emph{features}, \emph{genes} or \emph{gene_sets}
containing the respective found/valid input of \code{color_to}.
}
\details{
Members of the \code{adjusting-check_*()}-family take their
arguments input and compare it to a variety of requirement settings by
running several logical tests. If the input turns out to be appropriate
for the main-function they return it the way it is supposed to be returned.
If not, depending on the degree of deviation from the optimum, they either adjust
the input in order not to interrupt the function or - if not adjustable - raise an
error. In both cases informative messages will be printed in order to let the user
know what has been adjusted or what part of the input was insufficient.
}
