% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_coordinate_variables}
\alias{check_coordinate_variables}
\title{Check coordinate variables}
\usage{
check_coordinate_variables(data, x = "x", y = "y")
}
\arguments{
\item{data}{A data.frame containing the variables of interest as well
as the variables needed to map onto the x- and y axis of the plot.}

\item{x}{The name of the numeric variable to be plotted on the x axis.}

\item{y}{The name of the numeric variable to be plotted on the y axis.}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
