% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_coords_df}
\alias{check_coords_df}
\title{Check coords data.frame}
\usage{
check_coords_df(coords_df)
}
\arguments{
\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\itemize{
\item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
\item{\emph{sample} Character. The sample belonging of every barcode-spot.}
\item{\emph{x} Numeric. The x-coordinates of every barcode.}
\item{\emph{y} Numeric. The y-coordinates of every barcode.}
}}
}
\description{
Check coords data.frame
}
