% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_coords_df}
\alias{check_coords_df}
\title{Check coords data.frame}
\usage{
check_coords_df(coords_df)
}
\arguments{
\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\itemize{
 \item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
 \item{\emph{sample} Character. The sample belonging of every barcode-spot.}
 \item{\emph{x} Numeric. The x-coordinates of every barcode.}
 \item{\emph{y} Numeric. The y-coordinates of every barcode.}
}}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
