% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_dea_df}
\alias{check_dea_df}
\title{Check de data.frame}
\usage{
check_dea_df(dea_df)
}
\arguments{
\item{dea_df}{A data.frame containing information about differentially expressed genes. Must contain the variables:

 \describe{
  \item{\emph{gene}}{Character. The differentially expressed genes.}
  \item{\emph{cluster}}{Character. The clusters (or experimental groups) across which the analysis was performed.}
  \item{\emph{avg_logFC}}{Numeric. The average log-fold change to which the belonging gene was differentially expressed..}
  \item{\emph{p_val}}{Numeric. The p-values.}
  \item{\emph{p_val_adj}}{Numeric. The adjusted p-values.}
 }

Hint: Use the resulting data.frame of \code{SPATA::findDE()} or it's descendants as input.}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
