% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_feature_df}
\alias{check_feature_df}
\title{Check feature data.frame}
\usage{
check_feature_df(feature_name, feature_df)
}
\arguments{
\item{feature_name}{Character value. The name of the feature that is to be added
to the object.}

\item{feature_df}{A data.frame that contains the feature and the key-variables.}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
