% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_genes}
\alias{check_genes}
\title{Check gene input}
\usage{
check_genes(
  object,
  genes,
  valid_genes = NULL,
  rna_assay = NULL,
  max_length = NULL,
  fdb_fn = "warning",
  ...
)
}
\arguments{
\item{genes}{The genes of interest specified as a character vector.}

\item{rna_assay}{The rna-assay you want to
look in. If set to NULL the whole rna_assay of the specified object will be used
with \code{getExpressionMatrix()}.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of gene names and checks which of the genes exist.

Returns an adjusted genes-vector or raises an error.
}
\details{
Members of the \code{adjusting-check_*()}-family take their
arguments input and compare it to a variety of requirement settings by
running several logical tests. If the input turns out to be appropriate
for the main-function they return it the way it is supposed to be returned.
If not, depending on the degree of deviation from the optimum, they either adjust
the input in order not to interrupt the function or - if not adjustable - raise an
error. In both cases informative messages will be printed in order to let the user
know what has been adjusted or what part of the input was insufficient.
}
