% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_pattern}
\alias{check_pattern}
\title{Check pattern input}
\usage{
check_pattern(object, patterns = "", method_pr = "hotspot", of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{gene_sets}{The gene sets of interest specified as a character vector.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a character
vector of pattern names and checks which of these exist.

Returns an adjusted gene-set-vector or raises an error.
}
\details{
Members of the \code{adjusting-check_*()}-family take their
arguments input and compare it to a variety of requirement settings by
running several logical tests. If the input turns out to be appropriate
for the main-function they return it the way it is supposed to be returned.
If not, depending on the degree of deviation from the optimum, they either adjust
the input in order not to interrupt the function or - if not adjustable - raise an
error. In both cases informative messages will be printed in order to let the user
know what has been adjusted or what part of the input was insufficient.
}
