% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-adjusting.R
\name{check_saving}
\alias{check_saving}
\title{Check saving input}
\usage{
check_saving(output_path, file_name)
}
\arguments{
\item{output_path}{Character vector or NULL. Specifies the folder in which to store
the object if the directory is valid. If set to NULL saving is skipped.}

\item{file_name}{Character value. The name-suffix for the file name under which the
spata-object is stored if \code{output_path} is a valid directory. Is prefixed with
 \emph{'spata-obj-'} and suffixed with \emph{'.RDS'}.}
}
\value{
The original input, an adjusted version of it or an error. In the latter two
cases they print an informative message about what was going on.
}
\description{
A member of the \code{adjusting-check_*()}-family. Takes a folder directory
and assembles the final output directory including the specified filename.

Returns the final directory if it is valid or raises an error if not or NULL if \code{output_path}
is set to NULL.
}
\details{
Members of the \code{adjusting-check_*()}-family take their
arguments input and compare it to a variety of requirement settings by
running several logical tests. If the input turns out to be appropriate
for the main-function they return it the way it is supposed to be returned.
If not, depending on the degree of deviation from the optimum, they either adjust
the input in order not to interrupt the function or - if not adjustable - raise an
error. In both cases informative messages will be printed in order to let the user
know what has been adjusted or what part of the input was insufficient.
}
