% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-family-lazy.R
\name{check_to}
\alias{check_to}
\title{Gives feedback about input validity}
\usage{
check_to(object, to)
}
\arguments{
\item{to}{Character value. Denotes the object to which
the directory leads. Must be one of \emph{'cell_data_set', 'seurat_object'}
or \emph{'spata_object'}.}
}
\value{
A logical value TRUE if nothing invalid has been detected or an informative
error message.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
