% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_trajectory}
\alias{check_trajectory}
\title{Check trajectory name input}
\usage{
check_trajectory(object, trajectory_name, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\description{
Check trajectory name input
}
