% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{close_area_df}
\alias{close_area_df}
\title{Close area encircling}
\usage{
close_area_df(df)
}
\description{
"Closes" the area described by the vertices of \code{df} by
adding the starting point (first row) to the end of the data.frame.
}
\keyword{internal}
