% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-correlation.R
\name{clusterSpCorResults}
\alias{clusterSpCorResults}
\title{Cluster genes according to their expression profile}
\usage{
clusterSpCorResults(
  object,
  of_sample = "",
  method_hclust = "complete",
  k = NULL,
  h = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{method_hclust}{Character value. Denotes the hierarchical clustering method  according
to which the clustering is performed. Valid input options are \emph{'ward.D', 'ward.D2', 'single',
 'complete', 'average', 'mcquitty', 'median'} and \emph{'centroid'}.}

\item{k, h}{Numeric values. Given to the respective arguments of \code{stats::cutree()}
and are used to determine the eventual cluster belonging of every gene. Use
\code{plotCorrelationDendrogam()} for orientation. \code{k} overrides \code{h}
if both are specified as numeric values.}
}
\value{
An updated spata-object.
}
\description{
Takes the distance matrix that was computed with
\code{runSpCorAnalysis()} and clusters it via hierarchical clustering such that
genes sharing similar expression profiles / patterns throughout the sample are
clustered together. See details for a more detailed description of what the result
is composed of.

(This function requires that \code{runSpatialCorrelationAnalysis()} has been run.)
}
\details{
The resulting value, a named list, is stored in the spata object in
the @spatial$correlation$clusters slot. The slot that the list occupies is
named according to the method denoted with the \code{method_hclust}-argument.
Apart from the belonging value for \code{k} or \code{h} the list contains three
main slots:

\itemize{
 \item{\emph{assessment_df}: A data.frame that attempts to evaluate all cluster's
  quality by providing the average distance between it's genes.}
  \item{\emph{distances_list}: A named list of data.frames. Each data.frame contains the
  gene-gene distances between all genes the cluster it corresponds to contains.}
  \item{\emph{gene_names_list}: A named list of character vectors. Each vector contains the
  unique gene names of the cluster it corresponds to.}
  }

Hierarchical clustering is an explorative approach not a deterministic an. The optimal input for
\code{method_hclust}, \code{k} and \code{h} depend on the data's properties as well as the
biological question at hand. Therefore, in the same fashion as with differential gene
expression analysis SPATA allows to store the results of different approaches within
one and the same object and to conveniently extract/plot them via the respective arguments.
While different inputs for \code{method_hclust} are stored accumulatively different inputs for \code{k}
and \code{h} overwrite the preexisting cluster results of the chosen method!
}
