% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-sub-functions.R
\name{compute_pattern_similarity}
\alias{compute_pattern_similarity}
\alias{compute_pattern_relation}
\title{Title}
\usage{
compute_pattern_similarity(bcx, bcy, pb = NULL, verbose = TRUE)

compute_pattern_relation(bcx, bcy, x, y, pb = NULL)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
