% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{computeMetaFeatures}
\alias{computeMetaFeatures}
\title{Compute meta features for molecular data}
\usage{
computeMetaFeatures(
  object,
  assay_name = activeAssay(object),
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest (and thus the
active \link[=concept_molecular_modalities]{molecular modality}. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
This function computes meta features for molecular data observation wise.
Meta features include the number of counts and the number of distinct molecules observed per barcode.
}
\details{
This function computes meta features such as the number of counts
and the number of distinct molecules per observation. The computed meta
features are added to the input object via \code{\link[=addFeatures]{addFeatures()}}.
}
