% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{computePixelScaleFactor}
\alias{computePixelScaleFactor}
\alias{computePixelScaleFactor,HistoImaging-method}
\title{Compute pixel scale factor}
\usage{
computePixelScaleFactor(object, ...)

\S4method{computePixelScaleFactor}{HistoImaging}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Computes the pixel scale factor. Currently, only possible for spatial methods
\emph{VisiumSmall} and \emph{VisiumLarge}.
}
