% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsAssay}
\alias{containsAssay}
\title{Check availability of an assay}
\usage{
containsAssay(object, assay_name, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest (and thus the
active \link[=concept_molecular_modalities]{molecular modality}. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if the provided object contains a specific assay.
}
\seealso{
\code{\link[=getAssayNames]{getAssayNames()}}, \code{\link{MolecularAssay}}
}
