% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsCNV}
\alias{containsCNV}
\title{Check availability CNV results}
\usage{
containsCNV(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Logical tests that check if content exists in the \code{SPATA2} object.
}
