% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{containsHistologyImaging}
\alias{containsHistologyImaging}
\title{Check availability of \code{HistologyImaging} object}
\usage{
containsHistologyImaging(object)
}
\description{
Deprecated in favor of \code{\link[=containsHistoImaging]{containsHistoImaging()}}.
}
\keyword{internal}
