% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{containsHistologyImaging}
\alias{containsHistologyImaging}
\title{Check availability of \code{HistologyImaging} object}
\usage{
containsHistologyImaging(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Logical value.
}
\description{
Checks if slot @images contains an object
of class \code{HistologyImaging} or if it is empty.
}
