% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{containsHistologyImaging}
\alias{containsHistologyImaging}
\title{Deprecated}
\usage{
containsHistologyImaging(...)
}
\description{
Deprecated in favor of \code{\link[=containsHistoImages]{containsHistoImages()}}.
}
\keyword{internal}
