% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{containsImage}
\alias{containsImage}
\title{Check availability of an image}
\usage{
containsImage(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Logical value.
}
\description{
Checks if slot @image of the \code{HistologyImage} object
in the \code{SPATA2} object contains an image or if it is empty.
}
