% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{containsImage}
\alias{containsImage}
\alias{containsImage,spata2-method}
\alias{containsImage,HistoImage-method}
\title{Check availability of an image}
\usage{
containsImage(object, ...)

\S4method{containsImage}{spata2}(object)

\S4method{containsImage}{HistoImage}(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Logical value.
}
\description{
Checks if slot @image of the \code{HistoImage} object
in the \code{SPATA2} object contains an image or if it is empty.
}
