% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsMethod}
\alias{containsMethod}
\alias{containsMethod,SPATA2-method}
\alias{containsMethod,SpatialData-method}
\title{Check availability of specific spatial methods (platforms)}
\usage{
containsMethod(object, ...)

\S4method{containsMethod}{SPATA2}(object, method_name, error = FALSE)

\S4method{containsMethod}{SpatialData}(object, method_name, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{method_name}{Character vector. The names of platforms to check.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value. TRUE if any of \code{method_name} is the name of the platform.
}
\description{
Tests if the input object is associated with
specific \link[=spatial_methods]{spatial methods / platforms}.
}
