% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{containsMethod}
\alias{containsMethod}
\alias{containsMethod,spata2-method}
\alias{containsMethod,HistoImaging-method}
\title{Check availability of specific methods}
\usage{
containsMethod(object, ...)

\S4method{containsMethod}{spata2}(object, method_name, error = FALSE)

\S4method{containsMethod}{HistoImaging}(object, method_name, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{method_name}{Character vector. Names of methods to check.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if the input object is associated with
specific methods.
}
