% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{containsPixelScaleFactor}
\alias{containsPixelScaleFactor}
\title{Check availability of pixel scale factor}
\usage{
containsPixelScaleFactor(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Logical value.
}
\description{
Checks if a pixel scale factor is present in the \code{SPATA2}
object
}
