% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsScaleFactor}
\alias{containsScaleFactor}
\alias{containsScaleFactor,SPATA2-method}
\alias{containsScaleFactor,SpatialData-method}
\title{Check availability of specific scale factors}
\usage{
containsScaleFactor(object, ...)

\S4method{containsScaleFactor}{SPATA2}(
  object,
  fct_name,
  img_name = activeImage(object),
  error = FALSE
)

\S4method{containsScaleFactor}{SpatialData}(
  object,
  fct_name,
  img_name = activeImage(object),
  error = FALSE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{fct_name}{Character value. The name of the scale factor of interest.
E.g. \emph{'pixel'} or \emph{'coords'}.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if specifics scale factors are set or not.
}
