% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsSpatialAnnotations}
\alias{containsSpatialAnnotations}
\alias{containsSpatialAnnotations,spata2-method}
\alias{containsSpatialAnnotations,HistoImaging-method}
\title{Check availability of spatial annotations}
\usage{
containsSpatialAnnotations(object, ...)

\S4method{containsSpatialAnnotations}{spata2}(object, error = FALSE)

\S4method{containsSpatialAnnotations}{HistoImaging}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Tests if the object contains spatial annotations
as created by \code{\link{createGroupAnnotations}} \code{\link{createImageAnnotations}} and
\code{\link[=createNumericAnnotations]{createNumericAnnotations()}}.
}
