% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{containsSpatialOutliers}
\alias{containsSpatialOutliers}
\alias{containsSpatialOutliers,ANY-method}
\title{Check if spatial outliers exist}
\usage{
containsSpatialOutliers(object, ...)

\S4method{containsSpatialOutliers}{ANY}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Logical value.
}
\description{
Checks if \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}} has identified any
spatial outliers.
}
\seealso{
\code{\link[=excludeSpatialOutliers]{excludeSpatialOutliers()}} to exclude spatial outliers from further
analysis.
}
