% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{createHistoImaging}
\alias{createHistoImaging}
\alias{createHistoImagingVisium}
\title{Create an object of class \code{HistoImaging}}
\usage{
createHistoImaging(
  sample,
  hist_img_ref,
  hist_imgs = list(),
  active = NULL,
  empty_image_slots = FALSE,
  coordinates = tibble::tibble(),
  coordinates_id = NULL,
  meta = list(),
  method = SpatialMethod(),
  misc = list(),
  verbose = TRUE,
  ...
)

createHistoImagingVisium(
  dir,
  sample,
  img_ref = "hires",
  img_active = "lowres",
  meta = list(),
  misc = misc(),
  verbose = TRUE
)
}
\arguments{
\item{sample}{Character value. The sample name of the tissue.}

\item{hist_img_ref}{The \code{HistoImaging} serving as the reference image.
Should be created with \code{createHistoImage()}.}

\item{hist_imgs}{List of additional \code{HistoImaging} objects for slot @images.}

\item{active}{Character value. Name of the \code{HistoImage} that is set
to the active image. Defaults to the reference image.}

\item{empty_image_slots}{Logical value. If \code{TRUE}, content of slot @image
of all \code{HistoImage} objects is emptied except for the active one.}

\item{coordinates}{Data.frame of at least three variables:

\itemize{
\item{\code{coordinates_id}: }{Character variable with unique IDs for each observation.}
\item{\emph{x_orig}: }{Numeric variable representing x-coordinates in a cartesian coordinate system.}
\item{\emph{y_orig}: }{Numeric variable representing y-coordinates in a cartesian coordinate system.}
}

Coordinates should align with the tissue outline of the reference \code{HistoImage} after being
multiplied withe its coordinate scale factor in slot @scale_factors$coords.}

\item{coordinates_id}{Character value. The name of the ID variable of the data.frame
from \code{coordinates} when it is extracted via \code{getCoordsDf()}.}

\item{meta}{List of meta data regarding the tissue.}

\item{misc}{List of miscellaneous information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{dir}{The directory to the output folder of the platform.}

\item{img_ref, img_active}{Character values specifying the active and reference images. Choose between \emph{'lowres'} and \emph{'hires'}.
Setting both \code{img_ref} and \code{img_active} to the same value indicates that only one image is used,
and the other image will not be read automatically. In such cases, manual registration of the
unselected image is required using the \code{registerHistoImage()} function.}
}
\value{
An object of class \code{HistoImaging}
}
\description{
Official constructor function of the S4 class \code{HistoImaging}.
Functions suffixed by the platform name are wrappers written for the
standardized output folder.
}
\seealso{
\code{\link[=createHistoImage]{createHistoImage()}}, \code{\link[=registerHistoImage]{registerHistoImage()}}
}
