% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createImageAnnotations}
\alias{createImageAnnotations}
\title{Add spatial annotations based on histo-morphological features}
\usage{
createImageAnnotations(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Opens an interface in which the user can interactively outline
histomorphological features of an image. The outline created this way is
used to create a \code{\link{SpatialAnnotation}} of subclass \code{\link{ImageAnnotation}}.

Not to confuse with \code{createSpatialSegmentation()}.
}
\seealso{
\code{\link[=addSpatialAnnotation]{addSpatialAnnotation()}}
}
