% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createImageAnnotations}
\alias{createImageAnnotations}
\title{Add image annotations}
\usage{
createImageAnnotations(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Functions to add image annotations the \code{SPATA2} object. For
interactive drawing use \code{createImageAnnotaions()}. To set them with code
use \code{addImageAnnotation()}.

Not to confuse with \code{createSegmentation()}.
}
\note{
The interface allows to zoom in on the sample. This is useful if your
spata object contains an HE-image as background and you want to classify
barcode spots based on the histology. As these images are displayed by pixels
the resolution decreases the more you zoom in. Many experiments (such as
the Visium output) contain high resolution images. You can use the function
\code{exchangeImage()} to read in images of higher resolution for a better
histological classification.
}
\seealso{
exchangeImage(), plotImageAnnotations(), getImageAnnotations()
}
