% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createImageAnnotations}
\alias{createImageAnnotations}
\title{Create spatial annotations based on histo-morphological features}
\usage{
createImageAnnotations(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Opens an interface in which the user can interactively outline
histomorphological features of an image. The outline created this way is
used to create a \code{\link{SpatialAnnotation}} of subclass \code{\link{ImageAnnotation}}.

Not to confuse with \code{\link[=createSpatialSegmentation]{createSpatialSegmentation()}}.
}
\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

object <- createImageAnnotations(object)

}
\seealso{
\code{\link[=addSpatialAnnotation]{addSpatialAnnotation()}}, \code{\link[=mergeWithTissueOutline]{mergeWithTissueOutline()}}
}
