% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createImageAnnotations}
\alias{createImageAnnotations}
\title{Interactive image annotations}
\usage{
createImageAnnotations(object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
An updated spata object.
}
\description{
This function gives access to an interactive interface in
which \code{ImageAnnotations} are created by encircling regions or
structures of interest in the histology image.

Not to confuse with \code{createSegmentation()}.
}
\note{
The interface allows to zoom in on the sample. This is useful if your
spata object contains an HE-image as background and you want to classify
barcode spots based on the histology. As these images are displayed by pixels
the resolution decreases the more you zoom in. Many experiments (such as
the Visium output) contain high resolution images. You can use the function
\code{exchangeImage()} to read in images of higher resolution for a better
histological classification.
}
\seealso{
exchangeImage(), plotImageAnnotations(), getImageAnnotations()
}
