% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{createSegmentation}
\alias{createSegmentation}
\title{Deprecated}
\usage{
createSegmentation(...)
}
\description{
Deprecated in favor of \code{\link[=createSpatialSegmentation]{createSpatialSegmentation()}}.
}
\keyword{internal}
