% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family.R
\name{createSegmentation}
\alias{createSegmentation}
\title{Spatial Segmentation}
\usage{
createSegmentation(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated version of the spata-object specified as \code{object}
now containing the information about all drawn segments.
}
\description{
The function \code{createSegmentation()} provides access to an
interactive mini-shiny application that allows to separate a sample into
several segments.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
