% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpatialSegmentation.R
\name{createSpatialSegmentation}
\alias{createSpatialSegmentation}
\title{Interactive sample segmentation}
\usage{
createSpatialSegmentation(
  object,
  height = 500,
  break_add = NULL,
  box_widths = c(4, 4, 4)
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\description{
Gives access to an interactive user interface where data points
can be interactively annotated.
}
\details{
Segmentation variables are grouping variables that are stored in
the meta data.frame of the \code{SPATA2} object (such as clustering variables).
They differ from clustering variables in so far as that they are not the result
of unsupervised cluster algorithms but from group assignment the researcher
conducts him/herself (e.g. histological classification).

Therefore, all segmentation variables can be extracted via \code{getFeatureNames()}
as they are part of those. To specifically extract variables that were created
with \code{createSpatialSegmentation()} use \code{getSegmentationNames()}. To remove
annotations you no longer need use \code{removeFeatures()}.
}
\note{
The interface allows to zoom in on the sample. This is useful if your
\code{SPATA2} object contains an HE-image as background and you want to classify
barcode spots based on the histology. As these images are displayed by pixels
the resolution decreases the more you zoom in. Many experiments (such as
the Visium output) contain high resolution images. You can use the function
\code{\link[=registerImage]{registerImage()}} to register images of higher resolution for a better
histological classification.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

if(FALSE){

 object <- createSpatialSegmentation(object)

 }

}
