% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createSpatialTrajectories}
\alias{createSpatialTrajectories}
\title{Create and add spatial trajectories interactively}
\usage{
createSpatialTrajectories(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\description{
Opens an interface in which the user can interactively draw spatial trajectories
on the surface of the sample.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

object <- createSpatialTrajectories(object)

}
\seealso{
\code{\link[=addSpatialTrajectories]{addSpatialTrajectories()}}
}
