% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createSpatialTrajectories}
\alias{createSpatialTrajectories}
\title{Create spatial trajectories}
\usage{
createSpatialTrajectories(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
An updated \code{SPATA2} object.
}
\description{
Provides access to an interactive shiny application
where trajectories can be drawn.
}
