% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{createTrajectoryManually}
\alias{createTrajectoryManually}
\title{Create spatial trajectories manually}
\usage{
createTrajectoryManually(
  object,
  trajectory_name,
  start,
  end,
  width,
  vertices = list(),
  comment = "",
  plot = FALSE,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{trajectory_name}{Character value. The name of the new trajectory.}

\item{start, end}{Numeric vectors of length two. Defining start and endpoint
of the trajectory. First value of each vector is used as the respective
x-coordinate. Second value of each vector is used as the respective y-coordinate.}

\item{width}{Numeric value. Denotes the trajectory width.}

\item{vertices}{List. Optional if you want to specify additional vertices
between start and endpoint to split the trajectory into parts.

Every slot of the input list must be a numeric vector which
is then handled in the same way that the input of arguments \code{start} and \code{end}
is handled - first value is taken for x- and second value for y-coordinate.

Ignored if not a list or a list of length 0.}

\item{comment}{Character value. Optional if you want to provide a reasoning
why you have drawn the trajectory.}

\item{plot}{Logical value. If set to TRUE the created trajectory is plotted
via \code{plotTrajectory()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata-object.
}
\description{
Manual version of \code{createTrajectories()}. Instead of
drawing them interactively you can provide the coordinates via the
arguments \code{width}, \code{vertices}, \code{start} and \code{end}.
}
\examples{


 object <-
  createTrajectoryManually(
     object = object,
     trajectory_name = "my_trajectory",
     start = c(x = 136, y = 181),
     end = c(x = 381, y = 398),
     vertices = list(va = c(x = 251, y = 283), vb = c(x = 344, y = 356)),
     width = 25,
     comment = 'This serves as an example.'
       )

}
