% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{crop_image}
\alias{crop_image}
\title{Crop image}
\usage{
crop_image(image, xrange = NULL, yrange = NULL, expand = 0, ...)
}
\arguments{
\item{image}{Object of class \code{Image} from the \code{ÈBIMage} package.}
}
\value{
Cropped input object.
}
\description{
Crops an image.
}
\keyword{internal}
