% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{cropSpataObject}
\alias{cropSpataObject}
\title{Subset by x- and y-range}
\usage{
cropSpataObject(
  object,
  xrange,
  yrange,
  adjust_capture_area = TRUE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}

\item{adjust_capture_area}{Logical. If \code{TRUE}, the capture area is adjusted
to the input of \code{xrange} and \code{yrange}. If \code{FALSE}, it stays as is. Defaults to \code{TRUE}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
Creates a subset of the original \code{SPATA2} object
based on x- and y-range. Data poitns that fall into the
rectangle given by \code{xrange} and \code{yrange} are kept.
}
\seealso{
\code{\link[=ggpLayerRect]{ggpLayerRect()}} to visualize the rectangle based on which
the subsetting is done.
}
