% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{discardExpressionMatrix}
\alias{discardExpressionMatrix}
\title{Discard an expression matrix}
\usage{
discardExpressionMatrix(...)
}
\value{
An updated spata-object.
}
\description{
Discards the expression matrix of choice.
}
