% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{discardExpressionMatrix}
\alias{discardExpressionMatrix}
\title{Discard an expression matrix}
\usage{
discardExpressionMatrix(object, mtr_name, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}
}
\value{
An updated spata-object.
}
\description{
Discards the expression matrix of choice.
}
