% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{discardFeatures}
\alias{discardFeatures}
\title{Discard features}
\usage{
discardFeatures(object, feature_names, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_names}{Character vector. Specifies the features to be discarded.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
An updated spata-object.
}
\description{
Discards the features of choice.
}
