% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{discardGeneFeatures}
\alias{discardGeneFeatures}
\title{Discard gene features}
\usage{
discardGeneFeatures(object, feature_names, mtr_name = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_names}{Character vector. Specifies the gene features to be discarded.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}
}
\value{
An updated spata-object.
}
\description{
Discards the features of choice of the gene meta data.
}
