% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-family.R
\name{discardGeneSets}
\alias{discardGeneSets}
\title{Discard gene sets}
\usage{
discardGeneSets(object, gs_names)
}
\arguments{
\item{object}{A valid spata-object.}

\item{gs_names}{Character vector. The gene sets to be discarded.}
}
\value{
An updated spata-object.
}
\description{
A member of the \code{lazy-check_*()}-family.
}
\details{
Members of the \code{lazy-check_*()}-family take the arguments
of their kind - that are used in the function they are called in - and
checks whether these arguments input fit the requirements. They stop and return an
error immediately once they stumble upon something invalid. They do not alter or adjust input
and return TRUE if the whole function has been executed without anything
invalid being found.
}
