% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{discardImageAnnotations}
\alias{discardImageAnnotations}
\title{Discard image annotations}
\usage{
discardImageAnnotations(object, ids)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{ids}{Character vector. The IDs of the image annotations to
be discarded.}
}
\value{
An updated spata object.
}
\description{
Discards image annotations drawn with \code{annotateImage()}.
}
