% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{distToEdge}
\alias{distToEdge}
\title{Distance to cover the whole tissue}
\usage{
distToEdge(object, id = idSA(object), unit = getDefaultUnit(object))
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{unit}{The output unit of the distance measure.}
}
\value{
Distance measure.
}
\description{
Computes the distance from the center of a spatial annotation
to the \strong{farest} point of the tissue outline.
}
