% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{downloadSpataObject}
\alias{downloadSpataObject}
\alias{downloadSpataObjects}
\title{Download \code{spata2} objects}
\usage{
downloadSpataObject(
  sample_name,
  overwrite = FALSE,
  folder = base::getwd(),
  file = NULL,
  in_shiny = FALSE,
  verbose = TRUE,
  ...
)

downloadSpataObjects(
  sample_names,
  files = NULL,
  folder = base::getwd(),
  overwrite = FALSE,
  in_shiny = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name of the sample you want to
download. Use \code{validSampleNames()} to obtain all valid input options.}

\item{overwrite}{Logical. Must be set to TRUE if file directories
under which downloaded files are about to be saved already exist.}

\item{folder}{Character value. If character, specifies the output
folder in which the spata object is saved. Defaults to the working directory.}

\item{file}{The filename of the spata object. Must end with \emph{'.RDS'}. By
default the file it is `NULL` which makes the function skip the saving step.
Set to character, if you want the object to be saved.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{sample_names}{Character vector. The sample names of the spata objects
to be downloaded. Use \code{validSampleNames()} to obtain all valid input options.}

\item{files}{Character vector or NULL. Specifies the file names under which the
spata objects are saved. If character, the input must be of the same length
as the input for argument \code{sample_names}. If NULL, the files are named
according to the sample name.}
}
\value{
The downloaded spata object.
}
\description{
Downloads a spata object and returns it. For convenient
downloads of multiple spata objects check out \code{downloadSpataObjects()}.
}
\note{
Imported from the package \code{SPATAData}.
}
\examples{
# only download
downloadSpataObject(sample_name = "275_T", folder = getwd())

# download AND assign
object <- downloadSpataObject(sample_name = "275_T", folder = getwd())

# only assign (default)
object <- downloadSpataObject(sample_name = "275_T", file = NULL)
}
