% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{downloadSpataObject}
\alias{downloadSpataObject}
\alias{downloadSpataObjects}
\title{Download \code{spata2} objects}
\usage{
downloadSpataObject(
  sample_name,
  overwrite = FALSE,
  file = FALSE,
  in_shiny = FALSE,
  verbose = TRUE,
  ...
)

downloadSpataObjects(
  sample_names,
  files = NULL,
  folder = base::getwd(),
  overwrite = FALSE,
  in_shiny = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name of the sample you want to
download. Use \code{validSampleNames()} to obtain all valid input options.}

\item{overwrite}{Logical. Must be set to \code{TRUE} if file directories
under which downloaded files are to be saved already exist.}

\item{file}{If you want to save the object on disc:
The filename of the \code{SPATA2} object. Must end with \emph{'.RDS'} if provided
as a character. If \code{NULL}, the function saves the object
under the sample name with an \emph{'.RDS'} suffix. If \code{FALSE}, the saving
is skipped and the object is simply returned.}

\item{sample_names}{Character vector. The sample names of the \code{SPATA2} objects
to be downloaded. Use \code{validSampleNames()} to obtain all valid input options.}

\item{files}{Character vector or NULL. Specifies the file names under which the
\code{SPATA2} objects are saved. If character, the input must be of the same length
as the input for argument \code{sample_names}. If NULL, the files are named
according to the sample name.}

\item{folder}{Character value. If character, specifies the output
folder in which the \code{SPATA2} object is saved. Defaults to the working directory.}
}
\value{
The downloaded \code{SPATA2} object.
}
\description{
Downloads a \code{SPATA2} object and returns it. For convenient
downloads of multiple \code{SPATA2} objects check out \code{downloadSpataObjects()}.
}
\note{
Imported from the package \code{SPATAData}.
}
\examples{
# download & assign (no saving on the disk)
object <- downloadSpataObject(sample_name = "UKF275T")

# download, assign and save on disk
# -> stores the file under ~/UKF275T.RDS (where '~' is your working directory)
object <- downloadSpataObject(sample_name = "UKF275T", file = TRUE)

# download, assign and save on disk in a specified directory
object <- downloadSpataObject(sample_name = "UKF275T", file = "my/path/spata_object.RDS")
}
\keyword{internal}
