% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.R
\name{downloadSpataObjects}
\alias{downloadSpataObjects}
\title{Download and save several SPATA2 objects}
\usage{
downloadSpataObjects(
  sample_names,
  files = NULL,
  folder = base::getwd(),
  overwrite = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_names}{Character vector. The sample names of the \code{SPATA2} objects
to be downloaded. Use \code{validSampleNames()} to obtain all valid input options.}

\item{files}{Character vector or NULL. Specifies the file names under which the
\code{SPATA2} objects are saved. If character, the input must be of the same length
as the input for argument \code{sample_names}. If NULL, the files are named
according to the sample name.}

\item{folder}{Character value. If character, specifies the output
folder in which the \code{SPATA2} object is saved. Defaults to the working directory.}

\item{overwrite}{Logical. Must be set to \code{TRUE} if file directories
under which downloaded files are to be saved already exist.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An invisible \code{TRUE}.
}
\description{
Main function that downloads several \code{SPATA2} objects
at the same time and saves each as an .RDS file.
}
\examples{
# downloads three objects and
# saves them as "spata_objects/UKF275T.RDS", "spata_objects/UKF313t.RDS", ... etc.

  downloadSpataObjects(
    sample_names = c("UKF275T", "UKF313T", "UKF334T"),
    folder = "spata_objects" # the folder in which to save the files
   )
}
