% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-data-frame-functions.R
\name{examineTrajectoryAssessment}
\alias{examineTrajectoryAssessment}
\title{Examine trajectory-moddeling results}
\usage{
examineTrajectoryAssessment(
  atdf,
  limits = c(0, 10),
  plot_type = "histogram",
  binwidth = 0.5,
  clrp = "milo",
  ...
)
}
\arguments{
\item{atdf}{A data.frame containing the results of trajectory-modelling. Must contain the variables:

 \describe{
  \item{\emph{variables}}{Character. The genes, gene-sets and features that have been assessed.}
  \item{\emph{pattern}}{Character. The respective trajectory pattern.}
  \item{\emph{auc}}{Numeric. The assessment value which is the residual's area under the curve.}
 }}

\item{limits}{The minimum and maximum auc-values to include. Given to
\code{ggplot2::scale_x_continuous()}.}

\item{plot_type}{One of \emph{'histogram', 'density', and 'ridgeplot'}.}

\item{...}{additional arguments given to \code{ggplot2::facet_wrap()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes the distribution of the assessment-scores
 (residuals-area-under-the-curve) of a trajectory.
}
