% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Data required for reproducible examples in SPATA2 tutorials.}
\format{
An object of class \code{list} of length 12.
}
\usage{
example_data
}
\description{
\describe{
\item{\emph{object_LMU_MCI_diet}:}{ Diet object}
\item{\emph{hgnc_symbol}:}{ Character. Gene names in HUGO format.}
\item{\emph{chromosome_name}:}{ Character. Name of the chromosome.}
\item{\emph{start_position}:}{ Integer. Starting position of the gene.}
\item{\emph{end_position}:}{ Integer. Ending positiong of the gene.}
}
}
\details{
@docType data
@usage data(example_data)
}
\keyword{datasets}
