% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{exchangeImage}
\alias{exchangeImage}
\title{Exchange HE-Image}
\usage{
exchangeImage(object, image_dir, resize = NULL, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{image_dir}{Character value. Directory to the image you want to
exchange the current image with. Should be .png.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata object.
}
\description{
Exchanges histology images and scales the coordinates
accordingly.
}
\details{
The function requires the spata object to already contain an
image. This is because images of different resolution (total number of pixels)
require the barcode spots x- and y-coordinates to be scaled. The scale
factor is computed by comparing the resolution of the old image with
the one from the image that is supposed to replace the old one.
}
