% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{exchangeImage}
\alias{exchangeImage}
\title{Deprecated}
\usage{
exchangeImage(...)
}
\description{
Deprecated in favor of \code{\link[=lastSpatialAnnotation]{lastSpatialAnnotation()}}.
}
\keyword{internal}
