% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{exclude}
\alias{exclude}
\alias{exclude,SpatialData-method}
\alias{excludeSpatialOutliers}
\alias{excludeSpatialOutliers,SPATA2-method}
\alias{excludeSpatialOutliers,SpatialData-method}
\alias{excludeTissueFragments}
\alias{excludeTissueFragments,SpatialData-method}
\title{Exclude data points}
\usage{
exclude(object, ...)

\S4method{exclude}{SpatialData}(object, barcodes, reason)

excludeSpatialOutliers(object, ...)

\S4method{excludeSpatialOutliers}{SPATA2}(object)

\S4method{excludeSpatialOutliers}{SpatialData}(object)

excludeTissueFragments(object, ...)

\S4method{excludeTissueFragments}{SpatialData}(object, fragments = "all")
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{barcodes}{Character vector. The barcodes to exclude.}

\item{reason}{Character value. The reasoning for exclusion.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Excludes data points from further integration in analysis
or plots by setting their \emph{exclude} value to \code{TRUE}. Depending on the
suffix of the function exclusion happens based on results
of previous algorithms.
}
\note{
\code{excludeTissueFragments()} requires the output of \code{\link[=identifyTissueOutline]{identifyTissueOutline()}} and
\code{excludeSpatialOutliers()} requires the output of \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}
}
\keyword{internal}
