% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e.R
\name{expandSpatialAnnotation}
\alias{expandSpatialAnnotation}
\title{Expand the borders of a Spatial Annotations}
\usage{
expandSpatialAnnotation(object, id, expand, new_id = FALSE, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The ID of the spatial annotation of interest.}

\item{expand}{Distance measure with which to expand the border. Negative
values shrink the outline.}

\item{new_id}{Character value or \code{FALSE}. If character, the resulting
spatial annotation is stored under a new ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Expands or shrinks the outer border of a spatial annotation.
}
\seealso{
\code{\link[=smoothSpatialAnnotation]{smoothSpatialAnnotation()}}, \code{\link[=shiftSpatialAnnotation]{shiftSpatialAnnotation()}}, \code{\link{SpatialAnnotation}}
}
