% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{findSDEGS}
\alias{findSDEGS}
\title{Spatial Differential Gene Expression}
\usage{
findSDEGS(
  object,
  id,
  distance = NA_integer_,
  binwidth = NA_integer_,
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  genes_rm = character(0),
  variable.features.n = 3000,
  method_de = "wilcox",
  base = 2,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. The spatial annotation of interest.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An S4 object of class \code{\link{SDEGS}} storing the results.
}
\description{
Identifies spatially differentially expressed genes (SDEGs) as
suggested by \emph{Zeng et al. (2023)}.
}
\details{
Groups the data points in spatial intervals depending on their distance
to the spatial annotation up to a specified distance. Then gene expression is
tested across the created groups as well as in 1v1 comparison to data points
outside the interval based groups (group \emph{control}).

\bold{How distance binning works:}
To bin data points according to their localisation to the spatial annotation
two of the following three parameters are required (the third one is calculated):

\itemize{
\item{\code{distance}: The distance from the border of the spatial annotation to
the \emph{horizon} in the periphery up to which the screening is conducted.
}
\item{\code{binwidth}: The width of every bin.}
\item{\code{n_bins_dist}: The number of bins that are created.}
}

These three parameters stand in the following relation to each other:

\enumerate{
\item{\code{n_bins_dist} = \code{distance} / \code{binwidth}}
\item{\code{distance} = \code{n_bins_dist} * \code{binwidth}}
\item{\code{binwidth} = \code{distance} / \code{n_bins_dist}}
}
}
\references{
Zeng, H., Huang, J., Zhou, H. et al. Integrative in situ mapping of single-cell
transcriptional states and tissue histopathology in a mouse model of Alzheimer's
disease. Nat Neurosci 26, 430-446 (2023).
}
\keyword{internal}
