% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{findSeuratClusters}
\alias{findSeuratClusters}
\title{Cluster sample via Seurat}
\usage{
findSeuratClusters(
  object,
  mtr_name = getActiveMatrixName(object, of_sample = of_sample),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}

\item{FindVariableFeatures}{A named list of arguments given to \code{Seurat::FindVariableFeatures()}, TRUE or FALSE.}

\item{RunPCA}{A named list of arguments given to \code{Seurat::RunPCA()}, TRUE or FALSE.}

\item{FindNeighbors}{A named list of arguments given to \code{Seurat::FindNeighbors()}, TRUE or FALSE.}

\item{FindClusters}{A named list of arguments given to \code{Seurat::FindClusters()}, TRUE or FALSE.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A tidy spata-data.frame containing the cluster variables.
}
\description{
Cluster sample via Seurat
}
