% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{flipAll}
\alias{flipAll}
\alias{flipCoordinates}
\alias{flipCoordsDf}
\alias{flipImage}
\alias{flipImageAnnotations}
\alias{flipSpatialTrajectories}
\title{Mirror invert image and coordinates}
\usage{
flipAll(object, axis, verbose = FALSE)

flipCoordinates(object, axis, verbose = FALSE)

flipCoordsDf(object, axis, verbose = NULL)

flipImage(object, axis, track = FALSE, verbose = FALSE)

flipImageAnnotations(object, axis, verbose = NULL)

flipSpatialTrajectories(object, axis, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{axis}{Character value. The axis around which the content is flipped.
Either \emph{'horizontal'}, \emph{'h'}, \emph{'vertical'} or \emph{'v'}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{track}{Logical value. If \code{TRUE}, changes regarding the image
justification (rotations and flipping) are tracked. Assuming that
image versions of different resolution are stored on your device with the same
justification as the primarily read image these changes in justification
can be automatically applied if the image is exchanged via \code{exchangeImage()}.}
}
\description{
The \verb{flip*()} family mirror inverts the current image
or coordinates of spatial aspects or everything. See details
for more information.
}
\details{
The \verb{flip*()} functions can be used to flip the complete \code{SPATA2}
object or to flip single aspects.

\itemize{
\item{\code{flipAll()}:}{ Flips the image as well as every single spatial aspect.
\strong{Always tracks the justification.}}
\item{\code{flipImage()}:}{ Flips only the image.}
\item{\code{flipCoordinates()}:}{ Flips the coordinates data.frame, image annotations
and spatial trajectories.}
\item{\code{flipCoordsDf()}:}{ Flips the coordinates data.frame.}
\item{\code{flipImageAnnotations()}:}{ Flips image annotations.}
\item{\code{flipSpatialTrajectories()}:}{ Flips spatial trajectories.}
}
}
