% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-handling.R
\name{flipCoords}
\alias{flipCoords}
\title{Flip coordinates}
\usage{
flipCoords(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
An updated spata-object.
}
\description{
Flips coordinate dimensions along the y-axis to align with image in case
of non matching image and coordinates.
}
