% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-handling.R
\name{flipCoords}
\alias{flipCoords}
\title{Flip coordinates}
\usage{
flipCoords(object, axis = "x", verbose = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{axis}{Character value. Either \emph{'x'} or \emph{'y'}. Denotes the axis
around which the coordinates are flipped.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata-object.
}
\description{
Flips coordinates to align with image in case
of non matching image and coordinates.
}
\note{
Make sure to flip coordinates \bold{before} adding image annotations
via \code{createImageAnnotations()} or adding spatial trajectories
via \code{createTrajectories()}!
}
