% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{flipImage}
\alias{flipImage}
\title{Flip coordinates}
\usage{
flipImage(object, axis, track = TRUE, verbose = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{axis}{Character value. Either \emph{'h'} to flip horizontally
or \emph{'v'} to flip vertically.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Function \code{flipCoordinates()} flips the coordinates of all spatial
aspects in the \code{SPATA2} object. Functions \verb{flip<spatial_aspect>()} do so for
every aspect isolated.

Flips image to align with coordinates in case
of non matching image and coordinates.
}
\seealso{
\code{\link[=flipImage]{flipImage()}}, \code{\link[=rotateCoordinates]{rotateCoordinates()}}, \code{\link[=rotateImage]{rotateImage()}}
}
