% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R, R/r.R
\name{flipImage}
\alias{flipImage}
\alias{rotateImage}
\title{Flip object image}
\usage{
flipImage(object, axis = "x")

rotateImage(object, angle)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{axis}{Character value. Either \emph{'x'} or \emph{'y'}. Denotes
the axis around which the image is flipped.}
}
\value{
An updated spata-object.
}
\description{
Flips image to align with coordinates in case
of non matching image and coordinates.
}
