% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-handling.R
\name{flipImageAndCoords}
\alias{flipImageAndCoords}
\title{Mirror invert the surface}
\usage{
flipImageAndCoords(object, axis)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{axis}{Character value. Either \emph{'x'} or \emph{'y'}. Denotes the axis
around which the iamge and the coordinates are flipped.}
}
\value{
An updated spata-object.
}
\description{
Flips both image and coordinates which results in mirror inverted plots.
}
\note{
Make sure to flip coordinates \bold{before} adding image annotations
via \code{createImageAnnotations()} or adding spatial trajectories
via \code{createTrajectories()}!
}
