% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getActiveMatrixName}
\alias{getActiveMatrixName}
\title{Deprecated}
\usage{
getActiveMatrixName(object, ...)
}
\description{
Deprecated in favor of \code{\link[=activeMatrix]{activeMatrix()}}.
}
\keyword{internal}
