% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getActiveMatrixName}
\alias{getActiveMatrixName}
\alias{getActiveExpressionMatrixName}
\title{Obtain name of currently active expression matrix}
\usage{
getActiveMatrixName(object, verbose = NULL, ...)

getActiveExpressionMatrixName(object, of_sample = NA, verbose = NULL)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
Character value.
}
\description{
Obtain name of currently active expression matrix
}
