% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getActiveMatrixName}
\alias{getActiveMatrixName}
\alias{getActiveExpressionMatrixName}
\title{Obtain name of currently active data matrix}
\usage{
getActiveMatrixName(object, ...)

getActiveExpressionMatrixName(...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character value.
}
\description{
Obtain name of currently active data matrix
}
