% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getActiveMatrixName}
\alias{getActiveMatrixName}
\alias{getActiveExpressionMatrixName}
\title{Obtain name of currently active expression matrix}
\usage{
getActiveMatrixName(object, verbose = NULL, ...)

getActiveExpressionMatrixName(object, verbose = NULL, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character value.
}
\description{
Obtain name of currently active expression matrix
}
