% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-functions.R
\name{getAutoencoderAssessment}
\alias{getAutoencoderAssessment}
\title{Obtain information about the optimal neural network set up}
\usage{
getAutoencoderAssessment(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A data.frame containing the total variance measured by \code{irlba::prcomp_irlba()} after each
combination of activations/bottlenecks.
}
\description{
Extracts the results from \code{assessAutoencoderOptions()}.
}
