% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-functions.R
\name{getAutoencoderSetUp}
\alias{getAutoencoderSetUp}
\title{Obtain information on neural network}
\usage{
getAutoencoderSetUp(object, mtr_name, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A named list.
}
\description{
Returns the argument input that was chosen to construct the
neural network that generated the matrix denoted in \code{mtr_name}.
}
