% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{getAutoencoderSetUp}
\alias{getAutoencoderSetUp}
\title{Obtain information on neural network}
\usage{
getAutoencoderSetUp(object, mtr_name, of_sample = NA)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}
}
\value{
A named list.
}
\description{
Returns the argument input that was chosen to construct the
neural network that generated the matrix denoted in \code{mtr_name}.
}
