% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-functions.R
\name{getAutoencoderSetUp}
\alias{getAutoencoderSetUp}
\title{Obtain information on neural network}
\usage{
getAutoencoderSetUp(object, mtr_name, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A named list.
}
\description{
Returns the argument input that was chosen to construct the
neural network that generated the matrix denoted in \code{mtr_name}.
}
