% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getBackgroundColor}
\alias{getBackgroundColor}
\alias{getBackgroundColor,SPATA2-method}
\alias{getBackgroundColor,SpatialData-method}
\alias{getBackgroundColor,HistoImage-method}
\title{Obtain background color}
\usage{
getBackgroundColor(object, ...)

\S4method{getBackgroundColor}{SPATA2}(object, img_name = NULL, default = "white", ...)

\S4method{getBackgroundColor}{SpatialData}(object, img_name = NULL, default = "white", ...)

\S4method{getBackgroundColor}{HistoImage}(object, default = "white")
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{default}{Color to default to if no background color is set.}
}
\value{
Character value.
}
\description{
Extracts results of \code{\link[=identifyBackgroundColor]{identifyBackgroundColor()}}.
}
\keyword{internal}
