% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getBarcodes}
\alias{getBarcodes}
\title{Obtain specific barcodes}
\usage{
getBarcodes(
  object,
  across = NULL,
  across_subset = NULL,
  of_sample = NA,
  simplify = TRUE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
Named character vector or list.
}
\description{
Returns a character vector of barcode names. See details for more.
}
\details{
If argument \code{across} is specified the output is named according
to the group membership the variable specified assigns the barcode spots to.
If simplify is set to FALSE a list is returned.

Not specifying \code{across} makes the function return an unnamed character
vector containing all barcodes.
}
\examples{
#Not run:

  # get barcodes of those barcode spots that are assigned
  # to groups 'cluster_1', 'cluster_3' and 'cluster_5' by
  # the variable 'my_cluster'

  getBarcodes(object = spata_obj,
              across = "my_cluster",
              across_subset = c("cluster_1", "cluster_3", "cluster_5"),
              simplify = TRUE)

}
