% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getBarcodesInPolygon}
\alias{getBarcodesInPolygon}
\alias{getBarcodesInPolygonList}
\title{Obtain barcodes in polygon}
\usage{
getBarcodesInPolygon(object, polygon_df, strictly = TRUE)

getBarcodesInPolygonList(object, polygon_list, strictly = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{polygon_df}{A data.frame that contains the vertices of the polygon
in form of two variables: \emph{x} and \emph{y}. Must be scaled to the dimensions
of the currently active image.}

\item{strictly}{Logical value. If \code{TRUE}, only barcode spots that are strictly
interior to the polygon are returned. If \code{FALSE}, barcodes that are
on the relative interior the polygon border or that are vertices themselves
are returned, too.}

\item{polygon_list}{A named list of data.frames with the numeric variables x and y.
Observations correspond to the vertices of the polygons that confine spatial areas.
Must contain a slot named \emph{outer} which sets the outer border of
the spatial area. Can contain multiple slots named \emph{inner} (suffixed with numbers)
that correspond to inner polygons - holes within the annotation. Like \emph{inner1},
\emph{inner2}.}
}
\value{
Character vector.
}
\description{
Extracts barcodes of barcode-spots that fall in a given
polygon. Works closely with \code{sp::point.in.polygon()}.
}
