% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getBarcodeSpotDistance}
\alias{getBarcodeSpotDistance}
\title{Deprecated}
\usage{
getBarcodeSpotDistance(...)
}
\description{
Deprecated in favor of \code{\link[=getObsDistances]{getObsDistances()}}.
}
\keyword{internal}
