% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getBarcodeSpotDistances}
\alias{getBarcodeSpotDistances}
\title{Obtain distances between barcodes}
\usage{
getBarcodeSpotDistances(
  object,
  barcodes = NULL,
  unit = "pixel",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{barcdoes}{Character vector or NULL. If character,
only input barcodes are considered.}
}
\value{
A data.frame in which each observation/row corresponds to a barcodes-spot ~
barcode-spot pair.
}
\description{
Obtain distances between barcodes
}
