% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getBarcodeSpotDistances}
\alias{getBarcodeSpotDistances}
\title{Obtain distances between barcodes}
\usage{
getBarcodeSpotDistances(
  object,
  barcodes = NULL,
  unit = "pixel",
  arrange = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \link{`validUnitsOfArea()'} for valid
input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{barcdoes}{Character vector or NULL. If character,
only input barcodes are considered.}
}
\value{
A data.frame in which each observation/row corresponds to a barcodes-spot ~
barcode-spot pair.
}
\description{
Obtain distances between barcodes
}
