% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCaptureArea}
\alias{getCaptureArea}
\title{Obtain capture area}
\usage{
getCaptureArea(object, unit = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{unit}{If character, forces the output unit of the capture area.}
}
\value{
List of two length two vectors named \emph{x} and \emph{y}. Values correspond
to the range of the capture area along the respective axis.
}
\description{
Extracts the frame in which data points are plotted
by default.
}
\seealso{
\code{\link[=setCaptureArea]{setCaptureArea()}}
}
