% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getChrRegionsDf}
\alias{getChrRegionsDf}
\title{Obtain chromosome information}
\usage{
getChrRegionsDf(object, format = "long")
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{format}{Character. If \emph{'long'} rows correspond to chromosome
arms if \emph{'wide'} rows correspond to chromosomes and information
about the respective arms is stored in separate columns.}
}
\value{
Data.frame.
}
\description{
Extracts information regarding
start, end and length of chromosomal arms.
}
