% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCnvFeatureNames}
\alias{getCnvFeatureNames}
\title{Obtain features names under which cnv-analysis results are stored.}
\usage{
getCnvFeatureNames(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector.
}
\description{
Returns a character vector of feature names referring to the
barcode-spots chromosomal gains and losses as computed by \code{runCnvAnalysis()}.
}
