% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-analysis.R
\name{getCnvFeatureNames}
\alias{getCnvFeatureNames}
\title{Obtain features names under which cnv-analysis results are stored.}
\usage{
getCnvFeatureNames(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
Character vector.
}
\description{
Returns a character vector of feature names referring to the
barcode-spots chromosomal gains and losses as computed by \code{runCnvAnalysis()}.
}
