% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-analysis.R, R/getA-G.R
\name{getCnvResults}
\alias{getCnvResults}
\title{Obtain copy-number-variations results}
\usage{
getCnvResults(object, ...)

getCnvResults(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A named list.

A named list.
}
\description{
Provides convenient access to the results of \code{runCnvAnalysis()}.

Provides convenient access to the results of \code{runCnvAnalysis()}.
}
