% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv-analysis.R
\name{getCnvResults}
\alias{getCnvResults}
\title{Obtain copy-number-variations results}
\usage{
getCnvResults(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A named list.
}
\description{
Provides convenient access to the results of \code{runCnvAnalysis()}.
}
