% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCoordsCenter}
\alias{getCoordsCenter}
\title{Obtain coordinate center}
\usage{
getCoordsCenter(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeric vector of length two.
}
\description{
Calculates and extracts center of the coordinate frame.
}
