% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/getA-G.R
\name{getSegmentDf}
\alias{getSegmentDf}
\alias{getCoordsDf}
\alias{getCoordsDf,spata2-method}
\alias{getCoordsDf,HistoImaging-method}
\title{Obtain spatial coordinates}
\usage{
getSegmentDf(object, segment_names, ...)

getCoordsDf(object, ...)

\S4method{getCoordsDf}{spata2}(object, type = "both", ...)

\S4method{getCoordsDf}{HistoImaging}(object, name = NULL, exclude = TRUE, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A data.frame containing the variables \emph{barcods, sample, x, y}
(and \emph{segmentation} in case of \code{getSegmentDf()}).
}
\description{
Obtain spatial coordinates
}
\keyword{internal}
