% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCoordsDf}
\alias{getCoordsDf}
\alias{getCoordsDf,SPATA2-method}
\alias{getCoordsDf,SpatialData-method}
\title{Obtain coordinates}
\usage{
getCoordsDf(object, ...)

\S4method{getCoordsDf}{SPATA2}(
  object,
  img_name = activeImage(object),
  exclude = TRUE,
  as_is = FALSE,
  ...
)

\S4method{getCoordsDf}{SpatialData}(
  object,
  img_name = activeImage(object),
  exclude = TRUE,
  scale = TRUE,
  wh = FALSE,
  as_is = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{The name of the image based on which the coordinates are supposed
to be aligned. If \code{NULL}, defaults to the active image.}
}
\value{
Data.frame that, among others, contains at least the
variables \emph{x_orig}, \emph{y_orig} and \emph{barcodes}
}
\description{
Extracts the coordinates data.frame of the identified
or known entities the analysis revolves around.
}
\seealso{
\code{\link[=activateImage]{activateImage()}}, \code{\link[=activeImage]{activeImage()}}
}
