% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCoordsDf}
\alias{getCoordsDf}
\alias{getCoordsDf,SPATA2-method}
\alias{getCoordsDf,SpatialData-method}
\title{Obtain coordinates}
\usage{
getCoordsDf(object, ...)

\S4method{getCoordsDf}{SPATA2}(
  object,
  variables = NULL,
  img_name = activeImage(object),
  exclude = TRUE,
  as_is = FALSE,
  ...
)

\S4method{getCoordsDf}{SpatialData}(
  object,
  img_name = activeImage(object),
  scale = TRUE,
  exclude = TRUE,
  wh = FALSE,
  as_is = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Additional arguments given to \code{\link[=joinWithVariables]{joinWithVariables()}} if argument
\code{variables} is specified.}

\item{variables}{Character or \code{NULL}. If character, specifies the
variables that are merged to the coordinates data.frame via \code{\link[=joinWithVariables]{joinWithVariables()}}.}

\item{img_name}{Only relevant if the \code{\link{SPATA2}} object contains images. If so,
specifies the name of the image to which the original coordinates are scaled.
If \code{NULL}, defaults to the active image.}

\item{exclude}{Logical. If \code{TRUE}, observations that are no longer part of the dataset
(but still exist in the coordinates data frame for potential use in spatial computations)
will be excluded and removed before returning the data frame.}

\item{as_is}{Logical. If \code{TRUE}, no processing and coordinate scaling is conducted at all.}
}
\value{
Data.frame that, among others, contains at least the
variables \emph{x_orig}, \emph{y_orig} and \emph{barcodes}
}
\description{
Extracts the coordinates data.frame of the identified
or known entities the analysis revolves around.
}
\seealso{
\code{\link[=activateImage]{activateImage()}}, \code{\link[=activeImage]{activeImage()}}
}
