% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/getA-G.R
\name{getSegmentDf}
\alias{getSegmentDf}
\alias{getCoordsDf}
\title{Obtain spatial coordinates}
\usage{
getSegmentDf(object, segment_names, ...)

getCoordsDf(object, type = "both", ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A data.frame containing the variables \emph{barcods, sample, x, y}
(and \emph{segmentation} in case of \code{getSegmentDf()}).
}
\description{
Obtain spatial coordinates
}
