% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCoordsDfSA}
\alias{getCoordsDfSA}
\alias{getCoordsDfST}
\title{Relate points to spatial annotations}
\usage{
getCoordsDfSA(
  object,
  ids = idSA(object),
  distance = "dte",
  binwidth = recBinwidth(object),
  core = TRUE,
  core0 = FALSE,
  periphery = TRUE,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  dist_unit = getDefaultUnit(object),
  coords_df = NULL,
  variables = NULL,
  format = "wide",
  verbose = NULL,
  ...
)

getCoordsDfST(
  object,
  id = idST(object),
  binwidth = recBinwidth(object),
  n_bins = NA_integer_,
  width = getTrajectoryLength(object, id = id),
  dist_unit = getDefaultUnit(object),
  outside = TRUE,
  variables = NULL,
  format = "wide",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{core}{Logical value. If \code{FALSE}, data points that lie inside the core of the
spatial annotation are removed.}

\item{core0}{Logical value. If \code{TRUE}, \emph{dist} valus of core data points are
set to 0.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{dist_unit}{Character value. Unit in which the distance is computed.
Defaults to \emph{pixel}.}

\item{coords_df}{Data.frame. If \code{NULL}, the default, the coordinates data.frame obtained
via \code{getCoordsDf()} is used. Else other data.frame of observations can be put in
relation to the spatial annotation. Requires numeric variables named \emph{x} and \emph{y} in
pixel units.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{\link[=joinWithVariables]{joinWithVariables()}}. Only used
if not empty and \code{coords_df} is \code{NULL}.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}
}
\value{
Data.frame.
}
\description{
Adds the spatial relation to a spatial
annotation to the coordinates data.frame. See details for more.
}
\details{
The coordinates data.frame as returned by \code{\link[=getCoordsDf]{getCoordsDf()}} with six
additional variables:

\itemize{
\item{\emph{dist}:}{ Numeric. The distance of the data point to the outline of the spatial annotation.}
\item{\emph{dist_unit}:}{ Character. The unit in which the distance is computed.}
\item{\emph{bins_dist}:}{ Factor. The bin the data point was assigned to based on its \emph{dist} value and the \code{binwidth}.}
\item{\emph{angle}:}{ Numeric. The angle of the data point to the center of the spatial annotation.}
\item{\emph{bins_angle}:}{ Factor. The bin the data point was assigned to based on its \emph{angle} value.}
\item{\emph{rel_loc}:}{ Character. Possible values are \emph{'core'}, if the data point lies inside the spatial annotation,
\emph{'periphery'} if the data point lies outside of the boundaries of the spatial annotation but inside
the area denoted via \code{distance} and \emph{outside}, if the data point lies beyond the screening area (it's
distance to the spatial annotation boundaries is bigger than the value denoted in \code{distance}).}
\item{\emph{id}}{ Character. The ID of the spatial annotation the data points lies closes to. (only relevant
in case of \code{length(ids) > 1})}
}
}
