% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getCountMatrix}
\alias{getCountMatrix}
\alias{getCountMatrix,SPATA2-method}
\alias{getCountMatrix,MolecularAssay-method}
\title{Obtain raw counts}
\usage{
getCountMatrix(object, ...)

\S4method{getCountMatrix}{SPATA2}(object, assay_name = activeAssay(object), ...)

\S4method{getCountMatrix}{MolecularAssay}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A matrix of unprocessed molecular counts with rownames corresponding
to the features and column names corresponding to the barcodes.
}
\description{
Extracts the unprocessed raw count matrix.
}
