% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R
\name{getDefaultInstructions}
\alias{getDefaultInstructions}
\alias{getDefault}
\alias{getDirectoryInstructions}
\title{Obtain default argument inputs}
\usage{
getDefaultInstructions(object)

getDefault(object, arg)

getDirectoryInstructions(
  object,
  to = c("cell_data_set", "seurat_object", "spata_object")
)

getDefault(object, arg)

getDefaultInstructions(object)

getDirectoryInstructions(
  object,
  to = c("cell_data_set", "seurat_object", "spata_object")
)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
S4 object containing all default argument inputs. Or the respective
default in case of \code{getDefault()}.

S4 object containing all default argument inputs. Or the respective
default in case of \code{getDefault()}.
}
\description{
Obtain default argument inputs

Obtain default argument inputs
}
