% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getDefault}
\alias{getDefault}
\alias{getDefaultInstructions}
\alias{getDirectoryInstructions}
\title{Obtain default argument inputs}
\usage{
getDefault(object, arg)

getDefaultInstructions(object)

getDirectoryInstructions(
  object,
  to = c("cell_data_set", "seurat_object", "spata_object")
)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
S4 object containing all default argument inputs. Or the respective
default in case of \code{getDefault()}.
}
\description{
Obtain default argument inputs
}
\keyword{internal}
