% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDefaultInstructions}
\alias{getDefaultInstructions}
\alias{getDirectoryInstructions}
\title{Obtain default argument inputs}
\usage{
getDefaultInstructions(object)

getDirectoryInstructions(
  object,
  to = c("cell_data_set", "seurat_object", "spata_object")
)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
S4 object containing all default argument inputs.
}
\description{
Obtain default argument inputs
}
