% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getN-Z.R
\name{getDimRedDf}
\alias{getDimRedDf}
\alias{getPcaDf}
\alias{getPcaMtr}
\alias{getTsneDf}
\alias{getUmapDf}
\title{Obtain dim red data.frame}
\usage{
getDimRedDf(object, method_dr = c("pca", "tsne", "umap"), ...)

getPcaDf(object, n_pcs = NULL, ...)

getPcaMtr(object, n_pcs = NULL, ...)

getTsneDf(object, ...)

getUmapDf(object, ...)
}
\description{
Obtain dim red data.frame
}
