% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getDimRedDf}
\alias{getDimRedDf}
\alias{getPcaDf}
\alias{getPcaMtr}
\alias{getUmapDf}
\alias{getTsneDf}
\title{Obtain dimensional reduction data}
\usage{
getDimRedDf(object, method_dr = c("pca", "tsne", "umap"), of_sample = NA)

getPcaDf(object, n_pcs = 30, of_sample = NA)

getPcaMtr(object, n_pcs = 30, of_sample = NA)

getUmapDf(object, of_sample = NA)

getTsneDf(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_dr}{Character value. The dimensional reduction method of
interest specified as a single character value. (Currently
either \emph{'pca'}, \emph{'umap'} or \emph{'tsne'}).}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be included.}
}
\value{
A data.frame that contains the unique identifiers
(keys): \emph{barcodes, sample} and:.

 \itemize{
  \item{ \code{getPcaDf()}: \emph{PC1, PC2, PC3, ...PCn}}
  \item{ \code{getTsneDf()}: \emph{tsne1, tsne2}}
  \item{ \code{getUmapDf()}: \emph{umap1, umap2}}
  }
}
\description{
Obtain dimensional reduction data
}
