% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R
\name{getExpressionMatrixNames}
\alias{getExpressionMatrixNames}
\title{Obtain names of stored expression matrices}
\usage{
getExpressionMatrixNames(object, ...)

getExpressionMatrixNames(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character vector.

Character vector.
}
\description{
Obtain names of stored expression matrices

Obtain names of stored expression matrices
}
