% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getExpressionMatrixNames}
\alias{getExpressionMatrixNames}
\title{Deprecated}
\usage{
getExpressionMatrixNames(object, assay_name = activeAssay(object), ...)
}
\description{
Deprecated in favor of \code{\link[=getProcessedMatrixNames]{getProcessedMatrixNames()}}.
}
\keyword{internal}
