% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R
\name{getFeatureDf}
\alias{getFeatureDf}
\title{Obtain feature data}
\usage{
getFeatureDf(object, of_sample = NA)

getFeatureDf(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
The feature data data.frame of the specified object and sample(s).

The feature data data.frame of the specified object and sample(s).
}
\description{
Obtain feature data

Obtain feature data
}
