% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureNames}
\alias{getFeatureNames}
\title{Obtain feature names}
\usage{
getFeatureNames(object, of_class = NULL, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_class}{Character vector. Specify the classes a feature must be of for
it's name to be returned.}
}
\value{
A named character vector of the variables in the feature data slot.
}
\description{
An easy way to obtain all features of interest along with their
class.
}
