% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getFeatureValues}
\alias{getFeatureValues}
\title{Obtain unique categorical feature values}
\usage{
getFeatureValues(object, features, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{features}{The features of interest specified as a character vector.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A vector or a named list according to the length of \code{features}.
}
\description{
Extracts the unique values of discrete features.
}
