% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneCounts}
\alias{getGeneCounts}
\title{Obtain total number of gene counts}
\usage{
getGeneCounts(object, of_sample = NA, return = "tibble")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{return}{Character value. One of \emph{'data.frame', 'tibble' or 'vector'}.
Specifies the output class.}
}
\value{
Depends on input for argument \code{return}.
}
\description{
Obtain total number of gene counts
}
