% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneDistMtr}
\alias{getGeneDistMtr}
\title{Obtain distance measurements of spatially correlated genes}
\usage{
getGeneDistMtr(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A data.frame or a distance matrix.
}
\description{
Obtain distance measurements of spatially correlated genes
}
