% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGeneFeatureNames}
\alias{getGeneFeatureNames}
\title{Obtain feature names of the gene meta data}
\usage{
getGeneFeatureNames(object, mtr_name = NULL, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\description{
Convenient way to obtain the column names of the output data.frame
of \code{getGeneMetaDf()}.
}
