% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getGeneMetaData}
\alias{getGeneMetaData}
\alias{getGeneMetaDf}
\title{Deprecated}
\usage{
getGeneMetaData(object, ...)

getGeneMetaDf(object, ...)
}
\description{
Deprecated in favor of \code{\link[=getMetaVarDf]{getMetaVarDf()}}.
}
\keyword{internal}
