% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGeneMetaData}
\alias{getGeneMetaData}
\alias{getGeneMetaDf}
\title{Obtain gene meta data}
\usage{
getGeneMetaData(object, ...)

getGeneMetaDf(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{only_df}{Logical. If set to TRUE only the data.frame is returned.
If set to FALSE (the default) the whole list is returned.}
}
\value{
A data.frame from \code{getMetaDataDf()} or a list from \code{getGeneMetaData()}.
}
\description{
Obtain gene meta data
}
