% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGeneSetDf}
\alias{getGeneSetDf}
\title{Obtain gene sets}
\usage{
getGeneSetDf(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Either a named list or a data.frame with variables \emph{ont} and \emph{gene}.
}
\description{
Extracts the gene sets (gene signatures) stored in the transcriptomic
assay.
}
\keyword{internal}
