% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGeneSetDf}
\alias{getGeneSetDf}
\title{Obtain gene sets}
\usage{
getGeneSetDf(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Either a named list or a data.frame with variables \emph{ont} and \emph{gene}.
}
\description{
Extracts the gene sets (gene signatures) stored in the transcriptomic
assay.
}
\keyword{internal}
