% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGeneSetDf}
\alias{getGeneSetDf}
\alias{getGeneSetList}
\title{Obtain gene set data.frame}
\usage{
getGeneSetDf(object)

getGeneSetList(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A data.frame.
}
\description{
Obtain gene set data.frame
}
