% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\title{Obtain variable names that group data points}
\usage{
getGroupingOptions(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character vector.
}
\description{
Extracts the names of the features of class \emph{factor} which
are valid input options for the arguments \code{grouping} and \code{across}.
}
