% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\title{Obtain variable names that group the barcode spots}
\usage{
getGroupingOptions(object, ...)

getGroupingOptions(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character vector of variables that assign the
barcode spots to groups.

Character vector of variables that assign the
barcode spots to groups.
}
\description{
Obtain variable names that group the barcode spots

Obtain variable names that group the barcode spots
}
