% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\title{Obtain variable names that group the barcode spots}
\usage{
getGroupingOptions(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
Character vector of variables that assign the
barcode spots to groups.
}
\description{
Obtain variable names that group the barcode spots
}
