% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getGroupingOptions}
\alias{getGroupingOptions}
\title{Obtain variable names that group the barcode spots}
\usage{
getGroupingOptions(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
Character vector of variables that assign the
barcode spots to groups.
}
\description{
Obtain variable names that group the barcode spots
}
