% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R
\name{getGroupNames}
\alias{getGroupNames}
\title{Obtain group names a grouping variable contains}
\usage{
getGroupNames(object, grouping, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector
}
\description{
Extracts the group names of a grouping variable.
}
\examples{
#Not run:

 # obtain all group names the variable 'my_cluster'
 # contains

 getGroupNames(object = object, grouping_variable = "my_cluster")

}
