% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getHistoImage}
\alias{getHistoImage}
\alias{getHistoImage,SPATA2-method}
\alias{getHistoImage,SpatialData-method}
\alias{getHistoImageRef}
\alias{getHistoImageRef,SPATA2-method}
\alias{getHistoImageRef,SpatialData-method}
\title{Obtain object of class \code{HistoImage}}
\usage{
getHistoImage(object, ...)

\S4method{getHistoImage}{SPATA2}(object, img_name = activeImage(object), ...)

\S4method{getHistoImage}{SpatialData}(object, img_name = activeImage(object), ...)

getHistoImageRef(object, ...)

\S4method{getHistoImageRef}{SPATA2}(object, ...)

\S4method{getHistoImageRef}{SpatialData}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\description{
Extracts the S4-containers of registered images. Note that
slot @image might be empty. Use \code{loadImage()} in that case.

\itemize{
\item{\code{getHistoImage()}:}{ Extracts object by name. By default, the active \code{HistoImage} image is returned.}
\item{\code{getHistoImageRef()}:}{ Extracts the reference \code{HistoImage} object.}
}
}
