% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getImage}
\alias{getImage}
\title{Obtain histology image}
\usage{
getImage(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
An image of class \emph{EBImage}.
}
\description{
Obtain histology image
}
