% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImage}
\alias{getImage}
\alias{getImage,SPATA2-method}
\alias{getImage,SpatialData-method}
\alias{getImage,HistoImage-method}
\title{Obtain object of class \code{Image}}
\usage{
getImage(object, ...)

\S4method{getImage}{SPATA2}(
  object,
  img_name = activeImage(object),
  xrange = NULL,
  yrange = NULL,
  expand = 0,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getImage}{SpatialData}(
  object,
  img_name = activeImage(object),
  xrange = NULL,
  yrange = NULL,
  expand = 0,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getImage}{HistoImage}(
  object,
  xrange = NULL,
  yrange = NULL,
  expand = 0,
  transform = TRUE,
  scale_fct = 1,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there are any images registered
in the \code{SPATA2} object.}
}
\value{
Object of class \code{Image}.
}
\description{
Extracts the image as an object of class \code{Image}
as specified in the package \code{EBImage}.
}
\seealso{
\code{\link[=getHistoImage]{getHistoImage()}},\code{\link[=getSpatialData]{getSpatialData()}}
}
