% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getImage}
\alias{getImage}
\alias{getImage,spata2-method}
\alias{getImage,HistoImaging-method}
\alias{getImage,HistoImage-method}
\title{Obtain \code{HistoImage} object}
\usage{
getImage(object, ...)

\S4method{getImage}{spata2}(object, xrange = NULL, yrange = NULL, expand = 0, ...)

\S4method{getImage}{HistoImaging}(
  object,
  name = NULL,
  xrange = NULL,
  yrange = NULL,
  expand = 0,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getImage}{HistoImage}(
  object,
  xrange = NULL,
  yrange = NULL,
  expand = 0,
  transform = TRUE,
  scale_fct = 1,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}
}
\description{
Extracts the image as an object of class \emph{EBImage}.
}
