% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotation}
\alias{getImageAnnotation}
\title{Obtain object of class \code{ImageAnnotation}}
\usage{
getImageAnnotation(object, id, add_image = TRUE, padding = 20)
}
\arguments{
\item{object}{A valid spata-object.}

\item{padding}{Numeric value. If bigger than 0 creates a padding around the cropped image that never
exceeds the actual limits of the image. Barcode spots are not affected.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts object of class \code{ImageAnnotaion} by
its id.
}
