% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotation}
\alias{getImageAnnotation}
\title{Obtain object of class \code{ImageAnnotation}}
\usage{
getImageAnnotation(object, id, add_image = TRUE, expand = 0, square = FALSE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{expand}{Numeric value or vector. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the xspan and the second
value is used to adjust the yspan. E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original xspan to the xspan of the image section and does not adjust
 the yspan.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts object of class \code{ImageAnnotaion} by
its id.
}
