% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImgAnns.R
\name{getImageAnnotation}
\alias{getImageAnnotation}
\alias{getImageAnnotation,spata2-method}
\alias{getImageAnnotation,HistoImaging-method}
\title{Obtain object of class \code{ImageAnnotation}}
\usage{
getImageAnnotation(object, ...)

\S4method{getImageAnnotation}{spata2}(
  object,
  id,
  img_name = NULL,
  add_barcodes = TRUE,
  strictly = FALSE,
  add_image = TRUE,
  expand = 0,
  square = FALSE
)

\S4method{getImageAnnotation}{HistoImaging}(
  object,
  id,
  img_name = NULL,
  add_barcodes = TRUE,
  strictly = FALSE,
  add_image = TRUE,
  expand = 0,
  square = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{img_name}{Character value or \code{NULL}. The image to which the outline
of the image annotation is scaled. Furthermore, if \code{add_image = TRUE}, the
image from which a section cropped to the outline is stored in slot @image
of the returned image annotation. If \code{NULL}, the active image is chosen.}

\item{add_barcodes}{Logical. If \code{TRUE}, barcodes of spots that fall into the
area of an image annotation are identified and added to slot @misc$barcodes
of the output image annotations.}

\item{strictly}{Logical. If \code{TRUE}, only barcodes of spots that are strictly interior
to the area of an image annotation are added to the output. If \code{FALSE},
barcodes of spots that are on the relative interior of the area or are
vertices of the border are added, too.}

\item{add_image}{Logical. If TRUE, the area of the histology image that
is occupied by the annotated structure is added to the \code{ImageAnnotation}
object in slot @image. Dimensions of the image can be adjusted with \code{square}
and \code{expand}.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}

\item{square}{Logical value. Most image annotations come in variable shapes and
have different horizontal and vertical diameters. Therefore, height and width of the image
section are usually not equal. If \code{square = TRUE}, the cropped section of the image that
contains the annotated structure is forced into a square: the bigger diameter of both is taken
as default. E.g. if the horizontal diameter of the image annotation is 1mm and the
vertical diameter is 1.5mm, the output image will have height and width of 1.5mm. That is,
in terms of coordinates, an x-range and a y-range of 1.5mm.

Processing of the image output depending on argument \code{expand} happens afterwards.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts object of class \code{ImageAnnotaion} by
its id.
}
\section{Expansion of cropped image sections}{


The argument \code{expand} is a versatile way, to specify how a cropped
image section is extracted. If you want the cropped image as is, specify
\code{expand = 0}. Else, there are multiple options. In general, \code{expand} takes
three kinds of values, namely percentages, distances and distance exclamations.

\itemize{
\item{Percentage:}{ A string suffixed with \emph{\%}. E.g. \code{expand = '50\%'}
adds 50\% of the distance from the center to the border of the image annotation
to the image frame.}
\item{Distance measures:}{ In pixel or European units of length. E.g. \code{expand =  list(x = '1mm')}
expands the x-axis on both sides with 1mm. \verb{expand = list(x = c('0.5mm', 1.5mm')}
expands the x-axis on the left side with 0.5mm and on the right side with 1.5mm.}
\item{Exclam distance measures:}{ Distance measure with an exclamation mark
suffix. E.g. \code{expand = '1mm!'} centers the image and forces an axis length of
1 millimeter. (Example 5) }
}

Depending on how the values are specified different parts of the image can be
expanded.

Single values, like \code{expand = 50}, are recycled: Every end of each image axis
is expanded by 50 pixel. (Example 2)

Vectors of length two, like \code{expand = c('1mm', '2mm')}, are recycled: The beginning
of each axis is expanded by 1 millimeter. The end of each axis is expanded by
2mm. (Example 3)

Named lists can be more precise. \code{expand = list(x = c('1mm', '0.5mm'), y = c('0.25mm', '1mm'))}.
Applies the vectors to expand the corresponding axis. (Example 4)

Using exclam input the side of the axis must not be specified as the
axis is fixed as a whole. E.g \code{expand = list(x = '1mm!', y = '2mm!')} results
in the same output as \verb{expand = list(x = c('1mm!', '1mm!'), y = c('2mm!', '2mm!')}.
}

