% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationAreaDf}
\alias{getImageAnnotationAreaDf}
\title{Obtain image annotation area data.frame}
\usage{
getImageAnnotationAreaDf(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  add_tags = FALSE,
  sep = " & ",
  last = " & "
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{test}{Character value. Specifies how the input of \code{tags} is used to
subset the image annotations. One of \emph{'any'}, \emph{'all'} or \emph{'identical'}.
See details for more information.}
}
\value{
A data.frame that contains the grouping variables \emph{id} and \emph{tags}
and the numeric variables \emph{x} and \emph{y}. The returned data.frame can contain
the spatial extent of more than just
one annotated structure, depending on the input of arguments \code{ids}, \code{tags}
and \code{test}. The variable \emph{ids} of the returned data.frame is used to
uniquely mark the belonging of each x- and y-coordinate.
}
\description{
Extracts the coordinates of the polygon that was drawn to
annotate structures in the histology image in a data.frame.
}
\details{
How to use arguments \code{tags} and \code{test} to specify
the image annotations of interest: Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are three options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags and must not be tagged with anything else.
}

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.
}
\note{
The variables \emph{x} and \emph{y} correspond to the coordinates
with which the annotated structure was denoted in \code{annotateImage()}. These
coordinates are not to be confused with coordinates of barcode spots that might
fall in to the area of the polygon! To obtain a data.frame of barcode spots that fall in to
the spatial extent of the annotated structure along with their coordinates
use \code{getImageAnnotationBarcodes()}.
}
