% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageAnnotationBarcodes}
\alias{getImageAnnotationBarcodes}
\title{Obtain barcodes by image annotation tag}
\usage{
getImageAnnotationBarcodes(object, ids = NULL, tags = NULL, test = "any")
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{test}{Character value. Specifies how the input of \code{tags} is used to
subset the image annotations. One of \emph{'any'}, \emph{'all'} or \emph{'identical'}.
See details for more information.}
}
\value{
Character vector.
}
\description{
Extracts the barcodes that are covered by the extent of the
annotated structures of interest.
}
