% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationCenter}
\alias{getImageAnnotationCenter}
\alias{getImageAnnotationCenter,spata2-method}
\alias{getImageAnnotationCenter,ImageAnnotation-method}
\title{Obtain center information an image annotation}
\usage{
getImageAnnotationCenter(object, ...)

\S4method{getImageAnnotationCenter}{spata2}(object, id)

\S4method{getImageAnnotationCenter}{ImageAnnotation}(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{id}{Character value. The ID of the image annotation of interest.}
}
\value{
Character vector.
}
\description{
\code{getImageAnnotationCenter()} computes the
x- and y- coordinates of the most central points within the
image annotation polygon. \code{getImageAnnotationBcsp()} returns
a data.frame that contains one row, namely the barcode spot
that lies closest to the most central point.
}
