% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationIds}
\alias{getImageAnnotationIds}
\title{Obtain image annotations ids}
\usage{
getImageAnnotationIds(object, tags = NULL, test = "any")
}
\arguments{
\item{object}{A valid spata-object.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}
}
\value{
Character vector.
}
\description{
Extracts image annotation IDs as a character vector.
}
