% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationIds}
\alias{getImageAnnotationIds}
\title{Obtain image annotations ids}
\usage{
getImageAnnotationIds(object, tags = NULL, test = "any")
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{test}{Character value. Specifies how the input of \code{tags} is used to
subset the image annotations. One of \emph{'any'}, \emph{'all'} or \emph{'identical'}.
See details for more information.}
}
\value{
Character vector.
}
\description{
Extracts image annotation IDs as a character vector.
}
