% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImageAnnotationIds}
\alias{getImageAnnotationIds}
\title{Deprecated}
\usage{
getImageAnnotationIds(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnIds]{getSpatAnnIds()}}.
}
\keyword{internal}
