% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotations}
\alias{getImageAnnotations}
\title{Obtain list of \code{ImageAnnotation}-objects}
\usage{
getImageAnnotations(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  add_barcodes = TRUE,
  add_image = TRUE,
  expand = 0,
  square = FALSE,
  flatten = FALSE,
  check = FALSE
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{add_image}{Logical. If TRUE, the area of the histology image that
is occupied by the annotated structure is added to the \code{ImageAnnotation}
object in slot @image.}

\item{expand}{Numeric value. Specifies the extent to which the cropped image
section is expanded. 0 as input leads to no expansion at all. Values lower than 1
are considered in relation to the x- and yrange - e.g. \code{expand} = 0.5 expands the plot to 150%. Values
bigger than 1 are considered as absolute - e.g. \code{expand} = 20 adds 20pixels on
both sides of each axis respectively. If length of \code{expand} is 2 the first value
is used to expand the x-axis, the second value is used to expand the y-axis.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts a list of objects of class \code{ImageAnnotaion}.
}
\details{
How to use arguments \code{tags} and \code{test} to specify
the image annotations of interest: Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are three options:

1. Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.

2. Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags.

3. Argument \code{test} set to \emph{'identical'} or \emph{3}: To be indluded, an image annotation
must be tagged with all of the input tags and must not be tagged with anything else.

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.
}
