% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotations}
\alias{getImageAnnotations}
\title{Obtain list of \code{ImageAnnotation}-objects}
\usage{
getImageAnnotations(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  add_barcodes = TRUE,
  add_image = TRUE,
  expand = 0,
  square = FALSE,
  flatten = FALSE,
  check = FALSE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{test}{Character value. Specifies how the input of \code{tags} is used to
subset the image annotations. One of \emph{'any'}, \emph{'all'} or \emph{'identical'}.
See details for more information.}

\item{add_image}{Logical. If TRUE, the area of the histology image that
is occupied by the annotated structure is added to the \code{ImageAnnotation}
object in slot @image.}

\item{expand}{Vector of length one or two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the x-span and the second
value is used to adjust the y-span E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original x-span to the x-span of the image section and does not adjust
the y-span.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. X- and yspan of the square are equal to the span
that is the biggest. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts a list of objects of class \code{ImageAnnotaion}.
}
\details{
How to use arguments \code{tags} and \code{test} to specify
the image annotations of interest: Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are three options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags and must not be tagged with anything else.
}

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.
}
