% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageAnnotations}
\alias{getImageAnnotations}
\title{Obtain list of \code{ImageAnnotation}-objects}
\usage{
getImageAnnotations(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  add_barcodes = TRUE,
  strictly = FALSE,
  add_image = TRUE,
  expand = 0,
  square = FALSE,
  flatten = FALSE,
  check = FALSE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{tags, test}{Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are multiple options:
\enumerate{
\item Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.
\item Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags. Can contain tags that are not specified.
\item Argument \code{test} set to \emph{'identical'} or \emph{3}: To be included, an image annotation
must be tagged with all of the input tags. Can not be tagged with anything else.
\item Argument \code{test} set to \emph{not_identical} or \emph{4}: To be included, an image
annotation must \strong{not} be tagged with the combination of input tags.
\item Argument \code{test} set to \emph{'none'} or \emph{5}: To be included, an image annotation
must \strong{not} contain any of the input tags.
}

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.}

\item{add_barcodes}{Logical. If \code{TRUE}, barcodes of spots that fall into the
area of an image annotation are identified and added to slot @misc$barcodes
of the output image annotations.}

\item{strictly}{Logical. If \code{TRUE}, only barcodes of spots that are strictly interior
to the area of an image annotation are added to the output. If \code{FALSE},
barcodes of spots that are on the relative interior of the area or are
vertices of the border are added, too.}

\item{add_image}{Logical. If TRUE, the area of the histology image that
is occupied by the annotated structure is added to the \code{ImageAnnotation}
object in slot @image. Dimensions of the image can be adjusted with \code{square}
and \code{expand}.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See details
for more information.}

\item{square}{Logical value. Most image annotations come in variable shapes and
have different horizontal and vertical diameters. Therefore, height and width of the image
section are usually not equal. If \code{square = TRUE}, the cropped section of the image that
contains the annotated structure is forced into a square: the bigger diameter of both is taken
as default. E.g. if the horizontal diameter of the image annotation is 1mm and the
vertical diameter is 1.5mm, the output image will have height and width of 1.5mm. That is,
in terms of coordinates, an x-range and a y-range of 1.5mm.

Processing of the image output depending on argument \code{expand} happens afterwards.}
}
\value{
An object of class \code{ImageAnnotation}.
}
\description{
Extracts a list of objects of class \code{ImageAnnotaion}.
}
\note{
To test how the extracted image section looks like depending
on input for argument \code{square} and \code{expand} use
\code{plotImageAnnotations(..., encircle = FALSE)}.
}
