% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{getImageAnnotationScreeningDf}
\alias{getImageAnnotationScreeningDf}
\title{Obtain IAS screending data.frame}
\usage{
getImageAnnotationScreeningDf(
  object,
  id,
  buffer,
  n_bins_circle,
  n_bins_angle = 12,
  variables = NULL,
  normalize = TRUE,
  smooth = FALSE,
  smooth_span = 0.2,
  remove_circle_bins = FALSE,
  remove_angle_bins = FALSE,
  rename_angle_bins = FALSE,
  drop = TRUE
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}.}

\item{n_bins_circle}{Numeric value (or vector of length two, see details
for more). Specifies how many times the area is buffere with the value
denoted in \code{buffer}. The combination of both results in area that is
screened.}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.}

\item{variables}{Character vector. All numeric variables, meaning genes,
gene-sets and numeric features that are supposed to be included in
the screening process.}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{remove_circle_bins, remove_angle_bins, remame_angle_bins}{Logical values.
Given to the corresponding arguments of \code{bin_by_area()}
and \code{bin_by_angle()}.}
}
\description{
Bins and annotates barcode-spots in the same way that
\code{imageAnnotationScreening()} does.
}
