% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{getImageAnnotationTags}
\alias{getImageAnnotationTags}
\title{Obtain image annotations tags}
\usage{
getImageAnnotationTags(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Character vector.
}
\description{
Extracts all unique tags with which image annotations
have been tagged.
}
