% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageCenter}
\alias{getImageCenter}
\alias{getImageCenter,spata2-method}
\alias{getImageCenter,HistoImaging-method}
\alias{getImageCenter,HistoImage-method}
\title{Obtain image center}
\usage{
getImageCenter(object, ...)

\S4method{getImageCenter}{spata2}(object)

\S4method{getImageCenter}{HistoImaging}(object, img_name = NULL)

\S4method{getImageCenter}{HistoImage}(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
Numeric vector of length two.
}
\description{
Computes and extracts center of the image frame.
}
