% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageCenter}
\alias{getImageCenter}
\title{Obtain image center}
\usage{
getImageCenter(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
Numeric vector of length two.
}
\description{
Computes and extracts center of the image frame.
}
