% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageDf}
\alias{getImageDf}
\alias{getImageDf,SPATA2-method}
\alias{getImageDf,SpatialData-method}
\alias{getImageDf,HistoImage-method}
\alias{getImageDf,SpatialAnnotation-method}
\alias{getImageDf,Image-method}
\title{Obtain image as a data.frame}
\usage{
getImageDf(object, ...)

\S4method{getImageDf}{SPATA2}(
  object,
  img_name = activeImage(object),
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getImageDf}{SpatialData}(
  object,
  img_name = activeImage(object),
  transform = TRUE,
  scale_fct = 1
)

\S4method{getImageDf}{HistoImage}(object, transform = TRUE, scale_fct = 1)

\S4method{getImageDf}{SpatialAnnotation}(object, rescale_axes = TRUE, scale_fct = 1)

\S4method{getImageDf}{Image}(object, scale_fct = 1)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there are any images registered
in the \code{SPATA2} object.}

\item{rescale_axes}{Logical value. If \code{TRUE}, rescales the pixel positions
(height/width) to the position in the original image.

The image annotation contains a crop of the original image that only shows
the area of the image annotation (plus \code{expand}, see \code{\link[=getSpatialAnnotation]{getSpatialAnnotation()}}).}
}
\value{
Data.frame with three variables.

\itemize{
\item{\emph{width}:}{ Numeric. Width value of the pixel (position on horizontal axis).}
\item{\emph{height}:}{ Numeric. Height value of the pixel (position on vertical axis).}
\item{\emph{color}:}{ Character. HEX-code of the color the pixel carries.}
}
}
\description{
Extracts a data.frame in which each row corresponds
to a pixel in the image. (Faster than \code{getPixelDf()}, though without
any further options.)
}
\seealso{
\code{\link[=getPixelDf]{getPixelDf()}}
}
