% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageDims}
\alias{getImageDims}
\alias{getImageRange}
\title{Obtain image dimensions/ranges}
\usage{
getImageDims(object, ...)

getImageRange(object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Similar output, different data structure:

\itemize{
\item{\code{getImageDims()}:}{ Vector of length three: image width, image height, image depth}
\item{\code{getImageRange()}:}{ Named list, names are \emph{x} and \emph{y}. Each slot contains a
vector of length two that describes the range of the x- and y-axis. Used for intersection
between histology image and scatterplots.}
}
}
\description{
Extracts information regarding the image.

\itemize{
\item{\code{getImageDims()}:}{ Extracts dimensions of the image, namely width, height and depth.}
\item{\code{getImageRange()}:} Extracts range of the image axis.
}
}
\details{
In case of confusion due to overlapping naming conventions: X-axis,
x and x-range in terms of coordinates, corresponds to image width in terms of
image analysis. Y-axis, y  and y-range, in terms of coordinates, refers to
image-height in terms of image analysis. \code{SPATA2} primarily uses coordinates
naming convention.
}
