% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageDims}
\alias{getImageDims}
\alias{getImageDims,SPATA2-method}
\alias{getImageDims,SpatialData-method}
\alias{getImageDims,HistoImage-method}
\alias{getImageRange}
\alias{getImageRange,SPATA2-method}
\alias{getImageRange,SpatialData-method}
\alias{getImageRange,HistoImage-method}
\title{Obtain image dimensions/ranges}
\usage{
getImageDims(object, ...)

\S4method{getImageDims}{SPATA2}(object, img_name = activeImage(object), ...)

\S4method{getImageDims}{SpatialData}(object, img_name = activeImage(object), ...)

\S4method{getImageDims}{HistoImage}(object, ...)

getImageRange(object, ...)

\S4method{getImageRange}{SPATA2}(object, img_name = activeImage(object), ...)

\S4method{getImageRange}{SpatialData}(object, img_name = activeImage(object))

\S4method{getImageRange}{HistoImage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
Similar output, different data structure:

\itemize{
\item{\code{getImageDims()}:}{ Vector of length three: image width, image height, image depth}
\item{\code{getImageRange()}:}{ Named list, names are \emph{x} and \emph{y}. Each slot contains a
vector of length two that describes the range of the x- and y-axis.}
}
}
\description{
Extracts information regarding the image.

\itemize{
\item{\code{getImageDims()}:}{ Extracts dimensions of the image, namely width, height and depth.}
\item{\code{getImageRange()}:} Extracts range of the image x- and y-axis.
}
}
\details{
In case of confusion due to overlapping naming conventions: X-axis,
x and x-range in terms of coordinates, corresponds to image width in terms of
image analysis. Y-axis, y  and y-range, in terms of coordinates, refers to
image-height in terms of image analysis. \code{SPATA2} primarily uses coordinates
naming convention.
}
