% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageDir}
\alias{getImageDir}
\title{Obtain directory to image}
\usage{
getImageDir(object, img_name = activeImage(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
Character value.
}
\description{
Extracts the directory from where the image was loaded when
it was registered via \code{\link[=registerImage]{registerImage()}}.
}
