% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getI-M.R
\name{getImageDirHighres}
\alias{getImageDirHighres}
\alias{getImageDirLowres}
\title{Obtain image directories}
\usage{
getImageDirHighres(object, check = TRUE)

getImageDirLowres(object, check = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{check}{Logical value. If set to TRUE the input directory is checked
for validity and it is checked if the file actually exists.}
}
\value{
Character value.
}
\description{
Extracts image directories.
}
