% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getImageDirLowres}
\alias{getImageDirLowres}
\alias{getImageDirHighres}
\title{Obtain image directories}
\usage{
getImageDirLowres(object, check = TRUE)

getImageDirHighres(object, check = TRUE)
}
\arguments{
\item{object}{A valid spata-object.}

\item{check}{Logical value. If set to TRUE the input directory is checked
for validity and it is checked if the file actually exists.}
}
\value{
Character value.
}
\description{
Extracts image directories.
}
