% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageDir}
\alias{getImageDir}
\alias{getImageDirDefault}
\alias{getImageDirectories}
\alias{getImageDirHighres}
\alias{getImageDirLowres}
\title{Obtain image directories}
\usage{
getImageDir(object, name)

getImageDirDefault(object, fdb_fn = "warning", check = FALSE, ...)

getImageDirectories(object)

getImageDirHighres(object, fdb_fn = "warning", check = FALSE, ...)

getImageDirLowres(object, fdb_fn = "warning", check = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{name}{Character value. The name of the image of interest. Should be one
of Get}

\item{check}{Logical value. If \code{TRUE}, it is checked if the file actually exists.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector.
}
\description{
Extracts image directories known to the \code{SPATA2} object.
}
\details{
\code{getImageDirectories()} returns all image directories known to
the \code{SPATA2} object. \code{getImageDirLowres()}, \code{getImageDirHighres()} and
\code{getImageDirDefault()} return the directories of the respective slot of
the \code{HistologyImaging} object. \code{getImageDir()} extracts specific directories
that were set with \code{setImageDir()} by name.
}
\seealso{
\code{\link[=setImageDir]{setImageDir()}} to set specific image directories. \code{\link[=loadImage]{loadImage()}},
\code{\link[=loadImageHighres]{loadImageHighres()}}, \code{\link[=loadImageLowres]{loadImageLowres()}}, \code{\link[=loadImageDefault]{loadImageDefault()}} to
exchange images.
}
