% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImageInfo}
\alias{getImageInfo}
\title{Obatain image information}
\usage{
getImageInfo(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
List that contains information of slots @image_info and @justification
of the \code{HistoImaging} object.
}
\description{
Extracts a list of information about the currently set
image.
}
