% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageNames}
\alias{getImageNames}
\alias{getImageNames,SPATA2-method}
\alias{getImageNames,HistoImaging-method}
\title{Obtain names of registered \code{HistoImage} objects}
\usage{
getImageNames(object, ...)

\S4method{getImageNames}{SPATA2}(object, ref = TRUE, ...)

\S4method{getImageNames}{HistoImaging}(object, ref = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{ref}{Logical value. If \code{FALSE}, name of the reference image is not
included.}
}
\value{
Character vector.
}
\description{
Extracts the names of the \code{HistoImage} objects currently
registered in the object.
}
