% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageObject}
\alias{getImageObject}
\title{Obtain object of class \code{HistologyImage}}
\usage{
getImageObject(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Object of class \code{HistologyImage}
}
\description{
Extracts the S4-object. Do not confuse with \code{getImage()}
}
