% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageOrigin}
\alias{getImageOrigin}
\title{Obtain image origin}
\usage{
getImageOrigin(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Either a directory or \emph{Global.Env.} if it was read in from
the global environment.
}
\description{
Extracts the origin of the image that is currently set.
}
