% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageRaster}
\alias{getImageRaster}
\alias{getImageRaster,SPATA2-method}
\alias{getImageRaster,HistoImage-method}
\alias{getImageRasterInfo}
\title{Obtain image raster-(information)}
\usage{
getImageRaster(object, ...)

\S4method{getImageRaster}{SPATA2}(
  object,
  img_name = activeImage(object),
  transform = TRUE,
  xrange = NULL,
  yrange = NULL,
  expand = 0
)

\S4method{getImageRaster}{HistoImage}(object, xrange = NULL, yrange = NULL, expand = 0)

getImageRasterInfo(object, xrange = NULL, yrange = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there are any images registered
in the \code{SPATA2} object.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}
}
\description{
Obtain image raster-(information)
}
