% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageSectionsByBarcode}
\alias{getImageSectionsByBarcode}
\title{Obtain image sections by barcode spot}
\usage{
getImageSectionsByBarcode(object, barcodes = NULL, expand = 0, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{barcodes}{Characte vector or NULL. If character, subsets the barcodes
of interest. If NULL, all barcodes are considered.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A named list. Each slot is named after one barcode. The content is
another list that contains the barcode specific image section as well
as the x- and y-ranges that were used to crop the section.
}
\description{
Cuts out the area of the image that is covered by each barcode.
}
