% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getImageTransformations}
\alias{getImageTransformations}
\alias{getImageTransformations,SPATA2-method}
\alias{getImageTransformations,SpatialData-method}
\alias{getImageTransformations,HistoImage-method}
\title{Obtain image transformation instructions}
\usage{
getImageTransformations(object, ...)

\S4method{getImageTransformations}{SPATA2}(object, img_name = activeImage(object), ...)

\S4method{getImageTransformations}{SpatialData}(object, img_name = activeImage(object), ...)

\S4method{getImageTransformations}{SpatialData}(object, img_name = activeImage(object), ...)

\S4method{getImageTransformations}{HistoImage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
A list with the following structure:
\itemize{
\item{\emph{angle}:}{ Numeric value that ranges from 0-359. Indicates the angle in degrees
b y which the image needs to be rotated in \strong{clockwise} direction. }
\item{\emph{flip}:}{ List of two logical values named \emph{horizontal} and \emph{vertical} indicating
whether the image is to be flipped along the respective axis.}
\item{\emph{stretch}:}{ List of two numeric values named \emph{horizontal} and \emph{vertical} indicating
whether (or how much) the image is to be stretched along the respective axis.}
\item{\emph{translate}:}{ Vector of two numeric values named \emph{horizontal} and \emph{vertical}. Indicate
the number of pixels the image needs to be translated. Positive values shift the image
\strong{downwards} or to the right, respectively. Negative values shift the image \strong{upwards}
or to the left, respectively. }
}
}
\description{
Extracts a list of instructions for required
image transformations to ensure alignment of the image with
the \code{\link[=concept_images]{reference image}}.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

plotImage(object)

object <- alignImage(object, img_name = "lowres", angle = 90)

# note that $angle contains instructions to rotate the image to 90°
getImageTransformations(object, img_name = "lowres")

plotImage(object)

object <- resetImageTransformations(object, img_name = "lowres")

getImageTransformations(object, img_name = "lowres")

plotImage(object)

}
