% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getImgAnnCenter}
\alias{getImgAnnCenter}
\title{Deprecated}
\usage{
getImgAnnCenter(...)
}
\description{
Deprecated in favor of \code{\link[=getSpatAnnCenter]{getSpatAnnCenter()}}.
}
\keyword{internal}
