% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/getH-M.R
\name{getImageAnnotationCenter}
\alias{getImageAnnotationCenter}
\alias{getImgAnnCenter}
\alias{getImgAnnCenter,spata2-method}
\alias{getImgAnnCenter,ImageAnnotation-method}
\alias{getImgAnnCenters}
\alias{getImgAnnCenters,spata2-method}
\alias{getImgAnnCenters,ImageAnnotation-method}
\title{Obtain center of an image annotation}
\usage{
getImageAnnotationCenter(...)

getImgAnnCenter(object, ...)

\S4method{getImgAnnCenter}{spata2}(object, id)

\S4method{getImgAnnCenter}{ImageAnnotation}(object)

getImgAnnCenters(object, ...)

\S4method{getImgAnnCenters}{spata2}(object, id, outer = TRUE, inner = TRUE)

\S4method{getImgAnnCenters}{ImageAnnotation}(object, outer = TRUE, inner = TRUE)
}
\arguments{
\item{...}{Used to absorb deprecated arguments or functions.}

\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{outer}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the outer border of the image annotation
is not included in the output.}

\item{inner}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the inner borders of the image annotation
are not included in the output.}
}
\value{
Numeric vector of length two or a list of these. Values are named \emph{x} and \emph{y}.
}
\description{
\code{getImgAnnCenter()} computes the
x- and y- coordinates of the center of the outer border, returns
a numeric vector of length two. \code{getImgAnnCenters()} computes the center of the outer
and every inner border and returns a list of numeric vectors of length two.
}
