% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImgAnns.R
\name{getImgAnnTags}
\alias{getImgAnnTags}
\alias{getImgAnnTags,spata2-method}
\alias{getImgAnnTags,HistoImaging-method}
\title{Obtain image annotations tags}
\usage{
getImgAnnTags(object, ...)

\S4method{getImgAnnTags}{spata2}(object)

\S4method{getImgAnnTags}{HistoImaging}(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector.
}
\description{
Extracts all unique tags with which image annotations
have been tagged.
}
