% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R
\name{getExpressionMatrix}
\alias{getExpressionMatrix}
\alias{getMatrix}
\alias{getMatrices}
\title{Obtain count and expression matrix}
\usage{
getExpressionMatrix(object, mtr_name = NULL, verbose = FALSE, ...)

getMatrix(object, mtr_name = NULL, verbose = NULL, ...)

getMatrices(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}
}
\value{
The matrix of the specified object. A list of all matrices
in case of \code{getMatrices()}.
}
\description{
Obtain count and expression matrix
}
