% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R
\name{getCountMatrix}
\alias{getCountMatrix}
\alias{getExpressionMatrix}
\alias{getMatrix}
\title{Obtain count and expression matrix}
\usage{
getCountMatrix(object, ...)

getExpressionMatrix(object, mtr_name = NULL, verbose = FALSE, ...)

getMatrix(object, mtr_name = NULL, verbose = NULL, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}
}
\value{
The matrix of the specified object and sample(s).
}
\description{
Obtain count and expression matrix
}
