% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getA-G.R, R/getH-M.R
\name{getMatrix}
\alias{getMatrix}
\alias{getExpressionMatrix}
\alias{getCountMatrix}
\alias{getMatrices}
\title{Obtain count and expression matrix}
\usage{
getMatrix(object, mtr_name = NULL, verbose = NULL, ...)

getExpressionMatrix(object, mtr_name = NULL, verbose = FALSE, ...)

getCountMatrix(object, ...)

getCountMatrix(object, ...)

getExpressionMatrix(object, mtr_name = NULL, verbose = FALSE, ...)

getMatrix(object, mtr_name = NULL, verbose = NULL, ...)

getMatrices(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{mtr_name}{Character value. The name of the matrix of interest.}
}
\value{
The active expression or count matrix of the specified object and sample(s).

The matrix of the specified object. A list of all matrices
in case of \code{getMatrices()}.
}
\description{
Obtain count and expression matrix

Obtain count and expression matrix
}
