% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getMatrix}
\alias{getMatrix}
\alias{getMatrix,SPATA2-method}
\alias{getMatrix,MolecularAssay-method}
\alias{getMatrices}
\alias{getMatrices,SPATA2-method}
\alias{getMatrices,MolecularAssay-method}
\title{Obtain a data matrix}
\usage{
getMatrix(object, mtr_name = NULL, ...)

\S4method{getMatrix}{SPATA2}(object, mtr_name = NULL, assay_name = activeAssay(object))

\S4method{getMatrix}{MolecularAssay}(object, mtr_name = NULL)

getMatrices(object, ...)

\S4method{getMatrices}{SPATA2}(object, assay_name = activeAssay(object), ...)

\S4method{getMatrices}{MolecularAssay}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{Character value. The name of the matrix to extract. If
\code{NULL}, defaults to the active matrix of the provided or specified \link[=MolecularAssay]{assay}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A numeric matrix with rownames corresponding to molecule names and
column names corresponding to barcodes. Or a named list of such in
case of \code{getMatrices()}.
}
\description{
Extracts a matrix or multiple matrices of molecular data.
}
\seealso{
\code{\link[=getCountMatrix]{getCountMatrix()}} to extract raw counts without additional argument
specification required, \code{\link[=getMatrixNames]{getMatrixNames()}} for a vector of all matrix names of
an assay, \code{\link[=getProcessedMatrixNames]{getProcessedMatrixNames()}} for a vector of all processed matrix of an assay.
}
