% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-G.R, R/getH-M.R, R/getN-Z.R
\name{getGenes}
\alias{getGenes}
\alias{getMetabolites}
\alias{getMolecules}
\alias{getProteins}
\title{Obtain molecule names}
\usage{
getGenes(object, signatures = NULL, simplify = TRUE, ...)

getMetabolites(object, signatures = NULL, simplify = TRUE, ...)

getMolecules(
  object,
  signatures = NULL,
  simplify = TRUE,
  assay_name = activeAssay(object)
)

getProteins(object, signatures = NULL, simplify = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{signatures}{Character vector or \code{NULL}. If character, specifies the name of
signatures to filter the molecules for.}

\item{simplify}{Only relevant if \code{signatures} is not \code{NULL}. If \code{TRUE}, all molecule
names are merged in to one character vector of unique molecule names. If \code{FALSE},
a named list of character vectors is returned (names correspond to signatures).}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
A character vector or a named list of character vectors.
}
\description{
Retrieves the names of molecules present in the given object,
optionally filtered by a specific \link[=concept_molecular_signatures]{signature}.
}
\details{
These functions retrieve molecule names from the provided object.

\itemize{
\item{\code{getMolecules()}}{: The molecules from the assay specified in \code{assay_name}.}
\item{\code{getGenes()}}{:  The molecules from the assay with @modality = 'gene' (\code{assay_name = 'gene'}).}
\item{\code{getMetabolites()}}{: The molecules from the assay with @modality = 'metabolite' (\code{assay_name = 'metabolite'}).}
\item{\code{getProteins()}}{: The molecules from the assay with @modality = 'protein' (\code{assay_name = 'protein'}).}
}

If 'signature' is \code{NULL}, it returns all molecules from the respective assay in the object.
}
\note{
Molecule names are picked from the currently active matrix of their assay!
If a processed matrix is active and it does not contain some molecules the raw count matrix
}
\examples{

library(SPATA2)
object <- example_data$object_UKFT269_diet

getAssayNames(object)

## no subsetting by signatures
# opt 1
genes <- getMolecules(object)

str(genes)

# opt 2
genes <- getMolecules(object, assay_name = "gene")

str(genes)

# opt 3
genes <- getGenes(object)

# opt 4 - fails cause no 'protein' assay

proteins <- getProteins(object)

## using signatures argument
# character vector of molecules of specific signatures (simplify defaults to TRUE)
genes_hyp <- getGenes(object, signatures = "HM_HYPOXIA")

genes_hyp

# list of gene sets
tcr_gs <- c("RCTM_TCR_SIGNALING", "RCTM_DOWNSTREAM_SIGNALING")

# simplify = TRUE (default) merges the output to a character vector of unique names
tcr_genes_vec <- getGenes(object, signatures = tcr_gs, simplify = TRUE)

str(tcr_genes_vec)

# simplify = FALSE returns a list
tcr_genes_lst <- getGenes(object, signatures = tcr_gs, simplify = FALSE)

str(tcr_genes_lst)

## DEPRECATED: of_gene_sets
tcr_genes <- getGenes(object, of_gene_sets = tcr_gs, simplify = FALSE)


}
\seealso{
\code{\link[=getMatrix]{getMatrix()}}, \code{\link[=getSignatures]{getSignatures()}}
}
