% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getPixelDf}
\alias{getPixelDf}
\alias{getPixelDf,spata2-method}
\alias{getPixelDf,HistoImaging-method}
\alias{getPixelDf,HistoImage-method}
\alias{getPixelDf,Image-method}
\title{Obtain pixel data.frame}
\usage{
getPixelDf(object, ...)

\S4method{getPixelDf}{spata2}(
  object,
  colors = FALSE,
  tissue = FALSE,
  xrange = NULL,
  yrange = NULL
)

\S4method{getPixelDf}{HistoImaging}(
  object,
  name = NULL,
  colors = FALSE,
  hex_code = FALSE,
  tissue = FALSE,
  use_greyscale = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  xrange = NULL,
  yrange = NULL,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getPixelDf}{HistoImage}(
  object,
  colors = FALSE,
  hex_code = FALSE,
  tissue = FALSE,
  use_greyscale = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  xrange = NULL,
  yrange = NULL,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getPixelDf}{Image}(
  object,
  colors = FALSE,
  hex_code = FALSE,
  tissue = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  use_greyscale = FALSE,
  dbscan_eps = 1,
  dbscan_minPts = 3,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{colors}{Logical value. If \code{TRUE}, adds all colors from the image
as variables named \emph{col1}-\emph{col.n} where n is the number of colors.}

\item{tissue}{Logical value. If \code{TRUE}, adds a variable called \emph{pxl_group}
that indicates whether the pixel is placed on a contiguous tissue section, on
artefact tissue fragments or on background.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}

\item{wh}{Logical value. If \code{TRUE}, a \emph{width} and a \emph{height} column are
added containing information about the position of each pixel in image
dimensions, where \emph{width} = \emph{x} and  \emph{height} = \code{range(y)[2]} - \emph{y} + \code{range(y)[1]}.}
}
\value{
Data.frame with \code{nrow()} equal to the number of pixels.
}
\description{
Extracts a data.frame in which each row corresponds
to a pixel in the current image with x- and y-coordinates.
}
