% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getPixelDf}
\alias{getPixelDf}
\alias{getPixelDf,spata2-method}
\alias{getPixelDf,HistoImaging-method}
\alias{getPixelDf,HistoImage-method}
\alias{getPixelDf,Image-method}
\title{Obtain pixel data.frame}
\usage{
getPixelDf(object, ...)

\S4method{getPixelDf}{spata2}(
  object,
  img_name = NULL,
  colors = FALSE,
  hex_code = FALSE,
  content = FALSE,
  transform = TRUE,
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1
)

\S4method{getPixelDf}{HistoImaging}(
  object,
  img_name = NULL,
  colors = FALSE,
  hex_code = FALSE,
  content = FALSE,
  xrange = NULL,
  yrange = NULL,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getPixelDf}{HistoImage}(
  object,
  colors = FALSE,
  hex_code = FALSE,
  content = FALSE,
  xrange = NULL,
  yrange = NULL,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getPixelDf}{Image}(
  object,
  colors = FALSE,
  hex_code = FALSE,
  use_greyscale = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  dbscan_eps = 1,
  dbscan_minPts = 3,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{colors}{Logical value. If \code{TRUE}, adds all colors from the image
as variables named \emph{col1}-\emph{col.n} where n is the number of colors.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}

\item{tissue}{Logical value. If \code{TRUE}, adds a variable called \emph{pxl_group}
that indicates whether the pixel is placed on a contiguous tissue section, on
artefact tissue fragments or on background.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame in which each row corresponds
to a pixel in the current image with x- and y-coordinates.
}
