% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getPixelDf}
\alias{getPixelDf}
\title{Obtain pixel data.frame}
\usage{
getPixelDf(object, xrange = NULL, yrange = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units. See info section \emph{Distance measures}
for more information.}
}
\value{
Data.frame with \code{nrow()} equal to the number of pixels.
}
\description{
Extracts a data.frame in which each row corresponds
to a pixel in the current image with x- and y-coordinates.
}
