% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getPixelDf}
\alias{getPixelDf}
\alias{getPixelDf,SPATA2-method}
\alias{getPixelDf,SpatialData-method}
\alias{getPixelDf,HistoImage-method}
\alias{getPixelDf,Image-method}
\title{Obtain pixel data.frame}
\usage{
getPixelDf(object, ...)

\S4method{getPixelDf}{SPATA2}(
  object,
  img_name = activeImage(object),
  colors = FALSE,
  hex_code = FALSE,
  content = FALSE,
  transform = TRUE,
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1
)

\S4method{getPixelDf}{SpatialData}(
  object,
  img_name = activeImage(object),
  colors = FALSE,
  hex_code = FALSE,
  content = FALSE,
  xrange = NULL,
  yrange = NULL,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getPixelDf}{HistoImage}(
  object,
  colors = FALSE,
  hex_code = FALSE,
  content = FALSE,
  xrange = NULL,
  yrange = NULL,
  transform = TRUE,
  scale_fct = 1,
  ...
)

\S4method{getPixelDf}{Image}(
  object,
  colors = FALSE,
  hex_code = FALSE,
  use_greyscale = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  eps = 1,
  minPts = 3,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{colors}{Logical value. If \code{TRUE}, adds all colors from the image
as variables named \emph{col1}-\emph{col.n} where n is the number of colors.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there is an image and \code{\link[=alignImage]{alignImage()}} has
been used.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{eps}{Distance measure. Given to \code{eps} of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines
the size (radius) of the epsilon neighborhood.}

\item{minPts}{Numeric value. Given to \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines the
number of minimum points required in the eps neighborhood for core points
(including the point itself)}

\item{tissue}{Logical value. If \code{TRUE}, adds a variable called \emph{pxl_group}
that indicates whether the pixel is placed on a contiguous tissue section, on
artefact tissue fragments or on background.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame in which each row corresponds
to a pixel in the current image with x- and y-coordinates.
}
