% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getPixelScaleFactor}
\alias{getPixelScaleFactor}
\alias{getPixelScaleFactor,SPATA2-method}
\alias{getPixelScaleFactor,SpatialData-method}
\alias{getPixelScaleFactor,HistoImage-method}
\title{Obtain scale factor for pixel to SI conversion}
\usage{
getPixelScaleFactor(object, ...)

\S4method{getPixelScaleFactor}{SPATA2}(
  object,
  unit,
  img_name = activeImage(object),
  switch = FALSE,
  add_attr = TRUE,
  verbose = NULL,
  ...
)

\S4method{getPixelScaleFactor}{SpatialData}(
  object,
  unit,
  img_name = activeImage(object),
  switch = FALSE,
  add_attr = TRUE,
  verbose = NULL,
  ...
)

\S4method{getPixelScaleFactor}{HistoImage}(
  object,
  unit,
  switch = FALSE,
  add_attr = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{unit}{Character value. The SI-unit of interest.
Determines the reference unit for the pixel size.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{switch}{Logical value. If \code{TRUE}, the unit of the output is switched.
See details for more.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A single numeric value with the unit defined in attribute \emph{unit}.
}
\description{
Extracts side length of pixel sides depending
on the resolution of the chosen image.
}
\details{
If \code{switch} is \code{FALSE}, the default, the output is to be interpreted as
unit/pixel. E.g. with \code{unit = 'um'} an output of \emph{15 'um/px'} means that under the current resolution
of the image height and width one pixel corresponds to \emph{15 um} in height and
width in the original tissue.

If \code{switch} is \code{TRUE}, the output is to be interpreted as pixel/unit.  E.g.
an output value of \emph{0.07 'px/um'} means that under the current image resolution
one micrometer corresponds to 0.07 pixel in the image.
}
\seealso{
\code{\link[=computePixelScaleFactor]{computePixelScaleFactor()}}, \code{\link[=setScaleFactor]{setScaleFactor()}}
}
