% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getPixelScaleFactor}
\alias{getPixelScaleFactor}
\title{Obtain scale factor for pixel to Euol conversion}
\usage{
getPixelScaleFactor(
  object,
  unit,
  switch = FALSE,
  square = FALSE,
  force = FALSE,
  add_attr = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{switch}{Logical value. If \code{TRUE}, the unit of the output is switched.
See details for more.}

\item{square}{Logical value. If \code{TRUE}, returns a scale factor for areas
instead of distances: The scale factor is squared and input for European units
of length is taken as a unit of area.}

\item{force}{Logical value. If \code{TRUE}, the scale factor is computed
regardless of what the function finds in the respective slot.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated Arguments or functions.}
}
\value{
A single numeric value with the unit defined in attribute \emph{unit}.

If \code{switch} is \code{FALSE}, the default, the output is to be interpreted as
unit/pixel. E.g. an output of \emph{15 'um/px'} means that under the current resolution
of the image height and width of one pixel corresponds to \emph{15 um} in height and
width in the original tissue.

If \code{switch} is \code{TRUE}, the output is to be interpreted as pixel/unit.  E.g.
an output value of \emph{0.07 'px/um'} means that under the current image resolution
one micrometer corresponds to 0.07 pixel in the image.
}
\description{
Extracts or computes the side length of pixel sides depending
on the current resolution of the image.
}
\seealso{
\code{setPixelScaleFactor()}
}
