% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getProcessedMatrixNames}
\alias{getProcessedMatrixNames}
\alias{getProcessedMatrixNames,SPATA2-method}
\alias{getProcessedMatrixNames,MolecularAssay-method}
\title{Obtain names of processed matrices}
\usage{
getProcessedMatrixNames(object, ...)

\S4method{getProcessedMatrixNames}{SPATA2}(object, assay_name = activeAssay(object), ...)

\S4method{getProcessedMatrixNames}{MolecularAssay}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Character vector.
}
\description{
Extract names of processed matrices.
}
\seealso{
\code{\link[=getMatrix]{getMatrix()}}
}
