% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getProjectionDf}
\alias{getProjectionDf}
\title{Obtain trajectory projection}
\usage{
getProjectionDf(object, id, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Given to \code{joinWith()}}
}
\value{
Data.frame that contains the projection length of each barcode-spot
in relation to the trajectory specified in \code{id}.
}
\description{
Extracts the projection data.frame of a trajectory. If \code{variables}
is specified
}
