% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSampleAreaSize}
\alias{getSampleAreaSize}
\title{Obtain sample area size}
\usage{
getSampleAreaSize(object, unit)
}
\arguments{
\item{unit}{Character value. Output unit. Must be one of \code{validUnitsOfArea()}.}
}
\value{
Single value. Numeric if \code{unit} is \emph{px}. Else value of class \code{unit}.
}
\description{
Computes and extracts the area size of the tissue sample.
}
