% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/getN-Z.R
\name{getSampleNames}
\alias{getSampleNames}
\alias{getSampleName}
\title{Obtain name of \code{SPATA2} object}
\usage{
getSampleNames(object)

getSampleName(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A character value.
}
\description{
Extracts the name/ID of the \code{SPATA2} object
in form of a single character value.
}
\keyword{internal}
