% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSampleName}
\alias{getSampleName}
\title{Obtain name of \code{SPATA2} object}
\usage{
getSampleName(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A character value.
}
\description{
Extracts the name/ID of the \code{SPATA2} object
in form of a single character value.
}
