% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{getSasBinAreas}
\alias{getSasBinAreas}
\title{Calculate SAS bin area}
\usage{
getSasBinAreas(
  object,
  area_unit,
  id = idSA(object),
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  use_outline = TRUE,
  remove_circle_bins = "Outside",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue section using \code{\link[=distToEdge]{distToEdge()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{use_outline}{Logical value. If \code{TRUE}, uses the outline variable
set with \code{setOutlineVarName()} or if none is set DBSCAN to identify the
outline of the tissue section or sections in case of multiple tissue sections
on one Visium slide to only compute the area of circle bins that covers the
tissue section.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Data.frame in which each observation corresponds to a circular bin.
}
\description{
Computes the area covered by each distance bin of the SAS algorithm.
}
\details{
Approximates the area each circular bin covers
by assigning each pixel to the circular bin it falls into.
Afterwards the number of pixels per bin is multiplied
with the area scale factor as is obtained by \code{getPixelScaleFactor(object, unit = unit)}
where unit is the squared unit of input for argument \code{binwidth}. E.g.
if \code{binwidth} = \emph{'0.1mm'} then \code{unit} = \emph{mm2}.
}
\keyword{internal}
