% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getH-M.R
\name{getSasDf}
\alias{getSasDf}
\alias{getSpatialAnnotationScreeningDf}
\title{Obtain spatial annotation screening data.frame}
\usage{
getSasDf(
  object,
  id,
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  angle_span = c(0, 360),
  n_bins_angle = 1,
  variables = NULL,
  method_gs = NULL,
  core = TRUE,
  periphery = TRUE,
  summarize_by = c("bins_circle", "bins_angle"),
  bcs_exclude = NULL,
  verbose = FALSE,
  ...
)

getSpatialAnnotationScreeningDf(...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{n_bins_dist}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}. Not required if \code{distance} and \code{binwidth} is
specified. (See details for more.)}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{bcs_exclude}{Character value containing name(s) of data points to be excluded from the analysis.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame of inferred gradients of numeric
variables as a fucntion of distance to spatial annotations.
}
\examples{

library(SPATA2)
library(SPATAData)

data("image_annotations")

necrotic_spat_ann <- image_annotations[["313_T"]][["necrotic_center"]]

object <- downloadSpataObject(sample_name = "313_T")

object <- setSpatialAnnotation(object = object, spat_ann = necrotic_spat_ann)

plotSurfaceIAS(
 object = object,
 id = "necrotic_center",
 distance = 200
 )

plotSurfaceIAS(
 object = object,
 id = "necrotic_center",
 distance = 200,
 binwidth = getCCD(object)*4, # lower resolution by increasing binwidth for visualization
 n_bins_angle = 12,
 display_angle = TRUE
 )

getSasDf(
  object = object,
  id = "necrotic_center",
  distance = 200,
  variables = "VEGFA"
  )

getSasDf(
  object = object,
   id = "necrotic_center",
   distance = 200,
   variables = "VEGFA"
   )

getSasDf(
  object = object,
   id = "necrotic_center",
   distance = 200,
   variables = "VEGFA",
   n_bins_angle = 12
   )

}
