% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getScaleFactor}
\alias{getScaleFactor}
\alias{getScaleFactor,HistoImaging-method}
\alias{getScaleFactor,HistoImage-method}
\title{Obtain scale factors}
\usage{
getScaleFactor(object, ...)

\S4method{getScaleFactor}{HistoImaging}(object, fct_name, img_name = NULL)

\S4method{getScaleFactor}{HistoImage}(object, fct_name)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{fct_name}{Character value. Name of the scale factor.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
Single value whose properties depend on \code{fct_name}.
}
\description{
Extracts scale factors.
}
