% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSegmentationNames}
\alias{getSegmentationNames}
\title{Obtain segmentation variable names}
\usage{
getSegmentationNames(object, fdb_fn = "message", ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector.
}
\description{
Extracts the names of the variables that have been created
via \code{createSpatialSegmentation()}.
}
