% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSegmentationNames}
\alias{getSegmentationNames}
\alias{getSegmentationVariableNames}
\alias{addSegmentationVariable}
\alias{discardSegmentationVariable}
\title{Obtain segmentation variable names}
\usage{
getSegmentationNames(object, fdb_fn = "message", ...)

getSegmentationVariableNames(object, fdb_fn = "message", ...)

addSegmentationVariable(object, name, verbose = NULL, ...)

discardSegmentationVariable(object, name, verbose = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Character vector.
}
\description{
Extracts the names of the variables that have been created
via \code{createSegmentation()}.
}
