% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSegmentNames}
\alias{getSegmentNames}
\title{Obtain segment names}
\usage{
getSegmentNames(object, simplify = TRUE, of_sample = NA, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of segments which were drawn for the
specific sample.
}
\description{
Obtain segment names
}
