% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{getSmrdResultsDf}
\alias{getSmrdResultsDf}
\alias{getSmrdResultsDf,ImageAnnotationScreening-method}
\title{Obtain IAS results (data.frame)}
\usage{
getSmrdResultsDf(object, ...)

\S4method{getSmrdResultsDf}{ImageAnnotationScreening}(
  object,
  var_pval = "pvalue_median",
  var_eval = "corr_median",
  threshold_pval = 1,
  threshold_eval = 0,
  model_subset = NULL,
  model_remove = NULL
)
}
\arguments{
\item{object}{Any object for which a method has been defined.}

\item{var_pval, var_eval}{Character value. Specifies the p-value- and the
evaluation variable based on which the thresholds are applied.}

\item{threshold_pval, threshold_eval}{Numeric value. Used to filter
the output accordingly.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}
}
\value{
Data.frame.
}
\description{
Extracts (and filters) the summarized IAS results in form
of a data.frame
}
