% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpataDf}
\alias{getSpataDf}
\title{Obtain a spata-data.frame}
\usage{
getSpataDf(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A tidy data.frame containing the character variables \emph{barcodes}
and \emph{sample}.
}
\description{
This function is the most basic start if you want
to extract data for your individual analysis.

(In order to extract the coordinates as well use \code{getCoordsDf()}.)
}
\seealso{
joinWith
}
