% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getSpataDf}
\alias{getSpataDf}
\title{Obtain a spata-data.frame}
\usage{
getSpataDf(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A tidy data.frame containing the character variables \emph{barcodes}
and \emph{sample}.
}
\description{
This function is the most basic start if you want
to extract data for your individual analysis.

(In order to extract the coordinates as well use \code{getCoordsDf()}.)
}
\seealso{
joinWith
}
