% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnArea}
\alias{getSpatAnnArea}
\alias{getSpatAnnArea,SPATA2-method}
\alias{getSpatAnnArea,SpatialData-method}
\title{Obtain area of spatial annotation}
\usage{
getSpatAnnArea(object, ...)

\S4method{getSpatAnnArea}{SPATA2}(
  object,
  ids = NULL,
  unit = "mm2",
  tags = NULL,
  test = "any",
  as_numeric = TRUE,
  verbose = NULL,
  ...
)

\S4method{getSpatAnnArea}{SpatialData}(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  unit = "mm2",
  as_numeric = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsOfArea]{validUnitsOfArea()}} for valid
input options.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of spatial annotations} for more information.}

\item{test}{Character value. One of \emph{c('any'. 'all', 'identical', 'not_identical', 'none')}.
Specifies how input for \code{tags} is used to select spatial annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Contains the area of the spatial annotations in the unit that is specified in \code{unit}.
The unit is attached to the output as an attribute named \emph{unit}. E.g. if
\verb{unit = *mm2*} the output value has the unit \emph{mm^2}.
}
\description{
Computes the area of spatial annotations.
}
\note{
The area is computed solely based on the outline of the annotation even
if the annotation transgresses the tissue outline! If you only want the area
of the annotation on the tissue section adjust the annotation with \code{\link[=mergeWithTissueOutline]{mergeWithTissueOutline()}}.
See examples.
}
\examples{

library(SPATA2)

object <- loadExampleObject("UKF313T")

ids <- c("necrotic_edge2", "necrotic_edge2_transgr")

plotSpatialAnnotations(object, ids = ids)

getSpatAnnArea(object, ids = ids)

}
\seealso{
\code{\link[=getSpatAnnOutlineDf]{getSpatAnnOutlineDf()}}, \code{\link[=getCCD]{getCCD()}}, \code{\link[=as_unit]{as_unit()}}
}
