% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImgAnns.R
\name{getSpatAnnCenter}
\alias{getSpatAnnCenter}
\alias{getSpatAnnCenter,spata2-method}
\alias{getSpatAnnCenter,SpatialAnnotation-method}
\alias{getSpatAnnCenters}
\alias{getSpatAnnCenters,spata2-method}
\alias{getSpatAnnCenters,SpatialAnnotation-method}
\title{Obtain center of an spatial annotation}
\usage{
getSpatAnnCenter(object, ...)

\S4method{getSpatAnnCenter}{spata2}(object, id)

\S4method{getSpatAnnCenter}{SpatialAnnotation}(object)

getSpatAnnCenters(object, ...)

\S4method{getSpatAnnCenters}{spata2}(object, id, outer = TRUE, inner = TRUE)

\S4method{getSpatAnnCenters}{SpatialAnnotation}(object, outer = TRUE, inner = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value. The ID of the spatial annotation of interest.}

\item{outer}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the outer border of the image annotation
is not included in the output.}

\item{inner}{Logical value. Only applies if an image annotation contains a secondary image annotation within its own area. If \code{FALSE}, the inner borders of the image annotation
are not included in the output.}
}
\value{
Numeric vector of length two or a list of these. Values are named \emph{x} and \emph{y}.
}
\description{
\code{getSpatAnnCenter()} computes the
x- and y- coordinates of the center of the outer border, returns
a numeric vector of length two. \code{getSpatAnnCenters()} computes the center of the outer
and every inner border and returns a list of numeric vectors of length two.
}
