% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatAnnSf}
\alias{getSpatAnnSf}
\title{Obtain the outline of a spatial anontation}
\usage{
getSpatAnnSf(object, id, img_name = activeImage(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}
}
\value{
An object of class \code{POLYGON} from the \code{sf} package.
}
\description{
Exracts an object as created by \code{sf::st_polygon()} that
corresponds to the spatial annotation.
}
