% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImgAnns.R
\name{getSpatAnnTags}
\alias{getSpatAnnTags}
\alias{getSpatAnnTags,spata2-method}
\alias{getSpatAnnTags,HistoImaging-method}
\title{Obtain spatial annotation tags}
\usage{
getSpatAnnTags(object, ...)

\S4method{getSpatAnnTags}{spata2}(object)

\S4method{getSpatAnnTags}{HistoImaging}(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character vector.
}
\description{
Extracts all unique tags with which spatial annotations
have been tagged.
}
