% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{getSpatialMethod}
\alias{getSpatialMethod}
\alias{getSpatialMethod,spata2-method}
\alias{getSpatialMethod,HistoImaging-method}
\title{Obtain spatial method}
\usage{
getSpatialMethod(object, ...)

\S4method{getSpatialMethod}{spata2}(object)

\S4method{getSpatialMethod}{HistoImaging}(object)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{SpatialMethod}.
}
\description{
Extracts an S4 object of class \code{SpatialMethod} that contains
meta data about the set up of the protocol that was followed to create
the data used for the \code{SPATA2} object.
}
\seealso{
\code{\linkS4class{SpatialMethod}}
}
