% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getSpatialMethod}
\alias{getSpatialMethod}
\alias{getSpatialMethod,ANY-method}
\title{Obtain spatial method}
\usage{
getSpatialMethod(object, ...)

\S4method{getSpatialMethod}{ANY}(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{SpatialMethod}.
}
\description{
Extracts an S4 object of class \code{SpatialMethod} that contains
meta data about the set up of the protocol that was followed to create
the data used for the object.
}
\seealso{
\code{\linkS4class{SpatialMethod}}
}
