% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{getSpatialTrajectory}
\alias{getSpatialTrajectory}
\title{Obtain object of class \code{SpatialTrajectory}.}
\usage{
getSpatialTrajectory(object, id)
}
\arguments{
\item{object}{A valid spata-object.}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}
}
\value{
An object of class \code{SpatialTrajectory.}
}
\description{
Obtain object of class \code{SpatialTrajectory}.
}
