% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getStsDf}
\alias{getStsDf}
\title{Obtain spatial trajectory screening data.frame}
\usage{
getStsDf(
  object,
  variables,
  id = idST(object),
  binwidth = recBinwidth(object),
  width = NULL,
  unit = getDefaultUnit(object),
  ro = c(0, 1),
  bcs_exclude = NULL,
  format = "wide",
  control = SPATA2::sgs_loess_control,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. The numeric variables of interest. Must be inside:

\itemize{
\item{ \strong{Gene sets} Must be in \code{getGeneSets()}}
\item{ \strong{Genes} Must be in \code{getGenes()}}
\item{ \strong{Features} Must be in \code{getFeatureNames(..., of_class = "numeric")}}
}}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}

\item{binwidth}{Distance value. The width of the bins to which
the barcode-spots are assigned. Defaults to the center-center
distance: \code{binwidth = getCCD(object)}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Data.frame.
}
\description{
Extracts a data.frame of inferred gradients related to the
course of a trajectory.
}
