% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTissueArea}
\alias{getTissueArea}
\title{Obtain tissue area size}
\usage{
getTissueArea(object, unit, method = NULL, img_name = activeImage(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{unit}{Character value. Output unit. Must be one of \code{validUnitsOfArea()}.}

\item{method}{Character value. Either \emph{'obs'} or \emph{'image'}. Decides whether
the tissue outline used based on the \link[=concept_observations]{observations}
or the image is used. If \code{method = NULL}, the function checks first if any \code{\link{HistoImage}}
is registered. If so, the outline from the image specified with \code{img_name} is returned.
If there are no images, the outline computed with \code{identifyTissueOutline(..., method = 'obs')}
is used.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
Single value. Numeric if \code{unit} is \emph{px}. Else value of class \code{unit}.
}
\description{
Computes and extracts the size of the area covered by the tissue.
}
