% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTissueOutlineCentroid}
\alias{getTissueOutlineCentroid}
\alias{getTissueOutlineCentroid,SpatialData-method}
\alias{getTissueOutlineCentroid,HistoImage-method}
\title{Obtain tissue outline centroid}
\usage{
getTissueOutlineCentroid(object, ...)

\S4method{getTissueOutlineCentroid}{SpatialData}(
  object,
  img_name = activeImage(object),
  transform = TRUE,
  ...
)

\S4method{getTissueOutlineCentroid}{HistoImage}(object, transform = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
Numeric vector of length two.
}
\description{
Extracts the centroid of the polygon used to outline
the whole tissue.
}
