% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTissueOutlineCentroid}
\alias{getTissueOutlineCentroid}
\alias{getTissueOutlineCentroid,SpatialData-method}
\alias{getTissueOutlineCentroid,HistoImage-method}
\title{Obtain tissue outline centroid}
\usage{
getTissueOutlineCentroid(object, ...)

\S4method{getTissueOutlineCentroid}{SpatialData}(
  object,
  method = NULL,
  img_name = activeImage(object),
  transform = TRUE,
  ...
)

\S4method{getTissueOutlineCentroid}{HistoImage}(object, transform = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{method}{Character value. Either \emph{'obs'} or \emph{'image'}. Decides whether
the tissue outline used based on the \link[=concept_observations]{observations}
or the image is used. If \code{method = NULL}, the function checks first if any \code{\link{HistoImage}}
is registered. If so, the outline from the image specified with \code{img_name} is returned.
If there are no images, the outline computed with \code{identifyTissueOutline(..., method = 'obs')}
is used.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there are any images registered
in the \code{SPATA2} object.}
}
\value{
Numeric vector of length two.
}
\description{
Extracts the centroid of the polygon used to outline
the whole tissue.
}
