% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTissueOutlineDf}
\alias{getTissueOutlineDf}
\title{Obtain outer barcode spots}
\usage{
getTissueOutlineDf(object, outline_nn = c(2, 3), force = FALSE, remove = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{outline_nn}{Numeric vector. Number of neighbors a barcode spot
might have to be considered an outline spot.}

\item{force}{Logical. If \code{TRUE}, forces computation.}

\item{remove}{Logical. If \code{TRUE}, none-outline spots are removed from
the output.}
}
\value{
Output of \code{getCoordsDf()} including a logical variable
called \emph{outline} which indicates whether the spot belongs to the outlining
spots or not and a character variable called \emph{section} that indicates to
which tissue section the spots belong.
}
\description{
Identifies the barcode spots that lie on the edge
of the tissue and returns a subset of the coordinates data.frame.
}
\examples{

 library(ggplot2)
 library(ggExtra)

 object <- downloadPubExample("MCI_LMU")

 to_df <- getTissueOutlineDf(object, force = TRUE, remove = FALSE)

 to_df[["outline"]] <- as.character(to_df[["outline"]])

 ggplot(to_df, mapping = aes(x = x, y = y)) +
   geom_point_fixed(mapping = aes(color = section, alpha = outline), size = getDefault(object, "pt_size")) +
   scale_alpha_manual(values = c("TRUE" = 1, "FALSE" = 0.4))


}
