% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTissueOutlineDf}
\alias{getTissueOutlineDf}
\title{Obtain outline barcode spots}
\usage{
getTissueOutlineDf(object, remove = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{remove}{Logical. If \code{TRUE}, none-outline spots are removed from
the output.}

\item{force}{Logical. If \code{TRUE}, forces computation.}
}
\value{
Output of \code{getCoordsDf()} filtered based on the \emph{outline} variable.
}
\description{
Identifies the barcode spots that lie on the edge
of the tissue and returns a subset of the coordinates data.frame. Requires
the results of \code{identifyTissueOutline()}.
}
\examples{

 library(ggplot2)
 library(ggExtra)

 object <- downloadPubExample("MCI_LMU")

 print(getTissueOutlineDf(object))

 to_df <- getTissueOutlineDf(object, remove = FALSE)

 to_df[["outline"]] <- as.character(to_df[["outline"]])

 ggplot(to_df, mapping = aes(x = x, y = y)) +
   geom_point_fixed(mapping = aes(color = section, alpha = outline), size = getDefault(object, "pt_size")) +
   scale_alpha_manual(values = c("TRUE" = 1, "FALSE" = 0.4))



}
