% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTissueOutlineDf}
\alias{getTissueOutlineDf}
\alias{getTissueOutlineDf,SPATA2-method}
\alias{getTissueOutlineDf,SpatialData-method}
\alias{getTissueOutlineDf,HistoImage-method}
\title{Obtain outline barcode spots}
\usage{
getTissueOutlineDf(object, ...)

\S4method{getTissueOutlineDf}{SPATA2}(
  object,
  img_name = activeImage(object),
  by_section = TRUE,
  transform = TRUE,
  ...
)

\S4method{getTissueOutlineDf}{SpatialData}(
  object,
  img_name = activeImage(object),
  by_section = TRUE,
  transform = TRUE
)

\S4method{getTissueOutlineDf}{HistoImage}(object, by_section = TRUE, transform = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{force}{Logical. If \code{TRUE}, forces computation.}
}
\value{
Output of \code{getCoordsDf()} filtered based on the \emph{outline} variable.
}
\description{
Extracts the polygons necessary to outline the tissue.
}
