% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{getTrajectoryDf}
\alias{getTrajectoryDf}
\title{Obtain a summarized trajectory data.frame}
\usage{
getTrajectoryDf(
  object,
  id,
  variables,
  binwidth = getBarcodeSpotDistance(object),
  n_bins = NA_integer_,
  method_gs = "mean",
  normalize = TRUE,
  summarize_with = "mean",
  format = "long",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{variables}{Character vector. The numeric variables of interest. Must be inside:

\itemize{
  \item{ \strong{Gene sets} Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} Must be in \code{getGenes()}}
  \item{ \strong{Features} Must be in \code{getFeatureNames(..., of_class = "numeric")}}
  }}

\item{binwidth}{The width of the the bins in which the barcode-spots
are binned according to their projection length values.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{shift_wider}{Logical. If set to TRUE the trajectory data.frame is
shifted to it's wider format. Formats can be changed via \code{shiftTrajectoryDf()}.}
}
\value{
Data.frame.
}
\description{
Computes the expression trends of all specified variables
along the direction of the spatial trajectory.
}
\details{
Initially the projection data.frame of the specified trajectory
is joined with the respective input of variables via \code{joinWithVariables()}.

The argument \code{binwidth} refers to the amount of which the barcode-spots of the
given trajectory will be summarized with regards to the trajectory's direction:
The amount of \code{binwidth} and the previously specified 'trajectory width' in \code{createTrajectories()}
determine the length and width of the sub-rectangles in which the rectangle the
trajectory embraces are splitted and in which all barcode-spots are binned.
Via \code{dplyr::summarize()} the variable-means of every sub-rectangle are calculated.
These mean-values are then arranged according to the trajectory's direction.

Eventually the data.frame is shifted via \code{tidyr::pivot_longer()} to a data.frame in which
every observation refers to the mean-value of one of the specified variable-elements (e.g. a specified
gene set) of the particular sub-rectangle. The returned data.frame contains the following variables:

\itemize{
 \item{\emph{trajectory_part}: Character. Specifies the trajectory's sub-part of the observation. (Negligible if there is
 only one trajectory part.)}
 \item{\emph{trajectory_part_order}: Numeric. Indicates the order within the trajectory-part. (Negligible if there is
 only one trajectory part.)}
 \item{\emph{trajectory_order}: Numeric. Indicates the order within the whole trajectory.}
 \item{\emph{variables}: Character. The respective gene sets, gene or feature the value refers to.}
 \item{\emph{values}: Numeric. The actual summarized values.}}
}
