% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrajectoryDf}
\alias{getTrajectoryDf}
\title{Obtain a summarized trajectory data.frame}
\usage{
getTrajectoryDf(
  object,
  trajectory_name,
  variables,
  method_gs = "mean",
  binwidth = 5,
  normalize = TRUE,
  shift_wider = FALSE,
  verbose = TRUE,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{variables}{Character vector. The variables of interest:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 \item{ \strong{Features}: Must be numeric ones of \code{getFeatureNames()}}
 }}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{shift_wider}{Logical. If set to TRUE the trajectory data.frame is
shifted to it's wider format. Formats can be changed via \code{shiftTrajectoryDf()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A summarized trajectory data.frame.
}
\description{
Computes the expression trends of all specified variables
along the direction of the spatial trajectory.
}
\details{
Initially the compiled trajectory data.frame of the specified trajectory
is joined with the respective input of variables via \code{joinWithVariables()}.

The argument \code{binwidth} refers to the amount of which the barcode-spots of the
given trajectory will be summarized with regards to the trajectory's direction:
The amount of \code{binwidth} and the previously specified 'trajectory width' in \code{createTrajectories()}
determine the length and width of the sub-rectangles in which the rectangle the
trajectory embraces are splitted and in which all barcode-spots are binned.
Via \code{dplyr::summarize()} the variable-means of every sub-rectangle are calculated.
These mean-values are then arranged according to the trajectory's direction.

Eventually the data.frame is shifted via \code{tidyr::pivot_longer()} to a data.frame in which
every observation refers to the mean-value of one of the specified variable-elements (e.g. a specified
gene set) of the particular sub-rectangle. The returned data.frame contains the following variables:

\itemize{
 \item{\emph{trajectory_part}: Character. Specifies the trajectory's sub-part of the observation. (Negligible if there is
 only one trajectory part.)}
 \item{\emph{trajectory_part_order}: Numeric. Indicates the order within the trajectory-part. (Negligible if there is
 only one trajectory part.)}
 \item{\emph{trajectory_order}: Numeric. Indicates the order within the whole trajectory.}
 \item{\emph{gene_sets, genes or features}: Character. The respective gene sets, gene or feature the value refers to.}
 \item{\emph{values}: Numeric. The actual summarized values.}}
}
