% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTrajectoryIds}
\alias{getTrajectoryIds}
\title{Obtain trajectory ids}
\usage{
getTrajectoryIds(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Character vector.
}
\description{
Extracts the ids of all objects of class \code{Trajectory}
in the SPATA2 object.
}
