% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTrajectoryIds}
\alias{getTrajectoryIds}
\title{Obtain trajectory IDs}
\usage{
getTrajectoryIds(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character vector.
}
\description{
Extracts the ids of all objects of class \code{\link{SpatialTrajectory}}
in the \code{\link{SPATA2}} object.
}
