% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getA-Z.R
\name{getTrajectoryLength}
\alias{getTrajectoryLength}
\title{Obtain length of trajectory}
\usage{
getTrajectoryLength(object, id, unit = "px", round = FALSE, as_numeric = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}

\item{unit}{Character value. Specifies the desired unit.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}

\item{as_numeric}{Logical value. If \code{TRUE}, forces the output to be numeric.
This means that the unit is not \bold{not} suffixed.}
}
\value{
Transformed input. Character vector of the same length as \code{input}.
}
\description{
Computes and returns the length of a trajectory.
}
\details{
Input to specify a distance can be provided in two different
ways.

\bold{Distance in pixel}:

There are two valid input options to specify the distance in pixel:

\itemize{
 \item{numeric:}{ Single numeric values, e.g. 2, 3.554, 69, 100.67. If no unit
 is specified the input will be interpreted as pixels.}
 \item{character:}{ Suffixed with \emph{'px'}, e.g. \emph{'2px'}, \emph{'3.554px'}}
 }

\bold{Distance in European units of length (eUOL)}:

 Specifying distances in European units of length e.g. \emph{'2mm'}, \emph{'400um'} etc.
 requires the input to be a character as the unit must be provided as suffix. Between the numeric
 value and the unit must be no empty space! Unit suffixes must be one of
 \emph{'m', 'dm', 'cm', 'mm', 'um', 'nm'}.
}
