% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTrajectoryLength}
\alias{getTrajectoryLength}
\title{Obtain length of spatial trajectory}
\usage{
getTrajectoryLength(object, id, unit = "px", round = FALSE, as_numeric = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}
}
\value{
The length of the spatial directory as a single \link[=concept_distance_measure]{distance value}.
}
\description{
Computes and returns the length of a spatial trajectory.
}
