% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrajectoryNames}
\alias{getTrajectoryNames}
\title{Obtain trajectory names}
\usage{
getTrajectoryNames(object, simplify = TRUE, of_sample = NA, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of trajectories which were drawn for the
specific sample.
}
\description{
Obtain trajectory names
}
