% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/get-family.R
\name{getTrajectoryNames}
\alias{getTrajectoryNames}
\title{Deprecated.}
\usage{
getTrajectoryNames(
  object,
  simplify = TRUE,
  of_sample = NA,
  verbose = FALSE,
  ...
)

getTrajectoryNames(
  object,
  simplify = TRUE,
  of_sample = NA,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{simplify}{Logical. If set to TRUE the output list is simplified to a vector if possible. If set
to FALSE a list is returned.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A list named according to the \code{of_sample} in which each element is
a character vector containing the names of trajectories which were drawn for the
specific sample.
}
\description{
This function is deprecated in favor of
getTrajectoryIds().
}
\keyword{internal}
