% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{getTrajectoryNames}
\alias{getTrajectoryNames}
\title{Deprecated.}
\usage{
getTrajectoryNames(object, ...)
}
\description{
This function is deprecated in favor of
getTrajectoryIds().
}
\keyword{internal}
