% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{getTrajectoryObject}
\alias{getTrajectoryObject}
\title{Obtain trajectory object}
\usage{
getTrajectoryObject(
  object,
  trajectory_name = getDefaultTrajectory(object, verbose = TRUE, "trajectory_name"),
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
An object of class \code{spatialTrajectory}.
}
\description{
Obtain trajectory object
}
