% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R, R/getN-Z.R
\name{getTrajectoryScreeningDf}
\alias{getTrajectoryScreeningDf}
\alias{getTrajectoryDf}
\title{Deprecated}
\usage{
getTrajectoryScreeningDf(...)

getTrajectoryDf(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  method_gs = NULL,
  normalize = TRUE,
  summarize_with = FALSE,
  smooth_span = 0,
  format = "wide",
  verbose = NULL,
  ...
)
}
\description{
Deprecated
}
