% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTrajectorySegmentDf}
\alias{getTrajectorySegmentDf}
\title{Obtain trjectory course}
\usage{
getTrajectorySegmentDf(
  object,
  id = getDefaultTrajectoryId(object, verbose = TRUE, "id"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{...}{Used to absorb deprecated Arguments or functions.}
}
\value{
Data.frame.
}
\description{
Extracts data.frame that contains the course
of a spatial trajectory.
}
