% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R
\name{getTrajectorySegmentDf}
\alias{getTrajectorySegmentDf}
\title{Obtain trjectory course}
\usage{
getTrajectorySegmentDf(
  object,
  id = getDefaultTrajectoryId(object, verbose = TRUE, "id"),
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Data.frame.
}
\description{
Extracts data.frame that contains the course
of a spatial trajectory.
}
