% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-family.R, R/getN-Z.R
\name{getTrajectorySegmentDf}
\alias{getTrajectorySegmentDf}
\title{Obtain trjectory course}
\usage{
getTrajectorySegmentDf(
  object,
  id = getDefaultTrajectoryId(object, verbose = TRUE, "id"),
  ...
)

getTrajectorySegmentDf(
  object,
  id = getDefaultTrajectoryId(object, verbose = TRUE, "id"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Data.frame.

Data.frame.
}
\description{
Extracts data.frame that contains the course
of a spatial trajectory.

Extracts data.frame that contains the course
of a spatial trajectory.
}
