% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getTrajectoryWidth}
\alias{getTrajectoryWidth}
\title{Obtain trajectory width}
\usage{
getTrajectoryWidth(object, id = idST(object), unit = "px", orig = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \link{`validUnitsOfArea()'} for valid
input options.}
}
\value{
Distance value.
}
\description{
Computes and extracts the default width of the trajectory.
}
