% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getVariableMolecules}
\alias{getVariableMolecules}
\title{Obtain variable molecules}
\usage{
getVariableMolecules(object, method, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{method}{The selection method of interest.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
Character vector.
}
\description{
Extracts results of \code{\link[=identifyVariableMolecules]{identifyVariableMolecules()}}.
}
\examples{

library(SPATA2)
data("example_data")

object <- loadExampleObject("UKF269T")

object <- identifyVariableMolecules(object, n_mol = 2000)
object <- identifyVariableMolecules(object, n_mol = 3000, method = "mean.var.plot")

vars_vst <- getVariableMolecules(object, method = "vst")
vars_mvp <- getVariableMolecules(object, method = "mean.var.plot")

length(vars_vst)
length(vars_mvp)

}
