% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R
\name{getVariableNames}
\alias{getVariableNames}
\title{Obtain variable names}
\usage{
getVariableNames(object, protected = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{protected}{Logical value. If \code{TRUE}, variable names that are protected
in \code{SPATA2} are returned, too, regardless of being in use or not.}
}
\value{
Character vector.
}
\description{
Extracts a character vector of variable names that are currently
known to the \code{SPATA2} object.
}
\note{
Molecule names are picked from the currently active matrix of their assay!
If a processed matrix is active and it does not contain some molecules the raw count matrix
}
