% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{getVarNames}
\alias{getVarNames}
\alias{getVarNames,ImageAnnotationScreening-method}
\title{Obtain IAS results (variable names)}
\usage{
getVarNames(object, ...)

\S4method{getVarNames}{ImageAnnotationScreening}(
  object,
  var_pval = "pvalue_median",
  var_eval = "corr_median",
  var_arrange = var_eval,
  threshold_pval = 1,
  threshold_eval = 0,
  model_subset = NULL,
  model_remove = NULL
)
}
\arguments{
\item{var_arrange}{The evaulation variable based on which the variables
are arranged before being returned. Defaults to input of argument \code{var_eval}.}
}
\value{
Data.frame.
}
\description{
Extracts (and filters) the summarized IAS results in form
of a character vector of variable names.
}
\details{
After the filtering the variable names are arranged according
to their values of \code{var_eval}.
}
