% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpInit}
\alias{ggpInit}
\title{Initiate ggplot2 layering}
\usage{
ggpInit(object = "object", theme = "void", data = "coords")
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{theme}{Character value. String that denotes the default
theme. Defaults to \code{void}}
}
\value{
An empty ggplot.
}
\description{
Initiates a ggplot object to which \code{ggpLayer}-
functions can be added for individual plotting ideas.
}
