% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{ggpLayer_dummy}
\alias{ggpLayer_dummy}
\title{ggpLayer}
\usage{
ggpLayer_dummy()
}
\value{
A list of \code{ggproto}-objects.
}
\description{
ggpLayer
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
