% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerCaptureArea}
\alias{ggpLayerCaptureArea}
\title{Add capture area to surface plot}
\usage{
ggpLayerCaptureArea(
  object,
  opt = c("rect"),
  rect_alpha = 0.9,
  rect_clr = "black",
  rect_line_type = "solid",
  rect_size = 1,
  expand_rect = 0.025,
  expand_x = ggplot2::waiver(),
  expand_y = ggplot2::waiver(),
  img_name = activeImage(object)
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{opt}{Combination of \emph{'rect'} and/or \emph{'crop'}.}

\item{rect_alpha}{The alpha transparency level of the rectangle's border}

\item{expand_x, expand_y}{Given to \code{expand} of \code{ggplot2::scale_x/y_continuous()}.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
List of ggpLayer outputs.
}
\description{
Adds the capture area as a rectangular and/or
crops the frame of the plot accordingly.
}
\examples{

library(SPATA2)

data("example_data")

object <- loadExampleObject("UKF275T")

plotImage(object) +
 ggpLayerCaptureArea(object, opt = "rect")

plotImage(object) +
 ggpLayerCaptureArea(object, opt = "crop")

}
\seealso{
\code{\link[=getCaptureArea]{getCaptureArea()}}
}
