% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{ggpLayerCaptureArea}
\alias{ggpLayerCaptureArea}
\title{Add capture area to surface plot}
\usage{
ggpLayerCaptureArea(
  object,
  img_name = NULL,
  opt = c("rect"),
  rect_alpha = 0.9,
  rect_clr = "black",
  rect_line_type = "solid",
  rect_size = 1,
  expand_rect = 1.01,
  expand_x = ggplot2::waiver(),
  expand_y = ggplot2::waiver()
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{opt}{Combination of \emph{'rect'} and/or \emph{'crop'}.}

\item{expand_x, expand_y}{Given to \code{expand} of \code{ggplot2::scale_x/y_continuous()}.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}
}
\value{
List of ggpLayer outputs.
}
\description{
Plots the capture area as a rectangular and/or
crops the frame of the plot accordingly.
}
