% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerEncirclingGroups}
\alias{ggpLayerEncirclingGroups}
\title{Add group encircling}
\usage{
ggpLayerEncirclingGroups(
  object,
  plot_type = "coords",
  grouping_variable,
  groups_subset = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{plot_type}{Character value. Either \emph{'surface', 'tsne'} or
\emph{'umap'}.}

\item{grouping_variable}{Character value. The grouping variable of choice.}

\item{groups_subset}{Character value or NULL. If character,
specifies the exact groups that are encircled. If NULL, all groups
are encircled.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Highlights groups of barcode-spots by encircling them.
Depending on the \code{plot_type} this can be added to a surface plot
or a dimensional reduction plot.
}
