% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerEncirclingIAS}
\alias{ggpLayerEncirclingIAS}
\title{Add IAS area expansion}
\usage{
ggpLayerEncirclingIAS(
  object,
  id,
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  linecolor = "black",
  linesize = 1
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{distance}{Distance value. Specifies the distance from the border of the
image annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
 (See details for more.)}

\item{binwidth}{Distance value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the center-center
distance: \code{binwidth = getCCD(object)}. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{linesize}{Numeric value. The size of the line(s) plotted.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Adds the circular expansion used by the IAS-algorithm
of the area of  an image annotation to a surface plot.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
