% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerEncirclingIAS}
\alias{ggpLayerEncirclingIAS}
\title{Add IAS area expansion}
\usage{
ggpLayerEncirclingIAS(
  object,
  id,
  buffer,
  n_bins_circle,
  linecolor = "black",
  linesize = 1
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}. (See details for more.)}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{buffer}.
 (See details for more.)}

\item{linesize}{Numeric value. The size of the line(s) plotted.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Adds the circular expansion used by the IAS-algorithm
of the area of  an image annotation to a surface plot.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
