% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerExprEstimatesSAS}
\alias{ggpLayerExprEstimatesSAS}
\title{Add IAS area expansion}
\usage{
ggpLayerExprEstimatesSAS(
  object,
  ids,
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  core = FALSE,
  alpha_core = 0,
  fill_core = NA,
  line_alpha = 1,
  line_color = "black",
  line_size = (line_size_core * 0.75),
  line_size_core = 1,
  incl_edge = TRUE,
  direction = "outwards",
  method = "normal",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. The size with which to display encircling lines
of the area expansion.}

\item{line_size_core}{Numeric. The size with which to display the core outline
of the spatial annotation.}

\item{incl_edge}{Logical. If \code{TRUE}, makes use of \code{SPATA2} automatic tissue outline algorithm.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds the circular expansion used by the IAS-algorithm
of the area of  an spatial annotation to a surface plot.
}
\examples{

object <- downloadPubExample("313_T")

plotImageGgplot(object) +
 ggpLayerEncirclingSAS(
   object = object,
   id = "necrotic_area",
   distance = "2.25mmm",
   binwidth = "112.5um"
 )
}
