% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByImage}
\title{Set plot limits}
\usage{
ggpLayerFrameByCoords(object = "object", ...)

ggpLayerFrameByImage(object = "object", ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Sets the limits on the x- and y-axis of a surface plot based on
the coordinate range of the objects \link[=concept_observations]{observations}
or the image.

Soft deprecated in favor of the arguments \code{xrange} and \code{yrange} in \code{\link[=plotSurface]{plotSurface()}}.
}
\examples{

library(SPATA2)
library(patchwork)

data("example_data")

object <- example_data$object_UKF275T_diet

plotImage(object) + plotSurface(object)

(plotImage(object) + ggpLayerFrameByCoords(object)) + plotSurface(object)


}
