% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByImage}
\title{Set plot limits}
\usage{
ggpLayerFrameByCoords(object = "object", opt = "coords")

ggpLayerFrameByImage(object = "object", opt = "coords")
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{opt}{Specifies the function with which the limits are set. If
\emph{'scale'} (the default), \code{ggplot2::scale_x|y_continuous()} is used.
If \emph{'coords'}, \code{ggplot2::coord_cartesian()} is used.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Sets the limits on the x- and y-axis of a ggplot based on the coordinate
range or the image range.
}
\note{
If \emph{'scale'}, always adds \code{ggplot2::coord_equal()}.
}
