% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByCoords}
\alias{ggpLayerFrameByImage}
\title{Set plot limits}
\usage{
ggpLayerFrameByCoords(object = "object", opt = "scale")

ggpLayerFrameByImage(object = "object", opt = "scale")
}
\arguments{
\item{object}{A valid spata-object.}

\item{opt}{Character value. Either \emph{'scale'} or \emph{'coords'}. If \emph{'scale'},
Depending on the input either functions \code{scale_x/y_continuous()} or
\code{coord_cartesian()} is used.}
}
\value{
List.
}
\description{
Sets the limits on the x- and y-axis of a ggplot based on the coordinate
range or the image range.
}
\note{
Always adds \code{ggplot2::coord_equal()}.
}
