% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerGridVisiumHD}
\alias{ggpLayerGridVisiumHD}
\title{Add a grid to VisiumHD surface plots}
\usage{
ggpLayerGridVisiumHD(
  object,
  res,
  line_alpha = 0.9,
  line_clr = "black",
  line_size = 0.5,
  img_name = activeImage(object)
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{res}{\link[=concept_distance_measure]{Distance measure.} A value specifying the desired resolution for the grid.
This resolution must be lower than or equal to the current resolution and divisible by the current resolution.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a grid overlay to VisiumHD spatial data at a specified resolution.
This function is designed to work specifically with VisiumHD data within a \code{SPATA2} object.
}
