% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerGroupOutline}
\alias{ggpLayerGroupOutline}
\title{Add group outline}
\usage{
ggpLayerGroupOutline(
  object,
  grouping,
  groups_subset = NULL,
  plot_type = "surface",
  line_color = "black",
  line_size = 1,
  outlier_rm = TRUE,
  minPts = 1,
  expand = getCCD(object, unit = "px"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{groups_subset}{Character value or NULL. If character,
specifies the exact groups that are encircled. If NULL, all groups
are encircled.}

\item{plot_type}{Character value. Either \emph{'surface', 'tsne'} or
\emph{'umap'}.}

\item{expand}{Specifies image expansion. An image that is cropped based on an image
annotation centers around the image annotation. If \code{expand = 0}, the default, the dimensions of the image,
that is width/x-axis and height/y-axis, are set to include only the image annotation area
and nothing more. Using \code{expand}, the cropped image section can be adjusted. See section
\emph{Expansion of cropped image sections} for more information.}

\item{...}{Additional arguments given to \code{ggforce::geom_mark_hull()}. Affects
the encircling.}

\item{grouping_variable}{Character value. The grouping variable of choice.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Highlights groups of barcode-spots by encircling them.
Depending on the \code{plot_type} this can be added to a surface plot
or a dimensional reduction plot.
}
