% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerGroupOutline}
\alias{ggpLayerGroupOutline}
\title{Add group outline}
\usage{
ggpLayerGroupOutline(
  object,
  grouping,
  groups_subset = NULL,
  plot_type = "surface",
  line_color = "black",
  line_size = 1,
  alpha = 0,
  bcsp_rm = character(0),
  outlier_rm = TRUE,
  eps = (getCCD(object, "px") * 1.25),
  minPts = 3,
  concavity = NULL,
  expand_outline = getCCD(object, "px") * 1.1,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{groups_subset}{Character value or NULL. If character,
specifies the exact groups that are encircled. If NULL, all groups
are encircled.}

\item{plot_type}{Character value. Either \emph{'surface', 'tsne'} or
\emph{'umap'}.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{bcsp_rm}{Character vector or \code{NULL.} If character, specifies barcode-spots that
are removed before analysis or plotting.}

\item{outlier_rm, minPts}{Logical. If \code{TRUE}, spatial outlier of the group to outline
are removed from the outline via \verb{dbscan::dbscan(..., minPts = minPts}). Ignored
if \code{plot_type} is not \emph{'surface'}.}

\item{concavity}{Given to \code{concavity} of \code{ggforce::geom_mark_hull()}.
Can be optimized for every sample via \code{object <- setDefault(object, concavity = ...)}.
Default by object initiation is 1.}

\item{expand_outline}{Distance measure. Determines the amount of
buffering applied to the outline.}

\item{...}{Additional arguments given to \code{ggforce::geom_mark_hull()}. Affects
the encircling.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Highlights groups of barcode-spots by encircling them.
Depending on the \code{plot_type} this can be added to a surface plot
or a dimensional reduction plot.
}
\examples{

 object <- downloadPubExample("269_T")

 plotImageGgplot(object) +
  ggpLayerGroupOutline(
    object = object,
    plot_type = "surface",
    grouping = "histology",
    groups_subset = "tumor",
    line_color = color_vector("npg")[1]
    )

}
