% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerGroupOutline}
\alias{ggpLayerGroupOutline}
\title{Add group outline}
\usage{
ggpLayerGroupOutline(
  object,
  grouping,
  group_subset = NULL,
  plot_type = "surface",
  line_alpha = 1,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  alpha = 0,
  fill = NA,
  incl_edge = FALSE,
  merdge_edge = FALSE,
  incr_vert = FALSE,
  bcsp_rm = character(0),
  outlier_rm = TRUE,
  eps = recDbscanEps(object),
  minPts = recDbscanMinPts(object),
  concavity = NULL,
  expand_outline = 0,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{group_subset}{Character value or \code{NULL}. If character,
specifies the exact groups that are outlined. If \code{NULL}, all groups
are encircled.}

\item{plot_type}{Character value. Either \emph{'surface'} or the name of a conducted
dimensional reduction.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{incl_edge}{Logical. If \code{TRUE}, include tissue section outline. See examples of \code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}}.}

\item{incr_vert}{Logical value. If \code{TRUE}, the number of vertices of the polygon
to be plotted is increased by linear interpolation. This can lead to smoother
visualization but also to increased runtime.}

\item{bcsp_rm}{Character vector or \code{NULL.} If character, specifies barcode-spots that
are removed before analysis or plotting. (Deprecated in favor of \code{bcs_rm}).}

\item{outlier_rm, minPts}{Logical. If \code{TRUE}, spatial outlier of the group to outline
are removed from the outline via \verb{dbscan::dbscan(..., minPts = minPts}). Ignored
if \code{plot_type} is not \emph{'surface'}.}

\item{eps}{Distance measure. Given to \code{eps} of \code{\link[dbscan:dbscan]{dbscan::dbscan()}}. Determines
the size (radius) of the epsilon neighborhood.}

\item{concavity}{Numeric value. Given to argument \code{concavity} of
\code{\link[concaveman:concaveman]{concaveman::concaveman()}}. Determines the relative measure of concavity.
1 results in a relatively detailed shape, Infinity results in a convex hull.
You can use values lower than 1, but they can produce pretty crazy shapes.}

\item{expand_outline}{Distance measure with which to expand the outline. Must be
provided in pixel units!}

\item{...}{Additional arguments given to \code{ggforce::geom_polygon()}. Affects
the encircling.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Highlights groups of barcode-spots by encircling them.
Depending on the \code{plot_type} this can be added to a surface plot
or a dimensional reduction plot.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- loadExampleObject("UKF275T", meta = TRUE)

plotSurface(object, color_by = "bayes_space") +
 ggpLayerGroupOutline(object, grouping = "bayes_space", group_subset = c("B1", "B2"))


}
