% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerHorizonSAS}
\alias{ggpLayerHorizonSAS}
\title{Add IAS area horizon}
\usage{
ggpLayerHorizonSAS(
  object,
  id,
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  alpha_core = 0,
  fill_core = NA,
  line_color = "black",
  line_size = (line_size_core * 0.75),
  line_size_core = 1,
  inc_outline = TRUE,
  direction = "outwards",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{n_bins_dist}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}. Not required if \code{distance} and \code{binwidth} is
specified. (See details for more.)}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. The size with which to display encircling lines
of the area expansion.}

\item{line_size_core}{Numeric. The size with which to display the core outline
of the spatial annotation.}

\item{inc_outline}{Logical. If \code{TRUE}, makes use of \code{SPATA2} automatic tissue outline algorithm.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds the last circular expansion used by the IAS-algorithm
of the area of  an spatial annotation to a surface plot in order to
visualize the border between screened tissue and everything beyond that
is not included in the screening.
}
\examples{

 object <- downloadSpataObject("313_T")

 object <-
  setImageAnnotation(
   object = object,
   img_ann = image_annotations$`313_T`$necrotic_center
   )

 plotSurface(object) +
  ggpLayerHorizonSAS(
   object = object,
   id = "necrotic_center",
   distance = "2.25mm",
   binwidth = "112.5um"
   )


}
