% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerImage}
\alias{ggpLayerImage}
\title{Add histology image}
\usage{
ggpLayerImage(object = "object")
}
\arguments{
\item{object}{An object of class \code{spata2}.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Creates ggplot2 layer with the histology image
as a raster annotation.
}
\note{
The returned list contains an additional \code{ggplot2::geom_point()}
layer with invisible barcode spots coordinates (\code{alpha} = 0) to enable the
image plotting.
}
