% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerImageAnnotation}
\alias{ggpLayerImageAnnotation}
\title{Add borders of annotated structures}
\usage{
ggpLayerImageAnnotation(
  object = "object",
  ids = NULL,
  tags = NULL,
  test = "any",
  alpha = 0.5,
  fill = NA,
  line_color = "black",
  line_size = 1.5,
  line_type = "solid",
  display_color = FALSE,
  clrp = NULL,
  clrp_adjust = NULL,
  ...
)
}
\arguments{
\item{alpha, size}{Numeric values. Given to \code{ggplot2::geom_polygon()}.}

\item{...}{Additional arguments given to \code{scale_color_add_on()}. Used to
set the color adjustments of the polygon (fill and color).}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds ggplot2 layer of polygons of structures that were annotated within the image
with \code{createImageAnnotations()}.
}
\note{
Adds two additional layers to set the scales for the color- and
fill aesthetic of the plot.
}
