% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-layer.R
\name{ggpLayerImageAnnotation}
\alias{ggpLayerImageAnnotation}
\title{Add polygons of annotated structures}
\usage{
ggpLayerImageAnnotation(
  object = "object",
  ids = NULL,
  tags = NULL,
  test = "any",
  alpha = 0.5,
  size = 1.5,
  linetype = "solid",
  clrp = NULL,
  clrp_adjust = NULL,
  ...
)
}
\arguments{
\item{alpha, size}{Numeric values. Given to \code{ggplot2::geom_polygon()}.}

\item{...}{Additional arguments given to \code{scale_color_add_on()}. Used to
set the color adjustments of the polygon (fill and color).}
}
\value{
List.
}
\description{
Adds ggplot2 layer of polygons of structures that were annotated within the image
with \code{annotateImage()}.
}
\details{
How to use arguments \code{tags} and \code{test} to specify
the image annotations of interest: Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are three options:

1. Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.

2. Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags.

3. Argument \code{test} set to \emph{'identical'} or \emph{3}: To be indluded, an image annotation
must be tagged with all of the input tags and must not be tagged with anything else.

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.
}
\note{
Adds two additional layers to set the scales for the color- and
fill aesthetic of the plot.
}
