% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerLineplotAid}
\alias{ggpLayerLineplotAid}
\title{Add horizontal and vertical lines}
\usage{
ggpLayerLineplotAid(object, xi, yi = 0.5, l = NULL, id = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{xi}{Distance measures of where to add vertical lines. Intercepts on x-axis.}

\item{yi}{Expression values of where to add horizontal lines. Intercepts on y-axis.}

\item{...}{Additional arguments given to \code{ggplot2::geom_h/vline()}}
}
\description{
Add horizontal and vertical lines
}
