% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerRect}
\alias{ggpLayerRect}
\title{Add a rectangular to the plot}
\usage{
ggpLayerRect(
  object = "object",
  xrange,
  yrange,
  alpha = 0,
  color = "black",
  size = 1,
  expand = 0,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{xrange, yrange}{Vector of length two. Specifies the x- and y-range
of the rectangle. E.g. \code{xrange = c(200, 500)} results in a rectangle
that ranges from 200px to 500px on the x-axis.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{alpha, color, fill, size}{Given to \code{ggplot2::geom_rect()}.}

\item{expand}{Vector of length one or two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the x-span and the second
value is used to adjust the y-span E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original x-span to the x-span of the image section and does not adjust
the y-span.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{...}{Additional arguments given to \code{ggplot2::geom_rect()}.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Adds a rectangular to the plot.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
