% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerSampleMask}
\alias{ggpLayerSampleMask}
\title{Add a hull that enricles the sample}
\usage{
ggpLayerSampleMask(
  object,
  line_color = "black",
  line_size = 0.5,
  expand = ggplot2::unit(2.25, "cm"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{expand}{Given to \code{ggforce::geom_mark_hull()}.}

\item{...}{Additional arguments given to \code{ggforce::geom_mark_hull()}}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a hull that encircles the sample. Usefull, if you want
to plot numeric variables by color against white.
}
