% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerSpatAnnOutline}
\alias{ggpLayerSpatAnnOutline}
\title{Add outline of spatial annotations}
\usage{
ggpLayerSpatAnnOutline(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  alpha = 0.5,
  fill = NA,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  use_colors = FALSE,
  clrp = NULL,
  clrp_adjust = NULL,
  inner = TRUE,
  incl_edge = FALSE,
  merge_edge = FALSE,
  incr_vert = FALSE,
  expand_outline = 0,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{tags}{Character vector or \code{NULL}. If character, the tags for the image annotation
selection. See section \emph{Selection of spatial annotations} for more information.}

\item{test}{Character value. One of \emph{c('any'. 'all', 'identical', 'not_identical', 'none')}.
Specifies how input for \code{tags} is used to select spatial annotations.
See section \emph{Selection of spatial annotations} for more information.}

\item{fill}{Character value or NA. If character, specifies the color with which
the outline of the spatial annotation is filled.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{use_colors}{Logical value. If \code{TRUE}, the color aesthetic is used to display
each outline in a different color while providing a legend.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{inner}{Logical value. If \code{FALSE}, only outer borders of the annotation
are displayed.}

\item{incl_edge}{Logical value. If \code{TRUE}, the function includes the tissue edge
(see \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}) in the visualization and removes the parts of
the outline that transgress the tissues edge.}

\item{merge_edge}{Logical value. If \code{incl_edge = TRUE} the outline of the
tissue edge is used to replace the part of the annotation outline that was
removed due to crossing the tissue edge.}

\item{incr_vert}{Logical value. If \code{TRUE}, the number of vertices of the polygon
to be plotted is increased by linear interpolation. This can lead to smoother
visualization but also to increased runtime.}

\item{expand_outline}{Distance measure by which the outline of the area is expanded.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a ggplot2 layer of polygons visualizing the outline
of spatial annotations.
}
\note{
Adds two additional layers to set the scales for the color- and
fill aesthetic of the plot.

\code{expand_outline} only works if \code{inner} is FALSE (or the spatial annotation
does not contain any inner borders).
}
\examples{

library(SPATA2)
library(patchwork)

object <- loadExampleObject("UKF275T")

object <-
 createNumericAnnotations(
   object = object,
   variable = "HM_HYPOXIA",
   threshold = "kmeans_high",
   id = "hypoxia_ann",
   inner_borders = FALSE,
   force1 = TRUE
   )

# range of hypoxia annotation
hr <- getSpatAnnRange(object, id = "hypoxia_ann")

hr_ggpL <- ggpLayerSpatAnnOutline(object, ids = "hypoxia_ann")

plotSurface(object, "HM_HYPOXIA") + hr_ggpL

p_rect <-
 plotImage(object) +
 hr_ggpL +
 ggpLayerRect(object, xrange = hr$x, yrange = hr$y)

p_zoomed <-
 plotImage(object) +
 ggpLayerZoom(object, xrange = hr$x, yrange = hr$y)

p_rect + p_zoomed

}
\seealso{
\code{\link[=ggpLayerSpatAnnPointer]{ggpLayerSpatAnnPointer()}}
}
