% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerThemeCoords}
\alias{ggpLayerThemeCoords}
\title{Add coordinates theme}
\usage{
ggpLayerThemeCoords(unit = NULL)
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a theme to the plot that displays the coordinates of
the tissue.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- loadExampleObject("UKF275T")

plotSurface(object) + ggpLayerThemeCoords()

}
