% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerThemeCoords}
\alias{ggpLayerThemeCoords}
\title{Add coordinates theme}
\usage{
ggpLayerThemeCoords()
}
\value{
List.
}
\description{
Adds a theme to the plot that displays the coordinates of
the tissue.
}
