% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerTissueOutline}
\alias{ggpLayerTissueOutline}
\title{Add a hull that outlines the tissue}
\usage{
ggpLayerTissueOutline(
  object,
  line_color = "grey",
  line_size = 0.5,
  expand_outline = getCCD(object, "px") * 1.25,
  concavity = NULL,
  inc_outline = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{expand_outline}{Distance measure by which the outline of the area is expanded.}

\item{concavity}{Given to \code{concavity} of \code{ggforce::geom_mark_hull()}.
Can be optimized for every sample via \code{object <- setDefault(object, concavity = ...)}.
Default by object initiation is 1.}

\item{inc_outline}{Logical. If \code{TRUE}, include tissue section outline. See examples of \code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}}.}

\item{...}{Additional arguments given to \code{ggforce::geom_mark_hull()}}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a hull that encircles the sample. Useful, if you want
to plot numeric variables by color against white.
}
\section{Distance measures}{


Several functions in \code{SPATA2} have arguments that take \emph{distance input}.
To specifically refer to a distance the unit must be specified. There are
three ways to create valid input for these arguments.

\strong{1. In pixel:}

There are two valid input options to specify the distance in pixel:

\itemize{
\item{numeric:}{ Single numeric values, e.g. \code{arg_input = c(2, 3.554, 69, 100.67)}. If no unit
is specified the input will be interpreted as pixels.}
\item{character:}{ Suffixed with \emph{'px'}, e.g. \code{arg_input = c('2px', '3.554px', '69px', '100.67px')}}
}

Note: The unit pixel (px) is used for distances as well as for areas. If pixel
refers to a distance the pixel side length is meant. If pixel refers to an area the
number of pixels is meant.

\strong{2. According to the Systeme international d`unites (SI):}

Specifying distances in SI units e.g. \code{arg_input = c('2mm', '4mm')} etc.
requires the input to be a character as the unit must be provided as suffix.
Between the numeric value and the unit must be no empty space! Valid suffixes
can be obtained using the function \code{validUnitsOfLengthSI()}.

\strong{3. As vectors of class \code{unit}:}

Behind the scenes \code{SPATA2} works with the \code{units} package. Input
is converted into vectors of class \code{units}. Therefore, input can be directly
provided this way: \code{arg_input = units::set_unit(x = c(2,4), value = 'mm')}
Note that \emph{pixel} is not a valid unit in the \code{units} package. If you want
to specify the input in pixel you have to use input option 1. In pixel.
}

\examples{

object <- donwloadPubExample("MCD_LMU")

plotImageGgplot(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = TRUE)

plotImageGgplot(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = FALSE)


}
