% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{ggpLayerTissueOutline}
\alias{ggpLayerTissueOutline}
\alias{ggpLayerTissueOutline,spata2-method}
\alias{ggpLayerTissueOutline,HistoImaging-method}
\alias{ggpLayerTissueOutline,HistoImage-method}
\title{Add a hull that outlines the tissue}
\usage{
ggpLayerTissueOutline(object, ...)

\S4method{ggpLayerTissueOutline}{spata2}(
  object,
  img_name = NULL,
  fragments = FALSE,
  line_alpha = 1,
  line_color = "grey",
  line_size = 0.5,
  line_type = "solid",
  transform = TRUE,
  smooth_with = "chaikin",
  scale_fct = 1,
  expand_outline = getCCD(object, "px") * 1.25,
  ...
)

\S4method{ggpLayerTissueOutline}{HistoImaging}(
  object,
  img_name = NULL,
  by_section = TRUE,
  fragments = FALSE,
  fill = NA,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  transform = TRUE,
  smooth_with = "chaikin",
  scale_fct = 1,
  expand_outline = 0.5,
  ...
)

\S4method{ggpLayerTissueOutline}{HistoImage}(
  object,
  by_section = TRUE,
  fragments = FALSE,
  fill = NA,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  transform = TRUE,
  smooth_with = "chaikin",
  scale_fct = 1,
  expand_outline = 0.5,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{smooth_with}{Character vaule. Sets the method with which to smooth
the tissue outline polygon. One of \code{c("chaikin", "densify", "ksmooth", "spline", "none")}.
If \emph{'none'}, no smoothing is conducted.}

\item{expand_outline}{Distance measure. Determines the amount of
buffering applied to the outline.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a hull that outlines the tissue.
}
\examples{

object <- download("MCD_LMU")

plotImage(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = TRUE)

plotImage(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = FALSE)

}
