% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerTissueOutline}
\alias{ggpLayerTissueOutline}
\alias{ggpLayerTissueOutline,spata2-method}
\alias{ggpLayerTissueOutline,HistoImaging-method}
\alias{ggpLayerTissueOutline,HistoImage-method}
\alias{ggpLayerTissueOutline,data.frame-method}
\title{Add a hull that outlines the tissue}
\usage{
ggpLayerTissueOutline(object, ...)

\S4method{ggpLayerTissueOutline}{spata2}(
  object,
  method = "image",
  img_name = NULL,
  by_section = TRUE,
  fragments = FALSE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  outline_fct = c(1.75, 2.75),
  use_scattermore = FALSE,
  transform = TRUE,
  scale_fct = 1,
  expand_outline = recBinwidth(object, "px")/1.25,
  ...
)

\S4method{ggpLayerTissueOutline}{HistoImaging}(
  object,
  method = "image",
  img_name = NULL,
  by_section = TRUE,
  fragments = FALSE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  transform = TRUE,
  outline_fct = c(1.75, 2.75),
  use_scattermore = FALSE,
  scale_fct = 1,
  expand_outline = 0,
  bcs_rm = character(0),
  ...
)

\S4method{ggpLayerTissueOutline}{HistoImage}(
  object,
  by_section = TRUE,
  fragments = FALSE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  transform = TRUE,
  smooth_with = "chaikin",
  scale_fct = 1,
  expand_outline = 0,
  ...
)

\S4method{ggpLayerTissueOutline}{data.frame}(
  object,
  method = "coords",
  by_section = TRUE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  outline_fct = c(1.75, 2.75),
  use_scattermore = FALSE,
  expand_outline = 0,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Additional arguments given to \code{ggforce::geom_mark_hull()}}

\item{method}{Character value. One of \code{c("coords", "image")}. If \emph{'coords'},
the outline is computed based on the coordinate position of the plotted data points
(cells, spots etc.). If \emph{'image'}, the outline is plotted solely based on
the image analysis results.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{use_scattermore}{Logical value. If \code{TRUE}, data points are plotted with
\code{scattermore::geom_scattermore()} which allows quick plotting of several
thousand data points. If the number of data points plotted is bigger than
10.000 it is used anyway.}

\item{expand_outline}{Distance measure with which to expand the outline. Must be
provided in pixel units!}

\item{bcs_rm}{Character vector or \code{NULL}. If character, specifies the observations
to be removed prior to analysis or visualization by their barcode.}

\item{smooth_with}{Character vaule. Sets the method with which to smooth
the tissue outline polygon. One of \code{c("chaikin", "densify", "ksmooth", "spline", "none")}.
If \emph{'none'}, no smoothing is conducted.}

\item{inc_outline}{Logical. If \code{TRUE}, include tissue section outline. See examples of \code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}}.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds a hull that outlines the tissue.
}
\examples{

object <- download("MCD_LMU")

plotImage(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = TRUE)

plotImage(object, unit = "mm") +
 ggpLayerTissueOutline(object, inc_outline = FALSE)

}
\seealso{
\code{\link[=identifyPixelContent]{identifyPixelContent()}},\code{\link[=identifyTissueOutline]{identifyTissueOutline()}},\code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}
}
