% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_a-i.R
\name{ggpLayerTrajectories}
\alias{ggpLayerTrajectories}
\title{Deprecated}
\usage{
ggpLayerTrajectories(...)
}
\description{
Deprecated in favor of \code{\link[=ggpLayerSpatialTrajectories]{ggpLayerSpatialTrajectories()}}.
}
\keyword{internal}
