% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerTrajectories}
\alias{ggpLayerTrajectories}
\title{Add trajectory layer}
\usage{
ggpLayerTrajectories(
  object = "object",
  ids,
  arrow = ggplot2::arrow(length = ggplot2::unit(x = 0.125, "inches")),
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{arrow}{A list of arguments given to \code{ggplot2::arrow()}. Based
on which the trajectories are plotted.}

\item{...}{Additional arguments given to \code{ggplot2::geom_segment()}.}

\item{trajectories}{Character vector. The name of the trajectories
that should be plotted.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Adds trajectories in form of arrows to a surface plot.
}
