% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerZoom}
\alias{ggpLayerZoom}
\title{Set plot limits manually}
\usage{
ggpLayerZoom(xrange = NULL, yrange = NULL)
}
\arguments{
\item{xrange, yrange}{Vector of length two. Specifies the x- and y-range
of zooming. E.g. \code{xrange = c(200, 500)} results in the plot
being cropped from x-coordinate 200px up to x-coordinate 500px.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}
}
\value{
A list of \code{ggproto}-objects.
}
\description{
Sets the limits on the x- and y-axis of a ggplot based on
manual input.
}
\details{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
