% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{ggpLayerZoom}
\alias{ggpLayerZoom}
\title{Set plot limits manually}
\usage{
ggpLayerZoom(
  object = NULL,
  xrange = NULL,
  yrange = NULL,
  expand_x = c(0, 0),
  expand_y = c(0, 0),
  round = 2,
  n_breaks = 5
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{xrange, yrange}{Vector of length two. Specifies the x- and y-range
of zooming. E.g. \code{xrange = c(200, 500)} results in the plot
being cropped from x-coordinate 200px up to x-coordinate 500px.

This argument works within the \code{SPATA2} distance framework.
If values are specified in SI units of length the input is
immediately converted to pixel units.}

\item{expand_x, expand_y}{Given to \code{expand} of \code{ggplot2::scale_x/y_continuous()}.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}
}
\value{
\code{ggpLayer*()}-functions return lists of \code{ggproto} objects
that can be added to ggplots via the \code{+} operator. In most of the cases
they are supposed to be added to plots created with the \code{plotSurface*()}
family.
}
\description{
Sets the limits on the x- and y-axis of a ggplot based on
manual input.
}
