% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{HistoImage-class}
\alias{HistoImage-class}
\alias{HistoImage}
\title{The \code{HistoImage} - Class}
\description{
S4 class that contains an histology image and information and data
about it. Usually live in slots @image_reference and @images_registered of \code{\link{HistoImaging}}
objects.
}
\section{Slots}{

\describe{
\item{\code{active}}{logical. If \code{TRUE}, it is the image used by default. Only one \code{\link{HistoImage}}
in the \code{\link{HistoImaging}} object can be the active at a time.}

\item{\code{aligned}}{logical. If \code{TRUE}, indicates that the image was aligned to
the reference image.}

\item{\code{bg_color}}{character. The color of the background.}

\item{\code{dir}}{character. The directory from where to load the image if slot @image is empty.}

\item{\code{image}}{Image. The image stored as class \code{Image} from the package \code{EBImage}.}

\item{\code{image_info}}{list. A list of miscellaneous slots that carrie information
regarding the image. Protected slots are:
\itemize{
\item{\emph{dims}:}{ Numeric vector of length three. Output of \code{base::dim()} on the read image.}
}}

\item{\code{name}}{character. The name of the image.}

\item{\code{outline}}{list. List of two data.frames in which each row corresponds to
a vertice of the polygon required to outline the whole tissue identified on
the image or single contiguous tissue sections.
\itemize{
\item{\emph{tissue_whole}:}{ Data.frame of two variables \emph{x} and \emph{y}.}
\item{\emph{tissue_sections}:} {Data.frame of two variables \emph{x}, \emph{y} and \emph{section} to
outline the tissue section outlined.}
}}

\item{\code{overlap}}{numeric. Numeric vector of length two. Quantifies the overlap
of the tissue outline of this image with the tissue outline of the reference image
with a value between 0-1 before and after alignment via \code{\link[=alignImage]{alignImage()}}.}

\item{\code{pixel_content}}{factor. Named factor where names correspond to the
pixels following the naming convention 'px_w1_h1' encoding the width and height value
of the pixel from the original \strong{not transformed} image. Values correspond to
the content the pixel displays. See \code{\link[=identifyPixelContent]{identifyPixelContent()}} for more information.}

\item{\code{reference}}{logical. \code{TRUE} if it is the \code{\link{HistoImage}} used as the reference
with which other histology images are aligned.}

\item{\code{sample}}{character. The name of the tissue portion to which this image belongs.}

\item{\code{scale_factors}}{list. List of single numeric values serving as scale factors for
multiple functionalities. Reserved slot names:
\itemize{
\item{\emph{coords}:} {Coordinate scale factor to be multiplied by the original x and y variables (\emph{x_orig}, \emph{y_orig}) upon
extraction of the coordinatse data.frame (resulting in the \emph{x} and \emph{y} variables) ensuring alignment with the image.}
\item{\emph{pixel}:} {Pixel scale factor used to convert pixel values into SI units. It must have an
attribute called "unit" conforming to the format "SI-unit/px".}
}}

\item{\code{transformations}}{list. List of transformations to apply upon extracting
the image to ensure alignment with additional images. In case of default values
no transformation is applied.
\itemize{
\item{\emph{angle}:}{ Numeric value that ranges from 0-359. Indicates the angle in degrees
by which the image needs to be rotated in \strong{clockwise} direction. Defaults to 0.}
\item{\emph{flip}:}{ List of two logical values named \emph{horizontal} and \emph{vertical}. Indicate
if the image is supposed to be flipped around either axis. Both default to \code{FALSE}}
\item{\emph{stretch}:}{ List of two numerical values named \emph{horizontal} and \emph{vertical} serving as scale factors
with which to stretch the respective axis. Both default to 1.}
\item{\emph{translate}:}{ List of two numeric values named \emph{horizontal} and \emph{vertical}. Indicate
the number of pixels the image needs to be translated. Positive values shift the image
\strong{downwards} or to the right, respectively. Negative values shift the image \strong{upwards}
or to the left, respectively. Both default to 0.}
}}
}}

