% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{HistoImaging-class}
\alias{HistoImaging-class}
\alias{HistoImaging}
\title{The \code{HistoImaging} - Class}
\description{
Deprecated in favor of \code{SpatialData}.
}
\keyword{internal}
