% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{HistologyImaging-class}
\alias{HistologyImaging-class}
\alias{HistologyImaging}
\title{The \code{HistologyImaging} - Class}
\description{
S4 class that represents a set of histological images from one
and the same tissue slide.
}
\section{Slots}{

\describe{
\item{\code{annotations}}{list. List of objects of class \code{ImageAnnotation}.}

\item{\code{coordinates}}{data.frame. A data.frame of observational units that underlie
the image in case of spatially resolved multi-omic studies. Should contain at least
the  two variables: \emph{x}, \emph{y} and a variable that identifies the observational
units (e.g. \emph{barcodes}).}

\item{\code{dir_add}}{list. Named list of directories that contain different versions
of tissue images. Can be arbitrarily expanded for convenient exchanging via
\code{loadImage()}.}

\item{\code{dir_default}}{character. Directory that leads to the default image for save
exchanging via \code{loadDefaultImage()}.}

\item{\code{dir_highres}}{character. Directory that leads to a high resolution version of the image
for save exchanging via \code{loadHighresImage()}.}

\item{\code{dir_lowres}}{character. Directory that leads to a low resolution version of the image
for save exchanging via \code{loadLowresImage()}.}

\item{\code{grid}}{list. That contains information about spatial grids.}

\item{\code{id}}{character. String to identify the imaged tissue.}

\item{\code{image}}{Image. Should be compatible with the \code{EBImage} package.}

\item{\code{image_info}}{list. Stores meta data and miscellaneous information regarding the
image that is currently stored in slot @image. Slots that should always exist:
\itemize{
\item{\emph{origin}:}{ Character string. Either the directory from where the current image was read
in or a substitute of the object name that was used from the global environment.}
\item{\emph{dim_input}:}{ The dimensions with which the image was given to argument \code{image} of
\code{createHistologyImaging()} or \code{exchangeImage()}.}
\item{\emph{dim_stored}:}{ The dimensions with which the image is currently stored.}
\item{\emph{img_scale_fct}:}{ The scale factor input that was used to resize the current image within
\code{createHistologyImaging()} or \code{exchangeImage()} before setting it. If 1, \emph{dim_stored} and \emph{dim_input}
should be identical. See argument \code{scale_fct} of \code{exchangeImage()} for more details on its interpretation.}
\item{\emph{pxl_scale_fct}:}{ Numeric value that gives the side length of one pixel an SI unit that is stated
in an attribute called \emph{unit} as nSI-units/px.}
}}

\item{\code{justification}}{list. List of two slots that track justification changes. See corresponding
section below the slot descriptions for more information.
\itemize{
\item{\emph{angle}:}{ Numeric value that ranges from 0-359.}
\item{\emph{flipped}:}{ List of two logical values named \emph{horizontal} and \emph{vertical}.}
}}

\item{\code{meta}}{list. List for meta data regarding the tissue.}

\item{\code{misc}}{list. A flexible list for miscellaneous input.}
}}

\section{Requirements}{

The \code{HistologyImaging} framework assumes that all read in images have the same
axes-ratio.
}

\section{Tracking changes in image justification}{

The histology image that is used while creating the object is considered the
default image. By default, the framework assumes that all related images (high resolution,
low resolution, fluorescent images, RAMAN spectroscopy etc.) have the same justification
in terms of angle rotation and axes-flipping. Flipping an image in the \code{SPATA2} object via
\code{flipImage()} or rotating images via \code{rotateImage()} changes their justification in space.
These changes in justification are tracked (if \code{track} is not set to \code{FALSE}) and applied
whenever an image is exchanged via \code{exchangeImage()} (if \code{adjust} is not set to \code{FALSE}).
This ensures consistent image exchanges using the different directories.
}

