% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_breaks}
\alias{hlpr_breaks}
\title{Calculates breaks for heatmap according to input matrix}
\usage{
hlpr_breaks(mtr, length_out)
}
\description{
Calculates breaks for heatmap according to input matrix
}
