% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_geom_trajectory_fit}
\alias{hlpr_geom_trajectory_fit}
\title{Easy switch between geom line and and geom smooth}
\usage{
hlpr_geom_trajectory_fit(smooth, smooth_span, plot_df)
}
\description{
To be used in plotTrajectoryFit()/-Customized()
}
