% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_image_add_on}
\alias{hlpr_image_add_on}
\alias{hlpr_image_add_on2}
\title{Provides the image as ggplot background}
\usage{
hlpr_image_add_on(object, display_image, of_sample)

hlpr_image_add_on2(image)
}
\arguments{
\item{object}{A valid spata-object.}

\item{display_image}{Logical value.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{image}{Image input.}
}
\value{
Either null or a ggplot2::geom_annotation_raster
}
\description{
Provides the image as ggplot background
}
