% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-analysis.R
\name{hlpr_rank_trajectory_trends}
\alias{hlpr_rank_trajectory_trends}
\alias{hlpr_rank_trajectory_trends_customized}
\title{Rank trajectory trends.}
\usage{
hlpr_rank_trajectory_trends(stdf, verbose = TRUE)

hlpr_rank_trajectory_trends_customized(
  stdf,
  verbose = TRUE,
  customized_trends_df
)
}
\arguments{
\item{stdf}{A summarized trajectory data.frame. (e.g. obtained by
\code{getTrajectoryDf()}).}
}
\value{
A nested data.frame with information about the dynamics of each gene
or gene set.
}
\description{
Analyze the expression dynamics along
a specified trajectory by fitting a variety of models to the genes or
gene sets expression trends.
}
