% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{hlpr_save_spata_object}
\alias{hlpr_save_spata_object}
\title{Save spata object inside functions}
\usage{
hlpr_save_spata_object(object, object_file, ref_step, verbose)
}
\arguments{
\item{object}{A valid spata-object.}

\item{object_file}{The directory under which to store the object.}

\item{ref_step}{Character value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Save spata object inside functions
}
