% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h.R
\name{hlpr_scatterplot}
\alias{hlpr_scatterplot}
\title{Color to + scatterplot helper}
\usage{
hlpr_scatterplot(
  object,
  spata_df,
  color_to,
  method_gs = "mean",
  display_title = FALSE,
  pt_size = 2,
  pt_alpha = 1,
  pt_clrsp = "inferno",
  pt_clrp = "milo",
  pt_clr = "black",
  smooth = FALSE,
  smooth_span = 0.02,
  normalize = TRUE,
  verbose = TRUE,
  complete = FALSE,
  ...
)
}
\arguments{
\item{color_to}{Named list.}
}
\value{
A named list. Slot \emph{data} contains a data.frame as input for \code{data} of \code{ggplot2::ggplot()} and
slot \emph{add_on} contains a list of ggplot-add-ons.
}
\description{
Color to + scatterplot helper
}
\keyword{internal}
