% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyBackgroundColor}
\alias{identifyBackgroundColor}
\alias{identifyBackgroundColor,SPATA2-method}
\alias{identifyBackgroundColor,SpatialData-method}
\alias{identifyBackgroundColor,HistoImage-method}
\title{Identifies the background color}
\usage{
identifyBackgroundColor(object, ...)

\S4method{identifyBackgroundColor}{SPATA2}(
  object,
  img_name = activeImage(object),
  verbose = NULL,
  ...
)

\S4method{identifyBackgroundColor}{SpatialData}(
  object,
  img_name = activeImage(object),
  verbose = TRUE,
  ...
)

\S4method{identifyBackgroundColor}{HistoImage}(object, verbose = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Identifies the background color based on the results
of \code{\link[=identifyPixelContent]{identifyPixelContent()}} by averaging the color values of
all pixels identified as background.
}
\keyword{internal}
