% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{identifyBackgroundColor}
\alias{identifyBackgroundColor}
\alias{identifyBackgroundColor,spata2-method}
\alias{identifyBackgroundColor,HistoImaging-method}
\alias{identifyBackgroundColor,HistoImage-method}
\title{Identifies the background color}
\usage{
identifyBackgroundColor(object, ...)

\S4method{identifyBackgroundColor}{spata2}(object, img_name = NULL, ...)

\S4method{identifyBackgroundColor}{HistoImaging}(object, img_name = NULL, ...)

\S4method{identifyBackgroundColor}{HistoImage}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\description{
Identifies the background color based on the results
of \code{\link[=identifyPixelContent]{identifyPixelContent()}} by averaging the color values of
all pixels identified as background.
}
