% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{identifyPixelContent}
\alias{identifyPixelContent}
\alias{identifyPixelContent,spata2-method}
\alias{identifyPixelContent,HistoImaging-method}
\alias{identifyPixelContent,HistoImage-method}
\alias{identifyPixelContent,Image-method}
\title{Identify pixel content}
\usage{
identifyPixelContent(object, ...)

\S4method{identifyPixelContent}{spata2}(
  object,
  img_name = NULL,
  percentile = 0,
  compactness_factor = 10,
  superpixel = 600,
  dbscan_eps = 0.005,
  dbscan_minPts = 0.005,
  frgmt_threshold = c(0.001, 0.05),
  verbose = TRUE
)

\S4method{identifyPixelContent}{HistoImaging}(
  object,
  img_name = NULL,
  percentile = 99,
  compactness_factor = 10,
  superpixel = 1000,
  dbscan_eps = 0.005,
  dbscan_minPts = 0.005,
  frgmt_threshold = c(0.001, 0.05),
  verbose = TRUE
)

\S4method{identifyPixelContent}{HistoImage}(
  object,
  percentile = 99,
  compactness_factor = 10,
  superpixel = 1000,
  dbscan_eps = 0.005,
  dbscan_minPts = 0.005,
  frgmt_threshold = c(0.001, 0.05),
  verbose = TRUE
)

\S4method{identifyPixelContent}{Image}(
  object,
  percentile = 99,
  compactness_factor = 10,
  superpixel = 1000,
  frgmt_threshold = c(0.001, 0.05),
  dbscan_eps = 0.005,
  dbscan_minPts = 0.005,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{percentile}{Numeric value between 0 and 100.
Specifies the percentile of colors to set to plain white, assuming that
this percentile of colors is responsible for the background.
If set to 0, the function is not called.}

\item{compactness_factor}{Numeric value controlling the compactness of superpixels.
Given as an argument to \verb{$spixel_segmentation()} function.}

\item{superpixel}{Numeric value specifying the number of superpixels to compute.
Given as an argument to \verb{$spixel_segmentation()} function.}

\item{dbscan_eps}{Numeric value specifying the value of \code{eps} parameter used in \code{dbscan::dbscan()}
when applied on the tissue pixels. If the value is less than 1, it is calculated
as a percentage of the width or height of the image, depending on which is larger.
If the value is greater than or equal to 1, it is taken as an absolute value.}

\item{dbscan_minPts}{Numeric value specifying the value of \code{minPts} parameter used in \code{dbscan::dbscan()}
when applied on the tissue pixels identified as potential tissue. If the value is less than 1,
it is calculated as a percentage of the width or height of the image, depending on which is larger.
If the value is greater than or equal to 1, it is taken as an absolute value.}

\item{frgmt_threshold}{Numeric vector of length 2 specifying the range of the number of pixels
an identified object must have to be considered a tissue fragment. Objects with a lower number
of pixels than the minimum threshold are considered artifacts, and objects with a higher number
of pixels than the maximum threshold are considered tissue sections. If a threshold value is less than 1,
it is calculated as a percentage of the total number of pixels in the image.
If a threshold value is greater than or equal to 1, it is taken as an absolute value.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The method for class \code{Image} returns a data.frame of the following
variables.

\itemize{
\item{\emph{pixel}:}{ character. Pixel index.}
\item{\emph{width}:}{ numeric. Pixel position on horizontal axis of the image.}
\item{\emph{height}:}{ numeric. Pixel position on the vertical axis of the image.}
\item{\emph{clusterK2}:}{ character. Either \emph{'background'} or \emph{'tissue'}.}
\item{\emph{colTiss#} :}{ numeric. Numeric variables that correspond to the color dimensions
of the image mask based on which the clustering of \emph{clusterK2} was conducted.}
\item{\emph{clusterDBSCAN}:}{ character. Cluster results of dbscan::dbscan() after removal
of background pixels.}
\item{\emph{clusterDBSCAN_size}:}{numeric. Size of each dbscan cluster.}
\item{\emph{content}:}{ character. The identified content of each pixel.}
}

Methods for S4-classes serving as containers return the input object with the
the results stored in the corresponding slots.
}
\description{
Determines the type of content displayed by each pixel in the image,
categorizing it as tissue from tissue segments or fragments, artifacts, or background.
}
\details{
If \code{img_name} specifies multiple images, the function
iterates over all of them. If it is \code{NULL} the active image is picked.
}
\seealso{
For subsequent image processing: \code{\link[=identifyTissueOutline]{identifyTissueOutline()}},\code{\link[=identifyBackgroundColor]{identifyBackgroundColor()}}.
For visualization of results: \code{\link[=plotImageMask]{plotImageMask()}}, \code{\link[=plotPixelContent]{plotPixelContent()}}.
For extraction of results: \code{\link[=getPixelDf]{getPixelDf()}}.
}
