% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyPixelContent}
\alias{identifyPixelContent}
\alias{identifyPixelContent,SPATA2-method}
\alias{identifyPixelContent,SpatialData-method}
\alias{identifyPixelContent,HistoImage-method}
\alias{identifyPixelContent,Image-method}
\title{Identify pixel content}
\usage{
identifyPixelContent(object, ...)

\S4method{identifyPixelContent}{SPATA2}(
  object,
  img_name = activeImage(object),
  percentile = 0,
  compactness_factor = 10,
  superpixel = 600,
  eps = 0.005,
  minPts = 0.005,
  frgmt_threshold = c(0.001, 0.05),
  verbose = TRUE
)

\S4method{identifyPixelContent}{SpatialData}(
  object,
  img_name = activeImage(object),
  percentile = 0,
  compactness_factor = 10,
  superpixel = 1000,
  eps = 0.005,
  minPts = 0.005,
  frgmt_threshold = c(0.001, 0.05),
  verbose = TRUE
)

\S4method{identifyPixelContent}{HistoImage}(
  object,
  percentile = 0,
  compactness_factor = 10,
  superpixel = 1000,
  eps = 0.005,
  minPts = 0.005,
  frgmt_threshold = c(0.001, 0.05),
  verbose = TRUE
)

\S4method{identifyPixelContent}{Image}(
  object,
  percentile = 0,
  compactness_factor = 10,
  superpixel = 1000,
  frgmt_threshold = c(0.001, 0.05),
  eps = 0.005,
  minPts = 0.005,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the \link[=concept_active]{active} image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{percentile}{Numeric value between 0 and 100 (inclusive) specifying the percentile
threshold for background color determination. If bigger than 0, it determines
the \strong{top} percentile of colors to identify as background based on the frequency
they appear in the image. This can improve identifying tissue pixels in images
where the edge between tissue and background is continuous rather than sharp
and thus difficult to identify using computational approaches. It follows the
hypothesis that the background consists of many pixels of equal color while
the tissue consists of pixels of heterogenous colors.

Values between 0-100 are valid. Usually values between 0.5-2.5 work well.
Test resuls with \code{\link[=plotPixelContent]{plotPixelContent()}}.}

\item{compactness_factor}{Numeric value controlling the compactness of superpixels.
Given as an argument to \verb{$spixel_segmentation()} function.}

\item{superpixel}{Numeric value specifying the number of superpixels to compute.
Given as an argument to \verb{$spixel_segmentation()} function. Increased values can
improve the output but increase runtime.}

\item{eps}{Numeric value specifying the value of \code{eps} parameter used in \code{dbscan::dbscan()}
when applied on the tissue pixels. If the value is less than 1, it is calculated
as a percentage of the width or height of the image, depending on which is larger.
If the value is greater than or equal to 1, it is taken as an absolute value.}

\item{minPts}{Numeric value specifying the value of \code{minPts} parameter used in \code{dbscan::dbscan()}
when applied on the tissue pixels identified as potential tissue. If the value is less than 1,
it is calculated as a percentage of the width or height of the image, depending on which is larger.
If the value is greater than or equal to 1, it is taken as an absolute value.}

\item{frgmt_threshold}{Numeric vector of length 2 specifying the range of the number of pixels
an identified non-background-object in the image must have to be considered a tissue fragment. Objects with a lower number
of pixels than the minimum threshold are considered artifacts, and objects with a higher number
of pixels than the maximum threshold are considered tissue sections. If a threshold value is less than 1,
it is calculated as a percentage of the total number of pixels in the image.
If a threshold value is greater than or equal to 1, it is taken as an absolute value.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The method for class \code{Image} returns a data.frame of the following
variables.

\itemize{
\item{\emph{pixel}:}{ character. Pixel index.}
\item{\emph{width}:}{ numeric. Pixel position on horizontal axis of the image.}
\item{\emph{height}:}{ numeric. Pixel position on the vertical axis of the image.}
\item{\emph{clusterK2}:}{ character. Either \emph{'background'} or \emph{'tissue'}.}
\item{\emph{colTiss#} :}{ numeric. Numeric variables that correspond to the color dimensions
of the image mask based on which the clustering of \emph{clusterK2} was conducted.}
\item{\emph{clusterDBSCAN}:}{ character. Cluster results of dbscan::dbscan() after removal
of background pixels.}
\item{\emph{clusterDBSCAN_size}:}{numeric. Size of each dbscan cluster.}
\item{\emph{content}:}{ character. The identified content of each pixel.}
}

Methods for S4-classes serving as containers return the input object with the
the results stored in the corresponding slots.
}
\description{
Determines the type of content displayed by each pixel in the image,
categorizing it as tissue from tissue segments or fragments, artifacts, or background.
}
\details{
If \code{percentile} is not 0, the \code{\link[=background_white]{background_white()}} function processes the
image by identifying the most frequently occurring colors setting their RGB values
to 1, effectively turning them white (assuming that the image is in grayscale,
with 1 representing white). The \code{percentile} parameter allows you to adjust the
sensitivity of the background removal, allowing for more precise isolation of the main subject.
}
\note{
If \code{img_name} specifies multiple images, the function
iterates over all of them. If it is \code{NULL} the active image is picked.
}
\seealso{
For subsequent image processing: \code{\link[=identifyTissueOutline]{identifyTissueOutline()}},\code{\link[=identifyBackgroundColor]{identifyBackgroundColor()}}.
For visualization of results: \code{\link[=plotImageMask]{plotImageMask()}}, \code{\link[=plotPixelContent]{plotPixelContent()}}.
For extraction of results: \code{\link[=getPixelDf]{getPixelDf()}}.
}
