% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifySpatialOutliers}
\alias{identifySpatialOutliers}
\alias{identifySpatialOutliers,SPATA2-method}
\alias{identifySpatialOutliers,SpatialData-method}
\title{Identify spatial outliers}
\usage{
identifySpatialOutliers(object, ...)

\S4method{identifySpatialOutliers}{SPATA2}(
  object,
  method,
  img_name = activeImage(object),
  buffer = NULL,
  eps = recDbscanEps(object),
  minPts = recDbscanMinPts(object),
  test = "any",
  verbose = NULL
)

\S4method{identifySpatialOutliers}{SpatialData}(
  object,
  method = c("outline", "dbscan"),
  img_name = activeImage(object),
  buffer = NULL,
  eps = NULL,
  minPts = 3,
  test = "any",
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{method}{Character vector. The method(s) to use. A combination of \emph{'outline'}
and/or \emph{'dbscan'}. See details for more.}

\item{img_name}{Character value. The name of the image whose tissue outline
is used if \code{method} contains \emph{'outline'}.}

\item{buffer}{Numeric value. Expands the tissue outline to include observations
that lie on the edge of the outline and are mistakenly removed.}

\item{eps, minPts}{Given to the corresponding arguments of
\code{\link[dbscan:dbscan]{dbscan::dbscan()}} if \code{method} contains \emph{'dbscan'}.}

\item{test}{Character value. Only required if \code{method = c('dbscan', 'outline')}. If
\emph{'any'}, spots are labeled as outliers if at least one method identifies them
as outliers. If \emph{'all'}, spots are labeled as outliers if both methods identify
them as outliers.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
Assigns data points to the tissue sections or
fragments they are located on or labels them as spatial outliers and saves
the results in a new variable of the coordinates data.frame called \emph{section}.
See details for more.
}
\details{
This function categorizes the data points of the object based on their spatial
proximity, grouping those that are close enough to be deemed part of a single
contiguous tissue section. Data points that are isolated and situated at a
significant distance from others are identified as spatial outliers.

The resulting classifications are saved in a \emph{section} variable within the
object's coordinates data.frame.

This function identifies spatial outliers using a combination of two methods:

Method \emph{outline}:
The \emph{outline} method involves the image based tissue outline from the
\code{identifyTissueOutline()} function. This function has created polygons that
outline the tissue or tissue sections identified in the image. For each data point,
the function checks which polygon it falls within and assigns it to the corresponding
group. If an observation does not fall within any of the tissue polygons, it is
considered a spatial outlier. As this method requires image processing steps, it does not
work for platforms that do not provide images of the analyzed tissue such as
\emph{MERFISH} or \emph{SlideSeq}.

Method \emph{dbscan}:
The \emph{dbscan} method applies the DBSCAN algorithm to the data points. Please
refer to the documentation of \code{dbscan::dbscan()} for a more detailed explanation.
The \code{eps} and \code{minPts} arguments are passed directly to the
corresponding arguments of the DBSCAN function.Data points that are not assigned
to any spatial cluster, indicated by being assigned to cluster 0, are considered
spatial outliers.

For objects derived from the Visium platform with a fixed center to center
distance, we recommend to set \code{eps = getCCD(object, unit = "px")*1.25}
and \code{minPts = 3} which has worked well for us. For objects derived
from platforms that do notrely on a fixed grid of data points (MERFISH, SlideSeq, etc.)
we recommend the average minimal distance between the data points times 10 for
\code{eps} and \code{minPts = 2}. The function
defaults to these recommendations using \code{\link[=recDbscanEps]{recDbscanEps()}} and \code{\link[=recDbscanMinPts]{recDbscanMinPts()}}
by default. This can, of course, be overwritten manually by the user by
specifying the parameters otherwise!

If \code{method = c('outline', 'dbscan')}, both algorithms are applied. Whether a
data point is considered a spatial outlier depends on the \code{test} argument:

\itemize{
\item{\code{test = 'any'}:} The data point is considered a spatial outlier if
either of the two tests classifies it as an outlier.
\item{\code{test = 'all'}:} The data point is considered a spatial outlier
only if both tests classify it as an outlier.
}

If \code{method = 'outline'} or \code{method = 'dbscan'} only one of the two
methods is applied. Note that for \code{method = 'outline'} the results from the
image processing pipeline must be available.

The results can be visualized using \code{plotSurface(object, color_by = "section")}.
In case of bad results the function can be run over and over again with
changing parameters as the results are simply overwritten.
}
\seealso{
\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}, \code{\link[=runImagePipeline]{runImagePipeline()}},
\code{\link[=mergeTissueSections]{mergeTissueSections()}}
}
