% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{identifyTissueOutline}
\alias{identifyTissueOutline}
\alias{identifyTissueOutline,spata2-method}
\alias{identifyTissueOutline,HistoImaging-method}
\alias{identifyTissueOutline,HistoImage-method}
\title{Identify tissue outline}
\usage{
identifyTissueOutline(object, ...)

\S4method{identifyTissueOutline}{spata2}(object, verbose = NULL)

\S4method{identifyTissueOutline}{HistoImaging}(
  object,
  name,
  use_greyscale = FALSE,
  frgmt_threshold = c(5e-04, 0.01),
  verbose = TRUE
)

\S4method{identifyTissueOutline}{HistoImage}(
  object,
  frgmt_threshold = c(5e-04, 0.01),
  use_greyscale = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated \code{spata2} object. The coordinates data.frame
as obtained by \code{getCoordsDf()} contains an additional, logical
variable named \emph{outline} indicating whether the spot belongs
to the outline spots of the respective tissue section indicated by
variable \emph{section}.
}
\description{
Identifies the barcode-spots that lie on the edge
of each tissue section and, thus, outline it. Requires \code{identifyTissueSections()}
results.
}
