% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{identifyTissueOutline}
\alias{identifyTissueOutline}
\alias{identifyTissueOutline,spata2-method}
\alias{identifyTissueOutline,HistoImaging-method}
\alias{identifyTissueOutline,HistoImage-method}
\title{Identify tissue outline}
\usage{
identifyTissueOutline(object, ...)

\S4method{identifyTissueOutline}{spata2}(object, verbose = NULL)

\S4method{identifyTissueOutline}{HistoImaging}(object, img_name, verbose = TRUE)

\S4method{identifyTissueOutline}{HistoImage}(object, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Identifies the outline of each tissue section on the \code{HistoImage}
as well as the outline of the whole tissue.
}
\note{
Requires results of \code{\link[=identifyPixelContent]{identifyPixelContent()}}
}
\seealso{
\code{\link[=excludeArtefacts]{excludeArtefacts()}}, \code{\link[=excludeTissueFragments]{excludeTissueFragments()}},
\code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}}, \code{\link[=ggpLayerTissueOutline]{ggpLayerTissueOutline()}}
}
