% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyTissueOutline}
\alias{identifyTissueOutline}
\alias{identifyTissueOutline,spata2-method}
\alias{identifyTissueOutline,HistoImaging-method}
\alias{identifyTissueOutline,HistoImage-method}
\title{Identify tissue outline}
\usage{
identifyTissueOutline(object, ...)

\S4method{identifyTissueOutline}{spata2}(object, img_name = NULL, use = "image", verbose = NULL)

\S4method{identifyTissueOutline}{HistoImaging}(object, img_name = NULL, verbose = TRUE)

\S4method{identifyTissueOutline}{HistoImage}(object, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{use}{Character value. Either \emph{'image'} (the default, which uses the
results of \code{\link[=identifyPixelContent]{identifyPixelContent()}}) or \emph{'coords'}. If
\code{\link[=identifyPixelContent]{identifyPixelContent()}} does not provide satisfying results, set \code{use = coords}. This way, the tissue
outline is created by using the coordinates data.frame.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
Identifies the outline of each tissue section on the image
as well as the outline of the whole tissue.
}
\details{
If \code{img_name} specifies multiple images, the function
iterates over all of them.
}
\note{
For \code{spata2} objects: If the \code{spata2} object contains a registered image
the results of \code{\link[=identifyPixelContent]{identifyPixelContent()}} is required.

If the \code{spata2} object does not contain a registered image because the
underlying spatial method does not come with an image (e.g. MERFISH, SlideSeq)
a workaround is applied and the tissue outline is identified by outlining all
data points instead of outlining pixels that were identified as \emph{tissue pixels}.
}
\seealso{
\code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}}, \code{\link[=ggpLayerTissueOutline]{ggpLayerTissueOutline()}}
}
