% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyTissueOutline}
\alias{identifyTissueOutline}
\alias{identifyTissueOutline,SPATA2-method}
\alias{identifyTissueOutline,SpatialData-method}
\alias{identifyTissueOutline,HistoImage-method}
\title{Identify tissue outline}
\usage{
identifyTissueOutline(object, ...)

\S4method{identifyTissueOutline}{SPATA2}(
  object,
  method,
  img_name = activeImage(object),
  verbose = NULL
)

\S4method{identifyTissueOutline}{SpatialData}(
  object,
  method,
  img_name = activeImage(object),
  verbose = TRUE
)

\S4method{identifyTissueOutline}{HistoImage}(object, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{method}{Character value. Defines the origin based on which the outline
is computed. Either \emph{'image'} or \emph{'obs'}.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Identifies the spatial boundaries of the tissue. Using \code{method},
the origin of the outline can be determined. See the vignette
on \link[=concept_tissue_outline]{tissue outlines} in SPATA2 for more information.
}
\details{
In case of \code{method = image} the object must contain an image named as
indicated by the input of argument \code{img_name}. Furthermore, the results
of \code{\link[=identifyPixelContent]{identifyPixelContent()}} for that image are required. If \code{img_name} specifies
multiple images, the function iterates over all of them. Since results of both methods
are stored in different locations, the object can contain results of both methods.
When extracting the tissue outline via \code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}} or \code{\link[=ggpLayerTissueOutline]{ggpLayerTissueOutline()}}
use argument \code{method} to decide on which results to use.
}
\seealso{
\code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}}, \code{\link[=ggpLayerTissueOutline]{ggpLayerTissueOutline()}}
}
