% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{identifyTissueSections}
\alias{identifyTissueSections}
\title{Identify tissue sections}
\usage{
identifyTissueSections(object, eps = getCCD(object, "px") * 1.25, minPts = 3)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{eps}{size (radius) of the epsilon neighborhood. Can be omitted if
\code{x} is a frNN object.}

\item{minPts}{number of minimum points required in the eps neighborhood for
core points (including the point itself).}
}
\value{
An updated \code{spata2} object. The coordinates data.frame
as obtained by \code{getCoordsDf()} contains an additional, character
variable named \emph{section} indicating the tissue section a barcode
spot was mapped to.
}
\description{
Identifies how many non-contiguous tissue sections
the data set contains and maps the barcode-spots to them.
}
