% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{image_annotations}
\alias{image_annotations}
\title{A list of lists of image annotations for specific \code{SPATA2}
objects - available via the \code{SPATAData} package. Names of the list slots
correspond to the sample names in variable  \emph{sample} of \code{SPATAData::source_df}.}
\format{
An object of class \code{list} of length 3.
}
\usage{
image_annotations
}
\description{
Use \code{setImageAnnotation()} to add one and \code{setImageAnnotations()} to add
multiple image annotations to the object.
}
\details{
@docType data
@usage data(image_annotations)
}
\keyword{datasets}
