% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\docType{class}
\name{ImageAnnotation-class}
\alias{ImageAnnotation-class}
\alias{ImageAnnotation}
\title{The \code{ImageAnnotation} - Class}
\description{
S4 class that contains information used to identify and
annotate structures in histology images.
}
\section{Slots}{

\describe{
\item{\code{area}}{data.frame. A data.frame that contains at least the numeric
variables \emph{x} and \emph{y}. Data corresponds to the polygong that
captures the spatial extent of the identified structure.}

\item{\code{id}}{character. String to identify the object in a list of multiple objects
of the same class.}

\item{\code{image}}{image. Cropped version of the annotated image that only contains
the area where the annotated structure is located (plus expand). This slot is
empty as long as the \code{ImageAnnotation} object is located in an
object of class \code{HistologyImage}. Extracting it with \code{getImageAnnotation()}
or \code{getImageAnnotations()} adds the cropped image to the slot.}

\item{\code{image_info}}{list. List of infos around the image of slot @image.}

\item{\code{misc}}{list. A flexible list for miscellaneous input.}

\item{\code{tags}}{character. Tags that can be used to group iamge annotations in different manners.
This can be a single or multiple strings.}
}}

