% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{imageAnnotationScreening}
\alias{imageAnnotationScreening}
\title{Implementation of the IAS-algorithm}
\usage{
imageAnnotationScreening(
  object,
  id,
  variables,
  buffer,
  n_bins_circle,
  n_bins_angle = 12,
  summarize_with = "mean",
  method_padj = "fdr",
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  with_de = TRUE,
  method_de = "wilcox",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{variables}{Character vector. All numeric variables, meaning genes,
gene-sets and numeric features that are supposed to be included in
the screening process.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}.}

\item{n_bins_circle}{Numeric value (or vector of length two, see details
for more). Specifies how many times the area is buffere with the value
denoted in \code{buffer}. The combination of both results in area that is
screened.}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}.
Specifies the function with which the bins are summarized.}

\item{method_padj}{Character value. The method with which adjusted p-values are
calculated. Use \code{validPadjMehods()} to obtain all valid input options.}

\item{model_subset}{Character value. Used as a regex to subset the models
that are known to \code{SPATA2}. Use \code{validModelNames()} to obtain all model names
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included in the screening process.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{ImageAnnotationScreening}. See documentation
with \code{?ImageAnnotationScreening} for more information.
}
\description{
Screens the sample for numeric variables that stand
in meaningful, spatial relation to annotated structures/areas of
the sample.
}
