% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{imageAnnotationToSegmentation}
\alias{imageAnnotationToSegmentation}
\title{Convert image annotation to segmentation}
\usage{
imageAnnotationToSegmentation(
  object,
  ids,
  segmentation_name,
  inside = "inside",
  outside = "outside",
  overwrite = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{ids}{Character vector. Specifies the image annotation(s) of interest.
Barcode-spots that fall into the area of these annotations are labeled
with the input for argument \code{inside}.}

\item{segmentation_name}{Character value. The name of the new segmentation variable.}

\item{inside}{Character value. The group name for the barcode-spots that
are located inside the area of the image annotation(s).}

\item{outside}{Character value. The group name for the barcode-spots that
are located outside the area of the image annotation(s).}

\item{overwrite}{Logical. Set to TRUE to overwrite existing variables with
the same name.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Converts one or more image annotations to a binary
segmentation variable in the feature data.frame.
}
