% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{include_tissue_outline}
\alias{include_tissue_outline}
\title{Include spatial extent of tissue sections in analysis}
\usage{
include_tissue_outline(
  coords_df,
  input_df,
  outline_var = NULL,
  img_ann_center = NULL,
  ias_circles = FALSE,
  ccd = NULL,
  remove = TRUE
)
}
\arguments{
\item{input_df}{A data.frame that contains at least numeric \emph{x} and \emph{y}
variables.}

\item{ias_circles}{Logical value. If \code{TRUE}, input data.frame is assumed
to contain polygon coordinates of the expanded image annotation encircling
and sorts them after filtering for those that lie inside the tissue section
in order to plot them via \code{ggplot2::geom_path()}.}
}
\value{
Filtered input data.frame.
}
\description{
Ensures section specific processing of observations
in relation to image annotations by identifying the outline of the
tissue section (or -sections in case of multiple tissue sections per sample)
and by relating the observations to it.
}
