% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{inferSingleCellGradient}
\alias{inferSingleCellGradient}
\title{Count cells depending on distance to image annotation}
\usage{
inferSingleCellGradient(
  object,
  sc_input,
  id,
  calculate = "density",
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  n_bins_angle = 1,
  remove_circle_bins = "Outside",
  normalize = TRUE,
  area_unit = NULL,
  format = "wide",
  as_models = FALSE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{sc_input}{Data.frame that contains the results from single cell deconvolution.
Must have at least three columns:
\itemize{
\item{\emph{x} : }{ numeric. Position of cell on the x axis in pixel.},
\item{\emph{y} :}{ numeric. Position of cell on the y axis in pixel.},
\item{\emph{cell_type} :}{ factor Cell type of the cell.}
}}

\item{id}{Character value. The ID of the spatial annotation of interest.}

\item{distance}{Distance value. Specifies the distance from the border of the
image annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{binwidth}.
(See details for more.)}

\item{binwidth}{Distance value. The width of the circular bins to which
the barcode-spots are assigned. We recommend to set it equal to the center-center
distance: \code{binwidth = getCCD(object)}. (See details for more.) - See details of \code{?is_dist} for more
information about distance values.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the image annotation.
(See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{as_models}{Adjusts the output to a list that is a valid input for
\code{models_add}-argument of \code{spatialAnnotationScreening()}.}
}
\value{
Data.frame as is returned by \code{getSasDf()} with cell types as variables.
}
\description{
Integration of single cell deconvolution and SPATA2s image annotations.
}
