% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObject}
\alias{initiateSpataObject}
\title{Initiate a \code{spata2} object}
\usage{
initiateSpataObject(
  count_mtr,
  coords_df,
  sample_name,
  dir_img_ref = NULL,
  name_img_ref = NULL,
  spatial_method = "Undefined",
  meta = list(),
  misc = list(),
  scale_factors = list(coords = 1),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{count_mtr}{A count matrix. Column names are barcodes. Rownames are the
feature (genes, proteins, metabolites etc.).}

\item{coords_df}{Data.frame with a variable called \emph{barcodes} as well as the
\emph{x_orig} and \emph{y_orig} or \emph{x} and \emph{y}.}

\item{dir_img_ref}{Character value or \code{NULL}. If character, the directory
to the reference image.}

\item{name_img_ref}{Character value or \code{NULL}. If character, the name of
the reference image. Ignored if \code{dir_img_ref} is \code{NULL}.}

\item{spatial_method}{Character value or object of class \code{\link{SpatialMethod}}.
If character, one of \code{validSpatialMethods()}.}

\item{meta, misc}{List of meta- and miscellaneous data for the \code{\link{HistoImaging}}
object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{spata2}.
}
\description{
Initiates a \code{spata2} object using the basic inputs, a coordinates
data.frame and a count matrix.
}
\details{
If \code{dir_img_ref} is not specified, the function creates a \code{\link{PseudoHistoImage}}.
Use \code{registerImage()} to register images later on.

If \code{dir_img_ref} is specified, the function creates a regular \code{\link{HistoImage}}.
Note that the x- and y-coordinates provided in the coordinates data.frame are
considered to be the original coordinates and are forced into names \emph{x_orig}
and \emph{y_orig}. Upon retrieval coordinates are scaled to image in use using the
\emph{coordinates scale factor}. This defaults to 1 (but can be adjusted via argument
\code{scale_factors} which defaults to \code{scale_factors = list(coords = 1)}). As the
default is 1, the default expects the coordinates and the image to align
perfectly.
}
