% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObject}
\alias{initiateSpataObject}
\title{Initiate an object of class \code{SPATA2}}
\usage{
initiateSpataObject(
  sample_name,
  count_mtr,
  modality,
  coords_df,
  img = NULL,
  img_dir = NULL,
  img_name = "image1",
  scale_factors = list(),
  spatial_method = "Undefined",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sample_name}{Character value. The name with which to identify the \code{SPATA2} object.}

\item{count_mtr}{A count matrix. Column names correspond to the barcodes of the \link[=concept_observations]{observations}.
Rownames correspond to the names of the molecular features (genes, proteins, metabolites etc.).}

\item{modality}{Character value. Should best describe the molecular type of the count matrix.
Additionally, it defines the \link[=MolecularAssay]{assay} name, that is created with the count matrix and further
referred to via the argument \code{assay_name}. Read more on molecular modalities in
SPATA2 \link[=concept_molecular_modalities]{here}.}

\item{coords_df}{Data.frame with a variable called \emph{barcodes} as well as the
\emph{x_orig} and \emph{y_orig} or \emph{x} and \emph{y}.}

\item{img}{The reference image. See details for more information on how or
how not to include an image.}

\item{img_dir}{Character value or \code{NULL}. If character, the file directory
to the reference image. See details for more information on how and why to include
a file directory.}

\item{img_name}{Character value. The name of the reference image. Only
required if at least \code{img} or \code{img_dir} is specified.}

\item{scale_factors}{A list of \link[=concept_scale_factors]{scale_factors}
set in slot @scale_factors of the \code{\link{HistoImage}} object (if the \code{SPATA2} object is initiated
with an image, see details) or slot @scale_factors of the \code{\link{SpatialData}} object, if
no image is provided.}

\item{spatial_method}{Character value or object of class \code{\link{SpatialMethod}}.
If character, one of \code{validSpatialMethods()}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{meta, misc}{List of meta- and miscellaneous data for the \code{\link{SpatialData}}
object.}
}
\value{
An object of class \code{SPATA2}.
}
\description{
Initiates a \code{\link{SPATA2}} object using the basic inputs: a coordinates
data.frame and a count matrix.
}
\note{
In contrast to \code{\link[=initiateSpataObjectVisium]{initiateSpataObjectVisium()}} or \code{\link[=initiateSpataObjectMERFISH]{initiateSpataObjectMERFISH()}},
a \code{SPATA2} object of this function the output does not contain a tissue outline yet!
Run \code{\link[=identifyTissueOutline]{identifyTissueOutline()}} with your choice of parameters afterwards.
}
\section{Initiating the object with an image}{

\code{SPATA2} allows to register multiple images with one object via file directories.
This facilitates exchanging them during the analysis while they must not
be loaded altogether in a \code{SPATA2} object, which saves storage space. By default,
the function takes a directory to the image you want to initiate your \code{SPATA2} object with,
then loads the image and saves the directory, too. If the image does not
exist in a file on your device but only in the global environment you can
use \code{img} directly. This way, no image directory is stored. In a scenario,
where you want to register an additional image and use it for further analysis,
you can not \emph{unload} the image with which you initiated the object because it
would be lost since there is not directory from which to retrieve it once
you want to use it again. Therefore, we recommend to initiate the object
with a file directory to the image and not with an image from the global
environment.

Lastly, if \code{img} and \code{img_dir} is specified the image is saved under
this directory as a .png file and the image is registered normally.
}

\section{Tutorials}{

Extensive tutorials for how to use this function can be found on our website
\url{https://themilolab.github.io/SPATA2/} .
}

