% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObject_Empty}
\alias{initiateSpataObject_Empty}
\title{Initiate an empty \code{spata2} object}
\usage{
initiateSpataObject_Empty(sample_name, spatial_method = "Visium")
}
\arguments{
\item{sample_name}{Character value. Denotes the name of the sample you are
analyzing with the spata-object. The future input for SPATA's \code{of_sample}-argument.}
}
\value{
An empty object of class \code{SPATA2}.
}
\description{
Initiate an empty \code{spata2} object
}
