% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{initiateSpataObjectVisium}
\alias{initiateSpataObjectVisium}
\title{Initiate \code{spata2} object from platform}
\usage{
initiateSpataObjectVisium(
  directory_visium,
  sample_name,
  mtr = "filtered",
  img_active = "lowres",
  img_ref = "lowres",
  directory_spata = NULL,
  gene_set_path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{directory_visium}{Character value. Directory to a visium folder. Should contain
the subdirectory \emph{'.../spatial'}.}

\item{sample_name}{Character value. Name of the sample.}

\item{mtr}{The matrix to load. One of \code{c("filtered", "raw")}.}

\item{img_ref, img_active}{Character values specifying which of the images to register and how to register
them. See details of \code{\link{HistoImaging}} for more information about the definitions
of the reference image and the active image. Setting both arguments to the same
value results in the function to register the specified image as the active
as well as the reference image. Additional images can be registered later on
at any time using the funciton \code{\link[=registerImage]{registerImage()}}. Valid input options depend
on the platform used:

\itemize{
\item{\emph{Visium}:}{ Either \emph{'lowres'} or \emph{'hires'}.}
}}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An object of class \code{spata2}.
}
\description{
A collection of functions that initiate \code{spata2} objects
from standardized output folders.
}
\seealso{
\code{\link{createHistoImagingVisium}}
}
