% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSpataObject.R
\name{initiateSpataObjectXenium}
\alias{initiateSpataObjectXenium}
\title{Initiate \code{spata2} object from platform Xenium}
\usage{
initiateSpataObjectXenium(directory_xenium, sample_name, verbose = TRUE)
}
\arguments{
\item{sample_name}{Character value. Name of the sample.}

\item{directory_visium}{Character value. Directory to a xenium folder. Should contain
the subdirectory \emph{'.../cell_feature_matrix'} and the file \emph{'.../cells.csv.gz'}..}
}
\value{
An object of class \code{spata2}.
}
\description{
Wrapper function around the necessary content to create a
\code{spata2} object from standardized output of the Xenium platform.
}
