% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{is_area}
\alias{is_area}
\alias{is_area_pixel}
\alias{is_area_si}
\title{Test area input}
\usage{
is_area(input, error = FALSE)

is_area_pixel(input, error = FALSE)

is_area_si(input, error = FALSE)
}
\arguments{
\item{input}{Character vector. Elements must match the requirements of
the \code{SPATA2} area framework. See details for more information.}
}
\value{
Logical vector of the same length as input and/or an error if \code{verbose}
is \code{TRUE}.
}
\description{
Tests if input refers to an area using international area
units according to the \code{SPATA2} area framework.

\itemize{
\item{\code{is_area()}:}{ Tests if input can be interpreted as an area}
\item{\code{is_area_si()}:} {Tests if input can be interpreted as an area in SI units.}
\item{\code{is_area_pixel()}:} {Tests if input can be interpreted as an area
in pixel.}
}
}
\examples{

library(SPATA2)

##### provide input as character vectors

# will return TRUE

is_area(input = c('2mm2', '4mm2'))

# will return FALSE

is_area(input = c('200 m2')) # space between value and unit

# will return TRUE

area_values <- c(200, 400)

area_values <- as_area(area_values, unit = "mm2")

is_area(input = area_values)

###### use units package

library(units)

area_values2 <- set_units(x = c(200, 300), value = "mm2")

is_area(area_values2)


}
