% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i.R
\name{is_dist}
\alias{is_dist}
\alias{is_dist_euol}
\alias{is_dist_pixel}
\title{Test distance input}
\usage{
is_dist(input, error = FALSE)

is_dist_euol(input, error = FALSE)

is_dist_pixel(input, error = FALSE)
}
\arguments{
\item{input}{Character or numeric vector. Elements must match the
requirements of the \code{SPATA2} distance framework. See details
for more information.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical vector of the same length as \code{input}. If \code{error} is \code{TRUE}
and one or more elements of the input values can not be interpreted approapriately
the functions throws an error.
}
\description{
Tests if input that refers to a distance is of valid input.

\itemize{
\item{\code{is_dist()}:}{ Tests if input can be interpreted as a distance.}
\item{\code{is_dist_euol()}:} {Tests if input can be interpreted as a distance in
European units of length.}
\item{\code{is_dist_pixel()}:} {Tests if input can be interpreted as a distance
in pixel.}
}
}
\details{
Several functions in \code{SPATA2} have arguments that take \emph{distance input}.
To specifically refer to a distance the unit must be specified. There are
three ways to create valid input for these arguments.

\bold{1. Distance in pixel}:

There are two valid input options to specify the distance in pixel:

\itemize{
\item{numeric:}{ Single numeric values, e.g. \code{arg_input <- c(2, 3.554, 69, 100.67)}. If no unit
is specified the input will be interpreted as pixels.}
\item{character:}{ Suffixed with \emph{'px'}, e.g. \code{arg_input <- c('2px', '3.554px', '69px', '100.67px')}}
}

\bold{2. Distance in European units of length (euol)}:

Specifying distances in European units of length e.g. \code{arg_input <- c('2mm', '4mm')} etc.
requires the input to be a character as the unit must be provided as suffix. Between the numeric
value and the unit must be no empty space! Unit suffixes must be one of
\emph{'m', 'dm', 'cm', 'mm', 'um', 'nm'}.

\strong{3. As vectors of class \code{unit}:}

Behind the scenes \code{SPATA2} works with the \code{units}-package. Input
is converted into vectors of class \code{units}. Therefore, input can be directly
provided this way: \code{arg_input <- units::set_unit(x = c(2,4), value = 'mm')}
Note that \emph{pixel} is not a valid unit in the \code{units} package. If you want
to specify the input in pixel you have to use input option 1. Distance in pixel.
}
\examples{

##### use numeric or character vectors

library(SPATA2)

# will return TRUE
is_dist(input = 200) # -> 200 pixel
is_dist(input = "20px") # > 20 pixel

is_dist(input = "40.5mm") # -> 40.5 mm

# will return FALSE
is_dist(input = "30.5 mm") # -> empty space between 30.5 and mm

is_dist(input = ".4mm") # -> must start with a number

##### use units package

library(units)

dist_input <- set_units(x = c(2, 3, 4.4), value = "mm")

is_dist(dist_input)

}
