% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinWithDimRed-family.R
\name{joinWithPca}
\alias{joinWithPca}
\alias{joinWithTsne}
\alias{joinWithUmap}
\title{Join barcodes with additional variables}
\usage{
joinWithPca(
  object,
  spata_df,
  n_pcs = NULL,
  verbose = NULL,
  force = FALSE,
  of_sample = NA
)

joinWithTsne(object, spata_df, verbose = NULL, force = FALSE, of_sample = NA)

joinWithUmap(object, spata_df, verbose = NULL, force = FALSE, of_sample = NA)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be included.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{force}{Logical. Only relevant if the spata data.frame provided
already contains the variables that would be added with the function.
If set to TRUE, the variables are added anyway.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{...}{Addtional arguments given to \code{dplyr::left_join()}.}
}
\value{
The input data.frame with the additional dimensional reduction
variables
}
\description{
These functions add dimensional reduction results
in form of additional variables to the provided spata data.frame.
}
