% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-add-ons.R
\name{legendBottom}
\alias{legendBottom}
\alias{legendLeft}
\alias{legendNone}
\alias{legendRight}
\alias{legendTop}
\alias{legendColor}
\title{Ggplot add on wrapper}
\usage{
legendBottom(...)

legendLeft(...)

legendNone(...)

legendRight(...)

legendTop(...)

legendColor(...)
}
\description{
Ggplot add on wrapper
}
