% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{legendBottom}
\alias{legendBottom}
\alias{legendColor}
\alias{legendLeft}
\alias{legendNone}
\alias{legendRight}
\alias{legendTop}
\title{ggplot2 legend manipulation}
\usage{
legendBottom(...)

legendColor(...)

legendLeft(...)

legendNone(...)

legendRight(...)

legendTop(...)
}
\description{
ggplot2 legend manipulation
}
