% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{labsNone}
\alias{labsNone}
\alias{legendBottom}
\alias{legendColor}
\alias{legendLeft}
\alias{legendNone}
\alias{legendRight}
\alias{legendTop}
\title{ggplot2 basic manipulation}
\usage{
labsNone()

legendBottom(...)

legendColor(...)

legendLeft(...)

legendNone(...)

legendRight(...)

legendTop(...)
}
\description{
ggplot2 basic manipulation
}
\examples{

library(SPATA2)
library(tidyverse)

object <- loadExampleObject("UKF275T")

plotSurface(object, color_by = "HM_HYPOXIA")
plotSurface(object, color_by = "HM_HYPOXIA") + legendBottom()

}
