% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{loadImageDefault}
\alias{loadImageDefault}
\alias{loadImageHighres}
\alias{loadImageLowres}
\title{Deprecated}
\usage{
loadImageDefault(object, ...)

loadImageHighres(object, ...)

loadImageLowres(object, ...)
}
\description{
Deprecated in favor of \code{\link[=activateImage]{activateImage()}} and/or \code{\link[=loadImage]{loadImage()}}.
}
\keyword{internal}
