% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.R
\name{loadImage}
\alias{loadImage}
\alias{loadImageDefault}
\alias{loadImageHighres}
\alias{loadImageLowres}
\title{Load known images}
\usage{
loadImage(object, name, ...)

loadImageDefault(object, ...)

loadImageHighres(object, ...)

loadImageLowres(object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{name}{Character value. Name of the image directory.}

\item{...}{Additional arguments given to \code{exchangeImage()}.}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Wrapper around the required \verb{getImageDir*()} function and
\code{exchangeImage()}. Exchanges the image of the \code{SPATA2} object by using
the directories that have been set with \code{setImageDir*()} family
or with \code{addImageDir()}.
}
\seealso{
\code{\link[=setImageDirLowres]{setImageDirLowres()}}, \code{\link[=setImageDirHighres]{setImageDirHighres()}},
\code{\link[=setImageDirDefault]{setImageDirDefault()}}, \code{\link[=addImageDir]{addImageDir()}},  \code{\link[=exchangeImage]{exchangeImage()}}, \code{\link[=getImagaDirectories]{getImagaDirectories()}}
}
