% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R, R/l.R
\name{loadImage}
\alias{loadImage}
\alias{loadImage,spata2-method}
\alias{loadImage,HistologyImage-method}
\alias{loadImage,HistologyImagingNew-method}
\alias{loadImageDefault}
\alias{loadImageHighres}
\alias{loadImageLowres}
\title{Load known images}
\usage{
loadImage(object, ...)

\S4method{loadImage}{spata2}(object, name, ...)

\S4method{loadImage}{HistologyImage}(object, verbose = TRUE)

\S4method{loadImage}{HistologyImagingNew}(object, name, verbose = TRUE)

loadImageDefault(object, ...)

loadImageHighres(object, ...)

loadImageLowres(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Additional arguments given to \code{exchangeImage()}.}

\item{name}{Character value. Name of the image directory.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Wrapper around the required \verb{getImageDir*()} function and
\code{exchangeImage()}. Exchanges the image of the \code{SPATA2} object by using
the directories that have been set with \code{setImageDir*()} family
or with \code{addImageDir()}.
}
\seealso{
\code{\link[=setImageDirLowres]{setImageDirLowres()}}, \code{\link[=setImageDirHighres]{setImageDirHighres()}},
\code{\link[=setImageDirDefault]{setImageDirDefault()}}, \code{\link[=addImageDir]{addImageDir()}},  \code{\link[=exchangeImage]{exchangeImage()}}, \code{\link[=getImagaDirectories]{getImagaDirectories()}}
}
