% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadImageLowres}
\alias{loadImageLowres}
\alias{loadImageHighres}
\title{Exchange images}
\usage{
loadImageLowres(object, verbose = NULL)

loadImageHighres(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata object.
}
\description{
Exchanges the image of the spata object by using
the directories that have been set with \code{setImageDirLowres()} or
\code{setImageDirHighres()} and scales the barcodes spots coordinates
accordingly.
}
