% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{loadSpataObject}
\alias{loadSpataObject}
\alias{loadCorrespondingCDS}
\alias{loadCorrespondingSeuratObject}
\title{Load corresponding objects}
\usage{
loadSpataObject(directory_spata, verbose = TRUE)

loadCorrespondingCDS(object, verbose = NULL)

loadCorrespondingSeuratObject(object, verbose = NULL)
}
\arguments{
\item{directory_spata}{Character value. The directory from which to load the spata-object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{object}{A valid spata-object.}
}
\value{
The load object of interest.
}
\description{
Family of functions to load corresponding objects of different analysis
platforms. See details and value for more information.
}
\details{
\code{loadSpataObject()} is a wrapper around \code{base::readRDS()} with which
you could load your spata-object as well. The other two functions take the spata-object
and use \code{getDirectoryInstructions()} to extract the directories of the corresponding object to be loaded under which
they were saved the last time \code{saveCorresponding*()} was used.
}
