% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{mergeGroups}
\alias{mergeGroups}
\title{Lump groups together}
\usage{
mergeGroups(object, grouping, grouping_new, merge, new_group, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable whose
groups are supposed to be merged.}

\item{grouping_new}{Character value or NULL. If character,
the results are stored in a new variable named accordingly. If NULL,
the grouping variable is updated - DEA results will be discarded.}

\item{merge}{Character vector or NULL. If character, specifies the groups
that are merged together.}

\item{new_group}{Character value. The new group name of the merge.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\description{
Merge groups into one group.
}
\details{
Only one argument of \code{keep} or \code{merge} must be specified.
If \code{grouping_new} is NULL DEA results of the specified
grouping variable is resetted.
}
\examples{

library(SPATA2)
library(tidyverse)

object <- loadExampleObject("UKF275T", meta = TRUE)

object <-
  mergeGroups(
    object = object,
    grouping = "bayes_space",
    grouping_new = "bayes_space_merged",
    merge = c("B1", "B6"),
    new_group = "B1B6_merged"
   )

plotSurface(object, color_by = "bayes_space")
plotSurface(object, color_by = "bayes_space_merged")

}
