% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{mergeTissueSections}
\alias{mergeTissueSections}
\title{Merge tissue sections}
\usage{
mergeTissueSections(object, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Collection of vectors that carry the names of the sections
to be merged.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Merges tissue sections that have been mistakenly identified
as two non-contiguous sections.
}
\examples{

# in this fictional example the algorithm identified 6 tissue sections
# this call merges sections 1,2,3 to section 1_2_3 and sections 4,5,6 to
# 4_5_6
\dontrun{ object <- mergeTissueSections(object, c(1,2,3), c(4,5,6)) }


}
\seealso{
\code{\link[=identifyTissueSections]{identifyTissueSections()}}
}
