% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{MolecularAssay-class}
\alias{MolecularAssay-class}
\alias{MolecularAssay}
\title{Assay Class}
\description{
A class to represent assay data, including analysis results, metrics, and
statistical summaries. The \code{Assay} class encapsulates various components
of assay data including raw and processed metrics, analytical results,
and associated metadata like the assay name and omic type.
}
\section{Slots}{

\describe{
\item{\code{analysis}}{List of analysis results where each element can represent
a different analysis aspect.}

\item{\code{mtr_counts}}{Matrix object storing raw counts metrics from the assay.}

\item{\code{mtr_proc}}{List of processed metrics, potentially including normalized
values or transformed data.}

\item{\code{name}}{Character vector storing the name of the assay.}

\item{\code{omic}}{Character vector indicating the type of omics data (e.g., "transcriptomics", "proteomics").}

\item{\code{stats}}{Data frame of statistical summaries or results derived from the assay data.}
}}

