% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nBarcodes}
\alias{nBarcodes}
\title{Number of barcodes}
\usage{
nBarcodes(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
Numeriv value.
}
\description{
Returns the number of barcodes in the sample.
}
