% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nBarcodes}
\alias{nBarcodes}
\title{Number of barcodes}
\usage{
nBarcodes(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeriv value.
}
\description{
Returns the number of barcodes in the sample.
}
