% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nGenes}
\alias{nGenes}
\title{Number of genes}
\usage{
nGenes(object, mtr_name = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Numeriv value.
}
\description{
Returns the number of genes in the active matrix.
}
