% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nImageAnnotations}
\alias{nImageAnnotations}
\title{Number of image annotations}
\usage{
nImageAnnotations(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Numeric value.
}
\description{
Returns the number of \code{ImageAnnotation}-objects in the sample.
}
