% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-functions.R
\name{nImageAnnotations}
\alias{nImageAnnotations}
\title{Number of image annotations}
\usage{
nImageAnnotations(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Numeric value.
}
\description{
Returns the number of \code{ImageAnnotation}-objects in the sample.
}
