% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nSpatialAnnotations}
\alias{nSpatialAnnotations}
\alias{nSpatialAnnotations,SPATA2-method}
\alias{nSpatialAnnotations,SpatialData-method}
\title{Number of image annotations}
\usage{
nSpatialAnnotations(object, ...)

\S4method{nSpatialAnnotations}{SPATA2}(object)

\S4method{nSpatialAnnotations}{SpatialData}(object)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Numeric value.
}
\description{
Returns the number of \code{ImageAnnotation}-objects in the sample.
}
