% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{nSpatialTrajectories}
\alias{nSpatialTrajectories}
\alias{nTrajectories}
\title{Number of spatial trajectories}
\usage{
nSpatialTrajectories(object)

nTrajectories(objcect)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}
}
\value{
Numeric value.
}
\description{
Returns the number of \code{SpatialTrajectries}-objects in the sample.
}
