% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-plotting-functions.R
\name{plotAutoencoderAssessment}
\alias{plotAutoencoderAssessment}
\title{Plot total variance of different neural networks}
\usage{
plotAutoencoderAssessment(
  object,
  activation_subset = NULL,
  clrp = NULL,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
ggplot_family return
}
\description{
Visualizes the results of \code{assessAutoencoderOptions()} by displaying the
total variance of each combination of an activation function (such as \emph{relu, sigmoid})
and the number of bottleneck neurons.

The results depend on further adjustments like number of layers, dropout and number of epochs.
}
