% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA-E.R
\name{plotClockplot}
\alias{plotClockplot}
\alias{plotClockplot,spata2-method}
\alias{plotClockplot,SpatialAnnotationScreening-method}
\title{Plota clockplot}
\usage{
plotClockplot(object, ...)

\S4method{plotClockplot}{spata2}(
  object,
  id,
  variables,
  distance = NA_integer_,
  n_bins_circle = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  n_angle_bins = 12,
  summarize_with = "mean",
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  layout = 1,
  switch = NULL,
  fill = "steelblue",
  ...
)

\S4method{plotClockplot}{SpatialAnnotationScreening}(
  object,
  variables,
  model_subset = NULL,
  model_remove = NULL,
  layout = 1,
  switch = NULL,
  fill = "steelblue",
  ...
)
}
\arguments{
\item{object}{Any object for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{variables}{Character vector. The numeric variables of interest. Must be inside:

\itemize{
\item{ \strong{Gene sets} Must be in \code{getGeneSets()}}
\item{ \strong{Genes} Must be in \code{getGenes()}}
\item{ \strong{Features} Must be in \code{getFeatureNames(..., of_class = "numeric")}}
}}

\item{distance}{Distance value. Specifies the distance from the border of the
spatial annotation to the \emph{horizon} in the periphery up to which the screening
is conducted. (See details for more.) - See details of \code{?is_dist} for more
information about distance values. Defaults to a distance that covers the whole
tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{switch}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}

\item{fill}{Character value. The color with which the columns are filled.}
}
\description{
Visualize the evaluation of the fit of a numeric variable
against models around the area of an image annotation.
}
