% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{plotCnvResults}
\alias{plotCnvResults}
\title{Deprecated}
\usage{
plotCnvResults(...)
}
\description{
Deprecated in favor of \code{\link[=plotCnvLineplot]{plotCnvLineplot()}} and/or \code{\link[=plotCnvHeatmap]{plotCnvHeatmap()}}.
}
\keyword{internal}
