% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotDeaDotPlot}
\alias{plotDeaDotPlot}
\title{Plot differentially expressed genes}
\usage{
plotDeaDotPlot(
  object,
  across,
  across_subset = NULL,
  relevel = NULL,
  method_de = NULL,
  by_group = TRUE,
  max_adj_pval = NULL,
  min_lfc = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  alpha_by = NULL,
  alpha_trans = "identity",
  color_by = "avg_logFC",
  size_by = "p_val_adj",
  size_trans = "reverse",
  pt_alpha = 0.9,
  pt_size = 2,
  pt_color = "blue4",
  pt_clrsp = "plasma",
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character value or NULL. If character, specifies the name of the grouping variable
that contains the group assignment across which results are displayed. If NULL,
no grouping is done.}

\item{across_subset}{Character vector or NULL. If character, specifies groups of the
grouping variable that are of interest. Prefixing group names with \emph{'-'} excludes them.
If NULL, all groups are considered.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across.subset} and
\code{variables} determines the order in which the groups of interest are displayed.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{by_group}{Logical value. If TRUE for every group in the grouping
variable a single dot plot is created. If FALSE one plot for all groups and all
gene sets is created.}

\item{alpha_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{alpha_trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{color_by}{Character value or NULL. If character, specifies the
variable that is displayed by color.}

\item{size_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{size_trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_color}{Character value. Specifies the color with which all points are displayed .}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{scales}{Affects the way the subplots
are displayed.}

\item{nrow}{Affects the way the subplots
are displayed.}

\item{ncol}{Affects the way the subplots
are displayed.}

\item{...}{Additional arguments given to \code{scale_color_add_on()}.}
}
\value{
A ggplot.
}
\description{
Visualizes results of DE analysis with
dot plots.
}
