% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA-G.R
\name{plotDeaVolcano}
\alias{plotDeaVolcano}
\title{Visualize gene expression testing}
\usage{
plotDeaVolcano(
  object,
  across = getDefaultGrouping(object),
  across_subset = NULL,
  relevel = TRUE,
  method_de = NULL,
  color_up = "tomato",
  color_down = "steelblue",
  color_insignif = "lightgrey",
  pt_alpha = 0.9,
  pt_size = 1,
  threshold_logFC = 1,
  threshold_pval = 0.01,
  label_genes = 5,
  label_size = 1,
  nrow = NULL,
  ncol = NULL,
  scales = "fixed",
  use_pseudolog = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{across}{Character value or NULL. Specifies the grouping variable of interest.

Use \code{getGroupingOptions()} to obtain all variable names that group the
barcode spots of your object in a certain manner.}

\item{across_subset}{Character vector or NULL. Specifies the particular groups
of interest the grouping variable specified in argument \code{across} contains.

If set to NULL all of them are chosen. You can prefix groups you are NOT interested in
with a \emph{'-'}. (Saves writing if there are more groups you are interested in
than groups you are not interested in.)

Use \code{getGroupNames()} to obtain all valid input options.}

\item{relevel}{Logical value. If set to TRUE the input order of \code{across_subset}
determines the order in which the groups of interest are displayed. Groups that
are not included are dropped which affects the colors with which they are displayed.}

\item{method_de}{Character value. Denotes the method to according to which the de-analysis is performed.
Given to argument \code{test.use} of the \code{Seurat::FindAllMarkers()}-function. Run \code{SPATA::dea_methods}
to obtain all valid input options.}

\item{color_up}{Character value. Color of points that represent upregulated genes.}

\item{color_down}{Character value. Color of poitns that represent downregulated genes.}

\item{color_insignif}{Character value. Color of points that fall below the threshold set
via argument \code{threshold_pval}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{threshold_logFC}{Numeric value. Denotes the threshold for the logFC. Genes with
a logFC below are displayed as not significant.}

\item{threshold_pval}{Numeric value. Denotes the threshold for the p-value. Genes with
a p-value above are displayed as not significant.}

\item{label_genes}{Specify which genes are labeled in the plot. If numeric,
specifies the number of genes that are labeled. E.g. if \code{label_genes} = 5,
the default, the top 5 genes are labeled. If character, specifies the genes
that are supposed to be labeled by name. If NULL, no genes are labeled.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{use_pseudolog}{Logical value. If TRUE, avglogFC is transformed with log10. Requires
package \code{ggallin} to be installed.}
}
\value{
A ggplot.
}
\description{
Plots a common volcano plot with p-value on the y- and logfold
change on the x-axis.
}
