% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust-implementation.R
\name{plotDendrogram}
\alias{plotDendrogram}
\alias{plotDendrogramCnv}
\title{Visualize hierarchical clustering results}
\usage{
plotDendrogram(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_size = 1,
  clrp = NULL,
  clrp_adjust = NULL,
  display_legend = NULL,
  display_title = NULL,
  ncol = NULL,
  nrow = NULL,
  of_sample = NA,
  verbose = NULL
)

plotDendrogramCnv(
  object,
  method_dist = NULL,
  method_aggl = NULL,
  k = NULL,
  h = NULL,
  type = "rectangle",
  direction = "bt",
  branch_size = 1,
  clrp = NULL,
  clrp_adjust = NULL,
  display_legend = NULL,
  display_title = NULL,
  ncol = NULL,
  nrow = NULL,
  of_sample = NA,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{method_dist}{Character value or vector (see details for more). Denotes the distance methods (e.g. \emph{'euclidean'}) to be used. Run \code{validDistanceMethods()}
to obtain all valid input options.}

\item{method_aggl}{Character value or vector (see details for more). Denotes the agglomerative methods (e.g. \emph{'ward.D'}) to be used. Run \code{validAgglomerativeMethods()}
to obtain all valid input options.}

\item{k}{Numeric value or vector or NULL (see details for more). Denotes the number of clusters
in which the hierarchical tree is supposed to be split.}

\item{h}{Numeric value or vector or NULL (see details for more). Denotes the height at which
the dendrogom is cut.}

\item{type}{Character value. Denotes the type of the dendrogram. Either \emph{'rectangle'}
or \emph{'triangle}.}

\item{direction}{Character value. Denotes the direction of the dendrogram.
Either \emph{'bt'} (bottom-top) or \emph{'lr'} (left-right).}

\item{branch_size}{Numeric value. Denotes the size of the branches.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{ncol}{Affects the way the subplots
are displayed.}

\item{nrow}{Affects the way the subplots
are displayed.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Visualizes one or more dendrograms (depending on the input of
arguments \code{method_dist} and \code{method_aggl}).
}
\details{
Not specifying arggments \code{k} and \code{h} results in pure dendrograms.
If you want highlight cluster you can only specify one the other argument must
be set to NULL.

Input for the \code{method_*}-arguments can be either a single character value or
a vector of several methods. In the latter case dendrograms for all combinations
are displayed via \code{gridExtra::grid.arrange()} and input for plot manipulating arguments
must be either of length one or of a length corresponding to the number of plots.
}
