% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotA-E.R
\name{plotExprVsDistSA}
\alias{plotExprVsDistSA}
\alias{plotExprVsDistST}
\title{Plot expression as a function of distance to a spatial references}
\usage{
plotExprVsDistSA(
  object,
  variables,
  ids = idSA(object),
  distance = "dte",
  resolution = recSgsRes(object),
  core = FALSE,
  pt_alpha = 0.1,
  pt_color = "black",
  pt_clrp = NULL,
  pt_size = 1.5,
  line_alpha = 0,
  line_color = "forestgreen",
  line_size = 1.75,
  border_linealpha = 1,
  border_linecolor = "black",
  border_linesize = 1,
  border_linetype = "solid",
  ee_linealpha = 0,
  ee_linecolor = "black",
  ee_lineend = "point",
  ee_linesize = 1,
  ee_linetype = "solid",
  se_fill = ggplot2::alpha("lightgrey", 0.5),
  unit = getDefaultUnit(object),
  ggpLayers = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)

plotExprVsDistST(
  object,
  variables,
  id = idST(object),
  resolution = recSgsRes(object),
  width = NULL,
  pt_alpha = 0.5,
  pt_color = "black",
  pt_clrp = NULL,
  pt_size = 1.5,
  line_alpha = 0.9,
  line_color = "forestgreen",
  line_size = 1,
  unit = getDefaultUnit(object),
  ggpLayers = NULL,
  ncol = NULL,
  nrow = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. All numeric variables hat are supposed to be plotted.}

\item{ids}{Character vector. Specifies the IDs of the spatial annotations of interest.}

\item{distance}{A numeric vector of distances to the annotation's edge.}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{se_fill}{The fill color for the smoothing line's standard error area (default: "lightgrey").}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value. The ID of the spatial trajectory.}
}
\value{
A ggplot.
}
\description{
Generates a scatterplot to visualize the relationship between gene expression and
the distance of data points to a spatial reference. Set \code{line_alpha} > 0 to visualize the
inferred expression pattern used for spatial gradient screening.
}
\examples{

library(SPATA2)
library(tidyverse)

data("example_data")

object <- example_data$object_UKF275T_diet

object <- normalizeCounts(object, activate = TRUE)

object <-
 createNumericAnnotations(
   object = object,
   variable = "HM_HYPOXIA",
   threshold = "kmeans_high",
   id = "hypoxia_ann",
   inner_borders = FALSE,
   force1 = TRUE
   )

 plotExprVsDistSA(object, variables = c("HM_HYPOXIA", "METRN"), ids = "hypoxia_ann", core = T)

}
