% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-plotting.R
\name{plotGenePatternBinarized}
\alias{plotGenePatternBinarized}
\alias{plotGenePatternDenoised}
\title{Title}
\usage{
plotGenePatternBinarized(
  object,
  genes,
  pt_clr = "forestgreen",
  pt_clrsp = NA,
  pt_alpha = NULL,
  pt_size = NULL,
  ncol = NULL,
  nrow = NULL,
  verbose = NULL,
  of_sample = NA
)

plotGenePatternDenoised(
  object,
  genes,
  pt_clr = "forestgreen",
  pt_clrp = NA,
  pt_alpha = NULL,
  pt_size = NULL,
  ncol = NULL,
  nrow = NULL,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
