% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploIasHeatmap.R
\name{plotIasHeatmap}
\alias{plotIasHeatmap}
\title{Plot Image Annotation Screening Heatmap}
\usage{
plotIasHeatmap(
  object,
  id,
  variables,
  buffer,
  n_bins_circle,
  include_core = FALSE,
  arrange_rows = "input",
  colors = NULL,
  method_gs = NULL,
  smooth_span = NULL,
  multiplier = 10,
  clrsp = NULL,
  .f = NULL,
  .cols = dplyr::everything(),
  summarize_with = "mean",
  verbose = NULL,
  ...
)
}
\description{
Plot Image Annotation Screening Heatmap
}
