% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurfaceBase.R
\name{plotImage}
\alias{plotImage}
\alias{plotImageGgplot}
\title{Plot histology image}
\usage{
plotImage(object, xrange = NULL, yrange = NULL, ...)

plotImageGgplot(object)
}
\arguments{
\item{object}{A valid spata-object.}

\item{xrange, yrange}{Numeric vectors of length two or NULL. If numeric, specifies the coordinates
with which the image is subsetted. E.g. xrange = c(200, 500) results in the image being cropped
from x-coordinate 200 up to x-coordinate 500.}
}
\value{
A plot that is immediately plotted.
}
\description{
Plots the histology image as a raster.
}
