% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH-M.R
\name{plotImage}
\alias{plotImage}
\alias{plotImage,SPATA2-method}
\alias{plotImage,SpatialData-method}
\alias{plotImage,HistoImage-method}
\alias{plotImage,Image-method}
\title{Plot histology image (ggplot2)}
\usage{
plotImage(object, ...)

\S4method{plotImage}{SPATA2}(
  object,
  img_name = activeImage(object),
  outline = FALSE,
  by_section = TRUE,
  fragments = TRUE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 0.5,
  line_type = "solid",
  transform = TRUE,
  img_alpha = 1,
  scale_fct = 1,
  xrange = NULL,
  yrange = NULL,
  ...
)

\S4method{plotImage}{SpatialData}(
  object,
  img_name = activeImage(object),
  outline = FALSE,
  by_section = TRUE,
  fragments = TRUE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 0.5,
  line_type = "solid",
  transform = TRUE,
  img_alpha = 1,
  scale_fct = 1,
  xrange = NULL,
  yrange = NULL,
  ...
)

\S4method{plotImage}{HistoImage}(
  object,
  outline = FALSE,
  by_section = TRUE,
  fragments = TRUE,
  line_alpha = 0.9,
  line_color = "black",
  line_size = 1,
  line_type = "solid",
  transform = TRUE,
  img_alpha = 1,
  scale_fct = 1,
  xrange = NULL,
  yrange = NULL,
  display_subtitle = FALSE,
  ...
)

\S4method{plotImage}{Image}(object, scale_fct = 1, img_alpha = 1, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Additional arguments given to \code{ggpLayerZoom()}.}

\item{img_name}{Character value. The name of the image of interest.
If \code{NULL}, the active histo image is chosen by default. Either way, must
be one of \code{\link[=getImageNames]{getImageNames()}}.}

\item{by_section}{Logical value. If \code{TRUE}, the outline is used which acknowledges
the potential for multiple tissue sections. If \code{FALSE}, the outline is used which
simply outlines everything with one single polygon.}

\item{line_alpha}{Numeric. Affects alpha of main lines of the plot.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{transform}{Logical value. Ignored if no images exist in the object. If \code{TRUE},
spatial transformation instructions saved during image alignment of
the image \code{img_name} are applied. Only relevant if there is an image and \code{\link[=alignImage]{alignImage()}} has
been used.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{unit}{Character value. Units of x- and y-axes. Defaults
to \emph{'px'}.}
}
\value{
A ggplot.
}
\description{
Plots the histology image with \code{ggplot2}.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\section{Image visualization with ggplot2}{


When comparing the output of \code{ggplot() + ggpLayerImage()} with other image plotting functions,
you may notice that the image appears horizontally flipped when plotted using \code{ggpLayerImage()}.
This behavior is due to the use of a Cartesian coordinate system in \code{SPATA2}, where a pixel
with coordinates c(width = 1, height = 1) is plotted on the left side at the bottom.
In contrast, functions like \code{EBImage::display()} or \code{graphics::plot()} use an \emph{image space} coordinate system,
where pixel heights start from the top. Consequently, in \emph{image space}, pixel c(width = 1, height = 1)
is displayed on the top resulting in mirror inverted visualization of the image.

We chose to use a Cartesian coordinate system in \code{SPATA2} because we believe it provides a more intuitive
framework for the spatial alignment of tissue, spatial annotations, spatial trajectories,
barcoded sots, single cells, etc. where coordinates in the corresponding data.frames are provided
in form of \emph{x}- and \emph{y}-variables. See \code{\link[=getCoordsDf]{getCoordsDf()}}, \code{\link[=getImgAnnOutlineDf]{getImgAnnOutlineDf()}}, \code{\link[=getTissueOutlineDf]{getTissueOutlineDf()}} etc.

If you prefer to view your image in the regular orientation, you can use the \code{flipAll()} function on your object,
specifying \code{axis = "horizontal"}, to reverse this effect.
}

