% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{plotImageAnnotations}
\alias{plotImageAnnotations}
\title{Plot image annotations}
\usage{
plotImageAnnotations(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  expand = 0.05,
  square = FALSE,
  encircle = TRUE,
  linecolor = "black",
  linesize = 1.5,
  linetype = "solid",
  fill = "orange",
  alpha = 0.25,
  display_title = FALSE,
  display_subtitle = TRUE,
  display_caption = TRUE,
  ggpLayers = list(),
  nrow = NULL,
  ncol = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{expand}{Numeric value or vector of length two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the xspan and the second
value is used to adjust the yspan. E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original xspan to the xspan of the image section and does not adjust
 the yspan.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. X- and yspan of the square are equal to the span
that is the biggest. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}

\item{encircle}{Logical value. If TRUE, are polygon is drawn around the
exact extent of the annotated structure (as was drawn in \code{annotateImage()}).}

\item{linesize}{Numeric value. The size of the line(s) plotted.}

\item{display_title}{Logical value. If TRUE, the number of each image annotation
is plotted in the title.}

\item{display_subtitle}{Logical value. If TRUE, the ID of each image annotation
is plotted in the subtitle.}

\item{display_caption}{Logial value. If TRUE, the tags of each image annotation
are plotted in the caption.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{plot}{Logical value. If TRUE, the plots are plotted immediately
via \code{gridExtra.grid.arrange()} and the list of plots is returned
invisibly. Else the list of plots is simply returned.}
}
\value{
A list of ggplots. Each slot contains a plot
that visualizes an image annotation.
}
\description{
Plots annotated
}
\details{
How to use arguments \code{tags} and \code{test} to specify
the image annotations of interest: Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are three options:

1. Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.

2. Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags.

3. Argument \code{test} set to \emph{'identical'} or \emph{3}: To be indluded, an image annotation
must be tagged with all of the input tags and must not be tagged with anything else.

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.
}
