% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-HistologyImage.R
\name{plotImageAnnotations}
\alias{plotImageAnnotations}
\title{Plot image annotations}
\usage{
plotImageAnnotations(
  object,
  ids = NULL,
  tags = NULL,
  test = "any",
  expand = 0.05,
  square = FALSE,
  encircle = TRUE,
  linecolor = "black",
  linesize = 1.5,
  linetype = "solid",
  fill = "orange",
  alpha = 0.25,
  display_title = FALSE,
  display_subtitle = TRUE,
  ggpLayers = list(),
  nrow = NULL,
  ncol = NULL,
  plot = TRUE,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{ids}{Character vector, numeric vector, or NULL. If character, the IDs of the image annotations of
interest. If numeric, the image annotations are picked by number. If NULL, all image annotations are included.}

\item{tags}{Character vector or NULL. If character, the image annotation tags of interest.}

\item{expand}{Numeric value. Specifies the extent to which the cropped image
section is expanded. 0 as input leads to no expansion at all. Values lower than 1
are considered in relation to the x- and yrange - e.g. \code{expand} = 0.5 expands the plot to 150%. Values
bigger than 1 are considered as absolute - e.g. \code{expand} = 20 adds 20pixels on
both sides of each axis respectively. If length of \code{expand} is 2 the first value
is used to expand the x-axis, the second value is used to expand the y-axis.}

\item{square}{Logical. If TRUE, the cropped section of the image that contains the annotated
structure is forced into a square. If FALSE, the section is cropped according to the extent
of the annotated structure and the input for argument \code{expand}.}

\item{encircle}{Logical value. If TRUE, are polygon is drawn around the
exact extent of the annotated structure (as was drawn in \code{annotateImage()}).}

\item{display_title}{Logical value. If TRUE, the ID of each image annotation
is plotted in the title.}

\item{display_subtitle}{Logical value. If TRUE, the tags of each image annotation
are plotted in the subtitle.}

\item{nrow}{Affects the way the subplots
are displayed.}

\item{ncol}{Affects the way the subplots
are displayed.}

\item{plot}{Logical value. If TRUE, the plots are plotted immediately
via \code{gridExtra.grid.arrange()} and the list of plots is returned
invisibly. Else the list of plots is simply returned.}
}
\value{
A list of ggplots. Each slot contains a plot
that visualizes an image annotation.
}
\description{
Plots annotated
}
\details{
How to use arguments \code{tags} and \code{test} to specify
the image annotations of interest: Input for argument \code{tags} specifies the tags of interest.
Argument \code{test} decides about how the specified tags are used to select
the image annotations of interest. There are three options:

1. Argument \code{test} set to \emph{'any'} or \emph{1}: To be included, an image annotation
must be tagged with at least one of the input tags.

2. Argument \code{test} set to \emph{'all'} or \emph{2}: To be included, an image annotation
must be tagged with all of the input tags.

3. Argument \code{test} set to \emph{'identical'} or \emph{3}: To be indluded, an image annotation
must be tagged with all of the input tags and must not be tagged with anything else.

Note that the filtering process happens in addition to / after the filtering by input for argument
\code{ids}.
}
