% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH-M.R
\name{plotImageBase}
\alias{plotImageBase}
\alias{plotImageBase,SPATA2-method}
\alias{plotImageBase,SpatialData-method}
\alias{plotImageBase,HistoImage-method}
\alias{plotImageBase,Image-method}
\title{Plot histology image}
\usage{
plotImageBase(object, ...)

\S4method{plotImageBase}{SPATA2}(object, xrange = NULL, yrange = NULL, axes = FALSE, ...)

\S4method{plotImageBase}{SpatialData}(
  object,
  img_name = activeImage(object),
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1,
  axes = TRUE,
  ...
)

\S4method{plotImageBase}{HistoImage}(
  object,
  img_name = activeImage(object),
  xrange = NULL,
  yrange = NULL,
  scale_fct = 1,
  axes = TRUE,
  ...
)

\S4method{plotImageBase}{Image}(
  object,
  scale_fct = 1,
  xrange = NULL,
  yrange = NULL,
  axes = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{xrange, yrange}{\link[=concept_distance_measure]{Distance vector} of length
two or \code{NULL}. If not \code{NULL}, specifies the x- and y-range to which
the spatial output is cropped. E.g. \code{xrange = c(200, 500)} results in
the two dimensional space being cropped from x-coordinate 200px up to
x-coordinate 500px. If \code{NULL}, the original range is used.}

\item{img_name}{Character value. The name of the \code{\link{HistoImage}} of interest.
If \code{NULL}, the active histo image is chosen by default.}
}
\value{
A plot that is immediately plotted.
}
\description{
Plots the histology image as a raster.
}
