% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotH-M.R
\name{plotImageGgplot}
\alias{plotImageGgplot}
\title{Plot histology image (ggplot2)}
\usage{
plotImageGgplot(object, unit = "px", xrange = NULL, yrange = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{unit}{Character value. Units of x- and y-axes. Defaults
to \emph{'px'}.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{...}{Additional arguments given to \code{ggpLayerAxesEUOL()} if
\code{unit} is not \emph{'px'}.}
}
\value{
A ggplot.
}
\description{
Plots the histology image with \code{ggplot2}.
}
