% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN-S.R
\name{plotOverview}
\alias{plotOverview}
\alias{plotOverview,SpatialAnnotationScreening-method}
\alias{plotOverview,SpatialTrajectoryScreening-method}
\title{Plot overview of S4 objects}
\usage{
plotOverview(object, ...)

\S4method{plotOverview}{SpatialAnnotationScreening}(
  object,
  eval = "ias_score",
  pval = "p_value_mean_adjusted",
  pt_alpha = 0.75,
  pt_color = "black",
  pt_size = 1,
  label_vars = NULL,
  label_alpha = 0.9,
  label_color = "black",
  label_size = 2,
  model_subset = NULL,
  nrow = NULL,
  ncol = NULL,
  ...
)

\S4method{plotOverview}{SpatialTrajectoryScreening}(
  object,
  eval = "sts_score",
  pval = "p_value",
  pt_alpha = 0.75,
  pt_color = "black",
  pt_size = 1,
  label_vars = NULL,
  label_alpha = 0.9,
  label_color = "black",
  label_size = 2,
  model_subset = NULL,
  model_remove = NULL,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{eval}{Character value. The variable to use for the x-axis.}

\item{pval}{Character value. The variable to use for the y-axis.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{label_vars}{Character value, numeric value or NULL. Useful to highlight
the exact position/evalation of variables.

If character, specifies the variables that are labeled. If numeric, specifies
the top n of variables that are labeled. If NULL, ignored.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}
}
\description{
Assigns every numeric variable to the model it fitted best
against and plots the p-value of the fit against the fit evaluation.
}
