% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN-S.R
\name{plotPcaVariation}
\alias{plotPcaVariation}
\title{Plot PCA Variation}
\usage{
plotPcaVariation(object, n_pcs = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{n_pcs}{Numeric value or \code{NULL}. If numeric, specifies the number of
principal components to use. If \code{NULL}, all principal components are used
(as computed with \code{\link[=runPCA]{runPCA()}}).}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Returns a ggplot that can be additionally customized according
to the rules of the ggplot2 framework.
}
\description{
Displays a scree plot of the current principal component
analysis data stored in the object.
}
