% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotPcaVariation}
\alias{plotPcaVariation}
\title{Plot Pca Variation}
\usage{
plotPcaVariation(object, n_pcs = NULL, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be included.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays a scree plot of the current principal component
analysis data stored in the object.
}
