% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN-S.R
\name{plotSasBarplot}
\alias{plotSasBarplot}
\title{Plot SAS barplot}
\usage{
plotSasBarplot(
  object,
  grouping,
  id = idSA(object),
  distance = distToEdge(object, id),
  resolution = getCCD(object) * 2,
  unit = getDefaultUnit(object),
  angle_span = c(0, 360),
  core = FALSE,
  round = 2,
  clrp = NULL,
  clrp_adjust = NULL,
  position = "fill",
  bar_width = 0.025,
  expand_x = c(0.025, 0),
  expand_y = c(0.0125, 0),
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping}{Character value. The grouping variable of interest. Use
\code{\link[=getGroupingOptions]{getGroupingOptions()}} to obtain all valid input options.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance measure}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. Defaults to a distance
that covers the whole tissue section the spatial annotation is located
on using \code{\link[=distToEdge]{distToEdge()}}. (This distance must not be exceeded.)}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{unit}{Character value. The unit in which the distance
to the spatial annotation is displayed on the x-axis.

If \code{FALSE}, plots the bin numbers instead.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of its closest spatial annotation.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{expand_x, expand_y}{Given to argument \code{expand} of \code{ggplot2:scale_x/y_continuous()}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Needed arguments that depend on the input/unit combination. If
one of both is \emph{'px'}, argument \code{object} must be specified.}
}
\value{
A ggplot.
}
\description{
Plots changes in grouping proportion against the distance to
a spatial annotation.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

object <-
 createNumericAnnotations(
   object = object,
   variable = "HM_HYPOXIA",
   threshold = "kmeans_high",
   id = "hypoxia_ann",
   inner_borders = FALSE,
   force1 = TRUE
   )

plotSurface(object, color_by = "bayes_space") +
  ggpLayerSpatAnnOutline(object, ids = "hypoxia_ann")

plotSasBarplot(object, grouping = "bayes_space", id = "hypoxia_ann")

}
