% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R, R/plotN-S.R
\name{plotSasLineplotSC}
\alias{plotSasLineplotSC}
\alias{plotSasLineplot}
\title{Plot SAS lineplot}
\usage{
plotSasLineplotSC(
  object,
  sc_input,
  id = idSA(object),
  distance = NA_integer_,
  n_bins_dist = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  n_bins_angle = 1,
  method_gs = NULL,
  smooth_span = 0.2,
  smooth_se = FALSE,
  unit = getSpatialMethod(object)@unit,
  round = 2,
  clrp = NULL,
  clrp_adjust = NULL,
  line_color = NULL,
  line_size = 1.5,
  facet_by = "variables",
  normalize_by = "sample",
  summarize_with = "mean",
  bcsp_exclude = NULL,
  nrow = NULL,
  ncol = NULL,
  include_area = FALSE,
  display_border = TRUE,
  border_linealpha = 0.75,
  border_linecolor = "black",
  border_linesize = 1,
  border_linetype = "dashed",
  x_nth = 3,
  xi = NULL,
  yi = NULL,
  model_aid = NULL,
  verbose = NULL,
  ...
)

plotSasLineplot(
  object,
  variables,
  ids = idSA(object),
  distance = "dte",
  binwidth = recBinwidth(object),
  core = FALSE,
  angle_span = c(0, 360),
  smooth_span = 0.2,
  smooth_se = TRUE,
  unit = getSpatialMethod(object)@unit,
  clrp = NULL,
  clrp_adjust = NULL,
  line_color = NULL,
  line_size = 1.5,
  display_facets = TRUE,
  nrow = NULL,
  ncol = NULL,
  border_linealpha = 0.75,
  border_linecolor = alpha("white", 0),
  border_linesize = 1,
  border_linetype = "solid",
  display_eval = FALSE,
  eval_size = line_size * 2.5,
  ggpLayers = list(),
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{sc_input}{Data.frame that contains the results from single cell deconvolution.
Must have at least three columns:
\itemize{
\item{\emph{x} : }{ numeric. Position of cell on the x axis in pixel.},
\item{\emph{y} :}{ numeric. Position of cell on the y axis in pixel.},
\item{\emph{cell_type} :}{ factor Cell type of the cell.}
}}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be}

\item{unit}{Character value. The unit in which the distance
to the spatial annotation is displayed on the x-axis.

If \code{FALSE}, plots the bin numbers instead.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric value. Specifies the thicknes of the lines with which
the trajectory dynamics are displayed.}

\item{facet_by}{Either \emph{'variables'} or \emph{'bins_angle'}.
If \emph{'bins_angle'} length of \code{variables} must be one.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{display_border}{Logical value. If \code{TRUE}, displays a vertical line
to highlight where the border of the spatial annotation runs.}

\item{border_linealpha, border_linecolor, border_linesize, border_linetype}{Given
to \code{ggplot2::geom_vline()}. Adjusts appearance of the vertical line that
represents the border of the spatial annotation.}

\item{x_nth}{Numeric value. If the number of breaks/labels on the
x-axis becomes too high \code{x_nth} can be used to reduce it. If \code{x_nth} is 1,
every label is kept. If 2, every second label is kept. If 3, every
third label is kept. And so on.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Needed arguments that depend on the input/unit combination. If
one of both is \emph{'px'}, argument \code{object} must be specified.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{display_facets}{Logical. If set to TRUE sub plots for every specified gene, gene-set
or feature are displayed via \code{ggplot2::facet_wrap()}}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}
}
\value{
A ggplot.
}
\description{
Plots expression changes against the distance to
a spatial annotation using lineplots.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\examples{
library(SPATA2)
library(ggplot2)

data("example_data")

object <- example_data$object_UKF313T_diet

ids <- getSpatAnnIds(object, tags = c("necrotic", "compr"), test = "identical")

object <- normalizeCounts(object, activate = T)

# visualize with lines
plotSasLineplot(object, ids = ids, variables = c("VEGFA", "HM_HYPOXIA", "RCTM_TCR_SIGNALING", "CD74")) +
  labs(x = "Distance to Necrosis")

# visualize with ridgeplots
plotSasRidgeplot(object, ids = ids, variables = c("VEGFA", "HM_HYPOXIA", "RCTM_TCR_SIGNALING", "CD74")) +
  labs(x = "Distance to Necrosis")

# visualize with a heatmap
plotSasHeatmap(object, ids = ids, variables = c("VEGFA", "HM_HYPOXIA", "RCTM_TCR_SIGNALING", "CD74")) +
  labs(x = "Distance to Necrosis")

}
