% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R, R/plotN-S.R
\name{plotSasRidgeplotSC}
\alias{plotSasRidgeplotSC}
\alias{plotSasRidgeplot}
\title{Plot SAS rideplot}
\usage{
plotSasRidgeplotSC(
  object,
  sc_input,
  id = idSA(object),
  distance = NA_integer_,
  n_bins_dist = NA_integer_,
  binwidth = getCCD(object),
  angle_span = c(0, 360),
  smooth_span = 0.3,
  unit = getSpatialMethod(object)@unit,
  round = 2,
  scale = FALSE,
  clrp = NULL,
  clrp_adjust = NULL,
  alpha = 1,
  fill = NULL,
  line_color = "black",
  line_size = 1.5,
  include_area = FALSE,
  display_border = TRUE,
  border_linealpha = 0.75,
  border_linecolor = "black",
  border_linesize = 1,
  border_linetype = "dashed",
  strip_pos = "right",
  overlap = 0.5,
  free_y = TRUE,
  x_nth = 7L,
  ncol = 1,
  nrow = NULL,
  verbose = NULL,
  ...
)

plotSasRidgeplot(
  object,
  variables,
  ids = idSA(object),
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  angle_span = c(0, 360),
  core = FALSE,
  smooth_span = 0.3,
  unit = getSpatialMethod(object)@unit,
  alpha = 1,
  fill = NULL,
  clrp = NULL,
  clrp_adjust = NULL,
  line_color = "black",
  line_size = 1.5,
  border_linealpha = 0.75,
  border_linecolor = "black",
  border_linesize = 1,
  border_linetype = "dashed",
  overlap = 0.5,
  strip_pos = "right",
  free_y = FALSE,
  ggpLayers = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{sc_input}{Data.frame that contains the results from single cell deconvolution.
Must have at least three columns:
\itemize{
\item{\emph{x} : }{ numeric. Position of cell on the x axis in pixel.},
\item{\emph{y} :}{ numeric. Position of cell on the y axis in pixel.},
\item{\emph{cell_type} :}{ factor Cell type of the cell.}
}}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{unit}{Character value. Specifies the desired unit.}

\item{round}{Numeric value or \code{FALSE}. If numeric, given to \code{digits}
of \code{base::round()}. Rounds transformed values before they are returned.}

\item{scale}{Logical value. If \code{TRUE}, density of cell types is scaled
to make them comparable. Else, the absolute values defined by count/\code{unit}^2
is plotted.}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric value. Specifies the thicknes of the lines with which
the trajectory dynamics are displayed.}

\item{display_border}{Logical value. If \code{TRUE}, displays a vertical line
to highlight where the border of the spatial annotation runs.}

\item{border_linealpha, border_linecolor, border_linesize, border_linetype}{Given
to \code{ggplot2::geom_vline()}. Adjusts appearance of the vertical line that
represents the border of the spatial annotation.}

\item{x_nth}{Numeric value. If the number of breaks/labels on the
x-axis becomes too high \code{x_nth} can be used to reduce it. If \code{x_nth} is 1,
every label is kept. If 2, every second label is kept. If 3, every
third label is kept. And so on.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Needed arguments that depend on the input/unit combination. If
one of both is \emph{'px'}, argument \code{object} must be specified.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}
}
\value{
A ggplot.
}
\description{
Plots gene expression changes against the distance to
to the spatial annotation using the design of ridgeplots.
}
\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

\examples{
library(SPATA2)
library(ggplot2)

data("example_data")

object <- example_data$object_UKF313T_diet

ids <- getSpatAnnIds(object, tags = c("necrotic", "compr"), test = "identical")

object <- normalizeCounts(object, activate = T)

# visualize with lines
plotSasLineplot(object, ids = ids, variables = c("VEGFA", "HM_HYPOXIA", "RCTM_TCR_SIGNALING", "CD74")) +
  labs(x = "Distance to Necrosis")

# visualize with ridgeplots
plotSasRidgeplot(object, ids = ids, variables = c("VEGFA", "HM_HYPOXIA", "RCTM_TCR_SIGNALING", "CD74")) +
  labs(x = "Distance to Necrosis")

# visualize with a heatmap
plotSasHeatmap(object, ids = ids, variables = c("VEGFA", "HM_HYPOXIA", "RCTM_TCR_SIGNALING", "CD74")) +
  labs(x = "Distance to Necrosis")

}
