% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotN-S.R
\name{plotScatterplot}
\alias{plotScatterplot}
\title{Plot numeric variables as a scatterplot}
\usage{
plotScatterplot(
  object,
  variables,
  smooth = NULL,
  smooth_clr = NULL,
  smooth_method = NULL,
  smooth_se = NULL,
  smooth_span = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_size = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. The names of the \emph{\link[=concept_variables]{data variables}}
of interest.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_clr}{Character value. The color with which to display the smoothed
model.}

\item{smooth_method}{The smoothing method that will be used specified as a
single character value (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{ggplot2::geom_smooth()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Use argument \code{variables} to denote the numeric variables
of interest. First one will be mapped on to the x-axis and the second
one on to the y-axis.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF275T_diet

plotScatterplot(object, variables = c("HM_HYPOXIA", "METRN"))
}
