% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(
  object,
  encircle = TRUE,
  segment_subset = NULL,
  pt_size = NULL,
  pt_clrp = NULL,
  clrp_adjust = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{encircle}{Logical. If set to TRUE the segments are enclosed in a polygon.}

\item{segment_subset}{Character vector or NULL. If character vector, denotes
the segments that are supposed to be highlighted.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to \code{confuns::scale_color_add_on()}.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the segmentation of a specified sample that was drawn with
\code{SPATA::createSegmentation()}.
}
