% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{plotSummary}
\alias{plotSummary}
\alias{plotSummary,ImageAnnotationScreening-method}
\title{Plot summary of spatial fitting}
\usage{
plotSummary(object, ...)

\S4method{plotSummary}{ImageAnnotationScreening}(
  object,
  x = "corr_mean",
  y = "pvalue_mean",
  pt_alpha = 0.75,
  pt_color = "black",
  pt_size = 1,
  display_labels = FALSE,
  model_subset = NULL,
  model_remove = NULL,
  pretty_names = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{x, y}{Character value. Specifies what to map to the x- and what to
map to the y-axis.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}
}
\description{
Assigns every numeric variable to the model it fitted best
against and plots the p-value of the fit against the fit evaluation.
}
