% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurface}
\alias{plotSurface}
\alias{plotSurface2}
\alias{plotSurfaceInteractive}
\title{Plot the surface of the sample}
\usage{
plotSurface(
  object,
  color_by = NULL,
  alpha_by = NULL,
  method_gs = NULL,
  normalize = NULL,
  smooth = NULL,
  smooth_span = NULL,
  pt_alpha = NULL,
  pt_clr = NULL,
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  pt_size_fixed = NULL,
  clrp_adjust = NULL,
  display_image = NULL,
  display_title = NULL,
  complete = NULL,
  verbose = NULL,
  highlight_groups = NULL,
  transform_with = NULL,
  na_rm = FALSE,
  of_sample = NA,
  ...
)

plotSurface2(
  coords_df,
  color_by,
  alpha_by = NULL,
  pt_alpha = 0.9,
  pt_clr = "lightgrey",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 2,
  image = NULL,
  clrp_adjust = NULL,
  ...
)

plotSurfaceInteractive(object)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{color_by}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{normalize}{Logical. If set to TRUE values will be scaled to 0-1.

Hint: Variables that are uniformly expressed can not be scaled and are discarded.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{display_title}{Logical value. If set to TRUE an informative title is displayed.}

\item{complete}{Logical. If the provided spata-object has been subsetted by
\code{subsetBySegment()} the original sample is completed with grey barcode
spots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\itemize{
 \item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
 \item{\emph{sample} Character. The sample belonging of every barcode-spot.}
 \item{\emph{x} Numeric. The x-coordinates of every barcode.}
 \item{\emph{y} Numeric. The y-coordinates of every barcode.}
}}

\item{image}{An image of class \emph{Image} to be displayed in the background.
Easily accessible via \code{SPATA::image()}.}
}
\description{
Displays the spatial dimension of the sample and colors the
surface according to the expression of genes, gene sets or features.

\itemize{

 \item{ \code{plotSurface()} Takes the spata-object as the starting point and creates the
 necessary data.frame from scratch according to additional parameters.}
 \item{ \code{plotSurface2()} Takes a data.frame as the starting point.}
 \item{ \code{plotSurfaceInteractive()} Takes only the spata-object and opens a shiny
 application which allows for interactive plotting.}

}
}
