% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurfaceAverage}
\alias{plotSurfaceAverage}
\title{Plot several surface plots colored by gene averages}
\usage{
plotSurfaceAverage(
  object,
  color_by,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  smooth = FALSE,
  smooth_span = NULL,
  use_scattermore = FALSE,
  sctm_pixels = c(1024, 1024),
  sctm_interpolate = FALSE,
  display_image = NULL,
  bcsp_rm = NULL,
  na_rm = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{color_by}{A character vector of gene names or a
named list in which each element is a vector of gene names.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{sctm_interpolate, sctm_pixels}{Given to the corresponding arguments
of \code{scattermore::geom_scattermore()}. Note: With increasing \code{sctm_pixels}
the point size must be adjusted with the argument \code{pt_size}.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{bcsp_rm}{Character vector or NULL. If character, specifies barcode-spots that
are removed before plotting.}
}
\description{
Displays the spatial dimension of the sample and colors the
surface according to the expression of averaged gene expression values.
For each element in the list specified in argument \code{color_by} a
surface plot is generated colored by the gene's average expression score.
}
\examples{
#Not run:

color_by_list <- list(Example1 = c("PGK1", "PDK1", "GBE1"),
                      Example2 = c("METRN", "GFAP")
                      )

plotSurfaceAverage(
   object = spata_obj,
   color_by = color_by_list
 )

}
