% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-surface.R
\name{plotSurfaceAverage}
\alias{plotSurfaceAverage}
\title{Plot several surface plots colored by gene averages}
\usage{
plotSurfaceAverage(
  object,
  color_by,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_size = NULL,
  smooth = NULL,
  smooth_span = NULL,
  of_sample = NA
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{color_by}{A character vector of gene names or a
named list in which each element is a vector of gene names.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.
displayed.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\description{
Displays the spatial dimension of the sample and colors the
surface according to the expression of averaged gene expression values.
For each element in the list specified in argument \code{color_by} a
surface plot is generated colored by the gene's average expression score.
}
\examples{
#Not run:

color_by_list <- list(Example1 = c("PGK1", "PDK1", "GBE1"),
                      Example2 = c("METRN", "GFAP")
                      )

plotSurfaceAverage(object = spata_obj,
                   color_by = color_by_list)

}
