% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurfaceBase.R
\name{plotSurfaceBase}
\alias{plotSurfaceBase}
\title{Plot surface with R base plotting}
\usage{
plotSurfaceBase(
  object,
  color_by = NULL,
  pt_alpha = 0.9,
  pt_color = "grey",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 1,
  clrp_adjust = NULL,
  smooth = FALSE,
  smooth_span = 0.2,
  display_axes = FALSE,
  display_image = NULL,
  highlight_barcodes = NULL,
  highlight_alpha = 0.75,
  highlight_color = "orange",
  xrange = NULL,
  yrange = NULL,
  adjust_pt_size = TRUE,
  padding = 0,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{color_by}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{smooth}{Logical. If set to TRUE values will be smoothed according to the
\code{smoooth_}-parameters.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{xrange}{Numeric vectors of length two or NULL. If numeric, specifies the coordinates
with which the image is subsetted. E.g. xrange = c(200, 500) results in the image being cropped
from x-coordinate 200 up to x-coordinate 500.}

\item{yrange}{Numeric vectors of length two or NULL. If numeric, specifies the coordinates
with which the image is subsetted. E.g. xrange = c(200, 500) results in the image being cropped
from x-coordinate 200 up to x-coordinate 500.}

\item{padding}{Numeric value. If bigger than 0 creates a padding around the cropped image that never
exceeds the actual limits of the image. Barcode spots are not affected.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Plots right into the plotting window.
}
\description{
Uses Rs base plotting device instead of ggplot2. This
is usually faster but can not make use of the mechanism ggplot2 offers.
}
