% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurfaceBase}
\alias{plotSurfaceBase}
\title{Plot surface with R base plotting}
\usage{
plotSurfaceBase(
  object,
  color_by = NULL,
  alpha_by = NULL,
  pt_alpha = 0.9,
  pt_color = "grey",
  pt_clrp = "milo",
  pt_clrsp = "inferno",
  pt_size = 1,
  clrp_adjust = NULL,
  smooth = FALSE,
  smooth_span = 0.2,
  display_axes = FALSE,
  display_image = NULL,
  highlight_barcodes = NULL,
  highlight_alpha = 0.75,
  highlight_color = "orange",
  xrange = NULL,
  yrange = NULL,
  adjust_pt_size = TRUE,
  expand = 0,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{color_by}{The variable to be displayed by color:

\itemize{
\item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
\item{ \strong{Genes} as a character vector. If more than one gene is specified the average
expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
\item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
}}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{smooth}{Logical. If TRUE, a loess fit is used to smooth the values.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{display_image}{Logical. If set to TRUE the histology image of the specified sample
is displayed underneath the plot.}

\item{xrange, yrange}{Vector of length two or \code{NULL}. If not \code{NULL},
specifies the x- and y-range to which the output image is cropped. E.g.
\code{xrange = c(200, 500)} results in the image being cropped from
x-coordinate 200px up to x-coordinate 500px. If \code{NULL}, the original image
ranges are taken.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{expand}{Vector of length one or two. Specifies the extent to which the  x- and y-span
of the cropped image section is expanded. There are two options to provide values:

Values from 0 to 1 are used to calculate the corresponding percentage of the image
span which is then \bold{added} to it. E.g. \code{expand} = 1 doubles the span,
\code{expand} = 0.5 results in 150 percent of the span and \code{expand} = 0 returns the
original span.

Values bigger than 1 are considered to be absolute values. The returned image
section has the exact x- and y- spans as specified in \code{expand}.
E.g. \code{expand} = 200 results in an image section with x- and y-span of
200. If the input value is bigger than 1 it must be at least as big as the
original span of the image annotation.

To adjust x- and y-span specifically you can provide a vector of length 2.
If you do so the first value is taken to adjust the x-span and the second
value is used to adjust the y-span E.g. \code{expand} = c(0.5, 0) adds 50 percent
of the original x-span to the x-span of the image section and does not adjust
the y-span.

This argument works within the \code{SPATA2} distance framework.
If values are specified in European units of length the input is
immediately converted to pixel units.

See details and examples of \code{?is_dist} and \code{?as_unit} for more information.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
Plots right into the plotting window.
}
\description{
Uses Rs base plotting device instead of ggplot2. This
is usually faster but can not make use of the mechanism ggplot2 offers.
}
