% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{plotSurfaceIAS}
\alias{plotSurfaceIAS}
\alias{plotSurfaceIAS,spata2-method}
\alias{plotSurfaceIAS,ImageAnnotationScreening-method}
\title{Visualize screening areaof IAS-algorithm}
\usage{
plotSurfaceIAS(object, ...)

\S4method{plotSurfaceIAS}{spata2}(
  object,
  id,
  buffer,
  n_bins_circle,
  n_bins_angle = 12,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = NULL,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)

\S4method{plotSurfaceIAS}{ImageAnnotationScreening}(
  object,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = 2.25,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}.}

\item{n_bins_circle}{Numeric value (or vector of length two, see details
for more). Specifies how many times the area is buffere with the value
denoted in \code{buffer}. The combination of both results in area that is
screened.}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.}

\item{color_core, color_outside}{Character value. Denotes
the colors with which the area of image annotation (\code{color_core})
and the area that is not included in the screening (\code{color_outside})
is displayed.}

\item{show_plots}{Logical value. If TRUE, the plots are immediately
plotted. If FALSE, only a list of plots is returned (invisibly).}

\item{display_angle, display_bins_angle, display_circle}{Logical value.
If TRUE, the plot is included. If FALSE, plotting is skipped.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}
}
\value{
An invisible list of ggplots.
}
\description{
Plots the surface of the sample three times with different
coloring to visualize how \code{imageAnnotationScreening()} screens
the sample depending on the input of arguments \code{buffer}, \code{n_bins_circle},
\code{n_bins_angle}.
}
\details{
The method for class \code{ImageAnnotationScreening} (the output of
the function \code{imageAnnotationScreening()}) can be used
to show the area on which the results base. Therefore, it does not have
arguments \code{buffer}, \code{n_bins_circle} and \code{n_bins_angle}.
}
