% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-annotation-screening.R
\name{plotSurfaceIAS}
\alias{plotSurfaceIAS}
\alias{plotSurfaceIAS,spata2-method}
\alias{plotSurfaceIAS,ImageAnnotationScreening-method}
\title{Visualize screening areaof IAS-algorithm}
\usage{
plotSurfaceIAS(object, ...)

\S4method{plotSurfaceIAS}{spata2}(
  object,
  id,
  n_bins_circle,
  buffer,
  angle_span = c(0, 360),
  n_bins_angle = 12,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = NULL,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)

\S4method{plotSurfaceIAS}{ImageAnnotationScreening}(
  object,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = 2.25,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{id}{Character value. The ID of the image annotation of interest.}

\item{n_bins_circle}{Numeric value or vector of length 2. Specifies how many times the area is buffered with the value
denoted in \code{buffer}.
 (See details for more.)}

\item{buffer}{The distance by which to consecutively expand the
area that covers the image annotation screening. Given to argument
\code{dist} of function \code{sf::st_buffer()}. (See details for more.)}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the image annotation.
 (See details fore more.)}

\item{n_bins_angle}{Numeric value. Number of bins that are created by angle.
(See details for more.)}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{color_core, color_outside}{Character value. Denotes
the colors with which the area of image annotation (\code{color_core})
and the area that is not included in the screening (\code{color_outside})
is displayed.}

\item{show_plots}{Logical value. If TRUE, the plots are immediately
plotted. If FALSE, only a list of plots is returned (invisibly).}

\item{display_angle, display_bins_angle, display_circle}{Logical value.
If TRUE, the plot is included. If FALSE, plotting is skipped.}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}
}
\value{
An invisible list of ggplots.
}
\description{
Plots the surface of the sample three times with different
coloring to visualize how \code{imageAnnotationScreening()} screens
the sample depending on the input of arguments \code{buffer}, \code{n_bins_circle},
\code{n_bins_angle}.
}
\details{
The method for class \code{ImageAnnotationScreening} (the output of
the function \code{imageAnnotationScreening()}) can be used
to show the area on which the results base. Therefore, it does not have
arguments \code{buffer}, \code{n_bins_circle} and \code{n_bins_angle}.
}
