% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurfaceSAS}
\alias{plotSurfaceSAS}
\alias{plotSurfaceSAS,SPATA2-method}
\alias{plotSurfaceSAS,SpatialAnnotationScreening-method}
\title{Plot screening area of SAS set up}
\usage{
plotSurfaceSAS(object, ...)

\S4method{plotSurfaceSAS}{SPATA2}(
  object,
  ids,
  distance = distToEdge(object, id),
  resolution = recBinwidth(object),
  color_by = c("dist", "bins_dist", "angle", "bins_angle"),
  unit = getDefaultUnit(object),
  angle_span = c(0, 360),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  ggpLayers = list(),
  bcs_exclude = NULL,
  verbose = NULL,
  ...
)

\S4method{plotSurfaceSAS}{SpatialAnnotationScreening}(
  object,
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_dist = TRUE,
  ggpLayers = list(),
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{ids}{Character vector. Specifies the IDs of the spatial annotations of interest.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance measure}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. Defaults to a distance
that covers the whole tissue section the spatial annotation is located
on using \code{\link[=distToEdge]{distToEdge()}}. (This distance must not be exceeded.)}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{color_by}{Character value. The \link[=concept_variable]{variables} by which
to color the data points.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of its closest spatial annotation.}

\item{show_plots}{Logical value. If TRUE, the plots are immediately
plotted. If FALSE, only a list of plots is returned (invisibly).}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{bcs_exclude}{Character value containing the barcodes of observations to be excluded
from the analysis.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{display_angle, display_bins_angle, display_circle}{Logical value.
If TRUE, the plot is included. If FALSE, plotting is skipped.}

\item{color_core, color_outside}{Character value. Denotes
the colors with which the area of image annotation (\code{color_core})
and the area that is not included in the screening (\code{color_outside})
is displayed.}
}
\value{
An invisible list of ggplots.
}
\description{
Plots the surface of the sample three times with different
coloring to visualize how \code{\link[=spatialAnnotationScreening]{spatialAnnotationScreening()}} screens
the sample depending on the input of arguments \code{binwidth}, \code{n_bins_dist},
\code{n_bins_angle}.
}
\details{
The method for class \code{SpatialAnnotationScreening} (the output of
the function \code{spatialAnnotationScreening()}) can be used
to show the area on which the results are based on. Therefore, it does not have
arguments \code{binwidth}, \code{n_bins_circle} and \code{n_bins_angle}.
}
