% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSurface.R
\name{plotSurfaceSAS}
\alias{plotSurfaceSAS}
\alias{plotSurfaceSAS,SPATA2-method}
\alias{plotSurfaceSAS,SpatialAnnotationScreening-method}
\title{Plot screening area of SAS-algorithm}
\usage{
plotSurfaceSAS(object, ...)

\S4method{plotSurfaceSAS}{SPATA2}(
  object,
  id,
  distance = distToEdge(object, id),
  binwidth = recBinwidth(object),
  n_bins_dist = NA_integer_,
  color_by = c("dist", "bins_dist", "angle", "bins_angle"),
  unit = extract_unit(binwidth),
  angle_span = c(0, 360),
  n_bins_angle = 1,
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  ggpLayers = list(),
  bcsp_exclude = NULL,
  verbose = NULL,
  ...
)

\S4method{plotSurfaceSAS}{SpatialAnnotationScreening}(
  object,
  pt_alpha = NA_integer_,
  pt_clrp = c("inferno", "default"),
  pt_clrsp = "inferno",
  pt_size = 2.25,
  color_core = ggplot2::alpha("grey", 0),
  color_outside = ggplot2::alpha("lightgrey", 0.25),
  show_plots = TRUE,
  display_angle = FALSE,
  display_bins_angle = TRUE,
  display_bins_circle = TRUE,
  ggpLayers = list(),
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{color_by}{Character value. The \link[=concept_variable]{variables} by which
to color the data points.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{show_plots}{Logical value. If TRUE, the plots are immediately
plotted. If FALSE, only a list of plots is returned (invisibly).}

\item{ggpLayers}{List of \code{ggproto}-objects that are added to each plot.
Skim \code{ggpLayer*()}-functions for more options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{color_core, color_outside}{Character value. Denotes
the colors with which the area of image annotation (\code{color_core})
and the area that is not included in the screening (\code{color_outside})
is displayed.}

\item{display_angle, display_bins_angle, display_circle}{Logical value.
If TRUE, the plot is included. If FALSE, plotting is skipped.}
}
\value{
An invisible list of ggplots.
}
\description{
Plots the surface of the sample three times with different
coloring to visualize how \code{\link[=spatialAnnotationScreening]{spatialAnnotationScreening()}} screens
the sample depending on the input of arguments \code{binwidth}, \code{n_bins_dist},
\code{n_bins_angle}.
}
\details{
The method for class \code{SpatialAnnotationScreening} (the output of
the function \code{spatialAnnotationScreening()}) can be used
to show the area on which the results are based on. Therefore, it does not have
arguments \code{binwidth}, \code{n_bins_circle} and \code{n_bins_angle}.
}
