% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{plotTrajectoryBarplot}
\alias{plotTrajectoryBarplot}
\title{Plot categorical trajectory dynamics}
\usage{
plotTrajectoryBarplot(
  object,
  id,
  grouping_variable,
  binwidth = ccDist(object),
  clrp = NULL,
  clrp_adjust = NULL,
  display_trajectory_parts = NULL,
  position = "fill",
  scales = "free_x",
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{display_trajectory_parts}{Logical. If set to TRUE the returned plot
visualizes the parts in which the trajectory has been partitioned while beeing
drawn.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()}.}

\item{feature}{Character value. The grouping feature of interest.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays discrete variables along a trajectory.
}
