% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotT-Z.R
\name{plotTrajectoryHeatmap}
\alias{plotTrajectoryHeatmap}
\title{Plot trajectory expression dynamic in heatmap}
\usage{
plotTrajectoryHeatmap(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  arrange_rows = "none",
  colors = NULL,
  method_gs = NULL,
  show_rownames = NULL,
  show_colnames = NULL,
  split_columns = NULL,
  smooth_span = NULL,
  multiplier = 10,
  with_ggplot = TRUE,
  display_trajectory_parts = FALSE,
  display_parts_with = "lines",
  line_alpha = 1,
  line_color = "red",
  line_size = 1,
  line_type = "dashed",
  clrsp = NULL,
  .f = NULL,
  .cols = dplyr::everything(),
  summarize_with = "mean",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{The variables of interest specified as a character vector:

\itemize{
\item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
\item{ \strong{Genes}: Must be in \code{getGenes()}}
}

All elements of the specified character vector must either belong to
gene sets or to genes.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{arrange_rows}{Alter the way the rows of the heatmap
are displayed in order to highlight patterns. Currently either \emph{'maxima'},
\emph{'minima'} or \emph{'input'}. If \emph{'input'}, variables are displayed
in the same order as they are provided in the argument \code{variables}.}

\item{colors}{A vector of colors to be used.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{show_rownames}{Logical. If set to TRUE the variable elements
will be displayed at the rownames of the heatmap.}

\item{split_columns}{Logial. If set to TRUE the heatmap is vertically
splitted according to the trajectory parts.}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{multiplier}{Numeric value. For better visualization the transient pattern
is smoothed with a loess fit. The total number of predicted values (via \code{stats::predict()})
is the number of bins multiplied with the input for this argument.}

\item{with_ggplot}{Logical value. If set to TRUE the heatmap is plotted with
\code{ggplot2::geom_tile()} and a ggplot is returned.}

\item{display_trajectory_parts}{Logical value. If TRUE and the trajectory
contains more than one part the parts are displayed by facetting the plot
or by adding vertical lines.}

\item{display_parts_with}{Character value. Either \emph{'facets'} or \emph{'lines'}.}

\item{line_alpha}{Numeric value. Specifies the transparency of the lines.}

\item{line_color}{Character. Affects color of the main lines of the plot.}

\item{line_size}{Numeric. Affects size of the main lines of the plot.}

\item{line_type}{Character. The line type. One of \emph{'blank'}, \emph{'solid'},
\emph{'dashed'}, \emph{'dotted'}, \emph{'dotdash'}, \emph{'longdash'} and \emph{'twodash'}.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{summarize_with}{Character value. Name of the function with which to summarize.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap()}. If argument
\code{with_ggplot} is TRUE given to \code{scale_color_add_on()}.}
}
\value{
A heatmap of class 'pheatmap' or a ggplot.
}
\description{
Displays variable-expression values along a trajectory
direction with a smoothed heatmap (from left to right).
}
