% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryHeatmap}
\alias{plotTrajectoryHeatmap}
\title{Plot trajectory expression dynamic in heatmap}
\usage{
plotTrajectoryHeatmap(
  object,
  trajectory_name = getDefaultTrajectory(object, verbose = TRUE, "trajectory_name"),
  variables,
  binwidth = 5,
  arrange_rows = "none",
  colors = NULL,
  method_gs = NULL,
  show_rownames = NULL,
  show_colnames = NULL,
  split_columns = NULL,
  smooth_span = NULL,
  multiplier = 10,
  with_ggplot = FALSE,
  display_trajectory_parts = FALSE,
  display_parts_with = "lines",
  linealpha = 1,
  linecolor = "red",
  linesize = 1,
  linetype = "dashed",
  clrsp = NULL,
  .f = NULL,
  .cols = dplyr::everything(),
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{variables}{The variables of interest specified as a character vector:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 }

All elements of the specified character vector must either belong to
gene sets or to genes.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{arrange_rows}{Alter the way the rows of the heatmap
are displayed in order to highlight patterns. Currently either \emph{'maxima'}
or \emph{'minima'}.}

\item{colors}{A vector of colors to be used.}

\item{show_rownames}{Logical. If set to TRUE the variable elements
will be displayed at the rownames of the heatmap.}

\item{split_columns}{Logial. If set to TRUE the heatmap is vertically
splitted according to the trajectory parts.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.}

\item{with_ggplot}{Logical value. If set to TRUE the heatmap is plotted with
\code{ggplot2::geom_tile()} and a ggplot is returned.}

\item{display_trajectory_parts}{Logical value. If TRUE and the trajectory
contains more than one part the parts are displayed by facetting the plot
or by adding vertical lines.}

\item{display_parts_with}{Character value. Either \emph{'facets'} or \emph{'lines'}.}

\item{clrsp}{Character value. Specifies the color spectrum to be used to represent
continuous values of numeric variables. Run \code{validColorSpectra()} to obtain
valid input options.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap()}. If argument
\code{with_ggplot} is TRUE given to \code{scale_color_add_on()}.}
}
\value{
A heatmap of class 'pheatmap' or a ggplot.
}
\description{
Displays variable-expression values along a trajectory
direction with a smoothed heatmap (from left to right).
}
