% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-trajectory.R
\name{plotTrajectoryHeatmap}
\alias{plotTrajectoryHeatmap}
\title{Plot trajectory expression dynamic in heatmap}
\usage{
plotTrajectoryHeatmap(
  object,
  trajectory_name,
  variables,
  binwidth = 5,
  arrange_rows = "none",
  colors = NULL,
  method_gs = NULL,
  show_rownames = NULL,
  show_colnames = NULL,
  split_columns = NULL,
  smooth_span = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{trajectory_name}{The trajectory of interest specified
as a single character value.}

\item{variables}{The variables of interest specified as a character vector:

\itemize{
 \item{ \strong{Gene sets}: Must be in \code{getGeneSets()}}
 \item{ \strong{Genes}: Must be in \code{getGenes()}}
 }

All elements of the specified character vector must either belong to
gene sets or to genes.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{arrange_rows}{Alter the way the rows of the heatmap
are displayed in order to highlight patterns. Currently either \emph{'maxima'}
or \emph{'minima'}.}

\item{colors}{A vector of colors to be used.}

\item{show_rownames}{Logical. If set to TRUE the variable elements
will be displayed at the rownames of the heatmap.}

\item{split_columns}{Logial. If set to TRUE the heatmap is vertically
splitted according to the trajectory parts.}

\item{smooth_span}{The amount of smoothing specified as a single numeric value.
displayed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional parameters given to \code{pheatmap::pheatmap()}}
}
\value{
A heatmap of class 'pheatmap'.
}
\description{
Displays variable-expression values along a trajectory
direction with a smoothed heatmap (from left to right).
}
