% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-screening.R
\name{plotTrajectoryLineplot}
\alias{plotTrajectoryLineplot}
\title{Plot continuous trajectory dynamics in lineplots}
\usage{
plotTrajectoryLineplot(
  object,
  id,
  variables,
  binwidth = 5,
  method_gs = NULL,
  smooth_method = NULL,
  smooth_span = NULL,
  smooth_se = NULL,
  clrp = NULL,
  clrp_adjust = NULL,
  display_trajectory_parts = NULL,
  display_facets = NULL,
  linesize = 1.5,
  vlinecolor = "grey",
  vlinesize = 1,
  vlinetype = "dashed",
  summarize_with = "mean",
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{binwidth}{Numeric value. Denotes the binwidth with which to sort all
relevant barcode spots into groups that are then aligned with respect to the
chosen trajectory's direction.#'}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{smooth_method}{The smoothing method that will be used specified as a
single character value (e.g. \emph{"lm", "glm", "gam", "loess"}).}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{smooth_se}{Logical. If set to TRUE the confidence interval will be}

\item{clrp}{Character value. Specifies the color palette to be used to represent
groups of discrete variables. Run \code{validColorPalettes()} to obtain valid
input options.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{display_trajectory_parts}{Logical. If set to TRUE the returned plot
visualizes the parts in which the trajectory has been partitioned while beeing
drawn.}

\item{display_facets}{Logical. If set to TRUE sub plots for every specified gene, gene-set
or feature are displayed via \code{ggplot2::facet_wrap()}}

\item{linesize}{Numeric value. Specifies the thicknes of the lines with which
the trajectory dynamics are displayed.}

\item{vlinesize, vlinecolor}{Adjusts size and color of vertical lines that
display the trajectory parts.}

\item{vlinetype}{Adjusts the type of the vertical lines that display the trajectory
parts.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{ggplot2::facet_wrap()} if argument
\code{display_facets} is set to TRUE.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays values along a trajectory direction with
a smoothed lineplot.
}
