% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{plotTrajectoryLineplotFitted}
\alias{plotTrajectoryLineplotFitted}
\title{Plot trajectory model fitting}
\usage{
plotTrajectoryLineplotFitted(
  object,
  id,
  variables,
  binwidth = getCCD(object),
  n_bins = NA_integer_,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  method_gs = NULL,
  smooth_span = 0,
  lineorder = c(1, 2, 3),
  linesizes = c(1, 1, 1),
  linecolors = c("forestgreen", "blue4", "red3"),
  linetypes = c("solid", "solid", "dotted"),
  display_residuals = TRUE,
  area_alpha = 0.25,
  display_points = TRUE,
  pt_alpha = 0.9,
  pt_size = 1.5,
  nrow = NULL,
  ncol = NULL,
  force_grid = FALSE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. The names of the \emph{\link[=concept_variables]{data variables}}
of interest.}

\item{n_bins}{Numeric value. Specifies the exact number of bins the barcodes
are binned into.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{method_gs}{Character value. The method according to which gene sets will
be handled specified as a character of length one. This can be either 'mean
or one of 'gsva', 'ssgsea', 'zscore', or 'plage'. The latter four will be given to gsva::GSVA().}

\item{smooth_span}{Numeric value. Controls the degree of smoothing.
Given to argument \code{span} of \code{stats::loess()}.}

\item{linecolors, linetypes}{The colors and types of the three lines. First value stands for the
values of the variable, second on for the models, third one for the residuals.}

\item{display_residuals}{Logical value. If TRUE, the residuals curve is displayed.}

\item{area_alpha}{Numeric value. The alpha value for the area under the curve
of the resiudals.}

\item{display_points}{Logical value. If set to TRUE points are used additionally
to display the results.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{nrow, ncol}{Numeric values or NULL. Used to arrange multiple plots.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A ggplot.
}
\description{
Plots a trajectory lineplot in combination with models
fitted to the course of the trajectory.
}
\keyword{internal}
