% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-family-miscellaneous.R
\name{plotUmap}
\alias{plotUmap}
\alias{plotTsne}
\alias{plotPca}
\title{Plot dimensional reduction}
\usage{
plotUmap(
  object,
  color_by = NULL,
  color_aes = "color",
  color_trans = "identity",
  alpha_by = NULL,
  order_by = NULL,
  order_desc = FALSE,
  shape_by = NULL,
  method_gs = NULL,
  pt_shape = 19,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_clrp = NULL,
  pt_clr = NULL,
  clrp_adjust = NULL,
  normalize = NULL,
  transform_with = list(),
  verbose = NULL,
  of_sample = NA,
  ...
)

plotTsne(
  object,
  color_by = NULL,
  color_aes = "color",
  color_trans = "identity",
  alpha_by = NULL,
  order_by = NULL,
  order_desc = FALSE,
  pt_shape = 19,
  shape_by = NULL,
  method_gs = NULL,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrsp = NULL,
  pt_clrp = NULL,
  pt_clr = NULL,
  clrp_adjust = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)

plotPca(
  object,
  color_by = NULL,
  n_pcs = NULL,
  method_gs = NULL,
  pt_size = NULL,
  pt_alpha = NULL,
  pt_clrp = NULL,
  pt_clrsp = NULL,
  pt_clr = NULL,
  normalize = NULL,
  verbose = NULL,
  of_sample = NA,
  ...
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{color_by}{The variable to be displayed by color:

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeaturenNames()}}
  }}

\item{color_aes}{Either \emph{'color'} or \emph{'fill'}. Specifies the
aesthetic that is used to visualize the variable that is specified in \code{color_by}.
Ignored if \code{color_by} is NULL.}

\item{color_trans}{Character value. Name of the transformation method applied to the continuous scale. Use
\code{validScaleTransformations()} to obtain all valid input options. Use \emph{'identity'}
for no transformation.}

\item{alpha_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{order_by}{Character value or NULL. If character, specifies the
numeric variable by which the points are arranged before plotting.}

\item{shape_by}{Character
value or NULL. If character, specifies the variable that is mapped to the respective
aesthetic of the plot.}

\item{method_gs}{Character value. The method according to which gene sets will be handled
specified as a character of length one. This can be either \emph{'mean'} or one
of \emph{'gsva', 'ssgsea', 'zscore', or 'plage'}. The latter four will be given to
\code{gsva::GSVA()}.}

\item{pt_shape}{Numeric or character value. If numeric, the respective
shape is taken for all points. If character, the respective variable is
mapped to the shape-aesthetic.}

\item{pt_size}{Numeric value. Specifies the size of all points.}

\item{pt_alpha}{Numeric value. Specifies the degree of transparency of all points.}

\item{pt_clrsp}{The color spectrum to be used if the specified variable displayed by
color is continuous. Run \code{validColorSpectra()} to see valid input.}

\item{pt_clrp}{The color palette to be used if the specified variable displayed by
color is categorical/discrete. Run \code{validColorPalettes()} to see valid input.}

\item{pt_clr}{Character value. Specifies the color of all points.}

\item{clrp_adjust}{Named character vector or NULL. If character, it adjusts the
color palette that is used to represent the groups. Names of the input vector must refer
to the group and the respective named element denotes the color with which to
represent the group.}

\item{transform_with}{List or NULL. If list, can be used to transform continuous variables before plotting.
Names of the list slots refer to the variable. The content of the slot refers to the transforming functions.
Slot content can either be a character vector of function names. Use \code{validScaleTransformations()} to obtain all valid character value inputs.
Or it can be a list of functions (and function names).

Useful if you want to apply more than one transformation on variables mapped to
plotting aesthetics. Input for \code{transform_with} is applied before the
respective \code{<aes>_trans} argument.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}

\item{n_pcs}{Numeric value. Determines the number of principal components to be plotted.
Must be an even number.}
}
\value{
Returns a ggplot-object that can be additionally customized according
to the rules of the ggplot2-framework.
}
\description{
Displays the dimensional reduction and maps gene, gene-set
or feature information onto the color-aesthetic.
}
