% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-functions.R
\name{printAutoencoderSummary}
\alias{printAutoencoderSummary}
\title{Print autoencoder summary}
\usage{
printAutoencoderSummary(object, mtr_name = "denoised", of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A human readable report of the issue of interest.
}
\description{
Prints a human readable summary about the set up of the last neural network that
was constructed to generate a denoised expression matrix.
}
