% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{printGeneSetOverview}
\alias{printGeneSetOverview}
\title{Deprecated}
\usage{
printGeneSetOverview(object, ...)
}
\description{
Deprecated in favor of \code{\link[=getGeneSetOverview]{getGeneSetOverview()}}
}
\keyword{internal}
