% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-seurat-object.R
\name{process_seurat_object}
\alias{process_seurat_object}
\title{Wrapper around Seurat processing functions}
\usage{
process_seurat_object(
  seurat_object,
  assay_name = NULL,
  calculate_rb_and_mt = TRUE,
  remove_stress_and_mt = TRUE,
  SCTransform = FALSE,
  NormalizeData = TRUE,
  FindVariableFeatures = TRUE,
  ScaleData = TRUE,
  RunPCA = TRUE,
  FindNeighbors = TRUE,
  FindClusters = TRUE,
  RunTSNE = TRUE,
  RunUMAP = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{A valid seurat-object.}

\item{SCTransform}{A named list of arguments given to \code{Seurat::SCTransform()}, TRUE or FALSE.}

\item{NormalizeData}{A named list of arguments given to \code{Seurat::NormalizeData()}, TRUE or FALSE.}

\item{FindVariableFeatures}{A named list of arguments given to \code{Seurat::FindVariableFeatures()}, TRUE or FALSE.}

\item{ScaleData}{A named list of arguments given to \code{Seurat::ScaleData()}, TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument \code{features} input
for argument \code{features} is set to \code{base::rownames(seurat_object)}.}

\item{RunPCA}{A named list of arguments given to \code{Seurat::RunPCA()}, TRUE or FALSE.}

\item{FindNeighbors}{A named list of arguments given to \code{Seurat::FindNeighbors()}, TRUE or FALSE.}

\item{FindClusters}{A named list of arguments given to \code{Seurat::FindClusters()}, TRUE or FALSE.}

\item{RunTSNE}{A named list of arguments given to \code{Seurat::RunTSNE()}, TRUE or FALSE.}

\item{RunUMAP}{A named list of arguments given to \code{Seurat::RunUMAP()}, TRUE or FALSE.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
A processed seurat-object.
}
\description{
Wrapper around Seurat processing functions
}
