% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_maybe.R
\name{processWithSCT}
\alias{processWithSCT}
\title{Apply SCTransform}
\usage{
processWithSCT(
  object,
  slot = "data",
  name = "sct_data",
  exchange_counts = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{slot}{The slot of the output assay in the \code{Seurat} object from where to
take the matrix. Defaults to \emph{data}.}

\item{name}{The name under which to store the matrix.}

\item{exchange_counts}{Logical. If \code{TRUE}, the counts matrix of the \code{spata2}
object is exchanged for the counts matrix in the output assay.}

\item{...}{Additional arguments given to \code{Seurat::SCTransform()}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Runs the pipeline suggested by \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}} and
extracts a matrix fromt he resulting assay object.
}
\keyword{internal}
