% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{processWithSeurat}
\alias{processWithSeurat}
\title{Process \code{spata2} object using \code{Seurat}}
\usage{
processWithSeurat(
  object,
  NormalizeData = TRUE,
  FindVariableFeatures = TRUE,
  ScaleData = TRUE,
  RunPCA = list(npcs = 30),
  FindNeighbors = list(dims = 1:30),
  FindClusters = TRUE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{NormalizeData}{A named list of arguments given to \code{Seurat::NormalizeData()}, TRUE or FALSE.}

\item{FindVariableFeatures}{A named list of arguments given to \code{Seurat::FindVariableFeatures()}, TRUE or FALSE.}

\item{ScaleData}{A named list of arguments given to \code{Seurat::ScaleData()}, TRUE or FALSE.

Hint: If set to TRUE or the argument-list provided does not specify the argument \code{features} input
for argument \code{features} is set to \code{base::rownames(seurat_object)}.}

\item{RunPCA}{A named list of arguments given to \code{Seurat::RunPCA()}, TRUE or FALSE.}

\item{FindNeighbors}{A named list of arguments given to \code{Seurat::FindNeighbors()}, TRUE or FALSE.}

\item{FindClusters}{A named list of arguments given to \code{Seurat::FindClusters()}, TRUE or FALSE.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
A wrapper around the most essential processing functions
of the \code{Seurat} package. A temporary \code{Seurat} object is created using the
data from the \code{spata2} object and is processed. Then the processed
data is transferred back to the \code{spata2} object.
}
\details{
By default, the function adds the matrix of @slot \code{data} from
the seurat assay to the \code{spata2} object under the name \emph{normalized} (if \code{NormalizeData} = TRUE)
and the matrix of @slot \code{scaled.data} from the Seurat assay to the \code{spata2}
object under the name \emph{scaled} (if \code{ScaleData = TRUE}).
}
