% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{refImage}
\alias{refImage}
\alias{refImage,spata2-method}
\alias{refImage,HistoImaging-method}
\title{Obtain name of reference content}
\usage{
refImage(object, ...)

\S4method{refImage}{spata2}(object)

\S4method{refImage}{HistoImaging}(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
Character value.
}
\description{
Handy functions to quickly access the name of reference content.
}
