% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{registerImage}
\alias{registerImage}
\alias{registerImage,SPATA2-method}
\alias{registerImage,SpatialData-method}
\alias{removeImage}
\alias{removeImage,SPATA2-method}
\alias{removeImage,SpatialData-method}
\title{Register or remove images}
\usage{
registerImage(object, ...)

\S4method{registerImage}{SPATA2}(
  object,
  img_name,
  img = NULL,
  img_dir = NULL,
  unload = TRUE,
  process = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

\S4method{registerImage}{SpatialData}(
  object,
  img_name,
  img = NULL,
  dir = NULL,
  unload = FALSE,
  process = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)

removeImage(object, ...)

\S4method{removeImage}{SPATA2}(object, img_name)

\S4method{removeImage}{SpatialData}(object, img_name)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{img_name}{Character value. The image to remove. Must neither be
the active nor the reference image.}

\item{img}{An image. Must be usable with \code{EBImage::as.Image()}.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{dir}{Character value. The directory from where to retrieve the image.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Use \code{registerImage()} to add a new image in form of a \code{HistoImage}
to the object.

Use \code{removeImage()} to savely discard images and their \code{HistoImage} container
that are no longer needed.

Do not confuse with \code{\link[=loadImage]{loadImage()}} and \code{\link[=unloadImage]{unloadImage()}}.
}
\section{Tutorials}{

Extensive tutorials for how to use this function can be found on our website
\url{https://themilolab.github.io/SPATA2/} .
}

