% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{relateToSpatialAnnotation}
\alias{relateToSpatialAnnotation}
\title{Deprecated}
\usage{
relateToSpatialAnnotation(object, input_df, ...)
}
\description{
Deprecated in favor of \code{\link[=getCoordsDfSA]{getCoordsDfSA()}}
}
\keyword{internal}
