% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{relevelGroups}
\alias{relevelGroups}
\title{Relevel groups of grouping variable}
\usage{
relevelGroups(object, grouping_variable, new_levels)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{grouping_variable}{Character value. The grouping variable of interest. Use
\code{getGroupingOptions()} to obtain all valid input options.}

\item{new_levels}{Character vector of group names in the order in which
the new ordering is supposed to be stored. Must contain all groups of the
grouping variable.}
}
\value{
An updated spata object.
}
\description{
Sets the ordering of the groups in a grouping variable. Affects the order
in which they appear in plots.
}
