% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeMetaFeatures}
\alias{removeMetaFeatures}
\title{Remove meta features}
\usage{
removeMetaFeatures(object, feature_names, ...)

removeMetaFeatures(object, feature_names, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{feature_names}{Character vector. Specifies the meta features to be removed.}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Remove meta \link[=concept_variables]{features} from the
\code{SPATA2} object.

Removes features from the meta feature data.frame.
}
\seealso{
\code{\link[=removeMolecules]{removeMolecules()}}, \code{\link[=getMetaFeatureNames]{getMetaFeatureNames()}}
}
