% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeGenes}
\alias{removeGenes}
\alias{removeGenesStress}
\alias{removeGenesZeroCounts}
\alias{removeMolecules}
\title{Remove molecules from the \code{SPATA2} object}
\usage{
removeGenes(object, genes, show_warnings = FALSE, verbose = NULL)

removeGenesStress(object, verbose = NULL)

removeGenesZeroCounts(object, verbose = NULL)

removeMolecules(
  object,
  molecules,
  show_warnings = FALSE,
  ref = "molecule",
  assay_name = activeAssay(object),
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{genes}{Character vector. Names of genes to remove.}

\item{show_warnings}{Logical value. If \code{TRUE}, \strong{warnings} about genes that were not found
although they were mentioned in the vector of genes that are to be discarded
are suppressed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Functions that removes molecules from the \code{SPATA2} object by removing
them from count matrix and all processed matrices of the respective \link[MolecularAssay]{assay}.

\itemize{
\item{\code{removeMolecules()}:}{ Removes user specified molecules}
\item{\code{removeMoleculesZeroCounts()}:}{ Removes molecules that do not have a single count
across all observations.}
}

Wrappers for transriptomic assay:

\itemize{
\item{\code{removeGenes()}:}{ Removes user specified genes.}
\item{\code{removeGenesZeroCounts()}:}{ Removes genes that do not have a single count
across all observations.}
\item{\code{removeGenesStress()}:}{ Removes mitochondrial and stress related genes.}
}
}
\details{
This step affects the matrices of the object and thus all subsequent
analysis steps. Analysis steps that have already been conducted are not affected!
This includes clustering, DEA, GSEA etc. It is advisable to integrate this
step as early as possible in the processing pipeline.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

genes <- getGenes(object)
head(genes)
length(genes)

object <- removeGenesZeroCounts(object)
object <- removeGenesStress(object)

genes_new <- getGenes(object)
length(genes_new)

}
