% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeObs}
\alias{removeObs}
\alias{removeObsZeroCounts}
\title{Remove observations}
\usage{
removeObs(object, barcodes, spatial_proc = TRUE, verbose = NULL)

removeObsZeroCounts(
  object,
  spatial_proc = TRUE,
  assay_name = activeAssay(object),
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{barcodes}{Character vector or barcodes that are \strong{removed}.}

\item{spatial_proc}{Logical value. Indicates whether the new sub-object is
processed spatially. If \code{TRUE}, a new tissue outline is identified based
on the remaining observations via \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}. Then,
spatial annotations are tested on being located on either of the remaining
tissue sections. If they are not, they are removed.

If \code{FALSE}, these processing steps are skipped. Generally speaking, this is
not recommended. Only set to \code{FALSE}, if you know what you're doing.

Only relevant, if \code{barcodes} is not \code{NULL}.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Remove unwanted \link[=concept_observations]{observations} from the object.

\itemize{
\item{\code{removeObs()}}{: Allows to specify the observations to remove manually.}
\item{\code{removeObsZeroCounts()}}{: Identifies and removes observations with no molecule counts from the \code{SPATA2} object.}
}
}
\details{
After removal of observations, unused levels of factor variables in the feature data.frame are dropped.
Analysis results are not affected.
}
\examples{
library(SPATA2)

data("example_data")
object <- example_data$object_UKF269T_diet

# the function tells you if / how many observations were removed
object <- removeObsZeroCounts(object, verbose = TRUE)

}
\seealso{
\code{\link[=subsetSpataObject]{subsetSpataObject()}}
}
