% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeProcessedMatrix}
\alias{removeProcessedMatrix}
\title{Remove a processed matrix}
\usage{
removeProcessedMatrix(object, mtr_name)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Removes a processed matrix from the \code{spata2} object.
}
