% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeProcessedMatrix}
\alias{removeProcessedMatrix}
\title{Remove a processed matrix}
\usage{
removeProcessedMatrix(object, mtr_name, assay_name = activeAssay(object))
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes a processed matrix from the \code{SPATA2} object.
}
\note{
If the removed matrix was the active matrix the new active matrix
is defined as the last element of \code{\link[=getMatrixNames]{getMatrixNames()}}.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

getMatrixNames(object)
getProcessedMatrixNames(object)

object <- normalizeCounts(object, method = "LogNormalize")

getProcessedMatrixNames(object)

object <- removeProcessedMatrix(object, mtr_name = "LogNormalize")

getProcessedMatrixNames(object)

}
