% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeSpatialAnnotations}
\alias{removeSpatialAnnotations}
\title{Remove spatial annotations}
\usage{
removeSpatialAnnotations(object, ids)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character value. The IDs of the spatial annotations to
remove.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes spatial annotations from the SPATA2 object.
}
