% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeSpatialOutliers}
\alias{removeSpatialOutliers}
\title{Remove spatial outliers}
\usage{
removeSpatialOutliers(
  object,
  spatial_proc = TRUE,
  rm_var = TRUE,
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{spatial_proc}{Logical value. Indicates whether the new sub-object is
processed spatially. If \code{TRUE}, a new tissue outline is identified based
on the remaining observations via \code{\link[=identifyTissueOutline]{identifyTissueOutline()}}. Then,
spatial annotations are tested on being located on either of the remaining
tissue sections. If they are not, they are removed.

If \code{FALSE}, these processing steps are skipped. Generally speaking, this is
not recommended. Only set to \code{FALSE}, if you know what you're doing.

Only relevant, if \code{barcodes} is not \code{NULL}.}

\item{rm_var}{Logical value. If \code{TRUE}, the variable \emph{sp_outlier} is removed
since it only contains \code{FALSE} after this function call and is of no value
any longer.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes data points that were identified as spatial outliers
and all their related data. If no spatial outliers exist, the input object
is returned as is.
}
\examples{

library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

# spatial outliers have not been labeled histologically (= NA)
plotSurface(object, color_by = "histology")

object <- identifyTissueOutline(object) # step 1

plotSurface(object, color_by = "tissue_section")

object <- identifySpatialOutliers(object) # step 2

plotSurface(object, color_by = "sp_outlier")

nObs(object) # before removal

object <- removeSpatialOutliers(object) # step 3

plotSurface(object, color_by = "histology")

nObs(object) # after removal


}
\seealso{
\code{\link[=identifyTissueOutline]{identifyTissueOutline()}}, \code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}, \code{\link[=containsSpatialOutliers]{containsSpatialOutliers()}},
\code{\link[=subsetSpataObject]{subsetSpataObject()}} is the working horse behind the removal.
}
