% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{removeSpatialOutliers}
\alias{removeSpatialOutliers}
\title{Remove spatial outliers}
\usage{
removeSpatialOutliers(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input object containing the added or computed
results.
}
\description{
Removes data points that were identified as spatial outliers
and all their related data. If no spatial outliers exist, the input object
is returned as is.
}
\seealso{
\code{\link[=identifySpatialOutliers]{identifySpatialOutliers()}}, \code{\link[=containsSpatialOutliers]{containsSpatialOutliers()}}
}
