% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameFeatures}
\alias{renameFeatures}
\title{Rename features}
\usage{
renameFeatures(object, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{...}{The features to be renamed specified according to the following
syntax: \emph{'new_feature_name'} \code{=} \emph{'old_feature_name'}.}
}
\value{
An upated spata-object.
}
\description{
Allows to rename features stored inside the @fdata slot.
}
\examples{
#Not run:

 object <- renameFeatures(object, "clusters_new" = "clusters")

}
