% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{renameGroups}
\alias{renameGroups}
\alias{renameSegments}
\title{Rename cluster/group names}
\usage{
renameGroups(
  object,
  grouping_variable,
  ...,
  keep_levels = NULL,
  of_sample = NA
)

renameSegments(object, ..., of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{grouping_variable}{Character value. The grouping variable of interest.}

\item{...}{The groups to be renamed specified according to the following
syntax: \emph{'new_group_name'} \code{=} \emph{'old_group_name'}.}

\item{of_sample}{This argument is currently inactive. It might be reactivated when
spata-objects can store more than one sample.}
}
\value{
An updated spata-object.
}
\description{
Allows to rename groups within a discrete grouping variable (such as
cluster variables) of the feature data in slot @fdata as well as in slot @dea
where differential gene expression analysis results are stored. Use \code{renameSegments()}
to rename already drawn segments.
}
\examples{
#Not run:

 object <-
    renameGroups(object = spata_object,
                 grouping_variable = "seurat_clusters",
                 "first_new_group" = "1",
                 "sec_new_group" = "2")


}
