% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-family.R
\name{renameGroups}
\alias{renameGroups}
\alias{renameSegments}
\title{Rename cluster/group names}
\usage{
renameGroups(object, discrete_feature, ..., of_sample = NA)

renameSegments(object, ..., of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{discrete_feature}{Character value. The grouping variable of interest.}

\item{...}{The groups to be renamed specified according to the following
syntax: \emph{'new_group_name'} \code{=} \emph{'old_group_name'}.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
An updated spata-object.
}
\description{
Allows to rename groups within a discrete grouping variable (such as
cluster variables) of the feature data in slot @fdata as well as in slot @dea
where differential gene expression analysis results are stored. Use \code{renameSegments()}
to rename already drawn segments.
}
\examples{
#Not run:

 object <-
    renameGroups(object = spata_object,
                 discrete_feature = "seurat_clusters",
                 "first_new_group" = "1",
                 "sec_new_group" = "2")


}
