% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{resetImageJustification}
\alias{resetImageJustification}
\title{Reset image justification}
\usage{
resetImageJustification(object)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Resets slot @justification of the \code{HistologyImaging} object.
}
