% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.R
\name{rotateAll}
\alias{rotateAll}
\alias{rotateImage}
\alias{rotateCoordinates}
\alias{rotateCoordsDf}
\alias{rotateSpatialAnnotations}
\alias{rotateSpatialTrajectories}
\title{Rotate image and coordinates}
\usage{
rotateAll(object, angle, clockwise = TRUE)

rotateImage(
  object,
  angle,
  img_name = activeImage(object),
  clockwise = TRUE,
  ...
)

rotateCoordinates(object, angle, clockwise = TRUE, verbose = NULL)

rotateCoordsDf(object, angle, clockwise = TRUE, verbose = NULL)

rotateSpatialAnnotations(
  object,
  angle,
  ids = getSpatAnnIds(object),
  clockwise = TRUE,
  verbose = NULL
)

rotateSpatialTrajectories(object, angle, clockwise = TRUE, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{angle}{Numeric value. The angle by which the coordinates
are rotated. Should range from 1-359.}

\item{clockwise}{Logical value. If \code{TRUE}, rotation is performed
in clockwise direction. If \code{FALSE}, the other way round.}

\item{img_name}{Character value. The name of the \code{HistoImage} of interest.
If \code{NULL}, the active histo image is chosen by default.}

\item{...}{Additional arguments given to \code{give_feedback()}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{ids}{Character vector or \code{NULL}. If character, specifies the IDs
of the image annotations of interest. If numeric, the image annotations are picked by number.
If \code{NULL}, all image annotations are included - subsequent selection with \code{tags} and
\code{test} is possible.}
}
\description{
The \verb{rotate*()} family rotates the current image
or coordinates of spatial aspects or everything. See details
for more information.

\strong{NOTE:} \code{rotateImage()} only rotates the image and lets everything else as
is. Only use it if you want to rotate the image because it is not aligned with
the spatial coordinates. If you want to rotate the image while maintaining
alignment with the spatial aspects in the \code{SPATA2} object
use \code{rotateAll()}!
}
\details{
The \verb{rotate*()} functions can be used to rotate the complete \code{SPATA2}
object content or to rotate single aspects.

\itemize{
\item{\code{rotateAll()}:}{ Rotates image as well as every single spatial aspect.
\strong{Always tracks the justification.}}
\item{\code{rotateImage()}:}{ Rotates the image.}
\item{\code{rotateCoordinates()}:}{ Rotates the coordinates data.frame, spatial annotations
and spatial trajectories.}
\item{\code{rotateCoordsDf()}:}{ Rotates the coordinates data.frame.}
\item{\code{rotateSpatialAnnotations()}:}{ Rotates spatial annotations.}
\item{\code{rotateSpatialTrajectories()}:}{ Rotates spatial trajectories.}
}
}
\seealso{
\code{\link[=flipAll]{flipAll()}}, \code{\link[=scaleAll]{scaleAll()}}
}
