% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runBayesSpaceClustering}
\alias{runBayesSpaceClustering}
\title{Clustering with BayesSpace}
\usage{
runBayesSpaceClustering(
  object,
  name = "bayes_space",
  n.Pcs = 15,
  n.HVGs = 2000,
  skip.PCA = FALSE,
  log.normalize = TRUE,
  assay.type = "logcounts",
  BSPARAM = BiocSingular::ExactParam(),
  qs = 3:7,
  burn.in = c(100, 1000),
  nrep = c(1000, 5000),
  use.dimred = "PCA",
  d = 15,
  init.method = "mclust",
  model = "t",
  gamma = 3,
  mu0 = NULL,
  lambda0 = NULL,
  alpha = 1,
  beta = 0.01,
  save.chain = FALSE,
  chain.fname = NULL,
  prefix = "",
  return_model = TRUE,
  empty_remove = FALSE,
  overwrite = FALSE,
  assign_sce = NULL,
  assign_envir = .GlobalEnv,
  seed = NULL,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{name}{Character value. The name the cluster variable has in
the feature data of the \code{SPATA2} object. Defaults to \emph{bayes_space}.}

\item{qs}{The values of q to evaluate.}

\item{burn.in, nrep}{Integers specifying the range of repetitions to
compute.}

\item{prefix}{Character value. Prefix of the cluster groups.}

\item{overwrite}{Logical value. If TRUE, \code{name} overwrites features
in feature data of the \code{SPATA2} object.}

\item{assign_sce}{Character value or \code{NULL}. If character, specifies the
name under which the bayes space output (object of class \code{SingleCellExperiment})
is assigned to the global environment. This makes the whole output
of the bayes space pipeline available instead of only adding the clustering
output as a grouping variable to the \code{SPATA2} object.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Additional arguments given to \code{BayesSpace::spatialCluster()}. Exception:
\code{sce}, \code{q} are specified within the function.}

\item{q_force}{Numeric value or \code{FALSE}. If numeric, it forces the number
of output clusters with input value. If \code{FALSE}, the optimal number
of clusters is chosen for \code{q} determined by the elbow point of \code{BayesSpace::qTune()}.}
}
\value{
An updated \code{SPATA2} object.
}
\description{
A wrapper around the BayesSpace clustering pipeline introduced
by \emph{Zhao et al. 2021}.
}
\details{
This function is a wrapper around \code{readVisium()},
\code{spatialPreprocess()}, \code{qTune()} and \code{spatialCluster()}
of the BayesSpace package. The results are stored in form of a grouping
variable in the feature data.frame of the returned \code{SPATA2} object.
}
\author{
Zhao E, Stone MR, Ren X, Guenthoer J, Smythe KS, Pulliam T,
Williams SR, Uytingco CR, Taylor SEB, Nghiem P, Bielas JH, Gottardo R.
Spatial transcriptomics at subspot resolution with BayesSpace.
Nat Biotechnol. 2021 Nov;39(11):1375-1384. doi: 10.1038/s41587-021-00935-2.
Epub 2021 Jun 3. PMID: 34083791; PMCID: PMC8763026.
}
