% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-set-enrichment.R
\name{runGSEA}
\alias{runGSEA}
\alias{runGsea}
\title{Compute gene set enrichment}
\usage{
runGSEA(
  object,
  across,
  methods_de,
  max_adj_pval = NULL,
  min_lfc = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  gene_set_list = NULL,
  gene_set_names = NULL,
  test = c("hypergeometric", "kstest"),
  power = 1,
  background = nGenes(object),
  absolute = FALSE,
  pval = 1,
  fdr = 1,
  reduce = TRUE,
  quiet = TRUE,
  chr_to_fct = TRUE,
  verbose = NULL
)

runGsea(
  object,
  across,
  methods_de,
  max_adj_pval = NULL,
  min_lfc = NULL,
  n_highest_lfc = NULL,
  n_lowest_pval = NULL,
  gene_set_list = NULL,
  gene_set_names = NULL,
  test = c("hypergeometric", "kstest"),
  power = 1,
  background = nGenes(object),
  absolute = FALSE,
  pval = 1,
  fdr = 1,
  reduce = TRUE,
  quiet = TRUE,
  chr_to_fct = TRUE,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{across}{Character vector. All grouping variables of interest.}

\item{methods_de}{Character vector. All differential expression methods
of interest.}

\item{max_adj_pval}{Numeric value. Sets the threshold for adjusted p-values. All genes
with adjusted p-values above that threshold are ignored.}

\item{min_lfc}{Numeric value. Sets the threshold for average log fold change. All genes
with an average log fold change below that threshold are ignored.}

\item{n_highest_lfc}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{n_lowest_pval}{Numeric value. Affects the total number of genes that are kept. See details.}

\item{gene_set_list}{A named list of character vectors. Names of slots correspond to the
gene set names. The slot contains the genes of the gene sets.Holds priority over
\code{gene_set_names}.}

\item{gene_set_names}{Character vector of gene set names that are taken
from the object's gene set data.frame.}

\item{test}{Choose an enrichment type e.g. c("hypergeometric", "kstest")}

\item{power}{Exponent for weights (kstest only)}

\item{background}{Size or character vector of background population genes}

\item{absolute}{Takes max-min score rather than the max deviation from null (kstest only)}

\item{pval}{Filter results to be less than pval cutoff}

\item{fdr}{Filter results to be less than fdr cutoff}

\item{reduce}{Logical value. If set to TRUE (the default) the return value
of \code{hypeR::hypeR()} is reduced to what is necessary for \code{SPATA2}s
function to work. If FALSE, the complete objects are stored. This will
grow the spata-objects size quickly!}

\item{quiet}{Use true to suppress logs and warnings}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata-object.
}
\description{
Computes gene set enrichment based on the results of
\code{runDeAnalysis()}. See details for more.
}
\details{
Computes gene set enrichment analysis using \code{hypeR::hypeR()}.
It does so by iterating about all possible combinations of \code{across} and
\code{methods_de}. Combinations for which no DE-results are found are silently
skipped.

If gene sets are provided via \code{gene_set_list} argument \code{gene_set_names}
is ignored. Else the latter determines the gene sets used which are then taken
from the spata-objects gene set data.frame.
}
