% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hspa-main-functions.R
\name{runHspa}
\alias{runHspa}
\title{Title}
\usage{
runHspa(
  object,
  genes_subset = NULL,
  include_genes = NULL,
  mtr_name = "scaled",
  method_binarization = "kmeans",
  method_kmeans = "Hartigan-Wong",
  kmeans = list(),
  method_csr = "MonteCarlo",
  n_quadrats = 10,
  method_padj = "fdr",
  save_intermediate_results = FALSE,
  verbose = NULL,
  of_sample = NA
)
}
\arguments{
\item{of_sample}{}
}
\value{

}
\description{
Title
}
