% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runPCA}
\alias{runPCA}
\alias{runPca}
\title{Run Principal Component Analysis}
\usage{
runPCA(
  object,
  n_pcs = 30,
  variables = NULL,
  mtr_name = activeMatrix(object),
  assay_name = activeAssay(object),
  ...
)

runPca(...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be computed.}

\item{variables}{Character vector or \code{NULL}. If character, subsets the data matrix
by variable/molecule names such that only the specified ones are used for dimensional reduction.
If \code{NULL}, the unsubstted matrix denoted in \code{mtr_name} is used.}

\item{mtr_name}{Character value. The name of the matrix of interest. Defaults
to the active matrix of the assay, as denoted by \code{\link[=activateMatrix]{activateMatrix()}}.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{...}{Additional arguments given to \code{irlba::prcomp_irlba()}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Takes the data matrix of choice and passes it to
\code{irlba::prcomp_irlba()}.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

object <- runPCA(object)
object <- runTSNE(object)
object <- runUMAP(object)

plotPCA(object, color_by = "histology")
plotTSNE(object, color_by = "histology")
plotUMAP(object, color_by = "histology")

}
\seealso{
\code{\link[=identifyVariableMolecules]{identifyVariableMolecules()}}, \code{\link[=plotPcaElbow]{plotPcaElbow()}}
}
