% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{runPca}
\alias{runPca}
\alias{runPca2}
\title{Run Principal Component Analysis}
\usage{
runPca(object, n_pcs = 30, mtr_name = NULL, of_sample = NA, ...)

runPca2(object, n_pcs = 30, mtr_name = NULL, of_sample = NA, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be computed.}

\item{mtr_name}{Character value. The name of the expression matrix of interest. If set to NULL
the currently active matrix is chosen.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to \code{irlba::prcomp_irlba()}.}
}
\value{
\itemize{
  \item{\code{runPca()}:}{ An updated spata-object containing the reduction variables in the pca data.frame.}
  \item{\code{runPca2()}:}{ The direct output-object of \code{irlba::prcomp_irlba()}}.
  }
}
\description{
Takes the expression matrix of choice and passes it to
\code{irlba::prcomp_irlba()}.
}
