% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runSeuratClustering}
\alias{runSeuratClustering}
\title{Clustering with Seurat}
\usage{
runSeuratClustering(
  object,
  name = "seurat_clusters",
  mtr_name = activeMatrix(object),
  assay_name = activeAssay(object),
  NormalizeData = list(),
  ScaleData = list(),
  FindVariableFeatures = list(selection.method = "vst", nfeatures = 2000),
  RunPCA = list(npcs = 60),
  FindNeighbors = list(dims = 1:30),
  FindClusters = list(resolution = 0.8),
  prefix = "S",
  overwrite = FALSE
)
}
\arguments{
\item{FindVariableFeatures, RunPCA, FindNeighbors, FindClusters}{Each argument
takes a list of arguments that is given to the equivalent function.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
A wrapper around the Seurat clustering pipeline suggested by
\emph{Hao and Hao et al., 2021}.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

object <- runSeuratClustering(object, name = "new_seurat_clst")

plotSurface(object, color_by = "new_seurat_clst")

}
