% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatially-variable-genes-wrapper.R
\name{runSparkx}
\alias{runSparkx}
\alias{getSparkxResults}
\alias{getSparkxGeneDf}
\alias{getSparkxGenes}
\title{Identify genes of interest with SPARKX}
\usage{
runSparkx(object, numCores = 1, option = "mixture", verbose = NULL)

getSparkxResults(object, test = TRUE)

getSparkxGeneDf(object, threshold_pval = 1, arrange_pval = TRUE)

getSparkxGenes(object, threshold_pval)
}
\value{
An updated spata object.
}
\description{
A wrapper around the algorithm introduced by \emph{Zhu et al. 2021}
to identify genes with spatial expression pattern with SPARK-X.
}
\author{
Zhu, J., Sun, S. & Zhou, X. SPARK-X: non-parametric modeling enables
scalable and robust detection of spatial expression patterns for large spatial
transcriptomic studies. Genome Biol 22, 184 (2021). https://doi.org/10.1186/s13059-021-02404-0
}
