% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R, R/runA-Z.R
\name{getSparkxGeneDf}
\alias{getSparkxGeneDf}
\alias{getSparkxGenes}
\alias{getSparkxResults}
\alias{runSparkx}
\title{Identify genes of interest with SPARKX}
\usage{
getSparkxGeneDf(object, threshold_pval = 1, arrange_pval = TRUE)

getSparkxGenes(object, threshold_pval)

getSparkxResults(object, test = TRUE)

runSparkx(object, numCores = 1, option = "mixture", verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata object.
}
\description{
A wrapper around the algorithm introduced by \emph{Zhu et al. 2021}
to identify genes with spatial expression pattern with SPARK-X.
}
\author{
Zhu, J., Sun, S. & Zhou, X. SPARK-X: non-parametric modeling enables
scalable and robust detection of spatial expression patterns for large spatial
transcriptomic studies. Genome Biol 22, 184 (2021). https://doi.org/10.1186/s13059-021-02404-0
}
