% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R, R/runA-Z.R
\name{getSparkxGeneDf}
\alias{getSparkxGeneDf}
\alias{getSparkxGenes}
\alias{getSparkxResults}
\alias{runSPARKX}
\title{Identify spatially significant features with SPARKX}
\usage{
getSparkxGeneDf(object, threshold_pval = 1, arrange_pval = TRUE)

getSparkxGenes(object, threshold_pval)

getSparkxResults(object, assay_name = activeAssay(object), error = TRUE, ...)

runSPARKX(
  object,
  assay_name = activeAssay(object),
  numCores = 1,
  option = "mixture",
  verbose = NULL
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{assay_name}{Only relevant if the \code{SPATA2} object contains more than
one assay: Denotes the assay of interest and thus the
\link[=concept_molecular_modalities]{molecular modality} to use. Defaults to the active assay,
as set by \code{\link[=activateAssay]{activateAssay()}}.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
A wrapper around the algorithm introduced by \emph{Zhu et al. 2021}
to identify features with non-random spatial expression pattern with SPARK-X.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

object <- runSPARKX(object)

sparkx_genes <- getSparkxGenes(object, threshold = 0.05)

sparkx_df <- getSparkxGenesDf(object)


}
\author{
Zhu, J., Sun, S. & Zhou, X. SPARK-X: non-parametric modeling enables
scalable and robust detection of spatial expression patterns for large spatial
transcriptomic studies. Genome Biol 22, 184 (2021). https://doi.org/10.1186/s13059-021-02404-0
}
