% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional-reduction.R
\name{runTsne}
\alias{runTsne}
\alias{runTsne2}
\title{Run t-Stochastic Neighbour Embedding}
\usage{
runTsne(object, n_pcs = 20, tsne_perplexity = 30, of_sample = NA, ...)

runTsne2(object, n_pcs = 20, tsne_perplexity = 30, of_sample = NA, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{n_pcs}{Numeric value. Denotes the number of principal components used. Must be
smaller or equal to the number of principal components the pca data.frame contains.}

\item{tsne_perplexity}{Numeric value. Given to argument \code{perplexity} of
\code{Rtsne::Rtsne()}.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}

\item{...}{Additional arguments given to \code{Rtsne::Rtsne()}.}
}
\value{
\itemize{
  \item{\code{runTsne()}:}{ An updated spata-object containing the reduction variables in the tsne data.frame.}
  \item{\code{runTsne2()}:}{ The direct output-object of \code{Rtsne::Rtsne()}}
  }
}
\description{
Takes the pca-data of the object up to the principal component denoted
in argument \code{n_pcs} and performs tSNE with it.
}
