% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runA-Z.R
\name{runUMAP}
\alias{runUMAP}
\alias{runUmap}
\title{Run Uniform Manifold Approximation and Projection}
\usage{
runUMAP(object, n_pcs = NULL, ...)

runUmap(...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{n_pcs}{Numeric value. Denotes the number of principal components used. Must be
smaller or equal to the number of principal components the pca data.frame contains.}

\item{...}{Additional arguments given to \code{umap::umap()}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Takes the pca data of the object up to the principal component denoted
in argument \code{n_pcs} and performs UMAP with it.
}
\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF269T_diet

object <- runPCA(object)
object <- runTSNE(object)
object <- runUMAP(object)

plotPCA(object, color_by = "histology")
plotTSNE(object, color_by = "histology")
plotUMAP(object, color_by = "histology")

}
\seealso{
\code{\link[=runPCA]{runPCA()}}
}
