% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-and-saving.R
\name{saveSpataObject}
\alias{saveSpataObject}
\alias{saveCorrespondingCDS}
\alias{saveCorrespondingSeuratObject}
\title{Save corresponding objects}
\usage{
saveSpataObject(
  object,
  directory_spata = NULL,
  combine_with_wd = FALSE,
  verbose = NULL
)

saveCorrespondingCDS(
  cds,
  object,
  directory_cds = NULL,
  combine_with_wd = FALSE,
  verbose = NULL
)

saveCorrespondingSeuratObject(
  seurat_object,
  object,
  directory_seurat = NULL,
  combine_with_wd = FALSE,
  verbose = NULL
)
}
\arguments{
\item{object}{A valid spata-object.}

\item{directory_spata, directory_cds_directory_seurat_object}{Character value or NULL. Set details for more.}

\item{combine_with_wd}{Character value or FALSE. If specified with a
character value (default: \emph{'/'}) the input of \code{new_directory}
is considered to be a relative directory and is combined with the
current working directory (\code{base::getwd()}) separated with the character string
specified. If set to FALSE the input of \code{new_directory}
is taken as is.}

\item{cds}{A valid cell-data-set. (from the monocle3 platform)}

\item{seurat_object}{A valid seurat-object. (from the Seurat platform)}
}
\value{
Apart from their side effect (saving the object of interest) all three functions
return the provided, updated spata-object.
}
\description{
Family of functions to save corresponding objects of different analysis
platforms. See details and value for more information.
}
\details{
If \code{directory_<platform>} is set to NULL (the default) all functions first check if the spata-object contains any
deposited default directories. If so the specified object to be saved is saved under
that direction. If \code{directory_<platform>} is specified as a character it's input is taken as the
directory under which to store the object and the deposited directory is overwritten
such that the next time you load the spata-object it contains the updated directory.
In order for that to work the \code{saveCorresponding*()}-functions - apart from saving the object of interest -  return the
updated spata-object while \code{saveSpataObject()} simply returns an invisible TRUE
as the  new directory (if provided) is stored inside the object before it is saved.
}
