% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setActiveMatrix}
\alias{setActiveMatrix}
\alias{setActiveExpressionMatrix}
\title{Denote the default expression matrix}
\usage{
setActiveMatrix(object, mtr_name, verbose = NULL)

setActiveExpressionMatrix(object, mtr_name, verbose = NULL, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{name}{Character value. The name of the matrix that is to be set as
the active expression matrix.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Denote the default expression matrix
}
