% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setActiveExpressionMatrix}
\alias{setActiveExpressionMatrix}
\title{Denote the default expression matrix}
\usage{
setActiveExpressionMatrix(object, mtr_name, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{name}{Character value. The name of the matrix that is to be set as
the active expression matrix.}
}
\value{
An updated spta-object.
}
\description{
Denote the default expression matrix
}
