% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder-preliminary-functions.R
\name{setAutoencoderAssessment}
\alias{setAutoencoderAssessment}
\title{Set results of autoencoder assessment}
\usage{
setAutoencoderAssessment(object, assessment_list, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{assessment_list}{Named list with slots \code{$df} and \code{$set_up}.}
}
\value{
A spata-object.
}
\description{
Set results of autoencoder assessment
}
