% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setCaptureArea}
\alias{setCaptureArea}
\title{Set capture area}
\usage{
setCaptureArea(object, x = NULL, y = NULL)
}
\arguments{
\item{object}{An object of class \code{SPATA2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{x, y}{Vectors of length two that correspond to the range of the
respective axis. If \code{NULL}, the respective range stays as is.}
}
\description{
Sets the capture area for objects from platforms with
a specific capture area / field of view.
}
\note{
The spatial methods \emph{VisiumSmall} and \emph{VisiumLarge} have a capture
area by default. You can override it but it is not recommended.
}
\seealso{
\code{\link[=getCaptureArea]{getCaptureArea()}}
}
