% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setCnvResults}
\alias{setCnvResults}
\title{Set cnv-results}
\usage{
setCnvResults(object, cnv_list, ...)
}
\arguments{
\item{object}{A valid spata-object.}

\item{cnv_list}{The list containing the results from \code{runCnvAnalysis()}.}
}
\value{
A spata object containing the set input.
}
\description{
Set cnv-results
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
