% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setCoordsDf}
\alias{setCoordsDf}
\title{Set the coordinates}
\usage{
setCoordsDf(object, coords_df, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{coords_df}{A data.frame containing information about every barcode-spot. Must contain the variables:
\itemize{
 \item{\emph{barcodes} Character. The barcode-sequences (+ the sample belonging) of every barcode spot.}
 \item{\emph{sample} Character. The sample belonging of every barcode-spot.}
 \item{\emph{x} Numeric. The x-coordinates of every barcode.}
 \item{\emph{y} Numeric. The y-coordinates of every barcode.}
}}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A spata object containing the set input.
}
\description{
Set the coordinates
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
