% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setCountMatrix}
\alias{setCountMatrix}
\alias{setDenoisedMatrix}
\alias{setScaledMatrix}
\title{Set data matrices}
\usage{
setCountMatrix(object, count_mtr, of_sample = NA)

setDenoisedMatrix(object, denoised_mtr, of_sample = NA)

setScaledMatrix(object, scaled_mtr, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{count_mtr, scaled_mtr, denoised_mtr}{Matrices whose column names refer
to the barcodes and whose rownames refer to the gene names.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A spata object containing the set input.
}
\description{
\code{SPATA} in general distinguishes between two types of data matrices.
There are \emph{count-matrices} containing the raw counts and
\emph{expression-matrices} containing scaled, denoised or in any other
way processed and normalized count data.

The majority of \code{SPATA}-functions leans on data carried in expression matrices.
They default to the one that is set as the \emph{active expression matrix} - use
\code{getActiveMatrixName()} to see which one it currently is. After
initiating a spata-object \emph{'scaled'} should be the default. After running
 \code{runAutoencoderDenoising()} \emph{'denoised'} becomes the default
expression matrix.

To set one of those three matrices use these functions.
To add additional matrices use \code{addExpressionMatrix()}.
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
