% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setDefault}
\alias{setDefault}
\title{Set object specific default}
\usage{
setDefault(object, ...)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{...}{Named arguments whoose default input you want to override.}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Sets object specific default for recurring arguments
such as \code{pt_alpha}, \code{pt_clrp}, \code{verbose}.
}
