% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setDefaultGrouping}
\alias{setDefaultGrouping}
\alias{getDefaultGrouping}
\title{Default grouping}
\usage{
setDefaultGrouping(object, grouping_variable, verbose = NULL)

getDefaultGrouping(object, verbose = FALSE, ...)
}
\arguments{
\item{grouping_variable}{Character value. The grouping variable that is
used by default within all functions that need one.}
}
\value{
\code{setDefaultGrouping()}: Updated spata object. \code{getDefaultGrouping()}: Character value. Name
of the default grouping variable.
}
\description{
Sets and extracts the default grouping. Useful to save typing
in functions that require a grouping variable as input.
}
