% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setDefaultInstructions}
\alias{setDefaultInstructions}
\alias{setDirectoryInstructions}
\title{Set default instructions}
\usage{
setDefaultInstructions(object)

setDirectoryInstructions(object)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A spata-object again containing the default spata-instructions.
Everything that previously has been adjusted with \code{adjustDefaultInstructions()}
is overwritten.
}
\description{
Set default instructions
}
