% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setDefaultTrajectory}
\alias{setDefaultTrajectory}
\alias{getDefaultTrajectory}
\title{Default trajecotry}
\usage{
setDefaultTrajectory(object, trajectory_name, verbose = NULL)

getDefaultTrajectory(object, verbose = FALSE, ...)
}
\arguments{
\item{trajectory_name}{Character value. The trajectory that is
used by default within all functions that need one.}
}
\value{
\code{setDefaultTrajectory()}: Updated spata object. \code{getDefaultTrajectory()}: Character value. Name
of the default trajectory.
}
\description{
Sets and extracts the default trajectory. Useful to save typing
in functions that require a trajectory name as input.
}
