% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setDefaultTrajectory}
\alias{setDefaultTrajectory}
\alias{getDefaultTrajectory}
\alias{getDefaultTrajectoryId}
\alias{setDefaultTrajectoryId}
\title{Default trajectory ID}
\usage{
setDefaultTrajectory(object, id, verbose = NULL)

getDefaultTrajectory(object, ...)

getDefaultTrajectoryId(object, ...)

setDefaultTrajectoryId(object, id, verbose = NULL)
}
\arguments{
\item{id}{Character value.}
}
\value{
\code{setDefaultTrajectory()}: Updated spata object. \code{getDefaultTrajectory()}: Character value. Id
of the default trajectory.
}
\description{
Sets and extracts the default trajectory id. Useful to save typing
in functions that require a trajectory name as input.
}
