% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setFeatureDf}
\alias{setFeatureDf}
\title{Set feature data.frame}
\usage{
setFeatureDf(object, feature_df, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{feature_df}{A data.frame that contains the feature and the key-variables.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
set_dummy return details
}
\description{
Set feature data.frame
}
