% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setHistoImage}
\alias{setHistoImage}
\alias{setHistoImage,SPATA2-method}
\alias{setHistoImage,SpatialData-method}
\title{Set HistoImage object}
\usage{
setHistoImage(object, ...)

\S4method{setHistoImage}{SPATA2}(object, hist_img, ...)

\S4method{setHistoImage}{SpatialData}(object, hist_img, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{hist_img}{An object of class \code{HistoImage}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets object of class \code{HistoImage}.
}
\seealso{
\code{\link[=registerImage]{registerImage()}}
}
\keyword{internal}
