% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setImageAnnotation}
\alias{setImageAnnotation}
\alias{setImageAnnotations}
\title{Set image annotations}
\usage{
setImageAnnotation(object, img_ann, align = TRUE, overwrite = FALSE)

setImageAnnotations(object, img_anns, align = TRUE, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{overwrite}{Logical Value. If TRUE and the ID of the
input image annotation is already used by an image annotation
it is overwritten.}
}
\description{
Sets image annotations in the correct slot.
}
