% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setImageDirDefault}
\alias{setImageDirDefault}
\alias{setImageDirHighres}
\alias{setImageDirLowres}
\title{Set image directories}
\usage{
setImageDirDefault(object, dir, check = TRUE, verbose = NULL, ...)

setImageDirHighres(object, dir, check = TRUE, verbose = NULL, ...)

setImageDirLowres(object, dir, check = TRUE, verbose = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{dir}{Character value. The chosen directory. See details for possible
requirements.}

\item{check}{Logical value. If set to TRUE the input directory is checked
for validity and it is checked if the file actually exists.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
Sets image directories that facilitate image exchanges.
}
\seealso{
\code{\link[=addImageDir]{addImageDir()}}
}
