% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setImageDirDefault}
\alias{setImageDirDefault}
\alias{setImageDirHighres}
\alias{setImageDirLowres}
\title{Set image directories}
\usage{
setImageDirDefault(object, dir, check = TRUE, verbose = NULL, ...)

setImageDirHighres(object, dir, check = TRUE, verbose = NULL, ...)

setImageDirLowres(object, dir, check = TRUE, verbose = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{dir}{Character value. Directory to specific image. Should end
with either \emph{.png}, \emph{.jpeg} or \emph{.tiff}. (Capital endings work, too.)}

\item{check}{Logical value. If set to TRUE the input directory is checked
for validity and it is checked if the file actually exists.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\description{
Sets image directories that facilitate image exchanges.
}
\seealso{
\code{\link[=addImageDir]{addImageDir()}}
}
