% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setImageDirHighres}
\alias{setImageDirHighres}
\alias{setImageDirLowres}
\title{Set image directories}
\usage{
setImageDirHighres(object, dir_highres, check = TRUE)

setImageDirLowres(object, dir_lowres, check = TRUE)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{check}{Logical value. If set to TRUE the input directory is checked
for validity and it is checked if the file actually exists.}

\item{dir_lowres, dir_highres}{Character value. The file directories to
the corresponding images.}
}
\value{
An updated spata object.
}
\description{
Sets image directories that facilitate image exchanges.
}
