% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN-Z.R, R/setA-Z.R
\name{getOutlineVarName}
\alias{getOutlineVarName}
\alias{setOutlineVarName}
\title{Set outline variable name}
\usage{
getOutlineVarName(object)

setOutlineVarName(object, name)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{name}{Name of the variable that contains the outline. Must be a factor
variable in the feature data.frame.}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Sets the name of the variable in the feature data.frame
that contains grouping of the barcode-spots according to the number
of coherent tissue sections on the capture frame. (E.g. if two brain slices
of a mouse were imaged and permeabilized on the same visium slide).
}
