% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-family.R
\name{setPcaDf}
\alias{setPcaDf}
\alias{setTsneDf}
\alias{setUmapDf}
\title{Set dimensional reductions data.frames}
\usage{
setPcaDf(object, pca_df, of_sample = "")

setTsneDf(object, tsne_df, of_sample = "")

setUmapDf(object, umap_df, of_sample = "")
}
\arguments{
\item{object}{A valid spata-object.}

\item{pca_df, tsne_df, umap_df}{The data.frame composed of the variables
\emph{barcodes}, \emph{sample} and the variables containing the barcode-
spot embedding.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
A spata object containing the set input.
}
\description{
Safely adds data.frames containing the barcode-spot embedding
of different dimensional reduction techniques. Cell embedding variables
must be named as follows:

 \itemize{
  \item{ \code{pca_df}: \emph{PC1, PC2, PC3, ...}}
  \item{ \code{tsne_df}: \emph{tsne1, tsne2, ...}}
  \item{ \code{umap_df}: \emph{umap1, umap2, ...}}
 }
}
\details{
All \code{set*()}-functions offer a save way to set certain
slots of your spata-object. They do check the input for validity but
effectively overwrite everything that is occupying the slot to be set -
use with caution.
}
