% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setPixelScaleFactor}
\alias{setPixelScaleFactor}
\title{Set pixel scale factor}
\usage{
setPixelScaleFactor(object, pxl_scale_fct = NULL, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{pxl_scale_fct}{Numeric value with an attribute named
\emph{unit} with the unit dist_si/px.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets pixel scale factor.
}
