% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{setScaleFactor}
\alias{setScaleFactor}
\alias{setScaleFactor,HistoImage-method}
\title{Set scale factors}
\usage{
setScaleFactor(object, ...)

\S4method{setScaleFactor}{HistoImage}(object, fct_name, value)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{fct_name}{Character value. Name of the scale factor.}

\item{value}{Value to set.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Sets scale factor values.
}
