% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{setSpataDir}
\alias{setSpataDir}
\title{Set SPATA2 directory}
\usage{
setSpataDir(object, dir)
}
\arguments{
\item{object}{An object of class \code{SPATA2}.}

\item{dir}{Character value. The chosen directory. See details for possible
requirements.}
}
\value{
The input \code{SPATA2} object containing the added or computed
results.
}
\description{
Sets a directory under which the \code{SPATA2} object is
always stored using the function \code{saveSpataObject()}.
}
