% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialData}
\alias{setSpatialData}
\title{Set spatial data}
\usage{
setSpatialData(object, sp_data)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{sp_data}{An object of class \code{\link{SpatialData}}.}
}
\description{
Sets the image container class \code{SpatialData}
in the corresponding slot of the \code{SPATA2} object.
}
