% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialMethod}
\alias{setSpatialMethod}
\alias{setSpatialMethod,SPATA2-method}
\title{Set spatial method}
\usage{
setSpatialMethod(object, ...)

\S4method{setSpatialMethod}{SPATA2}(object, method, ...)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{method}{An object of class \code{SpatialMethod}.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets the \code{\link{SpatialMethod}} object containing information
of the platform used.
}
