% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setSpatialMethodSlot}
\alias{setSpatialMethodSlot}
\title{Set slot content of SpatialMethod object}
\usage{
setSpatialMethodSlot(object, slot, content)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{slot}{Name of the slot.}

\item{content}{Content to be set.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Sets content of slot in the \code{SpatialMethod} object. Use with
caution.
}
