% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setTissueOutline}
\alias{setTissueOutline}
\title{Set tissue outline}
\usage{
setTissueOutline(object, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
\code{spata2} object with additional variables in coordinates data.frame.

\itemize{
\item{\emph{section} :}{ character. The identified tissue section. 0 means probable artefact spot.}
\item{\emph{outline} :}{logical. \code{TRUE} if identified as a spot that lies on the edge of the tissue.}
}
}
\description{
Sets tissue outline by calling \code{identifyTissueSections()}
and \code{identifyTissueOutline()}.
}
\examples{

 library(ggplot2)
 library(ggExtra)

 object <- downloadPubExample("MCI_LMU")

 print(getTissueOutlineDf(object))

 to_df <- getTissueOutlineDf(object, remove = FALSE)

 to_df[["outline"]] <- as.character(to_df[["outline"]])

 ggplot(to_df, mapping = aes(x = x, y = y)) +
   geom_point_fixed(mapping = aes(color = section, alpha = outline), size = getDefault(object, "pt_size")) +
   scale_alpha_manual(values = c("TRUE" = 1, "FALSE" = 0.4))



}
\keyword{internal}
