% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setTrajectory}
\alias{setTrajectory}
\alias{setTrajectories}
\title{Set trajectories}
\usage{
setTrajectory(object, trajectory, overwrite = FALSE)

setTrajectories(object, trajectories, overwrite = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{trajectory}{An object of class \code{Trajectory.}}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{trajectories}{List of objects of class \code{Trajectory}.}
}
\value{
The input object containing the added or computed
results.
}
\description{
Sets trajectories in the correct slot.
}
