% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{setTrajectory}
\alias{setTrajectory}
\alias{setTrajectories}
\title{Deprecated}
\usage{
setTrajectory(...)

setTrajectories(...)
}
\description{
Deprecated in favor of \code{\link[=setSpatialTrajectory]{setSpatialTrajectory()}} and \code{\link[=setSpatialTrajectories]{setSpatialTrajectories()}}.
}
\keyword{internal}
