% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setA-Z.R
\name{setTrajectory}
\alias{setTrajectory}
\alias{setTrajectories}
\title{Set trajectories}
\usage{
setTrajectory(object, trajectory, overwrite = FALSE)

setTrajectories(object, trajectories, overwrite = FALSE)
}
\description{
Set trajectories
}
