% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{shinyModuleZoomingServer}
\alias{shinyModuleZoomingServer}
\alias{shinyModuleZoomingUI}
\title{Shiny module \emph{Zooming}}
\usage{
shinyModuleZoomingServer(
  id = "m1",
  default = list(),
  brushed_area = NULL,
  object = NULL,
  persp = "ccs"
)

shinyModuleZoomingUI(id = "m1")
}
\arguments{
\item{id}{Character value. ID of the module.}

\item{brushed_area}{Reactive expression which, if called with \verb{<reactiveExpression>()}
returns a list of four slots named \emph{xmin}, \emph{xmax}, \emph{ymin} and \emph{ymax} each being
a numeric value.}

\item{object}{An object for which a method for \code{getImageRange()} is defined.}
}
\value{
A reactive expression which, if called with \verb{<reactiveExpression>()}
returns a list of two numeric vectors of length two named \emph{x} and \emph{y}.
}
\description{
The server logical for shiny module \emph{Zooming}.
}
