% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{showColors}
\alias{showColors}
\alias{showColorPalettes}
\alias{showColorSpectra}
\alias{showModels}
\title{Show color palettes and spectra}
\usage{
showColors(input, n = 20, title_size = 10)

showColorPalettes(input = validColorPalettes(flatten = TRUE))

showColorSpectra(input = validColorSpectra(flatten = TRUE))

showModels(
  input = 100,
  linecolor = "black",
  linesize = 0.5,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  pretty_names = FALSE,
  x_axis_arrow = TRUE,
  ...
)
}
\arguments{
\item{input}{Character vector of input options for \code{clrsp} and
\code{clrp}.}

\item{n}{Numnber of colors.}

\item{title_size}{Size of plot titles.}
}
\value{
A plot of ggplots arranged with \code{gridExtra::arrange.grid()}.
}
\description{
Simple visualization of available color palettes and
spectra from left to right.
}
\examples{

 showColors(input = c("inferno", "Reds", "npg", "uc"), n = 10)

 showColors(input = validColorPalettes()[[1]])

}
