% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showColors.R
\name{showColors}
\alias{showColors}
\alias{showModels}
\title{Show color palettes and spectra}
\usage{
showColors(input, n = 20, title_size = 10)

showModels(
  input = 100,
  pattern_subset = NULL,
  pattern_remove = NULL,
  pattern_add = NULL,
  ...
)
}
\arguments{
\item{input}{Character vector of input options for \code{clrsp} and
\code{clrp}.}

\item{n}{Numnber of colors.}

\item{title_size}{Size of plot titles.}
}
\value{
A plot of ggplots arranged with \code{gridExtra::arrange.grid()}.
}
\description{
Simple visualization of available color palettes and
spectra from left to right.
}
\examples{

 showColors(input = c("inferno", "Reds", "npg", "uc"), n = 10)

 showColors(input = validColorPalettes()[[1]])

}
