% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{showModels}
\alias{showModels}
\title{Show spatial gradient screening models}
\usage{
showModels(
  input = 100,
  linecolor = "black",
  linesize = 0.5,
  model_subset = NULL,
  model_remove = NULL,
  model_add = NULL,
  pretty_names = FALSE,
  x_axis_arrow = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{linesize}{Numeric value. The size of the line(s) plotted.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
A ggplot.
}
\description{
Display the models used for spatial gradient screening.
}
