% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{smoothSpatialAnnotation}
\alias{smoothSpatialAnnotation}
\title{Smooth the Borders of a Spatial Annotation}
\usage{
smoothSpatialAnnotation(
  object,
  id,
  method,
  outer = TRUE,
  inner = TRUE,
  new_id = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}

\item{method}{The smoothing method to be applied. Options include "chaikin", "densify", "ksmooth", and "spline".}

\item{outer}{Logical; if TRUE, the outer border of the annotation is smoothed.}

\item{inner}{Logical or numeric; if TRUE, all inner borders are smoothed.
If a numeric value is provided, only the specified inner border is smoothed.}

\item{new_id}{Character value or \code{FALSE}. If character, the resulting
spatial annotation is stored under a new ID.}

\item{overwrite}{Logical value. Must be \code{TRUE} to allow overwriting.}

\item{...}{Additional arguments passed to the smoothing function.}
}
\value{
The input object containing the added or computed
results.
}
\description{
This function applies a smoothing algorithm to the borders of a
\code{\link{SpatialAnnotation}} object. It can smooth both outer and inner borders using various methods.
}
\seealso{
\code{\link[=expandSpatialAnnotation]{expandSpatialAnnotation()}}, \code{\link[=shiftSpatialAnnotation]{shiftSpatialAnnotation()}}, \code{\link{SpatialAnnotation}}
}
