% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SPATA2-class}
\alias{SPATA2-class}
\alias{SPATA2}
\title{The \code{SPATA2} class}
\description{
A class to represent spatial transcriptomic and other -omic
studies.
}
\section{Slots}{

\describe{
\item{\code{compatibility}}{list. A list for miscellaneous short-term solutions, mainly
to maintain compatibility.}

\item{\code{coordinates}}{list. A list of data.frames that contain coordinates for
observations that are not equal to the observational unit of the method (e.g.
single cell deconvolution for Visium, molecules for MERFISH).}

\item{\code{fdata}}{data.frame. A data.frame in which data variables of the obesrvations
of the underlying spatial method are stored that are not molecule counts as stored
in \code{Assay} objects within slot \verb{@omics}. (e.g. clustering, histological grouping,
number of counts, etc.)}

\item{\code{signatures}}{list. A named list of character vectors. Each character vector
corresponds to a signature represented by molecules that are known to work together
in a way described by the name of the signature (e.g. HM_HYPOXIA: a signature
of multiple genes known to be upregulated in case of hypoxic environments)}

\item{\code{information}}{list. A list for miscellaneous information around the object.}

\item{\code{logfile}}{data.frame. A data.frame of prompts used on the object that return
the object (e.g. \code{\link[=runBayesSpaceClustering]{runBayesSpaceClustering()}}) including the argument input. Used to
keep track of the analysis progress as well as aid for debugging.}

\item{\code{meta}}{list. A list of meta data regarding the object, the tissue, and the
experimental desgin, etc.}

\item{\code{method}}{SpatialMethod. An object of class \code{\link{SpatialMethod}} that contains
information around the spatial method to which the object corresponds.}

\item{\code{omics}}{list. A list of \code{\link{AssayData}} objects.}

\item{\code{sample}}{character. The name of the object.}

\item{\code{spatial}}{HistoImaging. An object of class \code{\link{HistoImaging}} storing multiple
spatial information.}

\item{\code{version}}{list. The version of the object.}
}}

