% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_segmentations}
\alias{spatial_segmentations}
\title{A list of lists of spatial segmentations. Every slot in list \code{spatial_segmentations}
is named by the sample the segmentations belong to. Every sample specific slot
is a data.frame with two variables: \emph{barcodes} and the spatial segmentation
variable (usually named \emph{histolgoy}). The segmentation variable can be
added to the corresponding sample using the function \code{addFeatures()}.}
\format{
An object of class \code{list} of length 1.
}
\usage{
spatial_segmentations
}
\description{
@docType data
@usage data(spatial_segmentations)
}
\keyword{datasets}
