% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{spatial_trajectory-class}
\alias{spatial_trajectory-class}
\alias{spatial_trajectory}
\title{spatial_trajectory object}
\value{
S4 object
}
\description{
spatial_trajectory object
}
\section{Slots}{

\describe{
\item{\code{compiled_trajectory_df}}{A data.frame containing the variables:

\describe{
 \item{\emph{barcodes}}{Character. The barcode-spots' sequences.}
 \item{\emph{sample}}{Character. The barcode-spots' sample belonging.}
 \item{\emph{x,y}}{Numeric. The barcode-spots' spatial coordinates.}
 \item{\emph{projection_length}}{Numeric. The distance between the barcode-spots'
  projection onto the trajectory-vector and the start of the trajectory.}
 \item{\emph{trajectory_part}}{Character. The part of the trajectory.}
 }}

\item{\code{segment_trajectory_df}}{A data.frame containing the numeric variables
\emph{x, y, xend, yend} that denote the start and the end of every trajectory
part.}

\item{\code{comment}}{Character value. The comment written down before saving the trajectory.}

\item{\code{name}}{Character value. The trajectory's name.}

\item{\code{sample}}{Character value. The sample the trajectory belongs to.}
}}

