% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{spatialAnnotationScreening}
\alias{spatialAnnotationScreening}
\title{The Spatial Annotation Screening algorithm}
\usage{
spatialAnnotationScreening(
  object,
  ids,
  variables,
  core,
  distance = "dte",
  binwidth = recBinwidth(object),
  angle_span = c(0, 360),
  unit = getDefaultUnit(object),
  bcs_exclude = character(0),
  sign_var = "fdr",
  sign_threshold = 0.05,
  force_comp = FALSE,
  skip_comp = FALSE,
  model_add = NULL,
  model_subset = NULL,
  model_remove = NULL,
  estimate_R2 = TRUE,
  control = SPATA2::sgs_loess_control,
  n_random = 10000,
  rm_zero_infl = TRUE,
  seed = 123,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance value}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. (See details for more.)
Defaults to a distance that covers the whole tissue section using \code{\link[=distToEdge]{distToEdge()}}.}

\item{binwidth}{Distance value. The width of the distance bins to which
each data point is assigned. Defaults to our platform dependent
recommendation using \code{\link[=recBinwidth]{recBinwidth()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of the spatial annotation.
(See details fore more.)}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \link{`validUnitsOfArea()'} for valid
input options.}

\item{bcs_exclude}{Character value containing name(s) of data points to be excluded from the analysis.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{SpatialAnnotationScreening}. See documentation
with \code{?SpatialAnnotationScreening} for more information.
}
\description{
Screens the sample for numeric variables that stand
in meaningful, spatial relation to annotated structures/areas, \link[=concept_spatial_annotation]{spatial annotations}.
For a detailed explanation on how to define the parameters \code{distance},
\code{n_bins_dist}, \code{binwidth}, \code{angle_span} and \code{n_bins_angle}
see details section.
}
\section{Tutorials}{

Extensive tutorials for how to use this function can be found on our website
\url{https://themilolab.github.io/SPATA2/} .
}

\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

ids <- getSpatAnnIds(object, tags = c("necrotic", "compr"), test = "identical")

# opt 1 prefiltering by SPARKX is recommended, but not required
object <- runSPARKX(object)
genes <- getSparkxGenes(object)

# opt 2
genes <- getGenes(object)


sas_out <-
 spatialAnnotationScreening(
   object = object,
   ids = ids,
   variables = genes,
   core = FALSE
   )

}
\seealso{
\code{\link[=createGroupAnnotations]{createGroupAnnotations()}}, \code{\link[=createImageAnnotations]{createImageAnnotations()}},
\code{\link[=createNumericAnnotations]{createNumericAnnotations()}} for how to create spatial annotations.

\code{\link[=getCoordsDfSA]{getCoordsDfSA()}} for how to obtain spatial relation of data points to
a spatial annotation.

\code{\link[=getSasDf]{getSasDf()}} for how to obtain inferred expression gradients as used in
spatial annotation screening.

\code{\link[=plotSasLineplot]{plotSasLineplot()}} for visualization of inferred expression gradients.
}
