% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{spatialAnnotationScreening}
\alias{spatialAnnotationScreening}
\title{The Spatial Annotation Screening algorithm}
\usage{
spatialAnnotationScreening(
  object,
  ids,
  variables,
  core,
  distance = "dte",
  resolution = recSgsRes(object),
  angle_span = c(0, 360),
  unit = getDefaultUnit(object),
  bcs_exclude = character(0),
  sign_var = "fdr",
  sign_threshold = 0.05,
  force_comp = FALSE,
  skip_comp = FALSE,
  model_add = NULL,
  model_subset = NULL,
  model_remove = NULL,
  estimate_R2 = TRUE,
  control = NULL,
  n_random = 10000,
  rm_zero_infl = TRUE,
  seed = 123,
  add_image = TRUE,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{ids}{Character vector. Specifies the IDs of the spatial annotations of interest.}

\item{variables}{Character vector. The numeric variables to be included in
the screening process. Makre sure that the correct matrix is active in the
respective assays.}

\item{distance}{\code{\link[=concept_distance_measure]{Distance measure}}. Specifies
the distance from the border of the spatial annotation to the \emph{horizon} in
the periphery up to which the screening is conducted. Defaults to a distance
that covers the whole tissue section the spatial annotation is located
on using \code{\link[=distToEdge]{distToEdge()}}. (This distance must not be exceeded.)}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{angle_span}{Numeric vector of length 2. Confines the area screened by
an angle span relative to the center of its closest spatial annotation.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}

\item{bcs_exclude}{Character value containing the barcodes of observations to be excluded
from the analysis.}

\item{sign_var}{Either \emph{p_value} or \emph{fdr}. Defaults to \emph{fdr}.}

\item{sign_threshold}{The significance threshold. Defaults to 0.05.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{control}{A list of arguments as taken from \code{\link[stats:loess.control]{stats::loess.control()}}.
Default setting is stored in \code{SPATA2::sgs_loess_control}.}

\item{n_random}{Number of random permutations for the significance testing of step 2.}

\item{seed}{Numeric value. Sets the random seed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{\link{SpatialAnnotationScreening}}.
}
\description{
Screens the sample for numeric variables that stand
in meaningful, spatial relation to annotated structures/areas, \link[=concept_spatial_annotation]{spatial annotations}.
For a detailed explanation on how to define the parameters \code{distance},
\code{resolution}, \code{angle_span} and \code{n_bins_angle} see details section.
}
\section{Tutorials}{

Extensive tutorials for how to use this function can be found on our website
\url{https://themilolab.github.io/SPATA2/} .
}

\examples{
library(SPATA2)

data("example_data")

object <- example_data$object_UKF313T_diet

object <- identifyTissueOutline(object)

ids <- getSpatAnnIds(object, tags = c("necrotic", "compr"), test = "all")

# opt 1 prefiltering by SPARKX is recommended, but not required
object <- runSPARKX(object)
genes <- getSparkxGenes(object, threshold_pval = 0.05)

# opt 2
genes <- getGenes(object)

sas_out <-
 spatialAnnotationScreening(
   object = object,
   ids = ids,
   variables = genes,
   core = FALSE
   )

}
\seealso{
\code{\link[=createGroupAnnotations]{createGroupAnnotations()}}, \code{\link[=createImageAnnotations]{createImageAnnotations()}},
\code{\link[=createNumericAnnotations]{createNumericAnnotations()}} for how to create spatial annotations.

\code{\link[=getCoordsDfSA]{getCoordsDfSA()}} for how to obtain spatial relation of data points to
a spatial annotation.

\code{\link[=getSasDf]{getSasDf()}} for how to obtain inferred expression gradients as used in
spatial annotation screening.

\code{\link[=plotSasLineplot]{plotSasLineplot()}} for visualization of inferred expression gradients.
}
