% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialMethod-class}
\alias{SpatialMethod-class}
\alias{SpatialMethod}
\title{The \code{SpatialMethod} - class}
\description{
Abstracts the concept of spatial biology experiments
such as \emph{Visium10X} or \emph{SlideSeq}.
}
\section{Slots}{

\describe{
\item{\code{image_frame}}{list. List of length two or three. Provides
measures of the sample length in units of length.}

\item{\code{observational_unit}}{character. Single word that describes
the observational unit of the experiment. E.g \emph{'barcode-spot'} in
case of @name == \emph{'Visium10X'}.}

\item{\code{meta}}{list. List of miscellaneous meta data for the method.}

\item{\code{name}}{character. The name of the spatial method. Should
be one of \code{validSpatialMethods()}.}
}}

