% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialMethod-class}
\alias{SpatialMethod-class}
\alias{SpatialMethod}
\title{The \code{SpatialMethod} - Class}
\description{
Defines the core features of spatial biology platforms
like \emph{Visium} and \emph{SlideSeq}.
}
\section{Slots}{

\describe{
\item{\code{info}}{list. List of miscellaneous meta data about the method.}

\item{\code{method_specific}}{list. List method specific data. Depending on the
method certain slot names are reserved. See section \emph{Method specifics:}
for more information.}

\item{\code{name}}{character. The name of the spatial method. (E.g. \emph{'VisiumSmall'}, \emph{'VisiumHD'}, \emph{'MERFISH'}, ...).
Should be one of \code{validSpatialMethods()}.}

\item{\code{observational_unit}}{character. Name with which to refer to
the data points the method focuses on. (E.g. \emph{'barcode_spot'}, or \emph{'cell'})}

\item{\code{unit}}{character. The SI to be used by default.}
}}

\section{Method specifics}{

Slot @method_specific contains a versatile list of information around
specific methods. Depending on the type several slots are reserved/should
be set:

For methods of type Visium (currently known \emph{VisiumSmall} and \emph{VisiumLarge}):

\itemize{
\item{\emph{ccd}:}{ Center to center distance as a spatial distance measure in SI units.}
\item{\emph{diameter}:}{ Diameter of each spot in micrometer.}
\item{\emph{fiducial_frame}:}{ List of length two, named \emph{x} and \emph{y}. Provides standardized
measures of the fiducial frame(in SI units).}
}

For methods that of type \emph{SlideSeq} (currently known \emph{SlideSeqV1}):

\itemize{
\item{\emph{diameter}:}{ Diameter of each bead in micrometer.}
}
}

\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

