% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialMethod-class}
\alias{SpatialMethod-class}
\alias{SpatialMethod}
\title{The \code{SpatialMethod} - Class}
\description{
Abstracts the concept of spatial biology experiments
such as \emph{Visium} or \emph{SlideSeq}.
}
\section{Slots}{

\describe{
\item{\code{capture_area}}{list. List of length two. Provides the measures
of the area in which identified entities are to be expected (in SI units).}

\item{\code{fiducial_frame}}{list. List of length two, named \emph{x} and \emph{y}.
Provides standardized measures of the sample image (in SI units).}

\item{\code{info}}{list. List of miscellaneous meta data for the method.}

\item{\code{method_specific}}{list. List method specific data.}

\item{\code{name}}{character. The name of the spatial method. (E.g. \emph{'Visium'})}

\item{\code{observational_unit}}{character. Name with which to refer to
the entity the method focuses on. (E.g. \emph{'barcode_spot'})}

\item{\code{unit}}{character. The SI to be used by default.}
}}

