% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialMethod-class}
\alias{SpatialMethod-class}
\alias{SpatialMethod}
\title{The \code{SpatialMethod} - class}
\description{
Abstracts the concept of spatial biology experiments
such as \emph{Visium1} or \emph{SlideSeq}.
}
\section{Slots}{

\describe{
\item{\code{fiducial_frame}}{list. List of length two or three. Provides
standardized measures of the sample image in units of length.}

\item{\code{info}}{list. List of miscellaneous meta data for the method.}

\item{\code{name}}{character. The name of the spatial method. Should
be one of \code{validSpatialMethods()}.}

\item{\code{observational_unit}}{character. Single word that describes
the observational unit of the experiment. E.g \emph{'barcode-spot'} in
case of @name == \emph{'Visium'}.}
}}

