% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{SpatialTrajectoryScreening-class}
\alias{SpatialTrajectoryScreening-class}
\alias{SpatialTrajectoryScreening}
\title{The \code{SpatialTrajectoryScreening} - Class}
\description{
S4 class that contains input for and output of the
function \code{\link[=spatialTrajectoryScreening]{spatialTrajectoryScreening()}}.
}
\details{
This class is an extension of class \code{\link{SpatialGradientScreening}} and inherits
all of its slots.
}
\section{Slots}{

\describe{
\item{\code{trajectory}}{The \link{SpatialTrajectory} object used for the screening.}
}}

