% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{spatialTrajectoryScreening}
\alias{spatialTrajectoryScreening}
\title{The Spatial Trajectory Screening algorithm}
\usage{
spatialTrajectoryScreening(
  object,
  id,
  variables,
  resolution = recSgsRes(object),
  width = getTrajectoryLength(object, id),
  unit = getDefaultUnit(object),
  bcs_exclude = character(0),
  sign_var = "fdr",
  sign_threshold = 0.05,
  force_comp = FALSE,
  model_add = NULL,
  model_subset = NULL,
  model_remove = NULL,
  estimate_R2 = TRUE,
  rm_zero_infl = TRUE,
  n_random = 10000,
  seed = 123,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{variables}{Character vector. All numeric variables to be included in
the screening process.}

\item{resolution}{\link[=concept_distance_measure]{Distance measure}. The resolution
with which the expression gradient is inferred. Defaults are platform specific.
See more in detail section of \code{\link[=recSgsRes]{recSgsRes()}}.}

\item{width}{Distance measure. The width of the trajectory frame. Defaults
to the trajectory length.}

\item{unit}{Character value. Specifies the desired unit in
which \link[=concept_distance_measure]{distance measures}
or  \link[=concept_area_measure]{area measures} are provided.
Run \code{\link[=validUnitsOfLength]{validUnitsOfLength()}} or \code{\link[=validUnitsofArea]{validUnitsofArea()}} for valid
input options.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}
}
\value{
An object of class \code{SpatialTrajectoryScreening}. See documentation
with \code{?ImageAnnotationScreening} for more information.
}
\description{
Screens the sample for numeric variables that follow specific expression
changes along the course of the spatial trajectory.
}
\examples{

library(SPATA2)

object <- example_data$object_UKF269T_diet

object <- identifyTissueOutline(object)

object <- runSPARKX(object)
genes <- getSparkxGenes(object, threshold_pval = 0.05)

id <- "horizontal_mid"

plotImage(object) +
  ggpLayerSpatialTrajectories(object, ids = id)

plotSpatialTrajectories(object, ids = id)

sts_out <-
  spatialTrajectoryScreening(
    object = object,
    id = id,
    variables = genes
    )

}
\seealso{
\code{\link[=createSpatialTrajectories]{createSpatialTrajectories()}}
}
