% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{spatialTrajectoryScreening}
\alias{spatialTrajectoryScreening}
\title{The Spatial Trajectory Screening algorithm}
\usage{
spatialTrajectoryScreening(
  object,
  id,
  variables,
  binwidth = recBinwidth(object),
  width = getTrajectoryLength(object, id),
  unit = getDefaultUnit(object),
  bcs_exclude = character(0),
  sign_var = "fdr",
  sign_threshold = 0.05,
  force_comp = FALSE,
  model_add = NULL,
  model_subset = NULL,
  model_remove = NULL,
  estimate_R2 = TRUE,
  rm_zero_infl = TRUE,
  n_random = 10000,
  seed = 123,
  verbose = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value. Denotes the spatial trajectory
of interest.}

\item{variables}{Character vector. All numeric variables (meaning genes,
gene-sets and numeric features) that are supposed to be included in
the screening process.}

\item{binwidth}{Distance value. The width of the bins to which
the barcode-spots are assigned. Defaults to the center-center
distance: \code{binwidth = getCCD(object)}.}

\item{model_add}{Named list. Every slot in the list must be either a formula
containing a function that takes a numeric vector as input and returns a numeric
vector with the same length as its input vector. Or a numeric vector with the
same length as the input vector. Test models with \code{showModels()}.}

\item{model_subset}{Character value. Used as a regex to subset models.
Use \code{validModelNames()} to obtain all model names that are known to \code{SPATA2}
and \code{showModels()} to visualize them.}

\item{model_remove}{Character value. Used as a regex to remove models
are not supposed to be included.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{n_bins}{Numeric value or vector of length 2. Specifies exactly how many bins are
created. (See details for more.)}

\item{summarize_with}{Character value. Either \emph{'mean'} or \emph{'median'}.
Specifies the function with which the bins are summarized.}
}
\value{
An object of class \code{SpatialTrajectoryScreening}. See documentation
with \code{?ImageAnnotationScreening} for more information.
}
\description{
Screens the sample for numeric variables that follow specific expression
changes along the course of the spatial trajectory.
}
\details{
\bold{How the algorithm works:} All barcode-spots that fall into the scope
of the trajectory are projected on the trajectory's course. These projection
values indicate if a barcode-spot is rather located at the beginning or at
the end of the trajectory. Barcode-spots are binned by their projection values.

How many bins area created depends on the input for argument \code{binwidth}
or \code{n_bins} as well as on the length of trajectory. As the length of
the trajectory is fixed only one argument of the latter two must be provided.
The other one is calculated based on the equation shown below.

\code{n_bins} = \emph{length_of_trajectory} / \code{binwidth}

\code{binwidth} = \emph{length_of_trajectory} / \code{n_bins}

and for every numeric variable included the mean-expression of each bin is calculated.
As the bins can be aligned in an ascending order (ascending in relation to the
directory of the trajectory), so can the bin-wise mean expression of each variable.
Doing so results in \emph{inferred expression changes along the trajectory}.
Use \code{plotTrajectoryLineplot()} to visualize this concept.

The inferred expression changes are fitted against predefined models to find
variables whose expression e.g. increases, decreases or peaks over the course
of the trajectory. Use \code{showModels()} to visualize the inbuilt models.

How good a model fits is evaluated by pearson correlation and the area under
the curve of the gene-model-residuals.
}
\seealso{
\code{\link[=createSpatialTrajectories]{createSpatialTrajectories()}}
}
