% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_j-z.R
\name{subsetByBarcodes}
\alias{subsetByBarcodes}
\title{Deprecated}
\usage{
subsetByBarcodes(...)
}
\description{
Deprecated in favor of \code{\link[=subsetSpataObject]{subsetSpataObject()}}.
}
\keyword{internal}
