% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{subsetByBarcodes}
\alias{subsetByBarcodes}
\title{Subsetting by barcodes}
\usage{
subsetByBarcodes(object, barcodes, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{barcodes}{Character vector. The barcodes of the data points that are
supposed to be \bold{kept}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Removes unwanted data points from the object without any significant
post processing.
}
\details{
Unused levels of factor variables in the feature data.frame are dropped.
}
\examples{
library(SPATA2)
library(dplyr)
library(patchwork)

data("example_data")

object <- example_data$object_UKF313T_diet

barcodes_keep <-
 getMetaDf(object) \%>\%
 filter(bayes_sapce \%in\% c("3", "2", "1")) \%>\%
 pull(barcodes)

object_sub <- subsetByBarcodes(object, barcodes = barcodes_keep)

plotSurface(object, color_by = "bayes_space") +
 plotSurface(object_sub, color_by = "bayes_space")

}
