% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{subsetByBarcodes}
\alias{subsetByBarcodes}
\title{Subsetting by barcodes}
\usage{
subsetByBarcodes(object, barcodes, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{barcodes}{Character vector. The barcodes of the barcode spots that are
supposed to be \bold{kept}.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated \code{spata2} object.
}
\description{
Removes unwanted barcode spots from the object without any significant
post processing.
}
\details{
Unused levels of factor variables in the feature data.frame are dropped
and directory settings are reset to NULL.
}
