% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{subsetByGenes}
\alias{subsetByGenes}
\title{Subset by genes}
\usage{
subsetByGenes(object, genes, verbose = NULL)
}
\arguments{
\item{object}{An object of class \code{spata2}.}

\item{genes}{Character vector of gene names that are kept.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Removes genes from the data set. This affects count- and expression matrices
and can drastically decrease object size.
}
\note{
Gene dependent analysis results such as DEA or SPARKX
are \strong{not} subsetted. Stored results are kept as they are. To update them run
the algorithms again.
}
