% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-add-ons.R
\name{surfaceFrame}
\alias{surfaceFrame}
\title{Add frame of the sample}
\usage{
surfaceFrame(object, of_sample = NA)
}
\arguments{
\item{object}{A valid spata-object.}

\item{of_sample}{The sample(s) of interest specified as a single character value or vector.
If set to \emph{""} (the default) the first sample is chosen.}
}
\value{
ggproto object refering to the x- and y-scales
}
\description{
Add frame of the sample
}
