% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.R
\name{theme_lineplot_gradient}
\alias{theme_lineplot_gradient}
\alias{theme_ridgeplot_gradient}
\alias{theme_trajectory_fit}
\alias{theme_transparent}
\title{ggplot2 themes}
\usage{
theme_lineplot_gradient(breaks_x, range_d)

theme_ridgeplot_gradient(overlap = 0.5)

theme_trajectory_fit()

theme_transparent()
}
\value{
gg theme
}
\description{
Miscellaneous \code{ggplot2} themes used throughout the package.
}
