% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{tissueOutlineIdentified}
\alias{tissueOutlineIdentified}
\alias{tissueOutlineIdentified,spata2-method}
\alias{containsSpatialOutliers}
\alias{containsSpatialOutliers,spata2-method}
\alias{containsSpatialOutliers,HistoImaging-method}
\title{Check availability of tissue information}
\usage{
tissueOutlineIdentified(object, ...)

\S4method{tissueOutlineIdentified}{spata2}(object, error = FALSE)

containsSpatialOutliers(object, ...)

\S4method{containsSpatialOutliers}{spata2}(object, error = FALSE)

\S4method{containsSpatialOutliers}{HistoImaging}(object, error = FALSE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{error}{Logical. If \code{TRUE} and the input is invalid the
function throws an error.}
}
\value{
Logical value.
}
\description{
Checks if \code{identifySpatialOutliers()} and \code{identifyTissueOutline()}
has been run successfully.
}
