% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajectories}
\alias{trajectories}
\title{A list of lists of spatial trajectories. Every slot in list \code{trajectories}
is named by the sample the trajectory belongs to. Every sample specific slot
is an S4 object of class \code{SpatialTrajectory}. Use \code{setSpatialTrajectory()}
to add them to the corresponding object.}
\format{
An object of class \code{list} of length 1.
}
\usage{
trajectories
}
\description{
@docType data
 @usage data(trajectories)
}
\keyword{datasets}
