% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{transform_outline}
\alias{transform_outline}
\title{Transform image}
\usage{
transform_outline(outline_df, transformations, center, ranges)
}
\arguments{
\item{transformations}{List of transformation instructions. See
slot @transformations of class \code{HistologyImage}.}

\item{image}{Image comptabible with the \code{EBImage}-package.}
}
\value{
Transformed input.
}
\description{
Transforms the image or the tissue outline.
}
