% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{unloadImage}
\alias{unloadImage}
\alias{unloadImage,HistoImage-method}
\alias{unloadImage,HistoImaging-method}
\alias{unloadImages}
\alias{unloadImages,spata2-method}
\alias{unloadImages,HistoImaging-method}
\title{Empty image slot}
\usage{
unloadImage(object, ...)

\S4method{unloadImage}{HistoImage}(object, verbose = TRUE, ...)

\S4method{unloadImage}{HistoImaging}(object, img_name, verbose = TRUE, ...)

unloadImages(object, ...)

\S4method{unloadImages}{spata2}(object, active = FALSE, verbose = TRUE)

\S4method{unloadImages}{HistoImaging}(object, active = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{...}{Used to absorb deprecated arguments or functions.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{img_name}{Character value. The name of the image to unload.}

\item{active.}{Logical value. If \code{FALSE}, the default,
the image from the active \code{HistoImage} is not unloaded.}
}
\value{
The input \code{spata2} object containing the added or computed
results.
}
\description{
Removes the image from slot @image of a \code{HistoImage}.
Useful for efficient data storing.
}
\seealso{
\code{\link[=loadImage]{loadImage()}},\code{\link[=loadImages]{loadImages()}}
}
