% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{unwindAggregation}
\alias{unwindAggregation}
\title{Map aggregated to pre-aggregated barcodes}
\usage{
unwindAggregation(object, var_names = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{var_names}{Optional. A character vector specifying the names of metadata variables to include in the output.
If \code{NULL}, only the original and aggregated barcodes are returned.}
}
\value{
A \code{data.frame} containing the original barcodes (\code{barcodes_orig}),
the corresponding aggregated barcodes (\code{barcodes_aggr}), and any additional
metadata variables specified in \code{var_names}.
}
\description{
Map aggregated to pre-aggregated barcodes
}
\details{
This function reconstructs the original barcodes before the aggregation
process was applied. It retrieves the pre-aggregation state of the data and,
if specified, adds selected metadata variables.
}
\seealso{
\code{\link{reduceResolutionVisiumHD}} for aggregating barcodes by reducing resolution
in VisiumHD data sets.
}
