% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{updateSpataObject}
\alias{updateSpataObject}
\title{Update SPATA2 object}
\usage{
updateSpataObject(object, method = NULL, verbose = TRUE, ....)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{method}{Character value or \code{NULL}. If \code{NULL}, the functions tests whether
the barcodes of the input object can be mapped to either of the \code{VisiumSmall} or \code{VisiumLarge}
platform. If this does not succeed you must specify the argument. In that case it
should be one of \code{base::names(spatial_methods)}.

Only relevant for updating from SPATA2v2 to SPATA2v3. v3.0.0 and above should not
face any problems regarding this.}

\item{verbose}{Logical. If \code{TRUE}, informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{....}{Used to absorb deprecated arguments or functions.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Updates the input object to the newest version of the package.
}
\note{
\code{SPATA2} objects of version < 2.0.0 can not be updated any longer. If you have such an object
and want to transfer the data, please raise an issue at github.
}
