% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u.R
\name{updateSpataObject}
\alias{updateSpataObject}
\title{Update \code{SPATA2} object}
\usage{
updateSpataObject(
  object,
  method = NULL,
  sample_name = NULL,
  n_pcs = 30,
  verbose = TRUE,
  ....
)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{method}{Character value or \code{NULL}. If \code{NULL}, the functions tests whether
the barcodes of the input object can be mapped to either of the VisiumSmall or VisiumLarge
platform. If this does not succeed you must specify the argument. In that case it
should be one of \code{base::names(spatial_methods)}.}

\item{sample_name}{Character value. The name of the input object.}

\item{n_pcs}{Numeric value or \code{NULL}. If numeric, specifies the number of
principal components to use. If \code{NULL}, all principal components are used
(as computed with \code{\link[=runPCA]{runPCA()}}).}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}

\item{....}{Used to absorb deprecated arguments or functions.}
}
\value{
The updated input object, containing the added, removed or computed results.
}
\description{
Updates the input object to the newest version of the package.
}
\note{
\code{n_pcs} and \code{sample_name} only are if the input object derives from
the old \code{SPATA} (not \code{SPATA2}) package!
}
