% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-spata-object.R
\name{updateSpataObject}
\alias{updateSpataObject}
\title{Update spata-object from SPATA to SPATA2}
\usage{
updateSpataObject(
  object,
  sample_name = NULL,
  chr_to_fct = TRUE,
  n_pcs = 60,
  verbose = TRUE
)
}
\arguments{
\item{object}{A spata-object that has been created within the package SPATA.}

\item{sample_name}{Character value. Denotes the sample name. Must be one of
\code{getSampleNames()}.}

\item{chr_to_fct}{Logical. SPATA2 recommends to store grouping variables as factors
in the slot @fdata. If set to TRUE, character variables (apart from \emph{barcodes, sample, segmentation})
of the old obejct's feature data are converted to factors.}

\item{n_pcs}{Numeric value. Denotes the number of principal components to be computed.}

\item{verbose}{Logical. If set to TRUE informative messages regarding
the computational progress will be printed.

(Warning messages will always be printed.)}
}
\value{
An updated spata-object.
}
\description{
A convenient function that takes the spata-object you
have initiated with the package SPATA and adjusts it's architecture
to the new version. All features remain.
}
\details{
Apart from transferring the data and the progress from the old object
to the new one principal component analysis (PCA) is run via the function \code{runPca()} and
gene meta data is compuated via \code{computeGeneMetaData()}.
}
