% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-spata-objects.R
\name{updateSpataObject}
\alias{updateSpataObject}
\title{Update a spata object}
\usage{
updateSpataObject(object, verbose = TRUE)
}
\arguments{
\item{object}{A valid spata-object.}
}
\value{
A valid spata object.
}
\description{
Checks if the provided spata object needs any updates regarding
it's structure by comparing it's @version to the newest one.
}
