% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{validActivationFunctions}
\alias{validActivationFunctions}
\alias{validAgglomerationMethods}
\alias{validAlluvialTypes}
\alias{validColorPalettes}
\alias{validColorSpectra}
\alias{validDeaMethods}
\alias{validDefaultInstructionSlots}
\alias{validDistanceMethods}
\alias{validHierarchicalClusterMethods}
\alias{validModelNames}
\alias{validPadjMethods}
\alias{validSpatialMethods}
\alias{validUnits}
\alias{validUnitsOfArea}
\alias{validUnitsOfAreaSI}
\alias{validUnitsOfLength}
\alias{validUnitsOfLengthSI}
\title{Obtain valid argument inputs}
\usage{
validActivationFunctions()

validAgglomerationMethods()

validAlluvialTypes()

validColorPalettes(flatten = FALSE)

validColorSpectra(flatten = FALSE)

validDeaMethods()

validDefaultInstructionSlots()

validDistanceMethods()

validHierarchicalClusterMethods()

validModelNames()

validPadjMethods()

validSpatialMethods()

validUnits()

validUnitsOfArea()

validUnitsOfAreaSI()

validUnitsOfLength()

validUnitsOfLengthSI()
}
\value{
Character vectors or named lists of such.
}
\description{
These function simply return valid input options
for recurring arguments.
}
