% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.R
\name{validActivationFunctions}
\alias{validActivationFunctions}
\alias{validAgglomerationMethods}
\alias{validAlluvialTypes}
\alias{validColorPalettes}
\alias{validColorSpectra}
\alias{validDeAnalysisMethods}
\alias{validDefaultInstructionSlots}
\alias{validDimRedMethods}
\alias{validDirectoryInstructionSlots}
\alias{validDistanceMethods}
\alias{validHierarchicalClusterMethods}
\alias{validImageClasses}
\alias{validModelNames}
\alias{validPatternRecognitionMethods}
\alias{validPadjMethods}
\alias{validPlotTypes}
\alias{validSpatialMethods}
\alias{validTrajectoryTrends}
\alias{validUnits}
\alias{validUnitsOfArea}
\alias{validUnitsOfAreaSI}
\alias{validUnitsOfLength}
\alias{validUnitsOfLengthSI}
\alias{validEuropeanUnitsOfLength}
\title{Obtain valid argument inputs}
\usage{
validActivationFunctions()

validAgglomerationMethods()

validAlluvialTypes()

validColorPalettes()

validColorSpectra()

validDeAnalysisMethods()

validDefaultInstructionSlots()

validDimRedMethods()

validDirectoryInstructionSlots()

validDistanceMethods()

validHierarchicalClusterMethods()

validImageClasses()

validModelNames()

validPatternRecognitionMethods()

validPadjMethods()

validPlotTypes(fn_name)

validSpatialMethods()

validTrajectoryTrends()

validUnits()

validUnitsOfArea()

validUnitsOfAreaSI()

validUnitsOfLength()

validUnitsOfLengthSI()

validEuropeanUnitsOfLength(name = T)
}
\value{
Character vectors or named lists of such.
}
\description{
These function simply return valid input options
for recurring arguments.
}
