% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid-input-options.R
\name{validActivationFunctions}
\alias{validActivationFunctions}
\alias{validColorPalettes}
\alias{validColorSpectra}
\alias{validDeAnalysisMethods}
\alias{validDefaultInstructionSlots}
\alias{validDimRedMethods}
\alias{validDirectoryInstructionSlots}
\alias{validHierarchicalClusterMethods}
\alias{validPatternRecognitionMethods}
\alias{validPadjMethods}
\alias{validPlotTypes}
\title{Obtain valid argument inputs}
\usage{
validActivationFunctions()

validColorPalettes()

validColorSpectra()

validDeAnalysisMethods()

validDefaultInstructionSlots()

validDimRedMethods()

validDirectoryInstructionSlots()

validHierarchicalClusterMethods()

validPatternRecognitionMethods()

validPadjMethods()

validPlotTypes(fn_name)
}
\value{
Character vectors or named lists of such.
}
\description{
These function simply return valid input options
for recurring arguments.
}
