% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{variable}
\alias{variable}
\title{variable}
\usage{
variable(variable)
}
\arguments{
\item{variable}{The variable of interest.

 \itemize{
  \item{ \strong{Gene set} as a single character value. Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} as a character vector. If more than one gene is specified the average
  expression of those genes will be calculated and displayed. Must be in \code{getGenes()}}
  \item{ \strong{Feature} as a single character value. Must be in \code{getFeatureNames()}}
  }}
}
\description{
variable
}
