% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{variable_num}
\alias{variable_num}
\title{variable_num}
\usage{
variable_num(variable)
}
\arguments{
\item{variable}{Character value. The numeric variable of interest. Must be inside:

\itemize{
  \item{ \strong{Gene sets} Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} Must be in \code{getGenes()}}
  \item{ \strong{Features} Must be in \code{getFeatureNames(..., of_class = "numeric")}}
  }}
}
\description{
variable_num
}
