% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation-dummies.R
\name{variables_num}
\alias{variables_num}
\title{variables_num}
\usage{
variables_num(variables)
}
\arguments{
\item{variables}{Character vector. The numeric variables of interest. Must be inside:

\itemize{
  \item{ \strong{Gene sets} Must be in \code{getGeneSets()}}
  \item{ \strong{Genes} Must be in \code{getGenes()}}
  \item{ \strong{Features} Must be in \code{getFeatureNames(..., of_class = "numeric")}}
  }}
}
\description{
variables_num
}
