% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-documentation.R
\docType{class}
\name{Visium-class}
\alias{Visium-class}
\alias{Visium}
\title{The \code{Visium} - Class}
\description{
Abstracts the concept of 10X Visium spatial biology experiments.
}
\section{Slots}{

\describe{
\item{\code{capture_area}}{list. List of length two. Provides
standardized measures of the active region where tissue can
be placed on a Visium slide.}

\item{\code{ccd}}{character. The center to center distance
of the barcoded spots provided in a valid \code{SPATA2} spatial measure.}

\item{\code{fiducial_frame}}{list. List of length two. Provides
standardized measures of spots help the sample microscopist see where
to place tissue and are also used by Space Ranger to determine
where the capture area is in an image.}

\item{\code{info}}{list. List of miscellaneous meta data for the method.}

\item{\code{name}}{character. The name of the spatial method.}

\item{\code{observational_unit}}{character. Single word that describes
the observational unit of the experiment.}
}}

\section{Distance measures}{


The vignette on distance measures in SPATA2 has been replaced. Click
\code{\link[=concept_distance_measure]{here}} to read it.
}

