% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devImaging.R
\name{whichTissueSection}
\alias{whichTissueSection}
\title{Tissue section belonging}
\usage{
whichTissueSection(object, id)
}
\arguments{
\item{object}{An object of class \code{spata2} or, in case of S4 generics,
objects of classes for which a method has been defined.}

\item{id}{Character value specifying the ID of the spatial annotation of interest.
If there is only one spatial annotation in the object, the function
will default to using it. However, if there are multiple annotations,
this argument must be explicitly specified to identify the target annotation.}
}
\value{
Character value.
}
\description{
Checks to which tissue section the spatial annotation
belongs. (Only required in case of multiple tissue sections per sample.)
}
