% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{whichTissueSection}
\alias{whichTissueSection}
\title{Tissue section belonging}
\usage{
whichTissueSection(object, id)
}
\arguments{
\item{object}{An object of class \code{\link{SPATA2}} or, in case of S4 generics,
objects of classes for which a method has been defined.}
}
\value{
Character value.
}
\description{
Checks to which tissue section the spatial annotation
belongs. (Only required in case of multiple tissue sections per sample.)
}
